mlr3hyperband: Hyperband for 'mlr3'

Successive Halving (Jamieson and Talwalkar (2016) doi:10.48550/arXiv.1502.07943 ) and Hyperband (Li et al. 2018 doi:10.48550/arXiv.1603.06560 ) optimization algorithm for the mlr3 ecosystem. The implementation in mlr3hyperband features improved scheduling and parallelizes the evaluation of configurations. The package includes tuners for hyperparameter optimization in mlr3tuning and optimizers for black-box optimization in bbotk.

See also

Author

Maintainer: Marc Becker marcbecker@posteo.de (ORCID)

Authors:

Sebastian Gruber gruber_sebastian@t-online.de (ORCID)

Jakob Richter jakob1richter@gmail.com (ORCID)

Julia Moosbauer ju.moosbauer@googlemail.com (ORCID)

Bernd Bischl bernd_bischl@gmx.net (ORCID)