Optimizer Asynchronous Hyperband

OptimizerAhb class that implements the asynchronous hyperband algorithm (AHB). AHB repeatedly runs ASHA (OptimizerAsha) with different minimum budgets in the base stage. Each run of ASHA within AHB is called a bracket. AHB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget (r_min) is increased in each bracket by a factor of eta until the maximum budget is allocated in the base stage. The bracket s = 0 is a random search with full budget. Each ASHA run uses 1 / s_max + 1 of the bbotk::Terminator.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, Rostamizadeh A, Gonina E, Ben-tzur J, Hardt M, Recht B, Talwalkar A (2020). “A System for Massively Parallel Hyperparameter Tuning.” Proceedings of Machine Learning and Systems, 2, 230-246. https://proceedings.mlsys.org/paper/2020/hash/f4b9ec30ad9f68f89b29639786cb62ef-Abstract.html.

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

stage (integer(1))
The stages of each bracket. Starts counting at 0.

bracket (integer(1))
The bracket index. Counts down to 0.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Parallelization

The points are asynchronously evaluated with the future package. To select a parallel backend, use future::plan().

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerAhb

Methods

Public methods

OptimizerAhb$new()

OptimizerAhb$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAhb$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAhb$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("ahb")

# optimize branin function
optimizer$optimize(instance)
#>          x1         x2 fidelity  x_domain        y
#> 1: 3.875469 0.04240107     0.01 <list[3]> 2.928651

# best scoring evaluation
instance$result
#>          x1         x2 fidelity  x_domain        y
#> 1: 3.875469 0.04240107     0.01 <list[3]> 2.928651

# all evaluations
as.data.table(instance$archive)
#>             x1          x2 fidelity asha_id stage bracket           timestamp
#>  1:  5.7181430  9.85095801     0.01       1     0       6 2021-12-18 15:36:06
#>  2:  9.0556827  5.45778435     0.01       2     0       6 2021-12-18 15:36:06
#>  3:  9.0556827  5.45778435     0.02       2     1       6 2021-12-18 15:36:06
#>  4:  6.8454122 11.53101864     0.01       3     0       6 2021-12-18 15:36:06
#>  5:  3.8754692  0.04240107     0.01       4     0       6 2021-12-18 15:36:06
#>  6:  3.8754692  0.04240107     0.02       4     1       6 2021-12-18 15:36:06
#>  7:  3.8754692  0.04240107     0.04       4     2       6 2021-12-18 15:36:06
#>  8:  9.0017025  4.27160250     0.01       5     0       6 2021-12-18 15:36:06
#>  9:  4.7002390  1.22196727     0.02       6     0       5 2021-12-18 15:36:06
#> 10:  3.7288299  6.47402677     0.02       7     0       5 2021-12-18 15:36:06
#> 11:  4.7002390  1.22196727     0.04       6     1       5 2021-12-18 15:36:06
#> 12: -4.8800778 11.48983483     0.02       8     0       5 2021-12-18 15:36:06
#> 13: -1.6287128  1.10290866     0.02       9     0       5 2021-12-18 15:36:06
#> 14: -4.8800778 11.48983483     0.04       8     1       5 2021-12-18 15:36:06
#> 15:  4.7002390  1.22196727     0.08       6     2       5 2021-12-18 15:36:06
#> 16:  8.9329917 11.27497355     0.04      10     0       4 2021-12-18 15:36:06
#> 17: -3.7805743  7.51985783     0.04      11     0       4 2021-12-18 15:36:06
#> 18: -3.7805743  7.51985783     0.08      11     1       4 2021-12-18 15:36:06
#> 19: -1.5313429 12.45554632     0.04      12     0       4 2021-12-18 15:36:06
#> 20: -1.5313429 12.45554632     0.08      12     1       4 2021-12-18 15:36:07
#> 21: -1.5313429 12.45554632     0.16      12     2       4 2021-12-18 15:36:07
#> 22: -1.3027266  3.40314670     0.04      13     0       4 2021-12-18 15:36:07
#> 23: -1.9091629  7.32222417     0.08      14     0       3 2021-12-18 15:36:07
#> 24:  5.6412191  1.38611179     0.08      15     0       3 2021-12-18 15:36:07
#> 25: -1.9091629  7.32222417     0.16      14     1       3 2021-12-18 15:36:07
#> 26: -1.4832958  2.33597855     0.08      16     0       3 2021-12-18 15:36:07
#> 27: -0.8897767  9.45096344     0.08      17     0       3 2021-12-18 15:36:07
#> 28: -0.8897767  9.45096344     0.16      17     1       3 2021-12-18 15:36:07
#> 29: -1.9091629  7.32222417     0.32      14     2       3 2021-12-18 15:36:07
#> 30:  5.2399719  0.46894455     0.16      18     0       2 2021-12-18 15:36:07
#> 31:  6.3636039  9.06811942     0.16      19     0       2 2021-12-18 15:36:07
#> 32:  5.2399719  0.46894455     0.32      18     1       2 2021-12-18 15:36:07
#> 33:  2.5153032  5.39593888     0.16      20     0       2 2021-12-18 15:36:07
#> 34:  2.5153032  5.39593888     0.32      20     1       2 2021-12-18 15:36:07
#> 35:  2.5153032  5.39593888     0.64      20     2       2 2021-12-18 15:36:07
#> 36: -4.6235964 11.85394574     0.16      21     0       2 2021-12-18 15:36:07
#> 37:  0.9005048 13.09374998     0.32      22     0       1 2021-12-18 15:36:07
#> 38:  2.2656542  9.74410909     0.32      23     0       1 2021-12-18 15:36:07
#> 39:  2.2656542  9.74410909     0.64      23     1       1 2021-12-18 15:36:07
#> 40: -1.4898081  2.31664904     0.32      24     0       1 2021-12-18 15:36:07
#> 41: -1.4898081  2.31664904     0.64      24     1       1 2021-12-18 15:36:07
#> 42:  6.9476976  1.26173623     0.32      25     0       1 2021-12-18 15:36:07
#> 43:  6.9476976  1.26173623     0.64      25     1       1 2021-12-18 15:36:07
#> 44:  1.4564808 10.01508528     0.64      26     0       0 2021-12-18 15:36:07
#> 45:  5.0023477  1.48897456     0.64      27     0       0 2021-12-18 15:36:07
#> 46:  9.4826524 12.75209814     0.64      28     0       0 2021-12-18 15:36:08
#> 47:  4.6279431  0.75899213     0.64      29     0       0 2021-12-18 15:36:08
#> 48:  6.4636678  9.93704136     0.64      30     0       0 2021-12-18 15:36:08
#> 49: -2.0219709  9.78315514     0.64      31     0       0 2021-12-18 15:36:08
#> 50:  2.6492795 14.90653346     0.64      32     0       0 2021-12-18 15:36:08
#> 51:  3.6812783 14.90545722     0.64      33     0       0 2021-12-18 15:36:08
#>             x1          x2 fidelity asha_id stage bracket           timestamp
#>     batch_nr    status          y x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1:        1 evaluated 161.264084   5.7181430  9.85095801              0.01
#>  2:        2 evaluated 130.891909   9.0556827  5.45778435              0.01
#>  3:        3 evaluated 129.029722   9.0556827  5.45778435              0.02
#>  4:        4 evaluated 243.466023   6.8454122 11.53101864              0.01
#>  5:        5 evaluated   2.928651   3.8754692  0.04240107              0.01
#>  6:        6 evaluated   2.936175   3.8754692  0.04240107              0.02
#>  7:        7 evaluated   2.952575   3.8754692  0.04240107              0.04
#>  8:        8 evaluated 104.315455   9.0017025  4.27160250              0.01
#>  9:        9 evaluated  13.938301   4.7002390  1.22196727              0.02
#> 10:       10 evaluated  37.707697   3.7288299  6.47402677              0.02
#> 11:       11 evaluated  13.762305   4.7002390  1.22196727              0.04
#> 12:       12 evaluated  20.721305  -4.8800778 11.48983483              0.02
#> 13:       13 evaluated  66.779246  -1.6287128  1.10290866              0.02
#> 14:       14 evaluated  21.011234  -4.8800778 11.48983483              0.04
#> 15:       15 evaluated  13.422027   4.7002390  1.22196727              0.08
#> 16:       16 evaluated 285.262988   8.9329917 11.27497355              0.04
#> 17:       17 evaluated  27.007276  -3.7805743  7.51985783              0.04
#> 18:       18 evaluated  27.578985  -3.7805743  7.51985783              0.08
#> 19:       19 evaluated  25.906438  -1.5313429 12.45554632              0.04
#> 20:       20 evaluated  25.832601  -1.5313429 12.45554632              0.08
#> 21:       21 evaluated  25.685456  -1.5313429 12.45554632              0.16
#> 22:       22 evaluated  34.883350  -1.3027266  3.40314670              0.04
#> 23:       23 evaluated  10.241910  -1.9091629  7.32222417              0.08
#> 24:       24 evaluated  27.809636   5.6412191  1.38611179              0.08
#> 25:       25 evaluated  10.350756  -1.9091629  7.32222417              0.16
#> 26:       26 evaluated  48.129342  -1.4832958  2.33597855              0.08
#> 27:       27 evaluated  20.066983  -0.8897767  9.45096344              0.08
#> 28:       28 evaluated  20.041620  -0.8897767  9.45096344              0.16
#> 29:       29 evaluated  10.573551  -1.9091629  7.32222417              0.32
#> 30:       30 evaluated  17.292691   5.2399719  0.46894455              0.16
#> 31:       31 evaluated 148.764819   6.3636039  9.06811942              0.16
#> 32:       32 evaluated  16.108015   5.2399719  0.46894455              0.32
#> 33:       33 evaluated  11.912911   2.5153032  5.39593888              0.16
#> 34:       34 evaluated  11.292852   2.5153032  5.39593888              0.32
#> 35:       35 evaluated  10.114218   2.5153032  5.39593888              0.64
#> 36:       36 evaluated  15.252758  -4.6235964 11.85394574              0.16
#> 37:       37 evaluated  87.830142   0.9005048 13.09374998              0.32
#> 38:       38 evaluated  53.356420   2.2656542  9.74410909              0.32
#> 39:       39 evaluated  51.071927   2.2656542  9.74410909              0.64
#> 40:       40 evaluated  49.096064  -1.4898081  2.31664904              0.32
#> 41:       41 evaluated  49.980433  -1.4898081  2.31664904              0.64
#> 42:       42 evaluated  28.888460   6.9476976  1.26173623              0.32
#> 43:       43 evaluated  20.876004   6.9476976  1.26173623              0.64
#> 44:       44 evaluated  48.739283   1.4564808 10.01508528              0.64
#> 45:       45 evaluated  13.996760   5.0023477  1.48897456              0.64
#> 46:       46 evaluated 181.719979   9.4826524 12.75209814              0.64
#> 47:       47 evaluated   9.206689   4.6279431  0.75899213              0.64
#> 48:       48 evaluated 126.178445   6.4636678  9.93704136              0.64
#> 49:       49 evaluated   5.847158  -2.0219709  9.78315514              0.64
#> 50:       50 evaluated 157.013187   2.6492795 14.90653346              0.64
#> 51:       51  proposed         NA          NA          NA                NA
#>     batch_nr    status          y x_domain_x1 x_domain_x2 x_domain_fidelity