Optimizer Asynchronous Successive Halving

OptimizerAsha class that implements the asynchronous successive halving algorithm (ASHA). ASHA parallelizes SHA (OptimizerSuccessiveHalving) by promoting candidates to the next stage as soon as possible instead of waiting for all candidates in the stage to finish. ASHA starts with sampling a candidate point for each available worker. When an evaluation finishes and the worker is available again, ASHA checks the stages from top to bottom for promotable candidates. Promotions are possible when the evaluated candidates belong to the top 1 / eta of each stage. If no promotions are possible, a new candidate is sampled and added to the base stage, which increases the number of possible promotions for all stages.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stage.

Usage,NULL

Source

Li L, Jamieson K, Rostamizadeh A, Gonina E, Ben-tzur J, Hardt M, Recht B, Talwalkar A (2020). “A System for Massively Parallel Hyperparameter Tuning.” Proceedings of Machine Learning and Systems, 2, 230-246. https://proceedings.mlsys.org/paper/2020/hash/f4b9ec30ad9f68f89b29639786cb62ef-Abstract.html.

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1/eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional column that is specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Parallelization

The points are asynchronously evaluated with the future package. To select a parallel backend, use future::plan().

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerAsha

Methods

Public methods

OptimizerAsha$new()

OptimizerAsha$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAsha$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAsha$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("asha")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.358618 13.88597     0.16 <list[3]> 4.747552

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.358618 13.88597     0.16 <list[3]> 4.747552

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity asha_id stage           timestamp batch_nr
#>  1:  6.01706670  5.8179636     0.01       1     0 2021-12-01 16:59:34        1
#>  2:  7.39859744  5.9832796     0.01       2     0 2021-12-01 16:59:34        2
#>  3:  6.01706670  5.8179636     0.02       1     1 2021-12-01 16:59:34        3
#>  4:  7.93187169  8.7873053     0.01       3     0 2021-12-01 16:59:34        4
#>  5:  1.14735275  4.0804220     0.01       4     0 2021-12-01 16:59:35        5
#>  6:  1.14735275  4.0804220     0.02       4     1 2021-12-01 16:59:35        6
#>  7:  1.14735275  4.0804220     0.04       4     2 2021-12-01 16:59:35        7
#>  8:  4.11656716  4.4628912     0.01       5     0 2021-12-01 16:59:35        8
#>  9:  4.11656716  4.4628912     0.02       5     1 2021-12-01 16:59:35        9
#> 10:  4.02605964  3.3794674     0.01       6     0 2021-12-01 16:59:35       10
#> 11:  4.02605964  3.3794674     0.02       6     1 2021-12-01 16:59:35       11
#> 12:  4.02605964  3.3794674     0.04       6     2 2021-12-01 16:59:35       12
#> 13:  1.14735275  4.0804220     0.08       4     3 2021-12-01 16:59:35       13
#> 14:  3.64999424 13.7995725     0.01       7     0 2021-12-01 16:59:35       14
#> 15: -3.50956447  2.1008864     0.01       8     0 2021-12-01 16:59:35       15
#> 16: -3.35861827 13.8859695     0.01       9     0 2021-12-01 16:59:35       16
#> 17: -3.35861827 13.8859695     0.02       9     1 2021-12-01 16:59:35       17
#> 18: -3.35861827 13.8859695     0.04       9     2 2021-12-01 16:59:35       18
#> 19: -3.35861827 13.8859695     0.08       9     3 2021-12-01 16:59:35       19
#> 20: -3.35861827 13.8859695     0.16       9     4 2021-12-01 16:59:35       20
#> 21:  7.20215972 13.1997050     0.01      10     0 2021-12-01 16:59:35       21
#> 22:  0.09175539  9.2838008     0.01      11     0 2021-12-01 16:59:35       22
#> 23:  0.09175539  9.2838008     0.02      11     1 2021-12-01 16:59:35       23
#> 24:  3.76641623 10.5969075     0.01      12     0 2021-12-01 16:59:35       24
#> 25:  8.74602096 13.6153664     0.01      13     0 2021-12-01 16:59:35       25
#> 26: -0.02878775 14.0362618     0.01      14     0 2021-12-01 16:59:35       26
#> 27: -0.02878775 14.0362618     0.02      14     1 2021-12-01 16:59:35       27
#> 28:  7.21676640 13.3978111     0.01      15     0 2021-12-01 16:59:36       28
#> 29: -1.29316511  3.9920024     0.01      16     0 2021-12-01 16:59:36       29
#> 30: -1.29316511  3.9920024     0.02      16     1 2021-12-01 16:59:36       30
#> 31:  4.11656716  4.4628912     0.04       5     2 2021-12-01 16:59:36       31
#> 32: -3.92044214  6.2307755     0.01      17     0 2021-12-01 16:59:36       32
#> 33: -3.92044214  6.2307755     0.02      17     1 2021-12-01 16:59:36       33
#> 34:  1.44950359  4.1956853     0.01      18     0 2021-12-01 16:59:36       34
#> 35:  1.44950359  4.1956853     0.02      18     1 2021-12-01 16:59:36       35
#> 36:  1.44950359  4.1956853     0.04      18     2 2021-12-01 16:59:36       36
#> 37:  1.44950359  4.1956853     0.08      18     3 2021-12-01 16:59:36       37
#> 38:  1.70856537 13.7271465     0.01      19     0 2021-12-01 16:59:36       38
#> 39: -4.60523270 10.5393613     0.01      20     0 2021-12-01 16:59:36       39
#> 40: -4.60523270 10.5393613     0.02      20     1 2021-12-01 16:59:36       40
#> 41: -4.60523270 10.5393613     0.04      20     2 2021-12-01 16:59:36       41
#> 42:  9.34192007  0.8227779     0.01      21     0 2021-12-01 16:59:36       42
#> 43:  9.34192007  0.8227779     0.02      21     1 2021-12-01 16:59:36       43
#> 44:  5.29714439  4.7560232     0.01      22     0 2021-12-01 16:59:36       44
#> 45:  5.29714439  4.7560232     0.02      22     1 2021-12-01 16:59:36       45
#> 46:  9.37779943  3.1873698     0.01      23     0 2021-12-01 16:59:36       46
#> 47:  2.47898189  8.0837232     0.01      24     0 2021-12-01 16:59:36       47
#> 48:  2.47898189  8.0837232     0.02      24     1 2021-12-01 16:59:36       48
#> 49: -1.29316511  3.9920024     0.04      16     2 2021-12-01 16:59:36       49
#> 50:  4.23671227 11.4653536     0.01      25     0 2021-12-01 16:59:36       50
#> 51:  5.83498431  0.8320983     0.01      26     0 2021-12-01 16:59:37       51
#>              x1         x2 fidelity asha_id stage           timestamp batch_nr
#>        status          y x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: evaluated  88.180500  6.01706670   5.8179636              0.01
#>  2: evaluated 116.359086  7.39859744   5.9832796              0.01
#>  3: evaluated  87.580691  6.01706670   5.8179636              0.02
#>  4: evaluated 191.833076  7.93187169   8.7873053              0.01
#>  5: evaluated  13.963285  1.14735275   4.0804220              0.01
#>  6: evaluated  13.963637  1.14735275   4.0804220              0.02
#>  7: evaluated  13.964353  1.14735275   4.0804220              0.04
#>  8: evaluated  24.889316  4.11656716   4.4628912              0.01
#>  9: evaluated  24.736983  4.11656716   4.4628912              0.02
#> 10: evaluated  14.791102  4.02605964   3.3794674              0.01
#> 11: evaluated  14.684453  4.02605964   3.3794674              0.02
#> 12: evaluated  14.472732  4.02605964   3.3794674              0.04
#> 13: evaluated  13.965825  1.14735275   4.0804220              0.08
#> 14: evaluated 176.026265  3.64999424  13.7995725              0.01
#> 15: evaluated  98.192101 -3.50956447   2.1008864              0.01
#> 16: evaluated   5.463447 -3.35861827  13.8859695              0.01
#> 17: evaluated   5.413939 -3.35861827  13.8859695              0.02
#> 18: evaluated   5.315687 -3.35861827  13.8859695              0.04
#> 19: evaluated   5.122237 -3.35861827  13.8859695              0.08
#> 20: evaluated   4.747552 -3.35861827  13.8859695              0.16
#> 21: evaluated 308.118525  7.20215972  13.1997050              0.01
#> 22: evaluated  31.323739  0.09175539   9.2838008              0.01
#> 23: evaluated  31.323681  0.09175539   9.2838008              0.02
#> 24: evaluated 105.501686  3.76641623  10.5969075              0.01
#> 25: evaluated 372.172578  8.74602096  13.6153664              0.01
#> 26: evaluated  83.444940 -0.02878775  14.0362618              0.01
#> 27: evaluated  83.444927 -0.02878775  14.0362618              0.02
#> 28: evaluated 315.404032  7.21676640  13.3978111              0.01
#> 29: evaluated  29.578214 -1.29316511   3.9920024              0.01
#> 30: evaluated  29.591985 -1.29316511   3.9920024              0.02
#> 31: evaluated  24.434040  4.11656716   4.4628912              0.04
#> 32: evaluated  45.062246 -3.92044214   6.2307755              0.01
#> 33: evaluated  45.261452 -3.92044214   6.2307755              0.02
#> 34: evaluated  11.354729  1.44950359   4.1956853              0.01
#> 35: evaluated  11.352888  1.44950359   4.1956853              0.02
#> 36: evaluated  11.349232  1.44950359   4.1956853              0.04
#> 37: evaluated  11.342026  1.44950359   4.1956853              0.08
#> 38: evaluated 115.975649  1.70856537  13.7271465              0.01
#> 39: evaluated  20.739673 -4.60523270  10.5393613              0.01
#> 40: evaluated  20.885621 -4.60523270  10.5393613              0.02
#> 41: evaluated  21.180217 -4.60523270  10.5393613              0.04
#> 42: evaluated  50.227266  9.34192007   0.8227779              0.01
#> 43: evaluated  49.003192  9.34192007   0.8227779              0.02
#> 44: evaluated  55.492383  5.29714439   4.7560232              0.01
#> 45: evaluated  55.137389  5.29714439   4.7560232              0.02
#> 46: evaluated  89.863780  9.37779943   3.1873698              0.01
#> 47: evaluated  36.578059  2.47898189   8.0837232              0.01
#> 48: evaluated  36.506274  2.47898189   8.0837232              0.02
#> 49: evaluated  29.619544 -1.29316511   3.9920024              0.04
#> 50: evaluated 141.709722  4.23671227  11.4653536              0.01
#> 51:  proposed         NA          NA          NA                NA
#>        status          y x_domain_x1 x_domain_x2 x_domain_fidelity