Tuner Asynchronous Successive Halving

TunerAsha class that implements the asynchronous successive halving algorithm (ASHA). ASHA parallelizes SHA (TunerSuccessiveHalving) by promoting candidates to the next stage as soon as possible instead of waiting for all candidates in the stage to finish. ASHA starts with sampling a candidate hyperparameter configuration for each available worker. When an evaluation finishes and the worker is available again, ASHA checks the stages from top to bottom for promotable candidates. Promotions are possible when the evaluated candidates belong to the top 1 / eta of each stage. If no promotions are possible, a new candidate is sampled and added to the base stage, which increases the number of possible promotions for all stages.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stage.

Usage,NULL

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta candidates are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional column that is specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Parallelization

The hyperparameter configurations are asynchronously evaluated with the future package. The resampling of each candidate is send to an available worker. To select a parallel backend, use future::plan().

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerAsha

Methods

Public methods

TunerAsha$new()

TunerAsha$optimize()

TunerAsha$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$print()

Method new()

Creates a new instance of this R6 class.

Usage

TunerAsha$new()

Method optimize()

Performs the tuning on a TuningInstanceSingleCrit / TuningInstanceMultiCrit until termination. The single evaluations and the final results will be written into the ArchiveTuning that resides in the TuningInstanceSingleCrit/TuningInstanceMultiCrit. The final result is returned.

Usage

TunerAsha$optimize(inst)

Arguments

inst

(TuningInstanceSingleCrit | TuningInstanceMultiCrit).

Returns

data.table::data.table.

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerAsha$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if(requireNamespace("xgboost")) {
  library(mlr3learners)

  # define hyperparameter and budget parameter
  search_space = ps(
    nrounds = p_int(lower = 1, upper = 16, tags = "budget"),
    eta = p_dbl(lower = 0, upper = 1),
    booster = p_fct(levels = c("gbtree", "gblinear", "dart"))
  )

  # \donttest{
  # hyperparameter tuning on the pima indians diabetes data set
  instance = tune(
    method = "asha",
    task = tsk("pima"),
    learner = lrn("classif.xgboost", eval_metric = "logloss"),
    resampling = rsmp("cv", folds = 3),
    measures = msr("classif.ce"),
    search_space = search_space,
    term_evals = 100
  )

  # best performing hyperparameter configuration
  instance$result
  # }
}
#>    nrounds        eta booster learner_param_vals  x_domain classif.ce
#> 1:       8 0.08163538  gbtree          <list[6]> <list[3]>  0.2304688