Best points w.r.t. non dominated sorting with hypervolume contrib.

Select best subset of points by non dominated sorting with hypervolume contribution for tie breaking. Works on an arbitrary dimension of size two or higher.

Value

Vector of indices of selected points

Parameters

points :: matrix()
Numeric matrix with each column corresponding to a point

n_select :: integer(1L)
Amount of points to select

ref_point :: integer()
Reference point for hypervolume

minimize :: logical()
Should the ranking be based on minimization? (Single bool for all dimensions, or vector of bools with each entry corresponding to each dimension)