Optimizer Asynchronous Hyperband

OptimizerAhb class that implements the asynchronous hyperband algorithm (AHB). AHB repeatedly runs ASHA (OptimizerAsha) with different minimum budgets in the base stage. Each run of ASHA within AHB is called a bracket. AHB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget (r_min) is increased in each bracket by a factor of eta until the maximum budget is allocated in the base stage. The bracket s = 0 is a random search with full budget. Each ASHA run uses 1 / s_max + 1 of the bbotk::Terminator.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Usage,NULL

Source

Li L, Jamieson K, Rostamizadeh A, Gonina E, Ben-tzur J, Hardt M, Recht B, Talwalkar A (2020). “A System for Massively Parallel Hyperparameter Tuning.” Proceedings of Machine Learning and Systems, 2, 230-246. https://proceedings.mlsys.org/paper/2020/hash/f4b9ec30ad9f68f89b29639786cb62ef-Abstract.html.

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

stage (integer(1))
The stages of each bracket. Starts counting at 0.

bracket (integer(1))
The bracket index. Counts down to 0.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Parallelization

The points are asynchronously evaluated with the future package. To select a parallel backend, use future::plan().

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerAhb

Methods

Public methods

OptimizerAhb$new()

OptimizerAhb$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAhb$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAhb$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("ahb")

# optimize branin function
optimizer$optimize(instance)
#>        x1       x2 fidelity  x_domain        y
#> 1: 3.8269 1.369495     0.64 <list[3]> 2.574928

# best scoring evaluation
instance$result
#>        x1       x2 fidelity  x_domain        y
#> 1: 3.8269 1.369495     0.64 <list[3]> 2.574928

# all evaluations
as.data.table(instance$archive)
#>              x1        x2 fidelity asha_id stage bracket           timestamp
#>  1: -4.75010460  3.728186     0.01       1     0       6 2021-12-01 16:59:31
#>  2:  8.43796537 10.243439     0.01       2     0       6 2021-12-01 16:59:31
#>  3: -4.75010460  3.728186     0.02       1     1       6 2021-12-01 16:59:31
#>  4:  7.87753774  1.558615     0.01       3     0       6 2021-12-01 16:59:31
#>  5:  7.87753774  1.558615     0.02       3     1       6 2021-12-01 16:59:31
#>  6:  7.87753774  1.558615     0.04       3     2       6 2021-12-01 16:59:31
#>  7: -0.57428632  4.352620     0.01       4     0       6 2021-12-01 16:59:31
#>  8: -0.57428632  4.352620     0.02       4     1       6 2021-12-01 16:59:31
#>  9: -2.50668659 13.880578     0.02       5     0       5 2021-12-01 16:59:31
#> 10:  7.23688172  1.451648     0.02       6     0       5 2021-12-01 16:59:31
#> 11: -2.50668659 13.880578     0.04       5     1       5 2021-12-01 16:59:31
#> 12:  0.25973629 10.942364     0.02       7     0       5 2021-12-01 16:59:31
#> 13:  2.59892536  3.225662     0.02       8     0       5 2021-12-01 16:59:31
#> 14:  2.59892536  3.225662     0.04       8     1       5 2021-12-01 16:59:31
#> 15:  2.59892536  3.225662     0.08       8     2       5 2021-12-01 16:59:31
#> 16:  5.42402088  5.917460     0.04       9     0       4 2021-12-01 16:59:32
#> 17:  5.96713707  8.825012     0.04      10     0       4 2021-12-01 16:59:32
#> 18:  5.42402088  5.917460     0.08       9     1       4 2021-12-01 16:59:32
#> 19:  8.83497113  8.517892     0.04      11     0       4 2021-12-01 16:59:32
#> 20:  1.63350394 12.814950     0.04      12     0       4 2021-12-01 16:59:32
#> 21:  1.63350394 12.814950     0.08      12     1       4 2021-12-01 16:59:32
#> 22:  5.42402088  5.917460     0.16       9     2       4 2021-12-01 16:59:32
#> 23: -3.11248461  7.643688     0.08      13     0       3 2021-12-01 16:59:32
#> 24:  5.91439338  1.482866     0.08      14     0       3 2021-12-01 16:59:32
#> 25: -3.11248461  7.643688     0.16      13     1       3 2021-12-01 16:59:32
#> 26:  8.78394718  2.254798     0.08      15     0       3 2021-12-01 16:59:32
#> 27:  0.45144188  3.275787     0.08      16     0       3 2021-12-01 16:59:32
#> 28:  0.45144188  3.275787     0.16      16     1       3 2021-12-01 16:59:32
#> 29: -3.11248461  7.643688     0.32      13     2       3 2021-12-01 16:59:32
#> 30:  2.84033394  7.060653     0.16      17     0       2 2021-12-01 16:59:32
#> 31:  9.52042925  3.872838     0.16      18     0       2 2021-12-01 16:59:32
#> 32:  2.84033394  7.060653     0.32      17     1       2 2021-12-01 16:59:32
#> 33:  0.87610711  8.992980     0.16      19     0       2 2021-12-01 16:59:32
#> 34: -2.53017710 12.030456     0.16      20     0       2 2021-12-01 16:59:32
#> 35: -2.53017710 12.030456     0.32      20     1       2 2021-12-01 16:59:32
#> 36: -2.53017710 12.030456     0.64      20     2       2 2021-12-01 16:59:32
#> 37:  3.82690010  1.369495     0.32      21     0       1 2021-12-01 16:59:32
#> 38:  5.94855316 10.726180     0.32      22     0       1 2021-12-01 16:59:32
#> 39:  3.82690010  1.369495     0.64      21     1       1 2021-12-01 16:59:33
#> 40:  7.06821408  1.842428     0.32      23     0       1 2021-12-01 16:59:33
#> 41:  9.04562815 13.867516     0.32      24     0       1 2021-12-01 16:59:33
#> 42:  7.06821408  1.842428     0.64      23     1       1 2021-12-01 16:59:33
#> 43: -0.42380069 12.897621     0.32      25     0       1 2021-12-01 16:59:33
#> 44:  0.05575769 14.039054     0.64      26     0       0 2021-12-01 16:59:33
#> 45: -4.34115474  5.939460     0.64      27     0       0 2021-12-01 16:59:33
#> 46:  9.69549027  8.626063     0.64      28     0       0 2021-12-01 16:59:33
#> 47:  5.62913553 12.450384     0.64      29     0       0 2021-12-01 16:59:33
#> 48:  2.34537782 10.252278     0.64      30     0       0 2021-12-01 16:59:33
#> 49: -4.12537611 12.623753     0.64      31     0       0 2021-12-01 16:59:33
#> 50: -4.72551135  7.259882     0.64      32     0       0 2021-12-01 16:59:33
#> 51:  7.75202156  2.719907     0.64      33     0       0 2021-12-01 16:59:33
#>              x1        x2 fidelity asha_id stage bracket           timestamp
#>     batch_nr    status          y x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1:        1 evaluated 120.883313 -4.75010460    3.728186              0.01
#>  2:        2 evaluated 245.693074  8.43796537   10.243439              0.01
#>  3:        3 evaluated 121.358238 -4.75010460    3.728186              0.02
#>  4:        4 evaluated  48.499412  7.87753774    1.558615              0.01
#>  5:        5 evaluated  47.730920  7.87753774    1.558615              0.02
#>  6:        6 evaluated  46.217042  7.87753774    1.558615              0.04
#>  7:        7 evaluated  24.673532 -0.57428632    4.352620              0.01
#>  8:        8 evaluated  24.675229 -0.57428632    4.352620              0.02
#>  9:        9 evaluated  15.922953 -2.50668659   13.880578              0.02
#> 10:       10 evaluated  44.032512  7.23688172    1.451648              0.02
#> 11:       11 evaluated  15.830238 -2.50668659   13.880578              0.04
#> 12:       12 evaluated  47.941537  0.25973629   10.942364              0.02
#> 13:       13 evaluated   3.102981  2.59892536    3.225662              0.02
#> 14:       14 evaluated   3.072057  2.59892536    3.225662              0.04
#> 15:       15 evaluated   3.011304  2.59892536    3.225662              0.08
#> 16:       16 evaluated  73.632840  5.42402088    5.917460              0.04
#> 17:       17 evaluated 143.235355  5.96713707    8.825012              0.04
#> 18:       18 evaluated  71.864126  5.42402088    5.917460              0.08
#> 19:       19 evaluated 197.709835  8.83497113    8.517892              0.04
#> 20:       20 evaluated  96.376362  1.63350394   12.814950              0.04
#> 21:       21 evaluated  96.177392  1.63350394   12.814950              0.08
#> 22:       22 evaluated  68.409790  5.42402088    5.917460              0.16
#> 23:       23 evaluated  13.872413 -3.11248461    7.643688              0.08
#> 24:       24 evaluated  31.881958  5.91439338    1.482866              0.08
#> 25:       25 evaluated  14.447306 -3.11248461    7.643688              0.16
#> 26:       26 evaluated  56.558145  8.78394718    2.254798              0.08
#> 27:       27 evaluated  22.693538  0.45144188    3.275787              0.08
#> 28:       28 evaluated  22.700106  0.45144188    3.275787              0.16
#> 29:       29 evaluated  15.633130 -3.11248461    7.643688              0.32
#> 30:       30 evaluated  28.043866  2.84033394    7.060653              0.16
#> 31:       31 evaluated  80.179758  9.52042925    3.872838              0.16
#> 32:       32 evaluated  26.713795  2.84033394    7.060653              0.32
#> 33:       33 evaluated  35.092462  0.87610711    8.992980              0.16
#> 34:       34 evaluated   5.076238 -2.53017710   12.030456              0.16
#> 35:       35 evaluated   4.735544 -2.53017710   12.030456              0.32
#> 36:       36 evaluated   4.117106 -2.53017710   12.030456              0.64
#> 37:       37 evaluated   2.884061  3.82690010    1.369495              0.32
#> 38:       38 evaluated 163.755867  5.94855316   10.726180              0.32
#> 39:       39 evaluated   2.574928  3.82690010    1.369495              0.64
#> 40:       40 evaluated  33.074125  7.06821408    1.842428              0.32
#> 41:       41 evaluated 298.909982  9.04562815   13.867516              0.32
#> 42:       42 evaluated  22.728148  7.06821408    1.842428              0.64
#> 43:       43 evaluated  57.343223 -0.42380069   12.897621              0.32
#> 44:       44 evaluated  85.643539  0.05575769   14.039054              0.64
#> 45:       45 evaluated  82.656668 -4.34115474    5.939460              0.64
#> 46:       46 evaluated  87.188037  9.69549027    8.626063              0.64
#> 47:       47 evaluated 172.789337  5.62913553   12.450384              0.64
#> 48:       48 evaluated  59.121984  2.34537782   10.252278              0.64
#> 49:       49 evaluated   7.015998 -4.12537611   12.623753              0.64
#> 50:       50 evaluated  79.712553 -4.72551135    7.259882              0.64
#> 51:       51  proposed         NA          NA          NA                NA
#>     batch_nr    status          y x_domain_x1 x_domain_x2 x_domain_fidelity