Optimizer Asynchronous Hyperband

OptimizerAhb class that implements the asynchronous hyperband algorithm (AHB). AHB repeatedly runs ASHA (OptimizerAsha) with different minimum budgets in the base stage. Each run of ASHA within AHB is called a bracket. AHB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget (r_min) is increased in each bracket by a factor of eta until the maximum budget is allocated in the base stage. The bracket s = 0 is a random search with full budget. Each ASHA run uses 1 / s_max + 1 of the bbotk::Terminator.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, Rostamizadeh A, Gonina E, Ben-tzur J, Hardt M, Recht B, Talwalkar A (2020). “A System for Massively Parallel Hyperparameter Tuning.” Proceedings of Machine Learning and Systems, 2, 230-246. https://proceedings.mlsys.org/paper/2020/hash/f4b9ec30ad9f68f89b29639786cb62ef-Abstract.html.

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

stage (integer(1))
The stages of each bracket. Starts counting at 0.

bracket (integer(1))
The bracket index. Counts down to 0.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Parallelization

The points are asynchronously evaluated with the future package. To select a parallel backend, use future::plan().

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerAhb

Methods

Public methods

OptimizerAhb$new()

OptimizerAhb$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAhb$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAhb$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("ahb")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -4.546681 12.97795     0.02 <list[3]> 9.231485

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -4.546681 12.97795     0.02 <list[3]> 9.231485

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity asha_id stage bracket           timestamp
#>  1:  8.6896931  1.4646145     0.01       1     0       6 2021-12-06 14:19:06
#>  2: -2.6037906  1.1101729     0.01       2     0       6 2021-12-06 14:19:06
#>  3:  8.6896931  1.4646145     0.02       1     1       6 2021-12-06 14:19:06
#>  4:  8.8259828  0.4190953     0.01       3     0       6 2021-12-06 14:19:06
#>  5:  8.8259828  0.4190953     0.02       3     1       6 2021-12-06 14:19:07
#>  6:  8.8259828  0.4190953     0.04       3     2       6 2021-12-06 14:19:07
#>  7:  0.5273539  9.0772593     0.01       4     0       6 2021-12-06 14:19:07
#>  8:  0.5273539  9.0772593     0.02       4     1       6 2021-12-06 14:19:07
#>  9: -4.5466805 12.9779473     0.02       5     0       5 2021-12-06 14:19:07
#> 10:  5.7091157  8.9337847     0.02       6     0       5 2021-12-06 14:19:07
#> 11: -4.5466805 12.9779473     0.04       5     1       5 2021-12-06 14:19:07
#> 12: -0.6521140 14.0290957     0.02       7     0       5 2021-12-06 14:19:07
#> 13:  0.1580415  4.8964100     0.02       8     0       5 2021-12-06 14:19:07
#> 14:  0.1580415  4.8964100     0.04       8     1       5 2021-12-06 14:19:07
#> 15: -4.5466805 12.9779473     0.08       5     2       5 2021-12-06 14:19:07
#> 16:  0.9434334  7.5097800     0.04       9     0       4 2021-12-06 14:19:07
#> 17: -4.8994303 11.4905118     0.04      10     0       4 2021-12-06 14:19:07
#> 18: -4.8994303 11.4905118     0.08      10     1       4 2021-12-06 14:19:07
#> 19:  2.7809088 11.7410413     0.04      11     0       4 2021-12-06 14:19:07
#> 20: -1.1213909 10.9168182     0.04      12     0       4 2021-12-06 14:19:07
#> 21: -1.1213909 10.9168182     0.08      12     1       4 2021-12-06 14:19:07
#> 22: -4.8994303 11.4905118     0.16      10     2       4 2021-12-06 14:19:07
#> 23: -4.5215658  7.8898362     0.08      13     0       3 2021-12-06 14:19:07
#> 24:  3.7530243 12.9148070     0.08      14     0       3 2021-12-06 14:19:07
#> 25: -4.5215658  7.8898362     0.16      13     1       3 2021-12-06 14:19:07
#> 26: -0.8041102 13.6990447     0.08      15     0       3 2021-12-06 14:19:07
#> 27:  0.5863778  1.1931760     0.08      16     0       3 2021-12-06 14:19:07
#> 28:  0.5863778  1.1931760     0.16      16     1       3 2021-12-06 14:19:08
#> 29:  0.5863778  1.1931760     0.32      16     2       3 2021-12-06 14:19:08
#> 30: -3.3685436  5.2164906     0.16      17     0       2 2021-12-06 14:19:08
#> 31: -3.5127703  7.7005640     0.16      18     0       2 2021-12-06 14:19:08
#> 32: -3.5127703  7.7005640     0.32      18     1       2 2021-12-06 14:19:08
#> 33:  5.8083297  1.2786812     0.16      19     0       2 2021-12-06 14:19:08
#> 34:  6.8981711  4.4966530     0.16      20     0       2 2021-12-06 14:19:08
#> 35:  5.8083297  1.2786812     0.32      19     1       2 2021-12-06 14:19:08
#> 36: -3.5127703  7.7005640     0.64      18     2       2 2021-12-06 14:19:08
#> 37:  8.9428663  2.4391640     0.32      21     0       1 2021-12-06 14:19:08
#> 38:  9.6191835  7.3192517     0.32      22     0       1 2021-12-06 14:19:08
#> 39:  8.9428663  2.4391640     0.64      21     1       1 2021-12-06 14:19:08
#> 40:  6.1911334  2.7854952     0.32      23     0       1 2021-12-06 14:19:08
#> 41:  9.0920242 13.2104149     0.32      24     0       1 2021-12-06 14:19:08
#> 42:  6.1911334  2.7854952     0.64      23     1       1 2021-12-06 14:19:08
#> 43:  3.8205464  9.2996655     0.32      25     0       1 2021-12-06 14:19:08
#> 44: -3.1751692  5.2067139     0.64      26     0       0 2021-12-06 14:19:08
#> 45: -0.6824385 10.8423853     0.64      27     0       0 2021-12-06 14:19:08
#> 46:  9.4152579 13.7200737     0.64      28     0       0 2021-12-06 14:19:08
#> 47:  0.7370639  0.3981352     0.64      29     0       0 2021-12-06 14:19:08
#> 48: -4.6148892 12.1568956     0.64      30     0       0 2021-12-06 14:19:08
#> 49: -4.9142990 12.4445296     0.64      31     0       0 2021-12-06 14:19:08
#> 50:  9.5265433 11.7168687     0.64      32     0       0 2021-12-06 14:19:08
#> 51:  2.3796559  0.2543819     0.64      33     0       0 2021-12-06 14:19:08
#>             x1         x2 fidelity asha_id stage bracket           timestamp
#>     batch_nr    status          y x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1:        1 evaluated  52.093735   8.6896931   1.4646145              0.01
#>  2:        2 evaluated  87.103814  -2.6037906   1.1101729              0.01
#>  3:        3 evaluated  51.039955   8.6896931   1.4646145              0.02
#>  4:        4 evaluated  39.458934   8.8259828   0.4190953              0.01
#>  5:        5 evaluated  38.512346   8.8259828   0.4190953              0.02
#>  6:        6 evaluated  36.655577   8.8259828   0.4190953              0.04
#>  7:        7 evaluated  33.571416   0.5273539   9.0772593              0.01
#>  8:        8 evaluated  33.569242   0.5273539   9.0772593              0.02
#>  9:        9 evaluated   9.231485  -4.5466805  12.9779473              0.02
#> 10:       10 evaluated 139.144212   5.7091157   8.9337847              0.02
#> 11:       11 evaluated   9.307861  -4.5466805  12.9779473              0.04
#> 12:       12 evaluated  66.323748  -0.6521140  14.0290957              0.02
#> 13:       13 evaluated  20.209778   0.1580415   4.8964100              0.02
#> 14:       14 evaluated  20.209864   0.1580415   4.8964100              0.04
#> 15:       15 evaluated   9.470868  -4.5466805  12.9779473              0.08
#> 16:       16 evaluated  24.527473   0.9434334   7.5097800              0.04
#> 17:       17 evaluated  21.418793  -4.8994303  11.4905118              0.04
#> 18:       18 evaluated  22.024041  -4.8994303  11.4905118              0.08
#> 19:       19 evaluated  99.231048   2.7809088  11.7410413              0.04
#> 20:       20 evaluated  23.721642  -1.1213909  10.9168182              0.04
#> 21:       21 evaluated  23.690578  -1.1213909  10.9168182              0.08
#> 22:       22 evaluated  23.289855  -4.8994303  11.4905118              0.16
#> 23:       23 evaluated  44.983402  -4.5215658   7.8898362              0.08
#> 24:       24 evaluated 155.010472   3.7530243  12.9148070              0.08
#> 25:       25 evaluated  46.994642  -4.5215658   7.8898362              0.16
#> 26:       26 evaluated  57.560331  -0.8041102  13.6990447              0.08
#> 27:       27 evaluated  33.101880   0.5863778   1.1931760              0.08
#> 28:       28 evaluated  33.123268   0.5863778   1.1931760              0.16
#> 29:       29 evaluated  33.166089   0.5863778   1.1931760              0.32
#> 30:       30 evaluated  44.965429  -3.3685436   5.2164906              0.16
#> 31:       31 evaluated  20.834177  -3.5127703   7.7005640              0.16
#> 32:       32 evaluated  22.629417  -3.5127703   7.7005640              0.32
#> 33:       33 evaluated  27.531020   5.8083297   1.2786812              0.16
#> 34:       34 evaluated  71.503422   6.8981711   4.4966530              0.16
#> 35:       35 evaluated  24.585234   5.8083297   1.2786812              0.32
#> 36:       36 evaluated  26.453776  -3.5127703   7.7005640              0.64
#> 37:       37 evaluated  34.887792   8.9428663   2.4391640              0.32
#> 38:       38 evaluated 120.861206   9.6191835   7.3192517              0.32
#> 39:       39 evaluated  11.858320   8.9428663   2.4391640              0.64
#> 40:       40 evaluated  37.998003   6.1911334   2.7854952              0.32
#> 41:       41 evaluated 277.249033   9.0920242  13.2104149              0.32
#> 42:       42 evaluated  28.969270   6.1911334   2.7854952              0.64
#> 43:       43 evaluated  74.559395   3.8205464   9.2996655              0.32
#> 44:       44 evaluated  46.455580  -3.1751692   5.2067139              0.64
#> 45:       45 evaluated  31.236870  -0.6824385  10.8423853              0.64
#> 46:       46 evaluated 209.038645   9.4152579  13.7200737              0.64
#> 47:       47 evaluated  37.174995   0.7370639   0.3981352              0.64
#> 48:       48 evaluated  19.130022  -4.6148892  12.1568956              0.64
#> 49:       49 evaluated  25.082613  -4.9142990  12.4445296              0.64
#> 50:       50 evaluated 154.750433   9.5265433  11.7168687              0.64
#> 51:       51  proposed         NA          NA          NA                NA
#>     batch_nr    status          y x_domain_x1 x_domain_x2 x_domain_fidelity