Optimizer Asynchronous Hyperband

OptimizerAhb class that implements the asynchronous hyperband algorithm. Asynchronous hyperband (AHB) repeatedly runs ASHA (OptimizerAsha) with different minimum budgets in the base stage. Each run of ASHA within AHB is called a bracket. AHB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget (r_min) is increased in each bracket by a factor of eta until the maximum budget is allocated in the base stage. The bracket s = 0 is a random search with full budget. Each ASHA run uses 1 / s_max + 1 of the bbotk::Terminator.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, Rostamizadeh A, Gonina E, Ben-tzur J, Hardt M, Recht B, Talwalkar A (2020). “A System for Massively Parallel Hyperparameter Tuning.” Proceedings of Machine Learning and Systems, 2, 230-246. https://proceedings.mlsys.org/paper/2020/hash/f4b9ec30ad9f68f89b29639786cb62ef-Abstract.html.

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

Custom sampler

All algorithms support custom paradox::Samplers to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerAhb

Methods

Public methods

OptimizerAhb$new()

OptimizerAhb$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAhb$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAhb$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# modified branin function
objective = ObjectiveRFunDt$new(
  fun = function(xdt) {
    a = 1
    b = 5.1 / (4 * (pi ^ 2))
    c = 5 / pi
    r = 6
    s = 10
    t = 1 / (8 * pi)
    data.table(y =
      (a * ((xdt[["x2"]] -
      b * (xdt[["x1"]] ^ 2L) +
      c * xdt[["x1"]] - r) ^ 2) +
      ((s * (1 - t)) * cos(xdt[["x1"]])) +
      s - (5 * xdt[["fidelity"]] * xdt[["x1"]])))
  },
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 100)
)

optimizer = opt("ahb")

# modifies the instance by reference
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 9.956037 2.235532     0.64 <list[3]> -29.61371

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 9.956037 2.235532     0.64 <list[3]> -29.61371

# all evaluations
as.data.table(instance$archive)
#>              x1        x2 fidelity asha_id stage bracket           timestamp
#>   1: -3.7887479 12.514996     0.01       1     0       6 2021-11-26 08:40:25
#>   2:  4.0114133  2.358127     0.01       2     0       6 2021-11-26 08:40:25
#>   3:  4.0114133  2.358127     0.02       2     1       6 2021-11-26 08:40:25
#>   4: -4.8890084  6.995902     0.01       3     0       6 2021-11-26 08:40:25
#>   5:  2.4666608  4.346509     0.01       4     0       6 2021-11-26 08:40:25
#>  ---                                                                        
#>  97:  0.3845054 11.623695     0.64      59     0       0 2021-11-26 08:40:29
#>  98:  3.7671288  9.509646     0.64      60     0       0 2021-11-26 08:40:29
#>  99:  7.8799923  8.503415     0.64      61     0       0 2021-11-26 08:40:29
#> 100: -1.2050447 13.782048     0.64      62     0       0 2021-11-26 08:40:29
#> 101:  8.0102531  3.728080     0.64      63     0       0 2021-11-26 08:40:29
#>      batch_nr    status          y
#>   1:        1 evaluated   4.404046
#>   2:        2 evaluated   4.046947
#>   3:        3 evaluated   3.846376
#>   4:        4 evaluated 109.408016
#>   5:        5 evaluated   4.588936
#>  ---                              
#>  97:       97 evaluated  56.316150
#>  98:       98 evaluated  49.019696
#>  99:       99 evaluated  33.859541
#> 100:      100 evaluated  49.513763
#> 101:      101  proposed         NA