Optimizer Asynchronous Successive Halving

OptimizerAsha class that implements the asynchronous successive halving algorithm (ASHA). ASHA parallelizes SHA (OptimizerSuccessiveHalving) by promoting candidates to the next stage as soon as possible instead of waiting for all candidates in the stage to finish. ASHA starts with sampling a candidate point for each available worker. When an evaluation finishes and the worker is available again, ASHA checks the stages from top to bottom for promotable candidates. Promotions are possible when the evaluated candidates belong to the top 1 / eta of each stage. If no promotions are possible, a new candidate is sampled and added to the base stage, which increases the number of possible promotions for all stages.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stage.

Source

Li L, Jamieson K, Rostamizadeh A, Gonina E, Ben-tzur J, Hardt M, Recht B, Talwalkar A (2020). “A System for Massively Parallel Hyperparameter Tuning.” Proceedings of Machine Learning and Systems, 2, 230-246. https://proceedings.mlsys.org/paper/2020/hash/f4b9ec30ad9f68f89b29639786cb62ef-Abstract.html.

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1/eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional column that is specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Parallelization

The points are asynchronously evaluated with the future package. To select a parallel backend, use future::plan().

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerAsha

Methods

Public methods

OptimizerAsha$new()

OptimizerAsha$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAsha$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAsha$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("asha")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.787338 3.534103     0.64 <list[3]> 2.547429

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.787338 3.534103     0.64 <list[3]> 2.547429

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity asha_id stage           timestamp batch_nr
#>  1:  6.5881723 14.9833518     0.01       1     0 2021-12-18 15:36:09        1
#>  2: -0.1385951 13.0583587     0.01       2     0 2021-12-18 15:36:09        2
#>  3: -0.1385951 13.0583587     0.02       2     1 2021-12-18 15:36:09        3
#>  4:  8.0379920  1.0794223     0.01       3     0 2021-12-18 15:36:09        4
#>  5:  8.0379920  1.0794223     0.02       3     1 2021-12-18 15:36:09        5
#>  6:  8.0379920  1.0794223     0.04       3     2 2021-12-18 15:36:09        6
#>  7: -2.0381866 10.7968492     0.01       4     0 2021-12-18 15:36:09        7
#>  8: -2.0381866 10.7968492     0.02       4     1 2021-12-18 15:36:09        8
#>  9: -2.0381866 10.7968492     0.04       4     2 2021-12-18 15:36:09        9
#> 10: -2.0381866 10.7968492     0.08       4     3 2021-12-18 15:36:09       10
#> 11:  3.7297025  2.5736275     0.01       5     0 2021-12-18 15:36:09       11
#> 12:  3.7297025  2.5736275     0.02       5     1 2021-12-18 15:36:09       12
#> 13:  3.7297025  2.5736275     0.04       5     2 2021-12-18 15:36:09       13
#> 14:  3.7297025  2.5736275     0.08       5     3 2021-12-18 15:36:09       14
#> 15:  3.7297025  2.5736275     0.16       5     4 2021-12-18 15:36:10       15
#> 16: -3.9958461 14.9329922     0.01       6     0 2021-12-18 15:36:10       16
#> 17: -3.9958461 14.9329922     0.02       6     1 2021-12-18 15:36:10       17
#> 18:  2.8436635  4.1097451     0.01       7     0 2021-12-18 15:36:10       18
#> 19:  2.8436635  4.1097451     0.02       7     1 2021-12-18 15:36:10       19
#> 20:  2.8436635  4.1097451     0.04       7     2 2021-12-18 15:36:10       20
#> 21:  2.8436635  4.1097451     0.08       7     3 2021-12-18 15:36:10       21
#> 22:  5.0384610  5.4105540     0.01       8     0 2021-12-18 15:36:10       22
#> 23: -4.3858724 13.1586824     0.01       9     0 2021-12-18 15:36:10       23
#> 24: -4.3858724 13.1586824     0.02       9     1 2021-12-18 15:36:10       24
#> 25: -4.3858724 13.1586824     0.04       9     2 2021-12-18 15:36:10       25
#> 26:  2.7873379  3.5341029     0.01      10     0 2021-12-18 15:36:10       26
#> 27:  2.7873379  3.5341029     0.02      10     1 2021-12-18 15:36:10       27
#> 28:  2.7873379  3.5341029     0.04      10     2 2021-12-18 15:36:10       28
#> 29:  2.7873379  3.5341029     0.08      10     3 2021-12-18 15:36:10       29
#> 30:  2.7873379  3.5341029     0.16      10     4 2021-12-18 15:36:10       30
#> 31:  2.7873379  3.5341029     0.32      10     5 2021-12-18 15:36:10       31
#> 32: -2.3125947  9.8593007     0.01      11     0 2021-12-18 15:36:10       32
#> 33: -2.3125947  9.8593007     0.02      11     1 2021-12-18 15:36:10       33
#> 34: -2.3125947  9.8593007     0.04      11     2 2021-12-18 15:36:10       34
#> 35: -2.3125947  9.8593007     0.08      11     3 2021-12-18 15:36:10       35
#> 36: -2.3125947  9.8593007     0.16      11     4 2021-12-18 15:36:10       36
#> 37: -2.3125947  9.8593007     0.32      11     5 2021-12-18 15:36:10       37
#> 38:  2.7873379  3.5341029     0.64      10     6 2021-12-18 15:36:10       38
#> 39:  3.1808891 10.3534515     0.01      12     0 2021-12-18 15:36:10       39
#> 40: -4.4948067 14.4128072     0.01      13     0 2021-12-18 15:36:11       40
#> 41:  3.9983266  7.5078131     0.01      14     0 2021-12-18 15:36:11       41
#> 42: -3.4279246  8.1587219     0.01      15     0 2021-12-18 15:36:11       42
#> 43: -2.6419161  4.7178409     0.01      16     0 2021-12-18 15:36:11       43
#> 44: -4.4948067 14.4128072     0.02      13     1 2021-12-18 15:36:11       44
#> 45: -0.4282345 12.8523821     0.01      17     0 2021-12-18 15:36:11       45
#> 46: -3.3265699  1.4327677     0.01      18     0 2021-12-18 15:36:11       46
#> 47: -3.4279246  8.1587219     0.02      15     1 2021-12-18 15:36:11       47
#> 48:  4.5948266  1.7582454     0.01      19     0 2021-12-18 15:36:11       48
#> 49:  5.7549467  0.4248941     0.01      20     0 2021-12-18 15:36:11       49
#> 50:  4.5948266  1.7582454     0.02      19     1 2021-12-18 15:36:11       50
#> 51:  7.9415599  2.8148937     0.01      21     0 2021-12-18 15:36:11       51
#>             x1         x2 fidelity asha_id stage           timestamp batch_nr
#>        status          y x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: evaluated 348.894639   6.5881723  14.9833518              0.01
#>  2: evaluated  66.257315  -0.1385951  13.0583587              0.01
#>  3: evaluated  66.257053  -0.1385951  13.0583587              0.02
#>  4: evaluated  43.314138   8.0379920   1.0794223              0.01
#>  5: evaluated  42.553069   8.0379920   1.0794223              0.02
#>  6: evaluated  41.055977   8.0379920   1.0794223              0.04
#>  7: evaluated   7.711681  -2.0381866  10.7968492              0.01
#>  8: evaluated   7.699837  -2.0381866  10.7968492              0.02
#>  9: evaluated   7.676253  -2.0381866  10.7968492              0.04
#> 10: evaluated   7.629500  -2.0381866  10.7968492              0.08
#> 11: evaluated   6.378170   3.7297025   2.5736275              0.01
#> 12: evaluated   6.320224   3.7297025   2.5736275              0.02
#> 13: evaluated   6.205493   3.7297025   2.5736275              0.04
#> 14: evaluated   5.980675   3.7297025   2.5736275              0.08
#> 15: evaluated   5.549615   3.7297025   2.5736275              0.16
#> 16: evaluated   8.067697  -3.9958461  14.9329922              0.01
#> 17: evaluated   8.001164  -3.9958461  14.9329922              0.02
#> 18: evaluated   6.540126   2.8436635   4.1097451              0.01
#> 19: evaluated   6.501514   2.8436635   4.1097451              0.02
#> 20: evaluated   6.424683   2.8436635   4.1097451              0.04
#> 21: evaluated   6.272590   2.8436635   4.1097451              0.08
#> 22: evaluated  57.474655   5.0384610   5.4105540              0.01
#> 23: evaluated   7.081990  -4.3858724  13.1586824              0.01
#> 24: evaluated   7.097837  -4.3858724  13.1586824              0.02
#> 25: evaluated   7.131749  -4.3858724  13.1586824              0.04
#> 26: evaluated   4.007053   2.7873379   3.5341029              0.01
#> 27: evaluated   3.980142   2.7873379   3.5341029              0.02
#> 28: evaluated   3.926682   2.7873379   3.5341029              0.04
#> 29: evaluated   3.821211   2.7873379   3.5341029              0.08
#> 30: evaluated   3.616063   2.7873379   3.5341029              0.16
#> 31: evaluated   3.228947   2.7873379   3.5341029              0.32
#> 32: evaluated   3.512968  -2.3125947   9.8593007              0.01
#> 33: evaluated   3.512812  -2.3125947   9.8593007              0.02
#> 34: evaluated   3.512671  -2.3125947   9.8593007              0.04
#> 35: evaluated   3.513077  -2.3125947   9.8593007              0.08
#> 36: evaluated   3.516634  -2.3125947   9.8593007              0.16
#> 37: evaluated   3.534730  -2.3125947   9.8593007              0.32
#> 38: evaluated   2.547429   2.7873379   3.5341029              0.64
#> 39: evaluated  83.408062   3.1808891  10.3534515              0.01
#> 40: evaluated   8.348753  -4.4948067  14.4128072              0.01
#> 41: evaluated  58.305846   3.9983266   7.5078131              0.01
#> 42: evaluated  14.123574  -3.4279246   8.1587219              0.01
#> 43: evaluated  34.034267  -2.6419161   4.7178409              0.01
#> 44: evaluated   8.322926  -4.4948067  14.4128072              0.02
#> 45: evaluated  56.745850  -0.4282345  12.8523821              0.01
#> 46: evaluated 104.513107  -3.3265699   1.4327677              0.01
#> 47: evaluated  14.209532  -3.4279246   8.1587219              0.02
#> 48: evaluated  14.797478   4.5948266   1.7582454              0.01
#> 49: evaluated  24.972860   5.7549467   0.4248941              0.01
#> 50: evaluated  14.695154   4.5948266   1.7582454              0.02
#> 51:  proposed         NA          NA          NA                NA
#>        status          y x_domain_x1 x_domain_x2 x_domain_fidelity