Optimizer Asynchronous Successive Halving

OptimizerAsha class that implements the asynchronous successive halving algorithm (ASHA). ASHA parallelizes SHA (OptimizerSuccessiveHalving) by promoting candidates to the next stage as soon as possible instead of waiting for all candidates in the stage to finish. ASHA starts with sampling a candidate point for each available worker. When an evaluation finishes and the worker is available again, ASHA checks the stages from top to bottom for promotable candidates. Promotions are possible when the evaluated candidates belong to the top 1 / eta of each stage. If no promotions are possible, a new candidate is sampled and added to the base stage, which increases the number of possible promotions for all stages.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stage.

Source

Li L, Jamieson K, Rostamizadeh A, Gonina E, Ben-tzur J, Hardt M, Recht B, Talwalkar A (2020). “A System for Massively Parallel Hyperparameter Tuning.” Proceedings of Machine Learning and Systems, 2, 230-246. https://proceedings.mlsys.org/paper/2020/hash/f4b9ec30ad9f68f89b29639786cb62ef-Abstract.html.

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1/eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional column that is specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Parallelization

The points are asynchronously evaluated with the future package. To select a parallel backend, use future::plan().

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerAsha

Methods

Public methods

OptimizerAsha$new()

OptimizerAsha$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAsha$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAsha$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("asha")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.723838 0.7351081     0.16 <list[3]> 1.981266

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.723838 0.7351081     0.16 <list[3]> 1.981266

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity asha_id stage           timestamp batch_nr
#>  1: -2.7558811  9.4272494     0.01       1     0 2021-12-06 14:19:10        1
#>  2: -4.8658290  6.0541437     0.01       2     0 2021-12-06 14:19:10        2
#>  3: -2.7558811  9.4272494     0.02       1     1 2021-12-06 14:19:10        3
#>  4: -4.5204498 11.1439033     0.01       3     0 2021-12-06 14:19:10        4
#>  5:  3.7238378  0.7351081     0.01       4     0 2021-12-06 14:19:10        5
#>  6:  3.7238378  0.7351081     0.02       4     1 2021-12-06 14:19:10        6
#>  7:  3.7238378  0.7351081     0.04       4     2 2021-12-06 14:19:10        7
#>  8:  3.3959702 13.3773871     0.01       5     0 2021-12-06 14:19:10        8
#>  9:  5.3350793  6.4702321     0.01       6     0 2021-12-06 14:19:10        9
#> 10: -4.5204498 11.1439033     0.02       3     1 2021-12-06 14:19:10       10
#> 11:  4.4563178  2.7347675     0.01       7     0 2021-12-06 14:19:10       11
#> 12:  2.7292973  3.0393647     0.01       8     0 2021-12-06 14:19:10       12
#> 13:  2.7292973  3.0393647     0.02       8     1 2021-12-06 14:19:10       13
#> 14:  2.7292973  3.0393647     0.04       8     2 2021-12-06 14:19:10       14
#> 15:  3.7238378  0.7351081     0.08       4     3 2021-12-06 14:19:10       15
#> 16:  0.2231772  7.5899598     0.01       9     0 2021-12-06 14:19:10       16
#> 17:  4.5796793 13.9651987     0.01      10     0 2021-12-06 14:19:10       17
#> 18:  4.4563178  2.7347675     0.02       7     1 2021-12-06 14:19:10       18
#> 19:  1.2368257  4.4516840     0.01      11     0 2021-12-06 14:19:10       19
#> 20:  1.2368257  4.4516840     0.02      11     1 2021-12-06 14:19:10       20
#> 21: -2.7558811  9.4272494     0.04       1     2 2021-12-06 14:19:10       21
#> 22:  3.5517375  7.9445841     0.01      12     0 2021-12-06 14:19:10       22
#> 23:  8.2901931  2.7682211     0.01      13     0 2021-12-06 14:19:10       23
#> 24:  2.7015692 14.9592523     0.01      14     0 2021-12-06 14:19:10       24
#> 25:  0.2231772  7.5899598     0.02       9     1 2021-12-06 14:19:10       25
#> 26:  1.7991726  3.6299995     0.01      15     0 2021-12-06 14:19:10       26
#> 27:  1.7991726  3.6299995     0.02      15     1 2021-12-06 14:19:11       27
#> 28:  1.7991726  3.6299995     0.04      15     2 2021-12-06 14:19:11       28
#> 29:  2.7292973  3.0393647     0.08       8     3 2021-12-06 14:19:11       29
#> 30:  3.7238378  0.7351081     0.16       4     4 2021-12-06 14:19:11       30
#> 31:  5.3407788  4.9655973     0.01      16     0 2021-12-06 14:19:11       31
#> 32: -3.7248026  8.1003416     0.01      17     0 2021-12-06 14:19:11       32
#> 33: -3.7248026  8.1003416     0.02      17     1 2021-12-06 14:19:11       33
#> 34:  4.5079269  4.3486535     0.01      18     0 2021-12-06 14:19:11       34
#> 35:  8.0992186 10.4575349     0.01      19     0 2021-12-06 14:19:11       35
#> 36:  3.8885072  7.4924852     0.01      20     0 2021-12-06 14:19:11       36
#> 37:  4.5079269  4.3486535     0.02      18     1 2021-12-06 14:19:11       37
#> 38:  1.2368257  4.4516840     0.04      11     2 2021-12-06 14:19:11       38
#> 39:  2.6336287  8.3312248     0.01      21     0 2021-12-06 14:19:11       39
#> 40:  6.0473492  4.1133161     0.01      22     0 2021-12-06 14:19:11       40
#> 41:  2.6336287  8.3312248     0.02      21     1 2021-12-06 14:19:11       41
#> 42: -4.2000327  9.7343427     0.01      23     0 2021-12-06 14:19:11       42
#> 43: -4.2000327  9.7343427     0.02      23     1 2021-12-06 14:19:11       43
#> 44: -4.5204498 11.1439033     0.04       3     2 2021-12-06 14:19:11       44
#> 45: -2.7558811  9.4272494     0.08       1     3 2021-12-06 14:19:11       45
#> 46:  2.7619040  0.8840193     0.01      24     0 2021-12-06 14:19:11       46
#> 47:  2.7619040  0.8840193     0.02      24     1 2021-12-06 14:19:11       47
#> 48:  2.7619040  0.8840193     0.04      24     2 2021-12-06 14:19:11       48
#> 49:  2.7619040  0.8840193     0.08      24     3 2021-12-06 14:19:11       49
#> 50:  2.7619040  0.8840193     0.16      24     4 2021-12-06 14:19:11       50
#> 51:  3.7238378  0.7351081     0.32       4     5 2021-12-06 14:19:12       51
#>             x1         x2 fidelity asha_id stage           timestamp batch_nr
#>        status          y x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: evaluated   2.514973  -2.7558811   9.4272494              0.01
#>  2: evaluated  82.106906  -4.8658290   6.0541437              0.01
#>  3: evaluated   2.533078  -2.7558811   9.4272494              0.02
#>  4: evaluated  15.283405  -4.5204498  11.1439033              0.01
#>  5: evaluated   2.039178   3.7238378   0.7351081              0.01
#>  6: evaluated   2.032626   3.7238378   0.7351081              0.02
#>  7: evaluated   2.020673   3.7238378   0.7351081              0.04
#>  8: evaluated 155.314603   3.3959702  13.3773871              0.01
#>  9: evaluated  81.244690   5.3350793   6.4702321              0.01
#> 10: evaluated  15.392837  -4.5204498  11.1439033              0.02
#> 11: evaluated  17.986591   4.4563178   2.7347675              0.01
#> 12: evaluated   2.544239   2.7292973   3.0393647              0.01
#> 13: evaluated   2.527038   2.7292973   3.0393647              0.02
#> 14: evaluated   2.492968   2.7292973   3.0393647              0.04
#> 15: evaluated   2.001384   3.7238378   0.7351081              0.08
#> 16: evaluated  23.141763   0.2231772   7.5899598              0.01
#> 17: evaluated 222.500480   4.5796793  13.9651987              0.01
#> 18: evaluated  17.858785   4.4563178   2.7347675              0.02
#> 19: evaluated  13.287402   1.2368257   4.4516840              0.01
#> 20: evaluated  13.286260   1.2368257   4.4516840              0.02
#> 21: evaluated   2.569634  -2.7558811   9.4272494              0.04
#> 22: evaluated  53.273243   3.5517375   7.9445841              0.01
#> 23: evaluated  68.163146   8.2901931   2.7682211              0.01
#> 24: evaluated 171.318500   2.7015692  14.9592523              0.01
#> 25: evaluated  23.141569   0.2231772   7.5899598              0.02
#> 26: evaluated   7.982746   1.7991726   3.6299995              0.01
#> 27: evaluated   7.980194   1.7991726   3.6299995              0.02
#> 28: evaluated   7.975154   1.7991726   3.6299995              0.04
#> 29: evaluated   2.426160   2.7292973   3.0393647              0.08
#> 30: evaluated   1.981266   3.7238378   0.7351081              0.16
#> 31: evaluated  59.266995   5.3407788   4.9655973              0.01
#> 32: evaluated  20.019164  -3.7248026   8.1003416              0.01
#> 33: evaluated  20.137194  -3.7248026   8.1003416              0.02
#> 34: evaluated  32.157012   4.5079269   4.3486535              0.01
#> 35: evaluated 243.838573   8.0992186  10.4575349              0.01
#> 36: evaluated  55.152279   3.8885072   7.4924852              0.01
#> 37: evaluated  31.957875   4.5079269   4.3486535              0.02
#> 38: evaluated  13.283991   1.2368257   4.4516840              0.04
#> 39: evaluated  41.469495   2.6336287   8.3312248              0.01
#> 40: evaluated  63.347743   6.0473492   4.1133161              0.01
#> 41: evaluated  41.381964   2.6336287   8.3312248              0.02
#> 42: evaluated  17.421794  -4.2000327   9.7343427              0.01
#> 43: evaluated  17.544976  -4.2000327   9.7343427              0.02
#> 44: evaluated  15.614207  -4.5204498  11.1439033              0.04
#> 45: evaluated   2.644130  -2.7558811   9.4272494              0.08
#> 46: evaluated   1.985247   2.7619040   0.8840193              0.01
#> 47: evaluated   1.999806   2.7619040   0.8840193              0.02
#> 48: evaluated   2.029275   2.7619040   0.8840193              0.04
#> 49: evaluated   2.089608   2.7619040   0.8840193              0.08
#> 50: evaluated   2.215860   2.7619040   0.8840193              0.16
#> 51:  proposed         NA          NA          NA                NA
#>        status          y x_domain_x1 x_domain_x2 x_domain_fidelity