Optimizer Asynchronous Successive Halving

OptimizerAsha class that implements the asynchronous successive halving algorithm (ASHA). ASHA parallelizes SHA (OptimizerSuccessiveHalving) by promoting candidates to the next stage as soon as possible instead of waiting for all candidates in the stage to finish. ASHA starts with sampling a candidate point for each available worker. When an evaluation finishes and the worker is available again, ASHA checks the stages from top to bottom for promotable candidates. Promotions are possible when the evaluated candidates belong to the top 1 / eta of each stage. If no promotions are possible, a new candidate is sampled and added to the base stage, which increases the number of possible promotions for all stages.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stage.

Source

Li L, Jamieson K, Rostamizadeh A, Gonina E, Ben-tzur J, Hardt M, Recht B, Talwalkar A (2020). “A System for Massively Parallel Hyperparameter Tuning.” Proceedings of Machine Learning and Systems, 2, 230-246. https://proceedings.mlsys.org/paper/2020/hash/f4b9ec30ad9f68f89b29639786cb62ef-Abstract.html.

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1/eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional column that is specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Parallelization

The points are asynchronously evaluated with the future package. To select a parallel backend, use future::plan().

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerAsha

Methods

Public methods

OptimizerAsha$new()

OptimizerAsha$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAsha$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAsha$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("asha")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain       y
#> 1: 3.169051 4.178957     0.16 <list[3]> 8.06818

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain       y
#> 1: 3.169051 4.178957     0.16 <list[3]> 8.06818

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity asha_id stage           timestamp batch_nr
#>  1:  3.1690509  4.1789573     0.01       1     0 2021-11-27 11:24:37        1
#>  2:  1.7005370  5.5726678     0.01       2     0 2021-11-27 11:24:37        2
#>  3:  3.1690509  4.1789573     0.02       1     1 2021-11-27 11:24:37        3
#>  4: -4.5790854  6.9898079     0.01       3     0 2021-11-27 11:24:37        4
#>  5:  0.8504708  0.3009783     0.01       4     0 2021-11-27 11:24:38        5
#>  6:  1.7005370  5.5726678     0.02       2     1 2021-11-27 11:24:38        6
#>  7:  3.1690509  4.1789573     0.04       1     2 2021-11-27 11:24:38        7
#>  8:  0.6545639  8.3986926     0.01       5     0 2021-11-27 11:24:38        8
#>  9:  7.8562538  5.7721457     0.01       6     0 2021-11-27 11:24:38        9
#> 10:  0.6545639  8.3986926     0.02       5     1 2021-11-27 11:24:38       10
#> 11:  2.9187555  9.0095629     0.01       7     0 2021-11-27 11:24:38       11
#> 12: -1.0794296  4.3507524     0.01       8     0 2021-11-27 11:24:38       12
#> 13: -1.0794296  4.3507524     0.02       8     1 2021-11-27 11:24:38       13
#> 14:  1.7005370  5.5726678     0.04       2     2 2021-11-27 11:24:38       14
#> 15:  3.1690509  4.1789573     0.08       1     3 2021-11-27 11:24:38       15
#> 16:  2.2011276 13.8000832     0.01       9     0 2021-11-27 11:24:38       16
#> 17:  1.0108028  3.1975907     0.01      10     0 2021-11-27 11:24:38       17
#> 18:  1.0108028  3.1975907     0.02      10     1 2021-11-27 11:24:38       18
#> 19:  5.0765022  0.8792117     0.01      11     0 2021-11-27 11:24:38       19
#> 20:  5.0765022  0.8792117     0.02      11     1 2021-11-27 11:24:38       20
#> 21:  1.0108028  3.1975907     0.04      10     2 2021-11-27 11:24:38       21
#> 22:  9.9560370  2.2355320     0.01      12     0 2021-11-27 11:24:38       22
#> 23:  2.7783495 12.6918008     0.01      13     0 2021-11-27 11:24:38       23
#> 24:  5.7740459  3.6197103     0.01      14     0 2021-11-27 11:24:38       24
#> 25:  0.8504708  0.3009783     0.02       4     1 2021-11-27 11:24:38       25
#> 26:  3.2056505 12.5220272     0.01      15     0 2021-11-27 11:24:38       26
#> 27: -4.5806596  7.0407645     0.01      16     0 2021-11-27 11:24:39       27
#> 28:  5.7740459  3.6197103     0.02      14     1 2021-11-27 11:24:39       28
#> 29:  5.0765022  0.8792117     0.04      11     2 2021-11-27 11:24:39       29
#> 30:  1.7005370  5.5726678     0.08       2     3 2021-11-27 11:24:39       30
#> 31:  3.1690509  4.1789573     0.16       1     4 2021-11-27 11:24:39       31
#> 32:  7.0852005 12.2107697     0.01      17     0 2021-11-27 11:24:39       32
#> 33:  1.0586650  3.2764651     0.01      18     0 2021-11-27 11:24:39       33
#> 34:  1.0586650  3.2764651     0.02      18     1 2021-11-27 11:24:39       34
#> 35:  1.0586650  3.2764651     0.04      18     2 2021-11-27 11:24:39       35
#> 36:  1.2754210 10.0330612     0.01      19     0 2021-11-27 11:24:39       36
#> 37:  1.2754210 10.0330612     0.02      19     1 2021-11-27 11:24:39       37
#> 38:  2.6147542  9.9053896     0.01      20     0 2021-11-27 11:24:39       38
#> 39:  2.6768697 12.5332866     0.01      21     0 2021-11-27 11:24:39       39
#> 40:  5.6317174 13.1130891     0.01      22     0 2021-11-27 11:24:39       40
#> 41:  2.9187555  9.0095629     0.02       7     1 2021-11-27 11:24:39       41
#> 42: -4.8278069 13.3237435     0.01      23     0 2021-11-27 11:24:39       42
#> 43: -4.8278069 13.3237435     0.02      23     1 2021-11-27 11:24:39       43
#> 44: -4.8278069 13.3237435     0.04      23     2 2021-11-27 11:24:39       44
#> 45: -4.8278069 13.3237435     0.08      23     3 2021-11-27 11:24:39       45
#> 46:  9.9452038  7.5028725     0.01      24     0 2021-11-27 11:24:39       46
#> 47:  0.3845054 11.6236953     0.01      25     0 2021-11-27 11:24:39       47
#> 48:  3.7671288  9.5096456     0.01      26     0 2021-11-27 11:24:40       48
#> 49: -4.5806596  7.0407645     0.02      16     1 2021-11-27 11:24:40       49
#> 50:  7.8799923  8.5034151     0.01      27     0 2021-11-27 11:24:40       50
#> 51: -1.2050447 13.7820482     0.01      28     0 2021-11-27 11:24:40       51
#>             x1         x2 fidelity asha_id stage           timestamp batch_nr
#>        status          y x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: evaluated   8.925098   3.1690509   4.1789573              0.01
#>  2: evaluated  13.561993   1.7005370   5.5726678              0.01
#>  3: evaluated   8.866558   3.1690509   4.1789573              0.02
#>  4: evaluated  56.761753  -4.5790854   6.9898079              0.01
#>  5: evaluated  35.407039   0.8504708   0.3009783              0.01
#>  6: evaluated  13.549325   1.7005370   5.5726678              0.02
#>  7: evaluated   8.750083   3.1690509   4.1789573              0.04
#>  8: evaluated  29.365453   0.6545639   8.3986926              0.01
#>  9: evaluated 118.403590   7.8562538   5.7721457              0.01
#> 10: evaluated  29.362516   0.6545639   8.3986926              0.02
#> 11: evaluated  55.362169   2.9187555   9.0095629              0.01
#> 12: evaluated  26.106789  -1.0794296   4.3507524              0.01
#> 13: evaluated  26.114719  -1.0794296   4.3507524              0.02
#> 14: evaluated  13.524037   1.7005370   5.5726678              0.04
#> 15: evaluated   8.519555   3.1690509   4.1789573              0.08
#> 16: evaluated 128.820021   2.2011276  13.8000832              0.01
#> 17: evaluated  16.599875   1.0108028   3.1975907              0.01
#> 18: evaluated  16.602378   1.0108028   3.1975907              0.02
#> 19: evaluated  18.175552   5.0765022   0.8792117              0.01
#> 20: evaluated  18.063812   5.0765022   0.8792117              0.02
#> 21: evaluated  16.607391   1.0108028   3.1975907              0.04
#> 22: evaluated  84.332865   9.9560370   2.2355320              0.01
#> 23: evaluated 119.413636   2.7783495  12.6918008              0.01
#> 24: evaluated  52.059612   5.7740459   3.6197103              0.01
#> 25: evaluated  35.413357   0.8504708   0.3009783              0.02
#> 26: evaluated 128.419575   3.2056505  12.5220272              0.01
#> 27: evaluated  56.113451  -4.5806596   7.0407645              0.01
#> 28: evaluated  51.673781   5.7740459   3.6197103              0.02
#> 29: evaluated  17.844317   5.0765022   0.8792117              0.04
#> 30: evaluated  13.473664   1.7005370   5.5726678              0.08
#> 31: evaluated   8.068180   3.1690509   4.1789573              0.16
#> 32: evaluated 271.779215   7.0852005  12.2107697              0.01
#> 33: evaluated  15.855526   1.0586650   3.2764651              0.01
#> 34: evaluated  15.857931   1.0586650   3.2764651              0.02
#> 35: evaluated  15.862749   1.0586650   3.2764651              0.04
#> 36: evaluated  48.961625   1.2754210  10.0330612              0.01
#> 37: evaluated  48.942062   1.2754210  10.0330612              0.02
#> 38: evaluated  63.487864   2.6147542   9.9053896              0.01
#> 39: evaluated 113.296902   2.6768697  12.5332866              0.01
#> 40: evaluated 246.216885   5.6317174  13.1130891              0.01
#> 41: evaluated  55.236196   2.9187555   9.0095629              0.02
#> 42: evaluated  12.236791  -4.8278069  13.3237435              0.01
#> 43: evaluated  12.286909  -4.8278069  13.3237435              0.02
#> 44: evaluated  12.390404  -4.8278069  13.3237435              0.04
#> 45: evaluated  12.610433  -4.8278069  13.3237435              0.08
#> 46: evaluated 207.468081   9.9452038   7.5028725              0.01
#> 47: evaluated  57.728760   0.3845054  11.6236953              0.01
#> 48: evaluated  84.605460   3.7671288   9.5096456              0.01
#> 49: evaluated  56.402733  -4.5806596   7.0407645              0.02
#> 50: evaluated 183.213053   7.8799923   8.5034151              0.01
#> 51:  proposed         NA          NA          NA                NA
#>        status          y x_domain_x1 x_domain_x2 x_domain_fidelity