Optimizer Asynchronous Successive Halving

OptimizerAsha class that implements the asynchronous successive halving algorithm. Asynchronous successive halving (ASHA) parallelizes SHA (OptimizerSuccessiveHalving) by promoting candidates to the next stage as soon as possible instead of waiting for all candidates in the stage to finish. ASHA starts with sampling a candidate points for each available worker. When an evaluation finishes and the worker is available again, ASHA checks the stages from top to bottom for promotable candidates. Promotions are possible when the evaluated candidates belong to the top 1 / eta of each stage. If no promotions are possible, a new candidate is sampled and added to the base stage, which increases the number of possible promotions for all stages.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stage.

Source

Li L, Jamieson K, Rostamizadeh A, Gonina E, Ben-tzur J, Hardt M, Recht B, Talwalkar A (2020). “A System for Massively Parallel Hyperparameter Tuning.” Proceedings of Machine Learning and Systems, 2, 230-246. https://proceedings.mlsys.org/paper/2020/hash/f4b9ec30ad9f68f89b29639786cb62ef-Abstract.html.

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1/eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional column that is specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

Custom sampler

All algorithms support custom paradox::Samplers to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerAsha

Methods

Public methods

OptimizerAsha$new()

OptimizerAsha$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerAsha$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerAsha$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# modified branin function
objective = ObjectiveRFunDt$new(
  fun = function(xdt) {
    a = 1
    b = 5.1 / (4 * (pi ^ 2))
    c = 5 / pi
    r = 6
    s = 10
    t = 1 / (8 * pi)
    data.table(y =
      (a * ((xdt[["x2"]] -
      b * (xdt[["x1"]] ^ 2L) +
      c * xdt[["x1"]] - r) ^ 2) +
      ((s * (1 - t)) * cos(xdt[["x1"]])) +
      s - (5 * xdt[["fidelity"]] * xdt[["x1"]])))
  },
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 100)
)

optimizer = opt("asha")

# modifies the instance by reference
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain         y
#> 1: 8.641614 0.5617675     0.64 <list[3]> -22.68421

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain         y
#> 1: 8.641614 0.5617675     0.64 <list[3]> -22.68421

# all evaluations
as.data.table(instance$archive)
#>              x1        x2 fidelity asha_id stage           timestamp batch_nr
#>   1:  1.0432182 11.544453     0.01       1     0 2021-11-26 09:42:06        1
#>   2: -3.2077194  2.920424     0.01       2     0 2021-11-26 09:42:06        2
#>   3:  1.0432182 11.544453     0.02       1     1 2021-11-26 09:42:06        3
#>   4: -2.5314613  9.948099     0.01       3     0 2021-11-26 09:42:06        4
#>   5: -2.5314613  9.948099     0.02       3     1 2021-11-26 09:42:06        5
#>  ---                                                                         
#>  97: -4.4071897 14.171622     0.04      42     2 2021-11-26 09:42:10       97
#>  98: -1.3260802 11.716839     0.01      43     0 2021-11-26 09:42:10       98
#>  99: -0.6764425 13.130369     0.01      44     0 2021-11-26 09:42:10       99
#> 100: -0.5637486 14.752881     0.01      45     0 2021-11-26 09:42:10      100
#> 101:  3.8475633 11.387376     0.01      46     0 2021-11-26 09:42:10      101
#>         status         y
#>   1: evaluated 64.684784
#>   2: evaluated 91.096575
#>   3: evaluated 64.632623
#>   4: evaluated  3.082678
#>   5: evaluated  3.209251
#>  ---                    
#>  97: evaluated  9.823609
#>  98: evaluated 23.811346
#>  99: evaluated 53.457609
#> 100: evaluated 79.212281
#> 101:  proposed        NA