Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.165954 0.0177719 0.015625 <list[3]> 1.967341

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.165954 0.0177719 0.015625 <list[3]> 1.967341

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition          y
#>  1:  9.94893737  4.15318466 0.015625     0       6          1 121.458804
#>  2:  8.91369996 13.72548137 0.015625     0       6          1 380.670767
#>  3:  2.42701127  8.47115375 0.015625     0       6          1  40.603213
#>  4:  8.93117766 10.46325267 0.015625     0       6          1 264.797918
#>  5: -0.05043194  5.04592040 0.015625     0       6          1  20.659935
#>  6:  0.45051433 10.64297132 0.015625     0       6          1  47.306648
#>  7:  0.14320403  3.42834149 0.015625     0       6          1  24.999907
#>  8:  7.05642278  0.91668398 0.015625     0       6          1  38.182384
#>  9:  8.37673950  9.53798786 0.015625     0       6          1 221.662784
#> 10:  6.16997106 10.18192290 0.015625     0       6          1 184.181438
#> 11:  1.20161164  6.56716476 0.015625     0       6          1  19.395144
#> 12: -1.23832261  5.05905923 0.015625     0       6          1  21.889292
#> 13:  6.88291829  7.54983932 0.015625     0       6          1 139.978483
#> 14: -4.52821928  2.59919603 0.015625     0       6          1 134.537558
#> 15: -3.43127442  4.70633660 0.015625     0       6          1  51.445531
#> 16:  2.22328180  2.20178305 0.015625     0       6          1   4.339477
#> 17:  7.27369021 10.23160208 0.015625     0       6          1 216.374388
#> 18: -3.01942277  9.72198574 0.015625     0       6          1   2.329660
#> 19: -3.76938679  9.48530313 0.015625     0       6          1  10.935543
#> 20:  2.87789617  3.54508763 0.015625     0       6          1   4.229494
#> 21: -1.54095043  4.75689373 0.015625     0       6          1  24.488998
#> 22: -1.99226873 13.22753880 0.015625     0       6          1  21.553641
#> 23:  4.01803848 11.75179194 0.015625     0       6          1 139.585219
#> 24: -4.97952242 13.97080042 0.015625     0       6          1  13.048374
#> 25: -4.80467339 10.56349605 0.015625     0       6          1  25.273009
#> 26:  2.92626691 14.19984777 0.015625     0       6          1 159.224900
#> 27: -1.13058665  8.55355765 0.015625     0       6          1  14.602781
#> 28: -0.85700706 12.87717410 0.015625     0       6          1  46.433465
#> 29:  3.16595430  0.01777190 0.015625     0       6          1   1.967341
#> 30: -4.93213370  9.44411853 0.015625     0       6          1  38.652316
#> 31:  3.57090440  1.81841246 0.015625     0       6          1   2.500506
#> 32: -1.28637796  7.07054303 0.015625     0       6          1  13.750452
#> 33:  8.09350139  6.80715083 0.015625     0       6          1 144.010910
#> 34: -2.15204658 13.39936018 0.015625     0       6          1  19.411025
#> 35:  5.33661364  0.02326276 0.015625     0       6          1  18.305305
#> 36:  3.53908667  9.93414678 0.015625     0       6          1  85.449548
#> 37:  4.31861039  5.98421072 0.015625     0       6          1  45.805125
#> 38:  6.19582266  0.70439226 0.015625     0       6          1  31.023904
#> 39:  9.83926073  6.51209852 0.015625     0       6          1 175.322089
#> 40:  3.20223018 14.01827302 0.015625     0       6          1 164.242537
#> 41:  1.87756154 14.87631569 0.015625     0       6          1 145.307620
#> 42: -1.87872045  1.98342531 0.015625     0       6          1  57.715504
#> 43: -0.91077299  2.33432174 0.015625     0       6          1  42.314431
#> 44:  7.41185228  0.63981174 0.015625     0       6          1  36.642769
#> 45: -3.34881624  2.05421688 0.015625     0       6          1  93.155493
#> 46:  3.88414974 13.34688577 0.015625     0       6          1 173.615599
#> 47: -1.65510178  1.34871275 0.015625     0       6          1  63.503790
#> 48: -4.59724354  2.05420444 0.015625     0       6          1 150.801197
#> 49: -1.08006719 14.75679728 0.015625     0       6          1  63.552471
#> 50:  6.63223765  9.57011388 0.015625     0       6          1 182.177520
#> 51: -1.69569590  8.63920155 0.015625     0       6          1   8.825717
#> 52:  6.10884516  5.70679251 0.015625     0       6          1  88.046507
#> 53:  8.66905875 14.68815647 0.015625     0       6          1 410.029298
#> 54:  6.70982554 13.06052827 0.015625     0       6          1 286.236014
#> 55:  4.58790308 14.92603997 0.015625     0       6          1 251.566754
#> 56:  3.23337643 13.41891008 0.015625     0       6          1 150.342504
#> 57:  3.01056903  4.28291244 0.015625     0       6          1   8.296160
#> 58:  0.63067069  0.05891415 0.015625     0       6          1  42.253236
#> 59: -4.91835829  7.79144560 0.015625     0       6          1  57.934119
#> 60:  3.73788699  0.32458681 0.015625     0       6          1   2.079325
#> 61:  2.03550400 13.43340606 0.015625     0       6          1 116.906760
#> 62: -1.77329424  7.38055280 0.015625     0       6          1  10.435586
#> 63: -1.44373709 13.45548009 0.015625     0       6          1  37.161648
#> 64:  8.98411372  6.77058884 0.015625     0       6          1 159.761018
#>              x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-04-13 04:30:19        1  9.94893737  4.15318466          0.015625
#>  2: 2022-04-13 04:30:19        1  8.91369996 13.72548137          0.015625
#>  3: 2022-04-13 04:30:19        1  2.42701127  8.47115375          0.015625
#>  4: 2022-04-13 04:30:19        1  8.93117766 10.46325267          0.015625
#>  5: 2022-04-13 04:30:19        1 -0.05043194  5.04592040          0.015625
#>  6: 2022-04-13 04:30:19        1  0.45051433 10.64297132          0.015625
#>  7: 2022-04-13 04:30:19        1  0.14320403  3.42834149          0.015625
#>  8: 2022-04-13 04:30:19        1  7.05642278  0.91668398          0.015625
#>  9: 2022-04-13 04:30:19        1  8.37673950  9.53798786          0.015625
#> 10: 2022-04-13 04:30:19        1  6.16997106 10.18192290          0.015625
#> 11: 2022-04-13 04:30:19        1  1.20161164  6.56716476          0.015625
#> 12: 2022-04-13 04:30:19        1 -1.23832261  5.05905923          0.015625
#> 13: 2022-04-13 04:30:19        1  6.88291829  7.54983932          0.015625
#> 14: 2022-04-13 04:30:19        1 -4.52821928  2.59919603          0.015625
#> 15: 2022-04-13 04:30:19        1 -3.43127442  4.70633660          0.015625
#> 16: 2022-04-13 04:30:19        1  2.22328180  2.20178305          0.015625
#> 17: 2022-04-13 04:30:19        1  7.27369021 10.23160208          0.015625
#> 18: 2022-04-13 04:30:19        1 -3.01942277  9.72198574          0.015625
#> 19: 2022-04-13 04:30:19        1 -3.76938679  9.48530313          0.015625
#> 20: 2022-04-13 04:30:19        1  2.87789617  3.54508763          0.015625
#> 21: 2022-04-13 04:30:19        1 -1.54095043  4.75689373          0.015625
#> 22: 2022-04-13 04:30:19        1 -1.99226873 13.22753880          0.015625
#> 23: 2022-04-13 04:30:19        1  4.01803848 11.75179194          0.015625
#> 24: 2022-04-13 04:30:19        1 -4.97952242 13.97080042          0.015625
#> 25: 2022-04-13 04:30:19        1 -4.80467339 10.56349605          0.015625
#> 26: 2022-04-13 04:30:19        1  2.92626691 14.19984777          0.015625
#> 27: 2022-04-13 04:30:19        1 -1.13058665  8.55355765          0.015625
#> 28: 2022-04-13 04:30:19        1 -0.85700706 12.87717410          0.015625
#> 29: 2022-04-13 04:30:19        1  3.16595430  0.01777190          0.015625
#> 30: 2022-04-13 04:30:19        1 -4.93213370  9.44411853          0.015625
#> 31: 2022-04-13 04:30:19        1  3.57090440  1.81841246          0.015625
#> 32: 2022-04-13 04:30:19        1 -1.28637796  7.07054303          0.015625
#> 33: 2022-04-13 04:30:19        1  8.09350139  6.80715083          0.015625
#> 34: 2022-04-13 04:30:19        1 -2.15204658 13.39936018          0.015625
#> 35: 2022-04-13 04:30:19        1  5.33661364  0.02326276          0.015625
#> 36: 2022-04-13 04:30:19        1  3.53908667  9.93414678          0.015625
#> 37: 2022-04-13 04:30:19        1  4.31861039  5.98421072          0.015625
#> 38: 2022-04-13 04:30:19        1  6.19582266  0.70439226          0.015625
#> 39: 2022-04-13 04:30:19        1  9.83926073  6.51209852          0.015625
#> 40: 2022-04-13 04:30:19        1  3.20223018 14.01827302          0.015625
#> 41: 2022-04-13 04:30:19        1  1.87756154 14.87631569          0.015625
#> 42: 2022-04-13 04:30:19        1 -1.87872045  1.98342531          0.015625
#> 43: 2022-04-13 04:30:19        1 -0.91077299  2.33432174          0.015625
#> 44: 2022-04-13 04:30:19        1  7.41185228  0.63981174          0.015625
#> 45: 2022-04-13 04:30:19        1 -3.34881624  2.05421688          0.015625
#> 46: 2022-04-13 04:30:19        1  3.88414974 13.34688577          0.015625
#> 47: 2022-04-13 04:30:19        1 -1.65510178  1.34871275          0.015625
#> 48: 2022-04-13 04:30:19        1 -4.59724354  2.05420444          0.015625
#> 49: 2022-04-13 04:30:19        1 -1.08006719 14.75679728          0.015625
#> 50: 2022-04-13 04:30:19        1  6.63223765  9.57011388          0.015625
#> 51: 2022-04-13 04:30:19        1 -1.69569590  8.63920155          0.015625
#> 52: 2022-04-13 04:30:19        1  6.10884516  5.70679251          0.015625
#> 53: 2022-04-13 04:30:19        1  8.66905875 14.68815647          0.015625
#> 54: 2022-04-13 04:30:19        1  6.70982554 13.06052827          0.015625
#> 55: 2022-04-13 04:30:19        1  4.58790308 14.92603997          0.015625
#> 56: 2022-04-13 04:30:19        1  3.23337643 13.41891008          0.015625
#> 57: 2022-04-13 04:30:19        1  3.01056903  4.28291244          0.015625
#> 58: 2022-04-13 04:30:19        1  0.63067069  0.05891415          0.015625
#> 59: 2022-04-13 04:30:19        1 -4.91835829  7.79144560          0.015625
#> 60: 2022-04-13 04:30:19        1  3.73788699  0.32458681          0.015625
#> 61: 2022-04-13 04:30:19        1  2.03550400 13.43340606          0.015625
#> 62: 2022-04-13 04:30:19        1 -1.77329424  7.38055280          0.015625
#> 63: 2022-04-13 04:30:19        1 -1.44373709 13.45548009          0.015625
#> 64: 2022-04-13 04:30:19        1  8.98411372  6.77058884          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity