Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.804554 9.463401 0.015625 <list[3]> 2.480751

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.804554 9.463401 0.015625 <list[3]> 2.480751

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1: -1.6360184 12.5235493 0.015625     0       6          1  24.100189
#>  2:  7.1029815  4.5095514 0.015625     0       6          1  84.830006
#>  3: -2.9073643  5.3843772 0.015625     0       6          1  30.940168
#>  4:  2.4946149 14.3834750 0.015625     0       6          1 150.263239
#>  5: -4.2432227 13.5395359 0.015625     0       6          1   5.712595
#>  6: -1.1025138  4.3887915 0.015625     0       6          1  25.916323
#>  7:  4.2196833  2.7987793 0.015625     0       6          1  14.262016
#>  8: -1.6972903  0.3731707 0.015625     0       6          1  79.629914
#>  9: -4.4731439  4.7272771 0.015625     0       6          1  88.853707
#> 10:  8.4550157 11.1253920 0.015625     0       6          1 273.003767
#> 11:  9.5034224  0.6083786 0.015625     0       6          1  48.822238
#> 12:  8.2554744 13.6677437 0.015625     0       6          1 356.358169
#> 13:  3.7758948 11.1671340 0.015625     0       6          1 117.576408
#> 14:  1.0651439  8.3636439 0.015625     0       6          1  30.843548
#> 15:  4.8267277  2.7581632 0.015625     0       6          1  24.962313
#> 16:  0.7170398 14.0421052 0.015625     0       6          1 101.280709
#> 17:  7.6436253  3.0539117 0.015625     0       6          1  67.101847
#> 18:  8.9049658  3.7080903 0.015625     0       6          1  90.828797
#> 19: -3.4398872  6.9525467 0.015625     0       6          1  24.695193
#> 20:  4.6628577  9.1530846 0.015625     0       6          1 107.648360
#> 21: -1.1222757  6.1735852 0.015625     0       6          1  16.890578
#> 22:  8.1169909 14.7341752 0.015625     0       6          1 392.722074
#> 23:  4.2730739 12.4216971 0.015625     0       6          1 166.219257
#> 24:  2.8360373  6.3915074 0.015625     0       6          1  22.538684
#> 25:  9.9568250  0.6734216 0.015625     0       6          1  57.555857
#> 26:  8.8033593  7.0898422 0.015625     0       6          1 163.939403
#> 27:  3.5662824 14.3952912 0.015625     0       6          1 188.397489
#> 28:  6.4598532  3.3524038 0.015625     0       6          1  59.781706
#> 29:  2.2729081  8.9329225 0.015625     0       6          1  44.650253
#> 30: -4.4459806  5.8863830 0.015625     0       6          1  68.271251
#> 31: -3.5959576  6.7584024 0.015625     0       6          1  30.126688
#> 32:  8.0689606  5.2228736 0.015625     0       6          1 109.218462
#> 33:  2.8564781  0.1649427 0.015625     0       6          1   3.156238
#> 34:  3.4098417  8.6348643 0.015625     0       6          1  60.097529
#> 35: -2.1059097 11.3827687 0.015625     0       6          1   8.693592
#> 36: -3.8437286  1.0823269 0.015625     0       6          1 134.675925
#> 37:  1.0520132  3.4669036 0.015625     0       6          1  15.558037
#> 38:  7.1695176  9.7708159 0.015625     0       6          1 201.058219
#> 39: -1.5960767  0.4448361 0.015625     0       6          1  76.567072
#> 40:  4.9378244  2.7153384 0.015625     0       6          1  26.772809
#> 41:  8.8121380  1.8252962 0.015625     0       6          1  57.835014
#> 42:  6.3639995 12.0419332 0.015625     0       6          1 242.334878
#> 43: -0.2033389  2.7884862 0.015625     0       6          1  31.910476
#> 44:  9.5606348  8.2981416 0.015625     0       6          1 216.691378
#> 45: -2.2109683  7.5656801 0.015625     0       6          1   8.689001
#> 46: -1.0063964 12.5395220 0.015625     0       6          1  39.211477
#> 47: -2.8045539  9.4634009 0.015625     0       6          1   2.480751
#> 48: -1.5055009 14.1510651 0.015625     0       6          1  42.949128
#> 49: -1.2496665  5.3095171 0.015625     0       6          1  20.469540
#> 50: -0.9317993  0.2478362 0.015625     0       6          1  68.461287
#> 51:  1.1312173  6.2889517 0.015625     0       6          1  18.288739
#> 52: -0.3261660  1.0825742 0.015625     0       6          1  48.687358
#> 53: -3.9052541 14.3911025 0.015625     0       6          1   5.977390
#> 54:  6.3497133  2.6091478 0.015625     0       6          1  49.560286
#> 55:  6.9000493  2.6489333 0.015625     0       6          1  55.862039
#> 56:  1.8583695 14.0884646 0.015625     0       6          1 126.959440
#> 57:  8.1385230 14.8357042 0.015625     0       6          1 397.446838
#> 58:  5.1283978 12.5453517 0.015625     0       6          1 207.056689
#> 59:  4.4319155 13.8319568 0.015625     0       6          1 211.307290
#> 60:  8.7173015 13.5290815 0.015625     0       6          1 366.237251
#> 61:  8.1493877  8.4177353 0.015625     0       6          1 185.317660
#> 62:  3.1543162  3.5886024 0.015625     0       6          1   5.702145
#> 63:  5.1057184 13.8803877 0.015625     0       6          1 244.868134
#> 64:  2.2214121  4.5445557 0.015625     0       6          1   7.902630
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-06-01 04:59:18        1  -1.6360184  12.5235493          0.015625
#>  2: 2022-06-01 04:59:18        1   7.1029815   4.5095514          0.015625
#>  3: 2022-06-01 04:59:18        1  -2.9073643   5.3843772          0.015625
#>  4: 2022-06-01 04:59:18        1   2.4946149  14.3834750          0.015625
#>  5: 2022-06-01 04:59:18        1  -4.2432227  13.5395359          0.015625
#>  6: 2022-06-01 04:59:18        1  -1.1025138   4.3887915          0.015625
#>  7: 2022-06-01 04:59:18        1   4.2196833   2.7987793          0.015625
#>  8: 2022-06-01 04:59:18        1  -1.6972903   0.3731707          0.015625
#>  9: 2022-06-01 04:59:18        1  -4.4731439   4.7272771          0.015625
#> 10: 2022-06-01 04:59:18        1   8.4550157  11.1253920          0.015625
#> 11: 2022-06-01 04:59:18        1   9.5034224   0.6083786          0.015625
#> 12: 2022-06-01 04:59:18        1   8.2554744  13.6677437          0.015625
#> 13: 2022-06-01 04:59:18        1   3.7758948  11.1671340          0.015625
#> 14: 2022-06-01 04:59:18        1   1.0651439   8.3636439          0.015625
#> 15: 2022-06-01 04:59:18        1   4.8267277   2.7581632          0.015625
#> 16: 2022-06-01 04:59:18        1   0.7170398  14.0421052          0.015625
#> 17: 2022-06-01 04:59:18        1   7.6436253   3.0539117          0.015625
#> 18: 2022-06-01 04:59:18        1   8.9049658   3.7080903          0.015625
#> 19: 2022-06-01 04:59:18        1  -3.4398872   6.9525467          0.015625
#> 20: 2022-06-01 04:59:18        1   4.6628577   9.1530846          0.015625
#> 21: 2022-06-01 04:59:18        1  -1.1222757   6.1735852          0.015625
#> 22: 2022-06-01 04:59:18        1   8.1169909  14.7341752          0.015625
#> 23: 2022-06-01 04:59:18        1   4.2730739  12.4216971          0.015625
#> 24: 2022-06-01 04:59:18        1   2.8360373   6.3915074          0.015625
#> 25: 2022-06-01 04:59:18        1   9.9568250   0.6734216          0.015625
#> 26: 2022-06-01 04:59:18        1   8.8033593   7.0898422          0.015625
#> 27: 2022-06-01 04:59:18        1   3.5662824  14.3952912          0.015625
#> 28: 2022-06-01 04:59:18        1   6.4598532   3.3524038          0.015625
#> 29: 2022-06-01 04:59:18        1   2.2729081   8.9329225          0.015625
#> 30: 2022-06-01 04:59:18        1  -4.4459806   5.8863830          0.015625
#> 31: 2022-06-01 04:59:18        1  -3.5959576   6.7584024          0.015625
#> 32: 2022-06-01 04:59:18        1   8.0689606   5.2228736          0.015625
#> 33: 2022-06-01 04:59:18        1   2.8564781   0.1649427          0.015625
#> 34: 2022-06-01 04:59:18        1   3.4098417   8.6348643          0.015625
#> 35: 2022-06-01 04:59:18        1  -2.1059097  11.3827687          0.015625
#> 36: 2022-06-01 04:59:18        1  -3.8437286   1.0823269          0.015625
#> 37: 2022-06-01 04:59:18        1   1.0520132   3.4669036          0.015625
#> 38: 2022-06-01 04:59:18        1   7.1695176   9.7708159          0.015625
#> 39: 2022-06-01 04:59:18        1  -1.5960767   0.4448361          0.015625
#> 40: 2022-06-01 04:59:18        1   4.9378244   2.7153384          0.015625
#> 41: 2022-06-01 04:59:18        1   8.8121380   1.8252962          0.015625
#> 42: 2022-06-01 04:59:18        1   6.3639995  12.0419332          0.015625
#> 43: 2022-06-01 04:59:18        1  -0.2033389   2.7884862          0.015625
#> 44: 2022-06-01 04:59:18        1   9.5606348   8.2981416          0.015625
#> 45: 2022-06-01 04:59:18        1  -2.2109683   7.5656801          0.015625
#> 46: 2022-06-01 04:59:18        1  -1.0063964  12.5395220          0.015625
#> 47: 2022-06-01 04:59:18        1  -2.8045539   9.4634009          0.015625
#> 48: 2022-06-01 04:59:18        1  -1.5055009  14.1510651          0.015625
#> 49: 2022-06-01 04:59:18        1  -1.2496665   5.3095171          0.015625
#> 50: 2022-06-01 04:59:18        1  -0.9317993   0.2478362          0.015625
#> 51: 2022-06-01 04:59:18        1   1.1312173   6.2889517          0.015625
#> 52: 2022-06-01 04:59:18        1  -0.3261660   1.0825742          0.015625
#> 53: 2022-06-01 04:59:18        1  -3.9052541  14.3911025          0.015625
#> 54: 2022-06-01 04:59:18        1   6.3497133   2.6091478          0.015625
#> 55: 2022-06-01 04:59:18        1   6.9000493   2.6489333          0.015625
#> 56: 2022-06-01 04:59:18        1   1.8583695  14.0884646          0.015625
#> 57: 2022-06-01 04:59:18        1   8.1385230  14.8357042          0.015625
#> 58: 2022-06-01 04:59:18        1   5.1283978  12.5453517          0.015625
#> 59: 2022-06-01 04:59:18        1   4.4319155  13.8319568          0.015625
#> 60: 2022-06-01 04:59:18        1   8.7173015  13.5290815          0.015625
#> 61: 2022-06-01 04:59:18        1   8.1493877   8.4177353          0.015625
#> 62: 2022-06-01 04:59:18        1   3.1543162   3.5886024          0.015625
#> 63: 2022-06-01 04:59:18        1   5.1057184  13.8803877          0.015625
#> 64: 2022-06-01 04:59:18        1   2.2214121   4.5445557          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity