Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.442641 10.41896 0.015625 <list[3]> 2.859148

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.442641 10.41896 0.015625 <list[3]> 2.859148

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition          y
#>  1: -0.88859910  2.68186562 0.015625     0       6          1  38.679964
#>  2:  1.72360387 13.43765881 0.015625     0       6          1 110.336762
#>  3: -4.73109765  8.32973577 0.015625     0       6          1  44.851227
#>  4:  1.20690575 11.87653561 0.015625     0       6          1  73.520306
#>  5:  6.12569771  9.66795610 0.015625     0       6          1 169.877958
#>  6: -4.47225792  0.85550383 0.015625     0       6          1 173.540914
#>  7:  0.42846235 10.91494645 0.015625     0       6          1  49.995888
#>  8: -1.37704469  5.88889254 0.015625     0       6          1  17.423346
#>  9:  1.01635438  6.86631674 0.015625     0       6          1  21.068175
#> 10:  6.90475225  9.26408408 0.015625     0       6          1 181.325241
#> 11:  4.15187642 13.21480002 0.015625     0       6          1 181.591121
#> 12:  9.15719130 11.11841303 0.015625     0       6          1 293.637889
#> 13:  2.28844738 11.22020125 0.015625     0       6          1  79.399094
#> 14: -1.38089790  9.26467824 0.015625     0       6          1  12.829115
#> 15:  9.47012565 10.81964745 0.015625     0       6          1 293.992951
#> 16:  9.73426515  8.64353038 0.015625     0       6          1 232.582654
#> 17: -3.44264107 10.41896437 0.015625     0       6          1   2.859148
#> 18: -2.55622700  4.65185896 0.015625     0       6          1  33.551868
#> 19: -3.43696369  2.73789537 0.015625     0       6          1  83.540162
#> 20: -4.23832129  3.04179740 0.015625     0       6          1 110.802499
#> 21: -0.38951215  9.02083165 0.015625     0       6          1  24.624822
#> 22:  0.16811558 12.91929510 0.015625     0       6          1  71.105198
#> 23:  0.21789012  0.14089624 0.015625     0       6          1  49.776854
#> 24: -1.63682969 11.86991166 0.015625     0       6          1  19.494316
#> 25:  9.63624307 10.34894754 0.015625     0       6          1 283.874956
#> 26:  0.05959758  4.32816395 0.015625     0       6          1  22.072287
#> 27:  7.74257349  0.56585872 0.015625     0       6          1  36.523023
#> 28: -3.06173074 14.87183258 0.015625     0       6          1  14.197848
#> 29: -2.12915852 13.57281825 0.015625     0       6          1  21.272996
#> 30:  3.22435759 14.71458116 0.015625     0       6          1 183.400879
#> 31:  3.47999341 10.62820616 0.015625     0       6          1  96.873323
#> 32:  7.55835359  4.89078555 0.015625     0       6          1  96.771653
#> 33:  8.85479368  8.09038020 0.015625     0       6          1 191.595715
#> 34: -0.35944853 12.81316281 0.015625     0       6          1  57.889993
#> 35: -2.37059241  5.08962309 0.015625     0       6          1  26.694754
#> 36: -0.78503328 12.52462924 0.015625     0       6          1  44.420479
#> 37:  0.11259544 10.65688155 0.015625     0       6          1  42.925237
#> 38:  5.13493125  9.86113289 0.015625     0       6          1 139.891291
#> 39: -3.20528409  2.61516198 0.015625     0       6          1  77.894253
#> 40: -1.86477298 13.10897274 0.015625     0       6          1  23.492258
#> 41: -4.43662877  2.42025241 0.015625     0       6          1 133.859770
#> 42:  8.26710956 14.65877805 0.015625     0       6          1 394.821638
#> 43: -4.39506834  9.37789567 0.015625     0       6          1  24.736461
#> 44:  2.56318603  0.08773453 0.015625     0       6          1   6.100365
#> 45: -0.51760122 10.03303614 0.015625     0       6          1  28.590788
#> 46:  7.65845873  3.46384058 0.015625     0       6          1  73.476736
#> 47:  1.81961485  9.82702131 0.015625     0       6          1  51.476013
#> 48:  2.01536356 10.74180555 0.015625     0       6          1  67.092800
#> 49:  9.22965178  3.03151207 0.015625     0       6          1  83.421591
#> 50: -4.79827404 13.07969372 0.015625     0       6          1  12.423632
#> 51: -4.10677823  9.14955388 0.015625     0       6          1  19.784090
#> 52: -4.54624865  9.19536323 0.015625     0       6          1  30.274267
#> 53:  0.88680185 11.31623651 0.015625     0       6          1  61.003699
#> 54: -4.36031597 12.23425012 0.015625     0       6          1   8.352817
#> 55: -3.70881697  0.80471812 0.015625     0       6          1 134.634658
#> 56:  9.32773886  5.79702115 0.015625     0       6          1 143.661304
#> 57:  8.91836812  3.18015367 0.015625     0       6          1  81.323710
#> 58:  0.13591633  6.70328540 0.015625     0       6          1  20.358183
#> 59: -2.93327744  5.23412897 0.015625     0       6          1  33.082688
#> 60:  8.02689107  0.91710904 0.015625     0       6          1  40.966262
#> 61:  1.94585213  1.19737862 0.015625     0       6          1   9.802633
#> 62:  9.06204371 10.31452052 0.015625     0       6          1 263.859651
#> 63:  2.36212186 12.62325372 0.015625     0       6          1 107.437348
#> 64:  0.85624201  2.52609898 0.015625     0       6          1  20.844721
#>              x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-03-20 04:33:01        1 -0.88859910  2.68186562          0.015625
#>  2: 2022-03-20 04:33:01        1  1.72360387 13.43765881          0.015625
#>  3: 2022-03-20 04:33:01        1 -4.73109765  8.32973577          0.015625
#>  4: 2022-03-20 04:33:01        1  1.20690575 11.87653561          0.015625
#>  5: 2022-03-20 04:33:01        1  6.12569771  9.66795610          0.015625
#>  6: 2022-03-20 04:33:01        1 -4.47225792  0.85550383          0.015625
#>  7: 2022-03-20 04:33:01        1  0.42846235 10.91494645          0.015625
#>  8: 2022-03-20 04:33:01        1 -1.37704469  5.88889254          0.015625
#>  9: 2022-03-20 04:33:01        1  1.01635438  6.86631674          0.015625
#> 10: 2022-03-20 04:33:01        1  6.90475225  9.26408408          0.015625
#> 11: 2022-03-20 04:33:01        1  4.15187642 13.21480002          0.015625
#> 12: 2022-03-20 04:33:01        1  9.15719130 11.11841303          0.015625
#> 13: 2022-03-20 04:33:01        1  2.28844738 11.22020125          0.015625
#> 14: 2022-03-20 04:33:01        1 -1.38089790  9.26467824          0.015625
#> 15: 2022-03-20 04:33:01        1  9.47012565 10.81964745          0.015625
#> 16: 2022-03-20 04:33:01        1  9.73426515  8.64353038          0.015625
#> 17: 2022-03-20 04:33:01        1 -3.44264107 10.41896437          0.015625
#> 18: 2022-03-20 04:33:01        1 -2.55622700  4.65185896          0.015625
#> 19: 2022-03-20 04:33:01        1 -3.43696369  2.73789537          0.015625
#> 20: 2022-03-20 04:33:01        1 -4.23832129  3.04179740          0.015625
#> 21: 2022-03-20 04:33:01        1 -0.38951215  9.02083165          0.015625
#> 22: 2022-03-20 04:33:01        1  0.16811558 12.91929510          0.015625
#> 23: 2022-03-20 04:33:01        1  0.21789012  0.14089624          0.015625
#> 24: 2022-03-20 04:33:01        1 -1.63682969 11.86991166          0.015625
#> 25: 2022-03-20 04:33:01        1  9.63624307 10.34894754          0.015625
#> 26: 2022-03-20 04:33:01        1  0.05959758  4.32816395          0.015625
#> 27: 2022-03-20 04:33:01        1  7.74257349  0.56585872          0.015625
#> 28: 2022-03-20 04:33:01        1 -3.06173074 14.87183258          0.015625
#> 29: 2022-03-20 04:33:01        1 -2.12915852 13.57281825          0.015625
#> 30: 2022-03-20 04:33:01        1  3.22435759 14.71458116          0.015625
#> 31: 2022-03-20 04:33:01        1  3.47999341 10.62820616          0.015625
#> 32: 2022-03-20 04:33:01        1  7.55835359  4.89078555          0.015625
#> 33: 2022-03-20 04:33:01        1  8.85479368  8.09038020          0.015625
#> 34: 2022-03-20 04:33:01        1 -0.35944853 12.81316281          0.015625
#> 35: 2022-03-20 04:33:01        1 -2.37059241  5.08962309          0.015625
#> 36: 2022-03-20 04:33:01        1 -0.78503328 12.52462924          0.015625
#> 37: 2022-03-20 04:33:01        1  0.11259544 10.65688155          0.015625
#> 38: 2022-03-20 04:33:01        1  5.13493125  9.86113289          0.015625
#> 39: 2022-03-20 04:33:01        1 -3.20528409  2.61516198          0.015625
#> 40: 2022-03-20 04:33:01        1 -1.86477298 13.10897274          0.015625
#> 41: 2022-03-20 04:33:01        1 -4.43662877  2.42025241          0.015625
#> 42: 2022-03-20 04:33:01        1  8.26710956 14.65877805          0.015625
#> 43: 2022-03-20 04:33:01        1 -4.39506834  9.37789567          0.015625
#> 44: 2022-03-20 04:33:01        1  2.56318603  0.08773453          0.015625
#> 45: 2022-03-20 04:33:01        1 -0.51760122 10.03303614          0.015625
#> 46: 2022-03-20 04:33:01        1  7.65845873  3.46384058          0.015625
#> 47: 2022-03-20 04:33:01        1  1.81961485  9.82702131          0.015625
#> 48: 2022-03-20 04:33:01        1  2.01536356 10.74180555          0.015625
#> 49: 2022-03-20 04:33:01        1  9.22965178  3.03151207          0.015625
#> 50: 2022-03-20 04:33:01        1 -4.79827404 13.07969372          0.015625
#> 51: 2022-03-20 04:33:01        1 -4.10677823  9.14955388          0.015625
#> 52: 2022-03-20 04:33:01        1 -4.54624865  9.19536323          0.015625
#> 53: 2022-03-20 04:33:01        1  0.88680185 11.31623651          0.015625
#> 54: 2022-03-20 04:33:01        1 -4.36031597 12.23425012          0.015625
#> 55: 2022-03-20 04:33:01        1 -3.70881697  0.80471812          0.015625
#> 56: 2022-03-20 04:33:01        1  9.32773886  5.79702115          0.015625
#> 57: 2022-03-20 04:33:01        1  8.91836812  3.18015367          0.015625
#> 58: 2022-03-20 04:33:01        1  0.13591633  6.70328540          0.015625
#> 59: 2022-03-20 04:33:01        1 -2.93327744  5.23412897          0.015625
#> 60: 2022-03-20 04:33:01        1  8.02689107  0.91710904          0.015625
#> 61: 2022-03-20 04:33:01        1  1.94585213  1.19737862          0.015625
#> 62: 2022-03-20 04:33:01        1  9.06204371 10.31452052          0.015625
#> 63: 2022-03-20 04:33:01        1  2.36212186 12.62325372          0.015625
#> 64: 2022-03-20 04:33:01        1  0.85624201  2.52609898          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity