Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1      x2 fidelity  x_domain        y
#> 1: -3.278268 12.8801 0.015625 <list[3]> 2.262005

# best scoring evaluation
instance$result
#>           x1      x2 fidelity  x_domain        y
#> 1: -3.278268 12.8801 0.015625 <list[3]> 2.262005

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  6.50964559  7.9245942 0.015625     0       6          1 139.963410
#>  2:  4.98724081 10.4016699 0.015625     0       6          1 146.571701
#>  3: -3.66891236 14.3653685 0.015625     0       6          1   6.163765
#>  4: -3.27826775 12.8800956 0.015625     0       6          1   2.262005
#>  5:  2.61585360  8.0204280 0.015625     0       6          1  37.374887
#>  6:  5.86338557  9.3156193 0.015625     0       6          1 153.106362
#>  7: -3.35264804  5.7317639 0.015625     0       6          1  36.010423
#>  8:  5.86734080  8.4163481 0.015625     0       6          1 133.188708
#>  9:  0.87238942 11.9937428 0.015625     0       6          1  70.325981
#> 10:  1.99574934  0.8846391 0.015625     0       6          1  10.291014
#> 11: -3.23819127  6.4626348 0.015625     0       6          1  25.577997
#> 12:  0.92352961  1.1153516 0.015625     0       6          1  27.630838
#> 13: -0.85105088  1.5134634 0.015625     0       6          1  50.707886
#> 14:  9.83408683  5.7039310 0.015625     0       6          1 154.501163
#> 15: -0.17703536 12.9978732 0.015625     0       6          1  64.545259
#> 16: -3.25178014 13.9393734 0.015625     0       6          1   6.404272
#> 17: -0.85274381  2.6317232 0.015625     0       6          1  38.859201
#> 18:  6.96939497  8.5913804 0.015625     0       6          1 166.026242
#> 19:  1.54978311  8.3237943 0.015625     0       6          1  32.447145
#> 20: -2.92797753  0.8988534 0.015625     0       6          1 101.112013
#> 21:  1.72187791  7.2211030 0.015625     0       6          1  23.534776
#> 22:  0.13409314  4.5477471 0.015625     0       6          1  21.052002
#> 23:  4.08845559 12.5740666 0.015625     0       6          1 162.322019
#> 24: -0.09430641 13.5155582 0.015625     0       6          1  73.805491
#> 25:  7.24298204  1.4175286 0.015625     0       6          1  43.939682
#> 26:  8.38789959  7.9468703 0.015625     0       6          1 177.598643
#> 27:  2.60236405  1.3313948 0.015625     0       6          1   2.300666
#> 28:  5.09640847  1.8195002 0.015625     0       6          1  23.407374
#> 29:  9.59323301  2.5530584 0.015625     0       6          1  81.381094
#> 30:  3.43372749 11.6712611 0.015625     0       6          1 116.877026
#> 31:  7.99216573 14.0978370 0.015625     0       6          1 364.143415
#> 32:  1.77784544  9.9542379 0.015625     0       6          1  52.736467
#> 33:  9.75640577  5.7413944 0.015625     0       6          1 153.257563
#> 34: -2.84516608  1.2050967 0.015625     0       6          1  92.440272
#> 35: -0.34476715 14.2345330 0.015625     0       6          1  78.052717
#> 36:  4.40196265  0.2559163 0.015625     0       6          1   7.510536
#> 37:  2.45028080  2.3085972 0.015625     0       6          1   2.602997
#> 38:  0.02107566  8.7619069 0.015625     0       6          1  27.414444
#> 39:  1.34966742  9.9300772 0.015625     0       6          1  48.372102
#> 40: -1.73418002 10.1590128 0.015625     0       6          1  10.145115
#> 41:  5.62564441 12.7838324 0.015625     0       6          1 235.583009
#> 42: -0.73028505  3.5698066 0.015625     0       6          1  30.177409
#> 43:  6.30989844  9.8543870 0.015625     0       6          1 180.196411
#> 44:  3.10475269  5.4526727 0.015625     0       6          1  17.195176
#> 45:  4.48553430  5.3943856 0.015625     0       6          1  42.824084
#> 46:  7.00526570  4.6772250 0.015625     0       6          1  86.387883
#> 47: -0.29936500 11.4325115 0.015625     0       6          1  43.710248
#> 48: -0.42772771  6.3083274 0.015625     0       6          1  18.879985
#> 49:  0.01928599 13.8896287 0.015625     0       6          1  82.331667
#> 50:  3.00633795 11.4574385 0.015625     0       6          1  99.772467
#> 51:  0.77395206  3.1184360 0.015625     0       6          1  19.649875
#> 52:  5.62997158  5.8830074 0.015625     0       6          1  79.543549
#> 53:  7.70863812  5.1666478 0.015625     0       6          1 103.709044
#> 54:  3.57961710  4.7570658 0.015625     0       6          1  17.789810
#> 55:  9.09576309 10.3113957 0.015625     0       6          1 264.779264
#> 56:  4.87295565 13.5411193 0.015625     0       6          1 223.719831
#> 57:  8.86497894 14.3302415 0.015625     0       6          1 402.782153
#> 58:  2.47762143  2.0703105 0.015625     0       6          1   2.468532
#> 59:  1.62657111  5.2492542 0.015625     0       6          1  12.550626
#> 60: -1.84529980 12.8932700 0.015625     0       6          1  22.232654
#> 61:  5.02435644  0.7662512 0.015625     0       6          1  16.893695
#> 62:  1.79470041  3.6336797 0.015625     0       6          1   8.020847
#> 63: -3.21615402  7.5072713 0.015625     0       6          1  15.865010
#> 64: -4.21157762  1.8586611 0.015625     0       6          1 135.113668
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-15 04:55:04        1  6.50964559   7.9245942          0.015625
#>  2: 2022-08-15 04:55:04        1  4.98724081  10.4016699          0.015625
#>  3: 2022-08-15 04:55:04        1 -3.66891236  14.3653685          0.015625
#>  4: 2022-08-15 04:55:04        1 -3.27826775  12.8800956          0.015625
#>  5: 2022-08-15 04:55:04        1  2.61585360   8.0204280          0.015625
#>  6: 2022-08-15 04:55:04        1  5.86338557   9.3156193          0.015625
#>  7: 2022-08-15 04:55:04        1 -3.35264804   5.7317639          0.015625
#>  8: 2022-08-15 04:55:04        1  5.86734080   8.4163481          0.015625
#>  9: 2022-08-15 04:55:04        1  0.87238942  11.9937428          0.015625
#> 10: 2022-08-15 04:55:04        1  1.99574934   0.8846391          0.015625
#> 11: 2022-08-15 04:55:04        1 -3.23819127   6.4626348          0.015625
#> 12: 2022-08-15 04:55:04        1  0.92352961   1.1153516          0.015625
#> 13: 2022-08-15 04:55:04        1 -0.85105088   1.5134634          0.015625
#> 14: 2022-08-15 04:55:04        1  9.83408683   5.7039310          0.015625
#> 15: 2022-08-15 04:55:04        1 -0.17703536  12.9978732          0.015625
#> 16: 2022-08-15 04:55:04        1 -3.25178014  13.9393734          0.015625
#> 17: 2022-08-15 04:55:04        1 -0.85274381   2.6317232          0.015625
#> 18: 2022-08-15 04:55:04        1  6.96939497   8.5913804          0.015625
#> 19: 2022-08-15 04:55:04        1  1.54978311   8.3237943          0.015625
#> 20: 2022-08-15 04:55:04        1 -2.92797753   0.8988534          0.015625
#> 21: 2022-08-15 04:55:04        1  1.72187791   7.2211030          0.015625
#> 22: 2022-08-15 04:55:04        1  0.13409314   4.5477471          0.015625
#> 23: 2022-08-15 04:55:04        1  4.08845559  12.5740666          0.015625
#> 24: 2022-08-15 04:55:04        1 -0.09430641  13.5155582          0.015625
#> 25: 2022-08-15 04:55:04        1  7.24298204   1.4175286          0.015625
#> 26: 2022-08-15 04:55:04        1  8.38789959   7.9468703          0.015625
#> 27: 2022-08-15 04:55:04        1  2.60236405   1.3313948          0.015625
#> 28: 2022-08-15 04:55:04        1  5.09640847   1.8195002          0.015625
#> 29: 2022-08-15 04:55:04        1  9.59323301   2.5530584          0.015625
#> 30: 2022-08-15 04:55:04        1  3.43372749  11.6712611          0.015625
#> 31: 2022-08-15 04:55:04        1  7.99216573  14.0978370          0.015625
#> 32: 2022-08-15 04:55:04        1  1.77784544   9.9542379          0.015625
#> 33: 2022-08-15 04:55:04        1  9.75640577   5.7413944          0.015625
#> 34: 2022-08-15 04:55:04        1 -2.84516608   1.2050967          0.015625
#> 35: 2022-08-15 04:55:04        1 -0.34476715  14.2345330          0.015625
#> 36: 2022-08-15 04:55:04        1  4.40196265   0.2559163          0.015625
#> 37: 2022-08-15 04:55:04        1  2.45028080   2.3085972          0.015625
#> 38: 2022-08-15 04:55:04        1  0.02107566   8.7619069          0.015625
#> 39: 2022-08-15 04:55:04        1  1.34966742   9.9300772          0.015625
#> 40: 2022-08-15 04:55:04        1 -1.73418002  10.1590128          0.015625
#> 41: 2022-08-15 04:55:04        1  5.62564441  12.7838324          0.015625
#> 42: 2022-08-15 04:55:04        1 -0.73028505   3.5698066          0.015625
#> 43: 2022-08-15 04:55:04        1  6.30989844   9.8543870          0.015625
#> 44: 2022-08-15 04:55:04        1  3.10475269   5.4526727          0.015625
#> 45: 2022-08-15 04:55:04        1  4.48553430   5.3943856          0.015625
#> 46: 2022-08-15 04:55:04        1  7.00526570   4.6772250          0.015625
#> 47: 2022-08-15 04:55:04        1 -0.29936500  11.4325115          0.015625
#> 48: 2022-08-15 04:55:04        1 -0.42772771   6.3083274          0.015625
#> 49: 2022-08-15 04:55:04        1  0.01928599  13.8896287          0.015625
#> 50: 2022-08-15 04:55:04        1  3.00633795  11.4574385          0.015625
#> 51: 2022-08-15 04:55:04        1  0.77395206   3.1184360          0.015625
#> 52: 2022-08-15 04:55:04        1  5.62997158   5.8830074          0.015625
#> 53: 2022-08-15 04:55:04        1  7.70863812   5.1666478          0.015625
#> 54: 2022-08-15 04:55:04        1  3.57961710   4.7570658          0.015625
#> 55: 2022-08-15 04:55:04        1  9.09576309  10.3113957          0.015625
#> 56: 2022-08-15 04:55:04        1  4.87295565  13.5411193          0.015625
#> 57: 2022-08-15 04:55:04        1  8.86497894  14.3302415          0.015625
#> 58: 2022-08-15 04:55:04        1  2.47762143   2.0703105          0.015625
#> 59: 2022-08-15 04:55:04        1  1.62657111   5.2492542          0.015625
#> 60: 2022-08-15 04:55:04        1 -1.84529980  12.8932700          0.015625
#> 61: 2022-08-15 04:55:04        1  5.02435644   0.7662512          0.015625
#> 62: 2022-08-15 04:55:04        1  1.79470041   3.6336797          0.015625
#> 63: 2022-08-15 04:55:04        1 -3.21615402   7.5072713          0.015625
#> 64: 2022-08-15 04:55:04        1 -4.21157762   1.8586611          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity