Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.317472 1.139506 0.015625 <list[3]> 0.5525861

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.317472 1.139506 0.015625 <list[3]> 0.5525861

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition           y
#>  1:  1.75559044  1.5598571 0.015625     0       6          1  11.2660500
#>  2:  5.38951279  0.8124558 0.015625     0       6          1  22.2513913
#>  3:  1.28453178 10.5726842 0.015625     0       6          1  55.8282663
#>  4:  2.32298484  9.7514551 0.015625     0       6          1  56.4769171
#>  5: -4.81505488 14.3139594 0.015625     0       6          1  10.9879600
#>  6: -0.27849384 13.8607668 0.015625     0       6          1  74.2165306
#>  7:  0.06141349  7.8725674 0.015625     0       6          1  23.4656751
#>  8: -4.30236046  6.3713703 0.015625     0       6          1  55.8068974
#>  9:  2.73861430 14.0220185 0.015625     0       6          1 148.7908979
#> 10:  3.55071000 12.6416853 0.015625     0       6          1 142.9234391
#> 11:  1.27237608  1.7812638 0.015625     0       6          1  17.8562558
#> 12: -2.11727164  7.5940494 0.015625     0       6          1   8.6715682
#> 13:  5.11396323  8.8066619 0.015625     0       6          1 116.6066752
#> 14:  0.65116649 13.6919313 0.015625     0       6          1  93.5930197
#> 15: -0.21535150 13.3545343 0.015625     0       6          1  68.5255453
#> 16:  8.48227584 13.7071975 0.015625     0       6          1 365.1640333
#> 17:  8.32581498  5.3334734 0.015625     0       6          1 114.9021429
#> 18:  6.88758749 14.7704371 0.015625     0       6          1 351.8316880
#> 19:  1.65053195 14.0652909 0.015625     0       6          1 121.7739968
#> 20: -0.53723416  5.5225269 0.015625     0       6          1  20.0487440
#> 21: -3.79502756 10.1731918 0.015625     0       6          1   7.7102072
#> 22: -0.27484616  8.4238731 0.015625     0       6          1  23.1784451
#> 23:  5.87121754  4.8251139 0.015625     0       6          1  69.3446022
#> 24:  7.02369516 13.7228955 0.015625     0       6          1 319.3130224
#> 25: -4.96789997 12.8550475 0.015625     0       6          1  15.7045584
#> 26:  5.79046663 14.8264668 0.015625     0       6          1 307.8459274
#> 27:  3.06561086  0.4219413 0.015625     0       6          1   1.4016308
#> 28:  0.33521687  7.8675366 0.015625     0       6          1  24.8161147
#> 29:  0.05377174  0.6572043 0.015625     0       6          1  47.2274816
#> 30: -2.86891439  1.0521931 0.015625     0       6          1  96.1448839
#> 31:  0.12231487  8.2068852 0.015625     0       6          1  25.2956326
#> 32:  8.40628451  7.8385382 0.015625     0       6          1 175.1290429
#> 33: -2.94822991 14.0864289 0.015625     0       6          1  10.3544675
#> 34:  6.44668213  2.3321654 0.015625     0       6          1  47.7184155
#> 35:  8.61954047 12.6441708 0.015625     0       6          1 330.1676660
#> 36: -1.16594096  3.0061359 0.015625     0       6          1  37.7070971
#> 37:  8.96453803  4.3297298 0.015625     0       6          1 103.9392828
#> 38:  8.88223134  2.7408006 0.015625     0       6          1  73.2055691
#> 39:  6.84324916  1.8065544 0.015625     0       6          1  45.7820924
#> 40: -3.14567901  3.2550564 0.015625     0       6          1  65.2922383
#> 41:  2.31591056 12.4361605 0.015625     0       6          1 102.6338057
#> 42:  3.12354601  2.5256407 0.015625     0       6          1   1.8321024
#> 43:  2.57423462  5.9940997 0.015625     0       6          1  17.0139609
#> 44: -1.60191655  1.1690228 0.015625     0       6          1  65.3439901
#> 45: -2.23977228  6.2066780 0.015625     0       6          1  16.3810097
#> 46: -3.51004841  0.3966640 0.015625     0       6          1 134.8741073
#> 47:  7.97536294 13.2730992 0.015625     0       6          1 333.2182093
#> 48:  7.27833627  3.3850153 0.015625     0       6          1  69.1034909
#> 49:  1.82322713  3.9448863 0.015625     0       6          1   8.1559738
#> 50:  3.67286007 11.6410492 0.015625     0       6          1 124.3064417
#> 51:  8.96902960 13.2505970 0.015625     0       6          1 364.3511036
#> 52:  3.31747231  1.1395059 0.015625     0       6          1   0.5525861
#> 53: -4.65602051  0.3613472 0.015625     0       6          1 197.5736819
#> 54:  3.62964612  8.0011877 0.015625     0       6          1  55.8782846
#> 55:  3.92456911 10.5665841 0.015625     0       6          1 110.0920235
#> 56:  7.75157031  0.5528747 0.015625     0       6          1  36.4073571
#> 57:  4.36933983  2.7376270 0.015625     0       6          1  16.4090974
#> 58: -4.76077557  5.6072598 0.015625     0       6          1  85.5749134
#> 59:  9.27865494  1.4432194 0.015625     0       6          1  57.7073799
#> 60:  8.32798925  3.2395065 0.015625     0       6          1  75.5307703
#> 61:  7.26393208  1.6804692 0.015625     0       6          1  46.9160133
#> 62:  7.40556253  2.0405829 0.015625     0       6          1  51.8706605
#> 63: -4.87374031  6.4507759 0.015625     0       6          1  76.1257946
#> 64: -1.56511502  4.2785209 0.015625     0       6          1  28.4393902
#>              x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-04-25 04:38:31        1  1.75559044   1.5598571          0.015625
#>  2: 2022-04-25 04:38:31        1  5.38951279   0.8124558          0.015625
#>  3: 2022-04-25 04:38:31        1  1.28453178  10.5726842          0.015625
#>  4: 2022-04-25 04:38:31        1  2.32298484   9.7514551          0.015625
#>  5: 2022-04-25 04:38:31        1 -4.81505488  14.3139594          0.015625
#>  6: 2022-04-25 04:38:31        1 -0.27849384  13.8607668          0.015625
#>  7: 2022-04-25 04:38:31        1  0.06141349   7.8725674          0.015625
#>  8: 2022-04-25 04:38:31        1 -4.30236046   6.3713703          0.015625
#>  9: 2022-04-25 04:38:31        1  2.73861430  14.0220185          0.015625
#> 10: 2022-04-25 04:38:31        1  3.55071000  12.6416853          0.015625
#> 11: 2022-04-25 04:38:31        1  1.27237608   1.7812638          0.015625
#> 12: 2022-04-25 04:38:31        1 -2.11727164   7.5940494          0.015625
#> 13: 2022-04-25 04:38:31        1  5.11396323   8.8066619          0.015625
#> 14: 2022-04-25 04:38:31        1  0.65116649  13.6919313          0.015625
#> 15: 2022-04-25 04:38:31        1 -0.21535150  13.3545343          0.015625
#> 16: 2022-04-25 04:38:31        1  8.48227584  13.7071975          0.015625
#> 17: 2022-04-25 04:38:31        1  8.32581498   5.3334734          0.015625
#> 18: 2022-04-25 04:38:31        1  6.88758749  14.7704371          0.015625
#> 19: 2022-04-25 04:38:31        1  1.65053195  14.0652909          0.015625
#> 20: 2022-04-25 04:38:31        1 -0.53723416   5.5225269          0.015625
#> 21: 2022-04-25 04:38:31        1 -3.79502756  10.1731918          0.015625
#> 22: 2022-04-25 04:38:31        1 -0.27484616   8.4238731          0.015625
#> 23: 2022-04-25 04:38:31        1  5.87121754   4.8251139          0.015625
#> 24: 2022-04-25 04:38:31        1  7.02369516  13.7228955          0.015625
#> 25: 2022-04-25 04:38:31        1 -4.96789997  12.8550475          0.015625
#> 26: 2022-04-25 04:38:31        1  5.79046663  14.8264668          0.015625
#> 27: 2022-04-25 04:38:31        1  3.06561086   0.4219413          0.015625
#> 28: 2022-04-25 04:38:31        1  0.33521687   7.8675366          0.015625
#> 29: 2022-04-25 04:38:31        1  0.05377174   0.6572043          0.015625
#> 30: 2022-04-25 04:38:31        1 -2.86891439   1.0521931          0.015625
#> 31: 2022-04-25 04:38:31        1  0.12231487   8.2068852          0.015625
#> 32: 2022-04-25 04:38:31        1  8.40628451   7.8385382          0.015625
#> 33: 2022-04-25 04:38:31        1 -2.94822991  14.0864289          0.015625
#> 34: 2022-04-25 04:38:31        1  6.44668213   2.3321654          0.015625
#> 35: 2022-04-25 04:38:31        1  8.61954047  12.6441708          0.015625
#> 36: 2022-04-25 04:38:31        1 -1.16594096   3.0061359          0.015625
#> 37: 2022-04-25 04:38:31        1  8.96453803   4.3297298          0.015625
#> 38: 2022-04-25 04:38:31        1  8.88223134   2.7408006          0.015625
#> 39: 2022-04-25 04:38:31        1  6.84324916   1.8065544          0.015625
#> 40: 2022-04-25 04:38:31        1 -3.14567901   3.2550564          0.015625
#> 41: 2022-04-25 04:38:31        1  2.31591056  12.4361605          0.015625
#> 42: 2022-04-25 04:38:31        1  3.12354601   2.5256407          0.015625
#> 43: 2022-04-25 04:38:31        1  2.57423462   5.9940997          0.015625
#> 44: 2022-04-25 04:38:31        1 -1.60191655   1.1690228          0.015625
#> 45: 2022-04-25 04:38:31        1 -2.23977228   6.2066780          0.015625
#> 46: 2022-04-25 04:38:31        1 -3.51004841   0.3966640          0.015625
#> 47: 2022-04-25 04:38:31        1  7.97536294  13.2730992          0.015625
#> 48: 2022-04-25 04:38:31        1  7.27833627   3.3850153          0.015625
#> 49: 2022-04-25 04:38:31        1  1.82322713   3.9448863          0.015625
#> 50: 2022-04-25 04:38:31        1  3.67286007  11.6410492          0.015625
#> 51: 2022-04-25 04:38:31        1  8.96902960  13.2505970          0.015625
#> 52: 2022-04-25 04:38:31        1  3.31747231   1.1395059          0.015625
#> 53: 2022-04-25 04:38:31        1 -4.65602051   0.3613472          0.015625
#> 54: 2022-04-25 04:38:31        1  3.62964612   8.0011877          0.015625
#> 55: 2022-04-25 04:38:31        1  3.92456911  10.5665841          0.015625
#> 56: 2022-04-25 04:38:31        1  7.75157031   0.5528747          0.015625
#> 57: 2022-04-25 04:38:31        1  4.36933983   2.7376270          0.015625
#> 58: 2022-04-25 04:38:31        1 -4.76077557   5.6072598          0.015625
#> 59: 2022-04-25 04:38:31        1  9.27865494   1.4432194          0.015625
#> 60: 2022-04-25 04:38:31        1  8.32798925   3.2395065          0.015625
#> 61: 2022-04-25 04:38:31        1  7.26393208   1.6804692          0.015625
#> 62: 2022-04-25 04:38:31        1  7.40556253   2.0405829          0.015625
#> 63: 2022-04-25 04:38:31        1 -4.87374031   6.4507759          0.015625
#> 64: 2022-04-25 04:38:31        1 -1.56511502   4.2785209          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity