Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.025103 1.088722 0.015625 <list[3]> 0.6058849

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.025103 1.088722 0.015625 <list[3]> 0.6058849

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -0.8236127  5.1730632 0.015625     0       6          1  21.1849733
#>  2:  2.0307060  8.7241811 0.015625     0       6          1  39.7193356
#>  3:  9.0780996  2.0958452 0.015625     0       6          1  65.1319780
#>  4: -2.6693754  4.1662522 0.015625     0       6          1  41.1548612
#>  5: -0.4872554  0.4219784 0.015625     0       6          1  58.9445660
#>  6:  3.7647755  3.8573012 0.015625     0       6          1  13.8536905
#>  7:  5.4581349  3.1178530 0.015625     0       6          1  40.4151312
#>  8: -1.4853824  5.5853477 0.015625     0       6          1  18.9220087
#>  9: -1.4678338  4.1000527 0.015625     0       6          1  29.4968939
#> 10:  6.2632710  1.3932701 0.015625     0       6          1  36.8676781
#> 11: -2.4345132  3.7627779 0.015625     0       6          1  42.3156101
#> 12:  9.1313226 13.9781113 0.015625     0       6          1 398.7049355
#> 13:  5.9772290  9.4404653 0.015625     0       6          1 159.6975449
#> 14: -3.1191161  9.4747845 0.015625     0       6          1   3.5993204
#> 15: -2.0122995 11.2818336 0.015625     0       6          1   9.7176995
#> 16:  8.5631646  2.8216220 0.015625     0       6          1  70.9166586
#> 17:  3.3430267  9.0145055 0.015625     0       6          1  64.4557005
#> 18:  6.9443570 12.0805506 0.015625     0       6          1 262.5040518
#> 19:  8.3449619 14.2963462 0.015625     0       6          1 383.2538576
#> 20:  0.8146697 12.8474545 0.015625     0       6          1  82.5815428
#> 21:  2.7550583 13.1443738 0.015625     0       6          1 128.7015263
#> 22:  7.7366437  2.4551981 0.015625     0       6          1  59.1222142
#> 23:  5.6430658  1.1839732 0.015625     0       6          1  27.8522193
#> 24:  5.6981841  0.4557713 0.015625     0       6          1  24.3879821
#> 25:  2.8810226 10.0219678 0.015625     0       6          1  70.4787392
#> 26:  0.2255964  5.8902292 0.015625     0       6          1  19.4201648
#> 27: -0.4560087 14.9711144 0.015625     0       6          1  86.5014113
#> 28:  5.1745732  7.5896429 0.015625     0       6          1  95.3166118
#> 29:  6.4402356  3.2366529 0.015625     0       6          1  58.0644805
#> 30: -1.3178843  4.7888441 0.015625     0       6          1  23.7059685
#> 31: -2.6761993  7.3527099 0.015625     0       6          1  11.1960281
#> 32: -0.1905861  2.0305145 0.015625     0       6          1  37.6947248
#> 33: -1.8049449  4.1527953 0.015625     0       6          1  31.0048763
#> 34: -0.8632206  0.4961509 0.015625     0       6          1  63.8598452
#> 35: -4.3829300 10.7454887 0.015625     0       6          1  14.8504053
#> 36:  6.7203356 12.3247828 0.015625     0       6          1 263.0555895
#> 37: -0.2694167  8.8607820 0.015625     0       6          1  25.1594643
#> 38:  9.8469507  7.9351459 0.015625     0       6          1 215.1534982
#> 39:  7.3165461 10.4109375 0.015625     0       6          1 222.5533246
#> 40: -1.3437784  1.4078228 0.015625     0       6          1  58.2162388
#> 41:  2.7502832  1.3967578 0.015625     0       6          1   1.3340235
#> 42:  2.5777973  8.5385872 0.015625     0       6          1  43.3184832
#> 43: -4.3040593  6.7464491 0.015625     0       6          1  50.7196232
#> 44:  8.3557537  7.3063406 0.015625     0       6          1 160.5888700
#> 45: -2.5209169  8.2648344 0.015625     0       6          1   5.9630090
#> 46:  6.6235665 10.3719165 0.015625     0       6          1 203.0530767
#> 47:  1.6411621  8.0196587 0.015625     0       6          1  30.0167960
#> 48:  6.6876796 13.9118630 0.015625     0       6          1 313.9959209
#> 49: -3.4222091  2.3959623 0.015625     0       6          1  89.3356296
#> 50:  3.9709083  8.7855145 0.015625     0       6          1  77.8294751
#> 51:  4.8782796 13.0735557 0.015625     0       6          1 210.5612697
#> 52: -4.9559057  7.9785626 0.015625     0       6          1  56.7265769
#> 53:  7.2807749 14.9175150 0.015625     0       6          1 371.4859231
#> 54:  2.7952390 10.7265958 0.015625     0       6          1  80.8044504
#> 55: -0.4020574  7.3369266 0.015625     0       6          1  19.3153695
#> 56:  5.1503278 14.8612518 0.015625     0       6          1 277.8960627
#> 57: -2.3243961  6.1429526 0.015625     0       6          1  17.2870256
#> 58:  3.8298805  3.1779734 0.015625     0       6          1  10.5500152
#> 59:  3.0251032  1.0887222 0.015625     0       6          1   0.6058849
#> 60: -4.8400673  1.1678494 0.015625     0       6          1 186.9347547
#> 61: -2.0789936  4.5438159 0.015625     0       6          1  29.3170099
#> 62:  0.7875977 12.1448597 0.015625     0       6          1  71.2286562
#> 63: -3.5469412  0.2165236 0.015625     0       6          1 140.7803890
#> 64:  2.2806801 10.6181352 0.015625     0       6          1  69.1579278
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-17 04:36:30        1  -0.8236127   5.1730632          0.015625
#>  2: 2022-05-17 04:36:30        1   2.0307060   8.7241811          0.015625
#>  3: 2022-05-17 04:36:30        1   9.0780996   2.0958452          0.015625
#>  4: 2022-05-17 04:36:30        1  -2.6693754   4.1662522          0.015625
#>  5: 2022-05-17 04:36:30        1  -0.4872554   0.4219784          0.015625
#>  6: 2022-05-17 04:36:30        1   3.7647755   3.8573012          0.015625
#>  7: 2022-05-17 04:36:30        1   5.4581349   3.1178530          0.015625
#>  8: 2022-05-17 04:36:30        1  -1.4853824   5.5853477          0.015625
#>  9: 2022-05-17 04:36:30        1  -1.4678338   4.1000527          0.015625
#> 10: 2022-05-17 04:36:30        1   6.2632710   1.3932701          0.015625
#> 11: 2022-05-17 04:36:30        1  -2.4345132   3.7627779          0.015625
#> 12: 2022-05-17 04:36:30        1   9.1313226  13.9781113          0.015625
#> 13: 2022-05-17 04:36:30        1   5.9772290   9.4404653          0.015625
#> 14: 2022-05-17 04:36:30        1  -3.1191161   9.4747845          0.015625
#> 15: 2022-05-17 04:36:30        1  -2.0122995  11.2818336          0.015625
#> 16: 2022-05-17 04:36:30        1   8.5631646   2.8216220          0.015625
#> 17: 2022-05-17 04:36:30        1   3.3430267   9.0145055          0.015625
#> 18: 2022-05-17 04:36:30        1   6.9443570  12.0805506          0.015625
#> 19: 2022-05-17 04:36:30        1   8.3449619  14.2963462          0.015625
#> 20: 2022-05-17 04:36:30        1   0.8146697  12.8474545          0.015625
#> 21: 2022-05-17 04:36:30        1   2.7550583  13.1443738          0.015625
#> 22: 2022-05-17 04:36:30        1   7.7366437   2.4551981          0.015625
#> 23: 2022-05-17 04:36:30        1   5.6430658   1.1839732          0.015625
#> 24: 2022-05-17 04:36:30        1   5.6981841   0.4557713          0.015625
#> 25: 2022-05-17 04:36:30        1   2.8810226  10.0219678          0.015625
#> 26: 2022-05-17 04:36:30        1   0.2255964   5.8902292          0.015625
#> 27: 2022-05-17 04:36:30        1  -0.4560087  14.9711144          0.015625
#> 28: 2022-05-17 04:36:30        1   5.1745732   7.5896429          0.015625
#> 29: 2022-05-17 04:36:30        1   6.4402356   3.2366529          0.015625
#> 30: 2022-05-17 04:36:30        1  -1.3178843   4.7888441          0.015625
#> 31: 2022-05-17 04:36:30        1  -2.6761993   7.3527099          0.015625
#> 32: 2022-05-17 04:36:30        1  -0.1905861   2.0305145          0.015625
#> 33: 2022-05-17 04:36:30        1  -1.8049449   4.1527953          0.015625
#> 34: 2022-05-17 04:36:30        1  -0.8632206   0.4961509          0.015625
#> 35: 2022-05-17 04:36:30        1  -4.3829300  10.7454887          0.015625
#> 36: 2022-05-17 04:36:30        1   6.7203356  12.3247828          0.015625
#> 37: 2022-05-17 04:36:30        1  -0.2694167   8.8607820          0.015625
#> 38: 2022-05-17 04:36:30        1   9.8469507   7.9351459          0.015625
#> 39: 2022-05-17 04:36:30        1   7.3165461  10.4109375          0.015625
#> 40: 2022-05-17 04:36:30        1  -1.3437784   1.4078228          0.015625
#> 41: 2022-05-17 04:36:30        1   2.7502832   1.3967578          0.015625
#> 42: 2022-05-17 04:36:30        1   2.5777973   8.5385872          0.015625
#> 43: 2022-05-17 04:36:30        1  -4.3040593   6.7464491          0.015625
#> 44: 2022-05-17 04:36:30        1   8.3557537   7.3063406          0.015625
#> 45: 2022-05-17 04:36:30        1  -2.5209169   8.2648344          0.015625
#> 46: 2022-05-17 04:36:30        1   6.6235665  10.3719165          0.015625
#> 47: 2022-05-17 04:36:30        1   1.6411621   8.0196587          0.015625
#> 48: 2022-05-17 04:36:30        1   6.6876796  13.9118630          0.015625
#> 49: 2022-05-17 04:36:30        1  -3.4222091   2.3959623          0.015625
#> 50: 2022-05-17 04:36:30        1   3.9709083   8.7855145          0.015625
#> 51: 2022-05-17 04:36:30        1   4.8782796  13.0735557          0.015625
#> 52: 2022-05-17 04:36:30        1  -4.9559057   7.9785626          0.015625
#> 53: 2022-05-17 04:36:30        1   7.2807749  14.9175150          0.015625
#> 54: 2022-05-17 04:36:30        1   2.7952390  10.7265958          0.015625
#> 55: 2022-05-17 04:36:30        1  -0.4020574   7.3369266          0.015625
#> 56: 2022-05-17 04:36:30        1   5.1503278  14.8612518          0.015625
#> 57: 2022-05-17 04:36:30        1  -2.3243961   6.1429526          0.015625
#> 58: 2022-05-17 04:36:30        1   3.8298805   3.1779734          0.015625
#> 59: 2022-05-17 04:36:30        1   3.0251032   1.0887222          0.015625
#> 60: 2022-05-17 04:36:30        1  -4.8400673   1.1678494          0.015625
#> 61: 2022-05-17 04:36:30        1  -2.0789936   4.5438159          0.015625
#> 62: 2022-05-17 04:36:30        1   0.7875977  12.1448597          0.015625
#> 63: 2022-05-17 04:36:30        1  -3.5469412   0.2165236          0.015625
#> 64: 2022-05-17 04:36:30        1   2.2806801  10.6181352          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity