Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.650473 12.02996 0.015625 <list[3]> 1.650546

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.650473 12.02996 0.015625 <list[3]> 1.650546

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  2.08863529  2.7079917 0.015625     0       6          1   5.257318
#>  2:  4.27628475  7.0797746 0.015625     0       6          1  59.576650
#>  3:  0.08734115  3.5511458 0.015625     0       6          1  24.901985
#>  4:  7.89471632 11.7159687 0.015625     0       6          1 277.404031
#>  5:  1.69137694  6.1131078 0.015625     0       6          1  16.227380
#>  6:  1.76503726  2.4236667 0.015625     0       6          1   8.891313
#>  7: -4.47262717  3.3050732 0.015625     0       6          1 116.471495
#>  8:  1.58943032 12.4020133 0.015625     0       6          1  88.214301
#>  9:  8.90938977  0.9451983 0.015625     0       6          1  46.325325
#> 10:  7.60170081  3.9894136 0.015625     0       6          1  81.472094
#> 11: -4.46010733  5.8975883 0.015625     0       6          1  68.638798
#> 12:  7.34680154  3.6783485 0.015625     0       6          1  74.132061
#> 13:  7.26282264  9.3505793 0.015625     0       6          1 191.918683
#> 14:  4.09837767  9.9367419 0.015625     0       6          1 103.332264
#> 15:  2.29423237  0.1490921 0.015625     0       6          1   9.219771
#> 16:  1.72981618  9.3756251 0.015625     0       6          1  44.921367
#> 17:  7.26328291 10.2045988 0.015625     0       6          1 215.355449
#> 18:  1.56310778  7.7411524 0.015625     0       6          1  27.327809
#> 19:  2.46942968 12.8438325 0.015625     0       6          1 114.561698
#> 20:  5.24684069 12.3083198 0.015625     0       6          1 205.675661
#> 21:  4.95847889  1.0780606 0.015625     0       6          1  17.239952
#> 22:  4.22039864 12.4621927 0.015625     0       6          1 165.019118
#> 23:  5.44605455  2.6932440 0.015625     0       6          1  36.222857
#> 24:  8.95960620 14.4661158 0.015625     0       6          1 411.787301
#> 25: -3.65047314 12.0299577 0.015625     0       6          1   1.650546
#> 26:  4.39663373 11.5943343 0.015625     0       6          1 150.956843
#> 27: -2.73282899  2.6521921 0.015625     0       6          1  64.024243
#> 28: -3.68317683 12.5014630 0.015625     0       6          1   1.821476
#> 29: -3.25541632  9.0745428 0.015625     0       6          1   6.376892
#> 30:  7.13084888  0.2517327 0.015625     0       6          1  32.654478
#> 31:  3.53196704  8.2158832 0.015625     0       6          1  56.676749
#> 32: -0.90871454  8.8397503 0.015625     0       6          1  17.774684
#> 33:  5.57101692 14.5645651 0.015625     0       6          1 288.754632
#> 34:  4.85398287  4.3168160 0.015625     0       6          1  39.633365
#> 35:  9.39792933  4.7942839 0.015625     0       6          1 122.193409
#> 36:  3.85492374 12.3893851 0.015625     0       6          1 148.370154
#> 37:  1.01045970 14.7281548 0.015625     0       6          1 121.295402
#> 38: -0.78894986 10.0454670 0.015625     0       6          1  24.442207
#> 39:  1.32908263 11.7758659 0.015625     0       6          1  73.714697
#> 40:  7.49007825 14.6472696 0.015625     0       6          1 368.482047
#> 41: -3.70839196 14.9024659 0.015625     0       6          1   8.543134
#> 42: -3.66466740 12.1089739 0.015625     0       6          1   1.700431
#> 43:  1.80487995  0.1833746 0.015625     0       6          1  17.040116
#> 44:  4.49667664  1.0003298 0.015625     0       6          1  10.301894
#> 45:  2.94577808  5.1867407 0.015625     0       6          1  13.601084
#> 46: -0.06917811 11.0111554 0.015625     0       6          1  43.598043
#> 47:  3.36620392 10.3864731 0.015625     0       6          1  88.915419
#> 48:  3.79259419 11.2339685 0.015625     0       6          1 119.603243
#> 49:  2.14702815  8.9092886 0.015625     0       6          1  43.018067
#> 50:  9.54759700  3.7348342 0.015625     0       6          1 103.036753
#> 51:  6.18623233  3.5817784 0.015625     0       6          1  58.629534
#> 52: -1.94211179  1.0568183 0.015625     0       6          1  72.940412
#> 53:  0.53034209  6.7808570 0.015625     0       6          1  20.895453
#> 54:  6.77338012 10.8183055 0.015625     0       6          1 219.766258
#> 55:  2.66695831  8.1760907 0.015625     0       6          1  39.924029
#> 56:  6.55812803  1.1369616 0.015625     0       6          1  37.322219
#> 57: -4.59768221  1.8030868 0.015625     0       6          1 156.871253
#> 58: -0.49439320  9.5054490 0.015625     0       6          1  25.802293
#> 59:  3.91877826  2.7489658 0.015625     0       6          1   9.473532
#> 60:  6.91784129  0.7755574 0.015625     0       6          1  36.344658
#> 61: -3.82344698  9.6938148 0.015625     0       6          1  10.615426
#> 62:  1.93762389  3.9859291 0.015625     0       6          1   7.466867
#> 63:  7.42967528  2.5208423 0.015625     0       6          1  58.152982
#> 64:  7.45021122 14.2012645 0.015625     0       6          1 350.568878
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-04 08:18:32        1  2.08863529   2.7079917          0.015625
#>  2: 2022-05-04 08:18:32        1  4.27628475   7.0797746          0.015625
#>  3: 2022-05-04 08:18:32        1  0.08734115   3.5511458          0.015625
#>  4: 2022-05-04 08:18:32        1  7.89471632  11.7159687          0.015625
#>  5: 2022-05-04 08:18:32        1  1.69137694   6.1131078          0.015625
#>  6: 2022-05-04 08:18:32        1  1.76503726   2.4236667          0.015625
#>  7: 2022-05-04 08:18:32        1 -4.47262717   3.3050732          0.015625
#>  8: 2022-05-04 08:18:32        1  1.58943032  12.4020133          0.015625
#>  9: 2022-05-04 08:18:32        1  8.90938977   0.9451983          0.015625
#> 10: 2022-05-04 08:18:32        1  7.60170081   3.9894136          0.015625
#> 11: 2022-05-04 08:18:32        1 -4.46010733   5.8975883          0.015625
#> 12: 2022-05-04 08:18:32        1  7.34680154   3.6783485          0.015625
#> 13: 2022-05-04 08:18:32        1  7.26282264   9.3505793          0.015625
#> 14: 2022-05-04 08:18:32        1  4.09837767   9.9367419          0.015625
#> 15: 2022-05-04 08:18:32        1  2.29423237   0.1490921          0.015625
#> 16: 2022-05-04 08:18:32        1  1.72981618   9.3756251          0.015625
#> 17: 2022-05-04 08:18:32        1  7.26328291  10.2045988          0.015625
#> 18: 2022-05-04 08:18:32        1  1.56310778   7.7411524          0.015625
#> 19: 2022-05-04 08:18:32        1  2.46942968  12.8438325          0.015625
#> 20: 2022-05-04 08:18:32        1  5.24684069  12.3083198          0.015625
#> 21: 2022-05-04 08:18:32        1  4.95847889   1.0780606          0.015625
#> 22: 2022-05-04 08:18:32        1  4.22039864  12.4621927          0.015625
#> 23: 2022-05-04 08:18:32        1  5.44605455   2.6932440          0.015625
#> 24: 2022-05-04 08:18:32        1  8.95960620  14.4661158          0.015625
#> 25: 2022-05-04 08:18:32        1 -3.65047314  12.0299577          0.015625
#> 26: 2022-05-04 08:18:32        1  4.39663373  11.5943343          0.015625
#> 27: 2022-05-04 08:18:32        1 -2.73282899   2.6521921          0.015625
#> 28: 2022-05-04 08:18:32        1 -3.68317683  12.5014630          0.015625
#> 29: 2022-05-04 08:18:32        1 -3.25541632   9.0745428          0.015625
#> 30: 2022-05-04 08:18:32        1  7.13084888   0.2517327          0.015625
#> 31: 2022-05-04 08:18:32        1  3.53196704   8.2158832          0.015625
#> 32: 2022-05-04 08:18:32        1 -0.90871454   8.8397503          0.015625
#> 33: 2022-05-04 08:18:32        1  5.57101692  14.5645651          0.015625
#> 34: 2022-05-04 08:18:32        1  4.85398287   4.3168160          0.015625
#> 35: 2022-05-04 08:18:32        1  9.39792933   4.7942839          0.015625
#> 36: 2022-05-04 08:18:32        1  3.85492374  12.3893851          0.015625
#> 37: 2022-05-04 08:18:32        1  1.01045970  14.7281548          0.015625
#> 38: 2022-05-04 08:18:32        1 -0.78894986  10.0454670          0.015625
#> 39: 2022-05-04 08:18:32        1  1.32908263  11.7758659          0.015625
#> 40: 2022-05-04 08:18:32        1  7.49007825  14.6472696          0.015625
#> 41: 2022-05-04 08:18:32        1 -3.70839196  14.9024659          0.015625
#> 42: 2022-05-04 08:18:32        1 -3.66466740  12.1089739          0.015625
#> 43: 2022-05-04 08:18:32        1  1.80487995   0.1833746          0.015625
#> 44: 2022-05-04 08:18:32        1  4.49667664   1.0003298          0.015625
#> 45: 2022-05-04 08:18:32        1  2.94577808   5.1867407          0.015625
#> 46: 2022-05-04 08:18:32        1 -0.06917811  11.0111554          0.015625
#> 47: 2022-05-04 08:18:32        1  3.36620392  10.3864731          0.015625
#> 48: 2022-05-04 08:18:32        1  3.79259419  11.2339685          0.015625
#> 49: 2022-05-04 08:18:32        1  2.14702815   8.9092886          0.015625
#> 50: 2022-05-04 08:18:32        1  9.54759700   3.7348342          0.015625
#> 51: 2022-05-04 08:18:32        1  6.18623233   3.5817784          0.015625
#> 52: 2022-05-04 08:18:32        1 -1.94211179   1.0568183          0.015625
#> 53: 2022-05-04 08:18:32        1  0.53034209   6.7808570          0.015625
#> 54: 2022-05-04 08:18:32        1  6.77338012  10.8183055          0.015625
#> 55: 2022-05-04 08:18:32        1  2.66695831   8.1760907          0.015625
#> 56: 2022-05-04 08:18:32        1  6.55812803   1.1369616          0.015625
#> 57: 2022-05-04 08:18:32        1 -4.59768221   1.8030868          0.015625
#> 58: 2022-05-04 08:18:32        1 -0.49439320   9.5054490          0.015625
#> 59: 2022-05-04 08:18:32        1  3.91877826   2.7489658          0.015625
#> 60: 2022-05-04 08:18:32        1  6.91784129   0.7755574          0.015625
#> 61: 2022-05-04 08:18:32        1 -3.82344698   9.6938148          0.015625
#> 62: 2022-05-04 08:18:32        1  1.93762389   3.9859291          0.015625
#> 63: 2022-05-04 08:18:32        1  7.42967528   2.5208423          0.015625
#> 64: 2022-05-04 08:18:32        1  7.45021122  14.2012645          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity