Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.628207 11.89862 0.015625 <list[3]> 3.895686

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.628207 11.89862 0.015625 <list[3]> 3.895686

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1: -1.60490555 11.6169979 0.015625     0       6          1  18.573909
#>  2: -3.57051788  4.6078090 0.015625     0       6          1  57.021231
#>  3:  6.71701525 14.8388918 0.015625     0       6          1 347.848426
#>  4: -2.63140922 12.6553692 0.015625     0       6          1   6.703211
#>  5:  2.00731986  6.2272571 0.015625     0       6          1  16.817877
#>  6:  9.69475342  3.1784078 0.015625     0       6          1  95.189690
#>  7:  1.66356633  5.3433634 0.015625     0       6          1  12.743022
#>  8:  1.82698976 10.5453226 0.015625     0       6          1  61.595747
#>  9:  0.17588986  4.4539709 0.015625     0       6          1  21.059361
#> 10:  6.30331671 13.7564183 0.015625     0       6          1 294.059837
#> 11:  9.84941259  8.1262799 0.015625     0       6          1 220.862637
#> 12:  6.92798818  0.1815256 0.015625     0       6          1  31.601981
#> 13:  9.52918625  9.4857963 0.015625     0       6          1 251.988922
#> 14: -2.67164666  1.6847194 0.015625     0       6          1  78.646700
#> 15:  3.57084322  9.4132083 0.015625     0       6          1  77.034355
#> 16: -2.37006283  7.7741814 0.015625     0       6          1   7.828291
#> 17:  6.42534061  1.6093075 0.015625     0       6          1  40.355113
#> 18:  9.39579244 14.9045310 0.015625     0       6          1 447.471711
#> 19: -1.51954420 13.8026454 0.015625     0       6          1  38.722224
#> 20:  0.61536158  2.1680156 0.015625     0       6          1  26.044669
#> 21: -1.05667296 13.4145053 0.015625     0       6          1  47.194121
#> 22:  5.57864407  0.9846969 0.015625     0       6          1  25.763669
#> 23:  4.31912586  5.8845230 0.015625     0       6          1  44.575203
#> 24: -0.09657184  6.2481836 0.015625     0       6          1  19.566245
#> 25:  7.89604818 14.1650463 0.015625     0       6          1 363.600259
#> 26:  6.99190956  5.5769090 0.015625     0       6          1 101.962166
#> 27:  3.23080564 11.8500074 0.015625     0       6          1 114.307462
#> 28: -0.15101040 14.5386856 0.015625     0       6          1  88.343732
#> 29:  1.56247336  4.8298726 0.015625     0       6          1  11.621395
#> 30: -2.77572662  3.5317707 0.015625     0       6          1  51.768126
#> 31: -2.05339697  8.4056495 0.015625     0       6          1   6.528025
#> 32: -1.96230805  5.2058792 0.015625     0       6          1  22.622253
#> 33:  2.46017982  0.3141793 0.015625     0       6          1   6.369771
#> 34: -3.03852444 14.5188879 0.015625     0       6          1  12.002387
#> 35:  0.45879729  1.0790398 0.015625     0       6          1  36.225889
#> 36:  4.76082327  2.8022437 0.015625     0       6          1  24.025216
#> 37:  3.30003330 10.8193653 0.015625     0       6          1  95.321286
#> 38:  7.30973389  8.9389640 0.015625     0       6          1 182.153687
#> 39:  3.91676090 14.7438350 0.015625     0       6          1 213.561246
#> 40: -4.80582376 14.0308780 0.015625     0       6          1  11.003424
#> 41: -3.17485157  1.6766228 0.015625     0       6          1  94.226259
#> 42:  0.33610279  7.8111886 0.015625     0       6          1  24.552807
#> 43:  1.95657935  5.6989350 0.015625     0       6          1  13.651093
#> 44:  2.47997483  0.4084068 0.015625     0       6          1   5.786371
#> 45:  6.18086820 13.0733230 0.015625     0       6          1 267.168814
#> 46:  6.88844179  7.4843209 0.015625     0       6          1 138.646820
#> 47:  8.57050542  4.6737782 0.015625     0       6          1 104.810425
#> 48:  1.93749694  9.3832852 0.015625     0       6          1  46.898662
#> 49: -0.50372961  5.0689939 0.015625     0       6          1  21.438824
#> 50: -0.23855499 14.6361287 0.015625     0       6          1  87.470370
#> 51:  7.77842302  0.3555420 0.015625     0       6          1  34.490202
#> 52:  7.35978075 10.4740305 0.015625     0       6          1 225.443951
#> 53:  1.03691894  7.5361600 0.015625     0       6          1  24.830260
#> 54:  4.34824049 12.5104819 0.015625     0       6          1 171.691926
#> 55:  0.58091430 11.7323490 0.015625     0       6          1  62.203314
#> 56:  6.85677918  1.8296669 0.015625     0       6          1  46.123418
#> 57: -2.62820695 11.8986213 0.015625     0       6          1   3.895686
#> 58: -1.35945331  1.4189453 0.015625     0       6          1  58.274884
#> 59:  7.12398049  9.4746122 0.015625     0       6          1 192.027743
#> 60:  7.17848533  3.2948967 0.015625     0       6          1  66.917952
#> 61: -2.92902136 13.5354890 0.015625     0       6          1   7.426235
#> 62:  5.37372655 10.3104992 0.015625     0       6          1 159.302083
#> 63:  0.79866705  0.4858509 0.015625     0       6          1  34.869162
#> 64:  3.27593717 10.4119992 0.015625     0       6          1  86.897152
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-06-09 04:38:58        1 -1.60490555  11.6169979          0.015625
#>  2: 2022-06-09 04:38:58        1 -3.57051788   4.6078090          0.015625
#>  3: 2022-06-09 04:38:58        1  6.71701525  14.8388918          0.015625
#>  4: 2022-06-09 04:38:58        1 -2.63140922  12.6553692          0.015625
#>  5: 2022-06-09 04:38:58        1  2.00731986   6.2272571          0.015625
#>  6: 2022-06-09 04:38:58        1  9.69475342   3.1784078          0.015625
#>  7: 2022-06-09 04:38:58        1  1.66356633   5.3433634          0.015625
#>  8: 2022-06-09 04:38:58        1  1.82698976  10.5453226          0.015625
#>  9: 2022-06-09 04:38:58        1  0.17588986   4.4539709          0.015625
#> 10: 2022-06-09 04:38:58        1  6.30331671  13.7564183          0.015625
#> 11: 2022-06-09 04:38:58        1  9.84941259   8.1262799          0.015625
#> 12: 2022-06-09 04:38:58        1  6.92798818   0.1815256          0.015625
#> 13: 2022-06-09 04:38:58        1  9.52918625   9.4857963          0.015625
#> 14: 2022-06-09 04:38:58        1 -2.67164666   1.6847194          0.015625
#> 15: 2022-06-09 04:38:58        1  3.57084322   9.4132083          0.015625
#> 16: 2022-06-09 04:38:58        1 -2.37006283   7.7741814          0.015625
#> 17: 2022-06-09 04:38:58        1  6.42534061   1.6093075          0.015625
#> 18: 2022-06-09 04:38:58        1  9.39579244  14.9045310          0.015625
#> 19: 2022-06-09 04:38:58        1 -1.51954420  13.8026454          0.015625
#> 20: 2022-06-09 04:38:58        1  0.61536158   2.1680156          0.015625
#> 21: 2022-06-09 04:38:58        1 -1.05667296  13.4145053          0.015625
#> 22: 2022-06-09 04:38:58        1  5.57864407   0.9846969          0.015625
#> 23: 2022-06-09 04:38:58        1  4.31912586   5.8845230          0.015625
#> 24: 2022-06-09 04:38:58        1 -0.09657184   6.2481836          0.015625
#> 25: 2022-06-09 04:38:58        1  7.89604818  14.1650463          0.015625
#> 26: 2022-06-09 04:38:58        1  6.99190956   5.5769090          0.015625
#> 27: 2022-06-09 04:38:58        1  3.23080564  11.8500074          0.015625
#> 28: 2022-06-09 04:38:58        1 -0.15101040  14.5386856          0.015625
#> 29: 2022-06-09 04:38:58        1  1.56247336   4.8298726          0.015625
#> 30: 2022-06-09 04:38:58        1 -2.77572662   3.5317707          0.015625
#> 31: 2022-06-09 04:38:58        1 -2.05339697   8.4056495          0.015625
#> 32: 2022-06-09 04:38:58        1 -1.96230805   5.2058792          0.015625
#> 33: 2022-06-09 04:38:58        1  2.46017982   0.3141793          0.015625
#> 34: 2022-06-09 04:38:58        1 -3.03852444  14.5188879          0.015625
#> 35: 2022-06-09 04:38:58        1  0.45879729   1.0790398          0.015625
#> 36: 2022-06-09 04:38:58        1  4.76082327   2.8022437          0.015625
#> 37: 2022-06-09 04:38:58        1  3.30003330  10.8193653          0.015625
#> 38: 2022-06-09 04:38:58        1  7.30973389   8.9389640          0.015625
#> 39: 2022-06-09 04:38:58        1  3.91676090  14.7438350          0.015625
#> 40: 2022-06-09 04:38:58        1 -4.80582376  14.0308780          0.015625
#> 41: 2022-06-09 04:38:58        1 -3.17485157   1.6766228          0.015625
#> 42: 2022-06-09 04:38:58        1  0.33610279   7.8111886          0.015625
#> 43: 2022-06-09 04:38:58        1  1.95657935   5.6989350          0.015625
#> 44: 2022-06-09 04:38:58        1  2.47997483   0.4084068          0.015625
#> 45: 2022-06-09 04:38:58        1  6.18086820  13.0733230          0.015625
#> 46: 2022-06-09 04:38:58        1  6.88844179   7.4843209          0.015625
#> 47: 2022-06-09 04:38:58        1  8.57050542   4.6737782          0.015625
#> 48: 2022-06-09 04:38:58        1  1.93749694   9.3832852          0.015625
#> 49: 2022-06-09 04:38:58        1 -0.50372961   5.0689939          0.015625
#> 50: 2022-06-09 04:38:58        1 -0.23855499  14.6361287          0.015625
#> 51: 2022-06-09 04:38:58        1  7.77842302   0.3555420          0.015625
#> 52: 2022-06-09 04:38:58        1  7.35978075  10.4740305          0.015625
#> 53: 2022-06-09 04:38:58        1  1.03691894   7.5361600          0.015625
#> 54: 2022-06-09 04:38:58        1  4.34824049  12.5104819          0.015625
#> 55: 2022-06-09 04:38:58        1  0.58091430  11.7323490          0.015625
#> 56: 2022-06-09 04:38:58        1  6.85677918   1.8296669          0.015625
#> 57: 2022-06-09 04:38:58        1 -2.62820695  11.8986213          0.015625
#> 58: 2022-06-09 04:38:58        1 -1.35945331   1.4189453          0.015625
#> 59: 2022-06-09 04:38:58        1  7.12398049   9.4746122          0.015625
#> 60: 2022-06-09 04:38:58        1  7.17848533   3.2948967          0.015625
#> 61: 2022-06-09 04:38:58        1 -2.92902136  13.5354890          0.015625
#> 62: 2022-06-09 04:38:58        1  5.37372655  10.3104992          0.015625
#> 63: 2022-06-09 04:38:58        1  0.79866705   0.4858509          0.015625
#> 64: 2022-06-09 04:38:58        1  3.27593717  10.4119992          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity