Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>         x1       x2 fidelity  x_domain        y
#> 1: 3.51674 1.343039 0.015625 <list[3]> 1.379103

# best scoring evaluation
instance$result
#>         x1       x2 fidelity  x_domain        y
#> 1: 3.51674 1.343039 0.015625 <list[3]> 1.379103

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1:  7.1708882 13.1295509 0.015625     0       6          1 303.746926
#>  2:  9.9340863  2.0500532 0.015625     0       6          1  79.520865
#>  3: -3.8625034 10.5897387 0.015625     0       6          1   6.852468
#>  4: -3.1331948  9.0653855 0.015625     0       6          1   5.340353
#>  5:  5.8845792 10.0097688 0.015625     0       6          1 170.401480
#>  6:  1.3885817  0.2567704 0.015625     0       6          1  24.646168
#>  7:  1.2373090  5.1851297 0.015625     0       6          1  14.369263
#>  8:  2.2249972  6.0767555 0.015625     0       6          1  16.168170
#>  9:  7.9348503  7.7122354 0.015625     0       6          1 163.109685
#> 10: -3.4323385  2.4548592 0.015625     0       6          1  88.599774
#> 11:  5.0872473 14.4193006 0.015625     0       6          1 260.639499
#> 12:  7.7379918 13.1908251 0.015625     0       6          1 323.170463
#> 13: -2.7279290 13.6766567 0.015625     0       6          1  10.856362
#> 14: -4.6269966  2.5101400 0.015625     0       6          1 141.712247
#> 15:  2.5255554 14.9476694 0.015625     0       6          1 165.263944
#> 16:  0.1370866  4.5475536 0.015625     0       6          1  21.036869
#> 17: -4.0040574  8.0557583 0.015625     0       6          1  26.887812
#> 18: -4.3922491 10.5441594 0.015625     0       6          1  16.216714
#> 19: -3.6234715 13.1518781 0.015625     0       6          1   2.454172
#> 20:  1.8713027  9.8134324 0.015625     0       6          1  51.834048
#> 21:  8.1249119  6.0804540 0.015625     0       6          1 128.032640
#> 22:  1.1303292  1.5556702 0.015625     0       6          1  21.301262
#> 23:  8.0186460  9.7491167 0.015625     0       6          1 219.668861
#> 24:  8.0107941  2.1508623 0.015625     0       6          1  56.488123
#> 25:  3.4988056  7.8539762 0.015625     0       6          1  50.651607
#> 26:  3.8407958 12.1865570 0.015625     0       6          1 142.974007
#> 27:  6.5347289  0.6543409 0.015625     0       6          1  33.300286
#> 28:  0.4028015  3.7578131 0.015625     0       6          1  21.413169
#> 29: -2.1117772  3.5027518 0.015625     0       6          1  40.999539
#> 30:  5.9285358  6.8323408 0.015625     0       6          1 103.409512
#> 31:  4.4070563 14.9270897 0.015625     0       6          1 242.550759
#> 32: -1.4166280  2.9228965 0.015625     0       6          1  40.563692
#> 33:  8.5602750  6.0713850 0.015625     0       6          1 134.696502
#> 34: -3.5260946  2.6704317 0.015625     0       6          1  88.032420
#> 35:  1.6553153  0.2743229 0.015625     0       6          1  19.272634
#> 36: -2.6490724  4.3767814 0.015625     0       6          1  38.203611
#> 37:  1.5703293 13.2665983 0.015625     0       6          1 103.901261
#> 38:  1.6990086  9.0040886 0.015625     0       6          1  40.349801
#> 39: -4.7008448  7.1055192 0.015625     0       6          1  59.669986
#> 40: -0.7777314  3.2339275 0.015625     0       6          1  33.021825
#> 41:  1.0569778 12.4227057 0.015625     0       6          1  79.853880
#> 42:  4.3262889  7.9638170 0.015625     0       6          1  74.840391
#> 43:  1.1753446  7.8058957 0.015625     0       6          1  26.905220
#> 44: -4.9629682 14.5971017 0.015625     0       6          1  12.384474
#> 45:  7.9315266  1.5821481 0.015625     0       6          1  48.585324
#> 46:  6.8397212  6.3351293 0.015625     0       6          1 113.850046
#> 47:  5.1852033 12.1483784 0.015625     0       6          1 198.631956
#> 48:  4.7966294 14.9951098 0.015625     0       6          1 264.310566
#> 49:  4.4694282  0.7877047 0.015625     0       6          1   9.345866
#> 50:  0.4386624  7.9794326 0.015625     0       6          1  25.830806
#> 51:  0.6284824  0.8821637 0.015625     0       6          1  34.821937
#> 52:  5.6979329  6.1791920 0.015625     0       6          1  86.058112
#> 53:  0.6645165  0.3015623 0.015625     0       6          1  39.222406
#> 54:  0.6309280  5.6183759 0.015625     0       6          1  18.125972
#> 55:  7.0429733  5.6014423 0.015625     0       6          1 103.182317
#> 56: -3.9295834 10.9823827 0.015625     0       6          1   6.278265
#> 57:  6.0667609 10.7089891 0.015625     0       6          1 194.487150
#> 58:  9.4001319  4.7931363 0.015625     0       6          1 122.216811
#> 59:  3.5167401  1.3430386 0.015625     0       6          1   1.379103
#> 60:  2.2448477  7.1619688 0.015625     0       6          1  24.981441
#> 61:  6.4636633  1.1409171 0.015625     0       6          1  36.615404
#> 62:  7.4742976  5.6616051 0.015625     0       6          1 110.377194
#> 63:  5.0235309 11.6377239 0.015625     0       6          1 178.241952
#> 64:  0.8063109  2.2791101 0.015625     0       6          1  22.686032
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-03-13 04:33:02        1   7.1708882  13.1295509          0.015625
#>  2: 2022-03-13 04:33:02        1   9.9340863   2.0500532          0.015625
#>  3: 2022-03-13 04:33:02        1  -3.8625034  10.5897387          0.015625
#>  4: 2022-03-13 04:33:02        1  -3.1331948   9.0653855          0.015625
#>  5: 2022-03-13 04:33:02        1   5.8845792  10.0097688          0.015625
#>  6: 2022-03-13 04:33:02        1   1.3885817   0.2567704          0.015625
#>  7: 2022-03-13 04:33:02        1   1.2373090   5.1851297          0.015625
#>  8: 2022-03-13 04:33:02        1   2.2249972   6.0767555          0.015625
#>  9: 2022-03-13 04:33:02        1   7.9348503   7.7122354          0.015625
#> 10: 2022-03-13 04:33:02        1  -3.4323385   2.4548592          0.015625
#> 11: 2022-03-13 04:33:02        1   5.0872473  14.4193006          0.015625
#> 12: 2022-03-13 04:33:02        1   7.7379918  13.1908251          0.015625
#> 13: 2022-03-13 04:33:02        1  -2.7279290  13.6766567          0.015625
#> 14: 2022-03-13 04:33:02        1  -4.6269966   2.5101400          0.015625
#> 15: 2022-03-13 04:33:02        1   2.5255554  14.9476694          0.015625
#> 16: 2022-03-13 04:33:02        1   0.1370866   4.5475536          0.015625
#> 17: 2022-03-13 04:33:02        1  -4.0040574   8.0557583          0.015625
#> 18: 2022-03-13 04:33:02        1  -4.3922491  10.5441594          0.015625
#> 19: 2022-03-13 04:33:02        1  -3.6234715  13.1518781          0.015625
#> 20: 2022-03-13 04:33:02        1   1.8713027   9.8134324          0.015625
#> 21: 2022-03-13 04:33:02        1   8.1249119   6.0804540          0.015625
#> 22: 2022-03-13 04:33:02        1   1.1303292   1.5556702          0.015625
#> 23: 2022-03-13 04:33:02        1   8.0186460   9.7491167          0.015625
#> 24: 2022-03-13 04:33:02        1   8.0107941   2.1508623          0.015625
#> 25: 2022-03-13 04:33:02        1   3.4988056   7.8539762          0.015625
#> 26: 2022-03-13 04:33:02        1   3.8407958  12.1865570          0.015625
#> 27: 2022-03-13 04:33:02        1   6.5347289   0.6543409          0.015625
#> 28: 2022-03-13 04:33:02        1   0.4028015   3.7578131          0.015625
#> 29: 2022-03-13 04:33:02        1  -2.1117772   3.5027518          0.015625
#> 30: 2022-03-13 04:33:02        1   5.9285358   6.8323408          0.015625
#> 31: 2022-03-13 04:33:02        1   4.4070563  14.9270897          0.015625
#> 32: 2022-03-13 04:33:02        1  -1.4166280   2.9228965          0.015625
#> 33: 2022-03-13 04:33:02        1   8.5602750   6.0713850          0.015625
#> 34: 2022-03-13 04:33:02        1  -3.5260946   2.6704317          0.015625
#> 35: 2022-03-13 04:33:02        1   1.6553153   0.2743229          0.015625
#> 36: 2022-03-13 04:33:02        1  -2.6490724   4.3767814          0.015625
#> 37: 2022-03-13 04:33:02        1   1.5703293  13.2665983          0.015625
#> 38: 2022-03-13 04:33:02        1   1.6990086   9.0040886          0.015625
#> 39: 2022-03-13 04:33:02        1  -4.7008448   7.1055192          0.015625
#> 40: 2022-03-13 04:33:02        1  -0.7777314   3.2339275          0.015625
#> 41: 2022-03-13 04:33:02        1   1.0569778  12.4227057          0.015625
#> 42: 2022-03-13 04:33:02        1   4.3262889   7.9638170          0.015625
#> 43: 2022-03-13 04:33:02        1   1.1753446   7.8058957          0.015625
#> 44: 2022-03-13 04:33:02        1  -4.9629682  14.5971017          0.015625
#> 45: 2022-03-13 04:33:02        1   7.9315266   1.5821481          0.015625
#> 46: 2022-03-13 04:33:02        1   6.8397212   6.3351293          0.015625
#> 47: 2022-03-13 04:33:02        1   5.1852033  12.1483784          0.015625
#> 48: 2022-03-13 04:33:02        1   4.7966294  14.9951098          0.015625
#> 49: 2022-03-13 04:33:02        1   4.4694282   0.7877047          0.015625
#> 50: 2022-03-13 04:33:02        1   0.4386624   7.9794326          0.015625
#> 51: 2022-03-13 04:33:02        1   0.6284824   0.8821637          0.015625
#> 52: 2022-03-13 04:33:02        1   5.6979329   6.1791920          0.015625
#> 53: 2022-03-13 04:33:02        1   0.6645165   0.3015623          0.015625
#> 54: 2022-03-13 04:33:02        1   0.6309280   5.6183759          0.015625
#> 55: 2022-03-13 04:33:02        1   7.0429733   5.6014423          0.015625
#> 56: 2022-03-13 04:33:02        1  -3.9295834  10.9823827          0.015625
#> 57: 2022-03-13 04:33:02        1   6.0667609  10.7089891          0.015625
#> 58: 2022-03-13 04:33:02        1   9.4001319   4.7931363          0.015625
#> 59: 2022-03-13 04:33:02        1   3.5167401   1.3430386          0.015625
#> 60: 2022-03-13 04:33:02        1   2.2448477   7.1619688          0.015625
#> 61: 2022-03-13 04:33:02        1   6.4636633   1.1409171          0.015625
#> 62: 2022-03-13 04:33:02        1   7.4742976   5.6616051          0.015625
#> 63: 2022-03-13 04:33:02        1   5.0235309  11.6377239          0.015625
#> 64: 2022-03-13 04:33:02        1   0.8063109   2.2791101          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity