Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.322836 1.422942 0.015625 <list[3]> 0.6934877

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.322836 1.422942 0.015625 <list[3]> 0.6934877

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1:  6.7744249  1.7701861 0.015625     0       6          1  44.8960267
#>  2: -2.6426658  4.8626097 0.015625     0       6          1  32.4603473
#>  3: -2.0245054  8.6484730 0.015625     0       6          1   6.2808785
#>  4:  6.6245816  2.6995811 0.015625     0       6          1  53.7824200
#>  5:  8.6983479  1.7028273 0.015625     0       6          1  54.9549677
#>  6:  7.1006678  4.3052205 0.015625     0       6          1  81.4681343
#>  7: -2.8163261  5.7818163 0.015625     0       6          1  25.3482788
#>  8: -4.0385993  2.2550138 0.015625     0       6          1 117.9453489
#>  9: -4.5286636  4.5568091 0.015625     0       6          1  94.3895291
#> 10:  8.9340683  2.7075620 0.015625     0       6          1  73.3128772
#> 11:  4.0828154  4.6194893 0.015625     0       6          1  25.5519156
#> 12:  1.8179525  9.8231479 0.015625     0       6          1  51.4076610
#> 13:  3.3228363  1.4229416 0.015625     0       6          1   0.6934877
#> 14:  0.7269730  1.0168060 0.015625     0       6          1  31.9388677
#> 15: -3.2211030  6.7165332 0.015625     0       6          1  22.7919215
#> 16: -2.3729912 13.0275697 0.015625     0       6          1  12.5694235
#> 17:  2.0054831 11.7550234 0.015625     0       6          1  83.8049020
#> 18:  5.9738327  3.7764603 0.015625     0       6          1  57.4234428
#> 19: -4.6662929  5.5320465 0.015625     0       6          1  82.9010661
#> 20:  1.1144252  5.1196625 0.015625     0       6          1  14.9628520
#> 21:  4.6119799 14.0896938 0.015625     0       6          1 227.3643424
#> 22: -4.7751365 14.4638322 0.015625     0       6          1  10.6286421
#> 23:  2.5685458  4.1467771 0.015625     0       6          1   6.0603886
#> 24:  0.5551436  9.2906419 0.015625     0       6          1  35.5048775
#> 25:  9.9992700  7.0384446 0.015625     0       6          1 194.5527864
#> 26: -0.4843398 12.5988343 0.015625     0       6          1  52.3790737
#> 27: -0.9424833 11.6745493 0.015625     0       6          1  32.8434432
#> 28:  1.6107724  2.5330035 0.015625     0       6          1  10.5828278
#> 29: -3.7524984  2.3500761 0.015625     0       6          1 103.2410092
#> 30: -0.9178584  0.8583801 0.015625     0       6          1  59.7683946
#> 31:  1.1111228  8.4737654 0.015625     0       6          1  31.9354303
#> 32:  2.0902134 10.5243954 0.015625     0       6          1  64.7818341
#> 33: -3.2024725  6.7880890 0.015625     0       6          1  21.7983505
#> 34:  0.2748789 12.1202858 0.015625     0       6          1  62.2155021
#> 35: -4.3234414  8.8608090 0.015625     0       6          1  27.4717448
#> 36:  9.7880365  3.3621749 0.015625     0       6          1 100.9161392
#> 37:  9.8676234 11.3879686 0.015625     0       6          1 328.8958857
#> 38:  8.7119171  0.1597127 0.015625     0       6          1  35.1296676
#> 39:  9.7513780  8.2780134 0.015625     0       6          1 222.1443105
#> 40:  7.6076880  9.3281131 0.015625     0       6          1 198.8433489
#> 41:  6.8972390  2.8951555 0.015625     0       6          1  58.9329930
#> 42:  2.7350447 14.7202914 0.015625     0       6          1 166.1295661
#> 43:  6.3668859  5.4265778 0.015625     0       6          1  88.6809939
#> 44: -4.9946746 12.9207339 0.015625     0       6          1  15.8987767
#> 45:  0.2570676 12.6339402 0.015625     0       6          1  68.8628864
#> 46:  4.8627633  9.6798027 0.015625     0       6          1 125.7588954
#> 47: -0.9805571  0.7372376 0.015625     0       6          1  62.3068060
#> 48:  5.4957776  2.9102174 0.015625     0       6          1  39.1333514
#> 49:  3.8610455  2.8687534 0.015625     0       6          1   9.3078457
#> 50:  2.5041513  2.9322801 0.015625     0       6          1   2.8090965
#> 51:  8.5443271  3.7910904 0.015625     0       6          1  87.5183363
#> 52: -3.6175298 10.1002475 0.015625     0       6          1   5.7069876
#> 53:  0.7839737  4.2853814 0.015625     0       6          1  17.0353689
#> 54: -2.3298064 14.7482230 0.015625     0       6          1  27.1410998
#> 55:  3.3973840  4.7690284 0.015625     0       6          1  15.3122154
#> 56: -2.9358267  7.8338774 0.015625     0       6          1  10.2330700
#> 57:  3.0651460 10.0535171 0.015625     0       6          1  75.1269770
#> 58:  2.1703322  7.4459278 0.015625     0       6          1  27.1946993
#> 59:  7.0989661  0.4604668 0.015625     0       6          1  34.2986759
#> 60:  0.3792758  7.5681977 0.015625     0       6          1  23.6173895
#> 61:  3.6217346  3.6243104 0.015625     0       6          1  10.3948604
#> 62: -2.6375105  5.8442559 0.015625     0       6          1  22.4530187
#> 63: -0.2110124  6.6453117 0.015625     0       6          1  19.4840611
#> 64: -2.0267074  5.4464034 0.015625     0       6          1  21.0252785
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-06-04 04:31:24        1   6.7744249   1.7701861          0.015625
#>  2: 2022-06-04 04:31:24        1  -2.6426658   4.8626097          0.015625
#>  3: 2022-06-04 04:31:24        1  -2.0245054   8.6484730          0.015625
#>  4: 2022-06-04 04:31:24        1   6.6245816   2.6995811          0.015625
#>  5: 2022-06-04 04:31:24        1   8.6983479   1.7028273          0.015625
#>  6: 2022-06-04 04:31:24        1   7.1006678   4.3052205          0.015625
#>  7: 2022-06-04 04:31:24        1  -2.8163261   5.7818163          0.015625
#>  8: 2022-06-04 04:31:24        1  -4.0385993   2.2550138          0.015625
#>  9: 2022-06-04 04:31:24        1  -4.5286636   4.5568091          0.015625
#> 10: 2022-06-04 04:31:24        1   8.9340683   2.7075620          0.015625
#> 11: 2022-06-04 04:31:24        1   4.0828154   4.6194893          0.015625
#> 12: 2022-06-04 04:31:24        1   1.8179525   9.8231479          0.015625
#> 13: 2022-06-04 04:31:24        1   3.3228363   1.4229416          0.015625
#> 14: 2022-06-04 04:31:24        1   0.7269730   1.0168060          0.015625
#> 15: 2022-06-04 04:31:24        1  -3.2211030   6.7165332          0.015625
#> 16: 2022-06-04 04:31:24        1  -2.3729912  13.0275697          0.015625
#> 17: 2022-06-04 04:31:24        1   2.0054831  11.7550234          0.015625
#> 18: 2022-06-04 04:31:24        1   5.9738327   3.7764603          0.015625
#> 19: 2022-06-04 04:31:24        1  -4.6662929   5.5320465          0.015625
#> 20: 2022-06-04 04:31:24        1   1.1144252   5.1196625          0.015625
#> 21: 2022-06-04 04:31:24        1   4.6119799  14.0896938          0.015625
#> 22: 2022-06-04 04:31:24        1  -4.7751365  14.4638322          0.015625
#> 23: 2022-06-04 04:31:24        1   2.5685458   4.1467771          0.015625
#> 24: 2022-06-04 04:31:24        1   0.5551436   9.2906419          0.015625
#> 25: 2022-06-04 04:31:24        1   9.9992700   7.0384446          0.015625
#> 26: 2022-06-04 04:31:24        1  -0.4843398  12.5988343          0.015625
#> 27: 2022-06-04 04:31:24        1  -0.9424833  11.6745493          0.015625
#> 28: 2022-06-04 04:31:24        1   1.6107724   2.5330035          0.015625
#> 29: 2022-06-04 04:31:24        1  -3.7524984   2.3500761          0.015625
#> 30: 2022-06-04 04:31:24        1  -0.9178584   0.8583801          0.015625
#> 31: 2022-06-04 04:31:24        1   1.1111228   8.4737654          0.015625
#> 32: 2022-06-04 04:31:24        1   2.0902134  10.5243954          0.015625
#> 33: 2022-06-04 04:31:24        1  -3.2024725   6.7880890          0.015625
#> 34: 2022-06-04 04:31:24        1   0.2748789  12.1202858          0.015625
#> 35: 2022-06-04 04:31:24        1  -4.3234414   8.8608090          0.015625
#> 36: 2022-06-04 04:31:24        1   9.7880365   3.3621749          0.015625
#> 37: 2022-06-04 04:31:24        1   9.8676234  11.3879686          0.015625
#> 38: 2022-06-04 04:31:24        1   8.7119171   0.1597127          0.015625
#> 39: 2022-06-04 04:31:24        1   9.7513780   8.2780134          0.015625
#> 40: 2022-06-04 04:31:24        1   7.6076880   9.3281131          0.015625
#> 41: 2022-06-04 04:31:24        1   6.8972390   2.8951555          0.015625
#> 42: 2022-06-04 04:31:24        1   2.7350447  14.7202914          0.015625
#> 43: 2022-06-04 04:31:24        1   6.3668859   5.4265778          0.015625
#> 44: 2022-06-04 04:31:24        1  -4.9946746  12.9207339          0.015625
#> 45: 2022-06-04 04:31:24        1   0.2570676  12.6339402          0.015625
#> 46: 2022-06-04 04:31:24        1   4.8627633   9.6798027          0.015625
#> 47: 2022-06-04 04:31:24        1  -0.9805571   0.7372376          0.015625
#> 48: 2022-06-04 04:31:24        1   5.4957776   2.9102174          0.015625
#> 49: 2022-06-04 04:31:24        1   3.8610455   2.8687534          0.015625
#> 50: 2022-06-04 04:31:24        1   2.5041513   2.9322801          0.015625
#> 51: 2022-06-04 04:31:24        1   8.5443271   3.7910904          0.015625
#> 52: 2022-06-04 04:31:24        1  -3.6175298  10.1002475          0.015625
#> 53: 2022-06-04 04:31:24        1   0.7839737   4.2853814          0.015625
#> 54: 2022-06-04 04:31:24        1  -2.3298064  14.7482230          0.015625
#> 55: 2022-06-04 04:31:24        1   3.3973840   4.7690284          0.015625
#> 56: 2022-06-04 04:31:24        1  -2.9358267   7.8338774          0.015625
#> 57: 2022-06-04 04:31:24        1   3.0651460  10.0535171          0.015625
#> 58: 2022-06-04 04:31:24        1   2.1703322   7.4459278          0.015625
#> 59: 2022-06-04 04:31:24        1   7.0989661   0.4604668          0.015625
#> 60: 2022-06-04 04:31:24        1   0.3792758   7.5681977          0.015625
#> 61: 2022-06-04 04:31:24        1   3.6217346   3.6243104          0.015625
#> 62: 2022-06-04 04:31:24        1  -2.6375105   5.8442559          0.015625
#> 63: 2022-06-04 04:31:24        1  -0.2110124   6.6453117          0.015625
#> 64: 2022-06-04 04:31:24        1  -2.0267074   5.4464034          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity