Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain         y
#> 1: 3.364084 0.9435164 0.015625 <list[3]> 0.6371077

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain         y
#> 1: 3.364084 0.9435164 0.015625 <list[3]> 0.6371077

# all evaluations
as.data.table(instance$archive)
#>               x1         x2 fidelity stage bracket repetition           y
#>  1:  2.151684063  6.4714274 0.015625     0       6          1  18.8201085
#>  2:  8.794266669  1.6284668 0.015625     0       6          1  54.7635675
#>  3:  5.791575565 12.0244218 0.015625     0       6          1 220.4082222
#>  4:  5.316225155 13.8996102 0.015625     0       6          1 255.4438260
#>  5:  9.045443840  9.3329994 0.015625     0       6          1 232.5325454
#>  6:  1.125029775 14.3767052 0.015625     0       6          1 116.7230048
#>  7: -2.815059524  8.7986832 0.015625     0       6          1   4.6119641
#>  8: -1.575590335  3.6402540 0.015625     0       6          1  34.3941852
#>  9:  0.989637661 13.0746050 0.015625     0       6          1  89.5681177
#> 10:  3.710383220  8.8779190 0.015625     0       6          1  71.7973885
#> 11: -4.041066894  0.6576782 0.015625     0       6          1 154.7270843
#> 12:  8.144918374 14.9035676 0.015625     0       6          1 400.3500362
#> 13:  2.006129586  2.6057083 0.015625     0       6          1   6.0564104
#> 14:  7.627084501  2.5785298 0.015625     0       6          1  60.1680923
#> 15:  0.773396731 10.3796527 0.015625     0       6          1  48.1429759
#> 16:  1.535517190  8.0313002 0.015625     0       6          1  29.7220642
#> 17: -2.217372942  3.7787768 0.015625     0       6          1  39.0423287
#> 18: -1.072156047  0.5483977 0.015625     0       6          1  66.3361172
#> 19:  5.273566049  4.8843076 0.015625     0       6          1  56.3567769
#> 20:  2.604755318 11.2732279 0.015625     0       6          1  86.5766543
#> 21: -0.002348342  2.4311656 0.015625     0       6          1  32.3653573
#> 22:  4.885300435 12.7292378 0.015625     0       6          1 201.2820806
#> 23:  9.764234809 14.7229450 0.015625     0       6          1 455.9906137
#> 24: -0.234495847 10.0493696 0.015625     0       6          1  32.8410273
#> 25: -2.682321421  1.3499484 0.015625     0       6          1  84.9383267
#> 26:  7.326664134  8.8761239 0.015625     0       6          1 180.8905403
#> 27:  8.142206842  8.6768605 0.015625     0       6          1 192.1543247
#> 28:  3.176887471  7.2256747 0.015625     0       6          1  36.0630433
#> 29:  7.045050645 12.4574405 0.015625     0       6          1 277.5740363
#> 30: -4.169616050  7.6628099 0.015625     0       6          1  35.3773341
#> 31:  3.214615063  5.9784454 0.015625     0       6          1  23.2425580
#> 32:  0.710706238  4.7431070 0.015625     0       6          1  17.2974195
#> 33:  4.660849525  1.4063404 0.015625     0       6          1  14.1553045
#> 34:  5.430489383 14.4960016 0.015625     0       6          1 279.8007133
#> 35: -0.557885049  8.7163127 0.015625     0       6          1  21.4543925
#> 36: -1.234085375 12.8279582 0.015625     0       6          1  36.3761063
#> 37: -1.957683421  9.0019895 0.015625     0       6          1   6.4306898
#> 38: -4.032667102  6.3542222 0.015625     0       6          1  47.0502616
#> 39:  3.037561908  9.6292117 0.015625     0       6          1  67.3613106
#> 40:  4.987518677  1.6728951 0.015625     0       6          1  20.7079704
#> 41:  8.822986921 11.9285924 0.015625     0       6          1 311.0465235
#> 42:  3.364084419  0.9435164 0.015625     0       6          1   0.6371077
#> 43: -2.616463749  5.2479885 0.015625     0       6          1  27.9750766
#> 44: -4.678735646  1.2912159 0.015625     0       6          1 174.2420093
#> 45: -3.549838868 14.0510016 0.015625     0       6          1   5.2424083
#> 46:  2.591186861  6.0634105 0.015625     0       6          1  17.6641178
#> 47:  7.975032410  0.8518933 0.015625     0       6          1  40.0774891
#> 48:  2.865204334 14.5644081 0.015625     0       6          1 166.4534808
#> 49: -0.893025832 11.9458338 0.015625     0       6          1  36.2712461
#> 50:  7.920762664  2.7886785 0.015625     0       6          1  65.0995449
#> 51:  5.629999689  4.6015267 0.015625     0       6          1  61.0189103
#> 52:  5.082027111  9.1628121 0.015625     0       6          1 122.8180165
#> 53:  1.390448355  9.1512759 0.015625     0       6          1  39.8632367
#> 54:  4.755771169 13.6704376 0.015625     0       6          1 221.9462965
#> 55: -4.922011958 10.2414508 0.015625     0       6          1  30.8081850
#> 56: -3.275054196  3.6732199 0.015625     0       6          1  62.4041629
#> 57:  3.556644619 10.8529123 0.015625     0       6          1 103.7196011
#> 58: -0.164481922  6.2728752 0.015625     0       6          1  19.4726213
#> 59:  3.264801401  9.0670553 0.015625     0       6          1  63.4415357
#> 60:  0.076051245 12.5917533 0.015625     0       6          1  64.6335549
#> 61:  1.767503016  9.1238707 0.015625     0       6          1  42.2392762
#> 62: -1.795563566 14.9249324 0.015625     0       6          1  43.4777976
#> 63:  7.265730607  9.4218328 0.015625     0       6          1 193.8831283
#> 64:  3.511971809  9.4239728 0.015625     0       6          1  75.5987189
#>               x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-04-19 04:40:16        1  2.151684063   6.4714274          0.015625
#>  2: 2022-04-19 04:40:16        1  8.794266669   1.6284668          0.015625
#>  3: 2022-04-19 04:40:16        1  5.791575565  12.0244218          0.015625
#>  4: 2022-04-19 04:40:16        1  5.316225155  13.8996102          0.015625
#>  5: 2022-04-19 04:40:16        1  9.045443840   9.3329994          0.015625
#>  6: 2022-04-19 04:40:16        1  1.125029775  14.3767052          0.015625
#>  7: 2022-04-19 04:40:16        1 -2.815059524   8.7986832          0.015625
#>  8: 2022-04-19 04:40:16        1 -1.575590335   3.6402540          0.015625
#>  9: 2022-04-19 04:40:16        1  0.989637661  13.0746050          0.015625
#> 10: 2022-04-19 04:40:16        1  3.710383220   8.8779190          0.015625
#> 11: 2022-04-19 04:40:16        1 -4.041066894   0.6576782          0.015625
#> 12: 2022-04-19 04:40:16        1  8.144918374  14.9035676          0.015625
#> 13: 2022-04-19 04:40:16        1  2.006129586   2.6057083          0.015625
#> 14: 2022-04-19 04:40:16        1  7.627084501   2.5785298          0.015625
#> 15: 2022-04-19 04:40:16        1  0.773396731  10.3796527          0.015625
#> 16: 2022-04-19 04:40:16        1  1.535517190   8.0313002          0.015625
#> 17: 2022-04-19 04:40:16        1 -2.217372942   3.7787768          0.015625
#> 18: 2022-04-19 04:40:16        1 -1.072156047   0.5483977          0.015625
#> 19: 2022-04-19 04:40:16        1  5.273566049   4.8843076          0.015625
#> 20: 2022-04-19 04:40:16        1  2.604755318  11.2732279          0.015625
#> 21: 2022-04-19 04:40:16        1 -0.002348342   2.4311656          0.015625
#> 22: 2022-04-19 04:40:16        1  4.885300435  12.7292378          0.015625
#> 23: 2022-04-19 04:40:16        1  9.764234809  14.7229450          0.015625
#> 24: 2022-04-19 04:40:16        1 -0.234495847  10.0493696          0.015625
#> 25: 2022-04-19 04:40:16        1 -2.682321421   1.3499484          0.015625
#> 26: 2022-04-19 04:40:16        1  7.326664134   8.8761239          0.015625
#> 27: 2022-04-19 04:40:16        1  8.142206842   8.6768605          0.015625
#> 28: 2022-04-19 04:40:16        1  3.176887471   7.2256747          0.015625
#> 29: 2022-04-19 04:40:16        1  7.045050645  12.4574405          0.015625
#> 30: 2022-04-19 04:40:16        1 -4.169616050   7.6628099          0.015625
#> 31: 2022-04-19 04:40:16        1  3.214615063   5.9784454          0.015625
#> 32: 2022-04-19 04:40:16        1  0.710706238   4.7431070          0.015625
#> 33: 2022-04-19 04:40:16        1  4.660849525   1.4063404          0.015625
#> 34: 2022-04-19 04:40:16        1  5.430489383  14.4960016          0.015625
#> 35: 2022-04-19 04:40:16        1 -0.557885049   8.7163127          0.015625
#> 36: 2022-04-19 04:40:16        1 -1.234085375  12.8279582          0.015625
#> 37: 2022-04-19 04:40:16        1 -1.957683421   9.0019895          0.015625
#> 38: 2022-04-19 04:40:16        1 -4.032667102   6.3542222          0.015625
#> 39: 2022-04-19 04:40:16        1  3.037561908   9.6292117          0.015625
#> 40: 2022-04-19 04:40:16        1  4.987518677   1.6728951          0.015625
#> 41: 2022-04-19 04:40:16        1  8.822986921  11.9285924          0.015625
#> 42: 2022-04-19 04:40:16        1  3.364084419   0.9435164          0.015625
#> 43: 2022-04-19 04:40:16        1 -2.616463749   5.2479885          0.015625
#> 44: 2022-04-19 04:40:16        1 -4.678735646   1.2912159          0.015625
#> 45: 2022-04-19 04:40:16        1 -3.549838868  14.0510016          0.015625
#> 46: 2022-04-19 04:40:16        1  2.591186861   6.0634105          0.015625
#> 47: 2022-04-19 04:40:16        1  7.975032410   0.8518933          0.015625
#> 48: 2022-04-19 04:40:16        1  2.865204334  14.5644081          0.015625
#> 49: 2022-04-19 04:40:16        1 -0.893025832  11.9458338          0.015625
#> 50: 2022-04-19 04:40:16        1  7.920762664   2.7886785          0.015625
#> 51: 2022-04-19 04:40:16        1  5.629999689   4.6015267          0.015625
#> 52: 2022-04-19 04:40:16        1  5.082027111   9.1628121          0.015625
#> 53: 2022-04-19 04:40:16        1  1.390448355   9.1512759          0.015625
#> 54: 2022-04-19 04:40:16        1  4.755771169  13.6704376          0.015625
#> 55: 2022-04-19 04:40:16        1 -4.922011958  10.2414508          0.015625
#> 56: 2022-04-19 04:40:16        1 -3.275054196   3.6732199          0.015625
#> 57: 2022-04-19 04:40:16        1  3.556644619  10.8529123          0.015625
#> 58: 2022-04-19 04:40:16        1 -0.164481922   6.2728752          0.015625
#> 59: 2022-04-19 04:40:16        1  3.264801401   9.0670553          0.015625
#> 60: 2022-04-19 04:40:16        1  0.076051245  12.5917533          0.015625
#> 61: 2022-04-19 04:40:16        1  1.767503016   9.1238707          0.015625
#> 62: 2022-04-19 04:40:16        1 -1.795563566  14.9249324          0.015625
#> 63: 2022-04-19 04:40:16        1  7.265730607   9.4218328          0.015625
#> 64: 2022-04-19 04:40:16        1  3.511971809   9.4239728          0.015625
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity