Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 3.406177 1.372236 0.015625 <list[3]> 0.922655

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 3.406177 1.372236 0.015625 <list[3]> 0.922655

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition          y
#>  1:  3.26981972  6.79217389 0.015625     0       6          1  32.597447
#>  2: -2.48924869 13.57499699 0.015625     0       6          1  14.084535
#>  3:  0.34721037  7.71435029 0.015625     0       6          1  24.151395
#>  4: -1.70472238 13.79370032 0.015625     0       6          1  33.629917
#>  5:  8.85308762 13.13537410 0.015625     0       6          1 355.952953
#>  6:  4.86169786  7.97009423 0.015625     0       6          1  92.086489
#>  7:  8.65553933  7.59425450 0.015625     0       6          1 173.833770
#>  8:  8.17937803  9.94878546 0.015625     0       6          1 229.227319
#>  9: -3.03087985  8.24242913 0.015625     0       6          1   8.658113
#> 10:  0.71049991  5.21331149 0.015625     0       6          1  17.386715
#> 11:  0.99579221 14.13458842 0.015625     0       6          1 109.097790
#> 12:  2.29864746  3.20801531 0.015625     0       6          1   4.107601
#> 13:  9.47647375 14.83042568 0.015625     0       6          1 447.797148
#> 14: -4.18498599  0.53210434 0.015625     0       6          1 165.620591
#> 15:  5.04628945 12.25714126 0.015625     0       6          1 195.547776
#> 16:  8.71369576  2.70368064 0.015625     0       6          1  70.579398
#> 17:  8.27056678 11.67304871 0.015625     0       6          1 286.104516
#> 18:  9.73977045 13.84141831 0.015625     0       6          1 418.091267
#> 19: -0.64907928  9.37708235 0.015625     0       6          1  23.083399
#> 20:  4.71854182  0.30966502 0.015625     0       6          1  11.347047
#> 21: -1.53968275 10.48025374 0.015625     0       6          1  14.128099
#> 22: -4.72378435 13.16598014 0.015625     0       6          1  11.187379
#> 23:  6.16534019  1.81555909 0.015625     0       6          1  39.420562
#> 24:  3.03152400 12.12701295 0.015625     0       6          1 114.289188
#> 25:  3.40617729  1.37223560 0.015625     0       6          1   0.922655
#> 26:  2.69986651  5.21712081 0.015625     0       6          1  12.143437
#> 27: -1.06034730  1.27196060 0.015625     0       6          1  56.296434
#> 28:  1.76425386  5.86433514 0.015625     0       6          1  14.792469
#> 29:  8.40117437  3.93603539 0.015625     0       6          1  88.485287
#> 30:  1.57056252  5.17427396 0.015625     0       6          1  12.556038
#> 31:  1.75983774  8.42187963 0.015625     0       6          1  34.487098
#> 32: -3.19728307  5.87750136 0.015625     0       6          1  30.943345
#> 33:  4.97059178  6.31591407 0.015625     0       6          1  68.210865
#> 34:  9.30053828  4.29642432 0.015625     0       6          1 109.446166
#> 35:  0.26764601 11.63089438 0.015625     0       6          1  55.919193
#> 36: -1.91674808 10.09774971 0.015625     0       6          1   7.616706
#> 37:  3.73333810  9.98732179 0.015625     0       6          1  92.291340
#> 38:  2.21184790 13.31521247 0.015625     0       6          1 118.427967
#> 39:  3.19820651 12.66199667 0.015625     0       6          1 131.232054
#> 40:  4.26229081 14.45175474 0.015625     0       6          1 221.231507
#> 41:  5.10073263  1.12345735 0.015625     0       6          1  19.597141
#> 42:  5.54597164 11.75545894 0.015625     0       6          1 203.061273
#> 43:  5.13054408  7.01717534 0.015625     0       6          1  84.012208
#> 44:  9.04083446 13.39897731 0.015625     0       6          1 372.613097
#> 45: -1.54529660  2.52103808 0.015625     0       6          1  46.386563
#> 46: -1.24972996  0.09888988 0.015625     0       6          1  76.044265
#> 47: -0.32164310  3.57305938 0.015625     0       6          1  27.765245
#> 48: -4.47364458  3.83595367 0.015625     0       6          1 105.727939
#> 49:  3.77700377  1.89598028 0.015625     0       6          1   4.428854
#> 50:  3.77325848  6.35791606 0.015625     0       6          1  37.362018
#> 51: -1.46298673  5.50366558 0.015625     0       6          1  19.388526
#> 52: -4.84884781 14.97738938 0.015625     0       6          1  11.594928
#> 53:  1.30614912 10.83380053 0.015625     0       6          1  59.573219
#> 54: -1.24224827  5.06716504 0.015625     0       6          1  21.844417
#> 55: -0.08326387  2.26291556 0.015625     0       6          1  34.544324
#> 56:  8.54201618  0.79249616 0.015625     0       6          1  41.651811
#> 57: -4.55214179  5.86320471 0.015625     0       6          1  72.770516
#> 58:  1.26444576 10.05209801 0.015625     0       6          1  49.080455
#> 59: -3.00602558 14.50280466 0.015625     0       6          1  12.324615
#> 60:  3.24558296 12.37463037 0.015625     0       6          1 126.254090
#> 61:  8.07664732 14.71037776 0.015625     0       6          1 390.436593
#> 62:  9.49994029  2.77976870 0.015625     0       6          1  83.681691
#> 63:  9.41978176  8.49495366 0.015625     0       6          1 218.218693
#> 64:  7.21295777 14.56785357 0.015625     0       6          1 356.069638
#>              x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-18 04:51:20        1  3.26981972  6.79217389          0.015625
#>  2: 2022-08-18 04:51:20        1 -2.48924869 13.57499699          0.015625
#>  3: 2022-08-18 04:51:20        1  0.34721037  7.71435029          0.015625
#>  4: 2022-08-18 04:51:20        1 -1.70472238 13.79370032          0.015625
#>  5: 2022-08-18 04:51:20        1  8.85308762 13.13537410          0.015625
#>  6: 2022-08-18 04:51:20        1  4.86169786  7.97009423          0.015625
#>  7: 2022-08-18 04:51:20        1  8.65553933  7.59425450          0.015625
#>  8: 2022-08-18 04:51:20        1  8.17937803  9.94878546          0.015625
#>  9: 2022-08-18 04:51:20        1 -3.03087985  8.24242913          0.015625
#> 10: 2022-08-18 04:51:20        1  0.71049991  5.21331149          0.015625
#> 11: 2022-08-18 04:51:20        1  0.99579221 14.13458842          0.015625
#> 12: 2022-08-18 04:51:20        1  2.29864746  3.20801531          0.015625
#> 13: 2022-08-18 04:51:20        1  9.47647375 14.83042568          0.015625
#> 14: 2022-08-18 04:51:20        1 -4.18498599  0.53210434          0.015625
#> 15: 2022-08-18 04:51:20        1  5.04628945 12.25714126          0.015625
#> 16: 2022-08-18 04:51:20        1  8.71369576  2.70368064          0.015625
#> 17: 2022-08-18 04:51:20        1  8.27056678 11.67304871          0.015625
#> 18: 2022-08-18 04:51:20        1  9.73977045 13.84141831          0.015625
#> 19: 2022-08-18 04:51:20        1 -0.64907928  9.37708235          0.015625
#> 20: 2022-08-18 04:51:20        1  4.71854182  0.30966502          0.015625
#> 21: 2022-08-18 04:51:20        1 -1.53968275 10.48025374          0.015625
#> 22: 2022-08-18 04:51:20        1 -4.72378435 13.16598014          0.015625
#> 23: 2022-08-18 04:51:20        1  6.16534019  1.81555909          0.015625
#> 24: 2022-08-18 04:51:20        1  3.03152400 12.12701295          0.015625
#> 25: 2022-08-18 04:51:20        1  3.40617729  1.37223560          0.015625
#> 26: 2022-08-18 04:51:20        1  2.69986651  5.21712081          0.015625
#> 27: 2022-08-18 04:51:20        1 -1.06034730  1.27196060          0.015625
#> 28: 2022-08-18 04:51:20        1  1.76425386  5.86433514          0.015625
#> 29: 2022-08-18 04:51:20        1  8.40117437  3.93603539          0.015625
#> 30: 2022-08-18 04:51:20        1  1.57056252  5.17427396          0.015625
#> 31: 2022-08-18 04:51:20        1  1.75983774  8.42187963          0.015625
#> 32: 2022-08-18 04:51:20        1 -3.19728307  5.87750136          0.015625
#> 33: 2022-08-18 04:51:20        1  4.97059178  6.31591407          0.015625
#> 34: 2022-08-18 04:51:20        1  9.30053828  4.29642432          0.015625
#> 35: 2022-08-18 04:51:20        1  0.26764601 11.63089438          0.015625
#> 36: 2022-08-18 04:51:20        1 -1.91674808 10.09774971          0.015625
#> 37: 2022-08-18 04:51:20        1  3.73333810  9.98732179          0.015625
#> 38: 2022-08-18 04:51:20        1  2.21184790 13.31521247          0.015625
#> 39: 2022-08-18 04:51:20        1  3.19820651 12.66199667          0.015625
#> 40: 2022-08-18 04:51:20        1  4.26229081 14.45175474          0.015625
#> 41: 2022-08-18 04:51:20        1  5.10073263  1.12345735          0.015625
#> 42: 2022-08-18 04:51:20        1  5.54597164 11.75545894          0.015625
#> 43: 2022-08-18 04:51:20        1  5.13054408  7.01717534          0.015625
#> 44: 2022-08-18 04:51:20        1  9.04083446 13.39897731          0.015625
#> 45: 2022-08-18 04:51:20        1 -1.54529660  2.52103808          0.015625
#> 46: 2022-08-18 04:51:20        1 -1.24972996  0.09888988          0.015625
#> 47: 2022-08-18 04:51:20        1 -0.32164310  3.57305938          0.015625
#> 48: 2022-08-18 04:51:20        1 -4.47364458  3.83595367          0.015625
#> 49: 2022-08-18 04:51:20        1  3.77700377  1.89598028          0.015625
#> 50: 2022-08-18 04:51:20        1  3.77325848  6.35791606          0.015625
#> 51: 2022-08-18 04:51:20        1 -1.46298673  5.50366558          0.015625
#> 52: 2022-08-18 04:51:20        1 -4.84884781 14.97738938          0.015625
#> 53: 2022-08-18 04:51:20        1  1.30614912 10.83380053          0.015625
#> 54: 2022-08-18 04:51:20        1 -1.24224827  5.06716504          0.015625
#> 55: 2022-08-18 04:51:20        1 -0.08326387  2.26291556          0.015625
#> 56: 2022-08-18 04:51:20        1  8.54201618  0.79249616          0.015625
#> 57: 2022-08-18 04:51:20        1 -4.55214179  5.86320471          0.015625
#> 58: 2022-08-18 04:51:20        1  1.26444576 10.05209801          0.015625
#> 59: 2022-08-18 04:51:20        1 -3.00602558 14.50280466          0.015625
#> 60: 2022-08-18 04:51:20        1  3.24558296 12.37463037          0.015625
#> 61: 2022-08-18 04:51:20        1  8.07664732 14.71037776          0.015625
#> 62: 2022-08-18 04:51:20        1  9.49994029  2.77976870          0.015625
#> 63: 2022-08-18 04:51:20        1  9.41978176  8.49495366          0.015625
#> 64: 2022-08-18 04:51:20        1  7.21295777 14.56785357          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity