Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain       y
#> 1: 2.723612 2.354039 0.015625 <list[3]> 1.43679

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain       y
#> 1: 2.723612 2.354039 0.015625 <list[3]> 1.43679

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1: -1.8231039 12.8404432 0.015625     0       6          1  22.323121
#>  2: -4.0750406  3.2504643 0.015625     0       6          1  99.265703
#>  3: -0.2805615  4.9811465 0.015625     0       6          1  21.381112
#>  4:  5.8788275  1.3312779 0.015625     0       6          1  31.969008
#>  5: -0.2327190 14.5777356 0.015625     0       6          1  86.676560
#>  6:  5.9691660 11.2759392 0.015625     0       6          1 206.291847
#>  7: -4.6248297  8.3194827 0.015625     0       6          1  41.636773
#>  8:  3.2895981  8.3747476 0.015625     0       6          1  53.466010
#>  9:  7.2219424  1.7512452 0.015625     0       6          1  47.501269
#> 10: -1.0647097  2.7769104 0.015625     0       6          1  39.181787
#> 11: -3.8415338 13.9443758 0.015625     0       6          1   4.550671
#> 12:  6.1837933  3.5544876 0.015625     0       6          1  58.250043
#> 13: -3.2822332  3.6362104 0.015625     0       6          1  63.201121
#> 14:  7.3970388  4.3540909 0.015625     0       6          1  85.545839
#> 15:  8.6146162  3.3824889 0.015625     0       6          1  81.019337
#> 16:  0.4368503 14.4214046 0.015625     0       6          1 101.707147
#> 17:  3.8004835  0.7405913 0.015625     0       6          1   2.527008
#> 18:  8.6387156 14.2743770 0.015625     0       6          1 392.438382
#> 19:  4.1075206 14.1623914 0.015625     0       6          1 205.639303
#> 20:  6.1531459  0.9626749 0.015625     0       6          1  32.420573
#> 21: -3.4046109  7.1071764 0.015625     0       6          1  22.516747
#> 22:  8.3161357  0.4817707 0.015625     0       6          1  36.976874
#> 23:  8.0609182  2.6724821 0.015625     0       6          1  64.336243
#> 24: -4.8887524  2.7997883 0.015625     0       6          1 148.943574
#> 25: -2.2499016  8.2063027 0.015625     0       6          1   6.310362
#> 26: -1.6786269  1.0439503 0.015625     0       6          1  68.477120
#> 27:  4.6874260 10.5039798 0.015625     0       6          1 137.194575
#> 28: -4.2404737 10.1499638 0.015625     0       6          1  15.569046
#> 29:  4.1166611 13.1037019 0.015625     0       6          1 177.127369
#> 30: -3.3257865 14.5829872 0.015625     0       6          1   9.261312
#> 31:  2.4238498  5.4143237 0.015625     0       6          1  12.323313
#> 32:  9.3401382  3.6286489 0.015625     0       6          1  96.700293
#> 33: -2.8617958  9.3148582 0.015625     0       6          1   2.996305
#> 34:  0.5474310  7.1260650 0.015625     0       6          1  22.151501
#> 35: -2.4109416 12.1493517 0.015625     0       6          1   7.400734
#> 36:  7.0093825 13.8098799 0.015625     0       6          1 321.857052
#> 37: -4.6260933  3.4944381 0.015625     0       6          1 119.973858
#> 38:  3.7314481  7.0454841 0.015625     0       6          1  45.003622
#> 39:  1.9657718  7.3961674 0.015625     0       6          1  25.717870
#> 40:  5.4657189  3.5744793 0.015625     0       6          1  45.243511
#> 41: -1.0076854 10.9076888 0.015625     0       6          1  25.836268
#> 42:  2.3262139  7.7889981 0.015625     0       6          1  31.771445
#> 43: -3.9656227  4.0146472 0.015625     0       6          1  80.572476
#> 44: -1.7144235 10.1871930 0.015625     0       6          1  10.497666
#> 45:  9.6724788  3.6584708 0.015625     0       6          1 104.244084
#> 46:  3.0326082  2.8929365 0.015625     0       6          1   2.518994
#> 47:  6.3662373  0.1036311 0.015625     0       6          1  28.507129
#> 48:  7.2145356  4.4264918 0.015625     0       6          1  84.759825
#> 49:  3.8422126  6.1629222 0.015625     0       6          1  36.579734
#> 50:  2.7236121  2.3540389 0.015625     0       6          1   1.436790
#> 51:  7.8874857 13.5448199 0.015625     0       6          1 340.383379
#> 52:  1.8358879  9.1888728 0.015625     0       6          1  43.570082
#> 53:  9.5224132 11.8576166 0.015625     0       6          1 332.593558
#> 54: -0.5440193 10.4094887 0.015625     0       6          1  30.708987
#> 55:  1.3244119 13.1070461 0.015625     0       6          1  96.265517
#> 56:  1.3506448  3.7058459 0.015625     0       6          1  12.137136
#> 57: -3.0284486  1.8045317 0.015625     0       6          1  86.900746
#> 58:  2.7780558  3.4587815 0.015625     0       6          1   3.724559
#> 59:  0.2174762 14.5044400 0.015625     0       6          1  97.682683
#> 60:  4.8320155  8.9228761 0.015625     0       6          1 109.064373
#> 61: -0.1877636  8.4331848 0.015625     0       6          1  23.984170
#> 62:  7.7473349  8.3133012 0.015625     0       6          1 174.813338
#> 63:  7.2669203  0.5435017 0.015625     0       6          1  35.438235
#> 64: -4.3250515 12.1808145 0.015625     0       6          1   8.006010
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-05 04:51:34        1  -1.8231039  12.8404432          0.015625
#>  2: 2022-08-05 04:51:34        1  -4.0750406   3.2504643          0.015625
#>  3: 2022-08-05 04:51:34        1  -0.2805615   4.9811465          0.015625
#>  4: 2022-08-05 04:51:34        1   5.8788275   1.3312779          0.015625
#>  5: 2022-08-05 04:51:34        1  -0.2327190  14.5777356          0.015625
#>  6: 2022-08-05 04:51:34        1   5.9691660  11.2759392          0.015625
#>  7: 2022-08-05 04:51:34        1  -4.6248297   8.3194827          0.015625
#>  8: 2022-08-05 04:51:34        1   3.2895981   8.3747476          0.015625
#>  9: 2022-08-05 04:51:34        1   7.2219424   1.7512452          0.015625
#> 10: 2022-08-05 04:51:34        1  -1.0647097   2.7769104          0.015625
#> 11: 2022-08-05 04:51:34        1  -3.8415338  13.9443758          0.015625
#> 12: 2022-08-05 04:51:34        1   6.1837933   3.5544876          0.015625
#> 13: 2022-08-05 04:51:34        1  -3.2822332   3.6362104          0.015625
#> 14: 2022-08-05 04:51:34        1   7.3970388   4.3540909          0.015625
#> 15: 2022-08-05 04:51:34        1   8.6146162   3.3824889          0.015625
#> 16: 2022-08-05 04:51:34        1   0.4368503  14.4214046          0.015625
#> 17: 2022-08-05 04:51:34        1   3.8004835   0.7405913          0.015625
#> 18: 2022-08-05 04:51:34        1   8.6387156  14.2743770          0.015625
#> 19: 2022-08-05 04:51:34        1   4.1075206  14.1623914          0.015625
#> 20: 2022-08-05 04:51:34        1   6.1531459   0.9626749          0.015625
#> 21: 2022-08-05 04:51:34        1  -3.4046109   7.1071764          0.015625
#> 22: 2022-08-05 04:51:34        1   8.3161357   0.4817707          0.015625
#> 23: 2022-08-05 04:51:34        1   8.0609182   2.6724821          0.015625
#> 24: 2022-08-05 04:51:34        1  -4.8887524   2.7997883          0.015625
#> 25: 2022-08-05 04:51:34        1  -2.2499016   8.2063027          0.015625
#> 26: 2022-08-05 04:51:34        1  -1.6786269   1.0439503          0.015625
#> 27: 2022-08-05 04:51:34        1   4.6874260  10.5039798          0.015625
#> 28: 2022-08-05 04:51:34        1  -4.2404737  10.1499638          0.015625
#> 29: 2022-08-05 04:51:34        1   4.1166611  13.1037019          0.015625
#> 30: 2022-08-05 04:51:34        1  -3.3257865  14.5829872          0.015625
#> 31: 2022-08-05 04:51:34        1   2.4238498   5.4143237          0.015625
#> 32: 2022-08-05 04:51:34        1   9.3401382   3.6286489          0.015625
#> 33: 2022-08-05 04:51:34        1  -2.8617958   9.3148582          0.015625
#> 34: 2022-08-05 04:51:34        1   0.5474310   7.1260650          0.015625
#> 35: 2022-08-05 04:51:34        1  -2.4109416  12.1493517          0.015625
#> 36: 2022-08-05 04:51:34        1   7.0093825  13.8098799          0.015625
#> 37: 2022-08-05 04:51:34        1  -4.6260933   3.4944381          0.015625
#> 38: 2022-08-05 04:51:34        1   3.7314481   7.0454841          0.015625
#> 39: 2022-08-05 04:51:34        1   1.9657718   7.3961674          0.015625
#> 40: 2022-08-05 04:51:34        1   5.4657189   3.5744793          0.015625
#> 41: 2022-08-05 04:51:34        1  -1.0076854  10.9076888          0.015625
#> 42: 2022-08-05 04:51:34        1   2.3262139   7.7889981          0.015625
#> 43: 2022-08-05 04:51:34        1  -3.9656227   4.0146472          0.015625
#> 44: 2022-08-05 04:51:34        1  -1.7144235  10.1871930          0.015625
#> 45: 2022-08-05 04:51:34        1   9.6724788   3.6584708          0.015625
#> 46: 2022-08-05 04:51:34        1   3.0326082   2.8929365          0.015625
#> 47: 2022-08-05 04:51:34        1   6.3662373   0.1036311          0.015625
#> 48: 2022-08-05 04:51:34        1   7.2145356   4.4264918          0.015625
#> 49: 2022-08-05 04:51:34        1   3.8422126   6.1629222          0.015625
#> 50: 2022-08-05 04:51:34        1   2.7236121   2.3540389          0.015625
#> 51: 2022-08-05 04:51:34        1   7.8874857  13.5448199          0.015625
#> 52: 2022-08-05 04:51:34        1   1.8358879   9.1888728          0.015625
#> 53: 2022-08-05 04:51:34        1   9.5224132  11.8576166          0.015625
#> 54: 2022-08-05 04:51:34        1  -0.5440193  10.4094887          0.015625
#> 55: 2022-08-05 04:51:34        1   1.3244119  13.1070461          0.015625
#> 56: 2022-08-05 04:51:34        1   1.3506448   3.7058459          0.015625
#> 57: 2022-08-05 04:51:34        1  -3.0284486   1.8045317          0.015625
#> 58: 2022-08-05 04:51:34        1   2.7780558   3.4587815          0.015625
#> 59: 2022-08-05 04:51:34        1   0.2174762  14.5044400          0.015625
#> 60: 2022-08-05 04:51:34        1   4.8320155   8.9228761          0.015625
#> 61: 2022-08-05 04:51:34        1  -0.1877636   8.4331848          0.015625
#> 62: 2022-08-05 04:51:34        1   7.7473349   8.3133012          0.015625
#> 63: 2022-08-05 04:51:34        1   7.2669203   0.5435017          0.015625
#> 64: 2022-08-05 04:51:34        1  -4.3250515  12.1808145          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity