Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain       y
#> 1: 2.705449 2.609524 0.015625 <list[3]> 1.77332

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain       y
#> 1: 2.705449 2.609524 0.015625 <list[3]> 1.77332

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  8.94716186  5.0489723 0.015625     0       6          1 118.706582
#>  2:  6.24017236  1.6897462 0.015625     0       6          1  39.165079
#>  3: -2.33842452  8.5197087 0.015625     0       6          1   5.209287
#>  4:  4.24639881  2.7034374 0.015625     0       6          1  14.138462
#>  5:  9.77632385 11.4626170 0.015625     0       6          1 327.995502
#>  6:  4.30750134  6.8891144 0.015625     0       6          1  57.686986
#>  7:  5.32363806 11.9279702 0.015625     0       6          1 198.555190
#>  8:  9.12833078  1.0583434 0.015625     0       6          1  50.160255
#>  9: -3.52800496  3.4937151 0.015625     0       6          1  73.423582
#> 10:  9.22957138  1.9479533 0.015625     0       6          1  64.870002
#> 11:  3.47263228 10.2360238 0.015625     0       6          1  89.130892
#> 12:  0.04339284  2.0610665 0.015625     0       6          1  34.569429
#> 13:  0.45248301  3.5519557 0.015625     0       6          1  21.643214
#> 14:  6.46474914 10.6372837 0.015625     0       6          1 205.527771
#> 15: -4.10429497 14.6987868 0.015625     0       6          1   7.232345
#> 16:  1.43568164 13.8434865 0.015625     0       6          1 112.599069
#> 17:  5.89176095  5.0265278 0.015625     0       6          1  72.696295
#> 18:  3.92029547  9.1224532 0.015625     0       6          1  82.183831
#> 19:  4.88618490 13.4998615 0.015625     0       6          1 223.141455
#> 20: -0.01904810 12.3390650 0.015625     0       6          1  59.400543
#> 21:  2.70544876  2.6095244 0.015625     0       6          1   1.773320
#> 22:  9.57271309  6.0549350 0.015625     0       6          1 156.074065
#> 23: -1.65169897  7.1377458 0.015625     0       6          1  11.704308
#> 24:  0.20123570  5.8746856 0.015625     0       6          1  19.445871
#> 25:  3.27754726 11.1269016 0.015625     0       6          1 100.746380
#> 26: -4.13510249  8.0217845 0.015625     0       6          1  30.618691
#> 27:  2.41287742  3.5232036 0.015625     0       6          1   4.239355
#> 28: -2.69437727  5.5072074 0.015625     0       6          1  26.384630
#> 29:  5.04620213  9.6052242 0.015625     0       6          1 130.945828
#> 30:  6.75004043  2.0071129 0.015625     0       6          1  47.188604
#> 31: -0.25959290  8.3585202 0.015625     0       6          1  23.056778
#> 32: -0.78126769 10.2475340 0.015625     0       6          1  25.729964
#> 33:  4.92119459  3.6988574 0.015625     0       6          1  34.901460
#> 34: -0.63033440  7.0249197 0.015625     0       6          1  17.756969
#> 35: -4.81565895  5.1720604 0.015625     0       6          1  95.728124
#> 36: -0.27647325  7.2368704 0.015625     0       6          1  19.868693
#> 37:  4.84771434 10.4887107 0.015625     0       6          1 143.120810
#> 38:  9.69336656  2.6626824 0.015625     0       6          1  85.402837
#> 39: -4.43084163 10.7627844 0.015625     0       6          1  15.700159
#> 40: -0.93362290  3.9966452 0.015625     0       6          1  28.075240
#> 41:  3.80826061  0.4938213 0.015625     0       6          1   2.465697
#> 42:  7.74239810  1.4019360 0.015625     0       6          1  45.658013
#> 43:  6.02776977 11.2524975 0.015625     0       6          1 207.771348
#> 44:  9.37677815  3.2210212 0.015625     0       6          1  89.545734
#> 45:  2.02127237  9.0967439 0.015625     0       6          1  44.111638
#> 46:  1.83318352 12.8565820 0.015625     0       6          1 101.034813
#> 47:  2.41564016 12.6248475 0.015625     0       6          1 108.703810
#> 48: -0.35700391  9.0650398 0.015625     0       6          1  25.211389
#> 49:  0.84722271  1.5214639 0.015625     0       6          1  26.293665
#> 50: -2.19233926 14.3986322 0.015625     0       6          1  27.081951
#> 51:  0.64225953 12.8401983 0.015625     0       6          1  79.306676
#> 52:  7.08313381  7.5813035 0.015625     0       6          1 144.648376
#> 53: -1.28220046  1.9298429 0.015625     0       6          1  50.695572
#> 54:  4.41019314  2.9591669 0.015625     0       6          1  18.567889
#> 55:  3.64022079  2.5349256 0.015625     0       6          1   5.257602
#> 56:  7.36328894  8.0841397 0.015625     0       6          1 161.810667
#> 57:  9.96742970 14.8179224 0.015625     0       6          1 469.499016
#> 58:  2.82872495 10.7848737 0.015625     0       6          1  82.601922
#> 59: -4.16255889  7.6919019 0.015625     0       6          1  34.857093
#> 60:  9.63611738 12.3820839 0.015625     0       6          1 356.440704
#> 61:  6.96328873  5.9025817 0.015625     0       6          1 107.604747
#> 62:  0.85172384  2.8788411 0.015625     0       6          1  19.521386
#> 63:  0.06437588  7.9774898 0.015625     0       6          1  23.907873
#> 64: -4.11735040  7.7072699 0.015625     0       6          1  33.421645
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-30 04:42:56        1  8.94716186   5.0489723          0.015625
#>  2: 2022-05-30 04:42:56        1  6.24017236   1.6897462          0.015625
#>  3: 2022-05-30 04:42:56        1 -2.33842452   8.5197087          0.015625
#>  4: 2022-05-30 04:42:56        1  4.24639881   2.7034374          0.015625
#>  5: 2022-05-30 04:42:56        1  9.77632385  11.4626170          0.015625
#>  6: 2022-05-30 04:42:56        1  4.30750134   6.8891144          0.015625
#>  7: 2022-05-30 04:42:56        1  5.32363806  11.9279702          0.015625
#>  8: 2022-05-30 04:42:56        1  9.12833078   1.0583434          0.015625
#>  9: 2022-05-30 04:42:56        1 -3.52800496   3.4937151          0.015625
#> 10: 2022-05-30 04:42:56        1  9.22957138   1.9479533          0.015625
#> 11: 2022-05-30 04:42:56        1  3.47263228  10.2360238          0.015625
#> 12: 2022-05-30 04:42:56        1  0.04339284   2.0610665          0.015625
#> 13: 2022-05-30 04:42:56        1  0.45248301   3.5519557          0.015625
#> 14: 2022-05-30 04:42:56        1  6.46474914  10.6372837          0.015625
#> 15: 2022-05-30 04:42:56        1 -4.10429497  14.6987868          0.015625
#> 16: 2022-05-30 04:42:56        1  1.43568164  13.8434865          0.015625
#> 17: 2022-05-30 04:42:56        1  5.89176095   5.0265278          0.015625
#> 18: 2022-05-30 04:42:56        1  3.92029547   9.1224532          0.015625
#> 19: 2022-05-30 04:42:56        1  4.88618490  13.4998615          0.015625
#> 20: 2022-05-30 04:42:56        1 -0.01904810  12.3390650          0.015625
#> 21: 2022-05-30 04:42:56        1  2.70544876   2.6095244          0.015625
#> 22: 2022-05-30 04:42:56        1  9.57271309   6.0549350          0.015625
#> 23: 2022-05-30 04:42:56        1 -1.65169897   7.1377458          0.015625
#> 24: 2022-05-30 04:42:56        1  0.20123570   5.8746856          0.015625
#> 25: 2022-05-30 04:42:56        1  3.27754726  11.1269016          0.015625
#> 26: 2022-05-30 04:42:56        1 -4.13510249   8.0217845          0.015625
#> 27: 2022-05-30 04:42:56        1  2.41287742   3.5232036          0.015625
#> 28: 2022-05-30 04:42:56        1 -2.69437727   5.5072074          0.015625
#> 29: 2022-05-30 04:42:56        1  5.04620213   9.6052242          0.015625
#> 30: 2022-05-30 04:42:56        1  6.75004043   2.0071129          0.015625
#> 31: 2022-05-30 04:42:56        1 -0.25959290   8.3585202          0.015625
#> 32: 2022-05-30 04:42:56        1 -0.78126769  10.2475340          0.015625
#> 33: 2022-05-30 04:42:56        1  4.92119459   3.6988574          0.015625
#> 34: 2022-05-30 04:42:56        1 -0.63033440   7.0249197          0.015625
#> 35: 2022-05-30 04:42:56        1 -4.81565895   5.1720604          0.015625
#> 36: 2022-05-30 04:42:56        1 -0.27647325   7.2368704          0.015625
#> 37: 2022-05-30 04:42:56        1  4.84771434  10.4887107          0.015625
#> 38: 2022-05-30 04:42:56        1  9.69336656   2.6626824          0.015625
#> 39: 2022-05-30 04:42:56        1 -4.43084163  10.7627844          0.015625
#> 40: 2022-05-30 04:42:56        1 -0.93362290   3.9966452          0.015625
#> 41: 2022-05-30 04:42:56        1  3.80826061   0.4938213          0.015625
#> 42: 2022-05-30 04:42:56        1  7.74239810   1.4019360          0.015625
#> 43: 2022-05-30 04:42:56        1  6.02776977  11.2524975          0.015625
#> 44: 2022-05-30 04:42:56        1  9.37677815   3.2210212          0.015625
#> 45: 2022-05-30 04:42:56        1  2.02127237   9.0967439          0.015625
#> 46: 2022-05-30 04:42:56        1  1.83318352  12.8565820          0.015625
#> 47: 2022-05-30 04:42:56        1  2.41564016  12.6248475          0.015625
#> 48: 2022-05-30 04:42:56        1 -0.35700391   9.0650398          0.015625
#> 49: 2022-05-30 04:42:56        1  0.84722271   1.5214639          0.015625
#> 50: 2022-05-30 04:42:56        1 -2.19233926  14.3986322          0.015625
#> 51: 2022-05-30 04:42:56        1  0.64225953  12.8401983          0.015625
#> 52: 2022-05-30 04:42:56        1  7.08313381   7.5813035          0.015625
#> 53: 2022-05-30 04:42:56        1 -1.28220046   1.9298429          0.015625
#> 54: 2022-05-30 04:42:56        1  4.41019314   2.9591669          0.015625
#> 55: 2022-05-30 04:42:56        1  3.64022079   2.5349256          0.015625
#> 56: 2022-05-30 04:42:56        1  7.36328894   8.0841397          0.015625
#> 57: 2022-05-30 04:42:56        1  9.96742970  14.8179224          0.015625
#> 58: 2022-05-30 04:42:56        1  2.82872495  10.7848737          0.015625
#> 59: 2022-05-30 04:42:56        1 -4.16255889   7.6919019          0.015625
#> 60: 2022-05-30 04:42:56        1  9.63611738  12.3820839          0.015625
#> 61: 2022-05-30 04:42:56        1  6.96328873   5.9025817          0.015625
#> 62: 2022-05-30 04:42:56        1  0.85172384   2.8788411          0.015625
#> 63: 2022-05-30 04:42:56        1  0.06437588   7.9774898          0.015625
#> 64: 2022-05-30 04:42:56        1 -4.11735040   7.7072699          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity