Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain         y
#> 1: 3.419852 0.8254997 0.015625 <list[3]> 0.7755558

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain         y
#> 1: 3.419852 0.8254997 0.015625 <list[3]> 0.7755558

# all evaluations
as.data.table(instance$archive)
#>              x1           x2 fidelity stage bracket repetition           y
#>  1:  2.85163275  9.747485718 0.015625     0       6          1  65.3755604
#>  2: -3.92656729  3.088041188 0.015625     0       6          1  96.0471828
#>  3: -0.00761159 13.809034150 0.015625     0       6          1  80.3937671
#>  4: -3.19949209 14.135417385 0.015625     0       6          1   7.8587889
#>  5:  1.54711929  3.325992379 0.015625     0       6          1  10.3087165
#>  6:  8.02073314  0.939541011 0.015625     0       6          1  41.2037731
#>  7:  6.50167110  6.805650849 0.015625     0       6          1 116.4683093
#>  8: -2.27450329  0.278082822 0.015625     0       6          1  94.0553433
#>  9: -3.32313454 10.927346338 0.015625     0       6          1   1.0472708
#> 10:  6.79181043  5.377347919 0.015625     0       6          1  95.2737590
#> 11:  7.82006047 14.681169307 0.015625     0       6          1 380.7687973
#> 12:  6.07432619  0.795515324 0.015625     0       6          1  30.4731506
#> 13:  5.04886302  0.630161004 0.015625     0       6          1  16.7118031
#> 14: -4.07279473  9.221840641 0.015625     0       6          1  18.4834545
#> 15:  7.43077138  8.752899425 0.015625     0       6          1 179.8791413
#> 16: -0.54820465  4.816130775 0.015625     0       6          1  22.4617554
#> 17: -4.57205033  8.436085907 0.015625     0       6          1  38.7232610
#> 18: -0.71662667  2.986288044 0.015625     0       6          1  34.6295487
#> 19: -4.37882404  4.751332565 0.015625     0       6          1  84.4251746
#> 20: -1.22148340  2.019858649 0.015625     0       6          1  48.9280544
#> 21: -1.71569974  0.829444699 0.015625     0       6          1  72.4804972
#> 22:  6.46087910  3.666097656 0.015625     0       6          1  63.8789408
#> 23: -4.86573653  4.873268791 0.015625     0       6          1 103.6024848
#> 24:  3.42588899  4.527079817 0.015625     0       6          1  13.8781871
#> 25:  8.47469377 12.393674942 0.015625     0       6          1 316.7609958
#> 26: -4.83200258  0.005232386 0.015625     0       6          1 218.5930043
#> 27: -1.71955878  1.074087792 0.015625     0       6          1  68.6949914
#> 28:  5.15164662  5.812043846 0.015625     0       6          1  65.8534357
#> 29:  3.41985182  0.825499742 0.015625     0       6          1   0.7755558
#> 30: -4.57865062  2.904013081 0.015625     0       6          1 130.3302023
#> 31:  7.85783780  1.897006613 0.015625     0       6          1  52.2734561
#> 32:  5.35645785 13.461940378 0.015625     0       6          1 243.9213868
#> 33: -2.52730582 10.639732021 0.015625     0       6          1   2.3305501
#> 34:  9.22926459 12.888899292 0.015625     0       6          1 360.0139136
#> 35:  1.38945246 12.791548694 0.015625     0       6          1  91.7192265
#> 36: -0.61457853 11.688684571 0.015625     0       6          1  39.9251137
#> 37:  9.07135794 12.226841489 0.015625     0       6          1 329.8405233
#> 38:  1.71899210  5.230438644 0.015625     0       6          1  12.0995012
#> 39:  0.76952104 10.664615566 0.015625     0       6          1  51.3669344
#> 40: -4.52381172 10.323867827 0.015625     0       6          1  20.4866506
#> 41:  1.35831736  8.434071679 0.015625     0       6          1  32.6290109
#> 42:  2.25451943 10.385260214 0.015625     0       6          1  65.0424406
#> 43: -0.34650409  0.924746117 0.015625     0       6          1  50.7330914
#> 44:  7.63443321  7.123828557 0.015625     0       6          1 143.9351398
#> 45: -3.40824942  8.182331573 0.015625     0       6          1  13.6916957
#> 46:  4.17307615  8.881253544 0.015625     0       6          1  85.8433913
#> 47:  2.28261022 11.498918838 0.015625     0       6          1  84.2175061
#> 48:  2.41110575  7.811665518 0.015625     0       6          1  32.7722042
#> 49:  8.48807121  1.555723114 0.015625     0       6          1  51.2292265
#> 50:  6.24310946 13.123476331 0.015625     0       6          1 271.1747783
#> 51: -0.63852626 12.433004568 0.015625     0       6          1  46.9158763
#> 52:  7.71203523 12.485520132 0.015625     0       6          1 298.0141960
#> 53:  0.54939455  6.285130262 0.015625     0       6          1  19.5121219
#> 54: -3.95544043  0.105657560 0.015625     0       6          1 163.9521474
#> 55:  1.48937912  4.999825610 0.015625     0       6          1  12.4762260
#> 56: -4.03833208 13.581921710 0.015625     0       6          1   4.4335266
#> 57:  9.73155082  3.081098524 0.015625     0       6          1  94.1136062
#> 58:  7.31178038  1.389333063 0.015625     0       6          1  43.9272087
#> 59:  1.57446201 13.431339272 0.015625     0       6          1 107.2032094
#> 60:  6.64215238 11.173349958 0.015625     0       6          1 226.0092661
#> 61:  7.73595022  5.913485246 0.015625     0       6          1 118.9909918
#> 62: -3.78249768  0.017998765 0.015625     0       6          1 157.1052400
#> 63: -3.23010188  7.520767618 0.015625     0       6          1  15.9661657
#> 64:  1.94663151  6.606115691 0.015625     0       6          1  19.3475997
#>              x1           x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1  x_domain_x2 x_domain_fidelity
#>  1: 2022-02-26 04:32:33        1  2.85163275  9.747485718          0.015625
#>  2: 2022-02-26 04:32:33        1 -3.92656729  3.088041188          0.015625
#>  3: 2022-02-26 04:32:33        1 -0.00761159 13.809034150          0.015625
#>  4: 2022-02-26 04:32:33        1 -3.19949209 14.135417385          0.015625
#>  5: 2022-02-26 04:32:33        1  1.54711929  3.325992379          0.015625
#>  6: 2022-02-26 04:32:33        1  8.02073314  0.939541011          0.015625
#>  7: 2022-02-26 04:32:33        1  6.50167110  6.805650849          0.015625
#>  8: 2022-02-26 04:32:33        1 -2.27450329  0.278082822          0.015625
#>  9: 2022-02-26 04:32:33        1 -3.32313454 10.927346338          0.015625
#> 10: 2022-02-26 04:32:33        1  6.79181043  5.377347919          0.015625
#> 11: 2022-02-26 04:32:33        1  7.82006047 14.681169307          0.015625
#> 12: 2022-02-26 04:32:33        1  6.07432619  0.795515324          0.015625
#> 13: 2022-02-26 04:32:33        1  5.04886302  0.630161004          0.015625
#> 14: 2022-02-26 04:32:33        1 -4.07279473  9.221840641          0.015625
#> 15: 2022-02-26 04:32:33        1  7.43077138  8.752899425          0.015625
#> 16: 2022-02-26 04:32:33        1 -0.54820465  4.816130775          0.015625
#> 17: 2022-02-26 04:32:33        1 -4.57205033  8.436085907          0.015625
#> 18: 2022-02-26 04:32:33        1 -0.71662667  2.986288044          0.015625
#> 19: 2022-02-26 04:32:33        1 -4.37882404  4.751332565          0.015625
#> 20: 2022-02-26 04:32:33        1 -1.22148340  2.019858649          0.015625
#> 21: 2022-02-26 04:32:33        1 -1.71569974  0.829444699          0.015625
#> 22: 2022-02-26 04:32:33        1  6.46087910  3.666097656          0.015625
#> 23: 2022-02-26 04:32:33        1 -4.86573653  4.873268791          0.015625
#> 24: 2022-02-26 04:32:33        1  3.42588899  4.527079817          0.015625
#> 25: 2022-02-26 04:32:33        1  8.47469377 12.393674942          0.015625
#> 26: 2022-02-26 04:32:33        1 -4.83200258  0.005232386          0.015625
#> 27: 2022-02-26 04:32:33        1 -1.71955878  1.074087792          0.015625
#> 28: 2022-02-26 04:32:33        1  5.15164662  5.812043846          0.015625
#> 29: 2022-02-26 04:32:33        1  3.41985182  0.825499742          0.015625
#> 30: 2022-02-26 04:32:33        1 -4.57865062  2.904013081          0.015625
#> 31: 2022-02-26 04:32:33        1  7.85783780  1.897006613          0.015625
#> 32: 2022-02-26 04:32:33        1  5.35645785 13.461940378          0.015625
#> 33: 2022-02-26 04:32:33        1 -2.52730582 10.639732021          0.015625
#> 34: 2022-02-26 04:32:33        1  9.22926459 12.888899292          0.015625
#> 35: 2022-02-26 04:32:33        1  1.38945246 12.791548694          0.015625
#> 36: 2022-02-26 04:32:33        1 -0.61457853 11.688684571          0.015625
#> 37: 2022-02-26 04:32:33        1  9.07135794 12.226841489          0.015625
#> 38: 2022-02-26 04:32:33        1  1.71899210  5.230438644          0.015625
#> 39: 2022-02-26 04:32:33        1  0.76952104 10.664615566          0.015625
#> 40: 2022-02-26 04:32:33        1 -4.52381172 10.323867827          0.015625
#> 41: 2022-02-26 04:32:33        1  1.35831736  8.434071679          0.015625
#> 42: 2022-02-26 04:32:33        1  2.25451943 10.385260214          0.015625
#> 43: 2022-02-26 04:32:33        1 -0.34650409  0.924746117          0.015625
#> 44: 2022-02-26 04:32:33        1  7.63443321  7.123828557          0.015625
#> 45: 2022-02-26 04:32:33        1 -3.40824942  8.182331573          0.015625
#> 46: 2022-02-26 04:32:33        1  4.17307615  8.881253544          0.015625
#> 47: 2022-02-26 04:32:33        1  2.28261022 11.498918838          0.015625
#> 48: 2022-02-26 04:32:33        1  2.41110575  7.811665518          0.015625
#> 49: 2022-02-26 04:32:33        1  8.48807121  1.555723114          0.015625
#> 50: 2022-02-26 04:32:33        1  6.24310946 13.123476331          0.015625
#> 51: 2022-02-26 04:32:33        1 -0.63852626 12.433004568          0.015625
#> 52: 2022-02-26 04:32:33        1  7.71203523 12.485520132          0.015625
#> 53: 2022-02-26 04:32:33        1  0.54939455  6.285130262          0.015625
#> 54: 2022-02-26 04:32:33        1 -3.95544043  0.105657560          0.015625
#> 55: 2022-02-26 04:32:33        1  1.48937912  4.999825610          0.015625
#> 56: 2022-02-26 04:32:33        1 -4.03833208 13.581921710          0.015625
#> 57: 2022-02-26 04:32:33        1  9.73155082  3.081098524          0.015625
#> 58: 2022-02-26 04:32:33        1  7.31178038  1.389333063          0.015625
#> 59: 2022-02-26 04:32:33        1  1.57446201 13.431339272          0.015625
#> 60: 2022-02-26 04:32:33        1  6.64215238 11.173349958          0.015625
#> 61: 2022-02-26 04:32:33        1  7.73595022  5.913485246          0.015625
#> 62: 2022-02-26 04:32:33        1 -3.78249768  0.017998765          0.015625
#> 63: 2022-02-26 04:32:33        1 -3.23010188  7.520767618          0.015625
#> 64: 2022-02-26 04:32:33        1  1.94663151  6.606115691          0.015625
#>               timestamp batch_nr x_domain_x1  x_domain_x2 x_domain_fidelity