Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1      x2 fidelity  x_domain         y
#> 1: -3.196447 11.4827 0.015625 <list[3]> 0.4189315

# best scoring evaluation
instance$result
#>           x1      x2 fidelity  x_domain         y
#> 1: -3.196447 11.4827 0.015625 <list[3]> 0.4189315

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -2.7456421  6.9733220 0.015625     0       6          1  14.3052947
#>  2: -4.7981369  3.7876506 0.015625     0       6          1 122.2659016
#>  3: -1.7672022 12.5429125 0.015625     0       6          1  21.3343257
#>  4: -3.7670934  8.2638729 0.015625     0       6          1  19.5878554
#>  5:  9.9550274 10.7441755 0.015625     0       6          1 309.4026313
#>  6:  9.6024556  8.2528323 0.015625     0       6          1 216.6543705
#>  7: -3.2233320 11.0392799 0.015625     0       6          1   0.5982682
#>  8:  7.0881554  1.4593884 0.015625     0       6          1  43.6513414
#>  9:  5.4388554  1.4135526 0.015625     0       6          1  26.3650741
#> 10:  7.7488109  0.5236520 0.015625     0       6          1  36.1090223
#> 11:  2.5836118 14.1000899 0.015625     0       6          1 146.0174752
#> 12: -0.2295242  3.7337518 0.015625     0       6          1  26.2838640
#> 13: -2.4374265  6.1598080 0.015625     0       6          1  17.9084791
#> 14:  0.1915207  4.9594343 0.015625     0       6          1  19.9695370
#> 15:  1.4411921  4.8966346 0.015625     0       6          1  12.5099969
#> 16:  9.2256902 13.6437941 0.015625     0       6          1 389.0698758
#> 17:  9.6975986 13.8310915 0.015625     0       6          1 415.8425151
#> 18:  1.8423751  9.6885288 0.015625     0       6          1  49.8875974
#> 19:  2.3732775 10.2689145 0.015625     0       6          1  65.0781902
#> 20:  1.4300267 12.5808859 0.015625     0       6          1  88.6810838
#> 21: -3.1609655  1.1590248 0.015625     0       6          1 104.0121999
#> 22:  7.5247778  8.9692226 0.015625     0       6          1 187.4581358
#> 23:  7.2602633  2.7465861 0.015625     0       6          1  60.0067702
#> 24:  8.0213590 10.9316502 0.015625     0       6          1 255.5095799
#> 25: -3.1210823 13.3579677 0.015625     0       6          1   4.7726043
#> 26:  5.4455219 11.3867119 0.015625     0       6          1 189.1317677
#> 27:  4.8109326  7.0904576 0.015625     0       6          1  75.5164154
#> 28:  1.1167778  8.4560320 0.015625     0       6          1  31.8100931
#> 29:  4.2884634 14.9637243 0.015625     0       6          1 237.8070281
#> 30:  0.9419265  0.9219113 0.015625     0       6          1  28.6532699
#> 31: -3.8139501 14.4555195 0.015625     0       6          1   6.2442632
#> 32:  2.3242107  3.9267518 0.015625     0       6          1   5.5617595
#> 33:  6.4191461 12.6125872 0.015625     0       6          1 261.6903292
#> 34:  4.7757261 11.7622301 0.015625     0       6          1 170.9281863
#> 35:  1.9684574 14.5078522 0.015625     0       6          1 139.0289576
#> 36:  5.2425284  3.0238762 0.015625     0       6          1  35.3089409
#> 37:  9.3149854  9.0222696 0.015625     0       6          1 230.8772811
#> 38:  3.5397761  3.9704800 0.015625     0       6          1  11.5107176
#> 39:  6.9760568 13.3186171 0.015625     0       6          1 303.8450515
#> 40:  2.3836942 12.7264418 0.015625     0       6          1 110.0556202
#> 41:  2.1022704  9.1686999 0.015625     0       6          1  45.8211568
#> 42: -1.1047834  7.4693265 0.015625     0       6          1  14.4211123
#> 43: -2.2572918 13.7927644 0.015625     0       6          1  20.2638267
#> 44:  4.4501540  2.9336077 0.015625     0       6          1  19.1207035
#> 45: -0.4309290 11.7082068 0.015625     0       6          1  43.8910702
#> 46:  9.6879086  7.3192323 0.015625     0       6          1 192.6129298
#> 47:  0.5220382  4.1254284 0.015625     0       6          1  19.4300725
#> 48: -2.0504513 11.5612515 0.015625     0       6          1  10.2716474
#> 49: -2.4193287  3.7738977 0.015625     0       6          1  41.9398523
#> 50: -1.3115655 12.1138091 0.015625     0       6          1  28.2500547
#> 51:  1.2851979  2.3456085 0.015625     0       6          1  15.4598941
#> 52:  3.6289954  4.6336971 0.015625     0       6          1  17.5520254
#> 53: -3.1964475 11.4827022 0.015625     0       6          1   0.4189315
#> 54:  3.4981862  7.8615782 0.015625     0       6          1  50.7446913
#> 55:  2.9481182 11.6764726 0.015625     0       6          1 102.6135819
#> 56:  9.1776093 14.4321373 0.015625     0       6          1 418.8505448
#> 57: -4.0885070  2.3665275 0.015625     0       6          1 117.9088469
#> 58:  0.4311058  7.4989311 0.015625     0       6          1  23.4730951
#> 59:  2.4636617  2.1047423 0.015625     0       6          1   2.5470529
#> 60:  3.2152967 11.2668235 0.015625     0       6          1 101.7535548
#> 61:  7.5724507 12.4327686 0.015625     0       6          1 292.2797246
#> 62: -1.2038692 10.8377055 0.015625     0       6          1  21.7226038
#> 63: -3.2387117 10.5389205 0.015625     0       6          1   1.3232846
#> 64:  7.1517821  7.2542104 0.015625     0       6          1 138.6135882
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-04-16 04:29:00        1  -2.7456421   6.9733220          0.015625
#>  2: 2022-04-16 04:29:00        1  -4.7981369   3.7876506          0.015625
#>  3: 2022-04-16 04:29:00        1  -1.7672022  12.5429125          0.015625
#>  4: 2022-04-16 04:29:00        1  -3.7670934   8.2638729          0.015625
#>  5: 2022-04-16 04:29:00        1   9.9550274  10.7441755          0.015625
#>  6: 2022-04-16 04:29:00        1   9.6024556   8.2528323          0.015625
#>  7: 2022-04-16 04:29:00        1  -3.2233320  11.0392799          0.015625
#>  8: 2022-04-16 04:29:00        1   7.0881554   1.4593884          0.015625
#>  9: 2022-04-16 04:29:00        1   5.4388554   1.4135526          0.015625
#> 10: 2022-04-16 04:29:00        1   7.7488109   0.5236520          0.015625
#> 11: 2022-04-16 04:29:00        1   2.5836118  14.1000899          0.015625
#> 12: 2022-04-16 04:29:00        1  -0.2295242   3.7337518          0.015625
#> 13: 2022-04-16 04:29:00        1  -2.4374265   6.1598080          0.015625
#> 14: 2022-04-16 04:29:00        1   0.1915207   4.9594343          0.015625
#> 15: 2022-04-16 04:29:00        1   1.4411921   4.8966346          0.015625
#> 16: 2022-04-16 04:29:00        1   9.2256902  13.6437941          0.015625
#> 17: 2022-04-16 04:29:00        1   9.6975986  13.8310915          0.015625
#> 18: 2022-04-16 04:29:00        1   1.8423751   9.6885288          0.015625
#> 19: 2022-04-16 04:29:00        1   2.3732775  10.2689145          0.015625
#> 20: 2022-04-16 04:29:00        1   1.4300267  12.5808859          0.015625
#> 21: 2022-04-16 04:29:00        1  -3.1609655   1.1590248          0.015625
#> 22: 2022-04-16 04:29:00        1   7.5247778   8.9692226          0.015625
#> 23: 2022-04-16 04:29:00        1   7.2602633   2.7465861          0.015625
#> 24: 2022-04-16 04:29:00        1   8.0213590  10.9316502          0.015625
#> 25: 2022-04-16 04:29:00        1  -3.1210823  13.3579677          0.015625
#> 26: 2022-04-16 04:29:00        1   5.4455219  11.3867119          0.015625
#> 27: 2022-04-16 04:29:00        1   4.8109326   7.0904576          0.015625
#> 28: 2022-04-16 04:29:00        1   1.1167778   8.4560320          0.015625
#> 29: 2022-04-16 04:29:00        1   4.2884634  14.9637243          0.015625
#> 30: 2022-04-16 04:29:00        1   0.9419265   0.9219113          0.015625
#> 31: 2022-04-16 04:29:00        1  -3.8139501  14.4555195          0.015625
#> 32: 2022-04-16 04:29:00        1   2.3242107   3.9267518          0.015625
#> 33: 2022-04-16 04:29:00        1   6.4191461  12.6125872          0.015625
#> 34: 2022-04-16 04:29:00        1   4.7757261  11.7622301          0.015625
#> 35: 2022-04-16 04:29:00        1   1.9684574  14.5078522          0.015625
#> 36: 2022-04-16 04:29:00        1   5.2425284   3.0238762          0.015625
#> 37: 2022-04-16 04:29:00        1   9.3149854   9.0222696          0.015625
#> 38: 2022-04-16 04:29:00        1   3.5397761   3.9704800          0.015625
#> 39: 2022-04-16 04:29:00        1   6.9760568  13.3186171          0.015625
#> 40: 2022-04-16 04:29:00        1   2.3836942  12.7264418          0.015625
#> 41: 2022-04-16 04:29:00        1   2.1022704   9.1686999          0.015625
#> 42: 2022-04-16 04:29:00        1  -1.1047834   7.4693265          0.015625
#> 43: 2022-04-16 04:29:00        1  -2.2572918  13.7927644          0.015625
#> 44: 2022-04-16 04:29:00        1   4.4501540   2.9336077          0.015625
#> 45: 2022-04-16 04:29:00        1  -0.4309290  11.7082068          0.015625
#> 46: 2022-04-16 04:29:00        1   9.6879086   7.3192323          0.015625
#> 47: 2022-04-16 04:29:00        1   0.5220382   4.1254284          0.015625
#> 48: 2022-04-16 04:29:00        1  -2.0504513  11.5612515          0.015625
#> 49: 2022-04-16 04:29:00        1  -2.4193287   3.7738977          0.015625
#> 50: 2022-04-16 04:29:00        1  -1.3115655  12.1138091          0.015625
#> 51: 2022-04-16 04:29:00        1   1.2851979   2.3456085          0.015625
#> 52: 2022-04-16 04:29:00        1   3.6289954   4.6336971          0.015625
#> 53: 2022-04-16 04:29:00        1  -3.1964475  11.4827022          0.015625
#> 54: 2022-04-16 04:29:00        1   3.4981862   7.8615782          0.015625
#> 55: 2022-04-16 04:29:00        1   2.9481182  11.6764726          0.015625
#> 56: 2022-04-16 04:29:00        1   9.1776093  14.4321373          0.015625
#> 57: 2022-04-16 04:29:00        1  -4.0885070   2.3665275          0.015625
#> 58: 2022-04-16 04:29:00        1   0.4311058   7.4989311          0.015625
#> 59: 2022-04-16 04:29:00        1   2.4636617   2.1047423          0.015625
#> 60: 2022-04-16 04:29:00        1   3.2152967  11.2668235          0.015625
#> 61: 2022-04-16 04:29:00        1   7.5724507  12.4327686          0.015625
#> 62: 2022-04-16 04:29:00        1  -1.2038692  10.8377055          0.015625
#> 63: 2022-04-16 04:29:00        1  -3.2387117  10.5389205          0.015625
#> 64: 2022-04-16 04:29:00        1   7.1517821   7.2542104          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity