Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.832355 12.00665 0.015625 <list[3]> 2.895349

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.832355 12.00665 0.015625 <list[3]> 2.895349

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1: -0.4959221  5.9431222 0.015625     0       6          1  19.174194
#>  2:  2.0528076 13.9594335 0.015625     0       6          1 128.692449
#>  3: -4.5548081  8.0387238 0.015625     0       6          1  42.696512
#>  4: -2.2884465  8.6543532 0.015625     0       6          1   5.005338
#>  5:  5.7613224  4.0543790 0.015625     0       6          1  56.804033
#>  6:  5.4842494 11.9449220 0.015625     0       6          1 205.721145
#>  7:  3.3708459  4.3201816 0.015625     0       6          1  11.775897
#>  8:  1.6290233 14.4620721 0.015625     0       6          1 129.851220
#>  9:  9.7594576  4.6999591 0.015625     0       6          1 128.712372
#> 10: -3.8323554 12.0066502 0.015625     0       6          1   2.895349
#> 11:  5.0127102 10.0200025 0.015625     0       6          1 138.849956
#> 12: -2.4262659  6.7200669 0.015625     0       6          1  13.790279
#> 13:  7.2175880 10.4874108 0.015625     0       6          1 222.284975
#> 14:  4.5687875  1.3411823 0.015625     0       6          1  12.510012
#> 15:  8.3147494  4.0999416 0.015625     0       6          1  90.510087
#> 16: -4.1618510  6.7888135 0.015625     0       6          1  45.522352
#> 17:  7.7182519 10.7553258 0.015625     0       6          1 242.572293
#> 18:  7.1449361  0.4386925 0.015625     0       6          1  34.234490
#> 19:  8.4453642  8.9593663 0.015625     0       6          1 206.501941
#> 20:  7.6848576  5.0957125 0.015625     0       6          1 102.069938
#> 21:  2.4213433  0.2047741 0.015625     0       6          1   7.284713
#> 22:  5.5614785 12.1366012 0.015625     0       6          1 214.244434
#> 23: -1.6898274 11.4117501 0.015625     0       6          1  15.800377
#> 24:  1.9739770  9.2171011 0.015625     0       6          1  45.157456
#> 25: -1.5907265  8.4320012 0.015625     0       6          1   9.840155
#> 26: -4.7924448  0.6444097 0.015625     0       6          1 198.161693
#> 27: -4.7175403 13.8994819 0.015625     0       6          1  10.135310
#> 28:  6.1063712  1.5847819 0.015625     0       6          1  36.732109
#> 29:  7.3022334  3.6051556 0.015625     0       6          1  72.603515
#> 30:  2.2481729 14.4213272 0.015625     0       6          1 144.262237
#> 31:  1.9681859  5.5304207 0.015625     0       6          1  12.754674
#> 32:  7.2812886 14.4356692 0.015625     0       6          1 353.545564
#> 33:  8.1478769  4.3459596 0.015625     0       6          1  93.197303
#> 34:  8.2492602 10.1551714 0.015625     0       6          1 237.097413
#> 35: -1.0429792  3.4079602 0.015625     0       6          1  33.201084
#> 36: -0.1400127  5.0870671 0.015625     0       6          1  20.799492
#> 37:  3.8897791  1.9395787 0.015625     0       6          1   5.735075
#> 38:  8.5980446 13.8864129 0.015625     0       6          1 375.894450
#> 39: -2.2079570  1.7959940 0.015625     0       6          1  66.192547
#> 40:  6.3494501 11.6614390 0.015625     0       6          1 230.624096
#> 41:  2.5538245  4.8790824 0.015625     0       6          1   9.533841
#> 42:  7.0731902  9.6756163 0.015625     0       6          1 196.175677
#> 43:  6.3752162 10.3784178 0.015625     0       6          1 195.793133
#> 44:  5.2731246 14.3071266 0.015625     0       6          1 266.158431
#> 45:  4.6199813  4.4508234 0.015625     0       6          1  35.610512
#> 46: -3.4805557  2.0060381 0.015625     0       6          1  99.071417
#> 47: -1.2525355  7.9718113 0.015625     0       6          1  13.009533
#> 48:  1.1597707  4.4818515 0.015625     0       6          1  13.918505
#> 49: -2.0835793  7.7233781 0.015625     0       6          1   8.269011
#> 50: -0.3224348  1.5953811 0.015625     0       6          1  43.323391
#> 51:  5.9558661 14.6253051 0.015625     0       6          1 308.558032
#> 52: -1.9936135  9.4597253 0.015625     0       6          1   6.087043
#> 53: -3.4211730  0.8812199 0.015625     0       6          1 120.096244
#> 54: -4.5754962 13.2396650 0.015625     0       6          1   9.160453
#> 55:  8.0406209  4.2816595 0.015625     0       6          1  90.862032
#> 56:  0.7251710 10.3524463 0.015625     0       6          1  47.330829
#> 57: -1.9753162  0.3163281 0.015625     0       6          1  86.277798
#> 58: -0.2136826  4.8663091 0.015625     0       6          1  21.559888
#> 59: -3.1829260  4.4409489 0.015625     0       6          1  48.518814
#> 60:  5.9175023  3.6346016 0.015625     0       6          1  54.679031
#> 61:  3.3895156  9.4664032 0.015625     0       6          1  73.073087
#> 62: -0.1312180  4.1565413 0.015625     0       6          1  23.733669
#> 63:  4.0094862 10.7694834 0.015625     0       6          1 117.353649
#> 64: -1.9044459 13.5622156 0.015625     0       6          1  26.388933
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-12 04:35:45        1  -0.4959221   5.9431222          0.015625
#>  2: 2022-08-12 04:35:45        1   2.0528076  13.9594335          0.015625
#>  3: 2022-08-12 04:35:45        1  -4.5548081   8.0387238          0.015625
#>  4: 2022-08-12 04:35:45        1  -2.2884465   8.6543532          0.015625
#>  5: 2022-08-12 04:35:45        1   5.7613224   4.0543790          0.015625
#>  6: 2022-08-12 04:35:45        1   5.4842494  11.9449220          0.015625
#>  7: 2022-08-12 04:35:45        1   3.3708459   4.3201816          0.015625
#>  8: 2022-08-12 04:35:45        1   1.6290233  14.4620721          0.015625
#>  9: 2022-08-12 04:35:45        1   9.7594576   4.6999591          0.015625
#> 10: 2022-08-12 04:35:45        1  -3.8323554  12.0066502          0.015625
#> 11: 2022-08-12 04:35:45        1   5.0127102  10.0200025          0.015625
#> 12: 2022-08-12 04:35:45        1  -2.4262659   6.7200669          0.015625
#> 13: 2022-08-12 04:35:45        1   7.2175880  10.4874108          0.015625
#> 14: 2022-08-12 04:35:45        1   4.5687875   1.3411823          0.015625
#> 15: 2022-08-12 04:35:45        1   8.3147494   4.0999416          0.015625
#> 16: 2022-08-12 04:35:45        1  -4.1618510   6.7888135          0.015625
#> 17: 2022-08-12 04:35:45        1   7.7182519  10.7553258          0.015625
#> 18: 2022-08-12 04:35:45        1   7.1449361   0.4386925          0.015625
#> 19: 2022-08-12 04:35:45        1   8.4453642   8.9593663          0.015625
#> 20: 2022-08-12 04:35:45        1   7.6848576   5.0957125          0.015625
#> 21: 2022-08-12 04:35:45        1   2.4213433   0.2047741          0.015625
#> 22: 2022-08-12 04:35:45        1   5.5614785  12.1366012          0.015625
#> 23: 2022-08-12 04:35:45        1  -1.6898274  11.4117501          0.015625
#> 24: 2022-08-12 04:35:45        1   1.9739770   9.2171011          0.015625
#> 25: 2022-08-12 04:35:45        1  -1.5907265   8.4320012          0.015625
#> 26: 2022-08-12 04:35:45        1  -4.7924448   0.6444097          0.015625
#> 27: 2022-08-12 04:35:45        1  -4.7175403  13.8994819          0.015625
#> 28: 2022-08-12 04:35:45        1   6.1063712   1.5847819          0.015625
#> 29: 2022-08-12 04:35:45        1   7.3022334   3.6051556          0.015625
#> 30: 2022-08-12 04:35:45        1   2.2481729  14.4213272          0.015625
#> 31: 2022-08-12 04:35:45        1   1.9681859   5.5304207          0.015625
#> 32: 2022-08-12 04:35:45        1   7.2812886  14.4356692          0.015625
#> 33: 2022-08-12 04:35:45        1   8.1478769   4.3459596          0.015625
#> 34: 2022-08-12 04:35:45        1   8.2492602  10.1551714          0.015625
#> 35: 2022-08-12 04:35:45        1  -1.0429792   3.4079602          0.015625
#> 36: 2022-08-12 04:35:45        1  -0.1400127   5.0870671          0.015625
#> 37: 2022-08-12 04:35:45        1   3.8897791   1.9395787          0.015625
#> 38: 2022-08-12 04:35:45        1   8.5980446  13.8864129          0.015625
#> 39: 2022-08-12 04:35:45        1  -2.2079570   1.7959940          0.015625
#> 40: 2022-08-12 04:35:45        1   6.3494501  11.6614390          0.015625
#> 41: 2022-08-12 04:35:45        1   2.5538245   4.8790824          0.015625
#> 42: 2022-08-12 04:35:45        1   7.0731902   9.6756163          0.015625
#> 43: 2022-08-12 04:35:45        1   6.3752162  10.3784178          0.015625
#> 44: 2022-08-12 04:35:45        1   5.2731246  14.3071266          0.015625
#> 45: 2022-08-12 04:35:45        1   4.6199813   4.4508234          0.015625
#> 46: 2022-08-12 04:35:45        1  -3.4805557   2.0060381          0.015625
#> 47: 2022-08-12 04:35:45        1  -1.2525355   7.9718113          0.015625
#> 48: 2022-08-12 04:35:45        1   1.1597707   4.4818515          0.015625
#> 49: 2022-08-12 04:35:45        1  -2.0835793   7.7233781          0.015625
#> 50: 2022-08-12 04:35:45        1  -0.3224348   1.5953811          0.015625
#> 51: 2022-08-12 04:35:45        1   5.9558661  14.6253051          0.015625
#> 52: 2022-08-12 04:35:45        1  -1.9936135   9.4597253          0.015625
#> 53: 2022-08-12 04:35:45        1  -3.4211730   0.8812199          0.015625
#> 54: 2022-08-12 04:35:45        1  -4.5754962  13.2396650          0.015625
#> 55: 2022-08-12 04:35:45        1   8.0406209   4.2816595          0.015625
#> 56: 2022-08-12 04:35:45        1   0.7251710  10.3524463          0.015625
#> 57: 2022-08-12 04:35:45        1  -1.9753162   0.3163281          0.015625
#> 58: 2022-08-12 04:35:45        1  -0.2136826   4.8663091          0.015625
#> 59: 2022-08-12 04:35:45        1  -3.1829260   4.4409489          0.015625
#> 60: 2022-08-12 04:35:45        1   5.9175023   3.6346016          0.015625
#> 61: 2022-08-12 04:35:45        1   3.3895156   9.4664032          0.015625
#> 62: 2022-08-12 04:35:45        1  -0.1312180   4.1565413          0.015625
#> 63: 2022-08-12 04:35:45        1   4.0094862  10.7694834          0.015625
#> 64: 2022-08-12 04:35:45        1  -1.9044459  13.5622156          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity