Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1      x2 fidelity  x_domain        y
#> 1: -2.462309 11.0388 0.015625 <list[3]> 3.400761

# best scoring evaluation
instance$result
#>           x1      x2 fidelity  x_domain        y
#> 1: -2.462309 11.0388 0.015625 <list[3]> 3.400761

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  3.15921019 11.8446531 0.015625     0       6          1 112.035895
#>  2:  0.89498951  2.6271613 0.015625     0       6          1  19.899296
#>  3: -2.84369619  2.9386888 0.015625     0       6          1  62.221085
#>  4:  1.97562149  9.4079661 0.015625     0       6          1  47.592064
#>  5: -4.08434448  6.6410263 0.015625     0       6          1  44.964706
#>  6:  4.92455006 13.5554568 0.015625     0       6          1 226.570323
#>  7:  2.25224096  6.1248745 0.015625     0       6          1  16.578559
#>  8:  7.38885382 12.9494437 0.015625     0       6          1 304.345919
#>  9:  3.08577407  7.0682466 0.015625     0       6          1  32.750666
#> 10: -2.80451652  8.7001921 0.015625     0       6          1   4.958944
#> 11:  1.18613563  0.7310061 0.015625     0       6          1  25.330060
#> 12: -1.39543699 12.1048782 0.015625     0       6          1  26.298939
#> 13: -3.97308367  4.8476133 0.015625     0       6          1  66.909442
#> 14: -1.00866454  0.4428685 0.015625     0       6          1  66.867926
#> 15: -2.46230870 11.0388046 0.015625     0       6          1   3.400761
#> 16: -0.72516454 14.5793244 0.015625     0       6          1  72.079713
#> 17:  7.69830880  7.7676125 0.015625     0       6          1 160.271716
#> 18:  5.91484221  4.2059792 0.015625     0       6          1  61.782615
#> 19:  3.07214865  7.1246562 0.015625     0       6          1  33.184512
#> 20:  6.27591782  6.6389035 0.015625     0       6          1 108.268566
#> 21:  0.64096740 11.6484884 0.015625     0       6          1  61.998433
#> 22:  5.45588031  5.0168503 0.015625     0       6          1  62.534426
#> 23: -3.87045508  9.7405550 0.015625     0       6          1  11.132268
#> 24: -0.08151188  7.3773429 0.015625     0       6          1  21.126259
#> 25:  7.32921206 14.6590472 0.015625     0       6          1 363.461943
#> 26:  8.32557820 12.8153885 0.015625     0       6          1 327.291891
#> 27:  1.62672334 10.6307592 0.015625     0       6          1  60.420106
#> 28: -3.90001510  1.1434495 0.015625     0       6          1 136.000101
#> 29:  5.30397538 10.5982305 0.015625     0       6          1 163.580400
#> 30:  6.77407388  2.0238546 0.015625     0       6          1  47.565709
#> 31:  5.79801199  9.0172669 0.015625     0       6          1 144.191013
#> 32:  2.97784236  3.3127692 0.015625     0       6          1   3.692964
#> 33:  1.31698523 13.2132481 0.015625     0       6          1  98.083947
#> 34:  6.07770837  0.7065399 0.015625     0       6          1  29.922123
#> 35: -4.01539904 12.1025994 0.015625     0       6          1   4.450710
#> 36:  5.81351769  2.6103637 0.015625     0       6          1  41.830545
#> 37: -2.17586850  3.4985458 0.015625     0       6          1  41.870521
#> 38: -1.19393794  7.7408751 0.015625     0       6          1  13.574865
#> 39: -4.81899556  4.0232558 0.015625     0       6          1 118.360489
#> 40:  3.62600659  3.4307046 0.015625     0       6          1   9.328154
#> 41:  2.53762668 12.4611595 0.015625     0       6          1 108.226175
#> 42:  9.49295579  6.0032133 0.015625     0       6          1 152.718445
#> 43:  9.54390428  2.0550365 0.015625     0       6          1  71.767184
#> 44:  9.64208175  7.7391585 0.015625     0       6          1 203.016063
#> 45: -2.03618360  2.3160640 0.015625     0       6          1  55.423003
#> 46:  2.86916250  6.4872768 0.015625     0       6          1  23.797568
#> 47: -1.73572105  8.3848289 0.015625     0       6          1   8.644716
#> 48:  5.82871295  5.8796576 0.015625     0       6          1  84.427863
#> 49:  7.87205337 11.8497216 0.015625     0       6          1 281.190014
#> 50: -3.20441575 14.3013332 0.015625     0       6          1   8.743680
#> 51: -0.04522924 11.8263666 0.015625     0       6          1  52.704481
#> 52: -2.72668466 13.5655562 0.015625     0       6          1  10.196429
#> 53: -2.22144702  4.8103823 0.015625     0       6          1  27.968031
#> 54:  6.38691563 12.6139052 0.015625     0       6          1 260.568171
#> 55:  9.92503137 14.3447738 0.015625     0       6          1 447.298635
#> 56: -4.08747121 11.1999199 0.015625     0       6          1   7.691944
#> 57:  1.01153952  2.2134539 0.015625     0       6          1  19.970131
#> 58: -0.11535956 14.0167037 0.015625     0       6          1  80.889388
#> 59:  1.49356332  2.1306783 0.015625     0       6          1  13.177054
#> 60:  0.73924102  0.7046642 0.015625     0       6          1  34.198955
#> 61:  9.05403142  8.8166711 0.015625     0       6          1 217.319916
#> 62:  2.92437581  9.3064841 0.015625     0       6          1  59.880039
#> 63:  8.94004474  4.2557814 0.015625     0       6          1 102.042145
#> 64:  3.63612611  5.1826367 0.015625     0       6          1  22.371038
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-03-19 04:34:04        1  3.15921019  11.8446531          0.015625
#>  2: 2022-03-19 04:34:04        1  0.89498951   2.6271613          0.015625
#>  3: 2022-03-19 04:34:04        1 -2.84369619   2.9386888          0.015625
#>  4: 2022-03-19 04:34:04        1  1.97562149   9.4079661          0.015625
#>  5: 2022-03-19 04:34:04        1 -4.08434448   6.6410263          0.015625
#>  6: 2022-03-19 04:34:04        1  4.92455006  13.5554568          0.015625
#>  7: 2022-03-19 04:34:04        1  2.25224096   6.1248745          0.015625
#>  8: 2022-03-19 04:34:04        1  7.38885382  12.9494437          0.015625
#>  9: 2022-03-19 04:34:04        1  3.08577407   7.0682466          0.015625
#> 10: 2022-03-19 04:34:04        1 -2.80451652   8.7001921          0.015625
#> 11: 2022-03-19 04:34:04        1  1.18613563   0.7310061          0.015625
#> 12: 2022-03-19 04:34:04        1 -1.39543699  12.1048782          0.015625
#> 13: 2022-03-19 04:34:04        1 -3.97308367   4.8476133          0.015625
#> 14: 2022-03-19 04:34:04        1 -1.00866454   0.4428685          0.015625
#> 15: 2022-03-19 04:34:04        1 -2.46230870  11.0388046          0.015625
#> 16: 2022-03-19 04:34:04        1 -0.72516454  14.5793244          0.015625
#> 17: 2022-03-19 04:34:04        1  7.69830880   7.7676125          0.015625
#> 18: 2022-03-19 04:34:04        1  5.91484221   4.2059792          0.015625
#> 19: 2022-03-19 04:34:04        1  3.07214865   7.1246562          0.015625
#> 20: 2022-03-19 04:34:04        1  6.27591782   6.6389035          0.015625
#> 21: 2022-03-19 04:34:04        1  0.64096740  11.6484884          0.015625
#> 22: 2022-03-19 04:34:04        1  5.45588031   5.0168503          0.015625
#> 23: 2022-03-19 04:34:04        1 -3.87045508   9.7405550          0.015625
#> 24: 2022-03-19 04:34:04        1 -0.08151188   7.3773429          0.015625
#> 25: 2022-03-19 04:34:04        1  7.32921206  14.6590472          0.015625
#> 26: 2022-03-19 04:34:04        1  8.32557820  12.8153885          0.015625
#> 27: 2022-03-19 04:34:04        1  1.62672334  10.6307592          0.015625
#> 28: 2022-03-19 04:34:04        1 -3.90001510   1.1434495          0.015625
#> 29: 2022-03-19 04:34:04        1  5.30397538  10.5982305          0.015625
#> 30: 2022-03-19 04:34:04        1  6.77407388   2.0238546          0.015625
#> 31: 2022-03-19 04:34:04        1  5.79801199   9.0172669          0.015625
#> 32: 2022-03-19 04:34:04        1  2.97784236   3.3127692          0.015625
#> 33: 2022-03-19 04:34:04        1  1.31698523  13.2132481          0.015625
#> 34: 2022-03-19 04:34:04        1  6.07770837   0.7065399          0.015625
#> 35: 2022-03-19 04:34:04        1 -4.01539904  12.1025994          0.015625
#> 36: 2022-03-19 04:34:04        1  5.81351769   2.6103637          0.015625
#> 37: 2022-03-19 04:34:04        1 -2.17586850   3.4985458          0.015625
#> 38: 2022-03-19 04:34:04        1 -1.19393794   7.7408751          0.015625
#> 39: 2022-03-19 04:34:04        1 -4.81899556   4.0232558          0.015625
#> 40: 2022-03-19 04:34:04        1  3.62600659   3.4307046          0.015625
#> 41: 2022-03-19 04:34:04        1  2.53762668  12.4611595          0.015625
#> 42: 2022-03-19 04:34:04        1  9.49295579   6.0032133          0.015625
#> 43: 2022-03-19 04:34:04        1  9.54390428   2.0550365          0.015625
#> 44: 2022-03-19 04:34:04        1  9.64208175   7.7391585          0.015625
#> 45: 2022-03-19 04:34:04        1 -2.03618360   2.3160640          0.015625
#> 46: 2022-03-19 04:34:04        1  2.86916250   6.4872768          0.015625
#> 47: 2022-03-19 04:34:04        1 -1.73572105   8.3848289          0.015625
#> 48: 2022-03-19 04:34:04        1  5.82871295   5.8796576          0.015625
#> 49: 2022-03-19 04:34:04        1  7.87205337  11.8497216          0.015625
#> 50: 2022-03-19 04:34:04        1 -3.20441575  14.3013332          0.015625
#> 51: 2022-03-19 04:34:04        1 -0.04522924  11.8263666          0.015625
#> 52: 2022-03-19 04:34:04        1 -2.72668466  13.5655562          0.015625
#> 53: 2022-03-19 04:34:04        1 -2.22144702   4.8103823          0.015625
#> 54: 2022-03-19 04:34:04        1  6.38691563  12.6139052          0.015625
#> 55: 2022-03-19 04:34:04        1  9.92503137  14.3447738          0.015625
#> 56: 2022-03-19 04:34:04        1 -4.08747121  11.1999199          0.015625
#> 57: 2022-03-19 04:34:04        1  1.01153952   2.2134539          0.015625
#> 58: 2022-03-19 04:34:04        1 -0.11535956  14.0167037          0.015625
#> 59: 2022-03-19 04:34:04        1  1.49356332   2.1306783          0.015625
#> 60: 2022-03-19 04:34:04        1  0.73924102   0.7046642          0.015625
#> 61: 2022-03-19 04:34:04        1  9.05403142   8.8166711          0.015625
#> 62: 2022-03-19 04:34:04        1  2.92437581   9.3064841          0.015625
#> 63: 2022-03-19 04:34:04        1  8.94004474   4.2557814          0.015625
#> 64: 2022-03-19 04:34:04        1  3.63612611   5.1826367          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity