Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.742366 1.359363 0.015625 <list[3]> 1.410295

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.742366 1.359363 0.015625 <list[3]> 1.410295

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  8.53179232  1.3901478 0.015625     0       6          1  49.282244
#>  2:  1.07686102  4.5131188 0.015625     0       6          1  14.588920
#>  3:  0.05841489 11.5678108 0.015625     0       6          1  51.628988
#>  4: -3.98821992  1.4328955 0.015625     0       6          1 133.680850
#>  5:  7.69778872 10.4065256 0.015625     0       6          1 231.600398
#>  6:  0.96522355 10.7981282 0.015625     0       6          1  55.227485
#>  7:  5.10391949  3.0065116 0.015625     0       6          1  32.401778
#>  8: -1.57717151  2.0887023 0.015625     0       6          1  52.161832
#>  9:  4.81678611 12.3124143 0.015625     0       6          1 186.965957
#> 10: -0.29919749  4.0786223 0.015625     0       6          1  24.937048
#> 11: -3.85609449  6.6911173 0.015625     0       6          1  37.594608
#> 12: -1.29209278 11.8486460 0.015625     0       6          1  26.635847
#> 13:  3.60428486 10.7109166 0.015625     0       6          1 102.367479
#> 14: -2.01518288 10.2689279 0.015625     0       6          1   6.749611
#> 15:  4.52414908 10.4809295 0.015625     0       6          1 130.350056
#> 16:  8.15601012  8.8511393 0.015625     0       6          1 197.212083
#> 17:  9.83220525  2.5331344 0.015625     0       6          1  85.993106
#> 18:  0.90827123  8.0044250 0.015625     0       6          1  27.634366
#> 19:  8.25173425 10.9219261 0.015625     0       6          1 261.046006
#> 20:  1.77753738  6.0837365 0.015625     0       6          1  15.956708
#> 21:  5.40642141 13.9721883 0.015625     0       6          1 261.943444
#> 22:  7.87333191  5.6759817 0.015625     0       6          1 115.920510
#> 23: -0.59012395  7.0761199 0.015625     0       6          1  17.994050
#> 24:  9.70381903  8.1647283 0.015625     0       6          1 217.258480
#> 25:  2.30085594  1.7807340 0.015625     0       6          1   4.114779
#> 26: -4.53974987 14.5653307 0.015625     0       6          1   8.849550
#> 27:  5.78837196  2.8791322 0.015625     0       6          1  44.068444
#> 28:  9.43366456 14.9026203 0.015625     0       6          1 449.010305
#> 29: -4.98904626 10.5375985 0.015625     0       6          1  29.995179
#> 30:  7.64234199 14.7272660 0.015625     0       6          1 376.622327
#> 31:  1.91853755 13.2317922 0.015625     0       6          1 110.198774
#> 32:  7.58442872 13.4915804 0.015625     0       6          1 329.179790
#> 33:  2.21005331  3.6743013 0.015625     0       6          1   5.356175
#> 34:  0.02702879  1.5431916 0.015625     0       6          1  39.080353
#> 35: -2.27410738 13.2967264 0.015625     0       6          1  16.168737
#> 36:  9.03821592  5.6670051 0.015625     0       6          1 134.279623
#> 37:  1.82881948  7.6922946 0.015625     0       6          1  27.800871
#> 38: -3.09288918 14.0228697 0.015625     0       6          1   8.284362
#> 39: -0.98439721 14.6235318 0.015625     0       6          1  64.692485
#> 40: -4.78002116  6.3334201 0.015625     0       6          1  74.277394
#> 41:  3.36092021 11.9555039 0.015625     0       6          1 120.689527
#> 42: -1.95959814  6.1299896 0.015625     0       6          1  16.012716
#> 43:  2.42322452 14.4757795 0.015625     0       6          1 150.439771
#> 44:  7.23400574 13.8100135 0.015625     0       6          1 329.374497
#> 45: -0.33453884 11.4395763 0.015625     0       6          1  43.116064
#> 46:  7.00298211  5.9103343 0.015625     0       6          1 108.385260
#> 47: -0.52248656  1.3219767 0.015625     0       6          1  48.769113
#> 48: -3.57091255  9.7468375 0.015625     0       6          1   6.691267
#> 49:  9.47565959  5.1229986 0.015625     0       6          1 131.358582
#> 50: -4.25283689  0.5821087 0.015625     0       6          1 168.114894
#> 51:  8.42390029 11.9699168 0.015625     0       6          1 300.490456
#> 52: -0.99581116  3.2198342 0.015625     0       6          1  34.542602
#> 53: -0.86679465 11.7843523 0.015625     0       6          1  35.414525
#> 54:  3.22986125  5.0592579 0.015625     0       6          1  15.481824
#> 55: -0.73813949  0.4030299 0.015625     0       6          1  63.186734
#> 56:  8.50072300 14.3638641 0.015625     0       6          1 391.175316
#> 57: -3.61760732  2.6076834 0.015625     0       6          1  92.711924
#> 58: -2.23357251 13.8612360 0.015625     0       6          1  21.339179
#> 59:  2.74236570  1.3593626 0.015625     0       6          1   1.410295
#> 60:  4.18400285 13.5551424 0.015625     0       6          1 192.190423
#> 61: -2.12266772 13.2862743 0.015625     0       6          1  19.174212
#> 62:  2.83039990 13.0768169 0.015625     0       6          1 129.346275
#> 63: -0.53262122 11.6507087 0.015625     0       6          1  41.257266
#> 64: -3.82106461  2.1250356 0.015625     0       6          1 110.800749
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-19 04:44:38        1  8.53179232   1.3901478          0.015625
#>  2: 2022-08-19 04:44:38        1  1.07686102   4.5131188          0.015625
#>  3: 2022-08-19 04:44:38        1  0.05841489  11.5678108          0.015625
#>  4: 2022-08-19 04:44:38        1 -3.98821992   1.4328955          0.015625
#>  5: 2022-08-19 04:44:38        1  7.69778872  10.4065256          0.015625
#>  6: 2022-08-19 04:44:38        1  0.96522355  10.7981282          0.015625
#>  7: 2022-08-19 04:44:38        1  5.10391949   3.0065116          0.015625
#>  8: 2022-08-19 04:44:38        1 -1.57717151   2.0887023          0.015625
#>  9: 2022-08-19 04:44:38        1  4.81678611  12.3124143          0.015625
#> 10: 2022-08-19 04:44:38        1 -0.29919749   4.0786223          0.015625
#> 11: 2022-08-19 04:44:38        1 -3.85609449   6.6911173          0.015625
#> 12: 2022-08-19 04:44:38        1 -1.29209278  11.8486460          0.015625
#> 13: 2022-08-19 04:44:38        1  3.60428486  10.7109166          0.015625
#> 14: 2022-08-19 04:44:38        1 -2.01518288  10.2689279          0.015625
#> 15: 2022-08-19 04:44:38        1  4.52414908  10.4809295          0.015625
#> 16: 2022-08-19 04:44:38        1  8.15601012   8.8511393          0.015625
#> 17: 2022-08-19 04:44:38        1  9.83220525   2.5331344          0.015625
#> 18: 2022-08-19 04:44:38        1  0.90827123   8.0044250          0.015625
#> 19: 2022-08-19 04:44:38        1  8.25173425  10.9219261          0.015625
#> 20: 2022-08-19 04:44:38        1  1.77753738   6.0837365          0.015625
#> 21: 2022-08-19 04:44:38        1  5.40642141  13.9721883          0.015625
#> 22: 2022-08-19 04:44:38        1  7.87333191   5.6759817          0.015625
#> 23: 2022-08-19 04:44:38        1 -0.59012395   7.0761199          0.015625
#> 24: 2022-08-19 04:44:38        1  9.70381903   8.1647283          0.015625
#> 25: 2022-08-19 04:44:38        1  2.30085594   1.7807340          0.015625
#> 26: 2022-08-19 04:44:38        1 -4.53974987  14.5653307          0.015625
#> 27: 2022-08-19 04:44:38        1  5.78837196   2.8791322          0.015625
#> 28: 2022-08-19 04:44:38        1  9.43366456  14.9026203          0.015625
#> 29: 2022-08-19 04:44:38        1 -4.98904626  10.5375985          0.015625
#> 30: 2022-08-19 04:44:38        1  7.64234199  14.7272660          0.015625
#> 31: 2022-08-19 04:44:38        1  1.91853755  13.2317922          0.015625
#> 32: 2022-08-19 04:44:38        1  7.58442872  13.4915804          0.015625
#> 33: 2022-08-19 04:44:38        1  2.21005331   3.6743013          0.015625
#> 34: 2022-08-19 04:44:38        1  0.02702879   1.5431916          0.015625
#> 35: 2022-08-19 04:44:38        1 -2.27410738  13.2967264          0.015625
#> 36: 2022-08-19 04:44:38        1  9.03821592   5.6670051          0.015625
#> 37: 2022-08-19 04:44:38        1  1.82881948   7.6922946          0.015625
#> 38: 2022-08-19 04:44:38        1 -3.09288918  14.0228697          0.015625
#> 39: 2022-08-19 04:44:38        1 -0.98439721  14.6235318          0.015625
#> 40: 2022-08-19 04:44:38        1 -4.78002116   6.3334201          0.015625
#> 41: 2022-08-19 04:44:38        1  3.36092021  11.9555039          0.015625
#> 42: 2022-08-19 04:44:38        1 -1.95959814   6.1299896          0.015625
#> 43: 2022-08-19 04:44:38        1  2.42322452  14.4757795          0.015625
#> 44: 2022-08-19 04:44:38        1  7.23400574  13.8100135          0.015625
#> 45: 2022-08-19 04:44:38        1 -0.33453884  11.4395763          0.015625
#> 46: 2022-08-19 04:44:38        1  7.00298211   5.9103343          0.015625
#> 47: 2022-08-19 04:44:38        1 -0.52248656   1.3219767          0.015625
#> 48: 2022-08-19 04:44:38        1 -3.57091255   9.7468375          0.015625
#> 49: 2022-08-19 04:44:38        1  9.47565959   5.1229986          0.015625
#> 50: 2022-08-19 04:44:38        1 -4.25283689   0.5821087          0.015625
#> 51: 2022-08-19 04:44:38        1  8.42390029  11.9699168          0.015625
#> 52: 2022-08-19 04:44:38        1 -0.99581116   3.2198342          0.015625
#> 53: 2022-08-19 04:44:38        1 -0.86679465  11.7843523          0.015625
#> 54: 2022-08-19 04:44:38        1  3.22986125   5.0592579          0.015625
#> 55: 2022-08-19 04:44:38        1 -0.73813949   0.4030299          0.015625
#> 56: 2022-08-19 04:44:38        1  8.50072300  14.3638641          0.015625
#> 57: 2022-08-19 04:44:38        1 -3.61760732   2.6076834          0.015625
#> 58: 2022-08-19 04:44:38        1 -2.23357251  13.8612360          0.015625
#> 59: 2022-08-19 04:44:38        1  2.74236570   1.3593626          0.015625
#> 60: 2022-08-19 04:44:38        1  4.18400285  13.5551424          0.015625
#> 61: 2022-08-19 04:44:38        1 -2.12266772  13.2862743          0.015625
#> 62: 2022-08-19 04:44:38        1  2.83039990  13.0768169          0.015625
#> 63: 2022-08-19 04:44:38        1 -0.53262122  11.6507087          0.015625
#> 64: 2022-08-19 04:44:38        1 -3.82106461   2.1250356          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity