Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.437987 0.3316157 0.015625 <list[3]> 1.130269

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.437987 0.3316157 0.015625 <list[3]> 1.130269

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  8.44080878 10.6769273 0.015625     0       6          1 258.137519
#>  2:  1.83385053 13.3035391 0.015625     0       6          1 109.893276
#>  3:  4.68457794 11.3786476 0.015625     0       6          1 157.589677
#>  4:  2.84825165  4.6385563 0.015625     0       6          1   9.347624
#>  5: -2.45271888  1.3088985 0.015625     0       6          1  79.670551
#>  6: -0.95164013  2.4578989 0.015625     0       6          1  41.424999
#>  7: -2.18765851 11.4396749 0.015625     0       6          1   7.724227
#>  8:  7.65780787  0.6758658 0.015625     0       6          1  37.481081
#>  9: -1.65910628  8.4828256 0.015625     0       6          1   9.211878
#> 10:  8.47701865  3.0974029 0.015625     0       6          1  74.613512
#> 11:  8.27489755 14.2518808 0.015625     0       6          1 379.200720
#> 12:  2.75833518 13.7209889 0.015625     0       6          1 142.159530
#> 13:  9.21873001  1.0931751 0.015625     0       6          1  51.755044
#> 14: -3.19102898  9.7243733 0.015625     0       6          1   3.189820
#> 15:  3.81902250  2.5941984 0.015625     0       6          1   7.463979
#> 16:  0.78390998  7.8483152 0.015625     0       6          1  26.268066
#> 17:  8.95837756  3.9487507 0.015625     0       6          1  96.270146
#> 18: -2.62105287  9.9674391 0.015625     0       6          1   1.842174
#> 19: -1.57169826  5.0522384 0.015625     0       6          1  22.418018
#> 20: -4.83358874  5.1177178 0.015625     0       6          1  97.530772
#> 21:  1.89866852  6.9401339 0.015625     0       6          1  21.738945
#> 22: -4.20685561  3.4895831 0.015625     0       6          1 100.412151
#> 23:  1.73215838  9.0164168 0.015625     0       6          1  40.730833
#> 24: -0.09974793  9.8940712 0.015625     0       6          1  33.504695
#> 25:  4.09547562  3.8945998 0.015625     0       6          1  19.631862
#> 26:  4.22253868 10.2148341 0.015625     0       6          1 113.372015
#> 27:  3.96459888 14.9292436 0.015625     0       6          1 221.204472
#> 28: -1.46873453  3.2368667 0.015625     0       6          1  37.676443
#> 29:  7.15926029  5.1487012 0.015625     0       6          1  96.555514
#> 30: -3.43637388 12.6928909 0.015625     0       6          1   1.552787
#> 31: -4.25252775 14.7830043 0.015625     0       6          1   7.866649
#> 32: -1.41568598  0.2023512 0.015625     0       6          1  77.294539
#> 33: -1.72332850  2.9202653 0.015625     0       6          1  43.514209
#> 34: -0.77724216  6.8518782 0.015625     0       6          1  17.007856
#> 35: -1.47766702 10.9465057 0.015625     0       6          1  17.281655
#> 36:  0.03717285  7.7177401 0.015625     0       6          1  22.752711
#> 37:  0.65836924 11.9670959 0.015625     0       6          1  66.617529
#> 38:  3.76784348 12.6586208 0.015625     0       6          1 151.519773
#> 39:  8.20237224 11.3437653 0.015625     0       6          1 273.378498
#> 40:  4.40270899 11.0145912 0.015625     0       6          1 137.620909
#> 41:  5.99887559 12.5192362 0.015625     0       6          1 243.026190
#> 42: -1.83190794  3.9740085 0.015625     0       6          1  32.970643
#> 43:  8.32412888  2.2023613 0.015625     0       6          1  59.234365
#> 44:  9.09054830  6.0276595 0.015625     0       6          1 143.847508
#> 45:  1.46735653  9.4501959 0.015625     0       6          1  43.702593
#> 46:  3.43798695  0.3316157 0.015625     0       6          1   1.130269
#> 47:  6.93297441  2.6304503 0.015625     0       6          1  55.920963
#> 48: -3.65521631 12.4831542 0.015625     0       6          1   1.701823
#> 49:  4.77210515  2.6847309 0.015625     0       6          1  23.386388
#> 50: -1.74854796  1.8497920 0.015625     0       6          1  57.682528
#> 51:  1.13191651 14.0253477 0.015625     0       6          1 109.874452
#> 52:  6.71566276  5.3641103 0.015625     0       6          1  93.812847
#> 53:  5.56126076 12.6769640 0.015625     0       6          1 229.697215
#> 54: -0.17230836 11.5830627 0.015625     0       6          1  47.633856
#> 55:  6.86356773 13.6833256 0.015625     0       6          1 312.446854
#> 56:  3.49164458  2.9208728 0.015625     0       6          1   5.403419
#> 57:  4.66463092  3.3870424 0.015625     0       6          1  26.697889
#> 58: -0.21523972  0.1625251 0.015625     0       6          1  57.591044
#> 59:  9.09183124  5.6995588 0.015625     0       6          1 136.137144
#> 60:  0.58343228  1.7676242 0.015625     0       6          1  28.998156
#> 61:  8.68568683  0.3269343 0.015625     0       6          1  36.904338
#> 62:  6.18287676 10.1249512 0.015625     0       6          1 183.135634
#> 63:  5.79710358  0.5430593 0.015625     0       6          1  25.976361
#> 64:  7.52646461  1.5870229 0.015625     0       6          1  47.008106
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-16 04:47:29        1  8.44080878  10.6769273          0.015625
#>  2: 2022-08-16 04:47:29        1  1.83385053  13.3035391          0.015625
#>  3: 2022-08-16 04:47:29        1  4.68457794  11.3786476          0.015625
#>  4: 2022-08-16 04:47:29        1  2.84825165   4.6385563          0.015625
#>  5: 2022-08-16 04:47:29        1 -2.45271888   1.3088985          0.015625
#>  6: 2022-08-16 04:47:29        1 -0.95164013   2.4578989          0.015625
#>  7: 2022-08-16 04:47:29        1 -2.18765851  11.4396749          0.015625
#>  8: 2022-08-16 04:47:29        1  7.65780787   0.6758658          0.015625
#>  9: 2022-08-16 04:47:29        1 -1.65910628   8.4828256          0.015625
#> 10: 2022-08-16 04:47:29        1  8.47701865   3.0974029          0.015625
#> 11: 2022-08-16 04:47:29        1  8.27489755  14.2518808          0.015625
#> 12: 2022-08-16 04:47:29        1  2.75833518  13.7209889          0.015625
#> 13: 2022-08-16 04:47:29        1  9.21873001   1.0931751          0.015625
#> 14: 2022-08-16 04:47:29        1 -3.19102898   9.7243733          0.015625
#> 15: 2022-08-16 04:47:29        1  3.81902250   2.5941984          0.015625
#> 16: 2022-08-16 04:47:29        1  0.78390998   7.8483152          0.015625
#> 17: 2022-08-16 04:47:29        1  8.95837756   3.9487507          0.015625
#> 18: 2022-08-16 04:47:29        1 -2.62105287   9.9674391          0.015625
#> 19: 2022-08-16 04:47:29        1 -1.57169826   5.0522384          0.015625
#> 20: 2022-08-16 04:47:29        1 -4.83358874   5.1177178          0.015625
#> 21: 2022-08-16 04:47:29        1  1.89866852   6.9401339          0.015625
#> 22: 2022-08-16 04:47:29        1 -4.20685561   3.4895831          0.015625
#> 23: 2022-08-16 04:47:29        1  1.73215838   9.0164168          0.015625
#> 24: 2022-08-16 04:47:29        1 -0.09974793   9.8940712          0.015625
#> 25: 2022-08-16 04:47:29        1  4.09547562   3.8945998          0.015625
#> 26: 2022-08-16 04:47:29        1  4.22253868  10.2148341          0.015625
#> 27: 2022-08-16 04:47:29        1  3.96459888  14.9292436          0.015625
#> 28: 2022-08-16 04:47:29        1 -1.46873453   3.2368667          0.015625
#> 29: 2022-08-16 04:47:29        1  7.15926029   5.1487012          0.015625
#> 30: 2022-08-16 04:47:29        1 -3.43637388  12.6928909          0.015625
#> 31: 2022-08-16 04:47:29        1 -4.25252775  14.7830043          0.015625
#> 32: 2022-08-16 04:47:29        1 -1.41568598   0.2023512          0.015625
#> 33: 2022-08-16 04:47:29        1 -1.72332850   2.9202653          0.015625
#> 34: 2022-08-16 04:47:29        1 -0.77724216   6.8518782          0.015625
#> 35: 2022-08-16 04:47:29        1 -1.47766702  10.9465057          0.015625
#> 36: 2022-08-16 04:47:29        1  0.03717285   7.7177401          0.015625
#> 37: 2022-08-16 04:47:29        1  0.65836924  11.9670959          0.015625
#> 38: 2022-08-16 04:47:29        1  3.76784348  12.6586208          0.015625
#> 39: 2022-08-16 04:47:29        1  8.20237224  11.3437653          0.015625
#> 40: 2022-08-16 04:47:29        1  4.40270899  11.0145912          0.015625
#> 41: 2022-08-16 04:47:29        1  5.99887559  12.5192362          0.015625
#> 42: 2022-08-16 04:47:29        1 -1.83190794   3.9740085          0.015625
#> 43: 2022-08-16 04:47:29        1  8.32412888   2.2023613          0.015625
#> 44: 2022-08-16 04:47:29        1  9.09054830   6.0276595          0.015625
#> 45: 2022-08-16 04:47:29        1  1.46735653   9.4501959          0.015625
#> 46: 2022-08-16 04:47:29        1  3.43798695   0.3316157          0.015625
#> 47: 2022-08-16 04:47:29        1  6.93297441   2.6304503          0.015625
#> 48: 2022-08-16 04:47:29        1 -3.65521631  12.4831542          0.015625
#> 49: 2022-08-16 04:47:29        1  4.77210515   2.6847309          0.015625
#> 50: 2022-08-16 04:47:29        1 -1.74854796   1.8497920          0.015625
#> 51: 2022-08-16 04:47:29        1  1.13191651  14.0253477          0.015625
#> 52: 2022-08-16 04:47:29        1  6.71566276   5.3641103          0.015625
#> 53: 2022-08-16 04:47:29        1  5.56126076  12.6769640          0.015625
#> 54: 2022-08-16 04:47:29        1 -0.17230836  11.5830627          0.015625
#> 55: 2022-08-16 04:47:29        1  6.86356773  13.6833256          0.015625
#> 56: 2022-08-16 04:47:29        1  3.49164458   2.9208728          0.015625
#> 57: 2022-08-16 04:47:29        1  4.66463092   3.3870424          0.015625
#> 58: 2022-08-16 04:47:29        1 -0.21523972   0.1625251          0.015625
#> 59: 2022-08-16 04:47:29        1  9.09183124   5.6995588          0.015625
#> 60: 2022-08-16 04:47:29        1  0.58343228   1.7676242          0.015625
#> 61: 2022-08-16 04:47:29        1  8.68568683   0.3269343          0.015625
#> 62: 2022-08-16 04:47:29        1  6.18287676  10.1249512          0.015625
#> 63: 2022-08-16 04:47:29        1  5.79710358   0.5430593          0.015625
#> 64: 2022-08-16 04:47:29        1  7.52646461   1.5870229          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity