Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.752112 12.19485 0.015625 <list[3]> 3.619254

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.752112 12.19485 0.015625 <list[3]> 3.619254

# all evaluations
as.data.table(instance$archive)
#>               x1         x2 fidelity stage bracket repetition          y
#>  1:  5.833042325 14.1923093 0.015625     0       6          1 288.581837
#>  2:  0.893804911  4.6214715 0.015625     0       6          1  16.015629
#>  3: -4.032476480 10.6898629 0.015625     0       6          1   8.926877
#>  4: -0.773055638 12.3289749 0.015625     0       6          1  42.681879
#>  5:  8.044127888  3.6809372 0.015625     0       6          1  80.332994
#>  6: -0.909772519  4.4209242 0.015625     0       6          1  25.212569
#>  7:  1.065388116 10.4809828 0.015625     0       6          1  52.369497
#>  8:  1.367223170 10.4661397 0.015625     0       6          1  55.299107
#>  9: -4.302382291  8.9072383 0.015625     0       6          1  26.506760
#> 10:  1.982720349  5.8356372 0.015625     0       6          1  14.394553
#> 11: -1.893903795  5.2987464 0.015625     0       6          1  21.587786
#> 12: -0.691481773  0.9714484 0.015625     0       6          1  55.142567
#> 13: -1.990214384  8.7237215 0.015625     0       6          1   6.409634
#> 14: -4.491862183  9.6856547 0.015625     0       6          1  24.576556
#> 15:  6.444202330  8.1587408 0.015625     0       6          1 143.536369
#> 16:  4.735973607 11.2373817 0.015625     0       6          1 156.280420
#> 17: -4.865195584 10.0827692 0.015625     0       6          1  30.718000
#> 18:  6.629820927  4.5889557 0.015625     0       6          1  79.702209
#> 19: -1.703032070  2.6749694 0.015625     0       6          1  46.245493
#> 20: -4.713916924  7.7560355 0.015625     0       6          1  51.354762
#> 21: -2.752111556 12.1948462 0.015625     0       6          1   3.619254
#> 22: -0.567709121  7.1598360 0.015625     0       6          1  18.156585
#> 23:  3.354268075  6.7086908 0.015625     0       6          1  33.118319
#> 24:  2.708334463  5.4647107 0.015625     0       6          1  13.884950
#> 25:  5.566951998 11.3389727 0.015625     0       6          1 192.703815
#> 26:  5.518956609 11.5611836 0.015625     0       6          1 196.715941
#> 27:  6.262427275  0.6825836 0.015625     0       6          1  31.459183
#> 28:  5.662097918  0.9100854 0.015625     0       6          1  26.428184
#> 29:  2.249305228  2.1498383 0.015625     0       6          1   4.154740
#> 30:  8.974576829 11.0284029 0.015625     0       6          1 284.786292
#> 31:  4.011932933  4.3188908 0.015625     0       6          1  21.528069
#> 32: -0.443412234 12.6332813 0.015625     0       6          1  53.737962
#> 33:  4.345484434 10.6828639 0.015625     0       6          1 127.958704
#> 34:  2.798020418 14.6969504 0.015625     0       6          1 167.612247
#> 35:  6.492951968 11.2947036 0.015625     0       6          1 224.806801
#> 36: -2.926230137  5.7418091 0.015625     0       6          1  27.438763
#> 37: -2.138996375  8.8154050 0.015625     0       6          1   5.365254
#> 38: -2.794026821 13.5208404 0.015625     0       6          1   9.003489
#> 39:  4.824409558  4.5542434 0.015625     0       6          1  41.509519
#> 40:  4.351218152  8.5948937 0.015625     0       6          1  86.493609
#> 41:  9.454659015  7.7392354 0.015625     0       6          1 197.476032
#> 42: -0.005661792  4.7853198 0.015625     0       6          1  21.099381
#> 43:  5.250921880  1.2898823 0.015625     0       6          1  22.760347
#> 44:  9.693034072  1.4949914 0.015625     0       6          1  65.272036
#> 45:  0.263302326  3.4843524 0.015625     0       6          1  23.675811
#> 46:  4.951186388  7.6108056 0.015625     0       6          1  88.608560
#> 47: -0.327742309 12.5365361 0.015625     0       6          1  55.230526
#> 48: -4.557984405 12.6999457 0.015625     0       6          1   9.946732
#> 49: -2.550136985  3.1178163 0.015625     0       6          1  53.020120
#> 50:  8.195594240  7.6387939 0.015625     0       6          1 165.978963
#> 51: -0.593326736  8.7728727 0.015625     0       6          1  21.265148
#> 52:  8.744864245  8.9918699 0.015625     0       6          1 214.481556
#> 53:  3.409419699  5.2043814 0.015625     0       6          1  19.000747
#> 54: -0.620447110  4.1124026 0.015625     0       6          1  26.146683
#> 55:  7.302593341 10.4972537 0.015625     0       6          1 224.701226
#> 56: -4.897814381  7.9292124 0.015625     0       6          1  55.376219
#> 57: -4.592846270  7.1594045 0.015625     0       6          1  55.080210
#> 58:  6.639909773  6.9434900 0.015625     0       6          1 122.134807
#> 59:  8.320644936 14.1330393 0.015625     0       6          1 376.128677
#> 60:  3.350881954 12.0021037 0.015625     0       6          1 121.386593
#> 61:  8.285125049  3.3539615 0.015625     0       6          1  77.044858
#> 62: -4.933106161 14.5655707 0.015625     0       6          1  12.103338
#> 63:  4.037225961  6.5262265 0.015625     0       6          1  45.607649
#> 64:  5.175082281  7.9842968 0.015625     0       6          1 102.594225
#>               x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-06-22 04:38:20        1  5.833042325  14.1923093          0.015625
#>  2: 2022-06-22 04:38:20        1  0.893804911   4.6214715          0.015625
#>  3: 2022-06-22 04:38:20        1 -4.032476480  10.6898629          0.015625
#>  4: 2022-06-22 04:38:20        1 -0.773055638  12.3289749          0.015625
#>  5: 2022-06-22 04:38:20        1  8.044127888   3.6809372          0.015625
#>  6: 2022-06-22 04:38:20        1 -0.909772519   4.4209242          0.015625
#>  7: 2022-06-22 04:38:20        1  1.065388116  10.4809828          0.015625
#>  8: 2022-06-22 04:38:20        1  1.367223170  10.4661397          0.015625
#>  9: 2022-06-22 04:38:20        1 -4.302382291   8.9072383          0.015625
#> 10: 2022-06-22 04:38:20        1  1.982720349   5.8356372          0.015625
#> 11: 2022-06-22 04:38:20        1 -1.893903795   5.2987464          0.015625
#> 12: 2022-06-22 04:38:20        1 -0.691481773   0.9714484          0.015625
#> 13: 2022-06-22 04:38:20        1 -1.990214384   8.7237215          0.015625
#> 14: 2022-06-22 04:38:20        1 -4.491862183   9.6856547          0.015625
#> 15: 2022-06-22 04:38:20        1  6.444202330   8.1587408          0.015625
#> 16: 2022-06-22 04:38:20        1  4.735973607  11.2373817          0.015625
#> 17: 2022-06-22 04:38:20        1 -4.865195584  10.0827692          0.015625
#> 18: 2022-06-22 04:38:20        1  6.629820927   4.5889557          0.015625
#> 19: 2022-06-22 04:38:20        1 -1.703032070   2.6749694          0.015625
#> 20: 2022-06-22 04:38:20        1 -4.713916924   7.7560355          0.015625
#> 21: 2022-06-22 04:38:20        1 -2.752111556  12.1948462          0.015625
#> 22: 2022-06-22 04:38:20        1 -0.567709121   7.1598360          0.015625
#> 23: 2022-06-22 04:38:20        1  3.354268075   6.7086908          0.015625
#> 24: 2022-06-22 04:38:20        1  2.708334463   5.4647107          0.015625
#> 25: 2022-06-22 04:38:20        1  5.566951998  11.3389727          0.015625
#> 26: 2022-06-22 04:38:20        1  5.518956609  11.5611836          0.015625
#> 27: 2022-06-22 04:38:20        1  6.262427275   0.6825836          0.015625
#> 28: 2022-06-22 04:38:20        1  5.662097918   0.9100854          0.015625
#> 29: 2022-06-22 04:38:20        1  2.249305228   2.1498383          0.015625
#> 30: 2022-06-22 04:38:20        1  8.974576829  11.0284029          0.015625
#> 31: 2022-06-22 04:38:20        1  4.011932933   4.3188908          0.015625
#> 32: 2022-06-22 04:38:20        1 -0.443412234  12.6332813          0.015625
#> 33: 2022-06-22 04:38:20        1  4.345484434  10.6828639          0.015625
#> 34: 2022-06-22 04:38:20        1  2.798020418  14.6969504          0.015625
#> 35: 2022-06-22 04:38:20        1  6.492951968  11.2947036          0.015625
#> 36: 2022-06-22 04:38:20        1 -2.926230137   5.7418091          0.015625
#> 37: 2022-06-22 04:38:20        1 -2.138996375   8.8154050          0.015625
#> 38: 2022-06-22 04:38:20        1 -2.794026821  13.5208404          0.015625
#> 39: 2022-06-22 04:38:20        1  4.824409558   4.5542434          0.015625
#> 40: 2022-06-22 04:38:20        1  4.351218152   8.5948937          0.015625
#> 41: 2022-06-22 04:38:20        1  9.454659015   7.7392354          0.015625
#> 42: 2022-06-22 04:38:20        1 -0.005661792   4.7853198          0.015625
#> 43: 2022-06-22 04:38:20        1  5.250921880   1.2898823          0.015625
#> 44: 2022-06-22 04:38:20        1  9.693034072   1.4949914          0.015625
#> 45: 2022-06-22 04:38:20        1  0.263302326   3.4843524          0.015625
#> 46: 2022-06-22 04:38:20        1  4.951186388   7.6108056          0.015625
#> 47: 2022-06-22 04:38:20        1 -0.327742309  12.5365361          0.015625
#> 48: 2022-06-22 04:38:20        1 -4.557984405  12.6999457          0.015625
#> 49: 2022-06-22 04:38:20        1 -2.550136985   3.1178163          0.015625
#> 50: 2022-06-22 04:38:20        1  8.195594240   7.6387939          0.015625
#> 51: 2022-06-22 04:38:20        1 -0.593326736   8.7728727          0.015625
#> 52: 2022-06-22 04:38:20        1  8.744864245   8.9918699          0.015625
#> 53: 2022-06-22 04:38:20        1  3.409419699   5.2043814          0.015625
#> 54: 2022-06-22 04:38:20        1 -0.620447110   4.1124026          0.015625
#> 55: 2022-06-22 04:38:20        1  7.302593341  10.4972537          0.015625
#> 56: 2022-06-22 04:38:20        1 -4.897814381   7.9292124          0.015625
#> 57: 2022-06-22 04:38:20        1 -4.592846270   7.1594045          0.015625
#> 58: 2022-06-22 04:38:20        1  6.639909773   6.9434900          0.015625
#> 59: 2022-06-22 04:38:20        1  8.320644936  14.1330393          0.015625
#> 60: 2022-06-22 04:38:20        1  3.350881954  12.0021037          0.015625
#> 61: 2022-06-22 04:38:20        1  8.285125049   3.3539615          0.015625
#> 62: 2022-06-22 04:38:20        1 -4.933106161  14.5655707          0.015625
#> 63: 2022-06-22 04:38:20        1  4.037225961   6.5262265          0.015625
#> 64: 2022-06-22 04:38:20        1  5.175082281   7.9842968          0.015625
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity