Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 3.651051 1.519879 0.015625 <list[3]> 2.465227

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 3.651051 1.519879 0.015625 <list[3]> 2.465227

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  0.43926415  3.9276620 0.015625     0       6          1  20.592617
#>  2: -3.20341827  3.8244784 0.015625     0       6          1  58.015870
#>  3:  3.75014717 11.5797411 0.015625     0       6          1 125.684314
#>  4:  8.11722923  1.0168338 0.015625     0       6          1  42.428315
#>  5:  8.52492920  7.5345254 0.015625     0       6          1 169.611674
#>  6: -2.25034548 13.9103583 0.015625     0       6          1  21.380537
#>  7:  5.16610664  9.0259960 0.015625     0       6          1 122.941745
#>  8: -3.22272152 10.0154609 0.015625     0       6          1   2.482948
#>  9:  9.36301451  8.7630294 0.015625     0       6          1 224.494980
#> 10: -2.13181442  4.6645869 0.015625     0       6          1  28.588975
#> 11: -4.37712758  7.7075938 0.015625     0       6          1  41.038810
#> 12:  4.71932781  2.5022457 0.015625     0       6          1  21.145473
#> 13:  2.27639314  7.5629879 0.015625     0       6          1  29.040354
#> 14:  3.13518887 12.0437108 0.015625     0       6          1 115.558754
#> 15:  2.23685201  4.6868543 0.015625     0       6          1   8.447905
#> 16:  3.55921183 14.2127543 0.015625     0       6          1 183.146921
#> 17: -3.57552443  4.9218157 0.015625     0       6          1  52.580474
#> 18:  0.87609738 12.5127853 0.015625     0       6          1  78.297002
#> 19: -4.53928766  9.9193898 0.015625     0       6          1  23.859169
#> 20: -2.58256872  0.3014105 0.015625     0       6          1 102.138729
#> 21: -2.42946166  5.4135639 0.015625     0       6          1  24.210273
#> 22:  8.72905280  1.0071465 0.015625     0       6          1  45.624496
#> 23:  3.65105138  1.5198785 0.015625     0       6          1   2.465227
#> 24: -4.05063668  4.2673872 0.015625     0       6          1  79.509406
#> 25: -2.13580343  7.4392462 0.015625     0       6          1   9.269860
#> 26:  1.02842115 13.6027012 0.015625     0       6          1  99.724538
#> 27: -4.05465513  3.8796900 0.015625     0       6          1  86.557901
#> 28:  4.71353084  3.9070519 0.015625     0       6          1  32.343691
#> 29:  8.94158270 14.0708491 0.015625     0       6          1 395.263190
#> 30: -2.20123008  2.5448187 0.015625     0       6          1  54.856580
#> 31:  8.80746761  9.0112013 0.015625     0       6          1 216.605975
#> 32:  2.61057712 10.2883717 0.015625     0       6          1  69.513812
#> 33: -2.46117969  0.1778663 0.015625     0       6          1 101.051107
#> 34:  4.15238066  1.2041804 0.015625     0       6          1   6.544874
#> 35:  2.40913039  4.9570484 0.015625     0       6          1   9.687502
#> 36: -2.88820076  6.0904412 0.015625     0       6          1  23.388345
#> 37:  3.96800769  3.1069180 0.015625     0       6          1  12.126745
#> 38:  7.52448621  3.4743590 0.015625     0       6          1  72.537443
#> 39: -3.57698670  7.2207680 0.015625     0       6          1  24.967654
#> 40:  6.20986662 11.4524182 0.015625     0       6          1 219.800176
#> 41:  8.08270622  8.8583355 0.015625     0       6          1 195.887209
#> 42:  3.31945159 10.0789500 0.015625     0       6          1  81.967996
#> 43:  5.19089229  5.4258558 0.015625     0       6          1  61.466199
#> 44: -2.32687327  6.7150385 0.015625     0       6          1  13.364828
#> 45:  8.28130998 11.1718829 0.015625     0       6          1 269.867376
#> 46:  5.55446386 14.0550883 0.015625     0       6          1 271.460136
#> 47:  9.05719787  4.1397104 0.015625     0       6          1 101.688849
#> 48: -3.46335374 14.1419123 0.015625     0       6          1   6.002829
#> 49:  9.18394923 14.3424388 0.015625     0       6          1 415.440556
#> 50:  7.41934000  7.4538305 0.015625     0       6          1 147.897990
#> 51:  2.56608613  8.4358824 0.015625     0       6          1  41.855085
#> 52:  7.02248709  1.1704164 0.015625     0       6          1  40.431551
#> 53:  6.15980877  9.9419921 0.015625     0       6          1 177.760138
#> 54: -2.28313111 10.1014212 0.015625     0       6          1   3.818548
#> 55:  8.36346461  8.4513885 0.015625     0       6          1 190.591860
#> 56:  1.70836398 10.8773775 0.015625     0       6          1  65.032084
#> 57:  1.40530394  1.9895077 0.015625     0       6          1  14.947602
#> 58:  0.70689615  9.5534848 0.015625     0       6          1  39.046521
#> 59:  0.04380747  3.4624735 0.015625     0       6          1  25.683252
#> 60: -0.35665510  4.1155184 0.015625     0       6          1  25.029919
#> 61: -0.97696770  4.5036767 0.015625     0       6          1  24.862616
#> 62:  1.26800831  3.6482904 0.015625     0       6          1  13.009908
#> 63: -0.89817974  1.0065768 0.015625     0       6          1  57.555609
#> 64:  8.35163746 12.9433487 0.015625     0       6          1 332.693073
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-02-22 04:29:09        1  0.43926415   3.9276620          0.015625
#>  2: 2022-02-22 04:29:09        1 -3.20341827   3.8244784          0.015625
#>  3: 2022-02-22 04:29:09        1  3.75014717  11.5797411          0.015625
#>  4: 2022-02-22 04:29:09        1  8.11722923   1.0168338          0.015625
#>  5: 2022-02-22 04:29:09        1  8.52492920   7.5345254          0.015625
#>  6: 2022-02-22 04:29:09        1 -2.25034548  13.9103583          0.015625
#>  7: 2022-02-22 04:29:09        1  5.16610664   9.0259960          0.015625
#>  8: 2022-02-22 04:29:09        1 -3.22272152  10.0154609          0.015625
#>  9: 2022-02-22 04:29:09        1  9.36301451   8.7630294          0.015625
#> 10: 2022-02-22 04:29:09        1 -2.13181442   4.6645869          0.015625
#> 11: 2022-02-22 04:29:09        1 -4.37712758   7.7075938          0.015625
#> 12: 2022-02-22 04:29:09        1  4.71932781   2.5022457          0.015625
#> 13: 2022-02-22 04:29:09        1  2.27639314   7.5629879          0.015625
#> 14: 2022-02-22 04:29:09        1  3.13518887  12.0437108          0.015625
#> 15: 2022-02-22 04:29:09        1  2.23685201   4.6868543          0.015625
#> 16: 2022-02-22 04:29:09        1  3.55921183  14.2127543          0.015625
#> 17: 2022-02-22 04:29:09        1 -3.57552443   4.9218157          0.015625
#> 18: 2022-02-22 04:29:09        1  0.87609738  12.5127853          0.015625
#> 19: 2022-02-22 04:29:09        1 -4.53928766   9.9193898          0.015625
#> 20: 2022-02-22 04:29:09        1 -2.58256872   0.3014105          0.015625
#> 21: 2022-02-22 04:29:09        1 -2.42946166   5.4135639          0.015625
#> 22: 2022-02-22 04:29:09        1  8.72905280   1.0071465          0.015625
#> 23: 2022-02-22 04:29:09        1  3.65105138   1.5198785          0.015625
#> 24: 2022-02-22 04:29:09        1 -4.05063668   4.2673872          0.015625
#> 25: 2022-02-22 04:29:09        1 -2.13580343   7.4392462          0.015625
#> 26: 2022-02-22 04:29:09        1  1.02842115  13.6027012          0.015625
#> 27: 2022-02-22 04:29:09        1 -4.05465513   3.8796900          0.015625
#> 28: 2022-02-22 04:29:09        1  4.71353084   3.9070519          0.015625
#> 29: 2022-02-22 04:29:09        1  8.94158270  14.0708491          0.015625
#> 30: 2022-02-22 04:29:09        1 -2.20123008   2.5448187          0.015625
#> 31: 2022-02-22 04:29:09        1  8.80746761   9.0112013          0.015625
#> 32: 2022-02-22 04:29:09        1  2.61057712  10.2883717          0.015625
#> 33: 2022-02-22 04:29:09        1 -2.46117969   0.1778663          0.015625
#> 34: 2022-02-22 04:29:09        1  4.15238066   1.2041804          0.015625
#> 35: 2022-02-22 04:29:09        1  2.40913039   4.9570484          0.015625
#> 36: 2022-02-22 04:29:09        1 -2.88820076   6.0904412          0.015625
#> 37: 2022-02-22 04:29:09        1  3.96800769   3.1069180          0.015625
#> 38: 2022-02-22 04:29:09        1  7.52448621   3.4743590          0.015625
#> 39: 2022-02-22 04:29:09        1 -3.57698670   7.2207680          0.015625
#> 40: 2022-02-22 04:29:09        1  6.20986662  11.4524182          0.015625
#> 41: 2022-02-22 04:29:09        1  8.08270622   8.8583355          0.015625
#> 42: 2022-02-22 04:29:09        1  3.31945159  10.0789500          0.015625
#> 43: 2022-02-22 04:29:09        1  5.19089229   5.4258558          0.015625
#> 44: 2022-02-22 04:29:09        1 -2.32687327   6.7150385          0.015625
#> 45: 2022-02-22 04:29:09        1  8.28130998  11.1718829          0.015625
#> 46: 2022-02-22 04:29:09        1  5.55446386  14.0550883          0.015625
#> 47: 2022-02-22 04:29:09        1  9.05719787   4.1397104          0.015625
#> 48: 2022-02-22 04:29:09        1 -3.46335374  14.1419123          0.015625
#> 49: 2022-02-22 04:29:09        1  9.18394923  14.3424388          0.015625
#> 50: 2022-02-22 04:29:09        1  7.41934000   7.4538305          0.015625
#> 51: 2022-02-22 04:29:09        1  2.56608613   8.4358824          0.015625
#> 52: 2022-02-22 04:29:09        1  7.02248709   1.1704164          0.015625
#> 53: 2022-02-22 04:29:09        1  6.15980877   9.9419921          0.015625
#> 54: 2022-02-22 04:29:09        1 -2.28313111  10.1014212          0.015625
#> 55: 2022-02-22 04:29:09        1  8.36346461   8.4513885          0.015625
#> 56: 2022-02-22 04:29:09        1  1.70836398  10.8773775          0.015625
#> 57: 2022-02-22 04:29:09        1  1.40530394   1.9895077          0.015625
#> 58: 2022-02-22 04:29:09        1  0.70689615   9.5534848          0.015625
#> 59: 2022-02-22 04:29:09        1  0.04380747   3.4624735          0.015625
#> 60: 2022-02-22 04:29:09        1 -0.35665510   4.1155184          0.015625
#> 61: 2022-02-22 04:29:09        1 -0.97696770   4.5036767          0.015625
#> 62: 2022-02-22 04:29:09        1  1.26800831   3.6482904          0.015625
#> 63: 2022-02-22 04:29:09        1 -0.89817974   1.0065768          0.015625
#> 64: 2022-02-22 04:29:09        1  8.35163746  12.9433487          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity