Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.788444 10.32356 0.015625 <list[3]> 1.115379

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.788444 10.32356 0.015625 <list[3]> 1.115379

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1: -4.92489170  4.3174059 0.015625     0       6          1 117.427362
#>  2:  5.08865134 12.2173558 0.015625     0       6          1 196.319247
#>  3:  1.37956244 10.9101800 0.015625     0       6          1  61.489576
#>  4:  9.29331630  8.1927538 0.015625     0       6          1 205.773419
#>  5:  4.85160222 10.1386441 0.015625     0       6          1 135.353712
#>  6: -4.43234681  4.3858870 0.015625     0       6          1  93.324500
#>  7: -3.14882084  7.2993930 0.015625     0       6          1  16.534185
#>  8:  0.66231911  5.9091869 0.015625     0       6          1  18.474065
#>  9:  2.18444542 11.6447915 0.015625     0       6          1  85.016281
#> 10:  3.49926697  7.4136965 0.015625     0       6          1  44.650881
#> 11:  6.37805403  9.9030883 0.015625     0       6          1 183.483462
#> 12: -4.62286733  7.0828934 0.015625     0       6          1  57.190290
#> 13:  3.63921879  7.5742014 0.015625     0       6          1  49.989952
#> 14:  7.05524706  9.7686470 0.015625     0       6          1 198.238206
#> 15:  0.10485988 14.6513150 0.015625     0       6          1  97.304142
#> 16:  4.06950368 11.3970034 0.015625     0       6          1 133.398113
#> 17: -2.39148294 11.4263013 0.015625     0       6          1   5.060933
#> 18:  1.20613219 12.6440780 0.015625     0       6          1  85.997264
#> 19:  2.33691424  6.5271848 0.015625     0       6          1  19.977296
#> 20:  4.35173303 10.9585508 0.015625     0       6          1 134.352950
#> 21: -0.03289886 14.3901149 0.015625     0       6          1  89.114515
#> 22:  4.87038055  5.9973802 0.015625     0       6          1  60.784043
#> 23:  9.07539216  1.2310600 0.015625     0       6          1  51.994568
#> 24:  3.27957385  8.9331340 0.015625     0       6          1  61.673384
#> 25:  1.50197278  5.2588814 0.015625     0       6          1  13.156674
#> 26:  2.27089074  3.9600693 0.015625     0       6          1   5.817786
#> 27:  6.03714888 14.5893362 0.015625     0       6          1 310.940829
#> 28: -3.68062680 10.1925149 0.015625     0       6          1   6.093789
#> 29:  0.05120686  7.5334296 0.015625     0       6          1  22.197258
#> 30:  8.45741499  4.8354737 0.015625     0       6          1 106.492387
#> 31: -3.07594999  4.4193484 0.015625     0       6          1  46.212074
#> 32: -2.33905409  5.1460003 0.015625     0       6          1  25.842127
#> 33:  4.95277722  8.1478266 0.015625     0       6          1  98.333809
#> 34: -2.51355904  3.1290463 0.015625     0       6          1  52.153758
#> 35:  3.97067247 14.6258275 0.015625     0       6          1 212.621749
#> 36: -0.61986128  4.0310612 0.015625     0       6          1  26.620551
#> 37: -0.85212974  7.2849724 0.015625     0       6          1  16.330608
#> 38:  9.11519243  7.5866022 0.015625     0       6          1 184.164661
#> 39:  6.63217839  8.8731497 0.015625     0       6          1 164.856863
#> 40: -2.14131633  5.9702121 0.015625     0       6          1  17.621862
#> 41:  4.49199949  9.1827956 0.015625     0       6          1 102.216445
#> 42: -2.78844417 10.3235634 0.015625     0       6          1   1.115379
#> 43:  3.92871278  9.4477199 0.015625     0       6          1  88.338975
#> 44:  8.51374313 11.0416062 0.015625     0       6          1 271.861984
#> 45:  7.49791806  2.9397881 0.015625     0       6          1  64.391572
#> 46:  2.92765218  2.4751797 0.015625     0       6          1   1.375692
#> 47:  8.61427483 10.9418921 0.015625     0       6          1 271.370629
#> 48:  1.77321520 13.7602676 0.015625     0       6          1 118.020543
#> 49:  4.30192236 13.9681418 0.015625     0       6          1 209.112449
#> 50:  2.41309728 11.5920799 0.015625     0       6          1  88.464283
#> 51:  6.12439166  4.2621659 0.015625     0       6          1  66.488425
#> 52:  6.41824759  5.5732228 0.015625     0       6          1  92.132193
#> 53:  2.82625543 10.7297112 0.015625     0       6          1  81.551948
#> 54:  3.99407874  4.8485455 0.015625     0       6          1  25.910273
#> 55:  7.02159883  4.9173488 0.015625     0       6          1  90.660137
#> 56:  2.39951326 10.1711553 0.015625     0       6          1  63.966657
#> 57:  0.62716343  8.3266539 0.015625     0       6          1  28.748912
#> 58: -4.96518813  0.5230792 0.015625     0       6          1 212.264093
#> 59:  7.39706235 12.9181174 0.015625     0       6          1 303.526682
#> 60:  5.41833831  6.7254669 0.015625     0       6          1  87.570050
#> 61:  5.94857121 14.5753030 0.015625     0       6          1 306.533095
#> 62:  1.95530869  8.2353872 0.015625     0       6          1  33.749035
#> 63: -0.85437198  6.1370340 0.015625     0       6          1  17.856119
#> 64: -0.62476767  0.4835673 0.015625     0       6          1  60.334961
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-04-11 04:33:41        1 -4.92489170   4.3174059          0.015625
#>  2: 2022-04-11 04:33:41        1  5.08865134  12.2173558          0.015625
#>  3: 2022-04-11 04:33:41        1  1.37956244  10.9101800          0.015625
#>  4: 2022-04-11 04:33:41        1  9.29331630   8.1927538          0.015625
#>  5: 2022-04-11 04:33:41        1  4.85160222  10.1386441          0.015625
#>  6: 2022-04-11 04:33:41        1 -4.43234681   4.3858870          0.015625
#>  7: 2022-04-11 04:33:41        1 -3.14882084   7.2993930          0.015625
#>  8: 2022-04-11 04:33:41        1  0.66231911   5.9091869          0.015625
#>  9: 2022-04-11 04:33:41        1  2.18444542  11.6447915          0.015625
#> 10: 2022-04-11 04:33:41        1  3.49926697   7.4136965          0.015625
#> 11: 2022-04-11 04:33:41        1  6.37805403   9.9030883          0.015625
#> 12: 2022-04-11 04:33:41        1 -4.62286733   7.0828934          0.015625
#> 13: 2022-04-11 04:33:41        1  3.63921879   7.5742014          0.015625
#> 14: 2022-04-11 04:33:41        1  7.05524706   9.7686470          0.015625
#> 15: 2022-04-11 04:33:41        1  0.10485988  14.6513150          0.015625
#> 16: 2022-04-11 04:33:41        1  4.06950368  11.3970034          0.015625
#> 17: 2022-04-11 04:33:41        1 -2.39148294  11.4263013          0.015625
#> 18: 2022-04-11 04:33:41        1  1.20613219  12.6440780          0.015625
#> 19: 2022-04-11 04:33:41        1  2.33691424   6.5271848          0.015625
#> 20: 2022-04-11 04:33:41        1  4.35173303  10.9585508          0.015625
#> 21: 2022-04-11 04:33:41        1 -0.03289886  14.3901149          0.015625
#> 22: 2022-04-11 04:33:41        1  4.87038055   5.9973802          0.015625
#> 23: 2022-04-11 04:33:41        1  9.07539216   1.2310600          0.015625
#> 24: 2022-04-11 04:33:41        1  3.27957385   8.9331340          0.015625
#> 25: 2022-04-11 04:33:41        1  1.50197278   5.2588814          0.015625
#> 26: 2022-04-11 04:33:41        1  2.27089074   3.9600693          0.015625
#> 27: 2022-04-11 04:33:41        1  6.03714888  14.5893362          0.015625
#> 28: 2022-04-11 04:33:41        1 -3.68062680  10.1925149          0.015625
#> 29: 2022-04-11 04:33:41        1  0.05120686   7.5334296          0.015625
#> 30: 2022-04-11 04:33:41        1  8.45741499   4.8354737          0.015625
#> 31: 2022-04-11 04:33:41        1 -3.07594999   4.4193484          0.015625
#> 32: 2022-04-11 04:33:41        1 -2.33905409   5.1460003          0.015625
#> 33: 2022-04-11 04:33:41        1  4.95277722   8.1478266          0.015625
#> 34: 2022-04-11 04:33:41        1 -2.51355904   3.1290463          0.015625
#> 35: 2022-04-11 04:33:41        1  3.97067247  14.6258275          0.015625
#> 36: 2022-04-11 04:33:41        1 -0.61986128   4.0310612          0.015625
#> 37: 2022-04-11 04:33:41        1 -0.85212974   7.2849724          0.015625
#> 38: 2022-04-11 04:33:41        1  9.11519243   7.5866022          0.015625
#> 39: 2022-04-11 04:33:41        1  6.63217839   8.8731497          0.015625
#> 40: 2022-04-11 04:33:41        1 -2.14131633   5.9702121          0.015625
#> 41: 2022-04-11 04:33:41        1  4.49199949   9.1827956          0.015625
#> 42: 2022-04-11 04:33:41        1 -2.78844417  10.3235634          0.015625
#> 43: 2022-04-11 04:33:41        1  3.92871278   9.4477199          0.015625
#> 44: 2022-04-11 04:33:41        1  8.51374313  11.0416062          0.015625
#> 45: 2022-04-11 04:33:41        1  7.49791806   2.9397881          0.015625
#> 46: 2022-04-11 04:33:41        1  2.92765218   2.4751797          0.015625
#> 47: 2022-04-11 04:33:41        1  8.61427483  10.9418921          0.015625
#> 48: 2022-04-11 04:33:41        1  1.77321520  13.7602676          0.015625
#> 49: 2022-04-11 04:33:41        1  4.30192236  13.9681418          0.015625
#> 50: 2022-04-11 04:33:41        1  2.41309728  11.5920799          0.015625
#> 51: 2022-04-11 04:33:41        1  6.12439166   4.2621659          0.015625
#> 52: 2022-04-11 04:33:41        1  6.41824759   5.5732228          0.015625
#> 53: 2022-04-11 04:33:41        1  2.82625543  10.7297112          0.015625
#> 54: 2022-04-11 04:33:41        1  3.99407874   4.8485455          0.015625
#> 55: 2022-04-11 04:33:41        1  7.02159883   4.9173488          0.015625
#> 56: 2022-04-11 04:33:41        1  2.39951326  10.1711553          0.015625
#> 57: 2022-04-11 04:33:41        1  0.62716343   8.3266539          0.015625
#> 58: 2022-04-11 04:33:41        1 -4.96518813   0.5230792          0.015625
#> 59: 2022-04-11 04:33:41        1  7.39706235  12.9181174          0.015625
#> 60: 2022-04-11 04:33:41        1  5.41833831   6.7254669          0.015625
#> 61: 2022-04-11 04:33:41        1  5.94857121  14.5753030          0.015625
#> 62: 2022-04-11 04:33:41        1  1.95530869   8.2353872          0.015625
#> 63: 2022-04-11 04:33:41        1 -0.85437198   6.1370340          0.015625
#> 64: 2022-04-11 04:33:41        1 -0.62476767   0.4835673          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity