Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.634314 9.777704 0.015625 <list[3]> 2.001856

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.634314 9.777704 0.015625 <list[3]> 2.001856

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1:  7.1060935  3.1303798 0.015625     0       6          1  63.967432
#>  2:  5.8101934  1.7697509 0.015625     0       6          1  34.380272
#>  3: -2.4456758  9.6794441 0.015625     0       6          1   2.788215
#>  4:  3.4417643  3.8373235 0.015625     0       6          1   9.534698
#>  5:  5.3014496  0.9563814 0.015625     0       6          1  21.734351
#>  6:  6.8575507 13.2125554 0.015625     0       6          1 296.309668
#>  7:  9.6323348  6.4775134 0.015625     0       6          1 168.438208
#>  8:  8.2430832  6.4018042 0.015625     0       6          1 137.044972
#>  9: -4.1182497  2.7686297 0.015625     0       6          1 110.863912
#> 10: -4.8940135  5.0121087 0.015625     0       6          1 102.238972
#> 11: -2.8541774  4.3648221 0.015625     0       6          1  42.113750
#> 12: -2.7537164  8.0306713 0.015625     0       6          1   7.794226
#> 13: -3.0645512  9.5174945 0.015625     0       6          1   3.144416
#> 14:  9.2843434  0.5455114 0.015625     0       6          1  45.003204
#> 15: -2.5655115 13.9445531 0.015625     0       6          1  15.336224
#> 16: -2.8000874 11.9834472 0.015625     0       6          1   2.605928
#> 17:  8.9584237  7.3668991 0.015625     0       6          1 174.533139
#> 18: -3.6444217  8.7997312 0.015625     0       6          1  13.207161
#> 19:  3.1542447  5.6122829 0.015625     0       6          1  19.117339
#> 20:  0.5271518  3.8754646 0.015625     0       6          1  19.973236
#> 21: -4.8265482 12.1708723 0.015625     0       6          1  16.053705
#> 22: -0.8753611  2.6534520 0.015625     0       6          1  38.841180
#> 23:  7.8137116 11.2198893 0.015625     0       6          1 259.349746
#> 24: -2.6343137  9.7777044 0.015625     0       6          1   2.001856
#> 25:  6.0697382 13.4079773 0.015625     0       6          1 273.324141
#> 26: -3.6255158 13.2383413 0.015625     0       6          1   2.632540
#> 27:  8.0580188  1.1703698 0.015625     0       6          1  44.038185
#> 28:  8.8746885  5.8349183 0.015625     0       6          1 134.934928
#> 29: -4.3110977  6.6307558 0.015625     0       6          1  52.516816
#> 30:  8.7291510  2.8622603 0.015625     0       6          1  73.395250
#> 31:  8.0780621 10.6302746 0.015625     0       6          1 247.512534
#> 32: -0.8707344  2.0083135 0.015625     0       6          1  45.355109
#> 33:  1.7009116 11.2674722 0.015625     0       6          1  70.936875
#> 34:  6.0752511 12.0789986 0.015625     0       6          1 232.941758
#> 35:  6.6823438  5.3934860 0.015625     0       6          1  93.770119
#> 36:  2.7807206  8.4789527 0.015625     0       6          1  45.463370
#> 37: -4.5054287  3.8187362 0.015625     0       6          1 107.547606
#> 38: -0.5718587  9.1089361 0.015625     0       6          1  22.864964
#> 39:  3.7988385  6.0440575 0.015625     0       6          1  34.279843
#> 40: -3.8969932 11.5462997 0.015625     0       6          1   4.270593
#> 41:  5.1195725  6.5894783 0.015625     0       6          1  76.713765
#> 42: -2.5630762  2.9560447 0.015625     0       6          1  55.619035
#> 43:  5.6654515 13.8755364 0.015625     0       6          1 270.812038
#> 44: -3.9788502 10.2420739 0.015625     0       6          1  10.213458
#> 45:  9.5860611 12.2943471 0.015625     0       6          1 351.171180
#> 46:  4.2343039 10.2872947 0.015625     0       6          1 115.310515
#> 47: -3.7619328 10.4562522 0.015625     0       6          1   6.052831
#> 48:  6.5235453  9.1623900 0.015625     0       6          1 169.056669
#> 49:  4.5864323 14.7835048 0.015625     0       6          1 247.071995
#> 50:  2.9144947  3.2888585 0.015625     0       6          1   3.420812
#> 51:  7.1032026  8.9400325 0.015625     0       6          1 177.682572
#> 52: -1.4506895 11.6773183 0.015625     0       6          1  22.065386
#> 53: -1.5194900  3.4389120 0.015625     0       6          1  35.999192
#> 54:  5.9398213  9.0351568 0.015625     0       6          1 149.090006
#> 55: -3.3017011  4.8101063 0.015625     0       6          1  46.487667
#> 56:  7.0349349  0.7680884 0.015625     0       6          1  36.753267
#> 57:  9.5011548  8.8881013 0.015625     0       6          1 232.502663
#> 58: -2.8272598 14.2576548 0.015625     0       6          1  13.203617
#> 59:  7.9918300  7.9851089 0.015625     0       6          1 171.006243
#> 60:  0.9738361  4.3285167 0.015625     0       6          1  15.420366
#> 61:  4.4096301  2.4875962 0.015625     0       6          1  15.592792
#> 62:  9.2520495 14.5719941 0.015625     0       6          1 427.588424
#> 63:  5.2782752 14.4295139 0.015625     0       6          1 270.302044
#> 64: -3.7327795  4.7597846 0.015625     0       6          1  59.932606
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-07-17 04:42:18        1   7.1060935   3.1303798          0.015625
#>  2: 2022-07-17 04:42:18        1   5.8101934   1.7697509          0.015625
#>  3: 2022-07-17 04:42:18        1  -2.4456758   9.6794441          0.015625
#>  4: 2022-07-17 04:42:18        1   3.4417643   3.8373235          0.015625
#>  5: 2022-07-17 04:42:18        1   5.3014496   0.9563814          0.015625
#>  6: 2022-07-17 04:42:18        1   6.8575507  13.2125554          0.015625
#>  7: 2022-07-17 04:42:18        1   9.6323348   6.4775134          0.015625
#>  8: 2022-07-17 04:42:18        1   8.2430832   6.4018042          0.015625
#>  9: 2022-07-17 04:42:18        1  -4.1182497   2.7686297          0.015625
#> 10: 2022-07-17 04:42:18        1  -4.8940135   5.0121087          0.015625
#> 11: 2022-07-17 04:42:18        1  -2.8541774   4.3648221          0.015625
#> 12: 2022-07-17 04:42:18        1  -2.7537164   8.0306713          0.015625
#> 13: 2022-07-17 04:42:18        1  -3.0645512   9.5174945          0.015625
#> 14: 2022-07-17 04:42:18        1   9.2843434   0.5455114          0.015625
#> 15: 2022-07-17 04:42:18        1  -2.5655115  13.9445531          0.015625
#> 16: 2022-07-17 04:42:18        1  -2.8000874  11.9834472          0.015625
#> 17: 2022-07-17 04:42:18        1   8.9584237   7.3668991          0.015625
#> 18: 2022-07-17 04:42:18        1  -3.6444217   8.7997312          0.015625
#> 19: 2022-07-17 04:42:18        1   3.1542447   5.6122829          0.015625
#> 20: 2022-07-17 04:42:18        1   0.5271518   3.8754646          0.015625
#> 21: 2022-07-17 04:42:18        1  -4.8265482  12.1708723          0.015625
#> 22: 2022-07-17 04:42:18        1  -0.8753611   2.6534520          0.015625
#> 23: 2022-07-17 04:42:18        1   7.8137116  11.2198893          0.015625
#> 24: 2022-07-17 04:42:18        1  -2.6343137   9.7777044          0.015625
#> 25: 2022-07-17 04:42:18        1   6.0697382  13.4079773          0.015625
#> 26: 2022-07-17 04:42:18        1  -3.6255158  13.2383413          0.015625
#> 27: 2022-07-17 04:42:18        1   8.0580188   1.1703698          0.015625
#> 28: 2022-07-17 04:42:18        1   8.8746885   5.8349183          0.015625
#> 29: 2022-07-17 04:42:18        1  -4.3110977   6.6307558          0.015625
#> 30: 2022-07-17 04:42:18        1   8.7291510   2.8622603          0.015625
#> 31: 2022-07-17 04:42:18        1   8.0780621  10.6302746          0.015625
#> 32: 2022-07-17 04:42:18        1  -0.8707344   2.0083135          0.015625
#> 33: 2022-07-17 04:42:18        1   1.7009116  11.2674722          0.015625
#> 34: 2022-07-17 04:42:18        1   6.0752511  12.0789986          0.015625
#> 35: 2022-07-17 04:42:18        1   6.6823438   5.3934860          0.015625
#> 36: 2022-07-17 04:42:18        1   2.7807206   8.4789527          0.015625
#> 37: 2022-07-17 04:42:18        1  -4.5054287   3.8187362          0.015625
#> 38: 2022-07-17 04:42:18        1  -0.5718587   9.1089361          0.015625
#> 39: 2022-07-17 04:42:18        1   3.7988385   6.0440575          0.015625
#> 40: 2022-07-17 04:42:18        1  -3.8969932  11.5462997          0.015625
#> 41: 2022-07-17 04:42:18        1   5.1195725   6.5894783          0.015625
#> 42: 2022-07-17 04:42:18        1  -2.5630762   2.9560447          0.015625
#> 43: 2022-07-17 04:42:18        1   5.6654515  13.8755364          0.015625
#> 44: 2022-07-17 04:42:18        1  -3.9788502  10.2420739          0.015625
#> 45: 2022-07-17 04:42:18        1   9.5860611  12.2943471          0.015625
#> 46: 2022-07-17 04:42:18        1   4.2343039  10.2872947          0.015625
#> 47: 2022-07-17 04:42:18        1  -3.7619328  10.4562522          0.015625
#> 48: 2022-07-17 04:42:18        1   6.5235453   9.1623900          0.015625
#> 49: 2022-07-17 04:42:18        1   4.5864323  14.7835048          0.015625
#> 50: 2022-07-17 04:42:18        1   2.9144947   3.2888585          0.015625
#> 51: 2022-07-17 04:42:18        1   7.1032026   8.9400325          0.015625
#> 52: 2022-07-17 04:42:18        1  -1.4506895  11.6773183          0.015625
#> 53: 2022-07-17 04:42:18        1  -1.5194900   3.4389120          0.015625
#> 54: 2022-07-17 04:42:18        1   5.9398213   9.0351568          0.015625
#> 55: 2022-07-17 04:42:18        1  -3.3017011   4.8101063          0.015625
#> 56: 2022-07-17 04:42:18        1   7.0349349   0.7680884          0.015625
#> 57: 2022-07-17 04:42:18        1   9.5011548   8.8881013          0.015625
#> 58: 2022-07-17 04:42:18        1  -2.8272598  14.2576548          0.015625
#> 59: 2022-07-17 04:42:18        1   7.9918300   7.9851089          0.015625
#> 60: 2022-07-17 04:42:18        1   0.9738361   4.3285167          0.015625
#> 61: 2022-07-17 04:42:18        1   4.4096301   2.4875962          0.015625
#> 62: 2022-07-17 04:42:18        1   9.2520495  14.5719941          0.015625
#> 63: 2022-07-17 04:42:18        1   5.2782752  14.4295139          0.015625
#> 64: 2022-07-17 04:42:18        1  -3.7327795   4.7597846          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity