Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 3.404457 1.709332 0.015625 <list[3]> 1.322667

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 3.404457 1.709332 0.015625 <list[3]> 1.322667

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  0.15026690  1.7266269 0.015625     0       6          1  35.774408
#>  2:  7.33773172 11.6613571 0.015625     0       6          1 260.734884
#>  3:  4.90735884  6.0638364 0.015625     0       6          1  62.749758
#>  4:  0.23888362  0.6953759 0.015625     0       6          1  43.596724
#>  5:  9.19978290  0.1684278 0.015625     0       6          1  39.179580
#>  6:  2.10304891 11.3442271 0.015625     0       6          1  78.321091
#>  7:  8.99154168  2.8076560 0.015625     0       6          1  75.801790
#>  8:  0.47476869 11.4502696 0.015625     0       6          1  56.967171
#>  9:  3.91455564 13.8466724 0.015625     0       6          1 188.242045
#> 10:  3.18957900  5.7147940 0.015625     0       6          1  20.464690
#> 11:  1.03247631 13.0957605 0.015625     0       6          1  90.721233
#> 12:  1.72264165  3.1410496 0.015625     0       6          1   8.591042
#> 13:  3.77187307  2.2355994 0.015625     0       6          1   5.487437
#> 14:  3.59579425  4.8317328 0.015625     0       6          1  18.652631
#> 15:  6.29251656 13.1669061 0.015625     0       6          1 274.460758
#> 16: -0.38055155  6.8475700 0.015625     0       6          1  18.971559
#> 17:  8.70955580 14.0061907 0.015625     0       6          1 384.387569
#> 18:  7.67550795  2.1834829 0.015625     0       6          1  55.106671
#> 19: -2.66002151 12.6137458 0.015625     0       6          1   6.166934
#> 20: -4.92036400 14.4283193 0.015625     0       6          1  12.004263
#> 21: -4.09382861  4.5690481 0.015625     0       6          1  76.033895
#> 22:  1.16564487  0.9553612 0.015625     0       6          1  24.225623
#> 23: -1.27628985  3.4235431 0.015625     0       6          1  34.482465
#> 24:  8.78403843  1.0904020 0.015625     0       6          1  47.168509
#> 25:  3.40445716  1.7093319 0.015625     0       6          1   1.322667
#> 26:  5.03564036 13.4689527 0.015625     0       6          1 229.250405
#> 27: -1.50498463  6.3761251 0.015625     0       6          1  14.994451
#> 28: -4.23710941  2.6070617 0.015625     0       6          1 119.850473
#> 29:  8.04105107 14.4706747 0.015625     0       6          1 379.946068
#> 30: -4.79295179  8.7391911 0.015625     0       6          1  42.080799
#> 31: -3.49094237  9.8355876 0.015625     0       6          1   5.367441
#> 32:  8.36857222  4.6173095 0.015625     0       6          1 100.981146
#> 33:  5.26608905  2.7602362 0.015625     0       6          1  33.443053
#> 34: -4.24680749  0.1299809 0.015625     0       6          1 179.495378
#> 35: -1.86809045 10.2205265 0.015625     0       6          1   8.486970
#> 36: -2.39913965 10.2971585 0.015625     0       6          1   3.016149
#> 37: -1.50763919  6.9232404 0.015625     0       6          1  12.996554
#> 38:  3.66333095  1.1726687 0.015625     0       6          1   2.024000
#> 39: -4.47033015  8.7010759 0.015625     0       6          1  32.980338
#> 40:  4.80570563 11.1029605 0.015625     0       6          1 155.889627
#> 41:  3.87415498 11.5056577 0.015625     0       6          1 128.527136
#> 42:  2.91413799  3.2476705 0.015625     0       6          1   3.284384
#> 43:  6.92647097 12.0394291 0.015625     0       6          1 260.672775
#> 44: -3.83455362  1.2498743 0.015625     0       6          1 130.417945
#> 45:  3.47141986  7.5940313 0.015625     0       6          1  46.456939
#> 46: -2.35705076  5.2453740 0.015625     0       6          1  25.077017
#> 47:  0.65852367  7.1586611 0.015625     0       6          1  22.405284
#> 48:  6.93257361  3.4662722 0.015625     0       6          1  66.957342
#> 49: -4.78277604  8.0389701 0.015625     0       6          1  50.068385
#> 50: -0.08128821  1.6935856 0.015625     0       6          1  39.248429
#> 51: -2.12180274  9.5227027 0.015625     0       6          1   4.972916
#> 52:  1.58356187 14.7051072 0.015625     0       6          1 134.162486
#> 53:  7.02860530  6.9415540 0.015625     0       6          1 129.606220
#> 54: -1.81389065 10.0351844 0.015625     0       6          1   8.785172
#> 55: -4.59773520  8.0180526 0.015625     0       6          1  44.297286
#> 56:  1.94192155  3.0896937 0.015625     0       6          1   6.521805
#> 57: -3.95979965 10.1963081 0.015625     0       6          1  10.134517
#> 58:  9.13803820 10.3228119 0.015625     0       6          1 266.446065
#> 59:  7.27496746  5.6847038 0.015625     0       6          1 108.104527
#> 60:  8.68259124  4.9734903 0.015625     0       6          1 112.634836
#> 61: -0.74094859 14.6169556 0.015625     0       6          1  72.153268
#> 62: -2.33728809 10.4477699 0.015625     0       6          1   3.653335
#> 63:  3.58578292  5.8413996 0.015625     0       6          1  27.883201
#> 64:  9.39446766 10.1187762 0.015625     0       6          1 267.951362
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-02-12 04:27:08        1  0.15026690   1.7266269          0.015625
#>  2: 2022-02-12 04:27:08        1  7.33773172  11.6613571          0.015625
#>  3: 2022-02-12 04:27:08        1  4.90735884   6.0638364          0.015625
#>  4: 2022-02-12 04:27:08        1  0.23888362   0.6953759          0.015625
#>  5: 2022-02-12 04:27:08        1  9.19978290   0.1684278          0.015625
#>  6: 2022-02-12 04:27:08        1  2.10304891  11.3442271          0.015625
#>  7: 2022-02-12 04:27:08        1  8.99154168   2.8076560          0.015625
#>  8: 2022-02-12 04:27:08        1  0.47476869  11.4502696          0.015625
#>  9: 2022-02-12 04:27:08        1  3.91455564  13.8466724          0.015625
#> 10: 2022-02-12 04:27:08        1  3.18957900   5.7147940          0.015625
#> 11: 2022-02-12 04:27:08        1  1.03247631  13.0957605          0.015625
#> 12: 2022-02-12 04:27:08        1  1.72264165   3.1410496          0.015625
#> 13: 2022-02-12 04:27:08        1  3.77187307   2.2355994          0.015625
#> 14: 2022-02-12 04:27:08        1  3.59579425   4.8317328          0.015625
#> 15: 2022-02-12 04:27:08        1  6.29251656  13.1669061          0.015625
#> 16: 2022-02-12 04:27:08        1 -0.38055155   6.8475700          0.015625
#> 17: 2022-02-12 04:27:08        1  8.70955580  14.0061907          0.015625
#> 18: 2022-02-12 04:27:08        1  7.67550795   2.1834829          0.015625
#> 19: 2022-02-12 04:27:08        1 -2.66002151  12.6137458          0.015625
#> 20: 2022-02-12 04:27:08        1 -4.92036400  14.4283193          0.015625
#> 21: 2022-02-12 04:27:08        1 -4.09382861   4.5690481          0.015625
#> 22: 2022-02-12 04:27:08        1  1.16564487   0.9553612          0.015625
#> 23: 2022-02-12 04:27:08        1 -1.27628985   3.4235431          0.015625
#> 24: 2022-02-12 04:27:08        1  8.78403843   1.0904020          0.015625
#> 25: 2022-02-12 04:27:08        1  3.40445716   1.7093319          0.015625
#> 26: 2022-02-12 04:27:08        1  5.03564036  13.4689527          0.015625
#> 27: 2022-02-12 04:27:08        1 -1.50498463   6.3761251          0.015625
#> 28: 2022-02-12 04:27:08        1 -4.23710941   2.6070617          0.015625
#> 29: 2022-02-12 04:27:08        1  8.04105107  14.4706747          0.015625
#> 30: 2022-02-12 04:27:08        1 -4.79295179   8.7391911          0.015625
#> 31: 2022-02-12 04:27:08        1 -3.49094237   9.8355876          0.015625
#> 32: 2022-02-12 04:27:08        1  8.36857222   4.6173095          0.015625
#> 33: 2022-02-12 04:27:08        1  5.26608905   2.7602362          0.015625
#> 34: 2022-02-12 04:27:08        1 -4.24680749   0.1299809          0.015625
#> 35: 2022-02-12 04:27:08        1 -1.86809045  10.2205265          0.015625
#> 36: 2022-02-12 04:27:08        1 -2.39913965  10.2971585          0.015625
#> 37: 2022-02-12 04:27:08        1 -1.50763919   6.9232404          0.015625
#> 38: 2022-02-12 04:27:08        1  3.66333095   1.1726687          0.015625
#> 39: 2022-02-12 04:27:08        1 -4.47033015   8.7010759          0.015625
#> 40: 2022-02-12 04:27:08        1  4.80570563  11.1029605          0.015625
#> 41: 2022-02-12 04:27:08        1  3.87415498  11.5056577          0.015625
#> 42: 2022-02-12 04:27:08        1  2.91413799   3.2476705          0.015625
#> 43: 2022-02-12 04:27:08        1  6.92647097  12.0394291          0.015625
#> 44: 2022-02-12 04:27:08        1 -3.83455362   1.2498743          0.015625
#> 45: 2022-02-12 04:27:08        1  3.47141986   7.5940313          0.015625
#> 46: 2022-02-12 04:27:08        1 -2.35705076   5.2453740          0.015625
#> 47: 2022-02-12 04:27:08        1  0.65852367   7.1586611          0.015625
#> 48: 2022-02-12 04:27:08        1  6.93257361   3.4662722          0.015625
#> 49: 2022-02-12 04:27:08        1 -4.78277604   8.0389701          0.015625
#> 50: 2022-02-12 04:27:08        1 -0.08128821   1.6935856          0.015625
#> 51: 2022-02-12 04:27:08        1 -2.12180274   9.5227027          0.015625
#> 52: 2022-02-12 04:27:08        1  1.58356187  14.7051072          0.015625
#> 53: 2022-02-12 04:27:08        1  7.02860530   6.9415540          0.015625
#> 54: 2022-02-12 04:27:08        1 -1.81389065  10.0351844          0.015625
#> 55: 2022-02-12 04:27:08        1 -4.59773520   8.0180526          0.015625
#> 56: 2022-02-12 04:27:08        1  1.94192155   3.0896937          0.015625
#> 57: 2022-02-12 04:27:08        1 -3.95979965  10.1963081          0.015625
#> 58: 2022-02-12 04:27:08        1  9.13803820  10.3228119          0.015625
#> 59: 2022-02-12 04:27:08        1  7.27496746   5.6847038          0.015625
#> 60: 2022-02-12 04:27:08        1  8.68259124   4.9734903          0.015625
#> 61: 2022-02-12 04:27:08        1 -0.74094859  14.6169556          0.015625
#> 62: 2022-02-12 04:27:08        1 -2.33728809  10.4477699          0.015625
#> 63: 2022-02-12 04:27:08        1  3.58578292   5.8413996          0.015625
#> 64: 2022-02-12 04:27:08        1  9.39446766  10.1187762          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity