Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.641596 10.75582 0.015625 <list[3]> 3.669224

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.641596 10.75582 0.015625 <list[3]> 3.669224

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition          y
#>  1:  4.35699502 11.44710360 0.015625     0       6          1 145.846785
#>  2:  3.83257522 13.46732675 0.015625     0       6          1 174.614873
#>  3: -0.28775866 14.95511683 0.015625     0       6          1  91.365334
#>  4:  1.51249376 13.48252916 0.015625     0       6          1 106.980122
#>  5: -4.79681305 11.05186793 0.015625     0       6          1  21.633575
#>  6: -1.24262401  4.84085140 0.015625     0       6          1  23.234803
#>  7:  0.00816391 11.78395980 0.015625     0       6          1  53.206434
#>  8:  4.28934739  6.94840218 0.015625     0       6          1  58.033606
#>  9: -3.09214929  4.94064555 0.015625     0       6          1  39.780816
#> 10: -3.82768372  2.43966278 0.015625     0       6          1 104.636424
#> 11:  3.97628686  2.11422569 0.015625     0       6          1   7.381173
#> 12:  8.67516543 13.87449989 0.015625     0       6          1 378.071020
#> 13:  0.31309789  6.50603515 0.015625     0       6          1  20.137961
#> 14:  3.16522200  3.75421002 0.015625     0       6          1   6.569703
#> 15:  8.16410675  0.96227687 0.015625     0       6          1  41.956237
#> 16:  0.44611811 14.18883879 0.015625     0       6          1  97.743140
#> 17: -4.92949576 11.01343033 0.015625     0       6          1  24.879419
#> 18:  6.86837267  0.08818969 0.015625     0       6          1  30.742652
#> 19:  7.16700017  8.97658827 0.015625     0       6          1 180.028955
#> 20:  3.09847835 13.61489916 0.015625     0       6          1 150.496062
#> 21: -1.10389383  3.82263815 0.015625     0       6          1  30.096775
#> 22:  2.56024512 14.75051741 0.015625     0       6          1 161.333896
#> 23:  6.92728540  7.67281501 0.015625     0       6          1 143.621997
#> 24:  5.30274609 13.66192238 0.015625     0       6          1 247.509010
#> 25:  5.40009453  0.30664317 0.015625     0       6          1  20.113300
#> 26: -2.38841885  0.33311384 0.015625     0       6          1  95.993501
#> 27: -3.00556078  1.00744133 0.015625     0       6          1 101.565630
#> 28: -1.73212405  8.56768440 0.015625     0       6          1   8.536768
#> 29: -1.99401028  0.04108989 0.015625     0       6          1  91.706480
#> 30:  1.50384485  4.90658756 0.015625     0       6          1  12.156514
#> 31:  8.93892721 14.08911025 0.015625     0       6          1 395.890249
#> 32:  9.18418860  4.21853597 0.015625     0       6          1 105.575816
#> 33: -2.70571299  1.32163593 0.015625     0       6          1  86.114919
#> 34: -3.82594246 13.76318122 0.015625     0       6          1   4.058010
#> 35:  4.98457155  5.54372549 0.015625     0       6          1  57.645461
#> 36: -3.29474567 14.21606417 0.015625     0       6          1   7.472195
#> 37:  3.23227638 10.89319886 0.015625     0       6          1  94.843706
#> 38:  1.97049733  7.31436137 0.015625     0       6          1  25.021997
#> 39: -4.02102887  0.88056093 0.015625     0       6          1 148.267942
#> 40:  5.07692550  3.75477394 0.015625     0       6          1  38.849535
#> 41: -4.31109083 13.50712486 0.015625     0       6          1   6.254813
#> 42:  7.34560216 11.70471925 0.015625     0       6          1 262.313426
#> 43: -3.64159583 10.75581545 0.015625     0       6          1   3.669224
#> 44:  3.70762199 10.41677381 0.015625     0       6          1  99.805990
#> 45:  2.36568095  9.43712778 0.015625     0       6          1  52.569214
#> 46:  3.39903949 13.35588633 0.015625     0       6          1 154.732149
#> 47:  3.97399495  5.33769768 0.015625     0       6          1  30.337420
#> 48:  9.64540967  9.81417321 0.015625     0       6          1 266.477297
#> 49: -1.48682874  0.10114731 0.015625     0       6          1  80.247296
#> 50:  5.83546033 14.56946383 0.015625     0       6          1 301.227466
#> 51:  0.08964452  8.24042993 0.015625     0       6          1  25.241562
#> 52:  5.82124368 13.74477839 0.015625     0       6          1 273.561720
#> 53:  5.65621291  6.92904477 0.015625     0       6          1  97.833726
#> 54:  1.37461684 11.11977623 0.015625     0       6          1  64.426177
#> 55:  0.99519518 11.35751450 0.015625     0       6          1  62.988237
#> 56:  5.44222489 14.22025796 0.015625     0       6          1 271.480415
#> 57:  5.36528397 12.82192681 0.015625     0       6          1 225.386293
#> 58:  7.77141141  4.02005897 0.015625     0       6          1  83.581577
#> 59:  9.47774904  8.46929792 0.015625     0       6          1 219.204859
#> 60:  5.67300934 12.49262088 0.015625     0       6          1 229.047350
#> 61:  7.94213928  4.47571548 0.015625     0       6          1  93.364142
#> 62:  9.74441125  5.04320687 0.015625     0       6          1 136.196524
#> 63:  2.66187675  0.39187738 0.015625     0       6          1   4.008146
#> 64:  6.06830139  5.90898592 0.015625     0       6          1  90.526300
#>              x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-07-03 04:34:36        1  4.35699502 11.44710360          0.015625
#>  2: 2022-07-03 04:34:36        1  3.83257522 13.46732675          0.015625
#>  3: 2022-07-03 04:34:36        1 -0.28775866 14.95511683          0.015625
#>  4: 2022-07-03 04:34:36        1  1.51249376 13.48252916          0.015625
#>  5: 2022-07-03 04:34:36        1 -4.79681305 11.05186793          0.015625
#>  6: 2022-07-03 04:34:36        1 -1.24262401  4.84085140          0.015625
#>  7: 2022-07-03 04:34:36        1  0.00816391 11.78395980          0.015625
#>  8: 2022-07-03 04:34:36        1  4.28934739  6.94840218          0.015625
#>  9: 2022-07-03 04:34:36        1 -3.09214929  4.94064555          0.015625
#> 10: 2022-07-03 04:34:36        1 -3.82768372  2.43966278          0.015625
#> 11: 2022-07-03 04:34:36        1  3.97628686  2.11422569          0.015625
#> 12: 2022-07-03 04:34:36        1  8.67516543 13.87449989          0.015625
#> 13: 2022-07-03 04:34:36        1  0.31309789  6.50603515          0.015625
#> 14: 2022-07-03 04:34:36        1  3.16522200  3.75421002          0.015625
#> 15: 2022-07-03 04:34:36        1  8.16410675  0.96227687          0.015625
#> 16: 2022-07-03 04:34:36        1  0.44611811 14.18883879          0.015625
#> 17: 2022-07-03 04:34:36        1 -4.92949576 11.01343033          0.015625
#> 18: 2022-07-03 04:34:36        1  6.86837267  0.08818969          0.015625
#> 19: 2022-07-03 04:34:36        1  7.16700017  8.97658827          0.015625
#> 20: 2022-07-03 04:34:36        1  3.09847835 13.61489916          0.015625
#> 21: 2022-07-03 04:34:36        1 -1.10389383  3.82263815          0.015625
#> 22: 2022-07-03 04:34:36        1  2.56024512 14.75051741          0.015625
#> 23: 2022-07-03 04:34:36        1  6.92728540  7.67281501          0.015625
#> 24: 2022-07-03 04:34:36        1  5.30274609 13.66192238          0.015625
#> 25: 2022-07-03 04:34:36        1  5.40009453  0.30664317          0.015625
#> 26: 2022-07-03 04:34:36        1 -2.38841885  0.33311384          0.015625
#> 27: 2022-07-03 04:34:36        1 -3.00556078  1.00744133          0.015625
#> 28: 2022-07-03 04:34:36        1 -1.73212405  8.56768440          0.015625
#> 29: 2022-07-03 04:34:36        1 -1.99401028  0.04108989          0.015625
#> 30: 2022-07-03 04:34:36        1  1.50384485  4.90658756          0.015625
#> 31: 2022-07-03 04:34:36        1  8.93892721 14.08911025          0.015625
#> 32: 2022-07-03 04:34:36        1  9.18418860  4.21853597          0.015625
#> 33: 2022-07-03 04:34:36        1 -2.70571299  1.32163593          0.015625
#> 34: 2022-07-03 04:34:36        1 -3.82594246 13.76318122          0.015625
#> 35: 2022-07-03 04:34:36        1  4.98457155  5.54372549          0.015625
#> 36: 2022-07-03 04:34:36        1 -3.29474567 14.21606417          0.015625
#> 37: 2022-07-03 04:34:36        1  3.23227638 10.89319886          0.015625
#> 38: 2022-07-03 04:34:36        1  1.97049733  7.31436137          0.015625
#> 39: 2022-07-03 04:34:36        1 -4.02102887  0.88056093          0.015625
#> 40: 2022-07-03 04:34:36        1  5.07692550  3.75477394          0.015625
#> 41: 2022-07-03 04:34:36        1 -4.31109083 13.50712486          0.015625
#> 42: 2022-07-03 04:34:36        1  7.34560216 11.70471925          0.015625
#> 43: 2022-07-03 04:34:36        1 -3.64159583 10.75581545          0.015625
#> 44: 2022-07-03 04:34:36        1  3.70762199 10.41677381          0.015625
#> 45: 2022-07-03 04:34:36        1  2.36568095  9.43712778          0.015625
#> 46: 2022-07-03 04:34:36        1  3.39903949 13.35588633          0.015625
#> 47: 2022-07-03 04:34:36        1  3.97399495  5.33769768          0.015625
#> 48: 2022-07-03 04:34:36        1  9.64540967  9.81417321          0.015625
#> 49: 2022-07-03 04:34:36        1 -1.48682874  0.10114731          0.015625
#> 50: 2022-07-03 04:34:36        1  5.83546033 14.56946383          0.015625
#> 51: 2022-07-03 04:34:36        1  0.08964452  8.24042993          0.015625
#> 52: 2022-07-03 04:34:36        1  5.82124368 13.74477839          0.015625
#> 53: 2022-07-03 04:34:36        1  5.65621291  6.92904477          0.015625
#> 54: 2022-07-03 04:34:36        1  1.37461684 11.11977623          0.015625
#> 55: 2022-07-03 04:34:36        1  0.99519518 11.35751450          0.015625
#> 56: 2022-07-03 04:34:36        1  5.44222489 14.22025796          0.015625
#> 57: 2022-07-03 04:34:36        1  5.36528397 12.82192681          0.015625
#> 58: 2022-07-03 04:34:36        1  7.77141141  4.02005897          0.015625
#> 59: 2022-07-03 04:34:36        1  9.47774904  8.46929792          0.015625
#> 60: 2022-07-03 04:34:36        1  5.67300934 12.49262088          0.015625
#> 61: 2022-07-03 04:34:36        1  7.94213928  4.47571548          0.015625
#> 62: 2022-07-03 04:34:36        1  9.74441125  5.04320687          0.015625
#> 63: 2022-07-03 04:34:36        1  2.66187675  0.39187738          0.015625
#> 64: 2022-07-03 04:34:36        1  6.06830139  5.90898592          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity