Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>         x1       x2 fidelity  x_domain        y
#> 1: 2.90609 2.642596 0.015625 <list[3]> 1.679223

# best scoring evaluation
instance$result
#>         x1       x2 fidelity  x_domain        y
#> 1: 2.90609 2.642596 0.015625 <list[3]> 1.679223

# all evaluations
as.data.table(instance$archive)
#>             x1          x2 fidelity stage bracket repetition          y
#>  1:  7.1952901  8.62724298 0.015625     0       6          1 171.804019
#>  2:  5.6440213  3.70355111 0.015625     0       6          1  50.274682
#>  3:  2.9209895 14.42670482 0.015625     0       6          1 164.810348
#>  4: -1.7739038  9.07350034 0.015625     0       6          1   8.086681
#>  5:  0.4430497  3.74676742 0.015625     0       6          1  21.090341
#>  6:  8.3940291 14.52554899 0.015625     0       6          1 393.887402
#>  7:  3.0368780  4.68201725 0.015625     0       6          1  10.894953
#>  8:  3.7838317 12.55568702 0.015625     0       6          1 149.632998
#>  9:  7.9910195  9.00678011 0.015625     0       6          1 198.066856
#> 10:  0.1412439  6.89754859 0.015625     0       6          1  20.764774
#> 11: -0.6948319 11.57365535 0.015625     0       6          1  37.204750
#> 12:  3.7801918  0.09302587 0.015625     0       6          1   2.399049
#> 13:  8.4857279 13.94395069 0.015625     0       6          1 374.329595
#> 14:  1.0334463 12.09124338 0.015625     0       6          1  74.254026
#> 15: -1.4294484  9.68940756 0.015625     0       6          1  13.179394
#> 16: -4.6445540  8.65206542 0.015625     0       6          1  38.544173
#> 17:  0.4453589 11.30315520 0.015625     0       6          1  54.735928
#> 18:  1.7215724  1.02734841 0.015625     0       6          1  13.957814
#> 19: -3.6853070  2.18389066 0.015625     0       6          1 103.773343
#> 20:  1.4714655  4.00889606 0.015625     0       6          1  11.033006
#> 21:  8.8758291 13.72066117 0.015625     0       6          1 379.128519
#> 22: -2.6080053  0.46732487 0.015625     0       6          1  99.595761
#> 23:  3.2333806 14.65549775 0.015625     0       6          1 182.152227
#> 24: -1.3657205 13.86055911 0.015625     0       6          1  43.647765
#> 25: -0.3132006  9.08378178 0.015625     0       6          1  25.803221
#> 26:  6.2082354  3.36167786 0.015625     0       6          1  56.266455
#> 27: -4.4518214 10.79095325 0.015625     0       6          1  15.957730
#> 28: -4.4436662 11.29445655 0.015625     0       6          1  13.138831
#> 29:  4.2049745  2.88487130 0.015625     0       6          1  14.537081
#> 30: -4.0471190  0.43194019 0.015625     0       6          1 160.644562
#> 31:  9.1338948 14.49818582 0.015625     0       6          1 419.824963
#> 32:  7.8534715  3.10365985 0.015625     0       6          1  69.394489
#> 33:  5.2130287  4.37695571 0.015625     0       6          1  48.693210
#> 34:  3.2161948  3.21162707 0.015625     0       6          1   4.474010
#> 35: -0.1932775  4.08697088 0.015625     0       6          1  24.359665
#> 36:  8.3008291 13.32628718 0.015625     0       6          1 345.105919
#> 37:  2.9060904  2.64259586 0.015625     0       6          1   1.679223
#> 38:  1.5551325  0.51040342 0.015625     0       6          1  19.691616
#> 39: -1.6845747  8.03149180 0.015625     0       6          1   9.452785
#> 40:  7.1984390 12.03552402 0.015625     0       6          1 268.630327
#> 41:  3.3239087 13.11665869 0.015625     0       6          1 146.172400
#> 42:  7.6283855  7.01251267 0.015625     0       6          1 141.293531
#> 43: -1.0635407  0.97812579 0.015625     0       6          1  60.217977
#> 44: -3.0591743  7.06909415 0.015625     0       6          1  17.137976
#> 45:  3.5221096  2.40352646 0.015625     0       6          1   3.734151
#> 46:  5.0082022  6.10430618 0.015625     0       6          1  66.146264
#> 47:  5.4328813  6.18240219 0.015625     0       6          1  79.086256
#> 48:  8.7782183  9.41276345 0.015625     0       6          1 227.769854
#> 49:  5.9246380  1.80800727 0.015625     0       6          1  36.281290
#> 50: -3.5788876  0.73606292 0.015625     0       6          1 130.208733
#> 51:  1.1538663 13.86192412 0.015625     0       6          1 107.154239
#> 52:  0.8026025 10.09180048 0.015625     0       6          1  45.287728
#> 53:  6.1873668  7.95344855 0.015625     0       6          1 132.424155
#> 54:  1.0428830  9.80540433 0.015625     0       6          1  44.340946
#> 55: -4.5922537  3.45170158 0.015625     0       6          1 119.215056
#> 56: -3.2686863  2.53085807 0.015625     0       6          1  81.474044
#> 57: -4.1859721  7.18699272 0.015625     0       6          1  41.343113
#> 58: -2.3170999  9.54862220 0.015625     0       6          1   3.573402
#> 59: -2.1020556  2.39569427 0.015625     0       6          1  55.342401
#> 60:  3.6850447 14.21436298 0.015625     0       6          1 188.425045
#> 61:  5.8966873  3.35655979 0.015625     0       6          1  51.069107
#> 62: -0.2446953  2.50300770 0.015625     0       6          1  34.434782
#> 63:  4.9650602  2.13944076 0.015625     0       6          1  23.182537
#> 64:  0.4418422 10.35340036 0.015625     0       6          1  44.188655
#>             x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-14 04:58:11        1   7.1952901  8.62724298          0.015625
#>  2: 2022-05-14 04:58:11        1   5.6440213  3.70355111          0.015625
#>  3: 2022-05-14 04:58:11        1   2.9209895 14.42670482          0.015625
#>  4: 2022-05-14 04:58:11        1  -1.7739038  9.07350034          0.015625
#>  5: 2022-05-14 04:58:11        1   0.4430497  3.74676742          0.015625
#>  6: 2022-05-14 04:58:11        1   8.3940291 14.52554899          0.015625
#>  7: 2022-05-14 04:58:11        1   3.0368780  4.68201725          0.015625
#>  8: 2022-05-14 04:58:11        1   3.7838317 12.55568702          0.015625
#>  9: 2022-05-14 04:58:11        1   7.9910195  9.00678011          0.015625
#> 10: 2022-05-14 04:58:11        1   0.1412439  6.89754859          0.015625
#> 11: 2022-05-14 04:58:11        1  -0.6948319 11.57365535          0.015625
#> 12: 2022-05-14 04:58:11        1   3.7801918  0.09302587          0.015625
#> 13: 2022-05-14 04:58:11        1   8.4857279 13.94395069          0.015625
#> 14: 2022-05-14 04:58:11        1   1.0334463 12.09124338          0.015625
#> 15: 2022-05-14 04:58:11        1  -1.4294484  9.68940756          0.015625
#> 16: 2022-05-14 04:58:11        1  -4.6445540  8.65206542          0.015625
#> 17: 2022-05-14 04:58:11        1   0.4453589 11.30315520          0.015625
#> 18: 2022-05-14 04:58:11        1   1.7215724  1.02734841          0.015625
#> 19: 2022-05-14 04:58:11        1  -3.6853070  2.18389066          0.015625
#> 20: 2022-05-14 04:58:11        1   1.4714655  4.00889606          0.015625
#> 21: 2022-05-14 04:58:11        1   8.8758291 13.72066117          0.015625
#> 22: 2022-05-14 04:58:11        1  -2.6080053  0.46732487          0.015625
#> 23: 2022-05-14 04:58:11        1   3.2333806 14.65549775          0.015625
#> 24: 2022-05-14 04:58:11        1  -1.3657205 13.86055911          0.015625
#> 25: 2022-05-14 04:58:11        1  -0.3132006  9.08378178          0.015625
#> 26: 2022-05-14 04:58:11        1   6.2082354  3.36167786          0.015625
#> 27: 2022-05-14 04:58:11        1  -4.4518214 10.79095325          0.015625
#> 28: 2022-05-14 04:58:11        1  -4.4436662 11.29445655          0.015625
#> 29: 2022-05-14 04:58:11        1   4.2049745  2.88487130          0.015625
#> 30: 2022-05-14 04:58:11        1  -4.0471190  0.43194019          0.015625
#> 31: 2022-05-14 04:58:11        1   9.1338948 14.49818582          0.015625
#> 32: 2022-05-14 04:58:11        1   7.8534715  3.10365985          0.015625
#> 33: 2022-05-14 04:58:11        1   5.2130287  4.37695571          0.015625
#> 34: 2022-05-14 04:58:11        1   3.2161948  3.21162707          0.015625
#> 35: 2022-05-14 04:58:11        1  -0.1932775  4.08697088          0.015625
#> 36: 2022-05-14 04:58:11        1   8.3008291 13.32628718          0.015625
#> 37: 2022-05-14 04:58:11        1   2.9060904  2.64259586          0.015625
#> 38: 2022-05-14 04:58:11        1   1.5551325  0.51040342          0.015625
#> 39: 2022-05-14 04:58:11        1  -1.6845747  8.03149180          0.015625
#> 40: 2022-05-14 04:58:11        1   7.1984390 12.03552402          0.015625
#> 41: 2022-05-14 04:58:11        1   3.3239087 13.11665869          0.015625
#> 42: 2022-05-14 04:58:11        1   7.6283855  7.01251267          0.015625
#> 43: 2022-05-14 04:58:11        1  -1.0635407  0.97812579          0.015625
#> 44: 2022-05-14 04:58:11        1  -3.0591743  7.06909415          0.015625
#> 45: 2022-05-14 04:58:11        1   3.5221096  2.40352646          0.015625
#> 46: 2022-05-14 04:58:11        1   5.0082022  6.10430618          0.015625
#> 47: 2022-05-14 04:58:11        1   5.4328813  6.18240219          0.015625
#> 48: 2022-05-14 04:58:11        1   8.7782183  9.41276345          0.015625
#> 49: 2022-05-14 04:58:11        1   5.9246380  1.80800727          0.015625
#> 50: 2022-05-14 04:58:11        1  -3.5788876  0.73606292          0.015625
#> 51: 2022-05-14 04:58:11        1   1.1538663 13.86192412          0.015625
#> 52: 2022-05-14 04:58:11        1   0.8026025 10.09180048          0.015625
#> 53: 2022-05-14 04:58:11        1   6.1873668  7.95344855          0.015625
#> 54: 2022-05-14 04:58:11        1   1.0428830  9.80540433          0.015625
#> 55: 2022-05-14 04:58:11        1  -4.5922537  3.45170158          0.015625
#> 56: 2022-05-14 04:58:11        1  -3.2686863  2.53085807          0.015625
#> 57: 2022-05-14 04:58:11        1  -4.1859721  7.18699272          0.015625
#> 58: 2022-05-14 04:58:11        1  -2.3170999  9.54862220          0.015625
#> 59: 2022-05-14 04:58:11        1  -2.1020556  2.39569427          0.015625
#> 60: 2022-05-14 04:58:11        1   3.6850447 14.21436298          0.015625
#> 61: 2022-05-14 04:58:11        1   5.8966873  3.35655979          0.015625
#> 62: 2022-05-14 04:58:11        1  -0.2446953  2.50300770          0.015625
#> 63: 2022-05-14 04:58:11        1   4.9650602  2.13944076          0.015625
#> 64: 2022-05-14 04:58:11        1   0.4418422 10.35340036          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity