Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.759752 10.67562 0.015625 <list[3]> 1.091847

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.759752 10.67562 0.015625 <list[3]> 1.091847

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1:  8.3850812 13.5262800 0.015625     0       6          1 355.191138
#>  2:  9.2883256 14.6526654 0.015625     0       6          1 432.414338
#>  3: -1.6077606 12.9296291 0.015625     0       6          1  28.060581
#>  4:  1.5572227  9.0823629 0.015625     0       6          1  40.228723
#>  5:  9.7828261  8.0717946 0.015625     0       6          1 217.068716
#>  6:  1.3057338  2.1734551 0.015625     0       6          1  15.758438
#>  7:  5.3725135  3.2966820 0.015625     0       6          1  40.488017
#>  8:  2.4308108 10.8527244 0.015625     0       6          1  75.651216
#>  9:  1.3100977 10.2440967 0.015625     0       6          1  51.868310
#> 10:  0.3848712 10.1537343 0.015625     0       6          1  41.573678
#> 11:  1.2757732  0.1062572 0.015625     0       6          1  28.106085
#> 12:  0.4457446  9.5072028 0.015625     0       6          1  36.392354
#> 13:  0.1371439  7.0041812 0.015625     0       6          1  21.004930
#> 14:  8.9706643 14.0416395 0.015625     0       6          1 395.179076
#> 15: -2.7597516 10.6756193 0.015625     0       6          1   1.091847
#> 16:  6.0670683 13.5643545 0.015625     0       6          1 278.222343
#> 17:  4.8735865 10.4239189 0.015625     0       6          1 142.647640
#> 18:  6.1665490  9.3230156 0.015625     0       6          1 162.774380
#> 19:  2.7435545  2.7759410 0.015625     0       6          1   1.978486
#> 20:  9.0494628  2.0992948 0.015625     0       6          1  64.810594
#> 21:  4.5741033  6.6821492 0.015625     0       6          1  62.240624
#> 22:  1.4489241 11.6870861 0.015625     0       6          1  74.029570
#> 23:  2.3520211  4.8106241 0.015625     0       6          1   8.921605
#> 24:  7.4929462  5.5962044 0.015625     0       6          1 109.340393
#> 25: -4.2384621  2.3457027 0.015625     0       6          1 125.572284
#> 26: -4.7371971  5.5073178 0.015625     0       6          1  86.314266
#> 27:  2.6442192  9.6547527 0.015625     0       6          1  60.055925
#> 28:  4.4796185 13.6982782 0.015625     0       6          1 209.732238
#> 29:  4.2768607  5.2503008 0.015625     0       6          1  36.140983
#> 30:  0.5738056 11.6881165 0.015625     0       6          1  61.508602
#> 31:  1.8212870  3.6906070 0.015625     0       6          1   7.857281
#> 32:  2.1853252  8.7683593 0.015625     0       6          1  41.668496
#> 33:  7.4148953  3.8243232 0.015625     0       6          1  77.046273
#> 34: -3.8588234  5.1815421 0.015625     0       6          1  57.787402
#> 35:  7.8607793 13.7005009 0.015625     0       6          1 345.242283
#> 36: -1.2436261  3.6262098 0.015625     0       6          1  32.451378
#> 37:  7.7570297  9.5212763 0.015625     0       6          1 207.402370
#> 38:  8.3141084  6.7998844 0.015625     0       6          1 147.508679
#> 39: -0.7754294  1.3864906 0.015625     0       6          1  51.268564
#> 40:  0.5487943 11.0183758 0.015625     0       6          1  52.796462
#> 41:  3.2631733  0.3163386 0.015625     0       6          1   1.137178
#> 42: -0.7294517  8.0770801 0.015625     0       6          1  17.968329
#> 43: -0.7642124 13.7113964 0.015625     0       6          1  58.885603
#> 44: -3.6653625  5.6238162 0.015625     0       6          1  45.547579
#> 45:  5.7872481  2.8791772 0.015625     0       6          1  44.049713
#> 46:  0.2286837 14.3979786 0.015625     0       6          1  96.095545
#> 47:  1.0557306  9.7951330 0.015625     0       6          1  44.335612
#> 48:  0.5480187  6.2895156 0.015625     0       6          1  19.524175
#> 49:  6.8439389 12.9571572 0.015625     0       6          1 287.401632
#> 50:  6.8695623  9.8238871 0.015625     0       6          1 195.051887
#> 51:  9.3762292  7.0353193 0.015625     0       6          1 176.103404
#> 52:  5.6923530  3.8596986 0.015625     0       6          1  53.057071
#> 53:  8.4871314 14.2117923 0.015625     0       6          1 384.753099
#> 54:  2.7445525  5.8329780 0.015625     0       6          1  16.901492
#> 55: -0.6437532 11.1085580 0.015625     0       6          1  34.255320
#> 56:  9.8282842  2.5609478 0.015625     0       6          1  86.419822
#> 57:  6.3514934 14.6618311 0.015625     0       6          1 326.886486
#> 58: -4.3251440  4.0988352 0.015625     0       6          1  93.983950
#> 59: -2.4569508  2.0728105 0.015625     0       6          1  66.932794
#> 60:  9.0769416 14.0556002 0.015625     0       6          1 399.723059
#> 61:  6.8822160  4.8342552 0.015625     0       6          1  87.340997
#> 62: -3.9621851  2.5300809 0.015625     0       6          1 108.692826
#> 63: -3.4092389 14.9323878 0.015625     0       6          1  10.656250
#> 64:  9.2743748  2.5290744 0.015625     0       6          1  75.242766
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-08 04:32:29        1   8.3850812  13.5262800          0.015625
#>  2: 2022-05-08 04:32:29        1   9.2883256  14.6526654          0.015625
#>  3: 2022-05-08 04:32:29        1  -1.6077606  12.9296291          0.015625
#>  4: 2022-05-08 04:32:29        1   1.5572227   9.0823629          0.015625
#>  5: 2022-05-08 04:32:29        1   9.7828261   8.0717946          0.015625
#>  6: 2022-05-08 04:32:29        1   1.3057338   2.1734551          0.015625
#>  7: 2022-05-08 04:32:29        1   5.3725135   3.2966820          0.015625
#>  8: 2022-05-08 04:32:29        1   2.4308108  10.8527244          0.015625
#>  9: 2022-05-08 04:32:29        1   1.3100977  10.2440967          0.015625
#> 10: 2022-05-08 04:32:29        1   0.3848712  10.1537343          0.015625
#> 11: 2022-05-08 04:32:29        1   1.2757732   0.1062572          0.015625
#> 12: 2022-05-08 04:32:29        1   0.4457446   9.5072028          0.015625
#> 13: 2022-05-08 04:32:29        1   0.1371439   7.0041812          0.015625
#> 14: 2022-05-08 04:32:29        1   8.9706643  14.0416395          0.015625
#> 15: 2022-05-08 04:32:29        1  -2.7597516  10.6756193          0.015625
#> 16: 2022-05-08 04:32:29        1   6.0670683  13.5643545          0.015625
#> 17: 2022-05-08 04:32:29        1   4.8735865  10.4239189          0.015625
#> 18: 2022-05-08 04:32:29        1   6.1665490   9.3230156          0.015625
#> 19: 2022-05-08 04:32:29        1   2.7435545   2.7759410          0.015625
#> 20: 2022-05-08 04:32:29        1   9.0494628   2.0992948          0.015625
#> 21: 2022-05-08 04:32:29        1   4.5741033   6.6821492          0.015625
#> 22: 2022-05-08 04:32:29        1   1.4489241  11.6870861          0.015625
#> 23: 2022-05-08 04:32:29        1   2.3520211   4.8106241          0.015625
#> 24: 2022-05-08 04:32:29        1   7.4929462   5.5962044          0.015625
#> 25: 2022-05-08 04:32:29        1  -4.2384621   2.3457027          0.015625
#> 26: 2022-05-08 04:32:29        1  -4.7371971   5.5073178          0.015625
#> 27: 2022-05-08 04:32:29        1   2.6442192   9.6547527          0.015625
#> 28: 2022-05-08 04:32:29        1   4.4796185  13.6982782          0.015625
#> 29: 2022-05-08 04:32:29        1   4.2768607   5.2503008          0.015625
#> 30: 2022-05-08 04:32:29        1   0.5738056  11.6881165          0.015625
#> 31: 2022-05-08 04:32:29        1   1.8212870   3.6906070          0.015625
#> 32: 2022-05-08 04:32:29        1   2.1853252   8.7683593          0.015625
#> 33: 2022-05-08 04:32:29        1   7.4148953   3.8243232          0.015625
#> 34: 2022-05-08 04:32:29        1  -3.8588234   5.1815421          0.015625
#> 35: 2022-05-08 04:32:29        1   7.8607793  13.7005009          0.015625
#> 36: 2022-05-08 04:32:29        1  -1.2436261   3.6262098          0.015625
#> 37: 2022-05-08 04:32:29        1   7.7570297   9.5212763          0.015625
#> 38: 2022-05-08 04:32:29        1   8.3141084   6.7998844          0.015625
#> 39: 2022-05-08 04:32:29        1  -0.7754294   1.3864906          0.015625
#> 40: 2022-05-08 04:32:29        1   0.5487943  11.0183758          0.015625
#> 41: 2022-05-08 04:32:29        1   3.2631733   0.3163386          0.015625
#> 42: 2022-05-08 04:32:29        1  -0.7294517   8.0770801          0.015625
#> 43: 2022-05-08 04:32:29        1  -0.7642124  13.7113964          0.015625
#> 44: 2022-05-08 04:32:29        1  -3.6653625   5.6238162          0.015625
#> 45: 2022-05-08 04:32:29        1   5.7872481   2.8791772          0.015625
#> 46: 2022-05-08 04:32:29        1   0.2286837  14.3979786          0.015625
#> 47: 2022-05-08 04:32:29        1   1.0557306   9.7951330          0.015625
#> 48: 2022-05-08 04:32:29        1   0.5480187   6.2895156          0.015625
#> 49: 2022-05-08 04:32:29        1   6.8439389  12.9571572          0.015625
#> 50: 2022-05-08 04:32:29        1   6.8695623   9.8238871          0.015625
#> 51: 2022-05-08 04:32:29        1   9.3762292   7.0353193          0.015625
#> 52: 2022-05-08 04:32:29        1   5.6923530   3.8596986          0.015625
#> 53: 2022-05-08 04:32:29        1   8.4871314  14.2117923          0.015625
#> 54: 2022-05-08 04:32:29        1   2.7445525   5.8329780          0.015625
#> 55: 2022-05-08 04:32:29        1  -0.6437532  11.1085580          0.015625
#> 56: 2022-05-08 04:32:29        1   9.8282842   2.5609478          0.015625
#> 57: 2022-05-08 04:32:29        1   6.3514934  14.6618311          0.015625
#> 58: 2022-05-08 04:32:29        1  -4.3251440   4.0988352          0.015625
#> 59: 2022-05-08 04:32:29        1  -2.4569508   2.0728105          0.015625
#> 60: 2022-05-08 04:32:29        1   9.0769416  14.0556002          0.015625
#> 61: 2022-05-08 04:32:29        1   6.8822160   4.8342552          0.015625
#> 62: 2022-05-08 04:32:29        1  -3.9621851   2.5300809          0.015625
#> 63: 2022-05-08 04:32:29        1  -3.4092389  14.9323878          0.015625
#> 64: 2022-05-08 04:32:29        1   9.2743748   2.5290744          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity