Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: -3.34073 13.16816 0.015625 <list[3]> 2.861937

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: -3.34073 13.16816 0.015625 <list[3]> 2.861937

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  3.97013679  2.5911827 0.015625     0       6          1   9.391139
#>  2:  8.55556851 10.1207747 0.015625     0       6          1 243.642638
#>  3:  6.31609780  5.2603468 0.015625     0       6          1  84.982499
#>  4:  1.51643215 11.8536644 0.015625     0       6          1  77.703323
#>  5:  1.12927905  3.2649397 0.015625     0       6          1  15.057561
#>  6:  8.18078379  6.4674880 0.015625     0       6          1 137.559251
#>  7:  9.18871647  7.2484169 0.015625     0       6          1 176.934051
#>  8:  9.48317657 10.7849701 0.015625     0       6          1 293.262735
#>  9: -1.82077654  6.0468973 0.015625     0       6          1  16.344154
#> 10:  0.05011658  7.7368686 0.015625     0       6          1  22.889926
#> 11: -3.99566128  1.4433592 0.015625     0       6          1 133.807684
#> 12: -4.27483437 11.4497295 0.015625     0       6          1   9.601467
#> 13: -1.85872206 11.5927477 0.015625     0       6          1  13.665513
#> 14:  4.34870354 10.9148812 0.015625     0       6          1 133.250220
#> 15:  7.78849299  8.3171372 0.015625     0       6          1 175.693729
#> 16: -3.34072964 13.1681624 0.015625     0       6          1   2.861937
#> 17: -2.17055800  5.5407437 0.015625     0       6          1  21.052895
#> 18: -0.45370391  7.8276154 0.015625     0       6          1  19.838893
#> 19:  2.76937759  8.4598756 0.015625     0       6          1  45.034371
#> 20: -0.03834837 11.1781850 0.015625     0       6          1  45.779832
#> 21: -3.33669307  6.6272089 0.015625     0       6          1  25.836969
#> 22: -4.44434323 13.7209671 0.015625     0       6          1   7.458525
#> 23:  7.10221301  1.5737222 0.015625     0       6          1  44.927296
#> 24:  2.73219860  7.8800446 0.015625     0       6          1  37.178788
#> 25: -1.43076392  8.8558930 0.015625     0       6          1  11.606287
#> 26:  2.43874763  9.0879464 0.015625     0       6          1  48.729653
#> 27:  1.26661997  6.6251306 0.015625     0       6          1  19.592765
#> 28:  0.19526087 11.7359277 0.015625     0       6          1  55.967990
#> 29:  5.31528878  7.1957076 0.015625     0       6          1  92.648686
#> 30: -2.45643742 14.6741286 0.015625     0       6          1  23.532646
#> 31:  6.77000415 11.7233606 0.015625     0       6          1 246.162418
#> 32:  6.27953694  5.9810756 0.015625     0       6          1  96.389326
#> 33:  7.06325858  4.4798395 0.015625     0       6          1  83.859427
#> 34: -0.61595137  0.5405658 0.015625     0       6          1  59.458252
#> 35:  8.60276101 13.7107437 0.015625     0       6          1 369.303774
#> 36: -0.76546686  7.2119640 0.015625     0       6          1  16.924281
#> 37:  8.72910539  4.7674204 0.015625     0       6          1 109.077626
#> 38:  2.88750448  8.2089746 0.015625     0       6          1  43.585393
#> 39:  9.30073358  1.2346688 0.015625     0       6          1  54.899433
#> 40:  0.81515146  3.0551705 0.015625     0       6          1  19.366638
#> 41:  5.26347848  2.2671402 0.015625     0       6          1  29.410153
#> 42: -4.19952199  9.1592615 0.015625     0       6          1  21.826827
#> 43: -4.10461792  9.9197222 0.015625     0       6          1  14.319994
#> 44:  2.17907252  3.6383170 0.015625     0       6          1   5.435246
#> 45:  6.45146641  0.2744780 0.015625     0       6          1  30.110847
#> 46:  4.42788026  8.4907885 0.015625     0       6          1  87.141712
#> 47: -1.36238814  9.5464016 0.015625     0       6          1  13.731808
#> 48: -4.39253794 14.5567124 0.015625     0       6          1   7.926662
#> 49: -2.38265782 11.5887373 0.015625     0       6          1   5.664102
#> 50: -0.72720546  5.6944198 0.015625     0       6          1  19.361217
#> 51:  2.63461927  7.3141685 0.015625     0       6          1  29.630758
#> 52:  2.46514375 12.3558740 0.015625     0       6          1 104.369300
#> 53:  0.40967121  7.8295837 0.015625     0       6          1  24.940288
#> 54: -0.58006165 12.7302494 0.015625     0       6          1  51.633298
#> 55:  4.04618186  0.8101415 0.015625     0       6          1   4.623018
#> 56:  1.90503873 14.8956686 0.015625     0       6          1 146.468928
#> 57:  5.46617957  7.7766168 0.015625     0       6          1 107.919824
#> 58: -2.08383967 13.3871378 0.015625     0       6          1  20.787651
#> 59:  4.63322432  8.8062752 0.015625     0       6          1  99.875651
#> 60:  5.79620116  8.0766166 0.015625     0       6          1 123.929634
#> 61: -4.16375460  9.0669078 0.015625     0       6          1  21.744684
#> 62: -1.29373444  5.8914837 0.015625     0       6          1  17.550529
#> 63: -0.26930407  9.8082839 0.015625     0       6          1  30.663140
#> 64:  6.53939942  3.3723679 0.015625     0       6          1  61.088615
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-07-21 04:42:19        1  3.97013679   2.5911827          0.015625
#>  2: 2022-07-21 04:42:19        1  8.55556851  10.1207747          0.015625
#>  3: 2022-07-21 04:42:19        1  6.31609780   5.2603468          0.015625
#>  4: 2022-07-21 04:42:19        1  1.51643215  11.8536644          0.015625
#>  5: 2022-07-21 04:42:19        1  1.12927905   3.2649397          0.015625
#>  6: 2022-07-21 04:42:19        1  8.18078379   6.4674880          0.015625
#>  7: 2022-07-21 04:42:19        1  9.18871647   7.2484169          0.015625
#>  8: 2022-07-21 04:42:19        1  9.48317657  10.7849701          0.015625
#>  9: 2022-07-21 04:42:19        1 -1.82077654   6.0468973          0.015625
#> 10: 2022-07-21 04:42:19        1  0.05011658   7.7368686          0.015625
#> 11: 2022-07-21 04:42:19        1 -3.99566128   1.4433592          0.015625
#> 12: 2022-07-21 04:42:19        1 -4.27483437  11.4497295          0.015625
#> 13: 2022-07-21 04:42:19        1 -1.85872206  11.5927477          0.015625
#> 14: 2022-07-21 04:42:19        1  4.34870354  10.9148812          0.015625
#> 15: 2022-07-21 04:42:19        1  7.78849299   8.3171372          0.015625
#> 16: 2022-07-21 04:42:19        1 -3.34072964  13.1681624          0.015625
#> 17: 2022-07-21 04:42:19        1 -2.17055800   5.5407437          0.015625
#> 18: 2022-07-21 04:42:19        1 -0.45370391   7.8276154          0.015625
#> 19: 2022-07-21 04:42:19        1  2.76937759   8.4598756          0.015625
#> 20: 2022-07-21 04:42:19        1 -0.03834837  11.1781850          0.015625
#> 21: 2022-07-21 04:42:19        1 -3.33669307   6.6272089          0.015625
#> 22: 2022-07-21 04:42:19        1 -4.44434323  13.7209671          0.015625
#> 23: 2022-07-21 04:42:19        1  7.10221301   1.5737222          0.015625
#> 24: 2022-07-21 04:42:19        1  2.73219860   7.8800446          0.015625
#> 25: 2022-07-21 04:42:19        1 -1.43076392   8.8558930          0.015625
#> 26: 2022-07-21 04:42:19        1  2.43874763   9.0879464          0.015625
#> 27: 2022-07-21 04:42:19        1  1.26661997   6.6251306          0.015625
#> 28: 2022-07-21 04:42:19        1  0.19526087  11.7359277          0.015625
#> 29: 2022-07-21 04:42:19        1  5.31528878   7.1957076          0.015625
#> 30: 2022-07-21 04:42:19        1 -2.45643742  14.6741286          0.015625
#> 31: 2022-07-21 04:42:19        1  6.77000415  11.7233606          0.015625
#> 32: 2022-07-21 04:42:19        1  6.27953694   5.9810756          0.015625
#> 33: 2022-07-21 04:42:19        1  7.06325858   4.4798395          0.015625
#> 34: 2022-07-21 04:42:19        1 -0.61595137   0.5405658          0.015625
#> 35: 2022-07-21 04:42:19        1  8.60276101  13.7107437          0.015625
#> 36: 2022-07-21 04:42:19        1 -0.76546686   7.2119640          0.015625
#> 37: 2022-07-21 04:42:19        1  8.72910539   4.7674204          0.015625
#> 38: 2022-07-21 04:42:19        1  2.88750448   8.2089746          0.015625
#> 39: 2022-07-21 04:42:19        1  9.30073358   1.2346688          0.015625
#> 40: 2022-07-21 04:42:19        1  0.81515146   3.0551705          0.015625
#> 41: 2022-07-21 04:42:19        1  5.26347848   2.2671402          0.015625
#> 42: 2022-07-21 04:42:19        1 -4.19952199   9.1592615          0.015625
#> 43: 2022-07-21 04:42:19        1 -4.10461792   9.9197222          0.015625
#> 44: 2022-07-21 04:42:19        1  2.17907252   3.6383170          0.015625
#> 45: 2022-07-21 04:42:19        1  6.45146641   0.2744780          0.015625
#> 46: 2022-07-21 04:42:19        1  4.42788026   8.4907885          0.015625
#> 47: 2022-07-21 04:42:19        1 -1.36238814   9.5464016          0.015625
#> 48: 2022-07-21 04:42:19        1 -4.39253794  14.5567124          0.015625
#> 49: 2022-07-21 04:42:19        1 -2.38265782  11.5887373          0.015625
#> 50: 2022-07-21 04:42:19        1 -0.72720546   5.6944198          0.015625
#> 51: 2022-07-21 04:42:19        1  2.63461927   7.3141685          0.015625
#> 52: 2022-07-21 04:42:19        1  2.46514375  12.3558740          0.015625
#> 53: 2022-07-21 04:42:19        1  0.40967121   7.8295837          0.015625
#> 54: 2022-07-21 04:42:19        1 -0.58006165  12.7302494          0.015625
#> 55: 2022-07-21 04:42:19        1  4.04618186   0.8101415          0.015625
#> 56: 2022-07-21 04:42:19        1  1.90503873  14.8956686          0.015625
#> 57: 2022-07-21 04:42:19        1  5.46617957   7.7766168          0.015625
#> 58: 2022-07-21 04:42:19        1 -2.08383967  13.3871378          0.015625
#> 59: 2022-07-21 04:42:19        1  4.63322432   8.8062752          0.015625
#> 60: 2022-07-21 04:42:19        1  5.79620116   8.0766166          0.015625
#> 61: 2022-07-21 04:42:19        1 -4.16375460   9.0669078          0.015625
#> 62: 2022-07-21 04:42:19        1 -1.29373444   5.8914837          0.015625
#> 63: 2022-07-21 04:42:19        1 -0.26930407   9.8082839          0.015625
#> 64: 2022-07-21 04:42:19        1  6.53939942   3.3723679          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity