Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.634644 1.358971 0.015625 <list[3]> 2.042846

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.634644 1.358971 0.015625 <list[3]> 2.042846

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition          y
#>  1:  1.07656276 10.51838230 0.015625     0       6          1  52.947102
#>  2:  4.57502206  7.53932768 0.015625     0       6          1  75.550768
#>  3:  0.75433784  4.30085215 0.015625     0       6          1  17.263637
#>  4:  1.20012434  7.36993245 0.015625     0       6          1  23.948073
#>  5: -2.66710806  4.90771482 0.015625     0       6          1  32.325977
#>  6: -1.53241441  1.46985572 0.015625     0       6          1  59.947808
#>  7:  0.09944000  3.55374319 0.015625     0       6          1  24.790981
#>  8:  3.52216846 13.65707451 0.015625     0       6          1 167.013893
#>  9:  2.42333969  9.76704801 0.015625     0       6          1  58.173468
#> 10:  5.13320565  7.82375946 0.015625     0       6          1  98.256116
#> 11:  0.71032415  8.24501912 0.015625     0       6          1  28.569591
#> 12:  5.99794148  9.37682478 0.015625     0       6          1 158.848808
#> 13:  3.46219104 11.33770338 0.015625     0       6          1 110.704849
#> 14:  0.58296233 14.69685588 0.015625     0       6          1 110.449412
#> 15:  2.35645558  2.26018794 0.015625     0       6          1   3.234150
#> 16:  9.95886763 10.37581112 0.015625     0       6          1 296.763257
#> 17: -3.21467347 14.35378571 0.015625     0       6          1   8.948344
#> 18:  7.10387516 10.40518819 0.015625     0       6          1 217.043238
#> 19:  0.14033065  9.89219826 0.015625     0       6          1  36.440420
#> 20: -1.48936582 10.44185515 0.015625     0       6          1  14.794062
#> 21: -1.11012490  1.95452781 0.015625     0       6          1  48.493254
#> 22:  6.79672641 10.93533833 0.015625     0       6          1 223.778111
#> 23: -4.13264737  8.77940877 0.015625     0       6          1  23.428433
#> 24: -0.05686675 13.17174753 0.015625     0       6          1  69.729161
#> 25:  7.96946235 14.46135980 0.015625     0       6          1 377.239475
#> 26:  2.20275332  3.15138777 0.015625     0       6          1   4.585847
#> 27: -3.37741231  5.02383774 0.015625     0       6          1  45.583254
#> 28:  9.82440990 10.95889438 0.015625     0       6          1 311.874678
#> 29: -2.71321444  1.69587490 0.015625     0       6          1  79.564756
#> 30:  7.81438619  6.11506477 0.015625     0       6          1 124.324878
#> 31: -4.98656218 12.42074068 0.015625     0       6          1  17.798951
#> 32:  0.69480221  3.52501977 0.015625     0       6          1  19.291641
#> 33: -1.90640198  1.19312640 0.015625     0       6          1  70.083903
#> 34:  7.85907572 12.75493399 0.015625     0       6          1 311.457855
#> 35: -4.43774314 12.85544689 0.015625     0       6          1   8.056751
#> 36: -1.51800968 11.70675869 0.015625     0       6          1  20.874508
#> 37:  5.15341005  6.78311642 0.015625     0       6          1  80.822451
#> 38:  8.64277332  7.12793469 0.015625     0       6          1 161.610030
#> 39:  1.98269807  9.63586248 0.015625     0       6          1  50.652074
#> 40: -2.49411893  6.64168090 0.015625     0       6          1  14.725311
#> 41:  2.28603076 13.49027216 0.015625     0       6          1 123.998284
#> 42: -4.21273754 14.30933085 0.015625     0       6          1   6.520664
#> 43:  9.62677741  8.76206922 0.015625     0       6          1 232.670532
#> 44:  2.77173288  7.32672521 0.015625     0       6          1  31.317677
#> 45: -3.09189825  0.09275271 0.015625     0       6          1 124.110683
#> 46:  5.11919935  9.36737097 0.015625     0       6          1 128.483509
#> 47:  8.71537583 14.26823181 0.015625     0       6          1 394.901266
#> 48:  7.98072987  2.26014832 0.015625     0       6          1  57.835699
#> 49:  1.69299649  2.01482300 0.015625     0       6          1  10.730676
#> 50:  6.19305152 10.39175823 0.015625     0       6          1 190.362903
#> 51:  1.76625008  9.42127537 0.015625     0       6          1  45.790932
#> 52:  1.89673385  4.25715952 0.015625     0       6          1   8.283332
#> 53:  2.50895272  4.22411752 0.015625     0       6          1   6.351520
#> 54:  2.63464424  1.35897090 0.015625     0       6          1   2.042846
#> 55: -0.24651253  7.24710513 0.015625     0       6          1  20.039273
#> 56: -2.44535569 11.09731980 0.015625     0       6          1   3.676249
#> 57: -0.71599787  6.59865364 0.015625     0       6          1  17.554070
#> 58:  9.76372167 10.35828202 0.015625     0       6          1 288.809949
#> 59:  8.77215423 10.31197412 0.015625     0       6          1 255.412986
#> 60:  7.95087237 13.65969350 0.015625     0       6          1 346.534816
#> 61:  2.64931907 14.79298466 0.015625     0       6          1 165.216720
#> 62:  5.69631716 13.02961949 0.015625     0       6          1 245.942363
#> 63:  2.33366054  6.74945308 0.015625     0       6          1  21.821826
#> 64:  6.35460038 12.54639940 0.015625     0       6          1 257.306753
#>              x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-04 04:42:31        1  1.07656276 10.51838230          0.015625
#>  2: 2022-08-04 04:42:31        1  4.57502206  7.53932768          0.015625
#>  3: 2022-08-04 04:42:31        1  0.75433784  4.30085215          0.015625
#>  4: 2022-08-04 04:42:31        1  1.20012434  7.36993245          0.015625
#>  5: 2022-08-04 04:42:31        1 -2.66710806  4.90771482          0.015625
#>  6: 2022-08-04 04:42:31        1 -1.53241441  1.46985572          0.015625
#>  7: 2022-08-04 04:42:31        1  0.09944000  3.55374319          0.015625
#>  8: 2022-08-04 04:42:31        1  3.52216846 13.65707451          0.015625
#>  9: 2022-08-04 04:42:31        1  2.42333969  9.76704801          0.015625
#> 10: 2022-08-04 04:42:31        1  5.13320565  7.82375946          0.015625
#> 11: 2022-08-04 04:42:31        1  0.71032415  8.24501912          0.015625
#> 12: 2022-08-04 04:42:31        1  5.99794148  9.37682478          0.015625
#> 13: 2022-08-04 04:42:31        1  3.46219104 11.33770338          0.015625
#> 14: 2022-08-04 04:42:31        1  0.58296233 14.69685588          0.015625
#> 15: 2022-08-04 04:42:31        1  2.35645558  2.26018794          0.015625
#> 16: 2022-08-04 04:42:31        1  9.95886763 10.37581112          0.015625
#> 17: 2022-08-04 04:42:31        1 -3.21467347 14.35378571          0.015625
#> 18: 2022-08-04 04:42:31        1  7.10387516 10.40518819          0.015625
#> 19: 2022-08-04 04:42:31        1  0.14033065  9.89219826          0.015625
#> 20: 2022-08-04 04:42:31        1 -1.48936582 10.44185515          0.015625
#> 21: 2022-08-04 04:42:31        1 -1.11012490  1.95452781          0.015625
#> 22: 2022-08-04 04:42:31        1  6.79672641 10.93533833          0.015625
#> 23: 2022-08-04 04:42:31        1 -4.13264737  8.77940877          0.015625
#> 24: 2022-08-04 04:42:31        1 -0.05686675 13.17174753          0.015625
#> 25: 2022-08-04 04:42:31        1  7.96946235 14.46135980          0.015625
#> 26: 2022-08-04 04:42:31        1  2.20275332  3.15138777          0.015625
#> 27: 2022-08-04 04:42:31        1 -3.37741231  5.02383774          0.015625
#> 28: 2022-08-04 04:42:31        1  9.82440990 10.95889438          0.015625
#> 29: 2022-08-04 04:42:31        1 -2.71321444  1.69587490          0.015625
#> 30: 2022-08-04 04:42:31        1  7.81438619  6.11506477          0.015625
#> 31: 2022-08-04 04:42:31        1 -4.98656218 12.42074068          0.015625
#> 32: 2022-08-04 04:42:31        1  0.69480221  3.52501977          0.015625
#> 33: 2022-08-04 04:42:31        1 -1.90640198  1.19312640          0.015625
#> 34: 2022-08-04 04:42:31        1  7.85907572 12.75493399          0.015625
#> 35: 2022-08-04 04:42:31        1 -4.43774314 12.85544689          0.015625
#> 36: 2022-08-04 04:42:31        1 -1.51800968 11.70675869          0.015625
#> 37: 2022-08-04 04:42:31        1  5.15341005  6.78311642          0.015625
#> 38: 2022-08-04 04:42:31        1  8.64277332  7.12793469          0.015625
#> 39: 2022-08-04 04:42:31        1  1.98269807  9.63586248          0.015625
#> 40: 2022-08-04 04:42:31        1 -2.49411893  6.64168090          0.015625
#> 41: 2022-08-04 04:42:31        1  2.28603076 13.49027216          0.015625
#> 42: 2022-08-04 04:42:31        1 -4.21273754 14.30933085          0.015625
#> 43: 2022-08-04 04:42:31        1  9.62677741  8.76206922          0.015625
#> 44: 2022-08-04 04:42:31        1  2.77173288  7.32672521          0.015625
#> 45: 2022-08-04 04:42:31        1 -3.09189825  0.09275271          0.015625
#> 46: 2022-08-04 04:42:31        1  5.11919935  9.36737097          0.015625
#> 47: 2022-08-04 04:42:31        1  8.71537583 14.26823181          0.015625
#> 48: 2022-08-04 04:42:31        1  7.98072987  2.26014832          0.015625
#> 49: 2022-08-04 04:42:31        1  1.69299649  2.01482300          0.015625
#> 50: 2022-08-04 04:42:31        1  6.19305152 10.39175823          0.015625
#> 51: 2022-08-04 04:42:31        1  1.76625008  9.42127537          0.015625
#> 52: 2022-08-04 04:42:31        1  1.89673385  4.25715952          0.015625
#> 53: 2022-08-04 04:42:31        1  2.50895272  4.22411752          0.015625
#> 54: 2022-08-04 04:42:31        1  2.63464424  1.35897090          0.015625
#> 55: 2022-08-04 04:42:31        1 -0.24651253  7.24710513          0.015625
#> 56: 2022-08-04 04:42:31        1 -2.44535569 11.09731980          0.015625
#> 57: 2022-08-04 04:42:31        1 -0.71599787  6.59865364          0.015625
#> 58: 2022-08-04 04:42:31        1  9.76372167 10.35828202          0.015625
#> 59: 2022-08-04 04:42:31        1  8.77215423 10.31197412          0.015625
#> 60: 2022-08-04 04:42:31        1  7.95087237 13.65969350          0.015625
#> 61: 2022-08-04 04:42:31        1  2.64931907 14.79298466          0.015625
#> 62: 2022-08-04 04:42:31        1  5.69631716 13.02961949          0.015625
#> 63: 2022-08-04 04:42:31        1  2.33366054  6.74945308          0.015625
#> 64: 2022-08-04 04:42:31        1  6.35460038 12.54639940          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity