Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 3.763546 1.264781 0.015625 <list[3]> 2.866955

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 3.763546 1.264781 0.015625 <list[3]> 2.866955

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  0.28925509  5.0218566 0.015625     0       6          1  19.473975
#>  2: -1.28673097 10.1564200 0.015625     0       6          1  16.924879
#>  3: -2.35463244  8.3200312 0.015625     0       6          1   5.774350
#>  4:  4.29860431  7.4500198 0.015625     0       6          1  65.788845
#>  5:  4.33501974  9.3872516 0.015625     0       6          1 100.723423
#>  6:  9.66262841 12.7991530 0.015625     0       6          1 373.427177
#>  7:  8.11111873  7.3525849 0.015625     0       6          1 157.350642
#>  8:  7.74368109 10.7783777 0.015625     0       6          1 243.897010
#>  9: -0.40333649  8.7589240 0.015625     0       6          1  23.292117
#> 10: -2.13948844 14.2048098 0.015625     0       6          1  26.534939
#> 11:  0.42200079  3.0117822 0.015625     0       6          1  24.151686
#> 12: -1.86509692  3.5745919 0.015625     0       6          1  37.473050
#> 13:  4.07292304  3.8260109 0.015625     0       6          1  18.696223
#> 14:  6.99965074 14.1991134 0.015625     0       6          1 335.255378
#> 15:  3.44879310  4.9145757 0.015625     0       6          1  17.151171
#> 16: -4.40181426  7.1167892 0.015625     0       6          1  49.116468
#> 17: -0.30353281 10.4684620 0.015625     0       6          1  35.023806
#> 18:  0.69028891  8.8507173 0.015625     0       6          1  32.885652
#> 19:  0.07316201  2.4703046 0.015625     0       6          1  31.227855
#> 20:  5.40570900  2.6716965 0.015625     0       6          1  35.291895
#> 21: -0.55144784  0.6174301 0.015625     0       6          1  57.486350
#> 22:  5.07727433  4.8830921 0.015625     0       6          1  51.510232
#> 23:  6.11127558  9.4765700 0.015625     0       6          1 164.774860
#> 24: -4.46574856  8.2011965 0.015625     0       6          1  38.119974
#> 25:  1.46365728 12.7953313 0.015625     0       6          1  93.091281
#> 26: -0.51323411 12.8165289 0.015625     0       6          1  54.264167
#> 27: -4.78341828 14.2010999 0.015625     0       6          1  10.694792
#> 28:  8.34112788 10.2390849 0.015625     0       6          1 241.901878
#> 29:  3.94430652  1.2343369 0.015625     0       6          1   4.397082
#> 30: -2.21650988 14.4005366 0.015625     0       6          1  26.517091
#> 31:  3.16567919 11.3696671 0.015625     0       6          1 102.408049
#> 32: -0.37755148  3.0023172 0.015625     0       6          1  31.907139
#> 33:  0.83392610 13.2311509 0.015625     0       6          1  89.332785
#> 34: -3.82091893  4.9505822 0.015625     0       6          1  59.978227
#> 35: -2.62687125  5.6673378 0.015625     0       6          1  23.973574
#> 36:  2.35883959  4.1247207 0.015625     0       6          1   6.109098
#> 37: -4.43879808  6.1074262 0.015625     0       6          1  64.603759
#> 38: -4.68872403  8.0183362 0.015625     0       6          1  47.226529
#> 39:  9.42061214  1.9201838 0.015625     0       6          1  67.389252
#> 40: -4.98775828  3.3129204 0.015625     0       6          1 142.348844
#> 41: -1.38631213  3.9545366 0.015625     0       6          1  30.345603
#> 42: -3.40605465  2.0340485 0.015625     0       6          1  95.668649
#> 43:  5.65810524  5.9452656 0.015625     0       6          1  81.244944
#> 44:  0.94291847  0.9739030 0.015625     0       6          1  28.262466
#> 45: -4.76545934  7.7239694 0.015625     0       6          1  53.526512
#> 46:  9.04679996 13.1240762 0.015625     0       6          1 362.304820
#> 47:  0.20525005 14.5423960 0.015625     0       6          1  98.037845
#> 48:  0.94178151 12.1878109 0.015625     0       6          1  74.316279
#> 49:  3.56154839  8.5045159 0.015625     0       6          1  61.805494
#> 50:  4.74660689 10.7079228 0.015625     0       6          1 144.185156
#> 51: -1.35813925  1.4784836 0.015625     0       6          1  57.451186
#> 52:  6.86494170  3.7783814 0.015625     0       6          1  70.661178
#> 53: -0.80022394 13.6698552 0.015625     0       6          1  57.348960
#> 54: -3.51501491  2.4025210 0.015625     0       6          1  92.676843
#> 55:  4.55701408 13.3724605 0.015625     0       6          1 204.140997
#> 56: -3.24871827  2.0401016 0.015625     0       6          1  89.848128
#> 57:  7.70154807 13.8095318 0.015625     0       6          1 344.272476
#> 58:  4.28827300  1.3359072 0.015625     0       6          1   8.594173
#> 59:  0.44969557 12.5389900 0.015625     0       6          1  71.187998
#> 60:  6.20319117 14.2345586 0.015625     0       6          1 305.996971
#> 61:  6.92701013 11.1191282 0.015625     0       6          1 232.843516
#> 62:  3.76354636  1.2647809 0.015625     0       6          1   2.866955
#> 63:  2.35793084  7.4321629 0.015625     0       6          1  28.338478
#> 64: -0.21916136  3.2111773 0.015625     0       6          1  29.226416
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-04-23 04:35:03        1  0.28925509   5.0218566          0.015625
#>  2: 2022-04-23 04:35:03        1 -1.28673097  10.1564200          0.015625
#>  3: 2022-04-23 04:35:03        1 -2.35463244   8.3200312          0.015625
#>  4: 2022-04-23 04:35:03        1  4.29860431   7.4500198          0.015625
#>  5: 2022-04-23 04:35:03        1  4.33501974   9.3872516          0.015625
#>  6: 2022-04-23 04:35:03        1  9.66262841  12.7991530          0.015625
#>  7: 2022-04-23 04:35:03        1  8.11111873   7.3525849          0.015625
#>  8: 2022-04-23 04:35:03        1  7.74368109  10.7783777          0.015625
#>  9: 2022-04-23 04:35:03        1 -0.40333649   8.7589240          0.015625
#> 10: 2022-04-23 04:35:03        1 -2.13948844  14.2048098          0.015625
#> 11: 2022-04-23 04:35:03        1  0.42200079   3.0117822          0.015625
#> 12: 2022-04-23 04:35:03        1 -1.86509692   3.5745919          0.015625
#> 13: 2022-04-23 04:35:03        1  4.07292304   3.8260109          0.015625
#> 14: 2022-04-23 04:35:03        1  6.99965074  14.1991134          0.015625
#> 15: 2022-04-23 04:35:03        1  3.44879310   4.9145757          0.015625
#> 16: 2022-04-23 04:35:03        1 -4.40181426   7.1167892          0.015625
#> 17: 2022-04-23 04:35:03        1 -0.30353281  10.4684620          0.015625
#> 18: 2022-04-23 04:35:03        1  0.69028891   8.8507173          0.015625
#> 19: 2022-04-23 04:35:03        1  0.07316201   2.4703046          0.015625
#> 20: 2022-04-23 04:35:03        1  5.40570900   2.6716965          0.015625
#> 21: 2022-04-23 04:35:03        1 -0.55144784   0.6174301          0.015625
#> 22: 2022-04-23 04:35:03        1  5.07727433   4.8830921          0.015625
#> 23: 2022-04-23 04:35:03        1  6.11127558   9.4765700          0.015625
#> 24: 2022-04-23 04:35:03        1 -4.46574856   8.2011965          0.015625
#> 25: 2022-04-23 04:35:03        1  1.46365728  12.7953313          0.015625
#> 26: 2022-04-23 04:35:03        1 -0.51323411  12.8165289          0.015625
#> 27: 2022-04-23 04:35:03        1 -4.78341828  14.2010999          0.015625
#> 28: 2022-04-23 04:35:03        1  8.34112788  10.2390849          0.015625
#> 29: 2022-04-23 04:35:03        1  3.94430652   1.2343369          0.015625
#> 30: 2022-04-23 04:35:03        1 -2.21650988  14.4005366          0.015625
#> 31: 2022-04-23 04:35:03        1  3.16567919  11.3696671          0.015625
#> 32: 2022-04-23 04:35:03        1 -0.37755148   3.0023172          0.015625
#> 33: 2022-04-23 04:35:03        1  0.83392610  13.2311509          0.015625
#> 34: 2022-04-23 04:35:03        1 -3.82091893   4.9505822          0.015625
#> 35: 2022-04-23 04:35:03        1 -2.62687125   5.6673378          0.015625
#> 36: 2022-04-23 04:35:03        1  2.35883959   4.1247207          0.015625
#> 37: 2022-04-23 04:35:03        1 -4.43879808   6.1074262          0.015625
#> 38: 2022-04-23 04:35:03        1 -4.68872403   8.0183362          0.015625
#> 39: 2022-04-23 04:35:03        1  9.42061214   1.9201838          0.015625
#> 40: 2022-04-23 04:35:03        1 -4.98775828   3.3129204          0.015625
#> 41: 2022-04-23 04:35:03        1 -1.38631213   3.9545366          0.015625
#> 42: 2022-04-23 04:35:03        1 -3.40605465   2.0340485          0.015625
#> 43: 2022-04-23 04:35:03        1  5.65810524   5.9452656          0.015625
#> 44: 2022-04-23 04:35:03        1  0.94291847   0.9739030          0.015625
#> 45: 2022-04-23 04:35:03        1 -4.76545934   7.7239694          0.015625
#> 46: 2022-04-23 04:35:03        1  9.04679996  13.1240762          0.015625
#> 47: 2022-04-23 04:35:03        1  0.20525005  14.5423960          0.015625
#> 48: 2022-04-23 04:35:03        1  0.94178151  12.1878109          0.015625
#> 49: 2022-04-23 04:35:03        1  3.56154839   8.5045159          0.015625
#> 50: 2022-04-23 04:35:03        1  4.74660689  10.7079228          0.015625
#> 51: 2022-04-23 04:35:03        1 -1.35813925   1.4784836          0.015625
#> 52: 2022-04-23 04:35:03        1  6.86494170   3.7783814          0.015625
#> 53: 2022-04-23 04:35:03        1 -0.80022394  13.6698552          0.015625
#> 54: 2022-04-23 04:35:03        1 -3.51501491   2.4025210          0.015625
#> 55: 2022-04-23 04:35:03        1  4.55701408  13.3724605          0.015625
#> 56: 2022-04-23 04:35:03        1 -3.24871827   2.0401016          0.015625
#> 57: 2022-04-23 04:35:03        1  7.70154807  13.8095318          0.015625
#> 58: 2022-04-23 04:35:03        1  4.28827300   1.3359072          0.015625
#> 59: 2022-04-23 04:35:03        1  0.44969557  12.5389900          0.015625
#> 60: 2022-04-23 04:35:03        1  6.20319117  14.2345586          0.015625
#> 61: 2022-04-23 04:35:03        1  6.92701013  11.1191282          0.015625
#> 62: 2022-04-23 04:35:03        1  3.76354636   1.2647809          0.015625
#> 63: 2022-04-23 04:35:03        1  2.35793084   7.4321629          0.015625
#> 64: 2022-04-23 04:35:03        1 -0.21916136   3.2111773          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity