Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain         y
#> 1: 3.364516 0.5752015 0.015625 <list[3]> 0.8102631

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain         y
#> 1: 3.364516 0.5752015 0.015625 <list[3]> 0.8102631

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -0.2546134 12.7637554 0.015625     0       6          1  59.6980517
#>  2:  8.0428088  4.6081557 0.015625     0       6          1  96.9294695
#>  3: -1.3172329  2.1637732 0.015625     0       6          1  48.2411408
#>  4: -2.8028307  2.7245399 0.015625     0       6          1  64.5896615
#>  5: -0.2290069 11.4144314 0.015625     0       6          1  44.8371908
#>  6:  3.6360931 10.0225650 0.015625     0       6          1  89.9659908
#>  7:  3.7182271  2.5549497 0.015625     0       6          1   6.1432252
#>  8: -0.5308017  7.7231486 0.015625     0       6          1  19.0372345
#>  9:  7.1483936  9.5093109 0.015625     0       6          1 193.5206763
#> 10:  3.6777690 12.9271095 0.015625     0       6          1 154.6281360
#> 11: -3.1076032  5.5985139 0.015625     0       6          1  32.2617825
#> 12:  7.5552388 13.8909511 0.015625     0       6          1 342.6261254
#> 13: -2.8211646 14.8217625 0.015625     0       6          1  17.5904029
#> 14: -3.6053683 11.5578240 0.015625     0       6          1   1.7485227
#> 15:  9.3303801  6.6052678 0.015625     0       6          1 163.7260371
#> 16: -4.2897761 14.2194232 0.015625     0       6          1   6.7443830
#> 17:  8.3140500 10.7145730 0.015625     0       6          1 256.0550402
#> 18:  3.8821512 11.5297457 0.015625     0       6          1 129.3625613
#> 19:  9.5842112  0.2939958 0.015625     0       6          1  45.7239153
#> 20:  2.0402948  0.1336936 0.015625     0       6          1  13.2020233
#> 21: -0.3993149  0.6774856 0.015625     0       6          1  54.4034226
#> 22:  0.5327688 11.6449574 0.015625     0       6          1  60.3156034
#> 23:  0.1546495 10.3721256 0.015625     0       6          1  40.8090317
#> 24:  6.3376306  7.1540869 0.015625     0       6          1 119.7032851
#> 25: -1.8452224 13.9212608 0.015625     0       6          1  31.2104191
#> 26:  3.6829403  6.8742884 0.015625     0       6          1  41.6982820
#> 27:  7.0527896 10.9609040 0.015625     0       6          1 231.7048290
#> 28:  8.2651416  8.6812857 0.015625     0       6          1 194.8196645
#> 29:  4.8846326  2.2880499 0.015625     0       6          1  22.7251264
#> 30: -0.6159893  8.5277192 0.015625     0       6          1  20.1955640
#> 31: -0.4142422 13.0333222 0.015625     0       6          1  59.3510761
#> 32: -1.0173792  7.8202533 0.015625     0       6          1  15.0754805
#> 33: -1.5423926 11.3765338 0.015625     0       6          1  18.3871841
#> 34:  0.7682305  1.2298504 0.015625     0       6          1  29.6189149
#> 35: -0.3010423  6.3571622 0.015625     0       6          1  19.1858479
#> 36: -3.9188596  8.5581325 0.015625     0       6          1  20.3871130
#> 37: -4.8844680  1.2760260 0.015625     0       6          1 186.7143718
#> 38:  5.4725670 10.0928909 0.015625     0       6          1 157.7961219
#> 39: -2.6935016  5.2740729 0.015625     0       6          1  28.7598284
#> 40:  8.6230752  5.0612770 0.015625     0       6          1 113.5521218
#> 41:  6.8382970  4.3323710 0.015625     0       6          1  78.6584565
#> 42:  3.3645162  0.5752015 0.015625     0       6          1   0.8102631
#> 43:  7.9918750 12.4386528 0.015625     0       6          1 304.3241945
#> 44:  8.6628235  0.5692522 0.015625     0       6          1  39.6454398
#> 45:  8.3925416  1.1794870 0.015625     0       6          1  45.6643798
#> 46:  6.9262997 10.7467774 0.015625     0       6          1 222.0389525
#> 47:  3.7488257  6.9565652 0.015625     0       6          1  44.2462079
#> 48:  1.6299132  6.3242362 0.015625     0       6          1  17.4792136
#> 49: -0.2193541  6.2299483 0.015625     0       6          1  19.3865838
#> 50: -2.7076684  2.1456394 0.015625     0       6          1  71.6659490
#> 51: -3.8035829 12.0512873 0.015625     0       6          1   2.6260604
#> 52:  5.0192425 13.6192434 0.015625     0       6          1 232.9186888
#> 53: -4.7203773 12.3224516 0.015625     0       6          1  13.5937011
#> 54:  1.6365984  8.7368313 0.015625     0       6          1  37.0278447
#> 55:  6.4448269  6.1626470 0.015625     0       6          1 103.0678478
#> 56:  8.7393326 11.8385928 0.015625     0       6          1 305.3037189
#> 57:  1.6435002  6.9170162 0.015625     0       6          1  21.2027788
#> 58:  1.9974595 10.0371565 0.015625     0       6          1  56.3445621
#> 59:  0.2424904  2.2303029 0.015625     0       6          1  30.7832653
#> 60:  5.2839705  2.6121128 0.015625     0       6          1  32.5278360
#> 61: -2.9470144  3.1354982 0.015625     0       6          1  61.7605090
#> 62:  3.0753377 13.6064256 0.015625     0       6          1 149.5074876
#> 63: -1.6330432 13.7354562 0.015625     0       6          1  34.9495446
#> 64: -0.7728147  9.0044681 0.015625     0       6          1  19.9586169
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-03-26 04:31:40        1  -0.2546134  12.7637554          0.015625
#>  2: 2022-03-26 04:31:40        1   8.0428088   4.6081557          0.015625
#>  3: 2022-03-26 04:31:40        1  -1.3172329   2.1637732          0.015625
#>  4: 2022-03-26 04:31:40        1  -2.8028307   2.7245399          0.015625
#>  5: 2022-03-26 04:31:40        1  -0.2290069  11.4144314          0.015625
#>  6: 2022-03-26 04:31:40        1   3.6360931  10.0225650          0.015625
#>  7: 2022-03-26 04:31:40        1   3.7182271   2.5549497          0.015625
#>  8: 2022-03-26 04:31:40        1  -0.5308017   7.7231486          0.015625
#>  9: 2022-03-26 04:31:40        1   7.1483936   9.5093109          0.015625
#> 10: 2022-03-26 04:31:40        1   3.6777690  12.9271095          0.015625
#> 11: 2022-03-26 04:31:40        1  -3.1076032   5.5985139          0.015625
#> 12: 2022-03-26 04:31:40        1   7.5552388  13.8909511          0.015625
#> 13: 2022-03-26 04:31:40        1  -2.8211646  14.8217625          0.015625
#> 14: 2022-03-26 04:31:40        1  -3.6053683  11.5578240          0.015625
#> 15: 2022-03-26 04:31:40        1   9.3303801   6.6052678          0.015625
#> 16: 2022-03-26 04:31:40        1  -4.2897761  14.2194232          0.015625
#> 17: 2022-03-26 04:31:40        1   8.3140500  10.7145730          0.015625
#> 18: 2022-03-26 04:31:40        1   3.8821512  11.5297457          0.015625
#> 19: 2022-03-26 04:31:40        1   9.5842112   0.2939958          0.015625
#> 20: 2022-03-26 04:31:40        1   2.0402948   0.1336936          0.015625
#> 21: 2022-03-26 04:31:40        1  -0.3993149   0.6774856          0.015625
#> 22: 2022-03-26 04:31:40        1   0.5327688  11.6449574          0.015625
#> 23: 2022-03-26 04:31:40        1   0.1546495  10.3721256          0.015625
#> 24: 2022-03-26 04:31:40        1   6.3376306   7.1540869          0.015625
#> 25: 2022-03-26 04:31:40        1  -1.8452224  13.9212608          0.015625
#> 26: 2022-03-26 04:31:40        1   3.6829403   6.8742884          0.015625
#> 27: 2022-03-26 04:31:40        1   7.0527896  10.9609040          0.015625
#> 28: 2022-03-26 04:31:40        1   8.2651416   8.6812857          0.015625
#> 29: 2022-03-26 04:31:40        1   4.8846326   2.2880499          0.015625
#> 30: 2022-03-26 04:31:40        1  -0.6159893   8.5277192          0.015625
#> 31: 2022-03-26 04:31:40        1  -0.4142422  13.0333222          0.015625
#> 32: 2022-03-26 04:31:40        1  -1.0173792   7.8202533          0.015625
#> 33: 2022-03-26 04:31:40        1  -1.5423926  11.3765338          0.015625
#> 34: 2022-03-26 04:31:40        1   0.7682305   1.2298504          0.015625
#> 35: 2022-03-26 04:31:40        1  -0.3010423   6.3571622          0.015625
#> 36: 2022-03-26 04:31:40        1  -3.9188596   8.5581325          0.015625
#> 37: 2022-03-26 04:31:40        1  -4.8844680   1.2760260          0.015625
#> 38: 2022-03-26 04:31:40        1   5.4725670  10.0928909          0.015625
#> 39: 2022-03-26 04:31:40        1  -2.6935016   5.2740729          0.015625
#> 40: 2022-03-26 04:31:40        1   8.6230752   5.0612770          0.015625
#> 41: 2022-03-26 04:31:40        1   6.8382970   4.3323710          0.015625
#> 42: 2022-03-26 04:31:40        1   3.3645162   0.5752015          0.015625
#> 43: 2022-03-26 04:31:40        1   7.9918750  12.4386528          0.015625
#> 44: 2022-03-26 04:31:40        1   8.6628235   0.5692522          0.015625
#> 45: 2022-03-26 04:31:40        1   8.3925416   1.1794870          0.015625
#> 46: 2022-03-26 04:31:40        1   6.9262997  10.7467774          0.015625
#> 47: 2022-03-26 04:31:40        1   3.7488257   6.9565652          0.015625
#> 48: 2022-03-26 04:31:40        1   1.6299132   6.3242362          0.015625
#> 49: 2022-03-26 04:31:40        1  -0.2193541   6.2299483          0.015625
#> 50: 2022-03-26 04:31:40        1  -2.7076684   2.1456394          0.015625
#> 51: 2022-03-26 04:31:40        1  -3.8035829  12.0512873          0.015625
#> 52: 2022-03-26 04:31:40        1   5.0192425  13.6192434          0.015625
#> 53: 2022-03-26 04:31:40        1  -4.7203773  12.3224516          0.015625
#> 54: 2022-03-26 04:31:40        1   1.6365984   8.7368313          0.015625
#> 55: 2022-03-26 04:31:40        1   6.4448269   6.1626470          0.015625
#> 56: 2022-03-26 04:31:40        1   8.7393326  11.8385928          0.015625
#> 57: 2022-03-26 04:31:40        1   1.6435002   6.9170162          0.015625
#> 58: 2022-03-26 04:31:40        1   1.9974595  10.0371565          0.015625
#> 59: 2022-03-26 04:31:40        1   0.2424904   2.2303029          0.015625
#> 60: 2022-03-26 04:31:40        1   5.2839705   2.6121128          0.015625
#> 61: 2022-03-26 04:31:40        1  -2.9470144   3.1354982          0.015625
#> 62: 2022-03-26 04:31:40        1   3.0753377  13.6064256          0.015625
#> 63: 2022-03-26 04:31:40        1  -1.6330432  13.7354562          0.015625
#> 64: 2022-03-26 04:31:40        1  -0.7728147   9.0044681          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity