Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1      x2 fidelity  x_domain       y
#> 1: 2.561666 2.43004 0.015625 <list[3]> 2.06101

# best scoring evaluation
instance$result
#>          x1      x2 fidelity  x_domain       y
#> 1: 2.561666 2.43004 0.015625 <list[3]> 2.06101

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  6.33889293  3.7947270 0.015625     0       6          1  63.782132
#>  2:  5.98433866 13.7678441 0.015625     0       6          1 281.328025
#>  3:  2.34564702  7.7783319 0.015625     0       6          1  31.823223
#>  4: -0.42350797 12.0287691 0.015625     0       6          1  47.367953
#>  5:  3.22561512  2.6000631 0.015625     0       6          1   2.430815
#>  6: -0.74401382  0.2270789 0.015625     0       6          1  65.702550
#>  7:  3.48036265  2.6278672 0.015625     0       6          1   4.164221
#>  8:  4.86225232  3.4878689 0.015625     0       6          1  31.678931
#>  9:  1.08995095 10.2079369 0.015625     0       6          1  49.323514
#> 10: -3.95419550  6.1855097 0.015625     0       6          1  46.806313
#> 11:  3.26919322  6.6205328 0.015625     0       6          1  30.671009
#> 12: -2.40176234  2.6449013 0.015625     0       6          1  57.003850
#> 13:  1.59469402  0.1056790 0.015625     0       6          1  21.566191
#> 14: -3.23497417  3.9558094 0.015625     0       6          1  56.908652
#> 15: -3.88039720 11.3691292 0.015625     0       6          1   4.513513
#> 16: -2.84366664  3.8144472 0.015625     0       6          1  49.262802
#> 17:  0.39316334 14.5945009 0.015625     0       6          1 103.794688
#> 18:  1.21374177  1.0628045 0.015625     0       6          1  22.663235
#> 19:  4.51870763  9.5340096 0.015625     0       6          1 110.120280
#> 20: -2.75906532  5.0686876 0.015625     0       6          1  31.967258
#> 21: -0.75551150  9.5394406 0.015625     0       6          1  22.369457
#> 22:  2.07296793 14.8063047 0.015625     0       6          1 148.740754
#> 23:  4.75739256 13.7954966 0.015625     0       6          1 225.677021
#> 24: -3.17506547  1.8042995 0.015625     0       6          1  91.776524
#> 25: -2.15316423  9.4801436 0.015625     0       6          1   4.726778
#> 26:  0.62602228  5.7502612 0.015625     0       6          1  18.320785
#> 27: -1.26501297 11.6337878 0.015625     0       6          1  25.644478
#> 28:  4.38362433  0.6803075 0.015625     0       6          1   8.036558
#> 29:  1.07333258  8.8733415 0.015625     0       6          1  35.249877
#> 30:  7.41265837  3.8524950 0.015625     0       6          1  77.473133
#> 31:  5.91417955  3.6010315 0.015625     0       6          1  54.219006
#> 32:  5.54749601  3.4948282 0.015625     0       6          1  46.038488
#> 33: -0.43654195 11.2374701 0.015625     0       6          1  39.284469
#> 34:  5.19404076  7.5425998 0.015625     0       6          1  95.082731
#> 35: -0.44834860  7.4454016 0.015625     0       6          1  19.179651
#> 36: -3.79426807 12.8550242 0.015625     0       6          1   2.511073
#> 37:  9.50774367  3.6220125 0.015625     0       6          1  99.923826
#> 38:  0.61854318 11.2820314 0.015625     0       6          1  56.944466
#> 39: -1.49575258  7.5341665 0.015625     0       6          1  11.557470
#> 40:  6.17779751 10.4526985 0.015625     0       6          1 191.460360
#> 41:  9.74415208 10.5938772 0.015625     0       6          1 296.131874
#> 42:  1.38448225  4.2289256 0.015625     0       6          1  11.918150
#> 43: -3.56723389 10.7848142 0.015625     0       6          1   2.902980
#> 44: -0.83978703  9.3436678 0.015625     0       6          1  20.352473
#> 45: -0.83727525 11.6776362 0.015625     0       6          1  35.121309
#> 46: -1.94310308  4.9808676 0.015625     0       6          1  24.381099
#> 47:  2.40992428  7.0564514 0.015625     0       6          1  25.071533
#> 48:  3.94149647 11.5736899 0.015625     0       6          1 132.566130
#> 49:  5.81302687  4.2456741 0.015625     0       6          1  60.271347
#> 50: -3.78239365  4.0633856 0.015625     0       6          1  72.801694
#> 51:  2.87754845  7.3218508 0.015625     0       6          1  32.619479
#> 52: -2.41623389 10.7506532 0.015625     0       6          1   3.341591
#> 53: -0.58076258 10.3768656 0.015625     0       6          1  29.876415
#> 54: -1.20734697  1.5802934 0.015625     0       6          1  54.195553
#> 55:  5.42125144  6.6035878 0.015625     0       6          1  85.608340
#> 56: -4.73998047  2.5308313 0.015625     0       6          1 147.245891
#> 57: -1.93883729  6.8064429 0.015625     0       6          1  12.280786
#> 58: -2.12247285  1.9816455 0.015625     0       6          1  61.741901
#> 59: -4.56302462 12.8033128 0.015625     0       6          1   9.779257
#> 60: -0.08897041 10.4279381 0.015625     0       6          1  37.934735
#> 61:  9.19329908 10.1814793 0.015625     0       6          1 263.561799
#> 62:  6.92826650  5.0692361 0.015625     0       6          1  91.977518
#> 63:  2.56166582  2.4300404 0.015625     0       6          1   2.061010
#> 64:  7.20601801  5.5774573 0.015625     0       6          1 105.090440
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-18 04:36:37        1  6.33889293   3.7947270          0.015625
#>  2: 2022-05-18 04:36:37        1  5.98433866  13.7678441          0.015625
#>  3: 2022-05-18 04:36:37        1  2.34564702   7.7783319          0.015625
#>  4: 2022-05-18 04:36:37        1 -0.42350797  12.0287691          0.015625
#>  5: 2022-05-18 04:36:37        1  3.22561512   2.6000631          0.015625
#>  6: 2022-05-18 04:36:37        1 -0.74401382   0.2270789          0.015625
#>  7: 2022-05-18 04:36:37        1  3.48036265   2.6278672          0.015625
#>  8: 2022-05-18 04:36:37        1  4.86225232   3.4878689          0.015625
#>  9: 2022-05-18 04:36:37        1  1.08995095  10.2079369          0.015625
#> 10: 2022-05-18 04:36:37        1 -3.95419550   6.1855097          0.015625
#> 11: 2022-05-18 04:36:37        1  3.26919322   6.6205328          0.015625
#> 12: 2022-05-18 04:36:37        1 -2.40176234   2.6449013          0.015625
#> 13: 2022-05-18 04:36:37        1  1.59469402   0.1056790          0.015625
#> 14: 2022-05-18 04:36:37        1 -3.23497417   3.9558094          0.015625
#> 15: 2022-05-18 04:36:37        1 -3.88039720  11.3691292          0.015625
#> 16: 2022-05-18 04:36:37        1 -2.84366664   3.8144472          0.015625
#> 17: 2022-05-18 04:36:37        1  0.39316334  14.5945009          0.015625
#> 18: 2022-05-18 04:36:37        1  1.21374177   1.0628045          0.015625
#> 19: 2022-05-18 04:36:37        1  4.51870763   9.5340096          0.015625
#> 20: 2022-05-18 04:36:37        1 -2.75906532   5.0686876          0.015625
#> 21: 2022-05-18 04:36:37        1 -0.75551150   9.5394406          0.015625
#> 22: 2022-05-18 04:36:37        1  2.07296793  14.8063047          0.015625
#> 23: 2022-05-18 04:36:37        1  4.75739256  13.7954966          0.015625
#> 24: 2022-05-18 04:36:37        1 -3.17506547   1.8042995          0.015625
#> 25: 2022-05-18 04:36:37        1 -2.15316423   9.4801436          0.015625
#> 26: 2022-05-18 04:36:37        1  0.62602228   5.7502612          0.015625
#> 27: 2022-05-18 04:36:37        1 -1.26501297  11.6337878          0.015625
#> 28: 2022-05-18 04:36:37        1  4.38362433   0.6803075          0.015625
#> 29: 2022-05-18 04:36:37        1  1.07333258   8.8733415          0.015625
#> 30: 2022-05-18 04:36:37        1  7.41265837   3.8524950          0.015625
#> 31: 2022-05-18 04:36:37        1  5.91417955   3.6010315          0.015625
#> 32: 2022-05-18 04:36:37        1  5.54749601   3.4948282          0.015625
#> 33: 2022-05-18 04:36:37        1 -0.43654195  11.2374701          0.015625
#> 34: 2022-05-18 04:36:37        1  5.19404076   7.5425998          0.015625
#> 35: 2022-05-18 04:36:37        1 -0.44834860   7.4454016          0.015625
#> 36: 2022-05-18 04:36:37        1 -3.79426807  12.8550242          0.015625
#> 37: 2022-05-18 04:36:37        1  9.50774367   3.6220125          0.015625
#> 38: 2022-05-18 04:36:37        1  0.61854318  11.2820314          0.015625
#> 39: 2022-05-18 04:36:37        1 -1.49575258   7.5341665          0.015625
#> 40: 2022-05-18 04:36:37        1  6.17779751  10.4526985          0.015625
#> 41: 2022-05-18 04:36:37        1  9.74415208  10.5938772          0.015625
#> 42: 2022-05-18 04:36:37        1  1.38448225   4.2289256          0.015625
#> 43: 2022-05-18 04:36:37        1 -3.56723389  10.7848142          0.015625
#> 44: 2022-05-18 04:36:37        1 -0.83978703   9.3436678          0.015625
#> 45: 2022-05-18 04:36:37        1 -0.83727525  11.6776362          0.015625
#> 46: 2022-05-18 04:36:37        1 -1.94310308   4.9808676          0.015625
#> 47: 2022-05-18 04:36:37        1  2.40992428   7.0564514          0.015625
#> 48: 2022-05-18 04:36:37        1  3.94149647  11.5736899          0.015625
#> 49: 2022-05-18 04:36:37        1  5.81302687   4.2456741          0.015625
#> 50: 2022-05-18 04:36:37        1 -3.78239365   4.0633856          0.015625
#> 51: 2022-05-18 04:36:37        1  2.87754845   7.3218508          0.015625
#> 52: 2022-05-18 04:36:37        1 -2.41623389  10.7506532          0.015625
#> 53: 2022-05-18 04:36:37        1 -0.58076258  10.3768656          0.015625
#> 54: 2022-05-18 04:36:37        1 -1.20734697   1.5802934          0.015625
#> 55: 2022-05-18 04:36:37        1  5.42125144   6.6035878          0.015625
#> 56: 2022-05-18 04:36:37        1 -4.73998047   2.5308313          0.015625
#> 57: 2022-05-18 04:36:37        1 -1.93883729   6.8064429          0.015625
#> 58: 2022-05-18 04:36:37        1 -2.12247285   1.9816455          0.015625
#> 59: 2022-05-18 04:36:37        1 -4.56302462  12.8033128          0.015625
#> 60: 2022-05-18 04:36:37        1 -0.08897041  10.4279381          0.015625
#> 61: 2022-05-18 04:36:37        1  9.19329908  10.1814793          0.015625
#> 62: 2022-05-18 04:36:37        1  6.92826650   5.0692361          0.015625
#> 63: 2022-05-18 04:36:37        1  2.56166582   2.4300404          0.015625
#> 64: 2022-05-18 04:36:37        1  7.20601801   5.5774573          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity