Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -2.975434 10.44301 0.015625 <list[3]> 0.8490731

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -2.975434 10.44301 0.015625 <list[3]> 0.8490731

# all evaluations
as.data.table(instance$archive)
#>             x1          x2 fidelity stage bracket repetition           y
#>  1:  3.9430840  6.57963902 0.015625     0       6          1  43.9891080
#>  2:  5.4597592 13.51893294 0.015625     0       6          1 250.3680482
#>  3:  0.2424483 12.60590665 0.015625     0       6          1  68.1809307
#>  4:  4.1104968  9.71997572 0.015625     0       6          1  99.4800948
#>  5:  8.0309299 13.26796437 0.015625     0       6          1 334.7093615
#>  6:  9.7612083  7.95114661 0.015625     0       6          1 212.8418748
#>  7:  3.6783413  0.57211648 0.015625     0       6          1   1.7482843
#>  8:  1.6639995  9.45738256 0.015625     0       6          1  45.3537887
#>  9:  3.4976203  2.11454349 0.015625     0       6          1   2.7031775
#> 10:  7.7102438  4.08644132 0.015625     0       6          1  84.1334662
#> 11:  3.3513691  5.77370818 0.015625     0       6          1  23.2872944
#> 12:  0.4948312  9.99211728 0.015625     0       6          1  41.2236200
#> 13:  7.5555652 11.30229982 0.015625     0       6          1 255.3138274
#> 14:  2.2469262  3.63802577 0.015625     0       6          1   5.1124437
#> 15: -4.7522632  8.63513327 0.015625     0       6          1  41.9977455
#> 16:  6.6683636  2.92025775 0.015625     0       6          1  56.9188577
#> 17:  7.9605811  1.53852291 0.015625     0       6          1  48.1624152
#> 18: -3.9005428  1.68268233 0.015625     0       6          1 123.8795128
#> 19: -1.3726882  6.90432280 0.015625     0       6          1  13.6809226
#> 20:  0.2683068 14.70023876 0.015625     0       6          1 102.5250773
#> 21:  8.3462589 10.08005240 0.015625     0       6          1 237.1618124
#> 22: -2.9754341 10.44301271 0.015625     0       6          1   0.8490731
#> 23: -3.2922961  0.07404527 0.015625     0       6          1 132.7355626
#> 24:  7.1884680  0.38763823 0.015625     0       6          1  33.9033056
#> 25:  9.6069691 11.98285247 0.015625     0       6          1 340.4081975
#> 26:  1.5498718 14.00102973 0.015625     0       6          1 118.2334358
#> 27:  8.2678815  2.85911360 0.015625     0       6          1  68.8021338
#> 28:  4.3033952 13.66462760 0.015625     0       6          1 200.6244105
#> 29:  6.8686666  5.78397094 0.015625     0       6          1 103.8466478
#> 30:  7.9866583  5.68346141 0.015625     0       6          1 117.5850152
#> 31: -0.7036304  0.72232855 0.015625     0       6          1  58.4450435
#> 32:  3.1995942  3.22882108 0.015625     0       6          1   4.4395306
#> 33: -0.1809169 14.97852650 0.015625     0       6          1  94.9542334
#> 34:  8.8768144  3.76120727 0.015625     0       6          1  91.4145669
#> 35:  9.6328856 14.11155846 0.015625     0       6          1 424.5400167
#> 36:  9.7929426 12.10898239 0.015625     0       6          1 352.4629734
#> 37:  8.6545471  1.52189766 0.015625     0       6          1  52.0108290
#> 38:  1.7666684  0.11861463 0.015625     0       6          1  18.1523065
#> 39:  5.0210319 10.47274398 0.015625     0       6          1 149.5450486
#> 40: -2.8545431  7.05071000 0.015625     0       6          1  14.8006419
#> 41: -4.2618214  2.12342201 0.015625     0       6          1 131.6606607
#> 42:  7.3810167 14.52786070 0.015625     0       6          1 360.3353237
#> 43:  4.3341540  2.63321815 0.015625     0       6          1  15.1782273
#> 44:  9.4432471  8.38366443 0.015625     0       6          1 215.6435485
#> 45:  8.1156928 14.60263735 0.015625     0       6          1 387.5327420
#> 46:  3.3582731  0.49163502 0.015625     0       6          1   0.8827959
#> 47: -2.5006680  2.94004351 0.015625     0       6          1  54.6076978
#> 48:  8.3705733  6.07565465 0.015625     0       6          1 131.6736355
#> 49:  3.4649275  9.42566697 0.015625     0       6          1  74.3597318
#> 50:  6.4933709 10.82846650 0.015625     0       6          1 211.6726791
#> 51: -1.7612056  5.09612781 0.015625     0       6          1  22.6401091
#> 52:  7.6518701  1.07509282 0.015625     0       6          1  41.6643639
#> 53:  9.6468823 10.88351859 0.015625     0       6          1 302.5459270
#> 54: -1.0346889 13.77879768 0.015625     0       6          1  52.1039853
#> 55:  0.8072024  6.76143616 0.015625     0       6          1  20.7451723
#> 56:  9.4119565  1.63224040 0.015625     0       6          1  62.6210024
#> 57: -0.7611227  9.54371130 0.015625     0       6          1  22.3095978
#> 58:  7.1085047 14.43627176 0.015625     0       6          1 347.6127014
#> 59:  4.2820834  6.32301070 0.015625     0       6          1  49.2168806
#> 60:  0.1776854 12.69183366 0.015625     0       6          1  68.0828373
#> 61:  9.6593403  0.75091628 0.015625     0       6          1  53.3023856
#> 62: -1.0752405  2.44365388 0.015625     0       6          1  42.6898536
#> 63:  1.9850653 14.44490065 0.015625     0       6          1 137.9958941
#> 64: -0.8696637 14.33991836 0.015625     0       6          1  64.2544318
#>             x1          x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-23 04:36:14        1   3.9430840  6.57963902          0.015625
#>  2: 2022-05-23 04:36:14        1   5.4597592 13.51893294          0.015625
#>  3: 2022-05-23 04:36:14        1   0.2424483 12.60590665          0.015625
#>  4: 2022-05-23 04:36:14        1   4.1104968  9.71997572          0.015625
#>  5: 2022-05-23 04:36:14        1   8.0309299 13.26796437          0.015625
#>  6: 2022-05-23 04:36:14        1   9.7612083  7.95114661          0.015625
#>  7: 2022-05-23 04:36:14        1   3.6783413  0.57211648          0.015625
#>  8: 2022-05-23 04:36:14        1   1.6639995  9.45738256          0.015625
#>  9: 2022-05-23 04:36:14        1   3.4976203  2.11454349          0.015625
#> 10: 2022-05-23 04:36:14        1   7.7102438  4.08644132          0.015625
#> 11: 2022-05-23 04:36:14        1   3.3513691  5.77370818          0.015625
#> 12: 2022-05-23 04:36:14        1   0.4948312  9.99211728          0.015625
#> 13: 2022-05-23 04:36:14        1   7.5555652 11.30229982          0.015625
#> 14: 2022-05-23 04:36:14        1   2.2469262  3.63802577          0.015625
#> 15: 2022-05-23 04:36:14        1  -4.7522632  8.63513327          0.015625
#> 16: 2022-05-23 04:36:14        1   6.6683636  2.92025775          0.015625
#> 17: 2022-05-23 04:36:14        1   7.9605811  1.53852291          0.015625
#> 18: 2022-05-23 04:36:14        1  -3.9005428  1.68268233          0.015625
#> 19: 2022-05-23 04:36:14        1  -1.3726882  6.90432280          0.015625
#> 20: 2022-05-23 04:36:14        1   0.2683068 14.70023876          0.015625
#> 21: 2022-05-23 04:36:14        1   8.3462589 10.08005240          0.015625
#> 22: 2022-05-23 04:36:14        1  -2.9754341 10.44301271          0.015625
#> 23: 2022-05-23 04:36:14        1  -3.2922961  0.07404527          0.015625
#> 24: 2022-05-23 04:36:14        1   7.1884680  0.38763823          0.015625
#> 25: 2022-05-23 04:36:14        1   9.6069691 11.98285247          0.015625
#> 26: 2022-05-23 04:36:14        1   1.5498718 14.00102973          0.015625
#> 27: 2022-05-23 04:36:14        1   8.2678815  2.85911360          0.015625
#> 28: 2022-05-23 04:36:14        1   4.3033952 13.66462760          0.015625
#> 29: 2022-05-23 04:36:14        1   6.8686666  5.78397094          0.015625
#> 30: 2022-05-23 04:36:14        1   7.9866583  5.68346141          0.015625
#> 31: 2022-05-23 04:36:14        1  -0.7036304  0.72232855          0.015625
#> 32: 2022-05-23 04:36:14        1   3.1995942  3.22882108          0.015625
#> 33: 2022-05-23 04:36:14        1  -0.1809169 14.97852650          0.015625
#> 34: 2022-05-23 04:36:14        1   8.8768144  3.76120727          0.015625
#> 35: 2022-05-23 04:36:14        1   9.6328856 14.11155846          0.015625
#> 36: 2022-05-23 04:36:14        1   9.7929426 12.10898239          0.015625
#> 37: 2022-05-23 04:36:14        1   8.6545471  1.52189766          0.015625
#> 38: 2022-05-23 04:36:14        1   1.7666684  0.11861463          0.015625
#> 39: 2022-05-23 04:36:14        1   5.0210319 10.47274398          0.015625
#> 40: 2022-05-23 04:36:14        1  -2.8545431  7.05071000          0.015625
#> 41: 2022-05-23 04:36:14        1  -4.2618214  2.12342201          0.015625
#> 42: 2022-05-23 04:36:14        1   7.3810167 14.52786070          0.015625
#> 43: 2022-05-23 04:36:14        1   4.3341540  2.63321815          0.015625
#> 44: 2022-05-23 04:36:14        1   9.4432471  8.38366443          0.015625
#> 45: 2022-05-23 04:36:14        1   8.1156928 14.60263735          0.015625
#> 46: 2022-05-23 04:36:14        1   3.3582731  0.49163502          0.015625
#> 47: 2022-05-23 04:36:14        1  -2.5006680  2.94004351          0.015625
#> 48: 2022-05-23 04:36:14        1   8.3705733  6.07565465          0.015625
#> 49: 2022-05-23 04:36:14        1   3.4649275  9.42566697          0.015625
#> 50: 2022-05-23 04:36:14        1   6.4933709 10.82846650          0.015625
#> 51: 2022-05-23 04:36:14        1  -1.7612056  5.09612781          0.015625
#> 52: 2022-05-23 04:36:14        1   7.6518701  1.07509282          0.015625
#> 53: 2022-05-23 04:36:14        1   9.6468823 10.88351859          0.015625
#> 54: 2022-05-23 04:36:14        1  -1.0346889 13.77879768          0.015625
#> 55: 2022-05-23 04:36:14        1   0.8072024  6.76143616          0.015625
#> 56: 2022-05-23 04:36:14        1   9.4119565  1.63224040          0.015625
#> 57: 2022-05-23 04:36:14        1  -0.7611227  9.54371130          0.015625
#> 58: 2022-05-23 04:36:14        1   7.1085047 14.43627176          0.015625
#> 59: 2022-05-23 04:36:14        1   4.2820834  6.32301070          0.015625
#> 60: 2022-05-23 04:36:14        1   0.1776854 12.69183366          0.015625
#> 61: 2022-05-23 04:36:14        1   9.6593403  0.75091628          0.015625
#> 62: 2022-05-23 04:36:14        1  -1.0752405  2.44365388          0.015625
#> 63: 2022-05-23 04:36:14        1   1.9850653 14.44490065          0.015625
#> 64: 2022-05-23 04:36:14        1  -0.8696637 14.33991836          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity