Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: -2.87793 11.02301 0.015625 <list[3]> 0.7650572

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: -2.87793 11.02301 0.015625 <list[3]> 0.7650572

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition           y
#>  1: -1.77917822 12.9730705 0.015625     0       6          1  24.3682189
#>  2: -1.64444236  4.1450639 0.015625     0       6          1  30.0441773
#>  3: -1.51035245 11.7971446 0.015625     0       6          1  21.6237274
#>  4:  1.22284618 14.6904918 0.015625     0       6          1 125.4376995
#>  5:  0.48202290  3.5086667 0.015625     0       6          1  21.5054862
#>  6:  9.94468609  1.2226644 0.015625     0       6          1  65.8162506
#>  7:  9.92028444  3.0651813 0.015625     0       6          1  98.1409274
#>  8:  1.27555536  6.4643650 0.015625     0       6          1  18.7692553
#>  9:  3.75747733 14.4691878 0.015625     0       6          1 198.5905876
#> 10:  0.04724336 12.3923417 0.015625     0       6          1  61.4194799
#> 11:  5.75231131 10.2553950 0.015625     0       6          1 171.8692634
#> 12:  8.78670085 11.3057466 0.015625     0       6          1 288.4503519
#> 13:  4.28347736  2.2584302 0.015625     0       6          1  12.3150195
#> 14:  2.20359644  1.7362233 0.015625     0       6          1   5.1420016
#> 15: -3.75178268 12.5646397 0.015625     0       6          1   2.1565177
#> 16:  2.17315513 14.5174120 0.015625     0       6          1 144.5295168
#> 17:  7.47677231  1.3096681 0.015625     0       6          1  43.6822867
#> 18:  4.68017626  2.9832340 0.015625     0       6          1  23.8169290
#> 19:  5.57221038 13.7388469 0.015625     0       6          1 262.2801267
#> 20:  1.86543657  5.9015350 0.015625     0       6          1  14.8483794
#> 21:  6.29955027  0.4677386 0.015625     0       6          1  30.3173471
#> 22:  6.02304532  6.0046944 0.015625     0       6          1  91.1090235
#> 23:  1.59402546  6.3054325 0.015625     0       6          1  17.4181999
#> 24:  5.57632187 12.5005485 0.015625     0       6          1 225.2220350
#> 25: -3.89500740  5.5174656 0.015625     0       6          1  54.0919171
#> 26: -4.49820812  6.5996555 0.015625     0       6          1  59.5344244
#> 27:  4.69934762  5.3307414 0.015625     0       6          1  47.4636031
#> 28:  5.65293996 13.9736709 0.015625     0       6          1 273.3754808
#> 29:  1.84566676  8.2702976 0.015625     0       6          1  33.4346698
#> 30: -2.87792954 11.0230125 0.015625     0       6          1   0.7650572
#> 31: -4.27134303 10.9033424 0.015625     0       6          1  11.9313168
#> 32: -2.26816154 14.6268743 0.015625     0       6          1  27.4414912
#> 33:  7.42421375  5.4761648 0.015625     0       6          1 106.1113828
#> 34:  5.53288442  3.5889631 0.015625     0       6          1  46.7652129
#> 35:  3.83906374 12.5258022 0.015625     0       6          1 151.0584540
#> 36:  3.85984794  7.1596251 0.015625     0       6          1  49.6197113
#> 37:  3.07252010  6.9609046 0.015625     0       6          1  31.3422660
#> 38:  5.77454932  0.6148906 0.015625     0       6          1  26.1155295
#> 39: -3.71199627  4.4484916 0.015625     0       6          1  64.0596646
#> 40: -1.19745479 12.3913178 0.015625     0       6          1  33.2283898
#> 41: -1.36836608 14.5885314 0.015625     0       6          1  52.2928682
#> 42:  0.15149735  9.6482884 0.015625     0       6          1  34.6141061
#> 43: -3.63688178  6.6500963 0.015625     0       6          1  32.2973292
#> 44: -0.61972068  6.0333237 0.015625     0       6          1  18.7473515
#> 45: -1.55399850 12.0408504 0.015625     0       6          1  22.3646674
#> 46: -1.34831498 11.4830634 0.015625     0       6          1  22.8853570
#> 47:  6.91315514  0.2886759 0.015625     0       6          1  32.3654928
#> 48: -0.52022454  4.1542157 0.015625     0       6          1  25.5253176
#> 49:  9.51391690 14.6315929 0.015625     0       6          1 441.0332467
#> 50: -3.10584820 14.9553310 0.015625     0       6          1  14.2098932
#> 51:  0.62450504  6.8467870 0.015625     0       6          1  21.1339781
#> 52:  5.34679116  7.9204414 0.015625     0       6          1 106.9150834
#> 53:  6.25133017  8.8610700 0.015625     0       6          1 154.3615855
#> 54:  8.47626007 14.4331052 0.015625     0       6          1 393.0610911
#> 55:  8.37682750  0.8528871 0.015625     0       6          1  41.5355579
#> 56:  1.27116414  4.3520110 0.015625     0       6          1  12.9401609
#> 57:  4.64027132 11.9199097 0.015625     0       6          1 169.1556562
#> 58: -0.13161716  2.1285396 0.015625     0       6          1  36.1774463
#> 59: -0.84732960 14.7144559 0.015625     0       6          1  70.2880727
#> 60:  2.99271824  2.4325370 0.015625     0       6          1   1.3508946
#> 61: -1.33213331  0.7057929 0.015625     0       6          1  68.0548321
#> 62: -4.64917274  4.9566712 0.015625     0       6          1  92.3364484
#> 63:  4.76260079  9.4508474 0.015625     0       6          1 117.2599989
#> 64:  3.18534887  2.0284372 0.015625     0       6          1   1.0250382
#>              x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-06-17 04:40:30        1 -1.77917822  12.9730705          0.015625
#>  2: 2022-06-17 04:40:30        1 -1.64444236   4.1450639          0.015625
#>  3: 2022-06-17 04:40:30        1 -1.51035245  11.7971446          0.015625
#>  4: 2022-06-17 04:40:30        1  1.22284618  14.6904918          0.015625
#>  5: 2022-06-17 04:40:30        1  0.48202290   3.5086667          0.015625
#>  6: 2022-06-17 04:40:30        1  9.94468609   1.2226644          0.015625
#>  7: 2022-06-17 04:40:30        1  9.92028444   3.0651813          0.015625
#>  8: 2022-06-17 04:40:30        1  1.27555536   6.4643650          0.015625
#>  9: 2022-06-17 04:40:30        1  3.75747733  14.4691878          0.015625
#> 10: 2022-06-17 04:40:30        1  0.04724336  12.3923417          0.015625
#> 11: 2022-06-17 04:40:30        1  5.75231131  10.2553950          0.015625
#> 12: 2022-06-17 04:40:30        1  8.78670085  11.3057466          0.015625
#> 13: 2022-06-17 04:40:30        1  4.28347736   2.2584302          0.015625
#> 14: 2022-06-17 04:40:30        1  2.20359644   1.7362233          0.015625
#> 15: 2022-06-17 04:40:30        1 -3.75178268  12.5646397          0.015625
#> 16: 2022-06-17 04:40:30        1  2.17315513  14.5174120          0.015625
#> 17: 2022-06-17 04:40:30        1  7.47677231   1.3096681          0.015625
#> 18: 2022-06-17 04:40:30        1  4.68017626   2.9832340          0.015625
#> 19: 2022-06-17 04:40:30        1  5.57221038  13.7388469          0.015625
#> 20: 2022-06-17 04:40:30        1  1.86543657   5.9015350          0.015625
#> 21: 2022-06-17 04:40:30        1  6.29955027   0.4677386          0.015625
#> 22: 2022-06-17 04:40:30        1  6.02304532   6.0046944          0.015625
#> 23: 2022-06-17 04:40:30        1  1.59402546   6.3054325          0.015625
#> 24: 2022-06-17 04:40:30        1  5.57632187  12.5005485          0.015625
#> 25: 2022-06-17 04:40:30        1 -3.89500740   5.5174656          0.015625
#> 26: 2022-06-17 04:40:30        1 -4.49820812   6.5996555          0.015625
#> 27: 2022-06-17 04:40:30        1  4.69934762   5.3307414          0.015625
#> 28: 2022-06-17 04:40:30        1  5.65293996  13.9736709          0.015625
#> 29: 2022-06-17 04:40:30        1  1.84566676   8.2702976          0.015625
#> 30: 2022-06-17 04:40:30        1 -2.87792954  11.0230125          0.015625
#> 31: 2022-06-17 04:40:30        1 -4.27134303  10.9033424          0.015625
#> 32: 2022-06-17 04:40:30        1 -2.26816154  14.6268743          0.015625
#> 33: 2022-06-17 04:40:30        1  7.42421375   5.4761648          0.015625
#> 34: 2022-06-17 04:40:30        1  5.53288442   3.5889631          0.015625
#> 35: 2022-06-17 04:40:30        1  3.83906374  12.5258022          0.015625
#> 36: 2022-06-17 04:40:30        1  3.85984794   7.1596251          0.015625
#> 37: 2022-06-17 04:40:30        1  3.07252010   6.9609046          0.015625
#> 38: 2022-06-17 04:40:30        1  5.77454932   0.6148906          0.015625
#> 39: 2022-06-17 04:40:30        1 -3.71199627   4.4484916          0.015625
#> 40: 2022-06-17 04:40:30        1 -1.19745479  12.3913178          0.015625
#> 41: 2022-06-17 04:40:30        1 -1.36836608  14.5885314          0.015625
#> 42: 2022-06-17 04:40:30        1  0.15149735   9.6482884          0.015625
#> 43: 2022-06-17 04:40:30        1 -3.63688178   6.6500963          0.015625
#> 44: 2022-06-17 04:40:30        1 -0.61972068   6.0333237          0.015625
#> 45: 2022-06-17 04:40:30        1 -1.55399850  12.0408504          0.015625
#> 46: 2022-06-17 04:40:30        1 -1.34831498  11.4830634          0.015625
#> 47: 2022-06-17 04:40:30        1  6.91315514   0.2886759          0.015625
#> 48: 2022-06-17 04:40:30        1 -0.52022454   4.1542157          0.015625
#> 49: 2022-06-17 04:40:30        1  9.51391690  14.6315929          0.015625
#> 50: 2022-06-17 04:40:30        1 -3.10584820  14.9553310          0.015625
#> 51: 2022-06-17 04:40:30        1  0.62450504   6.8467870          0.015625
#> 52: 2022-06-17 04:40:30        1  5.34679116   7.9204414          0.015625
#> 53: 2022-06-17 04:40:30        1  6.25133017   8.8610700          0.015625
#> 54: 2022-06-17 04:40:30        1  8.47626007  14.4331052          0.015625
#> 55: 2022-06-17 04:40:30        1  8.37682750   0.8528871          0.015625
#> 56: 2022-06-17 04:40:30        1  1.27116414   4.3520110          0.015625
#> 57: 2022-06-17 04:40:30        1  4.64027132  11.9199097          0.015625
#> 58: 2022-06-17 04:40:30        1 -0.13161716   2.1285396          0.015625
#> 59: 2022-06-17 04:40:30        1 -0.84732960  14.7144559          0.015625
#> 60: 2022-06-17 04:40:30        1  2.99271824   2.4325370          0.015625
#> 61: 2022-06-17 04:40:30        1 -1.33213331   0.7057929          0.015625
#> 62: 2022-06-17 04:40:30        1 -4.64917274   4.9566712          0.015625
#> 63: 2022-06-17 04:40:30        1  4.76260079   9.4508474          0.015625
#> 64: 2022-06-17 04:40:30        1  3.18534887   2.0284372          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity