Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain        y
#> 1: 4.067086 0.6574276 0.015625 <list[3]> 4.611543

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain        y
#> 1: 4.067086 0.6574276 0.015625 <list[3]> 4.611543

# all evaluations
as.data.table(instance$archive)
#>               x1          x2 fidelity stage bracket repetition          y
#>  1: -1.144220036 11.22418047 0.015625     0       6          1  25.281645
#>  2: -4.949166446  1.35688023 0.015625     0       6          1 188.427236
#>  3:  9.093427819  3.22412670 0.015625     0       6          1  84.721105
#>  4:  8.295514286  2.92061036 0.015625     0       6          1  70.016168
#>  5:  5.591619618  5.92100707 0.015625     0       6          1  79.159951
#>  6:  3.489443667 10.79759507 0.015625     0       6          1 100.510217
#>  7:  6.490471546 10.40804273 0.015625     0       6          1 200.102575
#>  8:  2.458793062 14.82725933 0.015625     0       6          1 160.170179
#>  9:  2.381172158  8.53600993 0.015625     0       6          1  40.882894
#> 10:  5.553070564  9.69770869 0.015625     0       6          1 151.426106
#> 11:  1.996488882  9.08780095 0.015625     0       6          1  43.768232
#> 12: -1.029751507  0.01196872 0.015625     0       6          1  73.613878
#> 13:  0.343928995  5.49937897 0.015625     0       6          1  19.041645
#> 14:  6.771952388  9.13024754 0.015625     0       6          1 174.680503
#> 15: -1.997511861  9.88437472 0.015625     0       6          1   6.365217
#> 16:  7.487432694  6.28342860 0.015625     0       6          1 123.194634
#> 17:  0.013542692  7.59168747 0.015625     0       6          1  22.203762
#> 18:  7.280505185  8.14538697 0.015625     0       6          1 161.689926
#> 19: -1.627253002  1.61864393 0.015625     0       6          1  59.197722
#> 20:  4.067086219  0.65742760 0.015625     0       6          1   4.611543
#> 21:  5.330809478  6.39699219 0.015625     0       6          1  79.686629
#> 22:  9.317687758  0.14271334 0.015625     0       6          1  40.178687
#> 23:  5.938836220  6.67782070 0.015625     0       6          1 100.856455
#> 24:  3.646868814 14.61991835 0.015625     0       6          1 198.022632
#> 25:  0.495465737  0.78964924 0.015625     0       6          1  38.066491
#> 26: -0.499982149  8.00186270 0.015625     0       6          1  19.862964
#> 27: -3.911112624  0.07458185 0.015625     0       6          1 162.379801
#> 28: -0.245665180  1.94627627 0.015625     0       6          1  39.085780
#> 29:  6.513983954  5.67291431 0.015625     0       6          1  95.657950
#> 30:  7.408544963  4.25052005 0.015625     0       6          1  83.926216
#> 31: -4.163480737  9.64826248 0.015625     0       6          1  17.317926
#> 32:  4.907829695  3.26195459 0.015625     0       6          1  30.634507
#> 33:  7.924775011  7.18135997 0.015625     0       6          1 150.052891
#> 34: -2.593004088 14.00147861 0.015625     0       6          1  15.260024
#> 35: -3.124765111  3.20298849 0.015625     0       6          1  65.531382
#> 36:  0.975062003  0.11827513 0.015625     0       6          1  34.389642
#> 37:  2.653793007 14.78130366 0.015625     0       6          1 165.061171
#> 38:  3.367061922  6.96948907 0.015625     0       6          1  36.398363
#> 39:  6.669342254 12.07488378 0.015625     0       6          1 253.653851
#> 40:  2.131234230  7.26281960 0.015625     0       6          1  25.282295
#> 41: -1.534960289  6.66302059 0.015625     0       6          1  13.775368
#> 42:  9.347133087  6.10649770 0.015625     0       6          1 151.632937
#> 43:  8.761831271  6.06299255 0.015625     0       6          1 138.094688
#> 44: -3.869796466  9.58164889 0.015625     0       6          1  12.061307
#> 45:  1.568080458  7.79799401 0.015625     0       6          1  27.818180
#> 46: -4.324630887  0.54836893 0.015625     0       6          1 173.025498
#> 47:  1.962740357  6.11283684 0.015625     0       6          1  16.055219
#> 48: -0.787325413  7.16093341 0.015625     0       6          1  16.788985
#> 49:  5.369744829  8.29206387 0.015625     0       6          1 114.903751
#> 50:  7.696998996  4.51623911 0.015625     0       6          1  91.511058
#> 51:  7.902025981 11.83571878 0.015625     0       6          1 281.534102
#> 52: -1.625497835 13.91592788 0.015625     0       6          1  37.012597
#> 53:  2.910019314 11.03901088 0.015625     0       6          1  89.203774
#> 54:  9.638369451 12.06868205 0.015625     0       6          1 344.803281
#> 55:  1.611977550 12.32201732 0.015625     0       6          1  87.179631
#> 56:  4.920699877  9.95893143 0.015625     0       6          1 133.999516
#> 57:  2.860378893  7.77557002 0.015625     0       6          1  37.698183
#> 58:  0.008042459 14.98281343 0.015625     0       6          1 100.522827
#> 59: -2.702571449 14.18545360 0.015625     0       6          1  14.701180
#> 60: -3.543420978  1.32981122 0.015625     0       6          1 115.562274
#> 61:  8.776214602 11.35040743 0.015625     0       6          1 289.649816
#> 62:  8.410531983  2.03930053 0.015625     0       6          1  57.491901
#> 63:  2.350153881 12.34743334 0.015625     0       6          1 101.618080
#> 64:  3.997573357 12.12592723 0.015625     0       6          1 147.631812
#>               x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-02-14 04:29:25        1 -1.144220036 11.22418047          0.015625
#>  2: 2022-02-14 04:29:25        1 -4.949166446  1.35688023          0.015625
#>  3: 2022-02-14 04:29:25        1  9.093427819  3.22412670          0.015625
#>  4: 2022-02-14 04:29:25        1  8.295514286  2.92061036          0.015625
#>  5: 2022-02-14 04:29:25        1  5.591619618  5.92100707          0.015625
#>  6: 2022-02-14 04:29:25        1  3.489443667 10.79759507          0.015625
#>  7: 2022-02-14 04:29:25        1  6.490471546 10.40804273          0.015625
#>  8: 2022-02-14 04:29:25        1  2.458793062 14.82725933          0.015625
#>  9: 2022-02-14 04:29:25        1  2.381172158  8.53600993          0.015625
#> 10: 2022-02-14 04:29:25        1  5.553070564  9.69770869          0.015625
#> 11: 2022-02-14 04:29:25        1  1.996488882  9.08780095          0.015625
#> 12: 2022-02-14 04:29:25        1 -1.029751507  0.01196872          0.015625
#> 13: 2022-02-14 04:29:25        1  0.343928995  5.49937897          0.015625
#> 14: 2022-02-14 04:29:25        1  6.771952388  9.13024754          0.015625
#> 15: 2022-02-14 04:29:25        1 -1.997511861  9.88437472          0.015625
#> 16: 2022-02-14 04:29:25        1  7.487432694  6.28342860          0.015625
#> 17: 2022-02-14 04:29:25        1  0.013542692  7.59168747          0.015625
#> 18: 2022-02-14 04:29:25        1  7.280505185  8.14538697          0.015625
#> 19: 2022-02-14 04:29:25        1 -1.627253002  1.61864393          0.015625
#> 20: 2022-02-14 04:29:25        1  4.067086219  0.65742760          0.015625
#> 21: 2022-02-14 04:29:25        1  5.330809478  6.39699219          0.015625
#> 22: 2022-02-14 04:29:25        1  9.317687758  0.14271334          0.015625
#> 23: 2022-02-14 04:29:25        1  5.938836220  6.67782070          0.015625
#> 24: 2022-02-14 04:29:25        1  3.646868814 14.61991835          0.015625
#> 25: 2022-02-14 04:29:25        1  0.495465737  0.78964924          0.015625
#> 26: 2022-02-14 04:29:25        1 -0.499982149  8.00186270          0.015625
#> 27: 2022-02-14 04:29:25        1 -3.911112624  0.07458185          0.015625
#> 28: 2022-02-14 04:29:25        1 -0.245665180  1.94627627          0.015625
#> 29: 2022-02-14 04:29:25        1  6.513983954  5.67291431          0.015625
#> 30: 2022-02-14 04:29:25        1  7.408544963  4.25052005          0.015625
#> 31: 2022-02-14 04:29:25        1 -4.163480737  9.64826248          0.015625
#> 32: 2022-02-14 04:29:25        1  4.907829695  3.26195459          0.015625
#> 33: 2022-02-14 04:29:25        1  7.924775011  7.18135997          0.015625
#> 34: 2022-02-14 04:29:25        1 -2.593004088 14.00147861          0.015625
#> 35: 2022-02-14 04:29:25        1 -3.124765111  3.20298849          0.015625
#> 36: 2022-02-14 04:29:25        1  0.975062003  0.11827513          0.015625
#> 37: 2022-02-14 04:29:25        1  2.653793007 14.78130366          0.015625
#> 38: 2022-02-14 04:29:25        1  3.367061922  6.96948907          0.015625
#> 39: 2022-02-14 04:29:25        1  6.669342254 12.07488378          0.015625
#> 40: 2022-02-14 04:29:25        1  2.131234230  7.26281960          0.015625
#> 41: 2022-02-14 04:29:25        1 -1.534960289  6.66302059          0.015625
#> 42: 2022-02-14 04:29:25        1  9.347133087  6.10649770          0.015625
#> 43: 2022-02-14 04:29:25        1  8.761831271  6.06299255          0.015625
#> 44: 2022-02-14 04:29:25        1 -3.869796466  9.58164889          0.015625
#> 45: 2022-02-14 04:29:25        1  1.568080458  7.79799401          0.015625
#> 46: 2022-02-14 04:29:25        1 -4.324630887  0.54836893          0.015625
#> 47: 2022-02-14 04:29:25        1  1.962740357  6.11283684          0.015625
#> 48: 2022-02-14 04:29:25        1 -0.787325413  7.16093341          0.015625
#> 49: 2022-02-14 04:29:25        1  5.369744829  8.29206387          0.015625
#> 50: 2022-02-14 04:29:25        1  7.696998996  4.51623911          0.015625
#> 51: 2022-02-14 04:29:25        1  7.902025981 11.83571878          0.015625
#> 52: 2022-02-14 04:29:25        1 -1.625497835 13.91592788          0.015625
#> 53: 2022-02-14 04:29:25        1  2.910019314 11.03901088          0.015625
#> 54: 2022-02-14 04:29:25        1  9.638369451 12.06868205          0.015625
#> 55: 2022-02-14 04:29:25        1  1.611977550 12.32201732          0.015625
#> 56: 2022-02-14 04:29:25        1  4.920699877  9.95893143          0.015625
#> 57: 2022-02-14 04:29:25        1  2.860378893  7.77557002          0.015625
#> 58: 2022-02-14 04:29:25        1  0.008042459 14.98281343          0.015625
#> 59: 2022-02-14 04:29:25        1 -2.702571449 14.18545360          0.015625
#> 60: 2022-02-14 04:29:25        1 -3.543420978  1.32981122          0.015625
#> 61: 2022-02-14 04:29:25        1  8.776214602 11.35040743          0.015625
#> 62: 2022-02-14 04:29:25        1  8.410531983  2.03930053          0.015625
#> 63: 2022-02-14 04:29:25        1  2.350153881 12.34743334          0.015625
#> 64: 2022-02-14 04:29:25        1  3.997573357 12.12592723          0.015625
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity