Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.795052 12.13156 0.015625 <list[3]> 2.499467

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.795052 12.13156 0.015625 <list[3]> 2.499467

# all evaluations
as.data.table(instance$archive)
#>             x1        x2 fidelity stage bracket repetition          y
#>  1:  2.4672319  7.762711 0.015625     0       6          1  32.774683
#>  2:  2.0233109  7.409841 0.015625     0       6          1  26.089222
#>  3:  5.6015131  2.409323 0.015625     0       6          1  36.462874
#>  4: -4.7281024  6.495601 0.015625     0       6          1  69.699196
#>  5:  2.2282252  6.159303 0.015625     0       6          1  16.755940
#>  6:  7.9175090 14.352592 0.015625     0       6          1 371.389072
#>  7:  1.1378150  8.365906 0.015625     0       6          1  31.143526
#>  8:  3.4838888  5.490784 0.015625     0       6          1  22.692675
#>  9:  6.2939917  1.182549 0.015625     0       6          1  35.455692
#> 10: -4.1318374  5.970525 0.015625     0       6          1  55.576189
#> 11:  2.5351280  8.776989 0.015625     0       6          1  45.857409
#> 12:  6.0131330  8.257071 0.015625     0       6          1 134.076614
#> 13: -1.0705110  5.782242 0.015625     0       6          1  18.434831
#> 14:  1.0377666  7.167644 0.015625     0       6          1  22.642107
#> 15: -2.9488163 13.670541 0.015625     0       6          1   7.919818
#> 16: -3.7950524 12.131557 0.015625     0       6          1   2.499467
#> 17: -3.7917293  6.188068 0.015625     0       6          1  41.904616
#> 18:  3.8116041 12.102070 0.015625     0       6          1 139.872533
#> 19:  5.7082910  7.903197 0.015625     0       6          1 117.785664
#> 20:  5.2570465  7.494463 0.015625     0       6          1  96.183511
#> 21:  2.7429172 14.490078 0.015625     0       6          1 160.522311
#> 22: -1.0289889  7.352313 0.015625     0       6          1  15.052748
#> 23: -0.4832235  5.287284 0.015625     0       6          1  20.719719
#> 24:  7.7640701 11.516939 0.015625     0       6          1 267.514965
#> 25:  6.9641594  1.227891 0.015625     0       6          1  40.697482
#> 26:  1.2505199  5.124902 0.015625     0       6          1  14.161692
#> 27:  7.9584795  6.612971 0.015625     0       6          1 137.409220
#> 28:  2.6368702 10.974687 0.015625     0       6          1  81.833951
#> 29:  9.9697815  1.616968 0.015625     0       6          1  72.823913
#> 30:  7.1627467  5.166754 0.015625     0       6          1  96.925987
#> 31:  0.4579189 12.747293 0.015625     0       6          1  74.408470
#> 32:  6.7862923 11.449012 0.015625     0       6          1 238.451436
#> 33:  7.8284852 14.179532 0.015625     0       6          1 361.980459
#> 34: -3.8198963 13.433885 0.015625     0       6          1   3.343690
#> 35: -0.5458506  8.522250 0.015625     0       6          1  20.910647
#> 36: -4.8345920  2.399275 0.015625     0       6          1 155.503689
#> 37:  8.1438367 14.754181 0.015625     0       6          1 394.412142
#> 38:  8.3343842 14.656706 0.015625     0       6          1 397.030233
#> 39:  5.6744205  6.063881 0.015625     0       6          1  83.567792
#> 40: -2.6125102  3.835429 0.015625     0       6          1  44.382567
#> 41:  8.9860583 11.670699 0.015625     0       6          1 307.195359
#> 42: -4.5836527 12.475316 0.015625     0       6          1  10.915794
#> 43:  3.8247482  6.362716 0.015625     0       6          1  38.555184
#> 44: -4.1297311  4.057881 0.015625     0       6          1  86.423096
#> 45:  1.3827916  5.276337 0.015625     0       6          1  13.806276
#> 46:  7.6860911  5.427922 0.015625     0       6          1 108.514886
#> 47: -0.4925698 10.749211 0.015625     0       6          1  34.124811
#> 48:  4.7996003 12.193874 0.015625     0       6          1 183.085756
#> 49:  4.8515695  6.568160 0.015625     0       6          1  68.575832
#> 50:  0.2831743  4.661644 0.015625     0       6          1  20.012034
#> 51: -1.4729917 10.941416 0.015625     0       6          1  17.340384
#> 52: -2.4308943  7.759553 0.015625     0       6          1   7.971300
#> 53:  9.7320840 14.918873 0.015625     0       6          1 462.917573
#> 54: -3.7795917  0.618003 0.015625     0       6          1 142.392649
#> 55:  5.6379452 14.767681 0.015625     0       6          1 298.683652
#> 56:  8.5717625  3.443516 0.015625     0       6          1  81.596238
#> 57: -4.0178457  2.056683 0.015625     0       6          1 121.235699
#> 58:  9.9423286  0.992726 0.015625     0       6          1  62.138643
#> 59:  7.6744105 10.488711 0.015625     0       6          1 233.487303
#> 60:  5.1074176 10.725979 0.015625     0       6          1 158.961086
#> 61: -2.2589037  3.814680 0.015625     0       6          1  39.154267
#> 62:  5.3007031  7.263277 0.015625     0       6          1  93.398301
#> 63:  1.9507856  3.667851 0.015625     0       6          1   6.868305
#> 64:  8.2603032 12.746215 0.015625     0       6          1 322.865066
#>             x1        x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-03-22 04:30:15        1   2.4672319    7.762711          0.015625
#>  2: 2022-03-22 04:30:15        1   2.0233109    7.409841          0.015625
#>  3: 2022-03-22 04:30:15        1   5.6015131    2.409323          0.015625
#>  4: 2022-03-22 04:30:15        1  -4.7281024    6.495601          0.015625
#>  5: 2022-03-22 04:30:15        1   2.2282252    6.159303          0.015625
#>  6: 2022-03-22 04:30:15        1   7.9175090   14.352592          0.015625
#>  7: 2022-03-22 04:30:15        1   1.1378150    8.365906          0.015625
#>  8: 2022-03-22 04:30:15        1   3.4838888    5.490784          0.015625
#>  9: 2022-03-22 04:30:15        1   6.2939917    1.182549          0.015625
#> 10: 2022-03-22 04:30:15        1  -4.1318374    5.970525          0.015625
#> 11: 2022-03-22 04:30:15        1   2.5351280    8.776989          0.015625
#> 12: 2022-03-22 04:30:15        1   6.0131330    8.257071          0.015625
#> 13: 2022-03-22 04:30:15        1  -1.0705110    5.782242          0.015625
#> 14: 2022-03-22 04:30:15        1   1.0377666    7.167644          0.015625
#> 15: 2022-03-22 04:30:15        1  -2.9488163   13.670541          0.015625
#> 16: 2022-03-22 04:30:15        1  -3.7950524   12.131557          0.015625
#> 17: 2022-03-22 04:30:15        1  -3.7917293    6.188068          0.015625
#> 18: 2022-03-22 04:30:15        1   3.8116041   12.102070          0.015625
#> 19: 2022-03-22 04:30:15        1   5.7082910    7.903197          0.015625
#> 20: 2022-03-22 04:30:15        1   5.2570465    7.494463          0.015625
#> 21: 2022-03-22 04:30:15        1   2.7429172   14.490078          0.015625
#> 22: 2022-03-22 04:30:15        1  -1.0289889    7.352313          0.015625
#> 23: 2022-03-22 04:30:15        1  -0.4832235    5.287284          0.015625
#> 24: 2022-03-22 04:30:15        1   7.7640701   11.516939          0.015625
#> 25: 2022-03-22 04:30:15        1   6.9641594    1.227891          0.015625
#> 26: 2022-03-22 04:30:15        1   1.2505199    5.124902          0.015625
#> 27: 2022-03-22 04:30:15        1   7.9584795    6.612971          0.015625
#> 28: 2022-03-22 04:30:15        1   2.6368702   10.974687          0.015625
#> 29: 2022-03-22 04:30:15        1   9.9697815    1.616968          0.015625
#> 30: 2022-03-22 04:30:15        1   7.1627467    5.166754          0.015625
#> 31: 2022-03-22 04:30:15        1   0.4579189   12.747293          0.015625
#> 32: 2022-03-22 04:30:15        1   6.7862923   11.449012          0.015625
#> 33: 2022-03-22 04:30:15        1   7.8284852   14.179532          0.015625
#> 34: 2022-03-22 04:30:15        1  -3.8198963   13.433885          0.015625
#> 35: 2022-03-22 04:30:15        1  -0.5458506    8.522250          0.015625
#> 36: 2022-03-22 04:30:15        1  -4.8345920    2.399275          0.015625
#> 37: 2022-03-22 04:30:15        1   8.1438367   14.754181          0.015625
#> 38: 2022-03-22 04:30:15        1   8.3343842   14.656706          0.015625
#> 39: 2022-03-22 04:30:15        1   5.6744205    6.063881          0.015625
#> 40: 2022-03-22 04:30:15        1  -2.6125102    3.835429          0.015625
#> 41: 2022-03-22 04:30:15        1   8.9860583   11.670699          0.015625
#> 42: 2022-03-22 04:30:15        1  -4.5836527   12.475316          0.015625
#> 43: 2022-03-22 04:30:15        1   3.8247482    6.362716          0.015625
#> 44: 2022-03-22 04:30:15        1  -4.1297311    4.057881          0.015625
#> 45: 2022-03-22 04:30:15        1   1.3827916    5.276337          0.015625
#> 46: 2022-03-22 04:30:15        1   7.6860911    5.427922          0.015625
#> 47: 2022-03-22 04:30:15        1  -0.4925698   10.749211          0.015625
#> 48: 2022-03-22 04:30:15        1   4.7996003   12.193874          0.015625
#> 49: 2022-03-22 04:30:15        1   4.8515695    6.568160          0.015625
#> 50: 2022-03-22 04:30:15        1   0.2831743    4.661644          0.015625
#> 51: 2022-03-22 04:30:15        1  -1.4729917   10.941416          0.015625
#> 52: 2022-03-22 04:30:15        1  -2.4308943    7.759553          0.015625
#> 53: 2022-03-22 04:30:15        1   9.7320840   14.918873          0.015625
#> 54: 2022-03-22 04:30:15        1  -3.7795917    0.618003          0.015625
#> 55: 2022-03-22 04:30:15        1   5.6379452   14.767681          0.015625
#> 56: 2022-03-22 04:30:15        1   8.5717625    3.443516          0.015625
#> 57: 2022-03-22 04:30:15        1  -4.0178457    2.056683          0.015625
#> 58: 2022-03-22 04:30:15        1   9.9423286    0.992726          0.015625
#> 59: 2022-03-22 04:30:15        1   7.6744105   10.488711          0.015625
#> 60: 2022-03-22 04:30:15        1   5.1074176   10.725979          0.015625
#> 61: 2022-03-22 04:30:15        1  -2.2589037    3.814680          0.015625
#> 62: 2022-03-22 04:30:15        1   5.3007031    7.263277          0.015625
#> 63: 2022-03-22 04:30:15        1   1.9507856    3.667851          0.015625
#> 64: 2022-03-22 04:30:15        1   8.2603032   12.746215          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity