Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.792066 13.26772 0.015625 <list[3]> 2.983275

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.792066 13.26772 0.015625 <list[3]> 2.983275

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1: -1.47351785  0.9688525 0.015625     0       6          1  66.332100
#>  2:  3.61322456  3.3743721 0.015625     0       6          1   8.864075
#>  3:  5.98845123  6.9270023 0.015625     0       6          1 106.709386
#>  4:  7.81919124 11.2242637 0.015625     0       6          1 259.627418
#>  5: -0.02359536  5.4000665 0.015625     0       6          1  20.005851
#>  6: -1.50582136  0.5060168 0.015625     0       6          1  73.989693
#>  7: -2.90488457 14.5772183 0.015625     0       6          1  14.314947
#>  8: -0.58483560 11.8966115 0.015625     0       6          1  42.561270
#>  9:  7.74078723  9.6799651 0.015625     0       6          1 211.518155
#> 10:  6.98476700  1.5165266 0.015625     0       6          1  43.682767
#> 11: -2.40609790  2.0120624 0.015625     0       6          1  66.805909
#> 12: -3.31521032 14.3368331 0.015625     0       6          1   7.954703
#> 13:  3.91914405  9.9772507 0.015625     0       6          1  98.073474
#> 14:  1.11898039 11.8433107 0.015625     0       6          1  71.735498
#> 15: -1.74622370  8.4872480 0.015625     0       6          1   8.472925
#> 16:  8.50940084  2.3942802 0.015625     0       6          1  63.607489
#> 17: -1.94997651  1.6383701 0.015625     0       6          1  63.932818
#> 18:  6.77793518  1.4227252 0.015625     0       6          1  41.467799
#> 19: -3.35060441  4.6007059 0.015625     0       6          1  50.692625
#> 20: -3.79206587 13.2677180 0.015625     0       6          1   2.983275
#> 21:  1.73440134  0.7729511 0.015625     0       6          1  14.985320
#> 22:  8.19434608 10.5162231 0.015625     0       6          1 246.838448
#> 23: -4.80682264 13.3815080 0.015625     0       6          1  11.864276
#> 24: -2.39956957  2.3515980 0.015625     0       6          1  61.360251
#> 25: -1.46413055 14.8784674 0.015625     0       6          1  53.042743
#> 26:  6.19382925 10.6957997 0.015625     0       6          1 198.428259
#> 27:  8.28137710 10.9456936 0.015625     0       6          1 262.572125
#> 28:  1.91406755 12.9430337 0.015625     0       6          1 104.317830
#> 29:  4.89555810 13.8745852 0.015625     0       6          1 234.630306
#> 30: -2.13721711  2.9143225 0.015625     0       6          1  48.772538
#> 31: -0.58158260 13.1295092 0.015625     0       6          1  56.382813
#> 32: -3.78804894  1.1701672 0.015625     0       6          1 130.023074
#> 33: -1.09247194  9.6045965 0.015625     0       6          1  17.765667
#> 34: -3.50573677  8.6131672 0.015625     0       6          1  12.211668
#> 35:  1.10837904  4.1129187 0.015625     0       6          1  14.309487
#> 36:  8.12762673  9.3419097 0.015625     0       6          1 210.363125
#> 37:  0.13389086  3.5420471 0.015625     0       6          1  24.558040
#> 38:  5.32021976 12.8747653 0.015625     0       6          1 224.918732
#> 39:  8.36505590  6.4029048 0.015625     0       6          1 139.048265
#> 40:  5.53441231  7.8351862 0.015625     0       6          1 111.156859
#> 41: -3.19559004  4.9621323 0.015625     0       6          1  41.857015
#> 42:  9.53366103  5.1969422 0.015625     0       6          1 134.449682
#> 43:  1.54407531  3.1052211 0.015625     0       6          1  10.517273
#> 44:  3.63631437 13.3611027 0.015625     0       6          1 163.905591
#> 45:  9.62589599  3.4925974 0.015625     0       6          1  99.867428
#> 46:  5.76690993  0.1629760 0.015625     0       6          1  23.727153
#> 47: -3.60499929 10.6060042 0.015625     0       6          1   3.754900
#> 48:  0.25471509  9.5031412 0.015625     0       6          1  34.553341
#> 49: -0.60365762  0.8698150 0.015625     0       6          1  55.141197
#> 50:  9.48695509  7.1583629 0.015625     0       6          1 182.397050
#> 51: -4.21376093  8.9678055 0.015625     0       6          1  23.765354
#> 52: -4.05923313  7.6357267 0.015625     0       6          1  32.588380
#> 53:  6.74985980  9.5798873 0.015625     0       6          1 185.547503
#> 54:  0.86732387 12.9386226 0.015625     0       6          1  85.032972
#> 55:  6.38999450  4.0797503 0.015625     0       6          1  68.467341
#> 56:  6.99885743  6.7501682 0.015625     0       6          1 125.054693
#> 57:  0.26439377 10.9799576 0.015625     0       6          1  48.413375
#> 58:  7.29183919  4.7541711 0.015625     0       6          1  91.237391
#> 59: -0.86457045  6.9110496 0.015625     0       6          1  16.469539
#> 60: -2.71152947  7.1636502 0.015625     0       6          1  12.682788
#> 61:  9.30699251  6.4396227 0.015625     0       6          1 158.943612
#> 62:  4.64118938 13.7992684 0.015625     0       6          1 220.253044
#> 63:  5.49368134 12.3129781 0.015625     0       6          1 216.375798
#> 64:  3.61460995  2.1577384 0.015625     0       6          1   3.728839
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-25 04:40:13        1 -1.47351785   0.9688525          0.015625
#>  2: 2022-05-25 04:40:13        1  3.61322456   3.3743721          0.015625
#>  3: 2022-05-25 04:40:13        1  5.98845123   6.9270023          0.015625
#>  4: 2022-05-25 04:40:13        1  7.81919124  11.2242637          0.015625
#>  5: 2022-05-25 04:40:13        1 -0.02359536   5.4000665          0.015625
#>  6: 2022-05-25 04:40:13        1 -1.50582136   0.5060168          0.015625
#>  7: 2022-05-25 04:40:13        1 -2.90488457  14.5772183          0.015625
#>  8: 2022-05-25 04:40:13        1 -0.58483560  11.8966115          0.015625
#>  9: 2022-05-25 04:40:13        1  7.74078723   9.6799651          0.015625
#> 10: 2022-05-25 04:40:13        1  6.98476700   1.5165266          0.015625
#> 11: 2022-05-25 04:40:13        1 -2.40609790   2.0120624          0.015625
#> 12: 2022-05-25 04:40:13        1 -3.31521032  14.3368331          0.015625
#> 13: 2022-05-25 04:40:13        1  3.91914405   9.9772507          0.015625
#> 14: 2022-05-25 04:40:13        1  1.11898039  11.8433107          0.015625
#> 15: 2022-05-25 04:40:13        1 -1.74622370   8.4872480          0.015625
#> 16: 2022-05-25 04:40:13        1  8.50940084   2.3942802          0.015625
#> 17: 2022-05-25 04:40:13        1 -1.94997651   1.6383701          0.015625
#> 18: 2022-05-25 04:40:13        1  6.77793518   1.4227252          0.015625
#> 19: 2022-05-25 04:40:13        1 -3.35060441   4.6007059          0.015625
#> 20: 2022-05-25 04:40:13        1 -3.79206587  13.2677180          0.015625
#> 21: 2022-05-25 04:40:13        1  1.73440134   0.7729511          0.015625
#> 22: 2022-05-25 04:40:13        1  8.19434608  10.5162231          0.015625
#> 23: 2022-05-25 04:40:13        1 -4.80682264  13.3815080          0.015625
#> 24: 2022-05-25 04:40:13        1 -2.39956957   2.3515980          0.015625
#> 25: 2022-05-25 04:40:13        1 -1.46413055  14.8784674          0.015625
#> 26: 2022-05-25 04:40:13        1  6.19382925  10.6957997          0.015625
#> 27: 2022-05-25 04:40:13        1  8.28137710  10.9456936          0.015625
#> 28: 2022-05-25 04:40:13        1  1.91406755  12.9430337          0.015625
#> 29: 2022-05-25 04:40:13        1  4.89555810  13.8745852          0.015625
#> 30: 2022-05-25 04:40:13        1 -2.13721711   2.9143225          0.015625
#> 31: 2022-05-25 04:40:13        1 -0.58158260  13.1295092          0.015625
#> 32: 2022-05-25 04:40:13        1 -3.78804894   1.1701672          0.015625
#> 33: 2022-05-25 04:40:13        1 -1.09247194   9.6045965          0.015625
#> 34: 2022-05-25 04:40:13        1 -3.50573677   8.6131672          0.015625
#> 35: 2022-05-25 04:40:13        1  1.10837904   4.1129187          0.015625
#> 36: 2022-05-25 04:40:13        1  8.12762673   9.3419097          0.015625
#> 37: 2022-05-25 04:40:13        1  0.13389086   3.5420471          0.015625
#> 38: 2022-05-25 04:40:13        1  5.32021976  12.8747653          0.015625
#> 39: 2022-05-25 04:40:13        1  8.36505590   6.4029048          0.015625
#> 40: 2022-05-25 04:40:13        1  5.53441231   7.8351862          0.015625
#> 41: 2022-05-25 04:40:13        1 -3.19559004   4.9621323          0.015625
#> 42: 2022-05-25 04:40:13        1  9.53366103   5.1969422          0.015625
#> 43: 2022-05-25 04:40:13        1  1.54407531   3.1052211          0.015625
#> 44: 2022-05-25 04:40:13        1  3.63631437  13.3611027          0.015625
#> 45: 2022-05-25 04:40:13        1  9.62589599   3.4925974          0.015625
#> 46: 2022-05-25 04:40:13        1  5.76690993   0.1629760          0.015625
#> 47: 2022-05-25 04:40:13        1 -3.60499929  10.6060042          0.015625
#> 48: 2022-05-25 04:40:13        1  0.25471509   9.5031412          0.015625
#> 49: 2022-05-25 04:40:13        1 -0.60365762   0.8698150          0.015625
#> 50: 2022-05-25 04:40:13        1  9.48695509   7.1583629          0.015625
#> 51: 2022-05-25 04:40:13        1 -4.21376093   8.9678055          0.015625
#> 52: 2022-05-25 04:40:13        1 -4.05923313   7.6357267          0.015625
#> 53: 2022-05-25 04:40:13        1  6.74985980   9.5798873          0.015625
#> 54: 2022-05-25 04:40:13        1  0.86732387  12.9386226          0.015625
#> 55: 2022-05-25 04:40:13        1  6.38999450   4.0797503          0.015625
#> 56: 2022-05-25 04:40:13        1  6.99885743   6.7501682          0.015625
#> 57: 2022-05-25 04:40:13        1  0.26439377  10.9799576          0.015625
#> 58: 2022-05-25 04:40:13        1  7.29183919   4.7541711          0.015625
#> 59: 2022-05-25 04:40:13        1 -0.86457045   6.9110496          0.015625
#> 60: 2022-05-25 04:40:13        1 -2.71152947   7.1636502          0.015625
#> 61: 2022-05-25 04:40:13        1  9.30699251   6.4396227          0.015625
#> 62: 2022-05-25 04:40:13        1  4.64118938  13.7992684          0.015625
#> 63: 2022-05-25 04:40:13        1  5.49368134  12.3129781          0.015625
#> 64: 2022-05-25 04:40:13        1  3.61460995   2.1577384          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity