Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.667018 12.05537 0.015625 <list[3]> 1.730876

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.667018 12.05537 0.015625 <list[3]> 1.730876

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  7.06113404 12.9801094 0.015625     0       6          1 295.236650
#>  2:  3.37527744  3.1299798 0.015625     0       6          1   5.288325
#>  3:  9.72144472 10.3550190 0.015625     0       6          1 287.150683
#>  4:  8.33302986  4.2312100 0.015625     0       6          1  93.157176
#>  5:  7.81517254  3.3413954 0.015625     0       6          1  72.809357
#>  6:  2.62354028 13.2955765 0.015625     0       6          1 128.432816
#>  7:  4.14387969 11.4974620 0.015625     0       6          1 138.572081
#>  8:  1.18976035 13.1024262 0.015625     0       6          1  93.717491
#>  9:  6.00859848  9.1462769 0.015625     0       6          1 153.783629
#> 10: -2.55770611  3.9803856 0.015625     0       6          1  41.571333
#> 11:  5.83859176 12.4809086 0.015625     0       6          1 235.499471
#> 12: -3.18865782  6.6800085 0.015625     0       6          1  22.569259
#> 13:  1.90536120 13.3761451 0.015625     0       6          1 112.875228
#> 14: -2.25258057 12.4523513 0.015625     0       6          1  11.299790
#> 15: -2.18989020  0.8493362 0.015625     0       6          1  81.576618
#> 16:  9.87517886 13.5177650 0.015625     0       6          1 410.858376
#> 17:  0.84376544 12.6567381 0.015625     0       6          1  80.026459
#> 18:  8.92367552 11.1359088 0.015625     0       6          1 286.848308
#> 19:  1.14477676  5.1297785 0.015625     0       6          1  14.798814
#> 20:  9.09937320  8.2202072 0.015625     0       6          1 201.308296
#> 21:  8.55289472  3.8845646 0.015625     0       6          1  89.342348
#> 22:  6.90877997 14.0459097 0.015625     0       6          1 326.628446
#> 23: -0.75474207 12.4857969 0.015625     0       6          1  44.736700
#> 24:  1.48292696  9.2070225 0.015625     0       6          1  41.087798
#> 25: -2.77653046  9.7972262 0.015625     0       6          1   1.768175
#> 26: -0.99556873  1.6586421 0.015625     0       6          1  50.701663
#> 27: -4.71835837  3.6333321 0.015625     0       6          1 121.585411
#> 28:  1.54614655  8.2472580 0.015625     0       6          1  31.715477
#> 29:  4.08964022 10.7674683 0.015625     0       6          1 120.221833
#> 30:  0.42065655  8.4222542 0.015625     0       6          1  28.290315
#> 31:  0.05332141  4.2200037 0.015625     0       6          1  22.462237
#> 32:  6.88903309  2.5096429 0.015625     0       6          1  54.069309
#> 33:  4.35209028  9.5892208 0.015625     0       6          1 105.287533
#> 34: -3.65996992  6.8879807 0.015625     0       6          1  30.270212
#> 35:  9.99222055  9.4336745 0.015625     0       6          1 266.513759
#> 36:  7.64993868  1.2917373 0.015625     0       6          1  44.067705
#> 37:  4.46467138  1.7513631 0.015625     0       6          1  12.682160
#> 38:  9.40096446 11.5569122 0.015625     0       6          1 317.299301
#> 39: -2.14505079 12.7584507 0.015625     0       6          1  15.043395
#> 40: -2.33119583 14.8346992 0.015625     0       6          1  27.957854
#> 41: -4.60392048  4.2537335 0.015625     0       6          1 103.542940
#> 42: -4.65467209 13.1088474 0.015625     0       6          1  10.378211
#> 43:  0.55223717 13.2161357 0.015625     0       6          1  83.552874
#> 44:  4.48018151  5.8769563 0.015625     0       6          1  48.625365
#> 45: -0.72792031 11.4964224 0.015625     0       6          1  35.844858
#> 46:  1.39469476  9.9547309 0.015625     0       6          1  49.071144
#> 47: -0.06943773  0.4500798 0.015625     0       6          1  51.621163
#> 48: -3.66701841 12.0553676 0.015625     0       6          1   1.730876
#> 49: -3.06355095 14.0048592 0.015625     0       6          1   8.495526
#> 50: -3.94399299  6.2531475 0.015625     0       6          1  45.605100
#> 51:  5.95179064 12.1037451 0.015625     0       6          1 228.956818
#> 52:  8.89848256 10.8255745 0.015625     0       6          1 275.709348
#> 53:  4.05237064 11.0917524 0.015625     0       6          1 125.914376
#> 54: -2.09019246  9.9153431 0.015625     0       6          1   5.440384
#> 55: -1.05546094 12.3723462 0.015625     0       6          1  36.431712
#> 56:  5.19703834  8.1160682 0.015625     0       6          1 105.809077
#> 57: -0.31138312 14.1300654 0.015625     0       6          1  77.380185
#> 58:  4.73075062  7.0417498 0.015625     0       6          1  72.315706
#> 59:  1.38874676  8.7015328 0.015625     0       6          1  35.285106
#> 60:  3.96238817 11.9515019 0.015625     0       6          1 142.107672
#> 61:  3.07189754 10.1613437 0.015625     0       6          1  77.163611
#> 62: -4.53740304  6.4860224 0.015625     0       6          1  62.623105
#> 63:  3.45725064  3.0728596 0.015625     0       6          1   5.746416
#> 64:  2.61736355  3.0953264 0.015625     0       6          1   2.790599
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-06-11 04:29:24        1  7.06113404  12.9801094          0.015625
#>  2: 2022-06-11 04:29:24        1  3.37527744   3.1299798          0.015625
#>  3: 2022-06-11 04:29:24        1  9.72144472  10.3550190          0.015625
#>  4: 2022-06-11 04:29:24        1  8.33302986   4.2312100          0.015625
#>  5: 2022-06-11 04:29:24        1  7.81517254   3.3413954          0.015625
#>  6: 2022-06-11 04:29:24        1  2.62354028  13.2955765          0.015625
#>  7: 2022-06-11 04:29:24        1  4.14387969  11.4974620          0.015625
#>  8: 2022-06-11 04:29:24        1  1.18976035  13.1024262          0.015625
#>  9: 2022-06-11 04:29:24        1  6.00859848   9.1462769          0.015625
#> 10: 2022-06-11 04:29:24        1 -2.55770611   3.9803856          0.015625
#> 11: 2022-06-11 04:29:24        1  5.83859176  12.4809086          0.015625
#> 12: 2022-06-11 04:29:24        1 -3.18865782   6.6800085          0.015625
#> 13: 2022-06-11 04:29:24        1  1.90536120  13.3761451          0.015625
#> 14: 2022-06-11 04:29:24        1 -2.25258057  12.4523513          0.015625
#> 15: 2022-06-11 04:29:24        1 -2.18989020   0.8493362          0.015625
#> 16: 2022-06-11 04:29:24        1  9.87517886  13.5177650          0.015625
#> 17: 2022-06-11 04:29:24        1  0.84376544  12.6567381          0.015625
#> 18: 2022-06-11 04:29:24        1  8.92367552  11.1359088          0.015625
#> 19: 2022-06-11 04:29:24        1  1.14477676   5.1297785          0.015625
#> 20: 2022-06-11 04:29:24        1  9.09937320   8.2202072          0.015625
#> 21: 2022-06-11 04:29:24        1  8.55289472   3.8845646          0.015625
#> 22: 2022-06-11 04:29:24        1  6.90877997  14.0459097          0.015625
#> 23: 2022-06-11 04:29:24        1 -0.75474207  12.4857969          0.015625
#> 24: 2022-06-11 04:29:24        1  1.48292696   9.2070225          0.015625
#> 25: 2022-06-11 04:29:24        1 -2.77653046   9.7972262          0.015625
#> 26: 2022-06-11 04:29:24        1 -0.99556873   1.6586421          0.015625
#> 27: 2022-06-11 04:29:24        1 -4.71835837   3.6333321          0.015625
#> 28: 2022-06-11 04:29:24        1  1.54614655   8.2472580          0.015625
#> 29: 2022-06-11 04:29:24        1  4.08964022  10.7674683          0.015625
#> 30: 2022-06-11 04:29:24        1  0.42065655   8.4222542          0.015625
#> 31: 2022-06-11 04:29:24        1  0.05332141   4.2200037          0.015625
#> 32: 2022-06-11 04:29:24        1  6.88903309   2.5096429          0.015625
#> 33: 2022-06-11 04:29:24        1  4.35209028   9.5892208          0.015625
#> 34: 2022-06-11 04:29:24        1 -3.65996992   6.8879807          0.015625
#> 35: 2022-06-11 04:29:24        1  9.99222055   9.4336745          0.015625
#> 36: 2022-06-11 04:29:24        1  7.64993868   1.2917373          0.015625
#> 37: 2022-06-11 04:29:24        1  4.46467138   1.7513631          0.015625
#> 38: 2022-06-11 04:29:24        1  9.40096446  11.5569122          0.015625
#> 39: 2022-06-11 04:29:24        1 -2.14505079  12.7584507          0.015625
#> 40: 2022-06-11 04:29:24        1 -2.33119583  14.8346992          0.015625
#> 41: 2022-06-11 04:29:24        1 -4.60392048   4.2537335          0.015625
#> 42: 2022-06-11 04:29:24        1 -4.65467209  13.1088474          0.015625
#> 43: 2022-06-11 04:29:24        1  0.55223717  13.2161357          0.015625
#> 44: 2022-06-11 04:29:24        1  4.48018151   5.8769563          0.015625
#> 45: 2022-06-11 04:29:24        1 -0.72792031  11.4964224          0.015625
#> 46: 2022-06-11 04:29:24        1  1.39469476   9.9547309          0.015625
#> 47: 2022-06-11 04:29:24        1 -0.06943773   0.4500798          0.015625
#> 48: 2022-06-11 04:29:24        1 -3.66701841  12.0553676          0.015625
#> 49: 2022-06-11 04:29:24        1 -3.06355095  14.0048592          0.015625
#> 50: 2022-06-11 04:29:24        1 -3.94399299   6.2531475          0.015625
#> 51: 2022-06-11 04:29:24        1  5.95179064  12.1037451          0.015625
#> 52: 2022-06-11 04:29:24        1  8.89848256  10.8255745          0.015625
#> 53: 2022-06-11 04:29:24        1  4.05237064  11.0917524          0.015625
#> 54: 2022-06-11 04:29:24        1 -2.09019246   9.9153431          0.015625
#> 55: 2022-06-11 04:29:24        1 -1.05546094  12.3723462          0.015625
#> 56: 2022-06-11 04:29:24        1  5.19703834   8.1160682          0.015625
#> 57: 2022-06-11 04:29:24        1 -0.31138312  14.1300654          0.015625
#> 58: 2022-06-11 04:29:24        1  4.73075062   7.0417498          0.015625
#> 59: 2022-06-11 04:29:24        1  1.38874676   8.7015328          0.015625
#> 60: 2022-06-11 04:29:24        1  3.96238817  11.9515019          0.015625
#> 61: 2022-06-11 04:29:24        1  3.07189754  10.1613437          0.015625
#> 62: 2022-06-11 04:29:24        1 -4.53740304   6.4860224          0.015625
#> 63: 2022-06-11 04:29:24        1  3.45725064   3.0728596          0.015625
#> 64: 2022-06-11 04:29:24        1  2.61736355   3.0953264          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity