Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.120386 11.73843 0.015625 <list[3]> 0.6235901

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.120386 11.73843 0.015625 <list[3]> 0.6235901

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition           y
#>  1:  4.19525888  8.0893893 0.015625     0       6          1  72.9067135
#>  2: -3.12038597 11.7384303 0.015625     0       6          1   0.6235901
#>  3:  2.63479334  1.5178675 0.015625     0       6          1   1.8570319
#>  4:  4.81122187  4.1379550 0.015625     0       6          1  36.7896877
#>  5: -1.51061723  6.2067724 0.015625     0       6          1  15.7195671
#>  6:  2.88437105 10.7897518 0.015625     0       6          1  83.9715374
#>  7: -1.65795693 13.0773748 0.015625     0       6          1  28.1226368
#>  8:  5.27307806 14.9024492 0.015625     0       6          1 285.3758404
#>  9:  2.81198349 13.9269436 0.015625     0       6          1 148.7616547
#> 10: -1.50368540  2.5410667 0.015625     0       6          1  45.7097915
#> 11:  4.29944319  2.7856668 0.015625     0       6          1  15.5106621
#> 12:  8.44075226  8.6095063 0.015625     0       6          1 196.5827956
#> 13: -1.51519672  1.1467826 0.015625     0       6          1  64.3404811
#> 14:  8.18831653 14.8288479 0.015625     0       6          1 398.8660573
#> 15:  6.93648949  7.3667412 0.015625     0       6          1 137.0267562
#> 16:  3.91485250 12.8087796 0.015625     0       6          1 161.0887739
#> 17:  0.03566722  2.7931371 0.015625     0       6          1  29.5193618
#> 18:  1.39196471 12.7340363 0.015625     0       6          1  90.7373240
#> 19:  7.67428328  5.0159045 0.015625     0       6          1 100.4351332
#> 20: -0.17026026  4.0762097 0.015625     0       6          1  24.2841923
#> 21:  0.20993509 10.2246581 0.015625     0       6          1  40.1614158
#> 22:  5.30104765  2.6108179 0.015625     0       6          1  32.8346477
#> 23:  9.98163557  0.3876138 0.015625     0       6          1  53.8393353
#> 24:  9.64610363  7.0431571 0.015625     0       6          1 183.8142905
#> 25: -3.78564485  4.8830137 0.015625     0       6          1  59.8183074
#> 26:  4.25048369  1.6330218 0.015625     0       6          1   9.1151478
#> 27:  2.97848439 13.0736784 0.015625     0       6          1 133.7272950
#> 28:  1.44642070  8.1437719 0.015625     0       6          1  30.3886918
#> 29:  2.78979296 11.9513731 0.015625     0       6          1 104.0524156
#> 30:  3.81533479  5.2484268 0.015625     0       6          1  26.2435473
#> 31:  4.34052987  3.1330520 0.015625     0       6          1  18.4961194
#> 32:  2.95359800  3.7822030 0.015625     0       6          1   5.4722791
#> 33: -4.93815423  9.4546204 0.015625     0       6          1  38.7180494
#> 34:  4.29956757  7.4338477 0.015625     0       6          1  65.5674809
#> 35:  0.28222533  5.9611823 0.015625     0       6          1  19.3886242
#> 36:  6.97898236 11.3455675 0.015625     0       6          1 241.0338185
#> 37:  0.06626334 11.4044800 0.015625     0       6          1  49.9390064
#> 38:  0.59617593  3.9734988 0.015625     0       6          1  19.1306602
#> 39:  6.24948578 11.7469790 0.015625     0       6          1 229.6288441
#> 40:  8.71503719 12.8644902 0.015625     0       6          1 341.2621400
#> 41:  7.81461979  5.7986501 0.015625     0       6          1 117.6729995
#> 42: -0.94668041 10.2675244 0.015625     0       6          1  23.0821039
#> 43: -0.95829573 12.3838230 0.015625     0       6          1  38.8532823
#> 44:  7.66496492 14.4724691 0.015625     0       6          1 367.6999785
#> 45:  3.72772261  7.7947303 0.015625     0       6          1  55.2953772
#> 46: -4.77921369  8.3897325 0.015625     0       6          1  45.6746900
#> 47: -4.90101411  7.5302398 0.015625     0       6          1  60.9196595
#> 48:  2.45317255  0.8625004 0.015625     0       6          1   4.5960286
#> 49: -2.47800367  0.2035980 0.015625     0       6          1 101.0220037
#> 50: -3.24000190  4.4822642 0.015625     0       6          1  49.4041615
#> 51: -4.09428744  0.5299873 0.015625     0       6          1 160.7279728
#> 52:  3.19221158  3.2324314 0.015625     0       6          1   4.4088876
#> 53:  5.25171468  2.6525072 0.015625     0       6          1  32.2605790
#> 54:  7.23754664  0.2655436 0.015625     0       6          1  32.9724146
#> 55:  8.92814207 12.8163186 0.015625     0       6          1 346.5885787
#> 56:  8.37945284  9.8370127 0.015625     0       6          1 230.6161243
#> 57:  6.74278778  1.3342144 0.015625     0       6          1  40.3938230
#> 58:  7.86221255  8.1357850 0.015625     0       6          1 172.4395437
#> 59:  0.87417955 10.7475009 0.015625     0       6          1  53.5579160
#> 60:  1.93060793  7.8599647 0.015625     0       6          1  29.8324418
#> 61:  3.93551241 14.1085396 0.015625     0       6          1 196.3639940
#> 62:  2.94056187  0.5633341 0.015625     0       6          1   1.6367327
#> 63:  3.40351474 14.4758663 0.015625     0       6          1 183.9637676
#> 64:  7.47114182  7.5870860 0.015625     0       6          1 151.9208023
#>              x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-27 04:46:32        1  4.19525888   8.0893893          0.015625
#>  2: 2022-05-27 04:46:32        1 -3.12038597  11.7384303          0.015625
#>  3: 2022-05-27 04:46:32        1  2.63479334   1.5178675          0.015625
#>  4: 2022-05-27 04:46:32        1  4.81122187   4.1379550          0.015625
#>  5: 2022-05-27 04:46:32        1 -1.51061723   6.2067724          0.015625
#>  6: 2022-05-27 04:46:32        1  2.88437105  10.7897518          0.015625
#>  7: 2022-05-27 04:46:32        1 -1.65795693  13.0773748          0.015625
#>  8: 2022-05-27 04:46:32        1  5.27307806  14.9024492          0.015625
#>  9: 2022-05-27 04:46:32        1  2.81198349  13.9269436          0.015625
#> 10: 2022-05-27 04:46:32        1 -1.50368540   2.5410667          0.015625
#> 11: 2022-05-27 04:46:32        1  4.29944319   2.7856668          0.015625
#> 12: 2022-05-27 04:46:32        1  8.44075226   8.6095063          0.015625
#> 13: 2022-05-27 04:46:32        1 -1.51519672   1.1467826          0.015625
#> 14: 2022-05-27 04:46:32        1  8.18831653  14.8288479          0.015625
#> 15: 2022-05-27 04:46:32        1  6.93648949   7.3667412          0.015625
#> 16: 2022-05-27 04:46:32        1  3.91485250  12.8087796          0.015625
#> 17: 2022-05-27 04:46:32        1  0.03566722   2.7931371          0.015625
#> 18: 2022-05-27 04:46:32        1  1.39196471  12.7340363          0.015625
#> 19: 2022-05-27 04:46:32        1  7.67428328   5.0159045          0.015625
#> 20: 2022-05-27 04:46:32        1 -0.17026026   4.0762097          0.015625
#> 21: 2022-05-27 04:46:32        1  0.20993509  10.2246581          0.015625
#> 22: 2022-05-27 04:46:32        1  5.30104765   2.6108179          0.015625
#> 23: 2022-05-27 04:46:32        1  9.98163557   0.3876138          0.015625
#> 24: 2022-05-27 04:46:32        1  9.64610363   7.0431571          0.015625
#> 25: 2022-05-27 04:46:32        1 -3.78564485   4.8830137          0.015625
#> 26: 2022-05-27 04:46:32        1  4.25048369   1.6330218          0.015625
#> 27: 2022-05-27 04:46:32        1  2.97848439  13.0736784          0.015625
#> 28: 2022-05-27 04:46:32        1  1.44642070   8.1437719          0.015625
#> 29: 2022-05-27 04:46:32        1  2.78979296  11.9513731          0.015625
#> 30: 2022-05-27 04:46:32        1  3.81533479   5.2484268          0.015625
#> 31: 2022-05-27 04:46:32        1  4.34052987   3.1330520          0.015625
#> 32: 2022-05-27 04:46:32        1  2.95359800   3.7822030          0.015625
#> 33: 2022-05-27 04:46:32        1 -4.93815423   9.4546204          0.015625
#> 34: 2022-05-27 04:46:32        1  4.29956757   7.4338477          0.015625
#> 35: 2022-05-27 04:46:32        1  0.28222533   5.9611823          0.015625
#> 36: 2022-05-27 04:46:32        1  6.97898236  11.3455675          0.015625
#> 37: 2022-05-27 04:46:32        1  0.06626334  11.4044800          0.015625
#> 38: 2022-05-27 04:46:32        1  0.59617593   3.9734988          0.015625
#> 39: 2022-05-27 04:46:32        1  6.24948578  11.7469790          0.015625
#> 40: 2022-05-27 04:46:32        1  8.71503719  12.8644902          0.015625
#> 41: 2022-05-27 04:46:32        1  7.81461979   5.7986501          0.015625
#> 42: 2022-05-27 04:46:32        1 -0.94668041  10.2675244          0.015625
#> 43: 2022-05-27 04:46:32        1 -0.95829573  12.3838230          0.015625
#> 44: 2022-05-27 04:46:32        1  7.66496492  14.4724691          0.015625
#> 45: 2022-05-27 04:46:32        1  3.72772261   7.7947303          0.015625
#> 46: 2022-05-27 04:46:32        1 -4.77921369   8.3897325          0.015625
#> 47: 2022-05-27 04:46:32        1 -4.90101411   7.5302398          0.015625
#> 48: 2022-05-27 04:46:32        1  2.45317255   0.8625004          0.015625
#> 49: 2022-05-27 04:46:32        1 -2.47800367   0.2035980          0.015625
#> 50: 2022-05-27 04:46:32        1 -3.24000190   4.4822642          0.015625
#> 51: 2022-05-27 04:46:32        1 -4.09428744   0.5299873          0.015625
#> 52: 2022-05-27 04:46:32        1  3.19221158   3.2324314          0.015625
#> 53: 2022-05-27 04:46:32        1  5.25171468   2.6525072          0.015625
#> 54: 2022-05-27 04:46:32        1  7.23754664   0.2655436          0.015625
#> 55: 2022-05-27 04:46:32        1  8.92814207  12.8163186          0.015625
#> 56: 2022-05-27 04:46:32        1  8.37945284   9.8370127          0.015625
#> 57: 2022-05-27 04:46:32        1  6.74278778   1.3342144          0.015625
#> 58: 2022-05-27 04:46:32        1  7.86221255   8.1357850          0.015625
#> 59: 2022-05-27 04:46:32        1  0.87417955  10.7475009          0.015625
#> 60: 2022-05-27 04:46:32        1  1.93060793   7.8599647          0.015625
#> 61: 2022-05-27 04:46:32        1  3.93551241  14.1085396          0.015625
#> 62: 2022-05-27 04:46:32        1  2.94056187   0.5633341          0.015625
#> 63: 2022-05-27 04:46:32        1  3.40351474  14.4758663          0.015625
#> 64: 2022-05-27 04:46:32        1  7.47114182   7.5870860          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity