Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1      x2 fidelity  x_domain       y
#> 1: 3.097997 2.14199 0.015625 <list[3]> 1.01153

# best scoring evaluation
instance$result
#>          x1      x2 fidelity  x_domain       y
#> 1: 3.097997 2.14199 0.015625 <list[3]> 1.01153

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1: -1.09573620  5.9616256 0.015625     0       6          1  17.701449
#>  2: -4.48405733  9.0236863 0.015625     0       6          1  30.210180
#>  3:  5.17878115  9.6527797 0.015625     0       6          1 136.872349
#>  4:  7.56226493  3.5105922 0.015625     0       6          1  73.413785
#>  5:  5.29401812  5.6440032 0.015625     0       6          1  67.229949
#>  6:  1.33695191 12.6607847 0.015625     0       6          1  88.501657
#>  7: -4.98576934 14.5033772 0.015625     0       6          1  12.630880
#>  8:  7.96791887  9.7710296 0.015625     0       6          1 219.167273
#>  9:  4.29120280 11.1653951 0.015625     0       6          1 136.693266
#> 10:  5.45408589  2.5161996 0.015625     0       6          1  34.822256
#> 11: -4.85754480  4.3686330 0.015625     0       6          1 113.154401
#> 12:  1.52937806 14.3795641 0.015625     0       6          1 125.782303
#> 13:  1.94078102  6.6324187 0.015625     0       6          1  19.527170
#> 14: -1.94833574  0.8579639 0.015625     0       6          1  76.343658
#> 15: -3.98175303  1.7732147 0.015625     0       6          1 125.725893
#> 16: -4.15951659  9.9937793 0.015625     0       6          1  14.932267
#> 17:  4.62526051 13.2359033 0.015625     0       6          1 203.473046
#> 18:  5.71041660  2.2514704 0.015625     0       6          1  36.881439
#> 19:  2.49270447 14.0965073 0.015625     0       6          1 143.310958
#> 20: -4.18841954  5.1122547 0.015625     0       6          1  70.695949
#> 21:  1.45545091  9.6368058 0.015625     0       6          1  45.774772
#> 22:  8.94185023  8.9725482 0.015625     0       6          1 218.926233
#> 23: -2.66311235 11.7072909 0.015625     0       6          1   3.040629
#> 24: -0.06033206  4.7098857 0.015625     0       6          1  21.506325
#> 25: -1.65960805 14.4514527 0.015625     0       6          1  41.928751
#> 26: -0.83687538  5.9395522 0.015625     0       6          1  18.430506
#> 27:  0.77122658 11.4322314 0.015625     0       6          1  60.993298
#> 28:  5.90432060  7.3951565 0.015625     0       6          1 113.405472
#> 29:  0.90882002  6.3065934 0.015625     0       6          1  18.886902
#> 30: -2.83311748 14.2277826 0.015625     0       6          1  12.905522
#> 31:  6.80284797  7.7260283 0.015625     0       6          1 142.215063
#> 32:  5.94250812  9.1290632 0.015625     0       6          1 151.325041
#> 33:  3.09799687  2.1419904 0.015625     0       6          1   1.011530
#> 34:  1.85533971  0.1476621 0.015625     0       6          1  16.336408
#> 35:  7.99783732  7.5926642 0.015625     0       6          1 161.266323
#> 36:  6.62987454 14.6969915 0.015625     0       6          1 339.343178
#> 37:  8.61181085  9.2278953 0.015625     0       6          1 218.131495
#> 38:  3.61483609  1.9087353 0.015625     0       6          1   3.041197
#> 39:  9.67077912 14.3763513 0.015625     0       6          1 437.174835
#> 40:  3.71528599 13.6603636 0.015625     0       6          1 174.831712
#> 41:  1.31660891  3.1080594 0.015625     0       6          1  13.136683
#> 42:  9.44933633 14.7200061 0.015625     0       6          1 441.976070
#> 43:  2.61561976  1.6277302 0.015625     0       6          1   1.871927
#> 44: -2.32223575  7.9705554 0.015625     0       6          1   7.021395
#> 45:  7.03474314  1.3770177 0.015625     0       6          1  42.533844
#> 46:  3.95763600  6.4775269 0.015625     0       6          1  43.044925
#> 47:  3.83095692 11.6676998 0.015625     0       6          1 130.588226
#> 48: -2.54282225 13.9133637 0.015625     0       6          1  15.519135
#> 49:  5.71899090  8.3528750 0.015625     0       6          1 127.301712
#> 50: -1.90834095 13.0762925 0.015625     0       6          1  22.242172
#> 51: -0.51340661  0.2987792 0.015625     0       6          1  60.958558
#> 52:  2.52922529  7.7312675 0.015625     0       6          1  33.055920
#> 53: -4.25305970  8.9983675 0.015625     0       6          1  24.463741
#> 54: -0.23330287  0.2206153 0.015625     0       6          1  57.193646
#> 55:  1.49254196  1.1457768 0.015625     0       6          1  17.239192
#> 56:  5.99714028  4.5713841 0.015625     0       6          1  68.356146
#> 57:  7.92992236 14.6422778 0.015625     0       6          1 382.907128
#> 58:  9.26630268  6.8019546 0.015625     0       6          1 167.177557
#> 59:  1.76227009  5.7090944 0.015625     0       6          1  14.021061
#> 60:  1.77960947 13.7799990 0.015625     0       6          1 118.573862
#> 61:  1.94275938  2.9667983 0.015625     0       6          1   6.513439
#> 62:  1.90778985 10.5197623 0.015625     0       6          1  62.241095
#> 63:  1.25602566  7.1913259 0.015625     0       6          1  22.843997
#> 64: -3.40782991  5.5224595 0.015625     0       6          1  39.903069
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-02-13 04:26:29        1 -1.09573620   5.9616256          0.015625
#>  2: 2022-02-13 04:26:29        1 -4.48405733   9.0236863          0.015625
#>  3: 2022-02-13 04:26:29        1  5.17878115   9.6527797          0.015625
#>  4: 2022-02-13 04:26:29        1  7.56226493   3.5105922          0.015625
#>  5: 2022-02-13 04:26:29        1  5.29401812   5.6440032          0.015625
#>  6: 2022-02-13 04:26:29        1  1.33695191  12.6607847          0.015625
#>  7: 2022-02-13 04:26:29        1 -4.98576934  14.5033772          0.015625
#>  8: 2022-02-13 04:26:29        1  7.96791887   9.7710296          0.015625
#>  9: 2022-02-13 04:26:29        1  4.29120280  11.1653951          0.015625
#> 10: 2022-02-13 04:26:29        1  5.45408589   2.5161996          0.015625
#> 11: 2022-02-13 04:26:29        1 -4.85754480   4.3686330          0.015625
#> 12: 2022-02-13 04:26:29        1  1.52937806  14.3795641          0.015625
#> 13: 2022-02-13 04:26:29        1  1.94078102   6.6324187          0.015625
#> 14: 2022-02-13 04:26:29        1 -1.94833574   0.8579639          0.015625
#> 15: 2022-02-13 04:26:29        1 -3.98175303   1.7732147          0.015625
#> 16: 2022-02-13 04:26:29        1 -4.15951659   9.9937793          0.015625
#> 17: 2022-02-13 04:26:29        1  4.62526051  13.2359033          0.015625
#> 18: 2022-02-13 04:26:29        1  5.71041660   2.2514704          0.015625
#> 19: 2022-02-13 04:26:29        1  2.49270447  14.0965073          0.015625
#> 20: 2022-02-13 04:26:29        1 -4.18841954   5.1122547          0.015625
#> 21: 2022-02-13 04:26:29        1  1.45545091   9.6368058          0.015625
#> 22: 2022-02-13 04:26:29        1  8.94185023   8.9725482          0.015625
#> 23: 2022-02-13 04:26:29        1 -2.66311235  11.7072909          0.015625
#> 24: 2022-02-13 04:26:29        1 -0.06033206   4.7098857          0.015625
#> 25: 2022-02-13 04:26:29        1 -1.65960805  14.4514527          0.015625
#> 26: 2022-02-13 04:26:29        1 -0.83687538   5.9395522          0.015625
#> 27: 2022-02-13 04:26:29        1  0.77122658  11.4322314          0.015625
#> 28: 2022-02-13 04:26:29        1  5.90432060   7.3951565          0.015625
#> 29: 2022-02-13 04:26:29        1  0.90882002   6.3065934          0.015625
#> 30: 2022-02-13 04:26:29        1 -2.83311748  14.2277826          0.015625
#> 31: 2022-02-13 04:26:29        1  6.80284797   7.7260283          0.015625
#> 32: 2022-02-13 04:26:29        1  5.94250812   9.1290632          0.015625
#> 33: 2022-02-13 04:26:29        1  3.09799687   2.1419904          0.015625
#> 34: 2022-02-13 04:26:29        1  1.85533971   0.1476621          0.015625
#> 35: 2022-02-13 04:26:29        1  7.99783732   7.5926642          0.015625
#> 36: 2022-02-13 04:26:29        1  6.62987454  14.6969915          0.015625
#> 37: 2022-02-13 04:26:29        1  8.61181085   9.2278953          0.015625
#> 38: 2022-02-13 04:26:29        1  3.61483609   1.9087353          0.015625
#> 39: 2022-02-13 04:26:29        1  9.67077912  14.3763513          0.015625
#> 40: 2022-02-13 04:26:29        1  3.71528599  13.6603636          0.015625
#> 41: 2022-02-13 04:26:29        1  1.31660891   3.1080594          0.015625
#> 42: 2022-02-13 04:26:29        1  9.44933633  14.7200061          0.015625
#> 43: 2022-02-13 04:26:29        1  2.61561976   1.6277302          0.015625
#> 44: 2022-02-13 04:26:29        1 -2.32223575   7.9705554          0.015625
#> 45: 2022-02-13 04:26:29        1  7.03474314   1.3770177          0.015625
#> 46: 2022-02-13 04:26:29        1  3.95763600   6.4775269          0.015625
#> 47: 2022-02-13 04:26:29        1  3.83095692  11.6676998          0.015625
#> 48: 2022-02-13 04:26:29        1 -2.54282225  13.9133637          0.015625
#> 49: 2022-02-13 04:26:29        1  5.71899090   8.3528750          0.015625
#> 50: 2022-02-13 04:26:29        1 -1.90834095  13.0762925          0.015625
#> 51: 2022-02-13 04:26:29        1 -0.51340661   0.2987792          0.015625
#> 52: 2022-02-13 04:26:29        1  2.52922529   7.7312675          0.015625
#> 53: 2022-02-13 04:26:29        1 -4.25305970   8.9983675          0.015625
#> 54: 2022-02-13 04:26:29        1 -0.23330287   0.2206153          0.015625
#> 55: 2022-02-13 04:26:29        1  1.49254196   1.1457768          0.015625
#> 56: 2022-02-13 04:26:29        1  5.99714028   4.5713841          0.015625
#> 57: 2022-02-13 04:26:29        1  7.92992236  14.6422778          0.015625
#> 58: 2022-02-13 04:26:29        1  9.26630268   6.8019546          0.015625
#> 59: 2022-02-13 04:26:29        1  1.76227009   5.7090944          0.015625
#> 60: 2022-02-13 04:26:29        1  1.77960947  13.7799990          0.015625
#> 61: 2022-02-13 04:26:29        1  1.94275938   2.9667983          0.015625
#> 62: 2022-02-13 04:26:29        1  1.90778985  10.5197623          0.015625
#> 63: 2022-02-13 04:26:29        1  1.25602566   7.1913259          0.015625
#> 64: 2022-02-13 04:26:29        1 -3.40782991   5.5224595          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity