Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>         x1       x2 fidelity  x_domain        y
#> 1: 2.30313 3.394009 0.015625 <list[3]> 4.383604

# best scoring evaluation
instance$result
#>         x1       x2 fidelity  x_domain        y
#> 1: 2.30313 3.394009 0.015625 <list[3]> 4.383604

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1: -4.07480843  5.4602301 0.015625     0       6          1  61.068890
#>  2:  8.00299797 12.2964516 0.015625     0       6          1 299.765940
#>  3:  4.76794885  7.3023077 0.015625     0       6          1  77.638017
#>  4:  1.58946335 12.8552415 0.015625     0       6          1  96.446079
#>  5:  3.92501830  5.8103984 0.015625     0       6          1  34.373242
#>  6:  7.52768149  1.0677648 0.015625     0       6          1  41.232907
#>  7: -0.40195410 13.5773750 0.015625     0       6          1  66.898825
#>  8:  6.78687318  0.1595361 0.015625     0       6          1  30.976087
#>  9:  9.82450349  7.4542837 0.015625     0       6          1 200.606587
#> 10: -0.80103372  9.6218046 0.015625     0       6          1  22.098550
#> 11: -2.86865494  3.8240356 0.015625     0       6          1  49.677662
#> 12: -1.63219429  0.7688604 0.015625     0       6          1  71.991092
#> 13: -4.23164022  8.4130747 0.015625     0       6          1  29.299556
#> 14:  0.16051277  6.1063320 0.015625     0       6          1  19.609006
#> 15:  8.73889813 12.9878365 0.015625     0       6          1 346.602295
#> 16: -0.53465697  5.4831502 0.015625     0       6          1  20.157026
#> 17: -4.56208257  5.0584103 0.015625     0       6          1  86.748344
#> 18: -0.48086542  5.5401443 0.015625     0       6          1  20.031715
#> 19: -3.72283087  2.6300062 0.015625     0       6          1  96.476450
#> 20:  5.53692692  2.3670365 0.015625     0       6          1  34.999836
#> 21: -4.89720749  6.1860667 0.015625     0       6          1  81.411526
#> 22:  4.89230352  4.8805261 0.015625     0       6          1  46.894430
#> 23: -3.53818212  7.2681058 0.015625     0       6          1  23.686667
#> 24: -4.89204218  5.9477975 0.015625     0       6          1  85.228734
#> 25:  2.73875388  9.4112945 0.015625     0       6          1  58.011037
#> 26: -2.11758883 12.3025094 0.015625     0       6          1  12.815979
#> 27:  2.14947739 10.5742484 0.015625     0       6          1  66.420987
#> 28:  7.59432634 12.0750673 0.015625     0       6          1 281.048789
#> 29:  4.23069100 12.5289236 0.015625     0       6          1 167.145024
#> 30: -1.80480984 10.3750594 0.015625     0       6          1   9.740328
#> 31: -4.97942632  7.9902862 0.015625     0       6          1  57.384630
#> 32:  2.62691122  9.6338273 0.015625     0       6          1  59.440027
#> 33:  6.64527383  3.7794530 0.015625     0       6          1  67.950908
#> 34:  2.30312956  3.3940093 0.015625     0       6          1   4.383604
#> 35:  2.48285420 14.0833561 0.015625     0       6          1 142.720470
#> 36:  5.36712635  5.8811887 0.015625     0       6          1  72.661663
#> 37:  3.78757039  3.5054165 0.015625     0       6          1  11.895756
#> 38:  2.64443691  0.2327589 0.015625     0       6          1   4.705612
#> 39:  9.75126220 10.0943134 0.015625     0       6          1 279.470720
#> 40: -4.68416028  0.8826990 0.015625     0       6          1 185.212116
#> 41: -2.33876834  8.4232681 0.015625     0       6          1   5.482237
#> 42: -4.24173411  8.4492128 0.015625     0       6          1  29.216040
#> 43: -4.06120236  4.5577816 0.015625     0       6          1  74.957498
#> 44:  4.16865181 12.5953423 0.015625     0       6          1 166.211911
#> 45: -0.99953806  7.8482678 0.015625     0       6          1  15.243184
#> 46: -0.66928378  8.0892345 0.015625     0       6          1  18.551245
#> 47:  8.34350414  9.9354341 0.015625     0       6          1 232.713812
#> 48:  0.05400778 13.7286646 0.015625     0       6          1  80.655006
#> 49:  0.43497279  4.8852313 0.015625     0       6          1  18.891424
#> 50:  9.20411569  6.4134147 0.015625     0       6          1 155.819122
#> 51:  5.51036124  7.7757651 0.015625     0       6          1 109.268418
#> 52:  4.91724607 12.2305764 0.015625     0       6          1 189.193965
#> 53:  4.95325205  5.0168502 0.015625     0       6          1  50.061684
#> 54:  8.67773156  0.4503055 0.015625     0       6          1  38.309873
#> 55: -0.81086137  1.2368472 0.015625     0       6          1  53.506846
#> 56:  1.31931058  7.9657132 0.015625     0       6          1  28.485112
#> 57:  1.96974914  8.5490152 0.015625     0       6          1  37.235606
#> 58:  9.63150723  0.1163779 0.015625     0       6          1  44.071609
#> 59: -1.04458804 13.5092401 0.015625     0       6          1  48.615762
#> 60:  2.00300158 11.5448840 0.015625     0       6          1  80.099735
#> 61:  9.79319470  9.7774919 0.015625     0       6          1 270.494502
#> 62:  6.69606955  2.9937042 0.015625     0       6          1  58.135910
#> 63:  8.57499508  8.1486822 0.015625     0       6          1 186.867349
#> 64: -2.74345670  7.5313087 0.015625     0       6          1  10.552074
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-07-05 04:40:30        1 -4.07480843   5.4602301          0.015625
#>  2: 2022-07-05 04:40:30        1  8.00299797  12.2964516          0.015625
#>  3: 2022-07-05 04:40:30        1  4.76794885   7.3023077          0.015625
#>  4: 2022-07-05 04:40:30        1  1.58946335  12.8552415          0.015625
#>  5: 2022-07-05 04:40:30        1  3.92501830   5.8103984          0.015625
#>  6: 2022-07-05 04:40:30        1  7.52768149   1.0677648          0.015625
#>  7: 2022-07-05 04:40:30        1 -0.40195410  13.5773750          0.015625
#>  8: 2022-07-05 04:40:30        1  6.78687318   0.1595361          0.015625
#>  9: 2022-07-05 04:40:30        1  9.82450349   7.4542837          0.015625
#> 10: 2022-07-05 04:40:30        1 -0.80103372   9.6218046          0.015625
#> 11: 2022-07-05 04:40:30        1 -2.86865494   3.8240356          0.015625
#> 12: 2022-07-05 04:40:30        1 -1.63219429   0.7688604          0.015625
#> 13: 2022-07-05 04:40:30        1 -4.23164022   8.4130747          0.015625
#> 14: 2022-07-05 04:40:30        1  0.16051277   6.1063320          0.015625
#> 15: 2022-07-05 04:40:30        1  8.73889813  12.9878365          0.015625
#> 16: 2022-07-05 04:40:30        1 -0.53465697   5.4831502          0.015625
#> 17: 2022-07-05 04:40:30        1 -4.56208257   5.0584103          0.015625
#> 18: 2022-07-05 04:40:30        1 -0.48086542   5.5401443          0.015625
#> 19: 2022-07-05 04:40:30        1 -3.72283087   2.6300062          0.015625
#> 20: 2022-07-05 04:40:30        1  5.53692692   2.3670365          0.015625
#> 21: 2022-07-05 04:40:30        1 -4.89720749   6.1860667          0.015625
#> 22: 2022-07-05 04:40:30        1  4.89230352   4.8805261          0.015625
#> 23: 2022-07-05 04:40:30        1 -3.53818212   7.2681058          0.015625
#> 24: 2022-07-05 04:40:30        1 -4.89204218   5.9477975          0.015625
#> 25: 2022-07-05 04:40:30        1  2.73875388   9.4112945          0.015625
#> 26: 2022-07-05 04:40:30        1 -2.11758883  12.3025094          0.015625
#> 27: 2022-07-05 04:40:30        1  2.14947739  10.5742484          0.015625
#> 28: 2022-07-05 04:40:30        1  7.59432634  12.0750673          0.015625
#> 29: 2022-07-05 04:40:30        1  4.23069100  12.5289236          0.015625
#> 30: 2022-07-05 04:40:30        1 -1.80480984  10.3750594          0.015625
#> 31: 2022-07-05 04:40:30        1 -4.97942632   7.9902862          0.015625
#> 32: 2022-07-05 04:40:30        1  2.62691122   9.6338273          0.015625
#> 33: 2022-07-05 04:40:30        1  6.64527383   3.7794530          0.015625
#> 34: 2022-07-05 04:40:30        1  2.30312956   3.3940093          0.015625
#> 35: 2022-07-05 04:40:30        1  2.48285420  14.0833561          0.015625
#> 36: 2022-07-05 04:40:30        1  5.36712635   5.8811887          0.015625
#> 37: 2022-07-05 04:40:30        1  3.78757039   3.5054165          0.015625
#> 38: 2022-07-05 04:40:30        1  2.64443691   0.2327589          0.015625
#> 39: 2022-07-05 04:40:30        1  9.75126220  10.0943134          0.015625
#> 40: 2022-07-05 04:40:30        1 -4.68416028   0.8826990          0.015625
#> 41: 2022-07-05 04:40:30        1 -2.33876834   8.4232681          0.015625
#> 42: 2022-07-05 04:40:30        1 -4.24173411   8.4492128          0.015625
#> 43: 2022-07-05 04:40:30        1 -4.06120236   4.5577816          0.015625
#> 44: 2022-07-05 04:40:30        1  4.16865181  12.5953423          0.015625
#> 45: 2022-07-05 04:40:30        1 -0.99953806   7.8482678          0.015625
#> 46: 2022-07-05 04:40:30        1 -0.66928378   8.0892345          0.015625
#> 47: 2022-07-05 04:40:30        1  8.34350414   9.9354341          0.015625
#> 48: 2022-07-05 04:40:30        1  0.05400778  13.7286646          0.015625
#> 49: 2022-07-05 04:40:30        1  0.43497279   4.8852313          0.015625
#> 50: 2022-07-05 04:40:30        1  9.20411569   6.4134147          0.015625
#> 51: 2022-07-05 04:40:30        1  5.51036124   7.7757651          0.015625
#> 52: 2022-07-05 04:40:30        1  4.91724607  12.2305764          0.015625
#> 53: 2022-07-05 04:40:30        1  4.95325205   5.0168502          0.015625
#> 54: 2022-07-05 04:40:30        1  8.67773156   0.4503055          0.015625
#> 55: 2022-07-05 04:40:30        1 -0.81086137   1.2368472          0.015625
#> 56: 2022-07-05 04:40:30        1  1.31931058   7.9657132          0.015625
#> 57: 2022-07-05 04:40:30        1  1.96974914   8.5490152          0.015625
#> 58: 2022-07-05 04:40:30        1  9.63150723   0.1163779          0.015625
#> 59: 2022-07-05 04:40:30        1 -1.04458804  13.5092401          0.015625
#> 60: 2022-07-05 04:40:30        1  2.00300158  11.5448840          0.015625
#> 61: 2022-07-05 04:40:30        1  9.79319470   9.7774919          0.015625
#> 62: 2022-07-05 04:40:30        1  6.69606955   2.9937042          0.015625
#> 63: 2022-07-05 04:40:30        1  8.57499508   8.1486822          0.015625
#> 64: 2022-07-05 04:40:30        1 -2.74345670   7.5313087          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity