Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.375017 11.86832 0.015625 <list[3]> 0.6797834

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.375017 11.86832 0.015625 <list[3]> 0.6797834

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1:  8.4087395  4.8320204 0.015625     0       6          1 105.7622670
#>  2:  6.2567034  3.7154106 0.015625     0       6          1  61.4548652
#>  3:  4.4420212  3.1032778 0.015625     0       6          1  20.1538201
#>  4:  2.4342344  1.4358220 0.015625     0       6          1   3.4622862
#>  5:  7.7195517  4.5599547 0.015625     0       6          1  92.5344919
#>  6:  8.1035086  3.8549396 0.015625     0       6          1  83.8943473
#>  7:  8.0057604 13.2234490 0.015625     0       6          1 332.3455079
#>  8:  8.4459597  0.9723287 0.015625     0       6          1  43.3450253
#>  9:  2.3838356  0.3050973 0.015625     0       6          1   7.3334868
#> 10: -1.0827985  3.2923169 0.015625     0       6          1  34.4566550
#> 11: -2.4812619  7.6206804 0.015625     0       6          1   8.7550131
#> 12: -2.3326223  3.0012763 0.015625     0       6          1  50.6860422
#> 13:  9.1846716  6.6658486 0.015625     0       6          1 161.7080589
#> 14:  5.3576773  2.3457689 0.015625     0       6          1  31.6966295
#> 15:  0.5328733  0.7417533 0.015625     0       6          1  37.7973183
#> 16: -2.4648846  4.7104580 0.015625     0       6          1  31.6666520
#> 17:  6.7217700 14.8764419 0.015625     0       6          1 349.3967629
#> 18:  4.1232277  1.3254316 0.015625     0       6          1   6.5277377
#> 19: -0.3561281 14.1802064 0.015625     0       6          1  76.9042791
#> 20: -1.4700487  5.5229519 0.015625     0       6          1  19.2783017
#> 21:  3.7350006  5.8745119 0.015625     0       6          1  31.0918045
#> 22:  5.9217811 12.8189360 0.015625     0       6          1 248.9749089
#> 23:  9.2518968 12.0613039 0.015625     0       6          1 330.1190868
#> 24: -3.3750173 11.8683198 0.015625     0       6          1   0.6797834
#> 25: -0.7734194 12.8427357 0.015625     0       6          1  48.1567791
#> 26: -3.9700174  7.4429978 0.015625     0       6          1  32.2391350
#> 27:  7.9677202 12.7522534 0.015625     0       6          1 314.5065400
#> 28:  9.9834715  0.9197549 0.015625     0       6          1  61.8336053
#> 29:  2.4505017 10.9971426 0.015625     0       6          1  78.5105267
#> 30:  4.3263279  2.0452555 0.015625     0       6          1  11.9319541
#> 31: -1.7614561 14.5665436 0.015625     0       6          1  40.3031528
#> 32:  0.9545803  8.8883038 0.015625     0       6          1  34.7299957
#> 33:  8.4422003  3.8117184 0.015625     0       6          1  86.6928094
#> 34: -0.7463375 14.9306361 0.015625     0       6          1  76.7357599
#> 35:  7.3716579  2.7086976 0.015625     0       6          1  60.2897390
#> 36:  8.3291272  7.8838911 0.015625     0       6          1 174.7908125
#> 37: -2.0291287 13.0494668 0.015625     0       6          1  19.3928028
#> 38:  8.1103341 10.6557193 0.015625     0       6          1 249.0959538
#> 39:  4.8127404  0.9835358 0.015625     0       6          1  14.6910129
#> 40: -0.4232826  9.8178963 0.015625     0       6          1  28.6061977
#> 41: -3.4446450  4.0100006 0.015625     0       6          1  62.2564267
#> 42: -0.2253507  4.0767400 0.015625     0       6          1  24.5736037
#> 43:  9.1800236  1.2377080 0.015625     0       6          1  53.3485697
#> 44: -3.3334602  1.3120568 0.015625     0       6          1 107.3856783
#> 45:  3.8229908 13.2403267 0.015625     0       6          1 168.3188050
#> 46: -2.4326946 14.8040715 0.015625     0       6          1  25.2771212
#> 47: -1.1561621 10.1356229 0.015625     0       6          1  18.9507326
#> 48:  2.4514693  8.4255071 0.015625     0       6          1  40.3237662
#> 49:  9.5735254  7.0617263 0.015625     0       6          1 182.2258672
#> 50:  6.0274888 10.8347443 0.015625     0       6          1 196.4655296
#> 51:  3.0253955  5.3783332 0.015625     0       6          1  15.7661602
#> 52:  3.0171500  0.9784589 0.015625     0       6          1   0.7216344
#> 53: -4.5064318 11.6777893 0.015625     0       6          1  12.5257587
#> 54:  9.3906838  3.5168249 0.015625     0       6          1  95.4881827
#> 55: -2.2917684  8.8944470 0.015625     0       6          1   4.4978193
#> 56:  0.6386387  1.1114399 0.015625     0       6          1  32.8003172
#> 57:  4.9914439  6.4033699 0.015625     0       6          1  70.1233042
#> 58:  6.3448272  1.5311326 0.015625     0       6          1  38.8688099
#> 59: -1.4257868  0.4758615 0.015625     0       6          1  73.1019544
#> 60:  7.3870307  2.8897603 0.015625     0       6          1  62.8863690
#> 61:  9.8665826 10.0923002 0.015625     0       6          1 283.6354601
#> 62:  4.8270147  1.1588985 0.015625     0       6          1  15.6135400
#> 63: -3.9973177  2.4007197 0.015625     0       6          1 112.9589597
#> 64: -0.5578589 10.1317330 0.015625     0       6          1  28.6070661
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-07-27 04:46:14        1   8.4087395   4.8320204          0.015625
#>  2: 2022-07-27 04:46:14        1   6.2567034   3.7154106          0.015625
#>  3: 2022-07-27 04:46:14        1   4.4420212   3.1032778          0.015625
#>  4: 2022-07-27 04:46:14        1   2.4342344   1.4358220          0.015625
#>  5: 2022-07-27 04:46:14        1   7.7195517   4.5599547          0.015625
#>  6: 2022-07-27 04:46:14        1   8.1035086   3.8549396          0.015625
#>  7: 2022-07-27 04:46:14        1   8.0057604  13.2234490          0.015625
#>  8: 2022-07-27 04:46:14        1   8.4459597   0.9723287          0.015625
#>  9: 2022-07-27 04:46:14        1   2.3838356   0.3050973          0.015625
#> 10: 2022-07-27 04:46:14        1  -1.0827985   3.2923169          0.015625
#> 11: 2022-07-27 04:46:14        1  -2.4812619   7.6206804          0.015625
#> 12: 2022-07-27 04:46:14        1  -2.3326223   3.0012763          0.015625
#> 13: 2022-07-27 04:46:14        1   9.1846716   6.6658486          0.015625
#> 14: 2022-07-27 04:46:14        1   5.3576773   2.3457689          0.015625
#> 15: 2022-07-27 04:46:14        1   0.5328733   0.7417533          0.015625
#> 16: 2022-07-27 04:46:14        1  -2.4648846   4.7104580          0.015625
#> 17: 2022-07-27 04:46:14        1   6.7217700  14.8764419          0.015625
#> 18: 2022-07-27 04:46:14        1   4.1232277   1.3254316          0.015625
#> 19: 2022-07-27 04:46:14        1  -0.3561281  14.1802064          0.015625
#> 20: 2022-07-27 04:46:14        1  -1.4700487   5.5229519          0.015625
#> 21: 2022-07-27 04:46:14        1   3.7350006   5.8745119          0.015625
#> 22: 2022-07-27 04:46:14        1   5.9217811  12.8189360          0.015625
#> 23: 2022-07-27 04:46:14        1   9.2518968  12.0613039          0.015625
#> 24: 2022-07-27 04:46:14        1  -3.3750173  11.8683198          0.015625
#> 25: 2022-07-27 04:46:14        1  -0.7734194  12.8427357          0.015625
#> 26: 2022-07-27 04:46:14        1  -3.9700174   7.4429978          0.015625
#> 27: 2022-07-27 04:46:14        1   7.9677202  12.7522534          0.015625
#> 28: 2022-07-27 04:46:14        1   9.9834715   0.9197549          0.015625
#> 29: 2022-07-27 04:46:14        1   2.4505017  10.9971426          0.015625
#> 30: 2022-07-27 04:46:14        1   4.3263279   2.0452555          0.015625
#> 31: 2022-07-27 04:46:14        1  -1.7614561  14.5665436          0.015625
#> 32: 2022-07-27 04:46:14        1   0.9545803   8.8883038          0.015625
#> 33: 2022-07-27 04:46:14        1   8.4422003   3.8117184          0.015625
#> 34: 2022-07-27 04:46:14        1  -0.7463375  14.9306361          0.015625
#> 35: 2022-07-27 04:46:14        1   7.3716579   2.7086976          0.015625
#> 36: 2022-07-27 04:46:14        1   8.3291272   7.8838911          0.015625
#> 37: 2022-07-27 04:46:14        1  -2.0291287  13.0494668          0.015625
#> 38: 2022-07-27 04:46:14        1   8.1103341  10.6557193          0.015625
#> 39: 2022-07-27 04:46:14        1   4.8127404   0.9835358          0.015625
#> 40: 2022-07-27 04:46:14        1  -0.4232826   9.8178963          0.015625
#> 41: 2022-07-27 04:46:14        1  -3.4446450   4.0100006          0.015625
#> 42: 2022-07-27 04:46:14        1  -0.2253507   4.0767400          0.015625
#> 43: 2022-07-27 04:46:14        1   9.1800236   1.2377080          0.015625
#> 44: 2022-07-27 04:46:14        1  -3.3334602   1.3120568          0.015625
#> 45: 2022-07-27 04:46:14        1   3.8229908  13.2403267          0.015625
#> 46: 2022-07-27 04:46:14        1  -2.4326946  14.8040715          0.015625
#> 47: 2022-07-27 04:46:14        1  -1.1561621  10.1356229          0.015625
#> 48: 2022-07-27 04:46:14        1   2.4514693   8.4255071          0.015625
#> 49: 2022-07-27 04:46:14        1   9.5735254   7.0617263          0.015625
#> 50: 2022-07-27 04:46:14        1   6.0274888  10.8347443          0.015625
#> 51: 2022-07-27 04:46:14        1   3.0253955   5.3783332          0.015625
#> 52: 2022-07-27 04:46:14        1   3.0171500   0.9784589          0.015625
#> 53: 2022-07-27 04:46:14        1  -4.5064318  11.6777893          0.015625
#> 54: 2022-07-27 04:46:14        1   9.3906838   3.5168249          0.015625
#> 55: 2022-07-27 04:46:14        1  -2.2917684   8.8944470          0.015625
#> 56: 2022-07-27 04:46:14        1   0.6386387   1.1114399          0.015625
#> 57: 2022-07-27 04:46:14        1   4.9914439   6.4033699          0.015625
#> 58: 2022-07-27 04:46:14        1   6.3448272   1.5311326          0.015625
#> 59: 2022-07-27 04:46:14        1  -1.4257868   0.4758615          0.015625
#> 60: 2022-07-27 04:46:14        1   7.3870307   2.8897603          0.015625
#> 61: 2022-07-27 04:46:14        1   9.8665826  10.0923002          0.015625
#> 62: 2022-07-27 04:46:14        1   4.8270147   1.1588985          0.015625
#> 63: 2022-07-27 04:46:14        1  -3.9973177   2.4007197          0.015625
#> 64: 2022-07-27 04:46:14        1  -0.5578589  10.1317330          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity