Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>         x1        x2 fidelity  x_domain         y
#> 1: 3.39856 0.7394596 0.015625 <list[3]> 0.7558951

# best scoring evaluation
instance$result
#>         x1        x2 fidelity  x_domain         y
#> 1: 3.39856 0.7394596 0.015625 <list[3]> 0.7558951

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -2.3142014 10.6319233 0.015625     0       6          1   4.1161023
#>  2:  2.5930858  3.7138686 0.015625     0       6          1   4.4769154
#>  3:  0.1440637  0.3555091 0.015625     0       6          1  48.8340135
#>  4: -4.5208811  3.5355646 0.015625     0       6          1 114.0165103
#>  5: -3.8127176 13.2883524 0.015625     0       6          1   3.0783001
#>  6:  8.4688125  6.4109248 0.015625     0       6          1 140.9832116
#>  7:  0.4017496  6.4182652 0.015625     0       6          1  19.9457676
#>  8:  4.5032669  8.8544671 0.015625     0       6          1  96.3310173
#>  9:  7.8252956  6.6133281 0.015625     0       6          1 135.3767706
#> 10:  1.9694723  0.5533211 0.015625     0       6          1  12.1843182
#> 11:  7.3580287 14.0703020 0.015625     0       6          1 342.7699475
#> 12: -4.8102939  6.7601013 0.015625     0       6          1  68.8076586
#> 13:  1.3573569  3.3468765 0.015625     0       6          1  12.3358633
#> 14:  9.9713144 14.8366663 0.015625     0       6          1 470.4940096
#> 15:  6.7583749 10.3488115 0.015625     0       6          1 206.2477151
#> 16:  1.1921185  9.2789093 0.015625     0       6          1  39.8926791
#> 17:  7.3580937  9.9957609 0.015625     0       6          1 211.7420862
#> 18:  0.5658838  7.8900054 0.015625     0       6          1  25.8380896
#> 19:  0.6424979 13.3537697 0.015625     0       6          1  87.6380144
#> 20:  7.5230903 11.4861995 0.015625     0       6          1 260.2186468
#> 21:  4.1739904  4.4889226 0.015625     0       6          1  26.2028767
#> 22:  5.3867658 10.4243493 0.015625     0       6          1 162.5378204
#> 23:  5.5420919 10.5801914 0.015625     0       6          1 172.2436074
#> 24:  2.0037521  6.2596354 0.015625     0       6          1  17.0287064
#> 25:  7.1838373  2.0878274 0.015625     0       6          1  51.1819484
#> 26: -4.0929109 14.5207340 0.015625     0       6          1   6.6497629
#> 27:  9.2011459 14.3577175 0.015625     0       6          1 416.7448088
#> 28:  1.8817865 10.5689109 0.015625     0       6          1  62.6385895
#> 29:  2.2745389 11.0098841 0.015625     0       6          1  75.5420864
#> 30:  1.6311380  7.1389151 0.015625     0       6          1  22.7664200
#> 31:  7.4702124  5.5935177 0.015625     0       6          1 108.9879640
#> 32:  3.2117047 14.6434798 0.015625     0       6          1 180.9988949
#> 33:  5.0973361  5.2840092 0.015625     0       6          1  57.1363528
#> 34:  2.1697661  7.7475779 0.015625     0       6          1  30.1507008
#> 35:  2.2674995  4.9699568 0.015625     0       6          1   9.6982170
#> 36: -2.1995505 10.3869106 0.015625     0       6          1   4.8964871
#> 37:  8.9750616 12.6072151 0.015625     0       6          1 340.4554348
#> 38:  7.4106784  3.6284691 0.015625     0       6          1  73.9389351
#> 39:  8.4681144  2.1502404 0.015625     0       6          1  59.5651431
#> 40: -2.9100103 14.4888097 0.015625     0       6          1  13.5927298
#> 41:  0.5630366  3.1723787 0.015625     0       6          1  21.8884339
#> 42: -3.0305086  9.6814626 0.015625     0       6          1   2.4851924
#> 43: -3.7063401 13.0627328 0.015625     0       6          1   2.4387470
#> 44:  5.6254867  0.4346213 0.015625     0       6          1  23.4305675
#> 45: -4.8406151  2.6837616 0.015625     0       6          1 149.0734778
#> 46:  2.4827415  7.8751260 0.015625     0       6          1  34.1834963
#> 47: -0.5593091  0.6134299 0.015625     0       6          1  57.6572408
#> 48: -0.6336475 14.6926596 0.015625     0       6          1  76.5951093
#> 49:  4.0929661  8.9462287 0.015625     0       6          1  84.4437596
#> 50:  3.2763429 11.5860920 0.015625     0       6          1 110.1163469
#> 51: -3.9679885 14.0796646 0.015625     0       6          1   5.1334523
#> 52:  7.7601838  8.4947198 0.015625     0       6          1 179.7391404
#> 53:  1.3447835 11.8403556 0.015625     0       6          1  74.9580434
#> 54:  3.7569692  2.3782130 0.015625     0       6          1   5.8596765
#> 55:  3.4576029 11.5195692 0.015625     0       6          1 114.4009753
#> 56:  3.3985599  0.7394596 0.015625     0       6          1   0.7558951
#> 57: -4.1900803  3.6792535 0.015625     0       6          1  96.0169432
#> 58:  2.5212307  1.7885217 0.015625     0       6          1   2.3425094
#> 59:  1.0593619  6.5384679 0.015625     0       6          1  19.4956007
#> 60:  4.6667214  3.6761764 0.015625     0       6          1  29.2209126
#> 61:  3.4927467  4.3105362 0.015625     0       6          1  13.1941421
#> 62: -3.3803604  8.9811863 0.015625     0       6          1   8.2337069
#> 63: -1.6020985  5.2187018 0.015625     0       6          1  21.3278280
#> 64:  9.1717444 10.0542768 0.015625     0       6          1 258.7909343
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-25 05:00:38        1  -2.3142014  10.6319233          0.015625
#>  2: 2022-08-25 05:00:38        1   2.5930858   3.7138686          0.015625
#>  3: 2022-08-25 05:00:38        1   0.1440637   0.3555091          0.015625
#>  4: 2022-08-25 05:00:38        1  -4.5208811   3.5355646          0.015625
#>  5: 2022-08-25 05:00:38        1  -3.8127176  13.2883524          0.015625
#>  6: 2022-08-25 05:00:38        1   8.4688125   6.4109248          0.015625
#>  7: 2022-08-25 05:00:38        1   0.4017496   6.4182652          0.015625
#>  8: 2022-08-25 05:00:38        1   4.5032669   8.8544671          0.015625
#>  9: 2022-08-25 05:00:38        1   7.8252956   6.6133281          0.015625
#> 10: 2022-08-25 05:00:38        1   1.9694723   0.5533211          0.015625
#> 11: 2022-08-25 05:00:38        1   7.3580287  14.0703020          0.015625
#> 12: 2022-08-25 05:00:38        1  -4.8102939   6.7601013          0.015625
#> 13: 2022-08-25 05:00:38        1   1.3573569   3.3468765          0.015625
#> 14: 2022-08-25 05:00:38        1   9.9713144  14.8366663          0.015625
#> 15: 2022-08-25 05:00:38        1   6.7583749  10.3488115          0.015625
#> 16: 2022-08-25 05:00:38        1   1.1921185   9.2789093          0.015625
#> 17: 2022-08-25 05:00:38        1   7.3580937   9.9957609          0.015625
#> 18: 2022-08-25 05:00:38        1   0.5658838   7.8900054          0.015625
#> 19: 2022-08-25 05:00:38        1   0.6424979  13.3537697          0.015625
#> 20: 2022-08-25 05:00:38        1   7.5230903  11.4861995          0.015625
#> 21: 2022-08-25 05:00:38        1   4.1739904   4.4889226          0.015625
#> 22: 2022-08-25 05:00:38        1   5.3867658  10.4243493          0.015625
#> 23: 2022-08-25 05:00:38        1   5.5420919  10.5801914          0.015625
#> 24: 2022-08-25 05:00:38        1   2.0037521   6.2596354          0.015625
#> 25: 2022-08-25 05:00:38        1   7.1838373   2.0878274          0.015625
#> 26: 2022-08-25 05:00:38        1  -4.0929109  14.5207340          0.015625
#> 27: 2022-08-25 05:00:38        1   9.2011459  14.3577175          0.015625
#> 28: 2022-08-25 05:00:38        1   1.8817865  10.5689109          0.015625
#> 29: 2022-08-25 05:00:38        1   2.2745389  11.0098841          0.015625
#> 30: 2022-08-25 05:00:38        1   1.6311380   7.1389151          0.015625
#> 31: 2022-08-25 05:00:38        1   7.4702124   5.5935177          0.015625
#> 32: 2022-08-25 05:00:38        1   3.2117047  14.6434798          0.015625
#> 33: 2022-08-25 05:00:38        1   5.0973361   5.2840092          0.015625
#> 34: 2022-08-25 05:00:38        1   2.1697661   7.7475779          0.015625
#> 35: 2022-08-25 05:00:38        1   2.2674995   4.9699568          0.015625
#> 36: 2022-08-25 05:00:38        1  -2.1995505  10.3869106          0.015625
#> 37: 2022-08-25 05:00:38        1   8.9750616  12.6072151          0.015625
#> 38: 2022-08-25 05:00:38        1   7.4106784   3.6284691          0.015625
#> 39: 2022-08-25 05:00:38        1   8.4681144   2.1502404          0.015625
#> 40: 2022-08-25 05:00:38        1  -2.9100103  14.4888097          0.015625
#> 41: 2022-08-25 05:00:38        1   0.5630366   3.1723787          0.015625
#> 42: 2022-08-25 05:00:38        1  -3.0305086   9.6814626          0.015625
#> 43: 2022-08-25 05:00:38        1  -3.7063401  13.0627328          0.015625
#> 44: 2022-08-25 05:00:38        1   5.6254867   0.4346213          0.015625
#> 45: 2022-08-25 05:00:38        1  -4.8406151   2.6837616          0.015625
#> 46: 2022-08-25 05:00:38        1   2.4827415   7.8751260          0.015625
#> 47: 2022-08-25 05:00:38        1  -0.5593091   0.6134299          0.015625
#> 48: 2022-08-25 05:00:38        1  -0.6336475  14.6926596          0.015625
#> 49: 2022-08-25 05:00:38        1   4.0929661   8.9462287          0.015625
#> 50: 2022-08-25 05:00:38        1   3.2763429  11.5860920          0.015625
#> 51: 2022-08-25 05:00:38        1  -3.9679885  14.0796646          0.015625
#> 52: 2022-08-25 05:00:38        1   7.7601838   8.4947198          0.015625
#> 53: 2022-08-25 05:00:38        1   1.3447835  11.8403556          0.015625
#> 54: 2022-08-25 05:00:38        1   3.7569692   2.3782130          0.015625
#> 55: 2022-08-25 05:00:38        1   3.4576029  11.5195692          0.015625
#> 56: 2022-08-25 05:00:38        1   3.3985599   0.7394596          0.015625
#> 57: 2022-08-25 05:00:38        1  -4.1900803   3.6792535          0.015625
#> 58: 2022-08-25 05:00:38        1   2.5212307   1.7885217          0.015625
#> 59: 2022-08-25 05:00:38        1   1.0593619   6.5384679          0.015625
#> 60: 2022-08-25 05:00:38        1   4.6667214   3.6761764          0.015625
#> 61: 2022-08-25 05:00:38        1   3.4927467   4.3105362          0.015625
#> 62: 2022-08-25 05:00:38        1  -3.3803604   8.9811863          0.015625
#> 63: 2022-08-25 05:00:38        1  -1.6020985   5.2187018          0.015625
#> 64: 2022-08-25 05:00:38        1   9.1717444  10.0542768          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity