Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.259641 2.834118 0.015625 <list[3]> 3.971237

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.259641 2.834118 0.015625 <list[3]> 3.971237

# all evaluations
as.data.table(instance$archive)
#>             x1          x2 fidelity stage bracket repetition          y
#>  1:  8.6943254  9.73496772 0.015625     0       6          1 235.355563
#>  2:  4.8642945  7.65417820 0.015625     0       6          1  86.594641
#>  3: -1.2868757  0.04897448 0.015625     0       6          1  77.493391
#>  4:  7.7955423  6.54998604 0.015625     0       6          1 133.514910
#>  5: -0.7297738 13.13281171 0.015625     0       6          1  52.618314
#>  6:  6.1271020 14.25816537 0.015625     0       6          1 303.592117
#>  7:  8.9143142  1.74882192 0.015625     0       6          1  57.768271
#>  8: -0.6161075 14.63372992 0.015625     0       6          1  76.229029
#>  9:  2.1957423  8.10699377 0.015625     0       6          1  34.121671
#> 10: -4.3616080  1.81675728 0.015625     0       6          1 143.821633
#> 11:  9.1972951 14.35673164 0.015625     0       6          1 416.551716
#> 12:  0.7609237  8.63086716 0.015625     0       6          1  31.577677
#> 13:  7.0560879 10.14995636 0.015625     0       6          1 208.675072
#> 14:  0.8726511  7.91981191 0.015625     0       6          1  26.965183
#> 15: -1.3592338  8.81783431 0.015625     0       6          1  12.373621
#> 16:  3.9076674  2.06293891 0.015625     0       6          1   6.366173
#> 17:  9.2621578  9.58860011 0.015625     0       6          1 246.765665
#> 18: -4.5152928  8.36961333 0.015625     0       6          1  37.752084
#> 19:  9.9660513  3.07904243 0.015625     0       6          1  99.516243
#> 20:  0.1495383  0.35149579 0.015625     0       6          1  48.775975
#> 21: -3.4171628 14.73064716 0.015625     0       6          1   9.362842
#> 22:  2.9903491 14.88585709 0.015625     0       6          1 179.269748
#> 23: -0.8029516  3.08188811 0.015625     0       6          1  34.443078
#> 24:  8.3326385  1.95158429 0.015625     0       6          1  55.683080
#> 25:  9.1046110 11.48819774 0.015625     0       6          1 304.687094
#> 26:  5.4779924 14.86741123 0.015625     0       6          1 294.317508
#> 27: -4.7586778 12.40767627 0.015625     0       6          1  13.912328
#> 28:  7.5363850 10.25502738 0.015625     0       6          1 223.341855
#> 29: -0.2310638 13.70537912 0.015625     0       6          1  73.163640
#> 30:  0.4581461 12.15399455 0.015625     0       6          1  65.900929
#> 31: -1.6946322  7.58506671 0.015625     0       6          1  10.254725
#> 32: -3.8477323 10.28776994 0.015625     0       6          1   7.944058
#> 33:  0.3663318  5.49031376 0.015625     0       6          1  18.969781
#> 34: -0.4981892 12.95285675 0.015625     0       6          1  56.286169
#> 35:  2.2596409  2.83411815 0.015625     0       6          1   3.971237
#> 36: -0.0153148 14.39790592 0.015625     0       6          1  89.716898
#> 37: -4.5445299  7.21602007 0.015625     0       6          1  52.642657
#> 38:  3.5513811 14.92986638 0.015625     0       6          1 202.670313
#> 39:  0.7128051  1.75900210 0.015625     0       6          1  27.012146
#> 40:  4.3261663 12.43229285 0.015625     0       6          1 168.745300
#> 41: -0.1726339 14.18714115 0.015625     0       6          1  82.050733
#> 42:  7.2030420 13.44594007 0.015625     0       6          1 315.615885
#> 43: -4.1494122  4.95390852 0.015625     0       6          1  71.779131
#> 44: -3.9584997  0.56567947 0.015625     0       6          1 152.664696
#> 45: -2.8536835  6.59923668 0.015625     0       6          1  18.373777
#> 46:  1.2250227 14.65465629 0.015625     0       6          1 124.729958
#> 47:  4.8593448 12.18076979 0.015625     0       6          1 185.345735
#> 48:  3.8933728  7.79776390 0.015625     0       6          1  59.659476
#> 49: -0.7047903 11.66403589 0.015625     0       6          1  37.808607
#> 50:  5.5653793  3.97969717 0.015625     0       6          1  51.865229
#> 51:  4.5695224  4.89316044 0.015625     0       6          1  39.144846
#> 52:  6.0268578  4.26652912 0.015625     0       6          1  64.739543
#> 53: -0.2316791  8.02302719 0.015625     0       6          1  22.076811
#> 54:  4.0960983  1.31215368 0.015625     0       6          1   6.180198
#> 55:  9.3474665  3.37738311 0.015625     0       6          1  91.969637
#> 56:  4.1688770  4.70734763 0.015625     0       6          1  28.150723
#> 57:  5.2695667 10.88804773 0.015625     0       6          1 169.359526
#> 58:  6.3295641  2.50574245 0.015625     0       6          1  48.189976
#> 59:  5.1762219  6.29589245 0.015625     0       6          1  73.744397
#> 60:  4.1805508  0.38700682 0.015625     0       6          1   5.383796
#> 61:  2.1476704  9.68727416 0.015625     0       6          1  53.254358
#> 62:  9.2111704 12.31326620 0.015625     0       6          1 337.872775
#> 63:  1.2058666  7.57586147 0.015625     0       6          1  25.331739
#> 64:  1.4850398  7.52203981 0.015625     0       6          1  25.397622
#>             x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-04-26 04:38:11        1   8.6943254  9.73496772          0.015625
#>  2: 2022-04-26 04:38:11        1   4.8642945  7.65417820          0.015625
#>  3: 2022-04-26 04:38:11        1  -1.2868757  0.04897448          0.015625
#>  4: 2022-04-26 04:38:11        1   7.7955423  6.54998604          0.015625
#>  5: 2022-04-26 04:38:11        1  -0.7297738 13.13281171          0.015625
#>  6: 2022-04-26 04:38:11        1   6.1271020 14.25816537          0.015625
#>  7: 2022-04-26 04:38:11        1   8.9143142  1.74882192          0.015625
#>  8: 2022-04-26 04:38:11        1  -0.6161075 14.63372992          0.015625
#>  9: 2022-04-26 04:38:11        1   2.1957423  8.10699377          0.015625
#> 10: 2022-04-26 04:38:11        1  -4.3616080  1.81675728          0.015625
#> 11: 2022-04-26 04:38:11        1   9.1972951 14.35673164          0.015625
#> 12: 2022-04-26 04:38:11        1   0.7609237  8.63086716          0.015625
#> 13: 2022-04-26 04:38:11        1   7.0560879 10.14995636          0.015625
#> 14: 2022-04-26 04:38:11        1   0.8726511  7.91981191          0.015625
#> 15: 2022-04-26 04:38:11        1  -1.3592338  8.81783431          0.015625
#> 16: 2022-04-26 04:38:11        1   3.9076674  2.06293891          0.015625
#> 17: 2022-04-26 04:38:11        1   9.2621578  9.58860011          0.015625
#> 18: 2022-04-26 04:38:11        1  -4.5152928  8.36961333          0.015625
#> 19: 2022-04-26 04:38:11        1   9.9660513  3.07904243          0.015625
#> 20: 2022-04-26 04:38:11        1   0.1495383  0.35149579          0.015625
#> 21: 2022-04-26 04:38:11        1  -3.4171628 14.73064716          0.015625
#> 22: 2022-04-26 04:38:11        1   2.9903491 14.88585709          0.015625
#> 23: 2022-04-26 04:38:11        1  -0.8029516  3.08188811          0.015625
#> 24: 2022-04-26 04:38:11        1   8.3326385  1.95158429          0.015625
#> 25: 2022-04-26 04:38:11        1   9.1046110 11.48819774          0.015625
#> 26: 2022-04-26 04:38:11        1   5.4779924 14.86741123          0.015625
#> 27: 2022-04-26 04:38:11        1  -4.7586778 12.40767627          0.015625
#> 28: 2022-04-26 04:38:11        1   7.5363850 10.25502738          0.015625
#> 29: 2022-04-26 04:38:11        1  -0.2310638 13.70537912          0.015625
#> 30: 2022-04-26 04:38:11        1   0.4581461 12.15399455          0.015625
#> 31: 2022-04-26 04:38:11        1  -1.6946322  7.58506671          0.015625
#> 32: 2022-04-26 04:38:11        1  -3.8477323 10.28776994          0.015625
#> 33: 2022-04-26 04:38:11        1   0.3663318  5.49031376          0.015625
#> 34: 2022-04-26 04:38:11        1  -0.4981892 12.95285675          0.015625
#> 35: 2022-04-26 04:38:11        1   2.2596409  2.83411815          0.015625
#> 36: 2022-04-26 04:38:11        1  -0.0153148 14.39790592          0.015625
#> 37: 2022-04-26 04:38:11        1  -4.5445299  7.21602007          0.015625
#> 38: 2022-04-26 04:38:11        1   3.5513811 14.92986638          0.015625
#> 39: 2022-04-26 04:38:11        1   0.7128051  1.75900210          0.015625
#> 40: 2022-04-26 04:38:11        1   4.3261663 12.43229285          0.015625
#> 41: 2022-04-26 04:38:11        1  -0.1726339 14.18714115          0.015625
#> 42: 2022-04-26 04:38:11        1   7.2030420 13.44594007          0.015625
#> 43: 2022-04-26 04:38:11        1  -4.1494122  4.95390852          0.015625
#> 44: 2022-04-26 04:38:11        1  -3.9584997  0.56567947          0.015625
#> 45: 2022-04-26 04:38:11        1  -2.8536835  6.59923668          0.015625
#> 46: 2022-04-26 04:38:11        1   1.2250227 14.65465629          0.015625
#> 47: 2022-04-26 04:38:11        1   4.8593448 12.18076979          0.015625
#> 48: 2022-04-26 04:38:11        1   3.8933728  7.79776390          0.015625
#> 49: 2022-04-26 04:38:11        1  -0.7047903 11.66403589          0.015625
#> 50: 2022-04-26 04:38:11        1   5.5653793  3.97969717          0.015625
#> 51: 2022-04-26 04:38:11        1   4.5695224  4.89316044          0.015625
#> 52: 2022-04-26 04:38:11        1   6.0268578  4.26652912          0.015625
#> 53: 2022-04-26 04:38:11        1  -0.2316791  8.02302719          0.015625
#> 54: 2022-04-26 04:38:11        1   4.0960983  1.31215368          0.015625
#> 55: 2022-04-26 04:38:11        1   9.3474665  3.37738311          0.015625
#> 56: 2022-04-26 04:38:11        1   4.1688770  4.70734763          0.015625
#> 57: 2022-04-26 04:38:11        1   5.2695667 10.88804773          0.015625
#> 58: 2022-04-26 04:38:11        1   6.3295641  2.50574245          0.015625
#> 59: 2022-04-26 04:38:11        1   5.1762219  6.29589245          0.015625
#> 60: 2022-04-26 04:38:11        1   4.1805508  0.38700682          0.015625
#> 61: 2022-04-26 04:38:11        1   2.1476704  9.68727416          0.015625
#> 62: 2022-04-26 04:38:11        1   9.2111704 12.31326620          0.015625
#> 63: 2022-04-26 04:38:11        1   1.2058666  7.57586147          0.015625
#> 64: 2022-04-26 04:38:11        1   1.4850398  7.52203981          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity