Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1      x2 fidelity  x_domain         y
#> 1: -3.392499 12.2585 0.015625 <list[3]> 0.9538816

# best scoring evaluation
instance$result
#>           x1      x2 fidelity  x_domain         y
#> 1: -3.392499 12.2585 0.015625 <list[3]> 0.9538816

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -2.0069032 10.5195199 0.015625     0       6          1   7.3877667
#>  2:  8.2422879  5.9699234 0.015625     0       6          1 127.3451738
#>  3: -2.3062730  7.1579948 0.015625     0       6          1  10.7190246
#>  4: -0.4768173  8.8783012 0.015625     0       6          1  22.9934651
#>  5:  9.7840820 13.3821564 0.015625     0       6          1 401.4371283
#>  6: -0.7431417 11.1359402 0.015625     0       6          1  32.5642555
#>  7: -4.9411895  1.9485630 0.015625     0       6          1 172.6125847
#>  8:  7.6231417  7.1744983 0.015625     0       6          1 144.9148305
#>  9:  9.3598638 11.0525572 0.015625     0       6          1 298.1734418
#> 10: -3.8363330  6.2837047 0.015625     0       6          1  41.9931481
#> 11: -4.9597330  0.5419594 0.015625     0       6          1 211.3878136
#> 12: -4.7232748  3.8889365 0.015625     0       6          1 116.4900072
#> 13: -3.7331722  7.9307152 0.015625     0       6          1  21.7372317
#> 14: -4.5904669  9.9233577 0.015625     0       6          1  25.0772002
#> 15: -3.3924993 12.2585014 0.015625     0       6          1   0.9538816
#> 16: -2.5078499  6.8156901 0.015625     0       6          1  13.6129844
#> 17:  5.8570452  7.6335475 0.015625     0       6          1 116.7642461
#> 18:  1.6489742  5.4766519 0.015625     0       6          1  13.3202816
#> 19: -4.8970783  6.3018270 0.015625     0       6          1  79.4875342
#> 20:  0.4409220 14.0265603 0.015625     0       6          1  94.7628246
#> 21: -0.4947686  1.9107887 0.015625     0       6          1  42.3058944
#> 22:  6.8485596  9.3046913 0.015625     0       6          1 180.9865744
#> 23:  9.7866739  3.4979050 0.015625     0       6          1 103.6157327
#> 24:  8.9087844 11.5701753 0.015625     0       6          1 301.2379501
#> 25:  0.7124457 14.3602306 0.015625     0       6          1 107.1088256
#> 26: -0.7105858  0.4428729 0.015625     0       6          1  62.2162556
#> 27: -4.5794437  9.4081209 0.015625     0       6          1  29.2036865
#> 28:  6.8087525  6.6489422 0.015625     0       6          1 119.5098957
#> 29:  4.0287601  7.6083393 0.015625     0       6          1  60.5045316
#> 30:  5.3809476 13.3547893 0.015625     0       6          1 241.8097233
#> 31:  5.2824720  6.0793443 0.015625     0       6          1  73.3790259
#> 32: -0.8213706  9.4680380 0.015625     0       6          1  21.1209185
#> 33:  5.3802357  6.9071066 0.015625     0       6          1  89.5627164
#> 34: -1.3721263  0.7958979 0.015625     0       6          1  67.3350236
#> 35:  9.6488817  8.2244004 0.015625     0       6          1 217.2721759
#> 36:  5.2113158 11.2023538 0.015625     0       6          1 174.9054774
#> 37: -3.8297096 14.9023807 0.015625     0       6          1   8.1348368
#> 38: -2.7812194 11.3250842 0.015625     0       6          1   1.4513364
#> 39:  2.0323858  6.7833018 0.015625     0       6          1  20.8629107
#> 40:  7.9535442  0.1548866 0.015625     0       6          1  32.7461503
#> 41:  1.4488799 11.0640525 0.015625     0       6          1  64.5376600
#> 42: -3.2594281  2.5369624 0.015625     0       6          1  81.0551448
#> 43:  6.0864404  5.9307203 0.015625     0       6          1  91.3031136
#> 44:  4.5932053  0.9746272 0.015625     0       6          1  11.5356203
#> 45: -2.9594081 11.7097184 0.015625     0       6          1   1.0902677
#> 46:  5.1350394  9.6131330 0.015625     0       6          1 134.3902119
#> 47: -3.5854158  9.0245135 0.015625     0       6          1  10.7967918
#> 48:  0.2832746  6.2923561 0.015625     0       6          1  19.7681111
#> 49:  4.3604824  7.6930383 0.015625     0       6          1  71.4661047
#> 50:  0.6915248 14.9849852 0.015625     0       6          1 118.8188397
#> 51:  2.6303060  0.8561279 0.015625     0       6          1   2.9955280
#> 52: -2.7607172  8.8992841 0.015625     0       6          1   4.0749842
#> 53: -2.9390608  5.7271616 0.015625     0       6          1  27.8017880
#> 54:  3.8711457  3.8333866 0.015625     0       6          1  15.3292005
#> 55: -0.8504421  9.5217116 0.015625     0       6          1  20.9391575
#> 56: -2.2533757  2.6108227 0.015625     0       6          1  54.8035981
#> 57: -1.4291925  0.3862328 0.015625     0       6          1  74.5767584
#> 58:  5.3325337  6.3877811 0.015625     0       6          1  79.5875408
#> 59: -0.2611272  4.9341849 0.015625     0       6          1  21.4773944
#> 60:  9.0643664 14.9178418 0.015625     0       6          1 434.4027048
#> 61: -2.8281980  1.8030657 0.015625     0       6          1  80.8622821
#> 62: -0.3459762  2.5573745 0.015625     0       6          1  35.0087003
#> 63:  6.6041261  4.6712230 0.015625     0       6          1  80.5931207
#> 64: -3.7425562  6.9512045 0.015625     0       6          1  31.6295294
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-07-26 04:59:26        1  -2.0069032  10.5195199          0.015625
#>  2: 2022-07-26 04:59:26        1   8.2422879   5.9699234          0.015625
#>  3: 2022-07-26 04:59:26        1  -2.3062730   7.1579948          0.015625
#>  4: 2022-07-26 04:59:26        1  -0.4768173   8.8783012          0.015625
#>  5: 2022-07-26 04:59:26        1   9.7840820  13.3821564          0.015625
#>  6: 2022-07-26 04:59:26        1  -0.7431417  11.1359402          0.015625
#>  7: 2022-07-26 04:59:26        1  -4.9411895   1.9485630          0.015625
#>  8: 2022-07-26 04:59:26        1   7.6231417   7.1744983          0.015625
#>  9: 2022-07-26 04:59:26        1   9.3598638  11.0525572          0.015625
#> 10: 2022-07-26 04:59:26        1  -3.8363330   6.2837047          0.015625
#> 11: 2022-07-26 04:59:26        1  -4.9597330   0.5419594          0.015625
#> 12: 2022-07-26 04:59:26        1  -4.7232748   3.8889365          0.015625
#> 13: 2022-07-26 04:59:26        1  -3.7331722   7.9307152          0.015625
#> 14: 2022-07-26 04:59:26        1  -4.5904669   9.9233577          0.015625
#> 15: 2022-07-26 04:59:26        1  -3.3924993  12.2585014          0.015625
#> 16: 2022-07-26 04:59:26        1  -2.5078499   6.8156901          0.015625
#> 17: 2022-07-26 04:59:26        1   5.8570452   7.6335475          0.015625
#> 18: 2022-07-26 04:59:26        1   1.6489742   5.4766519          0.015625
#> 19: 2022-07-26 04:59:26        1  -4.8970783   6.3018270          0.015625
#> 20: 2022-07-26 04:59:26        1   0.4409220  14.0265603          0.015625
#> 21: 2022-07-26 04:59:26        1  -0.4947686   1.9107887          0.015625
#> 22: 2022-07-26 04:59:26        1   6.8485596   9.3046913          0.015625
#> 23: 2022-07-26 04:59:26        1   9.7866739   3.4979050          0.015625
#> 24: 2022-07-26 04:59:26        1   8.9087844  11.5701753          0.015625
#> 25: 2022-07-26 04:59:26        1   0.7124457  14.3602306          0.015625
#> 26: 2022-07-26 04:59:26        1  -0.7105858   0.4428729          0.015625
#> 27: 2022-07-26 04:59:26        1  -4.5794437   9.4081209          0.015625
#> 28: 2022-07-26 04:59:26        1   6.8087525   6.6489422          0.015625
#> 29: 2022-07-26 04:59:26        1   4.0287601   7.6083393          0.015625
#> 30: 2022-07-26 04:59:26        1   5.3809476  13.3547893          0.015625
#> 31: 2022-07-26 04:59:26        1   5.2824720   6.0793443          0.015625
#> 32: 2022-07-26 04:59:26        1  -0.8213706   9.4680380          0.015625
#> 33: 2022-07-26 04:59:26        1   5.3802357   6.9071066          0.015625
#> 34: 2022-07-26 04:59:26        1  -1.3721263   0.7958979          0.015625
#> 35: 2022-07-26 04:59:26        1   9.6488817   8.2244004          0.015625
#> 36: 2022-07-26 04:59:26        1   5.2113158  11.2023538          0.015625
#> 37: 2022-07-26 04:59:26        1  -3.8297096  14.9023807          0.015625
#> 38: 2022-07-26 04:59:26        1  -2.7812194  11.3250842          0.015625
#> 39: 2022-07-26 04:59:26        1   2.0323858   6.7833018          0.015625
#> 40: 2022-07-26 04:59:26        1   7.9535442   0.1548866          0.015625
#> 41: 2022-07-26 04:59:26        1   1.4488799  11.0640525          0.015625
#> 42: 2022-07-26 04:59:26        1  -3.2594281   2.5369624          0.015625
#> 43: 2022-07-26 04:59:26        1   6.0864404   5.9307203          0.015625
#> 44: 2022-07-26 04:59:26        1   4.5932053   0.9746272          0.015625
#> 45: 2022-07-26 04:59:26        1  -2.9594081  11.7097184          0.015625
#> 46: 2022-07-26 04:59:26        1   5.1350394   9.6131330          0.015625
#> 47: 2022-07-26 04:59:26        1  -3.5854158   9.0245135          0.015625
#> 48: 2022-07-26 04:59:26        1   0.2832746   6.2923561          0.015625
#> 49: 2022-07-26 04:59:26        1   4.3604824   7.6930383          0.015625
#> 50: 2022-07-26 04:59:26        1   0.6915248  14.9849852          0.015625
#> 51: 2022-07-26 04:59:26        1   2.6303060   0.8561279          0.015625
#> 52: 2022-07-26 04:59:26        1  -2.7607172   8.8992841          0.015625
#> 53: 2022-07-26 04:59:26        1  -2.9390608   5.7271616          0.015625
#> 54: 2022-07-26 04:59:26        1   3.8711457   3.8333866          0.015625
#> 55: 2022-07-26 04:59:26        1  -0.8504421   9.5217116          0.015625
#> 56: 2022-07-26 04:59:26        1  -2.2533757   2.6108227          0.015625
#> 57: 2022-07-26 04:59:26        1  -1.4291925   0.3862328          0.015625
#> 58: 2022-07-26 04:59:26        1   5.3325337   6.3877811          0.015625
#> 59: 2022-07-26 04:59:26        1  -0.2611272   4.9341849          0.015625
#> 60: 2022-07-26 04:59:26        1   9.0643664  14.9178418          0.015625
#> 61: 2022-07-26 04:59:26        1  -2.8281980   1.8030657          0.015625
#> 62: 2022-07-26 04:59:26        1  -0.3459762   2.5573745          0.015625
#> 63: 2022-07-26 04:59:26        1   6.6041261   4.6712230          0.015625
#> 64: 2022-07-26 04:59:26        1  -3.7425562   6.9512045          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity