Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 2.866654 1.718234 0.015625 <list[3]> 0.7593076

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 2.866654 1.718234 0.015625 <list[3]> 0.7593076

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -3.5558715  1.1845120 0.015625     0       6          1 119.2280798
#>  2:  7.5268222  4.7054783 0.015625     0       6          1  93.0607122
#>  3:  2.4740964 10.5996676 0.015625     0       6          1  72.1663424
#>  4:  0.1000685  1.0254021 0.015625     0       6          1  42.7444826
#>  5: -1.4612968 11.3379108 0.015625     0       6          1  19.7313547
#>  6: -1.7297846  4.9915045 0.015625     0       6          1  23.3295165
#>  7:  5.7029392  1.8212823 0.015625     0       6          1  33.2232894
#>  8:  5.5206671  0.4983457 0.015625     0       6          1  22.4547847
#>  9:  3.7635729 13.5785869 0.015625     0       6          1 174.6713736
#> 10:  0.1306523  9.2299915 0.015625     0       6          1  31.3360364
#> 11: -4.0924497 14.1953070 0.015625     0       6          1   5.7832349
#> 12:  8.6810862  1.8283507 0.015625     0       6          1  56.6267453
#> 13:  2.4011142  5.4020095 0.015625     0       6          1  12.1917832
#> 14:  7.6907837 11.8969335 0.015625     0       6          1 277.8557304
#> 15: -4.7054148  8.5723226 0.015625     0       6          1  41.2633054
#> 16:  7.7022273  5.5336405 0.015625     0       6          1 110.8140156
#> 17:  3.0516497 14.9330735 0.015625     0       6          1 182.7836787
#> 18: -3.9982233 14.8898281 0.015625     0       6          1   7.8517317
#> 19:  3.0519494 14.5694018 0.015625     0       6          1 173.1049860
#> 20:  7.5266864  2.4965660 0.015625     0       6          1  58.4311183
#> 21: -2.6275289 13.6953460 0.015625     0       6          1  12.5370416
#> 22:  7.9179070 12.6227623 0.015625     0       6          1 308.5684038
#> 23: -0.1336558  7.6271436 0.015625     0       6          1  21.5155159
#> 24: -1.0116038  7.8539239 0.015625     0       6          1  15.1390703
#> 25:  9.8681142 14.8588284 0.015625     0       6          1 466.6034076
#> 26:  0.6115922 13.7290353 0.015625     0       6          1  93.3935666
#> 27:  3.9524748  7.2059536 0.015625     0       6          1  52.6122914
#> 28: -2.8548608  8.7764822 0.015625     0       6          1   4.8612858
#> 29:  1.7161309  0.9052212 0.015625     0       6          1  14.6316264
#> 30:  1.4588381  1.2370619 0.015625     0       6          1  17.3556401
#> 31:  1.8811044  1.4864807 0.015625     0       6          1   9.7198209
#> 32:  8.1650757  6.9611381 0.015625     0       6          1 148.8232443
#> 33: -1.2904971  2.6872108 0.015625     0       6          1  42.0098338
#> 34:  6.3352383  7.5071852 0.015625     0       6          1 126.8356654
#> 35:  0.9790927 11.4097823 0.015625     0       6          1  63.4997462
#> 36:  2.3921614 11.9042823 0.015625     0       6          1  93.8977610
#> 37:  1.7015644  4.0213677 0.015625     0       6          1   9.1581262
#> 38:  9.0317805 11.5597935 0.015625     0       6          1 304.8020360
#> 39:  0.3799332 12.4454457 0.015625     0       6          1  68.5591263
#> 40:  4.2062048  6.1525265 0.015625     0       6          1  45.0713833
#> 41:  1.4951176  1.9941209 0.015625     0       6          1  13.5991994
#> 42:  8.8578823 12.5589701 0.015625     0       6          1 334.7527073
#> 43: -0.8994886 14.4363060 0.015625     0       6          1  64.6909016
#> 44:  5.6534948 13.2310236 0.015625     0       6          1 250.2042088
#> 45:  8.6153660  6.3400116 0.015625     0       6          1 141.8989920
#> 46:  3.4117557 13.4383810 0.015625     0       6          1 157.2578249
#> 47: -1.6340776  6.6423271 0.015625     0       6          1  13.5563674
#> 48:  0.7518151  8.7491466 0.015625     0       6          1  32.4455536
#> 49: -1.1730229 11.2590732 0.015625     0       6          1  24.9409784
#> 50: -1.5126792  6.1905864 0.015625     0       6          1  15.7893490
#> 51:  1.5623426  0.9358086 0.015625     0       6          1  17.1179722
#> 52:  6.9324674 11.0821089 0.015625     0       6          1 231.9134378
#> 53:  5.4111288  8.3242805 0.015625     0       6          1 116.8994392
#> 54:  2.8666542  1.7182341 0.015625     0       6          1   0.7593076
#> 55:  6.8548648 12.9624670 0.015625     0       6          1 287.9396052
#> 56:  7.3576053 14.3792622 0.015625     0       6          1 354.0456403
#> 57: -1.8086106 11.0999233 0.015625     0       6          1  12.2359682
#> 58:  6.4262049 11.4109355 0.015625     0       6          1 225.9672476
#> 59:  7.7457030  8.2450786 0.015625     0       6          1 173.0410028
#> 60: -0.5836380  4.3830127 0.015625     0       6          1  24.5475341
#> 61:  4.0879862  2.4818912 0.015625     0       6          1  10.5085956
#> 62: -0.9890023  4.0995686 0.015625     0       6          1  27.5584721
#> 63: -2.0849357  2.3704643 0.015625     0       6          1  55.4250341
#> 64:  5.2615778 12.1553959 0.015625     0       6          1 202.1068010
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-06-25 04:35:50        1  -3.5558715   1.1845120          0.015625
#>  2: 2022-06-25 04:35:50        1   7.5268222   4.7054783          0.015625
#>  3: 2022-06-25 04:35:50        1   2.4740964  10.5996676          0.015625
#>  4: 2022-06-25 04:35:50        1   0.1000685   1.0254021          0.015625
#>  5: 2022-06-25 04:35:50        1  -1.4612968  11.3379108          0.015625
#>  6: 2022-06-25 04:35:50        1  -1.7297846   4.9915045          0.015625
#>  7: 2022-06-25 04:35:50        1   5.7029392   1.8212823          0.015625
#>  8: 2022-06-25 04:35:50        1   5.5206671   0.4983457          0.015625
#>  9: 2022-06-25 04:35:50        1   3.7635729  13.5785869          0.015625
#> 10: 2022-06-25 04:35:50        1   0.1306523   9.2299915          0.015625
#> 11: 2022-06-25 04:35:50        1  -4.0924497  14.1953070          0.015625
#> 12: 2022-06-25 04:35:50        1   8.6810862   1.8283507          0.015625
#> 13: 2022-06-25 04:35:50        1   2.4011142   5.4020095          0.015625
#> 14: 2022-06-25 04:35:50        1   7.6907837  11.8969335          0.015625
#> 15: 2022-06-25 04:35:50        1  -4.7054148   8.5723226          0.015625
#> 16: 2022-06-25 04:35:50        1   7.7022273   5.5336405          0.015625
#> 17: 2022-06-25 04:35:50        1   3.0516497  14.9330735          0.015625
#> 18: 2022-06-25 04:35:50        1  -3.9982233  14.8898281          0.015625
#> 19: 2022-06-25 04:35:50        1   3.0519494  14.5694018          0.015625
#> 20: 2022-06-25 04:35:50        1   7.5266864   2.4965660          0.015625
#> 21: 2022-06-25 04:35:50        1  -2.6275289  13.6953460          0.015625
#> 22: 2022-06-25 04:35:50        1   7.9179070  12.6227623          0.015625
#> 23: 2022-06-25 04:35:50        1  -0.1336558   7.6271436          0.015625
#> 24: 2022-06-25 04:35:50        1  -1.0116038   7.8539239          0.015625
#> 25: 2022-06-25 04:35:50        1   9.8681142  14.8588284          0.015625
#> 26: 2022-06-25 04:35:50        1   0.6115922  13.7290353          0.015625
#> 27: 2022-06-25 04:35:50        1   3.9524748   7.2059536          0.015625
#> 28: 2022-06-25 04:35:50        1  -2.8548608   8.7764822          0.015625
#> 29: 2022-06-25 04:35:50        1   1.7161309   0.9052212          0.015625
#> 30: 2022-06-25 04:35:50        1   1.4588381   1.2370619          0.015625
#> 31: 2022-06-25 04:35:50        1   1.8811044   1.4864807          0.015625
#> 32: 2022-06-25 04:35:50        1   8.1650757   6.9611381          0.015625
#> 33: 2022-06-25 04:35:50        1  -1.2904971   2.6872108          0.015625
#> 34: 2022-06-25 04:35:50        1   6.3352383   7.5071852          0.015625
#> 35: 2022-06-25 04:35:50        1   0.9790927  11.4097823          0.015625
#> 36: 2022-06-25 04:35:50        1   2.3921614  11.9042823          0.015625
#> 37: 2022-06-25 04:35:50        1   1.7015644   4.0213677          0.015625
#> 38: 2022-06-25 04:35:50        1   9.0317805  11.5597935          0.015625
#> 39: 2022-06-25 04:35:50        1   0.3799332  12.4454457          0.015625
#> 40: 2022-06-25 04:35:50        1   4.2062048   6.1525265          0.015625
#> 41: 2022-06-25 04:35:50        1   1.4951176   1.9941209          0.015625
#> 42: 2022-06-25 04:35:50        1   8.8578823  12.5589701          0.015625
#> 43: 2022-06-25 04:35:50        1  -0.8994886  14.4363060          0.015625
#> 44: 2022-06-25 04:35:50        1   5.6534948  13.2310236          0.015625
#> 45: 2022-06-25 04:35:50        1   8.6153660   6.3400116          0.015625
#> 46: 2022-06-25 04:35:50        1   3.4117557  13.4383810          0.015625
#> 47: 2022-06-25 04:35:50        1  -1.6340776   6.6423271          0.015625
#> 48: 2022-06-25 04:35:50        1   0.7518151   8.7491466          0.015625
#> 49: 2022-06-25 04:35:50        1  -1.1730229  11.2590732          0.015625
#> 50: 2022-06-25 04:35:50        1  -1.5126792   6.1905864          0.015625
#> 51: 2022-06-25 04:35:50        1   1.5623426   0.9358086          0.015625
#> 52: 2022-06-25 04:35:50        1   6.9324674  11.0821089          0.015625
#> 53: 2022-06-25 04:35:50        1   5.4111288   8.3242805          0.015625
#> 54: 2022-06-25 04:35:50        1   2.8666542   1.7182341          0.015625
#> 55: 2022-06-25 04:35:50        1   6.8548648  12.9624670          0.015625
#> 56: 2022-06-25 04:35:50        1   7.3576053  14.3792622          0.015625
#> 57: 2022-06-25 04:35:50        1  -1.8086106  11.0999233          0.015625
#> 58: 2022-06-25 04:35:50        1   6.4262049  11.4109355          0.015625
#> 59: 2022-06-25 04:35:50        1   7.7457030   8.2450786          0.015625
#> 60: 2022-06-25 04:35:50        1  -0.5836380   4.3830127          0.015625
#> 61: 2022-06-25 04:35:50        1   4.0879862   2.4818912          0.015625
#> 62: 2022-06-25 04:35:50        1  -0.9890023   4.0995686          0.015625
#> 63: 2022-06-25 04:35:50        1  -2.0849357   2.3704643          0.015625
#> 64: 2022-06-25 04:35:50        1   5.2615778  12.1553959          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity