Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.133444 9.618601 0.015625 <list[3]> 3.188169

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.133444 9.618601 0.015625 <list[3]> 3.188169

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1: -1.98439530  5.3462186 0.015625     0       6          1  21.610280
#>  2:  9.13119276 12.7733059 0.015625     0       6          1 352.086499
#>  3: -3.02587000  4.3622322 0.015625     0       6          1  45.823759
#>  4:  2.01537688  1.5041645 0.015625     0       6          1   7.867324
#>  5:  6.39354043  1.2236220 0.015625     0       6          1  36.703190
#>  6:  8.00305106  1.0602949 0.015625     0       6          1  42.542253
#>  7:  6.57264589 11.3225377 0.015625     0       6          1 228.148738
#>  8: -4.30828322  7.1831245 0.015625     0       6          1  45.217334
#>  9:  2.34351758  5.9365552 0.015625     0       6          1  15.529504
#> 10: -1.16322317  9.8887142 0.015625     0       6          1  17.789261
#> 11:  5.32462180 14.9845238 0.015625     0       6          1 290.653362
#> 12:  3.56855998  4.6835874 0.015625     0       6          1  17.033322
#> 13: -4.72189010 10.0520830 0.015625     0       6          1  27.301957
#> 14:  3.21732337  5.1264910 0.015625     0       6          1  15.860507
#> 15:  2.60524107 10.0632160 0.015625     0       6          1  65.760586
#> 16:  9.71235600  7.2476146 0.015625     0       6          1 191.368806
#> 17:  5.69610263  2.1350373 0.015625     0       6          1  35.660063
#> 18:  9.39699739  9.3874310 0.015625     0       6          1 244.640554
#> 19:  5.60862071  5.9938023 0.015625     0       6          1  80.749292
#> 20: -1.86742362 11.3764978 0.015625     0       6          1  12.470362
#> 21: -4.56461836  5.0470261 0.015625     0       6          1  87.057966
#> 22:  6.67208983  8.3499754 0.015625     0       6          1 153.449210
#> 23:  3.63570842  2.5362680 0.015625     0       6          1   5.218434
#> 24: -0.06733061  5.6880553 0.015625     0       6          1  19.756121
#> 25:  8.56687639 10.0761481 0.015625     0       6          1 242.551959
#> 26:  0.17933913 10.2999453 0.015625     0       6          1  40.464683
#> 27:  8.29806869 10.5408396 0.015625     0       6          1 250.185586
#> 28:  0.64356640  1.6572309 0.015625     0       6          1  28.778448
#> 29:  6.91822702  3.1839532 0.015625     0       6          1  62.929476
#> 30:  6.90255271  0.4229027 0.015625     0       6          1  33.371314
#> 31:  4.50014125  6.7992416 0.015625     0       6          1  61.834827
#> 32:  3.76686883  9.6649648 0.015625     0       6          1  87.293894
#> 33:  1.30157767  8.4440218 0.015625     0       6          1  32.476414
#> 34:  9.96732642  5.3201133 0.015625     0       6          1 148.888573
#> 35:  6.11311531 11.7679896 0.015625     0       6          1 225.337110
#> 36:  8.92309641  5.1825661 0.015625     0       6          1 121.176929
#> 37:  7.93126986  7.3969597 0.015625     0       6          1 155.325795
#> 38:  9.08336308  4.3385352 0.015625     0       6          1 106.184785
#> 39:  0.34574116 12.5052158 0.015625     0       6          1  68.761851
#> 40:  7.58967815  3.2194513 0.015625     0       6          1  69.174283
#> 41: -3.13344391  9.6186013 0.015625     0       6          1   3.188169
#> 42: -2.87715747  3.1776365 0.015625     0       6          1  59.346398
#> 43:  8.66522297  1.5876626 0.015625     0       6          1  53.023705
#> 44:  4.83430384  9.1220077 0.015625     0       6          1 113.126612
#> 45: -3.34931009  5.3946120 0.015625     0       6          1  40.053957
#> 46:  5.09607348  2.1875613 0.015625     0       6          1  25.842459
#> 47: -0.89354799  6.0093698 0.015625     0       6          1  18.083020
#> 48:  6.65501390  2.7473588 0.015625     0       6          1  54.675039
#> 49: -3.94000544  2.7285394 0.015625     0       6          1 103.689190
#> 50:  5.39850970 12.6022702 0.015625     0       6          1 220.521199
#> 51:  7.35559866  3.8150153 0.015625     0       6          1  76.341998
#> 52:  2.44931194 11.9909026 0.015625     0       6          1  96.788567
#> 53:  3.37284033 13.7617698 0.015625     0       6          1 163.982710
#> 54: -3.78968708 13.9704436 0.015625     0       6          1   4.587028
#> 55:  6.46981610 13.4775311 0.015625     0       6          1 291.274311
#> 56:  8.82019851 12.0777297 0.015625     0       6          1 316.223118
#> 57:  6.75211420 11.1647751 0.015625     0       6          1 229.085677
#> 58:  6.47862841 12.2136140 0.015625     0       6          1 251.498285
#> 59: -1.06385452 11.1569882 0.015625     0       6          1  26.420007
#> 60:  0.67155066  6.7233214 0.015625     0       6          1  20.679302
#> 61:  2.66277247 11.3969231 0.015625     0       6          1  90.154811
#> 62: -0.14313852 12.8276967 0.015625     0       6          1  63.054076
#> 63:  4.64330346 13.7903640 0.015625     0       6          1 220.094887
#> 64: -0.85698852  1.4513051 0.015625     0       6          1  51.513453
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-24 04:57:54        1 -1.98439530   5.3462186          0.015625
#>  2: 2022-08-24 04:57:54        1  9.13119276  12.7733059          0.015625
#>  3: 2022-08-24 04:57:54        1 -3.02587000   4.3622322          0.015625
#>  4: 2022-08-24 04:57:54        1  2.01537688   1.5041645          0.015625
#>  5: 2022-08-24 04:57:54        1  6.39354043   1.2236220          0.015625
#>  6: 2022-08-24 04:57:54        1  8.00305106   1.0602949          0.015625
#>  7: 2022-08-24 04:57:54        1  6.57264589  11.3225377          0.015625
#>  8: 2022-08-24 04:57:54        1 -4.30828322   7.1831245          0.015625
#>  9: 2022-08-24 04:57:54        1  2.34351758   5.9365552          0.015625
#> 10: 2022-08-24 04:57:54        1 -1.16322317   9.8887142          0.015625
#> 11: 2022-08-24 04:57:54        1  5.32462180  14.9845238          0.015625
#> 12: 2022-08-24 04:57:54        1  3.56855998   4.6835874          0.015625
#> 13: 2022-08-24 04:57:54        1 -4.72189010  10.0520830          0.015625
#> 14: 2022-08-24 04:57:54        1  3.21732337   5.1264910          0.015625
#> 15: 2022-08-24 04:57:54        1  2.60524107  10.0632160          0.015625
#> 16: 2022-08-24 04:57:54        1  9.71235600   7.2476146          0.015625
#> 17: 2022-08-24 04:57:54        1  5.69610263   2.1350373          0.015625
#> 18: 2022-08-24 04:57:54        1  9.39699739   9.3874310          0.015625
#> 19: 2022-08-24 04:57:54        1  5.60862071   5.9938023          0.015625
#> 20: 2022-08-24 04:57:54        1 -1.86742362  11.3764978          0.015625
#> 21: 2022-08-24 04:57:54        1 -4.56461836   5.0470261          0.015625
#> 22: 2022-08-24 04:57:54        1  6.67208983   8.3499754          0.015625
#> 23: 2022-08-24 04:57:54        1  3.63570842   2.5362680          0.015625
#> 24: 2022-08-24 04:57:54        1 -0.06733061   5.6880553          0.015625
#> 25: 2022-08-24 04:57:54        1  8.56687639  10.0761481          0.015625
#> 26: 2022-08-24 04:57:54        1  0.17933913  10.2999453          0.015625
#> 27: 2022-08-24 04:57:54        1  8.29806869  10.5408396          0.015625
#> 28: 2022-08-24 04:57:54        1  0.64356640   1.6572309          0.015625
#> 29: 2022-08-24 04:57:54        1  6.91822702   3.1839532          0.015625
#> 30: 2022-08-24 04:57:54        1  6.90255271   0.4229027          0.015625
#> 31: 2022-08-24 04:57:54        1  4.50014125   6.7992416          0.015625
#> 32: 2022-08-24 04:57:54        1  3.76686883   9.6649648          0.015625
#> 33: 2022-08-24 04:57:54        1  1.30157767   8.4440218          0.015625
#> 34: 2022-08-24 04:57:54        1  9.96732642   5.3201133          0.015625
#> 35: 2022-08-24 04:57:54        1  6.11311531  11.7679896          0.015625
#> 36: 2022-08-24 04:57:54        1  8.92309641   5.1825661          0.015625
#> 37: 2022-08-24 04:57:54        1  7.93126986   7.3969597          0.015625
#> 38: 2022-08-24 04:57:54        1  9.08336308   4.3385352          0.015625
#> 39: 2022-08-24 04:57:54        1  0.34574116  12.5052158          0.015625
#> 40: 2022-08-24 04:57:54        1  7.58967815   3.2194513          0.015625
#> 41: 2022-08-24 04:57:54        1 -3.13344391   9.6186013          0.015625
#> 42: 2022-08-24 04:57:54        1 -2.87715747   3.1776365          0.015625
#> 43: 2022-08-24 04:57:54        1  8.66522297   1.5876626          0.015625
#> 44: 2022-08-24 04:57:54        1  4.83430384   9.1220077          0.015625
#> 45: 2022-08-24 04:57:54        1 -3.34931009   5.3946120          0.015625
#> 46: 2022-08-24 04:57:54        1  5.09607348   2.1875613          0.015625
#> 47: 2022-08-24 04:57:54        1 -0.89354799   6.0093698          0.015625
#> 48: 2022-08-24 04:57:54        1  6.65501390   2.7473588          0.015625
#> 49: 2022-08-24 04:57:54        1 -3.94000544   2.7285394          0.015625
#> 50: 2022-08-24 04:57:54        1  5.39850970  12.6022702          0.015625
#> 51: 2022-08-24 04:57:54        1  7.35559866   3.8150153          0.015625
#> 52: 2022-08-24 04:57:54        1  2.44931194  11.9909026          0.015625
#> 53: 2022-08-24 04:57:54        1  3.37284033  13.7617698          0.015625
#> 54: 2022-08-24 04:57:54        1 -3.78968708  13.9704436          0.015625
#> 55: 2022-08-24 04:57:54        1  6.46981610  13.4775311          0.015625
#> 56: 2022-08-24 04:57:54        1  8.82019851  12.0777297          0.015625
#> 57: 2022-08-24 04:57:54        1  6.75211420  11.1647751          0.015625
#> 58: 2022-08-24 04:57:54        1  6.47862841  12.2136140          0.015625
#> 59: 2022-08-24 04:57:54        1 -1.06385452  11.1569882          0.015625
#> 60: 2022-08-24 04:57:54        1  0.67155066   6.7233214          0.015625
#> 61: 2022-08-24 04:57:54        1  2.66277247  11.3969231          0.015625
#> 62: 2022-08-24 04:57:54        1 -0.14313852  12.8276967          0.015625
#> 63: 2022-08-24 04:57:54        1  4.64330346  13.7903640          0.015625
#> 64: 2022-08-24 04:57:54        1 -0.85698852   1.4513051          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity