Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.753684 1.731819 0.015625 <list[3]> 1.125395

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.753684 1.731819 0.015625 <list[3]> 1.125395

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition          y
#>  1: -3.09686068  0.69142645 0.015625     0       6          1 111.336014
#>  2:  7.23741840  9.14967230 0.015625     0       6          1 186.059621
#>  3: -2.16333264  9.99168295 0.015625     0       6          1   4.801350
#>  4: -3.23329002 13.72618110 0.015625     0       6          1   5.540433
#>  5:  2.22182403 13.55719925 0.015625     0       6          1 123.898800
#>  6:  2.70467268 13.21874141 0.015625     0       6          1 128.954652
#>  7:  3.92144637  2.56070259 0.015625     0       6          1   8.597255
#>  8:  4.36910044 12.46387699 0.015625     0       6          1 171.391991
#>  9: -1.50191063 10.97093150 0.015625     0       6          1  16.967098
#> 10:  0.23625128  0.43240276 0.015625     0       6          1  46.305833
#> 11:  7.89963860 12.49788616 0.015625     0       6          1 303.746205
#> 12: -2.73557637  0.66592660 0.015625     0       6          1  99.544652
#> 13:  7.36165805  6.92418237 0.015625     0       6          1 134.974406
#> 14:  8.71772861 14.89472520 0.015625     0       6          1 420.194330
#> 15: -4.70063669  3.58652726 0.015625     0       6          1 121.699885
#> 16:  1.75899636  1.41724689 0.015625     0       6          1  11.731755
#> 17:  8.50894580  2.85452942 0.015625     0       6          1  70.912824
#> 18: -3.52414684  1.08929166 0.015625     0       6          1 119.933119
#> 19:  7.38030140  4.09153752 0.015625     0       6          1  81.013043
#> 20: -0.83678446  0.16457442 0.015625     0       6          1  68.109976
#> 21: -0.58839548  2.77631179 0.015625     0       6          1  35.382867
#> 22:  0.44083451 14.75367246 0.015625     0       6          1 107.973517
#> 23:  5.80490274 11.78017575 0.015625     0       6          1 214.045747
#> 24:  8.20907224  8.71126878 0.015625     0       6          1 194.472352
#> 25:  6.11920285  5.88138768 0.015625     0       6          1  91.198797
#> 26:  2.73137692  0.01138914 0.015625     0       6          1   4.694702
#> 27: -2.37597125  2.91894382 0.015625     0       6          1  52.584128
#> 28:  5.47508591  4.84472433 0.015625     0       6          1  60.682348
#> 29:  4.12435681  0.47465440 0.015625     0       6          1   4.939474
#> 30:  1.29640680 11.03154472 0.015625     0       6          1  62.207950
#> 31: -4.05949431 12.22703879 0.015625     0       6          1   4.715232
#> 32: -4.63740971 14.06947708 0.015625     0       6          1   9.281476
#> 33:  8.55394162  4.37280165 0.015625     0       6          1  98.624906
#> 34: -0.30933400  0.41987630 0.015625     0       6          1  56.056680
#> 35:  4.33815597  8.54758870 0.015625     0       6          1  85.225885
#> 36: -1.34760972  2.77309675 0.015625     0       6          1  41.583385
#> 37:  9.13412328  9.31287531 0.015625     0       6          1 234.431649
#> 38: -3.44969504 14.44412085 0.015625     0       6          1   7.547058
#> 39:  0.83878099  3.58040501 0.015625     0       6          1  17.641589
#> 40: -2.42609930  7.35855612 0.015625     0       6          1   9.954736
#> 41: -0.46207963  8.71727085 0.015625     0       6          1  22.496857
#> 42: -3.89435546 10.30438669 0.015625     0       6          1   8.561865
#> 43:  6.89888749 13.27144885 0.015625     0       6          1 299.675004
#> 44:  6.43486988 11.41658442 0.015625     0       6          1 226.415471
#> 45:  5.24454897 14.95416634 0.015625     0       6          1 285.652911
#> 46: -2.36127425  4.76226517 0.015625     0       6          1  29.876383
#> 47:  7.17969487  9.54714526 0.015625     0       6          1 195.261302
#> 48:  5.30258092  3.14698165 0.015625     0       6          1  37.638945
#> 49:  0.06280547 10.53789610 0.015625     0       6          1  41.091747
#> 50:  1.65798414  2.26344434 0.015625     0       6          1  10.561736
#> 51:  0.91659333  8.69982798 0.015625     0       6          1  32.923178
#> 52: -3.19943731  1.01524503 0.015625     0       6          1 108.398389
#> 53:  4.36906315  9.05438485 0.015625     0       6          1  95.523800
#> 54:  2.75368387  1.73181914 0.015625     0       6          1   1.125395
#> 55: -1.38615308  0.90009592 0.015625     0       6          1  66.007800
#> 56: -2.06994873  1.28018845 0.015625     0       6          1  71.760613
#> 57:  5.65615950 10.69879242 0.015625     0       6          1 179.502413
#> 58: -2.96829642  0.65975570 0.015625     0       6          1 107.360964
#> 59:  0.43184075  3.00129454 0.015625     0       6          1  24.089762
#> 60:  8.63798208  1.63310083 0.015625     0       6          1  53.441043
#> 61:  0.13875543  8.77830917 0.015625     0       6          1  28.501148
#> 62:  4.29828995  3.15948691 0.015625     0       6          1  17.917591
#> 63: -1.71275152  5.14374501 0.015625     0       6          1  22.127883
#> 64:  7.92630227  1.88823000 0.015625     0       6          1  52.492259
#>              x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-04-08 04:28:43        1 -3.09686068  0.69142645          0.015625
#>  2: 2022-04-08 04:28:43        1  7.23741840  9.14967230          0.015625
#>  3: 2022-04-08 04:28:43        1 -2.16333264  9.99168295          0.015625
#>  4: 2022-04-08 04:28:43        1 -3.23329002 13.72618110          0.015625
#>  5: 2022-04-08 04:28:43        1  2.22182403 13.55719925          0.015625
#>  6: 2022-04-08 04:28:43        1  2.70467268 13.21874141          0.015625
#>  7: 2022-04-08 04:28:43        1  3.92144637  2.56070259          0.015625
#>  8: 2022-04-08 04:28:43        1  4.36910044 12.46387699          0.015625
#>  9: 2022-04-08 04:28:43        1 -1.50191063 10.97093150          0.015625
#> 10: 2022-04-08 04:28:43        1  0.23625128  0.43240276          0.015625
#> 11: 2022-04-08 04:28:43        1  7.89963860 12.49788616          0.015625
#> 12: 2022-04-08 04:28:43        1 -2.73557637  0.66592660          0.015625
#> 13: 2022-04-08 04:28:43        1  7.36165805  6.92418237          0.015625
#> 14: 2022-04-08 04:28:43        1  8.71772861 14.89472520          0.015625
#> 15: 2022-04-08 04:28:43        1 -4.70063669  3.58652726          0.015625
#> 16: 2022-04-08 04:28:43        1  1.75899636  1.41724689          0.015625
#> 17: 2022-04-08 04:28:43        1  8.50894580  2.85452942          0.015625
#> 18: 2022-04-08 04:28:43        1 -3.52414684  1.08929166          0.015625
#> 19: 2022-04-08 04:28:43        1  7.38030140  4.09153752          0.015625
#> 20: 2022-04-08 04:28:43        1 -0.83678446  0.16457442          0.015625
#> 21: 2022-04-08 04:28:43        1 -0.58839548  2.77631179          0.015625
#> 22: 2022-04-08 04:28:43        1  0.44083451 14.75367246          0.015625
#> 23: 2022-04-08 04:28:43        1  5.80490274 11.78017575          0.015625
#> 24: 2022-04-08 04:28:43        1  8.20907224  8.71126878          0.015625
#> 25: 2022-04-08 04:28:43        1  6.11920285  5.88138768          0.015625
#> 26: 2022-04-08 04:28:43        1  2.73137692  0.01138914          0.015625
#> 27: 2022-04-08 04:28:43        1 -2.37597125  2.91894382          0.015625
#> 28: 2022-04-08 04:28:43        1  5.47508591  4.84472433          0.015625
#> 29: 2022-04-08 04:28:43        1  4.12435681  0.47465440          0.015625
#> 30: 2022-04-08 04:28:43        1  1.29640680 11.03154472          0.015625
#> 31: 2022-04-08 04:28:43        1 -4.05949431 12.22703879          0.015625
#> 32: 2022-04-08 04:28:43        1 -4.63740971 14.06947708          0.015625
#> 33: 2022-04-08 04:28:43        1  8.55394162  4.37280165          0.015625
#> 34: 2022-04-08 04:28:43        1 -0.30933400  0.41987630          0.015625
#> 35: 2022-04-08 04:28:43        1  4.33815597  8.54758870          0.015625
#> 36: 2022-04-08 04:28:43        1 -1.34760972  2.77309675          0.015625
#> 37: 2022-04-08 04:28:43        1  9.13412328  9.31287531          0.015625
#> 38: 2022-04-08 04:28:43        1 -3.44969504 14.44412085          0.015625
#> 39: 2022-04-08 04:28:43        1  0.83878099  3.58040501          0.015625
#> 40: 2022-04-08 04:28:43        1 -2.42609930  7.35855612          0.015625
#> 41: 2022-04-08 04:28:43        1 -0.46207963  8.71727085          0.015625
#> 42: 2022-04-08 04:28:43        1 -3.89435546 10.30438669          0.015625
#> 43: 2022-04-08 04:28:43        1  6.89888749 13.27144885          0.015625
#> 44: 2022-04-08 04:28:43        1  6.43486988 11.41658442          0.015625
#> 45: 2022-04-08 04:28:43        1  5.24454897 14.95416634          0.015625
#> 46: 2022-04-08 04:28:43        1 -2.36127425  4.76226517          0.015625
#> 47: 2022-04-08 04:28:43        1  7.17969487  9.54714526          0.015625
#> 48: 2022-04-08 04:28:43        1  5.30258092  3.14698165          0.015625
#> 49: 2022-04-08 04:28:43        1  0.06280547 10.53789610          0.015625
#> 50: 2022-04-08 04:28:43        1  1.65798414  2.26344434          0.015625
#> 51: 2022-04-08 04:28:43        1  0.91659333  8.69982798          0.015625
#> 52: 2022-04-08 04:28:43        1 -3.19943731  1.01524503          0.015625
#> 53: 2022-04-08 04:28:43        1  4.36906315  9.05438485          0.015625
#> 54: 2022-04-08 04:28:43        1  2.75368387  1.73181914          0.015625
#> 55: 2022-04-08 04:28:43        1 -1.38615308  0.90009592          0.015625
#> 56: 2022-04-08 04:28:43        1 -2.06994873  1.28018845          0.015625
#> 57: 2022-04-08 04:28:43        1  5.65615950 10.69879242          0.015625
#> 58: 2022-04-08 04:28:43        1 -2.96829642  0.65975570          0.015625
#> 59: 2022-04-08 04:28:43        1  0.43184075  3.00129454          0.015625
#> 60: 2022-04-08 04:28:43        1  8.63798208  1.63310083          0.015625
#> 61: 2022-04-08 04:28:43        1  0.13875543  8.77830917          0.015625
#> 62: 2022-04-08 04:28:43        1  4.29828995  3.15948691          0.015625
#> 63: 2022-04-08 04:28:43        1 -1.71275152  5.14374501          0.015625
#> 64: 2022-04-08 04:28:43        1  7.92630227  1.88823000          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity