Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1      x2 fidelity  x_domain        y
#> 1: -3.71645 13.0463 0.015625 <list[3]> 2.440674

# best scoring evaluation
instance$result
#>          x1      x2 fidelity  x_domain        y
#> 1: -3.71645 13.0463 0.015625 <list[3]> 2.440674

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  2.51273601  9.8374536 0.015625     0       6          1  60.642043
#>  2:  3.57103295 14.9620449 0.015625     0       6          1 204.429921
#>  3:  6.57724016  5.6235208 0.015625     0       6          1  95.952179
#>  4:  6.66071950  0.9313154 0.015625     0       6          1  36.298814
#>  5: -3.53281306  0.1645257 0.015625     0       6          1 141.353200
#>  6:  3.27240837  6.0006897 0.015625     0       6          1  24.290721
#>  7: -0.73578603 11.3787455 0.015625     0       6          1  34.683062
#>  8:  6.76784168  6.0081647 0.015625     0       6          1 106.315546
#>  9:  7.88584071  6.9245156 0.015625     0       6          1 143.401096
#> 10:  7.36291794  3.5444699 0.015625     0       6          1  72.227892
#> 11:  6.24971943  3.4647999 0.015625     0       6          1  58.168165
#> 12:  8.84649848 14.5945940 0.015625     0       6          1 412.749198
#> 13:  3.15217228 13.8380827 0.015625     0       6          1 157.886181
#> 14:  7.28494613  6.3495584 0.015625     0       6          1 121.514306
#> 15: -1.81619747 13.7247335 0.015625     0       6          1  30.066078
#> 16:  1.38036620  0.8894970 0.015625     0       6          1  20.651280
#> 17: -0.30954438  5.4879660 0.015625     0       6          1  20.161077
#> 18:  0.05057673  4.2058794 0.015625     0       6          1  22.526615
#> 19:  0.05191865 12.5193298 0.015625     0       6          1  63.173968
#> 20:  3.88290262  9.3205021 0.015625     0       6          1  84.580715
#> 21:  1.21883328  3.6000969 0.015625     0       6          1  13.566023
#> 22: -2.80593286  8.6066642 0.015625     0       6          1   5.348775
#> 23:  6.28681264 14.5783433 0.015625     0       6          1 321.279735
#> 24: -2.24521770  8.7069649 0.015625     0       6          1   5.047274
#> 25:  8.33820039  5.8593706 0.015625     0       6          1 126.361258
#> 26: -2.39762853  2.0283406 0.015625     0       6          1  66.365840
#> 27: -4.17973267  7.3506546 0.015625     0       6          1  39.214846
#> 28:  3.11528942  3.7337356 0.015625     0       6          1   6.129917
#> 29:  0.71903737 12.9623982 0.015625     0       6          1  82.687436
#> 30:  2.15901578  8.7091371 0.015625     0       6          1  40.695916
#> 31:  8.92974156  2.7511225 0.015625     0       6          1  73.995769
#> 32:  9.37402550 10.7109571 0.015625     0       6          1 286.979960
#> 33:  4.49699790 12.8891111 0.015625     0       6          1 188.165191
#> 34:  7.21953828  3.7666281 0.015625     0       6          1  74.279651
#> 35:  1.44271563 12.8233851 0.015625     0       6          1  93.229302
#> 36:  1.90846619  5.4059819 0.015625     0       6          1  12.254404
#> 37:  2.43335460  1.2885500 0.015625     0       6          1   3.748947
#> 38: -4.76473608  9.2376568 0.015625     0       6          1  35.941289
#> 39: -2.11637969 12.9477499 0.015625     0       6          1  16.862664
#> 40: -2.25275083 14.4235712 0.015625     0       6          1  25.870387
#> 41: -1.73148635 14.3902307 0.015625     0       6          1  39.180798
#> 42:  0.32623926  7.7714899 0.015625     0       6          1  24.328039
#> 43: -0.34587427 14.1660619 0.015625     0       6          1  76.974609
#> 44: -3.69716158  5.1676138 0.015625     0       6          1  52.777133
#> 45:  5.41152526  6.3334756 0.015625     0       6          1  80.913945
#> 46:  2.50357962  4.4376766 0.015625     0       6          1   7.257602
#> 47:  8.30313520  7.8082786 0.015625     0       6          1 172.327605
#> 48:  3.51838641  9.5237348 0.015625     0       6          1  77.508074
#> 49:  7.07072590 10.4383155 0.015625     0       6          1 217.125691
#> 50:  0.28186193  8.4136476 0.015625     0       6          1  27.401673
#> 51:  4.92211137 10.9773611 0.015625     0       6          1 157.592987
#> 52:  7.94621971  6.6716058 0.015625     0       6          1 138.550956
#> 53:  0.42891383 10.8754544 0.015625     0       6          1  49.561874
#> 54:  5.07696668  0.9664152 0.015625     0       6          1  18.504810
#> 55:  9.45618940  9.5372230 0.015625     0       6          1 251.239582
#> 56:  1.73153273 11.2944002 0.015625     0       6          1  71.793538
#> 57: -3.71645033 13.0462998 0.015625     0       6          1   2.440674
#> 58: -2.45301022 11.0256545 0.015625     0       6          1   3.462887
#> 59:  1.84339422  4.1175539 0.015625     0       6          1   8.311450
#> 60: -2.74796298  3.9417503 0.015625     0       6          1  45.540439
#> 61:  2.68394877  6.7163003 0.015625     0       6          1  24.105018
#> 62:  2.23028956 10.5144685 0.015625     0       6          1  66.702812
#> 63:  5.33662404  5.9018564 0.015625     0       6          1  72.158043
#> 64:  7.50575878  6.0532979 0.015625     0       6          1 118.685903
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-03-18 04:28:11        1  2.51273601   9.8374536          0.015625
#>  2: 2022-03-18 04:28:11        1  3.57103295  14.9620449          0.015625
#>  3: 2022-03-18 04:28:11        1  6.57724016   5.6235208          0.015625
#>  4: 2022-03-18 04:28:11        1  6.66071950   0.9313154          0.015625
#>  5: 2022-03-18 04:28:11        1 -3.53281306   0.1645257          0.015625
#>  6: 2022-03-18 04:28:11        1  3.27240837   6.0006897          0.015625
#>  7: 2022-03-18 04:28:11        1 -0.73578603  11.3787455          0.015625
#>  8: 2022-03-18 04:28:11        1  6.76784168   6.0081647          0.015625
#>  9: 2022-03-18 04:28:11        1  7.88584071   6.9245156          0.015625
#> 10: 2022-03-18 04:28:11        1  7.36291794   3.5444699          0.015625
#> 11: 2022-03-18 04:28:11        1  6.24971943   3.4647999          0.015625
#> 12: 2022-03-18 04:28:11        1  8.84649848  14.5945940          0.015625
#> 13: 2022-03-18 04:28:11        1  3.15217228  13.8380827          0.015625
#> 14: 2022-03-18 04:28:11        1  7.28494613   6.3495584          0.015625
#> 15: 2022-03-18 04:28:11        1 -1.81619747  13.7247335          0.015625
#> 16: 2022-03-18 04:28:11        1  1.38036620   0.8894970          0.015625
#> 17: 2022-03-18 04:28:11        1 -0.30954438   5.4879660          0.015625
#> 18: 2022-03-18 04:28:11        1  0.05057673   4.2058794          0.015625
#> 19: 2022-03-18 04:28:11        1  0.05191865  12.5193298          0.015625
#> 20: 2022-03-18 04:28:11        1  3.88290262   9.3205021          0.015625
#> 21: 2022-03-18 04:28:11        1  1.21883328   3.6000969          0.015625
#> 22: 2022-03-18 04:28:11        1 -2.80593286   8.6066642          0.015625
#> 23: 2022-03-18 04:28:11        1  6.28681264  14.5783433          0.015625
#> 24: 2022-03-18 04:28:11        1 -2.24521770   8.7069649          0.015625
#> 25: 2022-03-18 04:28:11        1  8.33820039   5.8593706          0.015625
#> 26: 2022-03-18 04:28:11        1 -2.39762853   2.0283406          0.015625
#> 27: 2022-03-18 04:28:11        1 -4.17973267   7.3506546          0.015625
#> 28: 2022-03-18 04:28:11        1  3.11528942   3.7337356          0.015625
#> 29: 2022-03-18 04:28:11        1  0.71903737  12.9623982          0.015625
#> 30: 2022-03-18 04:28:11        1  2.15901578   8.7091371          0.015625
#> 31: 2022-03-18 04:28:11        1  8.92974156   2.7511225          0.015625
#> 32: 2022-03-18 04:28:11        1  9.37402550  10.7109571          0.015625
#> 33: 2022-03-18 04:28:11        1  4.49699790  12.8891111          0.015625
#> 34: 2022-03-18 04:28:11        1  7.21953828   3.7666281          0.015625
#> 35: 2022-03-18 04:28:11        1  1.44271563  12.8233851          0.015625
#> 36: 2022-03-18 04:28:11        1  1.90846619   5.4059819          0.015625
#> 37: 2022-03-18 04:28:11        1  2.43335460   1.2885500          0.015625
#> 38: 2022-03-18 04:28:11        1 -4.76473608   9.2376568          0.015625
#> 39: 2022-03-18 04:28:11        1 -2.11637969  12.9477499          0.015625
#> 40: 2022-03-18 04:28:11        1 -2.25275083  14.4235712          0.015625
#> 41: 2022-03-18 04:28:11        1 -1.73148635  14.3902307          0.015625
#> 42: 2022-03-18 04:28:11        1  0.32623926   7.7714899          0.015625
#> 43: 2022-03-18 04:28:11        1 -0.34587427  14.1660619          0.015625
#> 44: 2022-03-18 04:28:11        1 -3.69716158   5.1676138          0.015625
#> 45: 2022-03-18 04:28:11        1  5.41152526   6.3334756          0.015625
#> 46: 2022-03-18 04:28:11        1  2.50357962   4.4376766          0.015625
#> 47: 2022-03-18 04:28:11        1  8.30313520   7.8082786          0.015625
#> 48: 2022-03-18 04:28:11        1  3.51838641   9.5237348          0.015625
#> 49: 2022-03-18 04:28:11        1  7.07072590  10.4383155          0.015625
#> 50: 2022-03-18 04:28:11        1  0.28186193   8.4136476          0.015625
#> 51: 2022-03-18 04:28:11        1  4.92211137  10.9773611          0.015625
#> 52: 2022-03-18 04:28:11        1  7.94621971   6.6716058          0.015625
#> 53: 2022-03-18 04:28:11        1  0.42891383  10.8754544          0.015625
#> 54: 2022-03-18 04:28:11        1  5.07696668   0.9664152          0.015625
#> 55: 2022-03-18 04:28:11        1  9.45618940   9.5372230          0.015625
#> 56: 2022-03-18 04:28:11        1  1.73153273  11.2944002          0.015625
#> 57: 2022-03-18 04:28:11        1 -3.71645033  13.0462998          0.015625
#> 58: 2022-03-18 04:28:11        1 -2.45301022  11.0256545          0.015625
#> 59: 2022-03-18 04:28:11        1  1.84339422   4.1175539          0.015625
#> 60: 2022-03-18 04:28:11        1 -2.74796298   3.9417503          0.015625
#> 61: 2022-03-18 04:28:11        1  2.68394877   6.7163003          0.015625
#> 62: 2022-03-18 04:28:11        1  2.23028956  10.5144685          0.015625
#> 63: 2022-03-18 04:28:11        1  5.33662404   5.9018564          0.015625
#> 64: 2022-03-18 04:28:11        1  7.50575878   6.0532979          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity