Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain       y
#> 1: -2.667464 9.446433 0.015625 <list[3]> 2.49289

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain       y
#> 1: -2.667464 9.446433 0.015625 <list[3]> 2.49289

# all evaluations
as.data.table(instance$archive)
#>             x1          x2 fidelity stage bracket repetition          y
#>  1:  1.5911489  8.44702802 0.015625     0       6          1  33.830036
#>  2:  0.5204545 10.03467627 0.015625     0       6          1  41.898615
#>  3:  1.3139115  6.27080229 0.015625     0       6          1  17.770510
#>  4: -0.3007965 13.43857548 0.015625     0       6          1  67.571684
#>  5: -4.4270091 10.63968358 0.015625     0       6          1  16.349147
#>  6:  2.9773310  9.71209222 0.015625     0       6          1  67.408536
#>  7:  5.4914930  0.04882915 0.015625     0       6          1  20.212365
#>  8:  1.7310673 12.74993629 0.015625     0       6          1  97.069934
#>  9:  0.5461680  5.62801972 0.015625     0       6          1  18.443452
#> 10:  9.9336561  8.84811171 0.015625     0       6          1 245.723234
#> 11:  6.6086646  9.13936834 0.015625     0       6          1 170.745531
#> 12: -2.6674637  9.44643336 0.015625     0       6          1   2.492890
#> 13:  9.8679129 12.42737146 0.015625     0       6          1 367.612531
#> 14:  0.7832883  8.05086532 0.015625     0       6          1  27.553528
#> 15:  8.4145074  6.18797511 0.015625     0       6          1 134.925226
#> 16:  4.4860177  4.02920053 0.015625     0       6          1  28.548791
#> 17:  6.5228052 13.30644187 0.015625     0       6          1 287.619590
#> 18:  6.8819160  9.27814894 0.015625     0       6          1 181.128798
#> 19:  3.1465204 13.36978689 0.015625     0       6          1 146.160551
#> 20:  3.0044799  8.71455822 0.015625     0       6          1  52.598881
#> 21:  8.2253424  6.01071388 0.015625     0       6          1 127.989570
#> 22:  6.0603097  8.99297545 0.015625     0       6          1 151.821731
#> 23:  4.7246179  7.23360226 0.015625     0       6          1  75.189552
#> 24:  7.9179941 12.40902144 0.015625     0       6          1 301.222445
#> 25: -2.9448793  2.48876859 0.015625     0       6          1  72.235911
#> 26:  3.9291307  9.33636329 0.015625     0       6          1  86.309774
#> 27:  2.9584953  8.72706112 0.015625     0       6          1  51.917593
#> 28:  0.8164434  4.28635517 0.015625     0       6          1  16.764683
#> 29:  1.3225716  1.25236594 0.015625     0       6          1  19.630077
#> 30:  9.6086475 10.22449314 0.015625     0       6          1 278.728069
#> 31: -1.7188875  9.46147557 0.015625     0       6          1   8.986349
#> 32: -2.7203337 13.00464515 0.015625     0       6          1   7.227929
#> 33:  6.7492864  4.38389402 0.015625     0       6          1  78.255039
#> 34: -2.6249868  3.64399438 0.015625     0       6          1  47.154965
#> 35: -0.1706369 14.05675511 0.015625     0       6          1  80.057718
#> 36: -0.9580271 14.54743876 0.015625     0       6          1  64.445137
#> 37:  6.6503677  8.34766410 0.015625     0       6          1 152.877711
#> 38:  4.1203058 13.71603867 0.015625     0       6          1 193.750937
#> 39:  6.8917319  2.31269814 0.015625     0       6          1  51.760880
#> 40:  7.0822798  7.51694544 0.015625     0       6          1 143.180173
#> 41: -3.6923784  1.90948904 0.015625     0       6          1 109.693301
#> 42:  5.9542763  8.23832510 0.015625     0       6          1 131.973004
#> 43: -2.0529489  3.34906283 0.015625     0       6          1  42.124632
#> 44:  9.2592569  7.82245534 0.015625     0       6          1 194.377975
#> 45:  3.9800523  7.00470930 0.015625     0       6          1  50.531237
#> 46:  3.8157679 14.03385522 0.015625     0       6          1 189.071255
#> 47:  5.6441072 12.50361133 0.015625     0       6          1 228.160488
#> 48: -3.3267380  0.89539958 0.015625     0       6          1 115.899954
#> 49:  7.4596649 10.28659820 0.015625     0       6          1 222.435269
#> 50:  8.1071075  4.38852894 0.015625     0       6          1  93.538208
#> 51:  3.6288495 13.11806724 0.015625     0       6          1 157.482182
#> 52: -1.6163024  3.42174109 0.015625     0       6          1  36.926648
#> 53:  6.0657614  1.96675880 0.015625     0       6          1  39.531017
#> 54:  1.7443340 13.00803489 0.015625     0       6          1 102.251198
#> 55:  6.9790595  9.12108434 0.015625     0       6          1 179.447911
#> 56: -1.7249196 12.76053112 0.015625     0       6          1  23.921782
#> 57:  3.6148679  8.96735507 0.015625     0       6          1  70.656060
#> 58: -1.2766080 10.43552423 0.015625     0       6          1  18.323834
#> 59: -3.8862340 12.86571309 0.015625     0       6          1   2.986017
#> 60:  2.8618144  5.47745579 0.015625     0       6          1  15.062352
#> 61: -4.5742202  4.04803065 0.015625     0       6          1 106.201062
#> 62:  3.6674998 13.77922812 0.015625     0       6          1 176.005953
#> 63: -0.3481936  7.61002315 0.015625     0       6          1  20.132869
#> 64:  1.7757583 13.88002807 0.015625     0       6          1 120.602506
#>             x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-07-25 04:46:53        1   1.5911489  8.44702802          0.015625
#>  2: 2022-07-25 04:46:53        1   0.5204545 10.03467627          0.015625
#>  3: 2022-07-25 04:46:53        1   1.3139115  6.27080229          0.015625
#>  4: 2022-07-25 04:46:53        1  -0.3007965 13.43857548          0.015625
#>  5: 2022-07-25 04:46:53        1  -4.4270091 10.63968358          0.015625
#>  6: 2022-07-25 04:46:53        1   2.9773310  9.71209222          0.015625
#>  7: 2022-07-25 04:46:53        1   5.4914930  0.04882915          0.015625
#>  8: 2022-07-25 04:46:53        1   1.7310673 12.74993629          0.015625
#>  9: 2022-07-25 04:46:53        1   0.5461680  5.62801972          0.015625
#> 10: 2022-07-25 04:46:53        1   9.9336561  8.84811171          0.015625
#> 11: 2022-07-25 04:46:53        1   6.6086646  9.13936834          0.015625
#> 12: 2022-07-25 04:46:53        1  -2.6674637  9.44643336          0.015625
#> 13: 2022-07-25 04:46:53        1   9.8679129 12.42737146          0.015625
#> 14: 2022-07-25 04:46:53        1   0.7832883  8.05086532          0.015625
#> 15: 2022-07-25 04:46:53        1   8.4145074  6.18797511          0.015625
#> 16: 2022-07-25 04:46:53        1   4.4860177  4.02920053          0.015625
#> 17: 2022-07-25 04:46:53        1   6.5228052 13.30644187          0.015625
#> 18: 2022-07-25 04:46:53        1   6.8819160  9.27814894          0.015625
#> 19: 2022-07-25 04:46:53        1   3.1465204 13.36978689          0.015625
#> 20: 2022-07-25 04:46:53        1   3.0044799  8.71455822          0.015625
#> 21: 2022-07-25 04:46:53        1   8.2253424  6.01071388          0.015625
#> 22: 2022-07-25 04:46:53        1   6.0603097  8.99297545          0.015625
#> 23: 2022-07-25 04:46:53        1   4.7246179  7.23360226          0.015625
#> 24: 2022-07-25 04:46:53        1   7.9179941 12.40902144          0.015625
#> 25: 2022-07-25 04:46:53        1  -2.9448793  2.48876859          0.015625
#> 26: 2022-07-25 04:46:53        1   3.9291307  9.33636329          0.015625
#> 27: 2022-07-25 04:46:53        1   2.9584953  8.72706112          0.015625
#> 28: 2022-07-25 04:46:53        1   0.8164434  4.28635517          0.015625
#> 29: 2022-07-25 04:46:53        1   1.3225716  1.25236594          0.015625
#> 30: 2022-07-25 04:46:53        1   9.6086475 10.22449314          0.015625
#> 31: 2022-07-25 04:46:53        1  -1.7188875  9.46147557          0.015625
#> 32: 2022-07-25 04:46:53        1  -2.7203337 13.00464515          0.015625
#> 33: 2022-07-25 04:46:53        1   6.7492864  4.38389402          0.015625
#> 34: 2022-07-25 04:46:53        1  -2.6249868  3.64399438          0.015625
#> 35: 2022-07-25 04:46:53        1  -0.1706369 14.05675511          0.015625
#> 36: 2022-07-25 04:46:53        1  -0.9580271 14.54743876          0.015625
#> 37: 2022-07-25 04:46:53        1   6.6503677  8.34766410          0.015625
#> 38: 2022-07-25 04:46:53        1   4.1203058 13.71603867          0.015625
#> 39: 2022-07-25 04:46:53        1   6.8917319  2.31269814          0.015625
#> 40: 2022-07-25 04:46:53        1   7.0822798  7.51694544          0.015625
#> 41: 2022-07-25 04:46:53        1  -3.6923784  1.90948904          0.015625
#> 42: 2022-07-25 04:46:53        1   5.9542763  8.23832510          0.015625
#> 43: 2022-07-25 04:46:53        1  -2.0529489  3.34906283          0.015625
#> 44: 2022-07-25 04:46:53        1   9.2592569  7.82245534          0.015625
#> 45: 2022-07-25 04:46:53        1   3.9800523  7.00470930          0.015625
#> 46: 2022-07-25 04:46:53        1   3.8157679 14.03385522          0.015625
#> 47: 2022-07-25 04:46:53        1   5.6441072 12.50361133          0.015625
#> 48: 2022-07-25 04:46:53        1  -3.3267380  0.89539958          0.015625
#> 49: 2022-07-25 04:46:53        1   7.4596649 10.28659820          0.015625
#> 50: 2022-07-25 04:46:53        1   8.1071075  4.38852894          0.015625
#> 51: 2022-07-25 04:46:53        1   3.6288495 13.11806724          0.015625
#> 52: 2022-07-25 04:46:53        1  -1.6163024  3.42174109          0.015625
#> 53: 2022-07-25 04:46:53        1   6.0657614  1.96675880          0.015625
#> 54: 2022-07-25 04:46:53        1   1.7443340 13.00803489          0.015625
#> 55: 2022-07-25 04:46:53        1   6.9790595  9.12108434          0.015625
#> 56: 2022-07-25 04:46:53        1  -1.7249196 12.76053112          0.015625
#> 57: 2022-07-25 04:46:53        1   3.6148679  8.96735507          0.015625
#> 58: 2022-07-25 04:46:53        1  -1.2766080 10.43552423          0.015625
#> 59: 2022-07-25 04:46:53        1  -3.8862340 12.86571309          0.015625
#> 60: 2022-07-25 04:46:53        1   2.8618144  5.47745579          0.015625
#> 61: 2022-07-25 04:46:53        1  -4.5742202  4.04803065          0.015625
#> 62: 2022-07-25 04:46:53        1   3.6674998 13.77922812          0.015625
#> 63: 2022-07-25 04:46:53        1  -0.3481936  7.61002315          0.015625
#> 64: 2022-07-25 04:46:53        1   1.7757583 13.88002807          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity