Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.791217 9.469611 0.015625 <list[3]> 2.450936

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.791217 9.469611 0.015625 <list[3]> 2.450936

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1:  2.7356316  3.6874464 0.015625     0       6          1   4.458901
#>  2:  3.5836909  2.7689166 0.015625     0       6          1   5.637740
#>  3:  0.9776168  1.1226040 0.015625     0       6          1  26.595823
#>  4:  1.5195377  7.4776838 0.015625     0       6          1  25.123430
#>  5:  6.0031251  3.0640462 0.015625     0       6          1  49.591087
#>  6: -3.7040120 13.3825872 0.015625     0       6          1   3.012490
#>  7:  5.2877308  0.1140025 0.015625     0       6          1  18.013611
#>  8:  9.0121364 14.6308262 0.015625     0       6          1 420.505542
#>  9: -3.6197338  6.8526961 0.015625     0       6          1  29.683072
#> 10:  3.7154253  9.5162685 0.015625     0       6          1  83.027860
#> 11:  2.1286383 14.5156797 0.015625     0       6          1 143.313344
#> 12:  3.6240865 14.4635072 0.015625     0       6          1 192.696269
#> 13: -2.7912170  9.4696106 0.015625     0       6          1   2.450936
#> 14:  8.5867072  1.7796853 0.015625     0       6          1  55.112763
#> 15: -3.9798287  8.4835734 0.015625     0       6          1  22.392455
#> 16:  4.6482970  1.0043333 0.015625     0       6          1  12.405603
#> 17:  0.1074143  5.0466274 0.015625     0       6          1  20.159504
#> 18:  4.6329234  8.2069476 0.015625     0       6          1  88.813460
#> 19:  6.9984266  8.0117408 0.015625     0       6          1 152.835349
#> 20:  0.7595646  8.3025538 0.015625     0       6          1  29.168771
#> 21:  6.0575721 13.2989986 0.015625     0       6          1 269.368236
#> 22:  0.7712128 14.4609224 0.015625     0       6          1 110.395375
#> 23: -2.1951010  6.8056574 0.015625     0       6          1  12.430754
#> 24:  8.7255819 10.6461225 0.015625     0       6          1 264.843770
#> 25:  0.5430350 12.0455907 0.015625     0       6          1  65.841709
#> 26:  2.3532210 11.4958622 0.015625     0       6          1  85.511066
#> 27:  7.0538316  6.8801049 0.015625     0       6          1 128.757028
#> 28:  9.8504718 11.1141585 0.015625     0       6          1 318.388045
#> 29:  9.3442687  8.9117073 0.015625     0       6          1 228.405370
#> 30:  4.5612880 14.7988345 0.015625     0       6          1 246.289321
#> 31:  1.9638070  8.8209636 0.015625     0       6          1  40.286881
#> 32:  6.9359112  4.8905771 0.015625     0       6          1  89.035457
#> 33: -4.4108261 10.5713036 0.015625     0       6          1  16.431887
#> 34:  7.5450724  1.7694266 0.015625     0       6          1  49.248938
#> 35: -2.5430416 13.7805739 0.015625     0       6          1  14.558849
#> 36:  7.9808441  9.0255275 0.015625     0       6          1 198.371562
#> 37:  9.6554197  3.0272821 0.015625     0       6          1  91.433115
#> 38:  9.8790341  8.3698708 0.015625     0       6          1 229.142474
#> 39: -0.7489641  0.0594163 0.015625     0       6          1  68.152812
#> 40: -2.7760589  8.2926143 0.015625     0       6          1   6.614018
#> 41:  5.0427103  8.0666265 0.015625     0       6          1  99.799562
#> 42:  4.6065703  6.4367561 0.015625     0       6          1  59.621472
#> 43: -1.9389802  6.3132944 0.015625     0       6          1  14.886189
#> 44:  4.6511472  6.5288360 0.015625     0       6          1  62.210144
#> 45:  7.0081185  3.9520497 0.015625     0       6          1  74.882511
#> 46:  0.7030387  5.3919988 0.015625     0       6          1  17.571011
#> 47:  4.0916329  9.3859116 0.015625     0       6          1  92.459330
#> 48:  4.3832904  6.4447597 0.015625     0       6          1  53.548783
#> 49: -3.3020855  1.0194753 0.015625     0       6          1 112.271880
#> 50:  3.9839700  1.6483932 0.015625     0       6          1   5.861149
#> 51: -2.7945243  8.2967764 0.015625     0       6          1   6.687122
#> 52:  1.3306566  4.6416736 0.015625     0       6          1  12.780828
#> 53: -0.8101956 10.5601969 0.015625     0       6          1  27.185341
#> 54: -0.1923207  7.9317870 0.015625     0       6          1  22.064283
#> 55: -3.8504945  5.4739496 0.015625     0       6          1  53.265734
#> 56:  5.5708897  9.9139766 0.015625     0       6          1 157.124018
#> 57:  8.5687982  0.8886645 0.015625     0       6          1  43.003278
#> 58:  7.0430268 10.0690583 0.015625     0       6          1 206.111570
#> 59:  1.7004559  2.0022959 0.015625     0       6          1  10.663572
#> 60: -0.4973503 10.0390367 0.015625     0       6          1  28.935596
#> 61: -4.3799620  4.1589335 0.015625     0       6          1  95.261159
#> 62:  3.0317803  6.1703203 0.015625     0       6          1  22.667446
#> 63:  1.6588097  6.8805303 0.015625     0       6          1  20.962092
#> 64:  7.0227328  5.9693232 0.015625     0       6          1 109.829621
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-02-23 04:26:34        1   2.7356316   3.6874464          0.015625
#>  2: 2022-02-23 04:26:34        1   3.5836909   2.7689166          0.015625
#>  3: 2022-02-23 04:26:34        1   0.9776168   1.1226040          0.015625
#>  4: 2022-02-23 04:26:34        1   1.5195377   7.4776838          0.015625
#>  5: 2022-02-23 04:26:34        1   6.0031251   3.0640462          0.015625
#>  6: 2022-02-23 04:26:34        1  -3.7040120  13.3825872          0.015625
#>  7: 2022-02-23 04:26:34        1   5.2877308   0.1140025          0.015625
#>  8: 2022-02-23 04:26:34        1   9.0121364  14.6308262          0.015625
#>  9: 2022-02-23 04:26:34        1  -3.6197338   6.8526961          0.015625
#> 10: 2022-02-23 04:26:34        1   3.7154253   9.5162685          0.015625
#> 11: 2022-02-23 04:26:34        1   2.1286383  14.5156797          0.015625
#> 12: 2022-02-23 04:26:34        1   3.6240865  14.4635072          0.015625
#> 13: 2022-02-23 04:26:34        1  -2.7912170   9.4696106          0.015625
#> 14: 2022-02-23 04:26:34        1   8.5867072   1.7796853          0.015625
#> 15: 2022-02-23 04:26:34        1  -3.9798287   8.4835734          0.015625
#> 16: 2022-02-23 04:26:34        1   4.6482970   1.0043333          0.015625
#> 17: 2022-02-23 04:26:34        1   0.1074143   5.0466274          0.015625
#> 18: 2022-02-23 04:26:34        1   4.6329234   8.2069476          0.015625
#> 19: 2022-02-23 04:26:34        1   6.9984266   8.0117408          0.015625
#> 20: 2022-02-23 04:26:34        1   0.7595646   8.3025538          0.015625
#> 21: 2022-02-23 04:26:34        1   6.0575721  13.2989986          0.015625
#> 22: 2022-02-23 04:26:34        1   0.7712128  14.4609224          0.015625
#> 23: 2022-02-23 04:26:34        1  -2.1951010   6.8056574          0.015625
#> 24: 2022-02-23 04:26:34        1   8.7255819  10.6461225          0.015625
#> 25: 2022-02-23 04:26:34        1   0.5430350  12.0455907          0.015625
#> 26: 2022-02-23 04:26:34        1   2.3532210  11.4958622          0.015625
#> 27: 2022-02-23 04:26:34        1   7.0538316   6.8801049          0.015625
#> 28: 2022-02-23 04:26:34        1   9.8504718  11.1141585          0.015625
#> 29: 2022-02-23 04:26:34        1   9.3442687   8.9117073          0.015625
#> 30: 2022-02-23 04:26:34        1   4.5612880  14.7988345          0.015625
#> 31: 2022-02-23 04:26:34        1   1.9638070   8.8209636          0.015625
#> 32: 2022-02-23 04:26:34        1   6.9359112   4.8905771          0.015625
#> 33: 2022-02-23 04:26:34        1  -4.4108261  10.5713036          0.015625
#> 34: 2022-02-23 04:26:34        1   7.5450724   1.7694266          0.015625
#> 35: 2022-02-23 04:26:34        1  -2.5430416  13.7805739          0.015625
#> 36: 2022-02-23 04:26:34        1   7.9808441   9.0255275          0.015625
#> 37: 2022-02-23 04:26:34        1   9.6554197   3.0272821          0.015625
#> 38: 2022-02-23 04:26:34        1   9.8790341   8.3698708          0.015625
#> 39: 2022-02-23 04:26:34        1  -0.7489641   0.0594163          0.015625
#> 40: 2022-02-23 04:26:34        1  -2.7760589   8.2926143          0.015625
#> 41: 2022-02-23 04:26:34        1   5.0427103   8.0666265          0.015625
#> 42: 2022-02-23 04:26:34        1   4.6065703   6.4367561          0.015625
#> 43: 2022-02-23 04:26:34        1  -1.9389802   6.3132944          0.015625
#> 44: 2022-02-23 04:26:34        1   4.6511472   6.5288360          0.015625
#> 45: 2022-02-23 04:26:34        1   7.0081185   3.9520497          0.015625
#> 46: 2022-02-23 04:26:34        1   0.7030387   5.3919988          0.015625
#> 47: 2022-02-23 04:26:34        1   4.0916329   9.3859116          0.015625
#> 48: 2022-02-23 04:26:34        1   4.3832904   6.4447597          0.015625
#> 49: 2022-02-23 04:26:34        1  -3.3020855   1.0194753          0.015625
#> 50: 2022-02-23 04:26:34        1   3.9839700   1.6483932          0.015625
#> 51: 2022-02-23 04:26:34        1  -2.7945243   8.2967764          0.015625
#> 52: 2022-02-23 04:26:34        1   1.3306566   4.6416736          0.015625
#> 53: 2022-02-23 04:26:34        1  -0.8101956  10.5601969          0.015625
#> 54: 2022-02-23 04:26:34        1  -0.1923207   7.9317870          0.015625
#> 55: 2022-02-23 04:26:34        1  -3.8504945   5.4739496          0.015625
#> 56: 2022-02-23 04:26:34        1   5.5708897   9.9139766          0.015625
#> 57: 2022-02-23 04:26:34        1   8.5687982   0.8886645          0.015625
#> 58: 2022-02-23 04:26:34        1   7.0430268  10.0690583          0.015625
#> 59: 2022-02-23 04:26:34        1   1.7004559   2.0022959          0.015625
#> 60: 2022-02-23 04:26:34        1  -0.4973503  10.0390367          0.015625
#> 61: 2022-02-23 04:26:34        1  -4.3799620   4.1589335          0.015625
#> 62: 2022-02-23 04:26:34        1   3.0317803   6.1703203          0.015625
#> 63: 2022-02-23 04:26:34        1   1.6588097   6.8805303          0.015625
#> 64: 2022-02-23 04:26:34        1   7.0227328   5.9693232          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity