Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.623138 12.07532 0.015625 <list[3]> 1.498808

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.623138 12.07532 0.015625 <list[3]> 1.498808

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1:  8.1751259 14.6537998 0.015625     0       6          1 391.521626
#>  2:  8.4459147  0.8101226 0.015625     0       6          1  41.352884
#>  3:  6.6198943  2.4566643 0.015625     0       6          1  50.930386
#>  4: -3.9920088 11.4551743 0.015625     0       6          1   5.593128
#>  5:  3.1749856  7.5359938 0.015625     0       6          1  39.831519
#>  6:  7.3191612  3.5199097 0.015625     0       6          1  71.468467
#>  7:  8.3027299 12.2549280 0.015625     0       6          1 306.842094
#>  8: -2.3887323  2.1736688 0.015625     0       6          1  63.888548
#>  9: -2.3903889  6.6907107 0.015625     0       6          1  13.802268
#> 10: -1.8411953  0.5824692 0.015625     0       6          1  78.873396
#> 11:  3.9865275 14.6907976 0.015625     0       6          1 215.238882
#> 12:  8.3770375  9.6250339 0.015625     0       6          1 224.238699
#> 13:  0.9496082 12.4521407 0.015625     0       6          1  78.564778
#> 14:  7.0390957  1.2587876 0.015625     0       6          1  41.374518
#> 15:  9.1650913  8.0904552 0.015625     0       6          1 199.372972
#> 16:  7.4655563  1.3353506 0.015625     0       6          1  43.924941
#> 17: -3.6231385 12.0753214 0.015625     0       6          1   1.498808
#> 18:  6.5185891 11.4977511 0.015625     0       6          1 231.502901
#> 19: -4.1593371  8.6158618 0.015625     0       6          1  25.530115
#> 20: -1.0331997 11.6714202 0.015625     0       6          1  30.870692
#> 21: -1.0221302 10.8093818 0.015625     0       6          1  24.933554
#> 22: -0.9270241 11.8398654 0.015625     0       6          1  34.581919
#> 23:  6.9717625 12.2657601 0.015625     0       6          1 269.182058
#> 24:  4.5311605  9.8213996 0.015625     0       6          1 116.464404
#> 25: -2.5981069  7.4675508 0.015625     0       6          1  10.047137
#> 26:  1.8348943 12.4889801 0.015625     0       6          1  94.091109
#> 27:  6.4472369  9.6991337 0.015625     0       6          1 180.310707
#> 28:  8.7569929  9.3184777 0.015625     0       6          1 224.405886
#> 29: -1.6323576 11.9537735 0.015625     0       6          1  20.127464
#> 30:  9.6985300  6.3158734 0.015625     0       6          1 166.120996
#> 31:  5.7290662  0.7737533 0.015625     0       6          1  26.475027
#> 32:  9.8729797 13.4822973 0.015625     0       6          1 409.327548
#> 33:  6.0047927  9.5596870 0.015625     0       6          1 163.423280
#> 34:  8.5732565  1.0551683 0.015625     0       6          1  45.147380
#> 35:  5.3296477 12.0278856 0.015625     0       6          1 201.523038
#> 36:  8.1817830 13.6315082 0.015625     0       6          1 352.681282
#> 37:  2.3354990  0.1992944 0.015625     0       6          1   8.420840
#> 38:  3.5115248 14.7023396 0.015625     0       6          1 194.590348
#> 39:  5.0243568  4.0858876 0.015625     0       6          1  41.103140
#> 40:  9.0135251  1.1357625 0.015625     0       6          1  49.964069
#> 41: -4.3534489 10.7985490 0.015625     0       6          1  13.986849
#> 42: -1.0705228 13.9087728 0.015625     0       6          1  52.671571
#> 43:  1.0504781  3.4193288 0.015625     0       6          1  15.662458
#> 44: -3.6052064  1.1218066 0.015625     0       6          1 122.756941
#> 45: -1.6362847  0.6855195 0.015625     0       6          1  73.388125
#> 46:  2.3351752  4.5196896 0.015625     0       6          1   7.633474
#> 47:  9.5415080  2.2032775 0.015625     0       6          1  74.248582
#> 48:  7.0238354 13.4721684 0.015625     0       6          1 310.662953
#> 49: -3.7588615  5.3449543 0.015625     0       6          1  52.181399
#> 50:  4.9745747  1.9329581 0.015625     0       6          1  22.032951
#> 51:  2.2548590  7.2003888 0.015625     0       6          1  25.394612
#> 52:  5.5096234 10.2989866 0.015625     0       6          1 164.114968
#> 53:  7.6265168 12.4470834 0.015625     0       6          1 294.294495
#> 54: -0.4570089  2.9551466 0.015625     0       6          1  32.894735
#> 55:  4.8055508 14.7775114 0.015625     0       6          1 257.877505
#> 56:  4.0754126  6.9447262 0.015625     0       6          1  52.179115
#> 57: -0.4537556  1.3317707 0.015625     0       6          1  47.755191
#> 58:  9.7254279  3.9940481 0.015625     0       6          1 112.434592
#> 59: -4.9474006 14.8602093 0.015625     0       6          1  12.290457
#> 60:  8.6195361  4.1215715 0.015625     0       6          1  94.655778
#> 61: -0.9142550 14.5891848 0.015625     0       6          1  66.390333
#> 62:  0.7043220 14.1891005 0.015625     0       6          1 103.710811
#> 63:  9.3856961  0.9953048 0.015625     0       6          1  52.599492
#> 64:  8.8593523 13.2592509 0.015625     0       6          1 360.846036
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-06-24 04:40:55        1   8.1751259  14.6537998          0.015625
#>  2: 2022-06-24 04:40:55        1   8.4459147   0.8101226          0.015625
#>  3: 2022-06-24 04:40:55        1   6.6198943   2.4566643          0.015625
#>  4: 2022-06-24 04:40:55        1  -3.9920088  11.4551743          0.015625
#>  5: 2022-06-24 04:40:55        1   3.1749856   7.5359938          0.015625
#>  6: 2022-06-24 04:40:55        1   7.3191612   3.5199097          0.015625
#>  7: 2022-06-24 04:40:55        1   8.3027299  12.2549280          0.015625
#>  8: 2022-06-24 04:40:55        1  -2.3887323   2.1736688          0.015625
#>  9: 2022-06-24 04:40:55        1  -2.3903889   6.6907107          0.015625
#> 10: 2022-06-24 04:40:55        1  -1.8411953   0.5824692          0.015625
#> 11: 2022-06-24 04:40:55        1   3.9865275  14.6907976          0.015625
#> 12: 2022-06-24 04:40:55        1   8.3770375   9.6250339          0.015625
#> 13: 2022-06-24 04:40:55        1   0.9496082  12.4521407          0.015625
#> 14: 2022-06-24 04:40:55        1   7.0390957   1.2587876          0.015625
#> 15: 2022-06-24 04:40:55        1   9.1650913   8.0904552          0.015625
#> 16: 2022-06-24 04:40:55        1   7.4655563   1.3353506          0.015625
#> 17: 2022-06-24 04:40:55        1  -3.6231385  12.0753214          0.015625
#> 18: 2022-06-24 04:40:55        1   6.5185891  11.4977511          0.015625
#> 19: 2022-06-24 04:40:55        1  -4.1593371   8.6158618          0.015625
#> 20: 2022-06-24 04:40:55        1  -1.0331997  11.6714202          0.015625
#> 21: 2022-06-24 04:40:55        1  -1.0221302  10.8093818          0.015625
#> 22: 2022-06-24 04:40:55        1  -0.9270241  11.8398654          0.015625
#> 23: 2022-06-24 04:40:55        1   6.9717625  12.2657601          0.015625
#> 24: 2022-06-24 04:40:55        1   4.5311605   9.8213996          0.015625
#> 25: 2022-06-24 04:40:55        1  -2.5981069   7.4675508          0.015625
#> 26: 2022-06-24 04:40:55        1   1.8348943  12.4889801          0.015625
#> 27: 2022-06-24 04:40:55        1   6.4472369   9.6991337          0.015625
#> 28: 2022-06-24 04:40:55        1   8.7569929   9.3184777          0.015625
#> 29: 2022-06-24 04:40:55        1  -1.6323576  11.9537735          0.015625
#> 30: 2022-06-24 04:40:55        1   9.6985300   6.3158734          0.015625
#> 31: 2022-06-24 04:40:55        1   5.7290662   0.7737533          0.015625
#> 32: 2022-06-24 04:40:55        1   9.8729797  13.4822973          0.015625
#> 33: 2022-06-24 04:40:55        1   6.0047927   9.5596870          0.015625
#> 34: 2022-06-24 04:40:55        1   8.5732565   1.0551683          0.015625
#> 35: 2022-06-24 04:40:55        1   5.3296477  12.0278856          0.015625
#> 36: 2022-06-24 04:40:55        1   8.1817830  13.6315082          0.015625
#> 37: 2022-06-24 04:40:55        1   2.3354990   0.1992944          0.015625
#> 38: 2022-06-24 04:40:55        1   3.5115248  14.7023396          0.015625
#> 39: 2022-06-24 04:40:55        1   5.0243568   4.0858876          0.015625
#> 40: 2022-06-24 04:40:55        1   9.0135251   1.1357625          0.015625
#> 41: 2022-06-24 04:40:55        1  -4.3534489  10.7985490          0.015625
#> 42: 2022-06-24 04:40:55        1  -1.0705228  13.9087728          0.015625
#> 43: 2022-06-24 04:40:55        1   1.0504781   3.4193288          0.015625
#> 44: 2022-06-24 04:40:55        1  -3.6052064   1.1218066          0.015625
#> 45: 2022-06-24 04:40:55        1  -1.6362847   0.6855195          0.015625
#> 46: 2022-06-24 04:40:55        1   2.3351752   4.5196896          0.015625
#> 47: 2022-06-24 04:40:55        1   9.5415080   2.2032775          0.015625
#> 48: 2022-06-24 04:40:55        1   7.0238354  13.4721684          0.015625
#> 49: 2022-06-24 04:40:55        1  -3.7588615   5.3449543          0.015625
#> 50: 2022-06-24 04:40:55        1   4.9745747   1.9329581          0.015625
#> 51: 2022-06-24 04:40:55        1   2.2548590   7.2003888          0.015625
#> 52: 2022-06-24 04:40:55        1   5.5096234  10.2989866          0.015625
#> 53: 2022-06-24 04:40:55        1   7.6265168  12.4470834          0.015625
#> 54: 2022-06-24 04:40:55        1  -0.4570089   2.9551466          0.015625
#> 55: 2022-06-24 04:40:55        1   4.8055508  14.7775114          0.015625
#> 56: 2022-06-24 04:40:55        1   4.0754126   6.9447262          0.015625
#> 57: 2022-06-24 04:40:55        1  -0.4537556   1.3317707          0.015625
#> 58: 2022-06-24 04:40:55        1   9.7254279   3.9940481          0.015625
#> 59: 2022-06-24 04:40:55        1  -4.9474006  14.8602093          0.015625
#> 60: 2022-06-24 04:40:55        1   8.6195361   4.1215715          0.015625
#> 61: 2022-06-24 04:40:55        1  -0.9142550  14.5891848          0.015625
#> 62: 2022-06-24 04:40:55        1   0.7043220  14.1891005          0.015625
#> 63: 2022-06-24 04:40:55        1   9.3856961   0.9953048          0.015625
#> 64: 2022-06-24 04:40:55        1   8.8593523  13.2592509          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity