Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1      x2 fidelity  x_domain        y
#> 1: -3.477665 12.4605 0.015625 <list[3]> 1.241711

# best scoring evaluation
instance$result
#>           x1      x2 fidelity  x_domain        y
#> 1: -3.477665 12.4605 0.015625 <list[3]> 1.241711

# all evaluations
as.data.table(instance$archive)
#>             x1          x2 fidelity stage bracket repetition          y
#>  1:  3.9282351 10.63423906 0.015625     0       6          1 111.623510
#>  2: -2.2545373 14.28867384 0.015625     0       6          1  24.583952
#>  3: -0.3934232 12.61179576 0.015625     0       6          1  54.639508
#>  4:  3.2921310 14.89956563 0.015625     0       6          1 191.109769
#>  5: -1.0564133  3.51444460 0.015625     0       6          1  32.374328
#>  6: -3.6900915  0.21522993 0.015625     0       6          1 147.646075
#>  7:  4.4960860  0.30807320 0.015625     0       6          1   8.648615
#>  8:  6.3592481 14.01115974 0.015625     0       6          1 304.806161
#>  9:  1.9763184  2.23508158 0.015625     0       6          1   6.758993
#> 10:  1.3893014 11.50431572 0.015625     0       6          1  70.349217
#> 11:  6.4204492 13.41951707 0.015625     0       6          1 287.505753
#> 12: -0.9965247 14.90259247 0.015625     0       6          1  68.302478
#> 13:  5.8113493  1.89177531 0.015625     0       6          1  35.382985
#> 14:  4.6091455  8.13778080 0.015625     0       6          1  86.807521
#> 15:  9.8421842 13.59559571 0.015625     0       6          1 412.561492
#> 16: -4.5570992  0.59277126 0.015625     0       6          1 185.367776
#> 17:  7.9082438 14.16744887 0.015625     0       6          1 364.081340
#> 18: -1.5723009  1.99096640 0.015625     0       6          1  53.379904
#> 19:  5.0673639  1.38699989 0.015625     0       6          1  20.425925
#> 20: -3.7740513 14.47585565 0.015625     0       6          1   6.381432
#> 21:  7.1353792  6.84684114 0.015625     0       6          1 129.482218
#> 22: -0.7102124  6.20191631 0.015625     0       6          1  18.171552
#> 23: -1.9413419 13.05219877 0.015625     0       6          1  21.319004
#> 24:  1.5134989  4.07773057 0.015625     0       6          1  10.723021
#> 25:  3.6485255  4.84802359 0.015625     0       6          1  19.630048
#> 26: -3.4776648 12.46049924 0.015625     0       6          1   1.241711
#> 27:  9.4302207  9.95785984 0.015625     0       6          1 263.882982
#> 28:  1.3761889  5.92477367 0.015625     0       6          1  16.087357
#> 29:  7.6949513  9.03165532 0.015625     0       6          1 192.637025
#> 30:  4.8887931  1.39469224 0.015625     0       6          1  17.641539
#> 31:  0.6516532  0.07443383 0.015625     0       6          1  41.659021
#> 32: -2.6198188  0.94421009 0.015625     0       6          1  90.721065
#> 33: -1.8422727 14.57803373 0.015625     0       6          1  38.134610
#> 34: -3.7033359 10.23086310 0.015625     0       6          1   6.220127
#> 35:  9.6491436 14.82744097 0.015625     0       6          1 455.257706
#> 36:  1.7007116  2.49191563 0.015625     0       6          1   9.548592
#> 37:  1.0433578  4.44389888 0.015625     0       6          1  14.837989
#> 38:  8.8083161  7.06724346 0.015625     0       6          1 163.469662
#> 39:  4.0074696  8.47764083 0.015625     0       6          1  73.700063
#> 40:  2.4594047  0.63052275 0.015625     0       6          1   5.240397
#> 41:  7.4662773  6.01210781 0.015625     0       6          1 117.284117
#> 42:  7.1496130  9.04475263 0.015625     0       6          1 181.392327
#> 43:  3.1432745  3.56017317 0.015625     0       6          1   5.501271
#> 44:  3.8032809  3.43232752 0.015625     0       6          1  11.670110
#> 45:  8.2936133  5.92144929 0.015625     0       6          1 127.050641
#> 46:  8.5476569 10.49411575 0.015625     0       6          1 255.136568
#> 47: -3.8538011 14.31370796 0.015625     0       6          1   5.702572
#> 48:  2.0558708  3.03502041 0.015625     0       6          1   5.554141
#> 49:  1.3983006 13.11981261 0.015625     0       6          1  97.862303
#> 50: -0.4658175  9.03382500 0.015625     0       6          1  23.803851
#> 51:  8.7294674 12.74611137 0.015625     0       6          1 337.387327
#> 52:  1.6083084  3.52215758 0.015625     0       6          1   9.639894
#> 53:  9.7068382  4.94010066 0.015625     0       6          1 132.842171
#> 54:  3.0884266  5.44740754 0.015625     0       6          1  16.972403
#> 55: -4.4500554 12.23669551 0.015625     0       6          1   9.625903
#> 56:  8.6430778 14.83209902 0.015625     0       6          1 414.913840
#> 57:  6.9016166  0.44688409 0.015625     0       6          1  33.557586
#> 58:  5.5846793  7.19244775 0.015625     0       6          1 100.560268
#> 59: -0.8662489  9.99261387 0.015625     0       6          1  22.931758
#> 60: -3.5823983 14.43281882 0.015625     0       6          1   6.775754
#> 61:  4.4272349 14.63240816 0.015625     0       6          1 234.582161
#> 62: -2.4044373  2.31608926 0.015625     0       6          1  62.003014
#> 63: -1.1410281  6.26439118 0.015625     0       6          1  16.534143
#> 64:  8.4862816  7.11365947 0.015625     0       6          1 158.226267
#>             x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-04-29 04:39:27        1   3.9282351 10.63423906          0.015625
#>  2: 2022-04-29 04:39:27        1  -2.2545373 14.28867384          0.015625
#>  3: 2022-04-29 04:39:27        1  -0.3934232 12.61179576          0.015625
#>  4: 2022-04-29 04:39:27        1   3.2921310 14.89956563          0.015625
#>  5: 2022-04-29 04:39:27        1  -1.0564133  3.51444460          0.015625
#>  6: 2022-04-29 04:39:27        1  -3.6900915  0.21522993          0.015625
#>  7: 2022-04-29 04:39:27        1   4.4960860  0.30807320          0.015625
#>  8: 2022-04-29 04:39:27        1   6.3592481 14.01115974          0.015625
#>  9: 2022-04-29 04:39:27        1   1.9763184  2.23508158          0.015625
#> 10: 2022-04-29 04:39:27        1   1.3893014 11.50431572          0.015625
#> 11: 2022-04-29 04:39:27        1   6.4204492 13.41951707          0.015625
#> 12: 2022-04-29 04:39:27        1  -0.9965247 14.90259247          0.015625
#> 13: 2022-04-29 04:39:27        1   5.8113493  1.89177531          0.015625
#> 14: 2022-04-29 04:39:27        1   4.6091455  8.13778080          0.015625
#> 15: 2022-04-29 04:39:27        1   9.8421842 13.59559571          0.015625
#> 16: 2022-04-29 04:39:27        1  -4.5570992  0.59277126          0.015625
#> 17: 2022-04-29 04:39:27        1   7.9082438 14.16744887          0.015625
#> 18: 2022-04-29 04:39:27        1  -1.5723009  1.99096640          0.015625
#> 19: 2022-04-29 04:39:27        1   5.0673639  1.38699989          0.015625
#> 20: 2022-04-29 04:39:27        1  -3.7740513 14.47585565          0.015625
#> 21: 2022-04-29 04:39:27        1   7.1353792  6.84684114          0.015625
#> 22: 2022-04-29 04:39:27        1  -0.7102124  6.20191631          0.015625
#> 23: 2022-04-29 04:39:27        1  -1.9413419 13.05219877          0.015625
#> 24: 2022-04-29 04:39:27        1   1.5134989  4.07773057          0.015625
#> 25: 2022-04-29 04:39:27        1   3.6485255  4.84802359          0.015625
#> 26: 2022-04-29 04:39:27        1  -3.4776648 12.46049924          0.015625
#> 27: 2022-04-29 04:39:27        1   9.4302207  9.95785984          0.015625
#> 28: 2022-04-29 04:39:27        1   1.3761889  5.92477367          0.015625
#> 29: 2022-04-29 04:39:27        1   7.6949513  9.03165532          0.015625
#> 30: 2022-04-29 04:39:27        1   4.8887931  1.39469224          0.015625
#> 31: 2022-04-29 04:39:27        1   0.6516532  0.07443383          0.015625
#> 32: 2022-04-29 04:39:27        1  -2.6198188  0.94421009          0.015625
#> 33: 2022-04-29 04:39:27        1  -1.8422727 14.57803373          0.015625
#> 34: 2022-04-29 04:39:27        1  -3.7033359 10.23086310          0.015625
#> 35: 2022-04-29 04:39:27        1   9.6491436 14.82744097          0.015625
#> 36: 2022-04-29 04:39:27        1   1.7007116  2.49191563          0.015625
#> 37: 2022-04-29 04:39:27        1   1.0433578  4.44389888          0.015625
#> 38: 2022-04-29 04:39:27        1   8.8083161  7.06724346          0.015625
#> 39: 2022-04-29 04:39:27        1   4.0074696  8.47764083          0.015625
#> 40: 2022-04-29 04:39:27        1   2.4594047  0.63052275          0.015625
#> 41: 2022-04-29 04:39:27        1   7.4662773  6.01210781          0.015625
#> 42: 2022-04-29 04:39:27        1   7.1496130  9.04475263          0.015625
#> 43: 2022-04-29 04:39:27        1   3.1432745  3.56017317          0.015625
#> 44: 2022-04-29 04:39:27        1   3.8032809  3.43232752          0.015625
#> 45: 2022-04-29 04:39:27        1   8.2936133  5.92144929          0.015625
#> 46: 2022-04-29 04:39:27        1   8.5476569 10.49411575          0.015625
#> 47: 2022-04-29 04:39:27        1  -3.8538011 14.31370796          0.015625
#> 48: 2022-04-29 04:39:27        1   2.0558708  3.03502041          0.015625
#> 49: 2022-04-29 04:39:27        1   1.3983006 13.11981261          0.015625
#> 50: 2022-04-29 04:39:27        1  -0.4658175  9.03382500          0.015625
#> 51: 2022-04-29 04:39:27        1   8.7294674 12.74611137          0.015625
#> 52: 2022-04-29 04:39:27        1   1.6083084  3.52215758          0.015625
#> 53: 2022-04-29 04:39:27        1   9.7068382  4.94010066          0.015625
#> 54: 2022-04-29 04:39:27        1   3.0884266  5.44740754          0.015625
#> 55: 2022-04-29 04:39:27        1  -4.4500554 12.23669551          0.015625
#> 56: 2022-04-29 04:39:27        1   8.6430778 14.83209902          0.015625
#> 57: 2022-04-29 04:39:27        1   6.9016166  0.44688409          0.015625
#> 58: 2022-04-29 04:39:27        1   5.5846793  7.19244775          0.015625
#> 59: 2022-04-29 04:39:27        1  -0.8662489  9.99261387          0.015625
#> 60: 2022-04-29 04:39:27        1  -3.5823983 14.43281882          0.015625
#> 61: 2022-04-29 04:39:27        1   4.4272349 14.63240816          0.015625
#> 62: 2022-04-29 04:39:27        1  -2.4044373  2.31608926          0.015625
#> 63: 2022-04-29 04:39:27        1  -1.1410281  6.26439118          0.015625
#> 64: 2022-04-29 04:39:27        1   8.4862816  7.11365947          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity