Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.421425 12.10384 0.015625 <list[3]> 0.8605227

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.421425 12.10384 0.015625 <list[3]> 0.8605227

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition           y
#>  1: -3.30737494  2.77392262 0.015625     0       6          1  78.4324249
#>  2:  4.11892278  0.08842897 0.015625     0       6          1   4.6450822
#>  3: -0.57538189  1.28124931 0.015625     0       6          1  49.9184212
#>  4:  9.78017220  2.96174657 0.015625     0       6          1  92.8963860
#>  5:  3.15067065  6.18098228 0.015625     0       6          1  24.3124685
#>  6: -2.55553150  8.21819800 0.015625     0       6          1   6.2021493
#>  7:  4.22648417 10.81276884 0.015625     0       6          1 126.2999699
#>  8:  6.56713182  4.13572276 0.015625     0       6          1  71.9485202
#>  9: -1.49484871  5.28288878 0.015625     0       6          1  20.7454225
#> 10:  7.58321565  6.60181395 0.015625     0       6          1 131.4385542
#> 11: -1.83572442 12.42917607 0.015625     0       6          1  19.0724285
#> 12:  7.04458816 10.49120493 0.015625     0       6          1 217.9424592
#> 13: -4.18489983  9.98472055 0.015625     0       6          1  15.4979592
#> 14:  6.31869114 12.77819767 0.015625     0       6          1 263.1778261
#> 15:  5.41928392  4.58679188 0.015625     0       6          1  56.0379584
#> 16:  6.92216436  2.74746128 0.015625     0       6          1  57.2861482
#> 17:  2.01694546 10.58984510 0.015625     0       6          1  64.7597927
#> 18:  2.11987206 11.45980011 0.015625     0       6          1  80.6005364
#> 19:  8.68530326  9.40487941 0.015625     0       6          1 225.1716136
#> 20:  9.97224043 12.30643925 0.015625     0       6          1 367.3794285
#> 21: -4.70897578  6.01493103 0.015625     0       6          1  76.5762440
#> 22: -2.74402890 10.50328363 0.015625     0       6          1   1.1559034
#> 23:  6.44417686 14.17548128 0.015625     0       6          1 313.7672866
#> 24:  8.91629873 11.93276885 0.015625     0       6          1 314.1632672
#> 25:  5.03203721  4.57562302 0.015625     0       6          1  46.7246307
#> 26:  9.87115650 12.96158578 0.015625     0       6          1 388.4852473
#> 27:  4.16896626 11.21438419 0.015625     0       6          1 133.0667278
#> 28:  4.32872430  8.35207471 0.015625     0       6          1  81.4935036
#> 29: -4.08530477  1.45686109 0.015625     0       6          1 137.9590381
#> 30:  2.52444387  1.02616570 0.015625     0       6          1   3.4962807
#> 31:  4.55045617  0.11069856 0.015625     0       6          1   8.9649851
#> 32: -3.51428991 12.48280994 0.015625     0       6          1   1.3170953
#> 33: -0.74309650 14.66769678 0.015625     0       6          1  72.8424525
#> 34: -3.86316092  4.60545357 0.015625     0       6          1  66.8202459
#> 35: -3.42142521 12.10384208 0.015625     0       6          1   0.8605227
#> 36:  4.98095461 13.13872170 0.015625     0       6          1 217.1330625
#> 37:  1.29749357 11.88226326 0.015625     0       6          1  74.9310688
#> 38:  2.82784355 14.68408567 0.015625     0       6          1 168.2808305
#> 39:  9.74999447 12.59424873 0.015625     0       6          1 369.1175019
#> 40: -3.41268476 14.15758968 0.015625     0       6          1   6.3561826
#> 41:  0.08983161 11.04522706 0.015625     0       6          1  46.4782248
#> 42:  2.09824433  5.78602146 0.015625     0       6          1  14.1077482
#> 43:  1.54989315  4.18735483 0.015625     0       6          1  10.5373633
#> 44:  7.40942996  1.55101293 0.015625     0       6          1  46.1139506
#> 45: -3.49465540  3.04979705 0.015625     0       6          1  79.9799744
#> 46:  1.32882627  1.09887398 0.015625     0       6          1  20.3694040
#> 47:  2.62218272  4.11611426 0.015625     0       6          1   5.9825226
#> 48:  5.66478974 11.57098771 0.015625     0       6          1 202.7869922
#> 49: -3.54470986 12.82784099 0.015625     0       6          1   1.8074473
#> 50:  6.97894140  9.23158280 0.015625     0       6          1 182.2708336
#> 51:  1.63992194 11.36231273 0.015625     0       6          1  71.5832003
#> 52:  6.05305267  5.89931836 0.015625     0       6          1  90.0181640
#> 53:  7.74504337 13.96981867 0.015625     0       6          1 351.5223874
#> 54: -3.45962326  9.57636627 0.015625     0       6          1   6.1593157
#> 55: -3.90295989  3.39898215 0.015625     0       6          1  89.1887199
#> 56:  9.72250674  8.96284702 0.015625     0       6          1 242.0096269
#> 57:  9.41733226 13.39111039 0.015625     0       6          1 386.6160177
#> 58:  7.53415908  1.92534691 0.015625     0       6          1  51.1004236
#> 59:  3.41872138 10.59515351 0.015625     0       6          1  94.4058199
#> 60:  1.05399839  5.43700752 0.015625     0       6          1  15.9115284
#> 61: -0.49641909  4.01907836 0.015625     0       6          1  26.1635519
#> 62:  6.49440046  0.65567103 0.015625     0       6          1  33.0417967
#> 63:  6.63010869  1.24812582 0.015625     0       6          1  38.8208392
#> 64:  6.06457599  7.55495725 0.015625     0       6          1 120.9030451
#>              x1          x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-10 04:27:37        1 -3.30737494  2.77392262          0.015625
#>  2: 2022-05-10 04:27:37        1  4.11892278  0.08842897          0.015625
#>  3: 2022-05-10 04:27:37        1 -0.57538189  1.28124931          0.015625
#>  4: 2022-05-10 04:27:37        1  9.78017220  2.96174657          0.015625
#>  5: 2022-05-10 04:27:37        1  3.15067065  6.18098228          0.015625
#>  6: 2022-05-10 04:27:37        1 -2.55553150  8.21819800          0.015625
#>  7: 2022-05-10 04:27:37        1  4.22648417 10.81276884          0.015625
#>  8: 2022-05-10 04:27:37        1  6.56713182  4.13572276          0.015625
#>  9: 2022-05-10 04:27:37        1 -1.49484871  5.28288878          0.015625
#> 10: 2022-05-10 04:27:37        1  7.58321565  6.60181395          0.015625
#> 11: 2022-05-10 04:27:37        1 -1.83572442 12.42917607          0.015625
#> 12: 2022-05-10 04:27:37        1  7.04458816 10.49120493          0.015625
#> 13: 2022-05-10 04:27:37        1 -4.18489983  9.98472055          0.015625
#> 14: 2022-05-10 04:27:37        1  6.31869114 12.77819767          0.015625
#> 15: 2022-05-10 04:27:37        1  5.41928392  4.58679188          0.015625
#> 16: 2022-05-10 04:27:37        1  6.92216436  2.74746128          0.015625
#> 17: 2022-05-10 04:27:37        1  2.01694546 10.58984510          0.015625
#> 18: 2022-05-10 04:27:37        1  2.11987206 11.45980011          0.015625
#> 19: 2022-05-10 04:27:37        1  8.68530326  9.40487941          0.015625
#> 20: 2022-05-10 04:27:37        1  9.97224043 12.30643925          0.015625
#> 21: 2022-05-10 04:27:37        1 -4.70897578  6.01493103          0.015625
#> 22: 2022-05-10 04:27:37        1 -2.74402890 10.50328363          0.015625
#> 23: 2022-05-10 04:27:37        1  6.44417686 14.17548128          0.015625
#> 24: 2022-05-10 04:27:37        1  8.91629873 11.93276885          0.015625
#> 25: 2022-05-10 04:27:37        1  5.03203721  4.57562302          0.015625
#> 26: 2022-05-10 04:27:37        1  9.87115650 12.96158578          0.015625
#> 27: 2022-05-10 04:27:37        1  4.16896626 11.21438419          0.015625
#> 28: 2022-05-10 04:27:37        1  4.32872430  8.35207471          0.015625
#> 29: 2022-05-10 04:27:37        1 -4.08530477  1.45686109          0.015625
#> 30: 2022-05-10 04:27:37        1  2.52444387  1.02616570          0.015625
#> 31: 2022-05-10 04:27:37        1  4.55045617  0.11069856          0.015625
#> 32: 2022-05-10 04:27:37        1 -3.51428991 12.48280994          0.015625
#> 33: 2022-05-10 04:27:37        1 -0.74309650 14.66769678          0.015625
#> 34: 2022-05-10 04:27:37        1 -3.86316092  4.60545357          0.015625
#> 35: 2022-05-10 04:27:37        1 -3.42142521 12.10384208          0.015625
#> 36: 2022-05-10 04:27:37        1  4.98095461 13.13872170          0.015625
#> 37: 2022-05-10 04:27:37        1  1.29749357 11.88226326          0.015625
#> 38: 2022-05-10 04:27:37        1  2.82784355 14.68408567          0.015625
#> 39: 2022-05-10 04:27:37        1  9.74999447 12.59424873          0.015625
#> 40: 2022-05-10 04:27:37        1 -3.41268476 14.15758968          0.015625
#> 41: 2022-05-10 04:27:37        1  0.08983161 11.04522706          0.015625
#> 42: 2022-05-10 04:27:37        1  2.09824433  5.78602146          0.015625
#> 43: 2022-05-10 04:27:37        1  1.54989315  4.18735483          0.015625
#> 44: 2022-05-10 04:27:37        1  7.40942996  1.55101293          0.015625
#> 45: 2022-05-10 04:27:37        1 -3.49465540  3.04979705          0.015625
#> 46: 2022-05-10 04:27:37        1  1.32882627  1.09887398          0.015625
#> 47: 2022-05-10 04:27:37        1  2.62218272  4.11611426          0.015625
#> 48: 2022-05-10 04:27:37        1  5.66478974 11.57098771          0.015625
#> 49: 2022-05-10 04:27:37        1 -3.54470986 12.82784099          0.015625
#> 50: 2022-05-10 04:27:37        1  6.97894140  9.23158280          0.015625
#> 51: 2022-05-10 04:27:37        1  1.63992194 11.36231273          0.015625
#> 52: 2022-05-10 04:27:37        1  6.05305267  5.89931836          0.015625
#> 53: 2022-05-10 04:27:37        1  7.74504337 13.96981867          0.015625
#> 54: 2022-05-10 04:27:37        1 -3.45962326  9.57636627          0.015625
#> 55: 2022-05-10 04:27:37        1 -3.90295989  3.39898215          0.015625
#> 56: 2022-05-10 04:27:37        1  9.72250674  8.96284702          0.015625
#> 57: 2022-05-10 04:27:37        1  9.41733226 13.39111039          0.015625
#> 58: 2022-05-10 04:27:37        1  7.53415908  1.92534691          0.015625
#> 59: 2022-05-10 04:27:37        1  3.41872138 10.59515351          0.015625
#> 60: 2022-05-10 04:27:37        1  1.05399839  5.43700752          0.015625
#> 61: 2022-05-10 04:27:37        1 -0.49641909  4.01907836          0.015625
#> 62: 2022-05-10 04:27:37        1  6.49440046  0.65567103          0.015625
#> 63: 2022-05-10 04:27:37        1  6.63010869  1.24812582          0.015625
#> 64: 2022-05-10 04:27:37        1  6.06457599  7.55495725          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity