Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -2.807863 10.70351 0.015625 <list[3]> 0.9277217

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -2.807863 10.70351 0.015625 <list[3]> 0.9277217

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition           y
#>  1:  1.14916859  5.8896001 0.015625     0       6          1  16.7451596
#>  2:  5.69883594  9.8780051 0.015625     0       6          1 160.7975291
#>  3:  5.16145105  8.7256173 0.015625     0       6          1 116.6072405
#>  4:  6.93624303  1.5490320 0.015625     0       6          1  43.7294084
#>  5:  6.42525030 10.9677639 0.015625     0       6          1 213.3974316
#>  6:  0.88297256 13.7396145 0.015625     0       6          1  99.2874699
#>  7:  0.98657401 12.4343679 0.015625     0       6          1  78.8906540
#>  8: -1.83141085  4.5555926 0.015625     0       6          1  27.4380435
#>  9:  4.26045435  5.3806294 0.015625     0       6          1  37.2030990
#> 10:  1.79234435  2.2265884 0.015625     0       6          1   8.9295778
#> 11: -1.31791060  7.6864198 0.015625     0       6          1  12.6182025
#> 12:  9.59754121  2.2977670 0.015625     0       6          1  76.9377848
#> 13:  6.70012106 11.2269524 0.015625     0       6          1 229.3266947
#> 14: -2.28794268 10.6923860 0.015625     0       6          1   4.4813493
#> 15:  0.08277883  3.0169025 0.015625     0       6          1  27.7006365
#> 16:  0.79003423  2.2424938 0.015625     0       6          1  23.1051347
#> 17:  0.32271084  9.8890135 0.015625     0       6          1  38.4613554
#> 18:  1.97171543  9.1887244 0.015625     0       6          1  44.7828719
#> 19:  3.36637057 10.3970402 0.015625     0       6          1  89.1187935
#> 20:  5.52177824  7.3761018 0.015625     0       6          1 102.0846173
#> 21:  6.96573997  7.6843880 0.015625     0       6          1 144.6625711
#> 22:  2.54961016  1.9338844 0.015625     0       6          1   2.0751497
#> 23:  6.65779016  1.6067203 0.015625     0       6          1  42.3620733
#> 24:  3.28172341  8.6388614 0.015625     0       6          1  57.2042074
#> 25:  4.17377050  2.3819611 0.015625     0       6          1  11.2702054
#> 26:  9.77062117 11.0931967 0.015625     0       6          1 314.5524011
#> 27: -2.58671479 10.2223667 0.015625     0       6          1   1.8485922
#> 28: -4.58952426  4.5977839 0.015625     0       6          1  96.3264637
#> 29:  0.46270600  3.7700565 0.015625     0       6          1  20.8427510
#> 30: -1.05912914  6.6391258 0.015625     0       6          1  15.8701046
#> 31: -3.62507868  6.4453922 0.015625     0       6          1  34.3102112
#> 32:  6.29520311  5.8282446 0.015625     0       6          1  94.0589963
#> 33: -0.63188912 10.6142959 0.015625     0       6          1  30.6817058
#> 34: -1.34418962 12.7660192 0.015625     0       6          1  33.0524659
#> 35:  0.73223467  0.6695104 0.015625     0       6          1  34.6265954
#> 36:  9.23888556  8.7670738 0.015625     0       6          1 220.9891291
#> 37: -1.66012890 11.7652853 0.015625     0       6          1  18.3750421
#> 38: -2.33791364  9.0358697 0.015625     0       6          1   4.0632993
#> 39: -2.80786339 10.7035115 0.015625     0       6          1   0.9277217
#> 40:  6.68742684  4.1363396 0.015625     0       6          1  73.6570658
#> 41:  1.47530025  2.0242738 0.015625     0       6          1  13.7873558
#> 42: -1.73564868 10.2278844 0.015625     0       6          1  10.3090535
#> 43: -2.40219546 12.6656481 0.015625     0       6          1  10.0075098
#> 44: -3.11864664 12.4777620 0.015625     0       6          1   1.8772177
#> 45:  3.61403220  8.2846356 0.015625     0       6          1  59.7421995
#> 46:  1.38063000  9.2649745 0.015625     0       6          1  41.0150655
#> 47:  3.37303562  1.8010150 0.015625     0       6          1   1.3255727
#> 48: -4.03881523  3.1094895 0.015625     0       6          1 100.4433653
#> 49: -3.36128549 13.8376271 0.015625     0       6          1   5.2107965
#> 50:  0.94659333 11.0079069 0.015625     0       6          1  57.6919179
#> 51: -2.52008752  4.5719803 0.015625     0       6          1  33.9369075
#> 52:  7.36667920  6.7926989 0.015625     0       6          1 132.1872875
#> 53:  9.79795238 11.3727228 0.015625     0       6          1 325.5967976
#> 54:  1.50692931 11.9020042 0.015625     0       6          1  78.3545524
#> 55:  8.75032794  3.8439372 0.015625     0       6          1  91.1664488
#> 56: -4.57996331  6.1162160 0.015625     0       6          1  69.8528518
#> 57:  0.52770305  8.1346278 0.015625     0       6          1  27.0926460
#> 58: -2.97462931  8.7969816 0.015625     0       6          1   5.4126438
#> 59: -2.31548158 11.0290401 0.015625     0       6          1   4.8822730
#> 60: -1.39146396  8.1589316 0.015625     0       6          1  11.7260219
#> 61: -2.53895878  8.9790490 0.015625     0       6          1   3.6770284
#> 62:  9.85180180  0.6550869 0.015625     0       6          1  55.2895916
#> 63:  0.62827240  8.6526524 0.015625     0       6          1  31.0213531
#> 64:  8.49866920  9.7310856 0.015625     0       6          1 230.3218019
#>              x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-07 04:32:22        1  1.14916859   5.8896001          0.015625
#>  2: 2022-05-07 04:32:22        1  5.69883594   9.8780051          0.015625
#>  3: 2022-05-07 04:32:22        1  5.16145105   8.7256173          0.015625
#>  4: 2022-05-07 04:32:22        1  6.93624303   1.5490320          0.015625
#>  5: 2022-05-07 04:32:22        1  6.42525030  10.9677639          0.015625
#>  6: 2022-05-07 04:32:22        1  0.88297256  13.7396145          0.015625
#>  7: 2022-05-07 04:32:22        1  0.98657401  12.4343679          0.015625
#>  8: 2022-05-07 04:32:22        1 -1.83141085   4.5555926          0.015625
#>  9: 2022-05-07 04:32:22        1  4.26045435   5.3806294          0.015625
#> 10: 2022-05-07 04:32:22        1  1.79234435   2.2265884          0.015625
#> 11: 2022-05-07 04:32:22        1 -1.31791060   7.6864198          0.015625
#> 12: 2022-05-07 04:32:22        1  9.59754121   2.2977670          0.015625
#> 13: 2022-05-07 04:32:22        1  6.70012106  11.2269524          0.015625
#> 14: 2022-05-07 04:32:22        1 -2.28794268  10.6923860          0.015625
#> 15: 2022-05-07 04:32:22        1  0.08277883   3.0169025          0.015625
#> 16: 2022-05-07 04:32:22        1  0.79003423   2.2424938          0.015625
#> 17: 2022-05-07 04:32:22        1  0.32271084   9.8890135          0.015625
#> 18: 2022-05-07 04:32:22        1  1.97171543   9.1887244          0.015625
#> 19: 2022-05-07 04:32:22        1  3.36637057  10.3970402          0.015625
#> 20: 2022-05-07 04:32:22        1  5.52177824   7.3761018          0.015625
#> 21: 2022-05-07 04:32:22        1  6.96573997   7.6843880          0.015625
#> 22: 2022-05-07 04:32:22        1  2.54961016   1.9338844          0.015625
#> 23: 2022-05-07 04:32:22        1  6.65779016   1.6067203          0.015625
#> 24: 2022-05-07 04:32:22        1  3.28172341   8.6388614          0.015625
#> 25: 2022-05-07 04:32:22        1  4.17377050   2.3819611          0.015625
#> 26: 2022-05-07 04:32:22        1  9.77062117  11.0931967          0.015625
#> 27: 2022-05-07 04:32:22        1 -2.58671479  10.2223667          0.015625
#> 28: 2022-05-07 04:32:22        1 -4.58952426   4.5977839          0.015625
#> 29: 2022-05-07 04:32:22        1  0.46270600   3.7700565          0.015625
#> 30: 2022-05-07 04:32:22        1 -1.05912914   6.6391258          0.015625
#> 31: 2022-05-07 04:32:22        1 -3.62507868   6.4453922          0.015625
#> 32: 2022-05-07 04:32:22        1  6.29520311   5.8282446          0.015625
#> 33: 2022-05-07 04:32:22        1 -0.63188912  10.6142959          0.015625
#> 34: 2022-05-07 04:32:22        1 -1.34418962  12.7660192          0.015625
#> 35: 2022-05-07 04:32:22        1  0.73223467   0.6695104          0.015625
#> 36: 2022-05-07 04:32:22        1  9.23888556   8.7670738          0.015625
#> 37: 2022-05-07 04:32:22        1 -1.66012890  11.7652853          0.015625
#> 38: 2022-05-07 04:32:22        1 -2.33791364   9.0358697          0.015625
#> 39: 2022-05-07 04:32:22        1 -2.80786339  10.7035115          0.015625
#> 40: 2022-05-07 04:32:22        1  6.68742684   4.1363396          0.015625
#> 41: 2022-05-07 04:32:22        1  1.47530025   2.0242738          0.015625
#> 42: 2022-05-07 04:32:22        1 -1.73564868  10.2278844          0.015625
#> 43: 2022-05-07 04:32:22        1 -2.40219546  12.6656481          0.015625
#> 44: 2022-05-07 04:32:22        1 -3.11864664  12.4777620          0.015625
#> 45: 2022-05-07 04:32:22        1  3.61403220   8.2846356          0.015625
#> 46: 2022-05-07 04:32:22        1  1.38063000   9.2649745          0.015625
#> 47: 2022-05-07 04:32:22        1  3.37303562   1.8010150          0.015625
#> 48: 2022-05-07 04:32:22        1 -4.03881523   3.1094895          0.015625
#> 49: 2022-05-07 04:32:22        1 -3.36128549  13.8376271          0.015625
#> 50: 2022-05-07 04:32:22        1  0.94659333  11.0079069          0.015625
#> 51: 2022-05-07 04:32:22        1 -2.52008752   4.5719803          0.015625
#> 52: 2022-05-07 04:32:22        1  7.36667920   6.7926989          0.015625
#> 53: 2022-05-07 04:32:22        1  9.79795238  11.3727228          0.015625
#> 54: 2022-05-07 04:32:22        1  1.50692931  11.9020042          0.015625
#> 55: 2022-05-07 04:32:22        1  8.75032794   3.8439372          0.015625
#> 56: 2022-05-07 04:32:22        1 -4.57996331   6.1162160          0.015625
#> 57: 2022-05-07 04:32:22        1  0.52770305   8.1346278          0.015625
#> 58: 2022-05-07 04:32:22        1 -2.97462931   8.7969816          0.015625
#> 59: 2022-05-07 04:32:22        1 -2.31548158  11.0290401          0.015625
#> 60: 2022-05-07 04:32:22        1 -1.39146396   8.1589316          0.015625
#> 61: 2022-05-07 04:32:22        1 -2.53895878   8.9790490          0.015625
#> 62: 2022-05-07 04:32:22        1  9.85180180   0.6550869          0.015625
#> 63: 2022-05-07 04:32:22        1  0.62827240   8.6526524          0.015625
#> 64: 2022-05-07 04:32:22        1  8.49866920   9.7310856          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity