Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -2.923452 11.29805 0.015625 <list[3]> 0.7717092

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -2.923452 11.29805 0.015625 <list[3]> 0.7717092

# all evaluations
as.data.table(instance$archive)
#>               x1         x2 fidelity stage bracket repetition           y
#>  1:  2.039940134  9.2786408 0.015625     0       6          1  46.5848447
#>  2:  6.124853663  1.9757304 0.015625     0       6          1  40.3696453
#>  3:  8.863513325  9.8192774 0.015625     0       6          1 242.4452174
#>  4:  1.217712816  4.2719314 0.015625     0       6          1  13.3473646
#>  5:  8.994867733  0.0569013 0.015625     0       6          1  35.9049896
#>  6:  1.569116123  0.6269176 0.015625     0       6          1  18.7272471
#>  7:  2.688131492 12.8476871 0.015625     0       6          1 120.2612068
#>  8: -2.966484706 12.3866265 0.015625     0       6          1   2.4900385
#>  9: -1.533287788  9.7918331 0.015625     0       6          1  11.9965431
#> 10: -4.352627320  0.7642765 0.015625     0       6          1 169.0713713
#> 11:  3.763342333  7.6127199 0.015625     0       6          1  53.5578890
#> 12:  4.578409299  9.3994689 0.015625     0       6          1 109.5534977
#> 13:  2.078225702 12.9976804 0.015625     0       6          1 108.8134214
#> 14:  6.079913240  7.7652423 0.015625     0       6          1 125.6005545
#> 15: -2.949771505  5.2826252 0.015625     0       6          1  32.8320453
#> 16:  1.773970486  5.6573427 0.015625     0       6          1  13.7456843
#> 17:  0.136509221 14.8955310 0.015625     0       6          1 102.5453237
#> 18:  9.644951686  0.3374017 0.015625     0       6          1  47.2454763
#> 19:  2.739284895 14.6849221 0.015625     0       6          1 165.3609044
#> 20: -3.447756522  0.9728996 0.015625     0       6          1 119.2158953
#> 21:  0.485427105  9.9562174 0.015625     0       6          1  40.7859028
#> 22:  6.356287100  6.2219919 0.015625     0       6          1 102.3152214
#> 23: -1.724954924  5.6287566 0.015625     0       6          1  18.8173931
#> 24: -2.482988511  9.0101569 0.015625     0       6          1   3.6858190
#> 25:  5.320939819 11.2381693 0.015625     0       6          1 180.2570069
#> 26:  1.226482396  2.8044209 0.015625     0       6          1  14.9048462
#> 27: -2.778486772 14.9807985 0.015625     0       6          1  19.6978802
#> 28:  9.351227116 11.2184241 0.015625     0       6          1 303.6252134
#> 29:  8.286428809  8.6680781 0.015625     0       6          1 194.9035541
#> 30: -0.426983187 12.3073239 0.015625     0       6          1  50.3486368
#> 31: -4.931905614  4.9917683 0.015625     0       6          1 104.3562608
#> 32: -3.119368248 12.5841525 0.015625     0       6          1   2.1435598
#> 33:  7.330248547  4.4916607 0.015625     0       6          1  87.1541848
#> 34:  4.586785858  5.6704078 0.015625     0       6          1  48.7854102
#> 35: -2.051655609  2.6176150 0.015625     0       6          1  51.4879753
#> 36:  7.494597733  9.9933279 0.015625     0       6          1 214.8558169
#> 37: -1.797304215 12.9606479 0.015625     0       6          1  23.8502152
#> 38:  8.992484087  6.9769806 0.015625     0       6          1 165.1881315
#> 39: -0.203405650  9.3395719 0.015625     0       6          1  28.4917881
#> 40: -1.655810188  6.6746348 0.015625     0       6          1  13.3665656
#> 41:  3.177165659  6.0776243 0.015625     0       6          1  23.6736687
#> 42:  3.365735081 11.0960184 0.015625     0       6          1 102.7378869
#> 43: -2.923452273 11.2980490 0.015625     0       6          1   0.7717092
#> 44:  1.302903797  5.6719349 0.015625     0       6          1  15.4091982
#> 45:  4.999127435 10.5106161 0.015625     0       6          1 149.5722047
#> 46: -3.787952693  8.5889603 0.015625     0       6          1  17.3964278
#> 47:  9.825028732 10.4776313 0.015625     0       6          1 295.1628632
#> 48: -3.531477429  7.7875936 0.015625     0       6          1  18.8963880
#> 49:  8.306757610  0.2095554 0.015625     0       6          1  33.9802234
#> 50:  5.871333058 13.8124332 0.015625     0       6          1 277.9133145
#> 51: -0.004061718  4.9618731 0.015625     0       6          1  20.6932055
#> 52: -4.432326789  9.5963422 0.015625     0       6          1  23.8453959
#> 53:  4.827060300  2.0525831 0.015625     0       6          1  20.2110269
#> 54:  5.566728620  8.9962928 0.015625     0       6          1 136.1213844
#> 55:  1.277374759  1.5753156 0.015625     0       6          1  18.7398375
#> 56: -4.928842797 10.5492901 0.015625     0       6          1  28.4012721
#> 57: -4.425955509 12.2146306 0.015625     0       6          1   9.3371452
#> 58: -2.339719825 10.4149875 0.015625     0       6          1   3.5964694
#> 59:  9.226561034 10.8130540 0.015625     0       6          1 285.5239045
#> 60:  9.698159418 10.6552210 0.015625     0       6          1 296.5405678
#> 61:  3.701790411  1.6321595 0.015625     0       6          1   3.0808729
#> 62:  9.358601399  5.9777642 0.015625     0       6          1 148.7594360
#> 63: -4.222450168  3.5084721 0.015625     0       6          1 100.7382787
#> 64: -0.092410777  2.1645292 0.015625     0       6          1  35.4239150
#>               x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-04-30 04:34:33        1  2.039940134   9.2786408          0.015625
#>  2: 2022-04-30 04:34:33        1  6.124853663   1.9757304          0.015625
#>  3: 2022-04-30 04:34:33        1  8.863513325   9.8192774          0.015625
#>  4: 2022-04-30 04:34:33        1  1.217712816   4.2719314          0.015625
#>  5: 2022-04-30 04:34:33        1  8.994867733   0.0569013          0.015625
#>  6: 2022-04-30 04:34:33        1  1.569116123   0.6269176          0.015625
#>  7: 2022-04-30 04:34:33        1  2.688131492  12.8476871          0.015625
#>  8: 2022-04-30 04:34:33        1 -2.966484706  12.3866265          0.015625
#>  9: 2022-04-30 04:34:33        1 -1.533287788   9.7918331          0.015625
#> 10: 2022-04-30 04:34:33        1 -4.352627320   0.7642765          0.015625
#> 11: 2022-04-30 04:34:33        1  3.763342333   7.6127199          0.015625
#> 12: 2022-04-30 04:34:33        1  4.578409299   9.3994689          0.015625
#> 13: 2022-04-30 04:34:33        1  2.078225702  12.9976804          0.015625
#> 14: 2022-04-30 04:34:33        1  6.079913240   7.7652423          0.015625
#> 15: 2022-04-30 04:34:33        1 -2.949771505   5.2826252          0.015625
#> 16: 2022-04-30 04:34:33        1  1.773970486   5.6573427          0.015625
#> 17: 2022-04-30 04:34:33        1  0.136509221  14.8955310          0.015625
#> 18: 2022-04-30 04:34:33        1  9.644951686   0.3374017          0.015625
#> 19: 2022-04-30 04:34:33        1  2.739284895  14.6849221          0.015625
#> 20: 2022-04-30 04:34:33        1 -3.447756522   0.9728996          0.015625
#> 21: 2022-04-30 04:34:33        1  0.485427105   9.9562174          0.015625
#> 22: 2022-04-30 04:34:33        1  6.356287100   6.2219919          0.015625
#> 23: 2022-04-30 04:34:33        1 -1.724954924   5.6287566          0.015625
#> 24: 2022-04-30 04:34:33        1 -2.482988511   9.0101569          0.015625
#> 25: 2022-04-30 04:34:33        1  5.320939819  11.2381693          0.015625
#> 26: 2022-04-30 04:34:33        1  1.226482396   2.8044209          0.015625
#> 27: 2022-04-30 04:34:33        1 -2.778486772  14.9807985          0.015625
#> 28: 2022-04-30 04:34:33        1  9.351227116  11.2184241          0.015625
#> 29: 2022-04-30 04:34:33        1  8.286428809   8.6680781          0.015625
#> 30: 2022-04-30 04:34:33        1 -0.426983187  12.3073239          0.015625
#> 31: 2022-04-30 04:34:33        1 -4.931905614   4.9917683          0.015625
#> 32: 2022-04-30 04:34:33        1 -3.119368248  12.5841525          0.015625
#> 33: 2022-04-30 04:34:33        1  7.330248547   4.4916607          0.015625
#> 34: 2022-04-30 04:34:33        1  4.586785858   5.6704078          0.015625
#> 35: 2022-04-30 04:34:33        1 -2.051655609   2.6176150          0.015625
#> 36: 2022-04-30 04:34:33        1  7.494597733   9.9933279          0.015625
#> 37: 2022-04-30 04:34:33        1 -1.797304215  12.9606479          0.015625
#> 38: 2022-04-30 04:34:33        1  8.992484087   6.9769806          0.015625
#> 39: 2022-04-30 04:34:33        1 -0.203405650   9.3395719          0.015625
#> 40: 2022-04-30 04:34:33        1 -1.655810188   6.6746348          0.015625
#> 41: 2022-04-30 04:34:33        1  3.177165659   6.0776243          0.015625
#> 42: 2022-04-30 04:34:33        1  3.365735081  11.0960184          0.015625
#> 43: 2022-04-30 04:34:33        1 -2.923452273  11.2980490          0.015625
#> 44: 2022-04-30 04:34:33        1  1.302903797   5.6719349          0.015625
#> 45: 2022-04-30 04:34:33        1  4.999127435  10.5106161          0.015625
#> 46: 2022-04-30 04:34:33        1 -3.787952693   8.5889603          0.015625
#> 47: 2022-04-30 04:34:33        1  9.825028732  10.4776313          0.015625
#> 48: 2022-04-30 04:34:33        1 -3.531477429   7.7875936          0.015625
#> 49: 2022-04-30 04:34:33        1  8.306757610   0.2095554          0.015625
#> 50: 2022-04-30 04:34:33        1  5.871333058  13.8124332          0.015625
#> 51: 2022-04-30 04:34:33        1 -0.004061718   4.9618731          0.015625
#> 52: 2022-04-30 04:34:33        1 -4.432326789   9.5963422          0.015625
#> 53: 2022-04-30 04:34:33        1  4.827060300   2.0525831          0.015625
#> 54: 2022-04-30 04:34:33        1  5.566728620   8.9962928          0.015625
#> 55: 2022-04-30 04:34:33        1  1.277374759   1.5753156          0.015625
#> 56: 2022-04-30 04:34:33        1 -4.928842797  10.5492901          0.015625
#> 57: 2022-04-30 04:34:33        1 -4.425955509  12.2146306          0.015625
#> 58: 2022-04-30 04:34:33        1 -2.339719825  10.4149875          0.015625
#> 59: 2022-04-30 04:34:33        1  9.226561034  10.8130540          0.015625
#> 60: 2022-04-30 04:34:33        1  9.698159418  10.6552210          0.015625
#> 61: 2022-04-30 04:34:33        1  3.701790411   1.6321595          0.015625
#> 62: 2022-04-30 04:34:33        1  9.358601399   5.9777642          0.015625
#> 63: 2022-04-30 04:34:33        1 -4.222450168   3.5084721          0.015625
#> 64: 2022-04-30 04:34:33        1 -0.092410777   2.1645292          0.015625
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity