Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.048463 10.92908 0.015625 <list[3]> 0.4829426

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.048463 10.92908 0.015625 <list[3]> 0.4829426

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -2.4442586  6.3645731 0.015625     0       6          1  16.3981933
#>  2: -4.9001474 13.6309025 0.015625     0       6          1  12.6135095
#>  3:  7.5727968 14.5134912 0.015625     0       6          1 366.2066280
#>  4:  6.8901521 13.4141524 0.015625     0       6          1 304.1896614
#>  5: -4.9621651  3.7406896 0.015625     0       6          1 131.4873327
#>  6: -2.7678442  2.3803842 0.015625     0       6          1  69.2937914
#>  7:  7.5568907  2.5365739 0.015625     0       6          1  59.1581780
#>  8: -2.7993991  2.3544679 0.015625     0       6          1  70.5413575
#>  9:  4.7369446  3.1918784 0.015625     0       6          1  26.5657558
#> 10:  5.2986995 12.9811640 0.015625     0       6          1 227.0457213
#> 11:  1.7020171  3.2594906 0.015625     0       6          1   8.7581931
#> 12:  6.1949345  0.4416098 0.015625     0       6          1  29.3064460
#> 13:  6.4310676  1.3259463 0.015625     0       6          1  37.8984406
#> 14: -2.7684952  8.6898625 0.015625     0       6          1   4.8687778
#> 15:  5.5924358 11.3648718 0.015625     0       6          1 194.3944088
#> 16: -0.2014717  9.6323262 0.015625     0       6          1  30.3668124
#> 17:  4.0510635  2.8846405 0.015625     0       6          1  12.0975772
#> 18:  0.5271259  2.9982300 0.015625     0       6          1  23.0135198
#> 19: -1.2387365  5.4812299 0.015625     0       6          1  19.5689098
#> 20:  8.0986366 13.1312650 0.015625     0       6          1 331.8186536
#> 21: -4.2354279  4.3712040 0.015625     0       6          1  85.1806357
#> 22:  4.3922610  9.0681066 0.015625     0       6          1  96.5728307
#> 23:  2.6830964  0.3361828 0.015625     0       6          1   3.9974549
#> 24:  5.3299594 12.2140406 0.015625     0       6          1 206.6481830
#> 25:  0.8030018  5.1859247 0.015625     0       6          1  16.8663883
#> 26:  6.3246208 14.1687789 0.015625     0       6          1 308.7577643
#> 27: -3.0726071  0.3532605 0.015625     0       6          1 117.6495365
#> 28: -3.7948386 12.1671743 0.015625     0       6          1   2.4742182
#> 29:  6.9270214  3.2106655 0.015625     0       6          1  63.3774878
#> 30:  4.2658158  9.4921115 0.015625     0       6          1 100.3676136
#> 31: -1.1745151  3.0774681 0.015625     0       6          1  37.0762559
#> 32:  3.8526027  5.2250861 0.015625     0       6          1  26.7376811
#> 33:  1.6363259  1.4766646 0.015625     0       6          1  13.3767048
#> 34:  1.4931959  6.8405202 0.015625     0       6          1  20.6571854
#> 35:  9.9824350  2.0612925 0.015625     0       6          1  80.7944760
#> 36: -3.0484633 10.9290803 0.015625     0       6          1   0.4829426
#> 37:  4.5310907  7.5315155 0.015625     0       6          1  74.0685155
#> 38:  9.0484083  2.9764245 0.015625     0       6          1  79.5703378
#> 39:  0.5200484  7.7663039 0.015625     0       6          1  25.0183617
#> 40:  0.8362438  6.5043502 0.015625     0       6          1  19.7256344
#> 41:  5.7833704  6.7084993 0.015625     0       6          1  97.3638386
#> 42: -2.8405316  5.3639916 0.015625     0       6          1  30.0431521
#> 43:  4.3201184  1.8102336 0.015625     0       6          1  10.7900769
#> 44:  1.5851930  9.9606628 0.015625     0       6          1  50.9026198
#> 45:  1.8209973  8.4504890 0.015625     0       6          1  35.1507905
#> 46:  9.3386487 10.5699973 0.015625     0       6          1 281.0449149
#> 47: -0.9247162  5.9553208 0.015625     0       6          1  18.1609908
#> 48: -2.7926034 14.4664797 0.015625     0       6          1  15.2812115
#> 49: -2.2174583  3.6814332 0.015625     0       6          1  40.2018575
#> 50:  3.8527165 13.1423771 0.015625     0       6          1 167.0205190
#> 51:  0.4080233 12.5213785 0.015625     0       6          1  70.1603703
#> 52:  5.5722880  8.4308227 0.015625     0       6          1 124.2888225
#> 53:  9.8124002 11.9788782 0.015625     0       6          1 348.3875665
#> 54:  0.4764284 12.7935901 0.015625     0       6          1  75.4578757
#> 55:  1.7872069  7.7662008 0.015625     0       6          1  28.3001149
#> 56:  3.9456028  8.2507065 0.015625     0       6          1  68.1671390
#> 57:  0.9377884  4.4492379 0.015625     0       6          1  15.6876170
#> 58:  1.3711211  2.9807325 0.015625     0       6          1  12.7053703
#> 59:  6.6242287 10.7349958 0.015625     0       6          1 213.0547634
#> 60:  1.9740884  8.2176534 0.015625     0       6          1  33.6860260
#> 61:  5.9026232  5.9524285 0.015625     0       6          1  87.3988699
#> 62:  7.5391198  6.4108248 0.015625     0       6          1 126.6539670
#> 63:  9.7952241 13.5565015 0.015625     0       6          1 408.9284419
#> 64:  2.9711019  1.5438588 0.015625     0       6          1   0.5371036
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-19 04:40:01        1  -2.4442586   6.3645731          0.015625
#>  2: 2022-05-19 04:40:01        1  -4.9001474  13.6309025          0.015625
#>  3: 2022-05-19 04:40:01        1   7.5727968  14.5134912          0.015625
#>  4: 2022-05-19 04:40:01        1   6.8901521  13.4141524          0.015625
#>  5: 2022-05-19 04:40:01        1  -4.9621651   3.7406896          0.015625
#>  6: 2022-05-19 04:40:01        1  -2.7678442   2.3803842          0.015625
#>  7: 2022-05-19 04:40:01        1   7.5568907   2.5365739          0.015625
#>  8: 2022-05-19 04:40:01        1  -2.7993991   2.3544679          0.015625
#>  9: 2022-05-19 04:40:01        1   4.7369446   3.1918784          0.015625
#> 10: 2022-05-19 04:40:01        1   5.2986995  12.9811640          0.015625
#> 11: 2022-05-19 04:40:01        1   1.7020171   3.2594906          0.015625
#> 12: 2022-05-19 04:40:01        1   6.1949345   0.4416098          0.015625
#> 13: 2022-05-19 04:40:01        1   6.4310676   1.3259463          0.015625
#> 14: 2022-05-19 04:40:01        1  -2.7684952   8.6898625          0.015625
#> 15: 2022-05-19 04:40:01        1   5.5924358  11.3648718          0.015625
#> 16: 2022-05-19 04:40:01        1  -0.2014717   9.6323262          0.015625
#> 17: 2022-05-19 04:40:01        1   4.0510635   2.8846405          0.015625
#> 18: 2022-05-19 04:40:01        1   0.5271259   2.9982300          0.015625
#> 19: 2022-05-19 04:40:01        1  -1.2387365   5.4812299          0.015625
#> 20: 2022-05-19 04:40:01        1   8.0986366  13.1312650          0.015625
#> 21: 2022-05-19 04:40:01        1  -4.2354279   4.3712040          0.015625
#> 22: 2022-05-19 04:40:01        1   4.3922610   9.0681066          0.015625
#> 23: 2022-05-19 04:40:01        1   2.6830964   0.3361828          0.015625
#> 24: 2022-05-19 04:40:01        1   5.3299594  12.2140406          0.015625
#> 25: 2022-05-19 04:40:01        1   0.8030018   5.1859247          0.015625
#> 26: 2022-05-19 04:40:01        1   6.3246208  14.1687789          0.015625
#> 27: 2022-05-19 04:40:01        1  -3.0726071   0.3532605          0.015625
#> 28: 2022-05-19 04:40:01        1  -3.7948386  12.1671743          0.015625
#> 29: 2022-05-19 04:40:01        1   6.9270214   3.2106655          0.015625
#> 30: 2022-05-19 04:40:01        1   4.2658158   9.4921115          0.015625
#> 31: 2022-05-19 04:40:01        1  -1.1745151   3.0774681          0.015625
#> 32: 2022-05-19 04:40:01        1   3.8526027   5.2250861          0.015625
#> 33: 2022-05-19 04:40:01        1   1.6363259   1.4766646          0.015625
#> 34: 2022-05-19 04:40:01        1   1.4931959   6.8405202          0.015625
#> 35: 2022-05-19 04:40:01        1   9.9824350   2.0612925          0.015625
#> 36: 2022-05-19 04:40:01        1  -3.0484633  10.9290803          0.015625
#> 37: 2022-05-19 04:40:01        1   4.5310907   7.5315155          0.015625
#> 38: 2022-05-19 04:40:01        1   9.0484083   2.9764245          0.015625
#> 39: 2022-05-19 04:40:01        1   0.5200484   7.7663039          0.015625
#> 40: 2022-05-19 04:40:01        1   0.8362438   6.5043502          0.015625
#> 41: 2022-05-19 04:40:01        1   5.7833704   6.7084993          0.015625
#> 42: 2022-05-19 04:40:01        1  -2.8405316   5.3639916          0.015625
#> 43: 2022-05-19 04:40:01        1   4.3201184   1.8102336          0.015625
#> 44: 2022-05-19 04:40:01        1   1.5851930   9.9606628          0.015625
#> 45: 2022-05-19 04:40:01        1   1.8209973   8.4504890          0.015625
#> 46: 2022-05-19 04:40:01        1   9.3386487  10.5699973          0.015625
#> 47: 2022-05-19 04:40:01        1  -0.9247162   5.9553208          0.015625
#> 48: 2022-05-19 04:40:01        1  -2.7926034  14.4664797          0.015625
#> 49: 2022-05-19 04:40:01        1  -2.2174583   3.6814332          0.015625
#> 50: 2022-05-19 04:40:01        1   3.8527165  13.1423771          0.015625
#> 51: 2022-05-19 04:40:01        1   0.4080233  12.5213785          0.015625
#> 52: 2022-05-19 04:40:01        1   5.5722880   8.4308227          0.015625
#> 53: 2022-05-19 04:40:01        1   9.8124002  11.9788782          0.015625
#> 54: 2022-05-19 04:40:01        1   0.4764284  12.7935901          0.015625
#> 55: 2022-05-19 04:40:01        1   1.7872069   7.7662008          0.015625
#> 56: 2022-05-19 04:40:01        1   3.9456028   8.2507065          0.015625
#> 57: 2022-05-19 04:40:01        1   0.9377884   4.4492379          0.015625
#> 58: 2022-05-19 04:40:01        1   1.3711211   2.9807325          0.015625
#> 59: 2022-05-19 04:40:01        1   6.6242287  10.7349958          0.015625
#> 60: 2022-05-19 04:40:01        1   1.9740884   8.2176534          0.015625
#> 61: 2022-05-19 04:40:01        1   5.9026232   5.9524285          0.015625
#> 62: 2022-05-19 04:40:01        1   7.5391198   6.4108248          0.015625
#> 63: 2022-05-19 04:40:01        1   9.7952241  13.5565015          0.015625
#> 64: 2022-05-19 04:40:01        1   2.9711019   1.5438588          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity