Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain         y
#> 1: 2.992037 0.9905126 0.015625 <list[3]> 0.7783604

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain         y
#> 1: 2.992037 0.9905126 0.015625 <list[3]> 0.7783604

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -2.0885418  8.0043073 0.015625     0       6          1   7.3613182
#>  2:  6.8911972  2.1704075 0.015625     0       6          1  50.1204577
#>  3: -1.7499999  2.7605744 0.015625     0       6          1  45.7281723
#>  4:  4.9178340  3.9738447 0.015625     0       6          1  37.5341450
#>  5: -1.6726606  3.1734883 0.015625     0       6          1  40.1003846
#>  6:  7.1777705  9.5262479 0.015625     0       6          1 194.6574039
#>  7:  1.2787814  1.0494746 0.015625     0       6          1  21.5588352
#>  8:  3.1603589 10.5432008 0.015625     0       6          1  86.2577982
#>  9:  0.5255736  3.4627602 0.015625     0       6          1  21.2277344
#> 10:  1.1414949  8.8654616 0.015625     0       6          1  35.5462389
#> 11: -1.5428256  7.7510828 0.015625     0       6          1  10.8731858
#> 12: -4.8334412 12.7096181 0.015625     0       6          1  14.0541427
#> 13:  2.9920365  0.9905126 0.015625     0       6          1   0.7783604
#> 14: -0.7405201  0.5300015 0.015625     0       6          1  61.5154946
#> 15: -4.5119861  3.2368684 0.015625     0       6          1 119.8161864
#> 16:  2.7356623  7.6845264 0.015625     0       6          1  34.9152711
#> 17:  4.5627892  5.7303329 0.015625     0       6          1  48.9182317
#> 18:  5.7691745  0.9387208 0.015625     0       6          1  27.9544843
#> 19:  6.3587466  6.6989527 0.015625     0       6          1 111.2744539
#> 20:  2.3166716  8.0909588 0.015625     0       6          1  34.9900578
#> 21:  9.5585616  1.6682523 0.015625     0       6          1  65.6399260
#> 22:  1.2086236  8.1560971 0.015625     0       6          1  29.6814385
#> 23: -4.8487271  3.4190529 0.015625     0       6          1 132.7631772
#> 24:  3.9952472  3.3380021 0.015625     0       6          1  13.9666576
#> 25:  2.2478368  0.3514575 0.015625     0       6          1   8.9410458
#> 26: -0.7473669  7.5659091 0.015625     0       6          1  17.1720388
#> 27:  9.7946614  5.5905394 0.015625     0       6          1 150.6079735
#> 28:  6.8284949 13.5437045 0.015625     0       6          1 306.4589020
#> 29:  5.5123989 10.7272031 0.015625     0       6          1 174.7966555
#> 30:  4.7693918  2.1571098 0.015625     0       6          1  19.8399609
#> 31: -1.8774522  9.2545692 0.015625     0       6          1   7.1263950
#> 32: -2.8002691 12.2931825 0.015625     0       6          1   3.4968331
#> 33:  9.0160273  7.8848350 0.015625     0       6          1 189.8364669
#> 34:  8.3504398 11.9459183 0.015625     0       6          1 297.5652980
#> 35: -1.7574504 13.3701809 0.015625     0       6          1  28.2719020
#> 36:  7.0683769 10.7431211 0.015625     0       6          1 225.7847440
#> 37: -1.3159400  0.3444959 0.015625     0       6          1  73.3096264
#> 38: -4.7896197  3.9952436 0.015625     0       6          1 117.5122114
#> 39: -0.5018152  1.0267093 0.015625     0       6          1  51.8231738
#> 40:  8.4002642  0.5309106 0.015625     0       6          1  37.8524968
#> 41:  3.8879695 11.1582305 0.015625     0       6          1 121.3542409
#> 42: -3.9806784  5.5879595 0.015625     0       6          1  55.9252664
#> 43: -3.5317214  3.9737163 0.015625     0       6          1  65.6115629
#> 44:  4.2322791  1.5193774 0.015625     0       6          1   8.4703723
#> 45: -3.2501778  4.6173523 0.015625     0       6          1  47.7925166
#> 46: -3.9310964 11.6612329 0.015625     0       6          1   4.3840791
#> 47:  3.0592177  4.8208197 0.015625     0       6          1  12.0037814
#> 48:  5.4040842 14.6947452 0.015625     0       6          1 285.0083205
#> 49:  9.4498884 11.1553855 0.015625     0       6          1 304.8905037
#> 50:  3.4835116  7.1250455 0.015625     0       6          1  40.5948009
#> 51:  7.6805065  3.5112231 0.015625     0       6          1  74.4052862
#> 52: -4.5884889  1.0621124 0.015625     0       6          1 174.9151487
#> 53: -3.9492952  0.5932226 0.015625     0       6          1 151.5169648
#> 54:  1.4747830  1.1021254 0.015625     0       6          1  17.7721295
#> 55:  8.9930606  6.9792611 0.015625     0       6          1 165.2595460
#> 56:  4.5141508  6.0536168 0.015625     0       6          1  51.8216898
#> 57:  8.5338942  0.8609467 0.015625     0       6          1  42.4505168
#> 58: -4.7918275  6.9746936 0.015625     0       6          1  64.8982759
#> 59:  1.6856692 13.6881104 0.015625     0       6          1 114.6511963
#> 60:  5.3593765  2.3640355 0.015625     0       6          1  31.8729611
#> 61:  7.4419881  0.5397292 0.015625     0       6          1  35.7582402
#> 62: -3.6300739 12.5911696 0.015625     0       6          1   1.6878089
#> 63: -0.3476950  9.1650010 0.015625     0       6          1  25.8287234
#> 64:  3.8636808  3.4950723 0.015625     0       6          1  12.9405706
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-31 04:37:53        1  -2.0885418   8.0043073          0.015625
#>  2: 2022-05-31 04:37:53        1   6.8911972   2.1704075          0.015625
#>  3: 2022-05-31 04:37:53        1  -1.7499999   2.7605744          0.015625
#>  4: 2022-05-31 04:37:53        1   4.9178340   3.9738447          0.015625
#>  5: 2022-05-31 04:37:53        1  -1.6726606   3.1734883          0.015625
#>  6: 2022-05-31 04:37:53        1   7.1777705   9.5262479          0.015625
#>  7: 2022-05-31 04:37:53        1   1.2787814   1.0494746          0.015625
#>  8: 2022-05-31 04:37:53        1   3.1603589  10.5432008          0.015625
#>  9: 2022-05-31 04:37:53        1   0.5255736   3.4627602          0.015625
#> 10: 2022-05-31 04:37:53        1   1.1414949   8.8654616          0.015625
#> 11: 2022-05-31 04:37:53        1  -1.5428256   7.7510828          0.015625
#> 12: 2022-05-31 04:37:53        1  -4.8334412  12.7096181          0.015625
#> 13: 2022-05-31 04:37:53        1   2.9920365   0.9905126          0.015625
#> 14: 2022-05-31 04:37:53        1  -0.7405201   0.5300015          0.015625
#> 15: 2022-05-31 04:37:53        1  -4.5119861   3.2368684          0.015625
#> 16: 2022-05-31 04:37:53        1   2.7356623   7.6845264          0.015625
#> 17: 2022-05-31 04:37:53        1   4.5627892   5.7303329          0.015625
#> 18: 2022-05-31 04:37:53        1   5.7691745   0.9387208          0.015625
#> 19: 2022-05-31 04:37:53        1   6.3587466   6.6989527          0.015625
#> 20: 2022-05-31 04:37:53        1   2.3166716   8.0909588          0.015625
#> 21: 2022-05-31 04:37:53        1   9.5585616   1.6682523          0.015625
#> 22: 2022-05-31 04:37:53        1   1.2086236   8.1560971          0.015625
#> 23: 2022-05-31 04:37:53        1  -4.8487271   3.4190529          0.015625
#> 24: 2022-05-31 04:37:53        1   3.9952472   3.3380021          0.015625
#> 25: 2022-05-31 04:37:53        1   2.2478368   0.3514575          0.015625
#> 26: 2022-05-31 04:37:53        1  -0.7473669   7.5659091          0.015625
#> 27: 2022-05-31 04:37:53        1   9.7946614   5.5905394          0.015625
#> 28: 2022-05-31 04:37:53        1   6.8284949  13.5437045          0.015625
#> 29: 2022-05-31 04:37:53        1   5.5123989  10.7272031          0.015625
#> 30: 2022-05-31 04:37:53        1   4.7693918   2.1571098          0.015625
#> 31: 2022-05-31 04:37:53        1  -1.8774522   9.2545692          0.015625
#> 32: 2022-05-31 04:37:53        1  -2.8002691  12.2931825          0.015625
#> 33: 2022-05-31 04:37:53        1   9.0160273   7.8848350          0.015625
#> 34: 2022-05-31 04:37:53        1   8.3504398  11.9459183          0.015625
#> 35: 2022-05-31 04:37:53        1  -1.7574504  13.3701809          0.015625
#> 36: 2022-05-31 04:37:53        1   7.0683769  10.7431211          0.015625
#> 37: 2022-05-31 04:37:53        1  -1.3159400   0.3444959          0.015625
#> 38: 2022-05-31 04:37:53        1  -4.7896197   3.9952436          0.015625
#> 39: 2022-05-31 04:37:53        1  -0.5018152   1.0267093          0.015625
#> 40: 2022-05-31 04:37:53        1   8.4002642   0.5309106          0.015625
#> 41: 2022-05-31 04:37:53        1   3.8879695  11.1582305          0.015625
#> 42: 2022-05-31 04:37:53        1  -3.9806784   5.5879595          0.015625
#> 43: 2022-05-31 04:37:53        1  -3.5317214   3.9737163          0.015625
#> 44: 2022-05-31 04:37:53        1   4.2322791   1.5193774          0.015625
#> 45: 2022-05-31 04:37:53        1  -3.2501778   4.6173523          0.015625
#> 46: 2022-05-31 04:37:53        1  -3.9310964  11.6612329          0.015625
#> 47: 2022-05-31 04:37:53        1   3.0592177   4.8208197          0.015625
#> 48: 2022-05-31 04:37:53        1   5.4040842  14.6947452          0.015625
#> 49: 2022-05-31 04:37:53        1   9.4498884  11.1553855          0.015625
#> 50: 2022-05-31 04:37:53        1   3.4835116   7.1250455          0.015625
#> 51: 2022-05-31 04:37:53        1   7.6805065   3.5112231          0.015625
#> 52: 2022-05-31 04:37:53        1  -4.5884889   1.0621124          0.015625
#> 53: 2022-05-31 04:37:53        1  -3.9492952   0.5932226          0.015625
#> 54: 2022-05-31 04:37:53        1   1.4747830   1.1021254          0.015625
#> 55: 2022-05-31 04:37:53        1   8.9930606   6.9792611          0.015625
#> 56: 2022-05-31 04:37:53        1   4.5141508   6.0536168          0.015625
#> 57: 2022-05-31 04:37:53        1   8.5338942   0.8609467          0.015625
#> 58: 2022-05-31 04:37:53        1  -4.7918275   6.9746936          0.015625
#> 59: 2022-05-31 04:37:53        1   1.6856692  13.6881104          0.015625
#> 60: 2022-05-31 04:37:53        1   5.3593765   2.3640355          0.015625
#> 61: 2022-05-31 04:37:53        1   7.4419881   0.5397292          0.015625
#> 62: 2022-05-31 04:37:53        1  -3.6300739  12.5911696          0.015625
#> 63: 2022-05-31 04:37:53        1  -0.3476950   9.1650010          0.015625
#> 64: 2022-05-31 04:37:53        1   3.8636808   3.4950723          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity