Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.297934 0.5410661 0.015625 <list[3]> 0.811506

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.297934 0.5410661 0.015625 <list[3]> 0.811506

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  3.29793397  0.5410661 0.015625     0       6          1   0.811506
#>  2:  8.83293284  9.5557904 0.015625     0       6          1 233.525372
#>  3:  5.87691300  0.9058082 0.015625     0       6          1  29.042990
#>  4: -2.07332393  8.3756014 0.015625     0       6          1   6.491130
#>  5: -4.33825545  6.6316367 0.015625     0       6          1  53.434959
#>  6:  2.61757174  6.3054002 0.015625     0       6          1  19.840142
#>  7:  8.03757729  0.7021351 0.015625     0       6          1  38.585027
#>  8:  6.91092475  6.8133637 0.015625     0       6          1 124.768588
#>  9:  1.94656828  8.8391501 0.015625     0       6          1  40.356725
#> 10: -3.24375697  0.5840885 0.015625     0       6          1 119.302210
#> 11:  1.93353378  3.7272432 0.015625     0       6          1   7.068395
#> 12:  1.19498591 13.6611780 0.015625     0       6          1 104.138460
#> 13:  2.24653393  0.5846901 0.015625     0       6          1   7.974233
#> 14: -3.53083047  3.6846449 0.015625     0       6          1  70.307959
#> 15: -0.04130809 14.0752668 0.015625     0       6          1  83.745538
#> 16:  6.87091121  9.2039706 0.015625     0       6          1 178.971612
#> 17: -4.07370226  3.3961929 0.015625     0       6          1  96.388821
#> 18:  9.45607048  3.0024829 0.015625     0       6          1  86.947892
#> 19: -4.35851352 10.6232414 0.015625     0       6          1  15.068832
#> 20:  9.83747817  4.2926374 0.015625     0       6          1 121.630515
#> 21:  7.00219105 12.8799460 0.015625     0       6          1 290.027665
#> 22: -4.47043649  7.4845990 0.015625     0       6          1  46.696806
#> 23:  1.98878514  4.0631379 0.015625     0       6          1   7.327231
#> 24:  3.97023146 10.8274208 0.015625     0       6          1 117.179487
#> 25: -4.86828632  7.4753436 0.015625     0       6          1  60.511684
#> 26:  5.13240934  9.0527036 0.015625     0       6          1 122.309887
#> 27:  7.04912700 12.7263725 0.015625     0       6          1 286.457522
#> 28:  0.97390059  7.2916729 0.015625     0       6          1  23.307407
#> 29:  2.10506595  0.8471831 0.015625     0       6          1   8.869234
#> 30:  5.70256779 13.9965502 0.015625     0       6          1 276.356967
#> 31:  7.01748592  8.4618384 0.015625     0       6          1 163.934171
#> 32:  6.86879182 14.8794120 0.015625     0       6          1 355.120444
#> 33:  8.60595347  5.2336757 0.015625     0       6          1 116.933312
#> 34:  4.27550832 12.2612409 0.015625     0       6          1 162.283948
#> 35: -0.89671073 12.0601231 0.015625     0       6          1  37.229356
#> 36:  5.59295998 13.8860027 0.015625     0       6          1 267.856094
#> 37:  8.33902508  5.3057689 0.015625     0       6          1 114.508562
#> 38:  1.81813049  2.0279356 0.015625     0       6          1   9.041744
#> 39:  1.67720309  1.9175461 0.015625     0       6          1  11.228979
#> 40:  1.07913644  9.0498731 0.015625     0       6          1  36.920778
#> 41:  2.70230949  7.9385849 0.015625     0       6          1  37.488742
#> 42:  3.25934546  7.5029573 0.015625     0       6          1  40.961452
#> 43:  6.57086781  1.6991255 0.015625     0       6          1  42.531036
#> 44:  6.39324725 10.3340873 0.015625     0       6          1 195.173168
#> 45:  9.70474661  0.4197926 0.015625     0       6          1  49.346500
#> 46:  6.08862184  5.1588920 0.015625     0       6          1  78.855829
#> 47:  0.68902771 14.2282682 0.015625     0       6          1 104.093076
#> 48: -3.61578520 10.5252570 0.015625     0       6          1   4.118923
#> 49: -4.89518515  4.0656236 0.015625     0       6          1 121.200822
#> 50:  2.54888636  9.8763819 0.015625     0       6          1  61.839671
#> 51:  5.89879601  3.3821982 0.015625     0       6          1  51.397774
#> 52:  8.33646780 10.7517853 0.015625     0       6          1 257.810354
#> 53: -0.14074847 11.6029669 0.015625     0       6          1  48.433805
#> 54:  0.94995115  9.0045580 0.015625     0       6          1  35.734249
#> 55:  8.16571602  7.5490432 0.015625     0       6          1 163.180127
#> 56:  6.27748141 12.0171922 0.015625     0       6          1 238.537591
#> 57:  9.32645042 14.3264237 0.015625     0       6          1 420.508697
#> 58:  6.23446766  1.2049989 0.015625     0       6          1  35.054250
#> 59:  7.71681992 11.6416073 0.015625     0       6          1 270.276466
#> 60: -0.15601154  7.8920470 0.015625     0       6          1  22.184938
#> 61: -1.61117551 14.7754815 0.015625     0       6          1  47.206449
#> 62:  7.33529885  2.4320170 0.015625     0       6          1  56.389912
#> 63:  0.82035802 12.2559572 0.015625     0       6          1  73.413502
#> 64: -0.46639801  8.5105998 0.015625     0       6          1  21.679838
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-03-17 04:28:17        1  3.29793397   0.5410661          0.015625
#>  2: 2022-03-17 04:28:17        1  8.83293284   9.5557904          0.015625
#>  3: 2022-03-17 04:28:17        1  5.87691300   0.9058082          0.015625
#>  4: 2022-03-17 04:28:17        1 -2.07332393   8.3756014          0.015625
#>  5: 2022-03-17 04:28:17        1 -4.33825545   6.6316367          0.015625
#>  6: 2022-03-17 04:28:17        1  2.61757174   6.3054002          0.015625
#>  7: 2022-03-17 04:28:17        1  8.03757729   0.7021351          0.015625
#>  8: 2022-03-17 04:28:17        1  6.91092475   6.8133637          0.015625
#>  9: 2022-03-17 04:28:17        1  1.94656828   8.8391501          0.015625
#> 10: 2022-03-17 04:28:17        1 -3.24375697   0.5840885          0.015625
#> 11: 2022-03-17 04:28:17        1  1.93353378   3.7272432          0.015625
#> 12: 2022-03-17 04:28:17        1  1.19498591  13.6611780          0.015625
#> 13: 2022-03-17 04:28:17        1  2.24653393   0.5846901          0.015625
#> 14: 2022-03-17 04:28:17        1 -3.53083047   3.6846449          0.015625
#> 15: 2022-03-17 04:28:17        1 -0.04130809  14.0752668          0.015625
#> 16: 2022-03-17 04:28:17        1  6.87091121   9.2039706          0.015625
#> 17: 2022-03-17 04:28:17        1 -4.07370226   3.3961929          0.015625
#> 18: 2022-03-17 04:28:17        1  9.45607048   3.0024829          0.015625
#> 19: 2022-03-17 04:28:17        1 -4.35851352  10.6232414          0.015625
#> 20: 2022-03-17 04:28:17        1  9.83747817   4.2926374          0.015625
#> 21: 2022-03-17 04:28:17        1  7.00219105  12.8799460          0.015625
#> 22: 2022-03-17 04:28:17        1 -4.47043649   7.4845990          0.015625
#> 23: 2022-03-17 04:28:17        1  1.98878514   4.0631379          0.015625
#> 24: 2022-03-17 04:28:17        1  3.97023146  10.8274208          0.015625
#> 25: 2022-03-17 04:28:17        1 -4.86828632   7.4753436          0.015625
#> 26: 2022-03-17 04:28:17        1  5.13240934   9.0527036          0.015625
#> 27: 2022-03-17 04:28:17        1  7.04912700  12.7263725          0.015625
#> 28: 2022-03-17 04:28:17        1  0.97390059   7.2916729          0.015625
#> 29: 2022-03-17 04:28:17        1  2.10506595   0.8471831          0.015625
#> 30: 2022-03-17 04:28:17        1  5.70256779  13.9965502          0.015625
#> 31: 2022-03-17 04:28:17        1  7.01748592   8.4618384          0.015625
#> 32: 2022-03-17 04:28:17        1  6.86879182  14.8794120          0.015625
#> 33: 2022-03-17 04:28:17        1  8.60595347   5.2336757          0.015625
#> 34: 2022-03-17 04:28:17        1  4.27550832  12.2612409          0.015625
#> 35: 2022-03-17 04:28:17        1 -0.89671073  12.0601231          0.015625
#> 36: 2022-03-17 04:28:17        1  5.59295998  13.8860027          0.015625
#> 37: 2022-03-17 04:28:17        1  8.33902508   5.3057689          0.015625
#> 38: 2022-03-17 04:28:17        1  1.81813049   2.0279356          0.015625
#> 39: 2022-03-17 04:28:17        1  1.67720309   1.9175461          0.015625
#> 40: 2022-03-17 04:28:17        1  1.07913644   9.0498731          0.015625
#> 41: 2022-03-17 04:28:17        1  2.70230949   7.9385849          0.015625
#> 42: 2022-03-17 04:28:17        1  3.25934546   7.5029573          0.015625
#> 43: 2022-03-17 04:28:17        1  6.57086781   1.6991255          0.015625
#> 44: 2022-03-17 04:28:17        1  6.39324725  10.3340873          0.015625
#> 45: 2022-03-17 04:28:17        1  9.70474661   0.4197926          0.015625
#> 46: 2022-03-17 04:28:17        1  6.08862184   5.1588920          0.015625
#> 47: 2022-03-17 04:28:17        1  0.68902771  14.2282682          0.015625
#> 48: 2022-03-17 04:28:17        1 -3.61578520  10.5252570          0.015625
#> 49: 2022-03-17 04:28:17        1 -4.89518515   4.0656236          0.015625
#> 50: 2022-03-17 04:28:17        1  2.54888636   9.8763819          0.015625
#> 51: 2022-03-17 04:28:17        1  5.89879601   3.3821982          0.015625
#> 52: 2022-03-17 04:28:17        1  8.33646780  10.7517853          0.015625
#> 53: 2022-03-17 04:28:17        1 -0.14074847  11.6029669          0.015625
#> 54: 2022-03-17 04:28:17        1  0.94995115   9.0045580          0.015625
#> 55: 2022-03-17 04:28:17        1  8.16571602   7.5490432          0.015625
#> 56: 2022-03-17 04:28:17        1  6.27748141  12.0171922          0.015625
#> 57: 2022-03-17 04:28:17        1  9.32645042  14.3264237          0.015625
#> 58: 2022-03-17 04:28:17        1  6.23446766   1.2049989          0.015625
#> 59: 2022-03-17 04:28:17        1  7.71681992  11.6416073          0.015625
#> 60: 2022-03-17 04:28:17        1 -0.15601154   7.8920470          0.015625
#> 61: 2022-03-17 04:28:17        1 -1.61117551  14.7754815          0.015625
#> 62: 2022-03-17 04:28:17        1  7.33529885   2.4320170          0.015625
#> 63: 2022-03-17 04:28:17        1  0.82035802  12.2559572          0.015625
#> 64: 2022-03-17 04:28:17        1 -0.46639801   8.5105998          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity