Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 2.863941 1.722404 0.015625 <list[3]> 0.7664327

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 2.863941 1.722404 0.015625 <list[3]> 0.7664327

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition           y
#>  1:  6.74898974  6.1236855 0.015625     0       6          1 108.1567208
#>  2:  0.95630559  4.7236958 0.015625     0       6          1  15.5833690
#>  3:  5.67778819  4.9903032 0.015625     0       6          1  67.3950288
#>  4:  6.70439129  6.2029488 0.015625     0       6          1 108.8471724
#>  5:  8.50667529  4.5871139 0.015625     0       6          1 102.1970972
#>  6: -1.59954106  4.5610313 0.015625     0       6          1  26.2351223
#>  7:  2.86129358 11.7209099 0.015625     0       6          1 101.2346492
#>  8:  9.31149736  4.3711782 0.015625     0       6          1 111.2350214
#>  9:  9.79860918 14.7179496 0.015625     0       6          1 457.3455093
#> 10:  4.68711818  1.5674210 0.015625     0       6          1  15.2879498
#> 11: -1.82478663  8.2677678 0.015625     0       6          1   8.1331981
#> 12:  7.40055683  5.4397864 0.015625     0       6          1 105.1067406
#> 13:  1.26123650  4.6313716 0.015625     0       6          1  13.2730248
#> 14: -1.62379716  5.1412890 0.015625     0       6          1  21.9108500
#> 15: -1.37225961 14.7265554 0.015625     0       6          1  53.9443867
#> 16:  1.74342038  0.2404661 0.015625     0       6          1  17.8262884
#> 17:  0.01647538 11.1875605 0.015625     0       6          1  46.7842444
#> 18:  0.89475696 11.5033329 0.015625     0       6          1  63.6563158
#> 19: -1.14830599  8.7995916 0.015625     0       6          1  14.8048065
#> 20:  5.07459114  7.0589350 0.015625     0       6          1  83.0182998
#> 21:  4.35753526 11.2608412 0.015625     0       6          1 141.5079796
#> 22: -1.32324213  4.6556183 0.015625     0       6          1  24.6324247
#> 23:  9.56612287  8.3671803 0.015625     0       6          1 218.8964410
#> 24:  6.02344644 12.4974014 0.015625     0       6          1 243.3346463
#> 25:  7.20160524 10.2066500 0.015625     0       6          1 213.8989519
#> 26:  0.18803593 11.1404442 0.015625     0       6          1  49.0115063
#> 27:  5.53454129  1.9952532 0.015625     0       6          1  31.9491551
#> 28:  5.51679779 10.6871535 0.015625     0       6          1 173.9591326
#> 29: -2.91021443  3.2048039 0.015625     0       6          1  59.7491725
#> 30:  7.19555927  0.6487366 0.015625     0       6          1  36.2051102
#> 31:  9.22171704 12.0125398 0.015625     0       6          1 327.2874448
#> 32:  2.61424745  5.0447476 0.015625     0       6          1  10.6742997
#> 33:  1.76884513  5.7748249 0.015625     0       6          1  14.3298973
#> 34:  4.14983600  3.7158122 0.015625     0       6          1  19.2502876
#> 35: -0.01598323  9.5764420 0.015625     0       6          1  32.2104594
#> 36:  8.50164911  4.9717886 0.015625     0       6          1 109.8903163
#> 37:  2.86394115  1.7224041 0.015625     0       6          1   0.7664327
#> 38:  9.47035647 13.4412679 0.015625     0       6          1 390.7145583
#> 39:  7.29116821  2.7619502 0.015625     0       6          1  60.4386100
#> 40: -3.95848213  9.7825446 0.015625     0       6          1  12.4236387
#> 41:  8.15653855  6.0505037 0.015625     0       6          1 127.8413817
#> 42:  2.03312267  6.9718313 0.015625     0       6          1  22.3680527
#> 43: -4.15650894  9.4310403 0.015625     0       6          1  18.7376336
#> 44:  4.91633970 14.7857916 0.015625     0       6          1 263.7135377
#> 45: -1.46189686  5.2821236 0.015625     0       6          1  20.7173677
#> 46:  9.62051927  1.0419935 0.015625     0       6          1  56.9476338
#> 47:  5.16045463  4.6421860 0.015625     0       6          1  50.5992209
#> 48: -0.36854396  7.5734001 0.015625     0       6          1  19.9229947
#> 49:  2.35876829  6.5232714 0.015625     0       6          1  20.0545109
#> 50:  6.83160203  4.7064496 0.015625     0       6          1  84.5235102
#> 51: -1.25342557 14.7787813 0.015625     0       6          1  58.3646559
#> 52:  6.68265026 14.0312153 0.015625     0       6          1 317.9247379
#> 53:  9.51927402  9.0381682 0.015625     0       6          1 237.6738713
#> 54:  8.57170527  7.3167871 0.015625     0       6          1 164.9742012
#> 55:  0.82758800  4.0634970 0.015625     0       6          1  16.9074356
#> 56: -3.55217436  5.1034146 0.015625     0       6          1  49.3319205
#> 57:  5.74379320  1.8098975 0.015625     0       6          1  33.7391684
#> 58:  0.35767611  8.6285969 0.015625     0       6          1  29.1955647
#> 59:  4.50464174  8.8709284 0.015625     0       6          1  96.6876478
#> 60:  9.15368325  4.2595207 0.015625     0       6          1 105.8473761
#> 61:  6.46601473 12.0438246 0.015625     0       6          1 245.9233344
#> 62:  0.17751842  4.3280158 0.015625     0       6          1  21.3844931
#> 63:  3.47714214  2.8112570 0.015625     0       6          1   4.8283147
#> 64:  8.60872834 14.4712037 0.015625     0       6          1 399.1829714
#>              x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-13 04:44:27        1  6.74898974   6.1236855          0.015625
#>  2: 2022-10-13 04:44:27        1  0.95630559   4.7236958          0.015625
#>  3: 2022-10-13 04:44:27        1  5.67778819   4.9903032          0.015625
#>  4: 2022-10-13 04:44:27        1  6.70439129   6.2029488          0.015625
#>  5: 2022-10-13 04:44:27        1  8.50667529   4.5871139          0.015625
#>  6: 2022-10-13 04:44:27        1 -1.59954106   4.5610313          0.015625
#>  7: 2022-10-13 04:44:27        1  2.86129358  11.7209099          0.015625
#>  8: 2022-10-13 04:44:27        1  9.31149736   4.3711782          0.015625
#>  9: 2022-10-13 04:44:27        1  9.79860918  14.7179496          0.015625
#> 10: 2022-10-13 04:44:27        1  4.68711818   1.5674210          0.015625
#> 11: 2022-10-13 04:44:27        1 -1.82478663   8.2677678          0.015625
#> 12: 2022-10-13 04:44:27        1  7.40055683   5.4397864          0.015625
#> 13: 2022-10-13 04:44:27        1  1.26123650   4.6313716          0.015625
#> 14: 2022-10-13 04:44:27        1 -1.62379716   5.1412890          0.015625
#> 15: 2022-10-13 04:44:27        1 -1.37225961  14.7265554          0.015625
#> 16: 2022-10-13 04:44:27        1  1.74342038   0.2404661          0.015625
#> 17: 2022-10-13 04:44:27        1  0.01647538  11.1875605          0.015625
#> 18: 2022-10-13 04:44:27        1  0.89475696  11.5033329          0.015625
#> 19: 2022-10-13 04:44:27        1 -1.14830599   8.7995916          0.015625
#> 20: 2022-10-13 04:44:27        1  5.07459114   7.0589350          0.015625
#> 21: 2022-10-13 04:44:27        1  4.35753526  11.2608412          0.015625
#> 22: 2022-10-13 04:44:27        1 -1.32324213   4.6556183          0.015625
#> 23: 2022-10-13 04:44:27        1  9.56612287   8.3671803          0.015625
#> 24: 2022-10-13 04:44:27        1  6.02344644  12.4974014          0.015625
#> 25: 2022-10-13 04:44:27        1  7.20160524  10.2066500          0.015625
#> 26: 2022-10-13 04:44:27        1  0.18803593  11.1404442          0.015625
#> 27: 2022-10-13 04:44:27        1  5.53454129   1.9952532          0.015625
#> 28: 2022-10-13 04:44:27        1  5.51679779  10.6871535          0.015625
#> 29: 2022-10-13 04:44:27        1 -2.91021443   3.2048039          0.015625
#> 30: 2022-10-13 04:44:27        1  7.19555927   0.6487366          0.015625
#> 31: 2022-10-13 04:44:27        1  9.22171704  12.0125398          0.015625
#> 32: 2022-10-13 04:44:27        1  2.61424745   5.0447476          0.015625
#> 33: 2022-10-13 04:44:27        1  1.76884513   5.7748249          0.015625
#> 34: 2022-10-13 04:44:27        1  4.14983600   3.7158122          0.015625
#> 35: 2022-10-13 04:44:27        1 -0.01598323   9.5764420          0.015625
#> 36: 2022-10-13 04:44:27        1  8.50164911   4.9717886          0.015625
#> 37: 2022-10-13 04:44:27        1  2.86394115   1.7224041          0.015625
#> 38: 2022-10-13 04:44:27        1  9.47035647  13.4412679          0.015625
#> 39: 2022-10-13 04:44:27        1  7.29116821   2.7619502          0.015625
#> 40: 2022-10-13 04:44:27        1 -3.95848213   9.7825446          0.015625
#> 41: 2022-10-13 04:44:27        1  8.15653855   6.0505037          0.015625
#> 42: 2022-10-13 04:44:27        1  2.03312267   6.9718313          0.015625
#> 43: 2022-10-13 04:44:27        1 -4.15650894   9.4310403          0.015625
#> 44: 2022-10-13 04:44:27        1  4.91633970  14.7857916          0.015625
#> 45: 2022-10-13 04:44:27        1 -1.46189686   5.2821236          0.015625
#> 46: 2022-10-13 04:44:27        1  9.62051927   1.0419935          0.015625
#> 47: 2022-10-13 04:44:27        1  5.16045463   4.6421860          0.015625
#> 48: 2022-10-13 04:44:27        1 -0.36854396   7.5734001          0.015625
#> 49: 2022-10-13 04:44:27        1  2.35876829   6.5232714          0.015625
#> 50: 2022-10-13 04:44:27        1  6.83160203   4.7064496          0.015625
#> 51: 2022-10-13 04:44:27        1 -1.25342557  14.7787813          0.015625
#> 52: 2022-10-13 04:44:27        1  6.68265026  14.0312153          0.015625
#> 53: 2022-10-13 04:44:27        1  9.51927402   9.0381682          0.015625
#> 54: 2022-10-13 04:44:27        1  8.57170527   7.3167871          0.015625
#> 55: 2022-10-13 04:44:27        1  0.82758800   4.0634970          0.015625
#> 56: 2022-10-13 04:44:27        1 -3.55217436   5.1034146          0.015625
#> 57: 2022-10-13 04:44:27        1  5.74379320   1.8098975          0.015625
#> 58: 2022-10-13 04:44:27        1  0.35767611   8.6285969          0.015625
#> 59: 2022-10-13 04:44:27        1  4.50464174   8.8709284          0.015625
#> 60: 2022-10-13 04:44:27        1  9.15368325   4.2595207          0.015625
#> 61: 2022-10-13 04:44:27        1  6.46601473  12.0438246          0.015625
#> 62: 2022-10-13 04:44:27        1  0.17751842   4.3280158          0.015625
#> 63: 2022-10-13 04:44:27        1  3.47714214   2.8112570          0.015625
#> 64: 2022-10-13 04:44:27        1  8.60872834  14.4712037          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity