Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.683623 11.92674 0.015625 <list[3]> 1.898931

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.683623 11.92674 0.015625 <list[3]> 1.898931

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1:  1.2433684 12.5561670 0.015625     0       6          1  85.126040
#>  2:  6.8810051  3.3057174 0.015625     0       6          1  64.195273
#>  3:  1.0201573 12.6631997 0.015625     0       6          1  83.166372
#>  4:  2.6680944  5.2786129 0.015625     0       6          1  12.384871
#>  5: -1.2538576  8.0782782 0.015625     0       6          1  12.993768
#>  6: -3.6836234 11.9267435 0.015625     0       6          1   1.898931
#>  7:  0.9666516  7.5810473 0.015625     0       6          1  25.007547
#>  8:  8.9625051  5.5350134 0.015625     0       6          1 129.762915
#>  9:  1.8445946 11.9414133 0.015625     0       6          1  84.361478
#> 10:  2.6083354  0.5960669 0.015625     0       6          1   3.868011
#> 11:  2.2403379  7.6346607 0.015625     0       6          1  29.502250
#> 12:  7.0986671  4.7896018 0.015625     0       6          1  89.481683
#> 13:  0.2952928  5.6310071 0.015625     0       6          1  19.196167
#> 14:  6.3654335  9.3088035 0.015625     0       6          1 168.260249
#> 15:  8.6690231 14.1996883 0.015625     0       6          1 390.557274
#> 16: -4.3414118  6.2871201 0.015625     0       6          1  58.387305
#> 17:  2.6043778  9.4303754 0.015625     0       6          1  56.020520
#> 18: -1.4952620  7.3193932 0.015625     0       6          1  11.999544
#> 19:  4.3874811  1.9387137 0.015625     0       6          1  12.362427
#> 20:  6.8891304  9.0154127 0.015625     0       6          1 174.656705
#> 21:  9.3864890 10.3080735 0.015625     0       6          1 273.914778
#> 22:  4.9795228  8.5938482 0.015625     0       6          1 107.726147
#> 23: -0.4210327 11.4746797 0.015625     0       6          1  41.795230
#> 24:  4.6171265 11.0680603 0.015625     0       6          1 147.407430
#> 25:  6.1865534 14.6713109 0.015625     0       6          1 320.257974
#> 26: -1.4059780  0.1452072 0.015625     0       6          1  78.051034
#> 27:  3.4052355  7.7022800 0.015625     0       6          1  46.499625
#> 28:  4.1053125  2.9067751 0.015625     0       6          1  13.062631
#> 29:  8.1993625  2.0684235 0.015625     0       6          1  56.465853
#> 30: -3.6037080  3.1364348 0.015625     0       6          1  82.375290
#> 31:  5.2784155  0.6007035 0.015625     0       6          1  19.750039
#> 32:  6.3183646  3.7164008 0.015625     0       6          1  62.432104
#> 33: -4.2778212  5.7392231 0.015625     0       6          1  64.201758
#> 34:  9.6984319  3.1184328 0.015625     0       6          1  94.107802
#> 35:  6.1098803  3.4083342 0.015625     0       6          1  55.274973
#> 36:  8.0453883  8.0982085 0.015625     0       6          1 174.909587
#> 37:  5.6611836 11.4500835 0.015625     0       6          1 199.372015
#> 38: -0.0789500  9.1730297 0.015625     0       6          1  28.857541
#> 39:  7.8666637  8.7552115 0.015625     0       6          1 188.705608
#> 40:  2.7665007  7.7837920 0.015625     0       6          1  36.485737
#> 41:  1.0555051  4.5180779 0.015625     0       6          1  14.758610
#> 42: -2.6364876  9.8254136 0.015625     0       6          1   1.938501
#> 43:  0.4115256 14.9697376 0.015625     0       6          1 111.335110
#> 44:  3.0728060  3.3736043 0.015625     0       6          1   4.316519
#> 45: -1.3125246 11.3009975 0.015625     0       6          1  22.432269
#> 46: -1.0015849 10.3073853 0.015625     0       6          1  22.370871
#> 47: -1.8157431  0.6615291 0.015625     0       6          1  77.055117
#> 48:  6.8280142  1.0357447 0.015625     0       6          1  38.187297
#> 49:  7.0068915 14.9613575 0.015625     0       6          1 363.289352
#> 50:  1.8170108 14.4898392 0.015625     0       6          1 134.902314
#> 51: -3.7866911 12.6098886 0.015625     0       6          1   2.347772
#> 52:  6.8930896  6.8615663 0.015625     0       6          1 125.435255
#> 53:  4.1813136 10.0840383 0.015625     0       6          1 109.202279
#> 54:  6.2214348  4.8358955 0.015625     0       6          1  76.548795
#> 55:  6.3087416  2.3786619 0.015625     0       6          1  46.593219
#> 56:  9.7359563  1.5101412 0.015625     0       6          1  66.322085
#> 57:  5.3046154 12.6187687 0.015625     0       6          1 216.890323
#> 58: -4.7678293  1.6039519 0.015625     0       6          1 171.396427
#> 59: -0.2220417  6.0154498 0.015625     0       6          1  19.481610
#> 60: -4.0142319 12.3603229 0.015625     0       6          1   4.102254
#> 61: -2.2544269  2.5503175 0.015625     0       6          1  55.688600
#> 62:  1.4200763  3.4565274 0.015625     0       6          1  11.561028
#> 63:  4.2849222  6.8172763 0.015625     0       6          1  56.038260
#> 64: -1.5287534  6.5847851 0.015625     0       6          1  14.090596
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-09-23 04:46:39        1   1.2433684  12.5561670          0.015625
#>  2: 2022-09-23 04:46:39        1   6.8810051   3.3057174          0.015625
#>  3: 2022-09-23 04:46:39        1   1.0201573  12.6631997          0.015625
#>  4: 2022-09-23 04:46:39        1   2.6680944   5.2786129          0.015625
#>  5: 2022-09-23 04:46:39        1  -1.2538576   8.0782782          0.015625
#>  6: 2022-09-23 04:46:39        1  -3.6836234  11.9267435          0.015625
#>  7: 2022-09-23 04:46:39        1   0.9666516   7.5810473          0.015625
#>  8: 2022-09-23 04:46:39        1   8.9625051   5.5350134          0.015625
#>  9: 2022-09-23 04:46:39        1   1.8445946  11.9414133          0.015625
#> 10: 2022-09-23 04:46:39        1   2.6083354   0.5960669          0.015625
#> 11: 2022-09-23 04:46:39        1   2.2403379   7.6346607          0.015625
#> 12: 2022-09-23 04:46:39        1   7.0986671   4.7896018          0.015625
#> 13: 2022-09-23 04:46:39        1   0.2952928   5.6310071          0.015625
#> 14: 2022-09-23 04:46:39        1   6.3654335   9.3088035          0.015625
#> 15: 2022-09-23 04:46:39        1   8.6690231  14.1996883          0.015625
#> 16: 2022-09-23 04:46:39        1  -4.3414118   6.2871201          0.015625
#> 17: 2022-09-23 04:46:39        1   2.6043778   9.4303754          0.015625
#> 18: 2022-09-23 04:46:39        1  -1.4952620   7.3193932          0.015625
#> 19: 2022-09-23 04:46:39        1   4.3874811   1.9387137          0.015625
#> 20: 2022-09-23 04:46:39        1   6.8891304   9.0154127          0.015625
#> 21: 2022-09-23 04:46:39        1   9.3864890  10.3080735          0.015625
#> 22: 2022-09-23 04:46:39        1   4.9795228   8.5938482          0.015625
#> 23: 2022-09-23 04:46:39        1  -0.4210327  11.4746797          0.015625
#> 24: 2022-09-23 04:46:39        1   4.6171265  11.0680603          0.015625
#> 25: 2022-09-23 04:46:39        1   6.1865534  14.6713109          0.015625
#> 26: 2022-09-23 04:46:39        1  -1.4059780   0.1452072          0.015625
#> 27: 2022-09-23 04:46:39        1   3.4052355   7.7022800          0.015625
#> 28: 2022-09-23 04:46:39        1   4.1053125   2.9067751          0.015625
#> 29: 2022-09-23 04:46:39        1   8.1993625   2.0684235          0.015625
#> 30: 2022-09-23 04:46:39        1  -3.6037080   3.1364348          0.015625
#> 31: 2022-09-23 04:46:39        1   5.2784155   0.6007035          0.015625
#> 32: 2022-09-23 04:46:39        1   6.3183646   3.7164008          0.015625
#> 33: 2022-09-23 04:46:39        1  -4.2778212   5.7392231          0.015625
#> 34: 2022-09-23 04:46:39        1   9.6984319   3.1184328          0.015625
#> 35: 2022-09-23 04:46:39        1   6.1098803   3.4083342          0.015625
#> 36: 2022-09-23 04:46:39        1   8.0453883   8.0982085          0.015625
#> 37: 2022-09-23 04:46:39        1   5.6611836  11.4500835          0.015625
#> 38: 2022-09-23 04:46:39        1  -0.0789500   9.1730297          0.015625
#> 39: 2022-09-23 04:46:39        1   7.8666637   8.7552115          0.015625
#> 40: 2022-09-23 04:46:39        1   2.7665007   7.7837920          0.015625
#> 41: 2022-09-23 04:46:39        1   1.0555051   4.5180779          0.015625
#> 42: 2022-09-23 04:46:39        1  -2.6364876   9.8254136          0.015625
#> 43: 2022-09-23 04:46:39        1   0.4115256  14.9697376          0.015625
#> 44: 2022-09-23 04:46:39        1   3.0728060   3.3736043          0.015625
#> 45: 2022-09-23 04:46:39        1  -1.3125246  11.3009975          0.015625
#> 46: 2022-09-23 04:46:39        1  -1.0015849  10.3073853          0.015625
#> 47: 2022-09-23 04:46:39        1  -1.8157431   0.6615291          0.015625
#> 48: 2022-09-23 04:46:39        1   6.8280142   1.0357447          0.015625
#> 49: 2022-09-23 04:46:39        1   7.0068915  14.9613575          0.015625
#> 50: 2022-09-23 04:46:39        1   1.8170108  14.4898392          0.015625
#> 51: 2022-09-23 04:46:39        1  -3.7866911  12.6098886          0.015625
#> 52: 2022-09-23 04:46:39        1   6.8930896   6.8615663          0.015625
#> 53: 2022-09-23 04:46:39        1   4.1813136  10.0840383          0.015625
#> 54: 2022-09-23 04:46:39        1   6.2214348   4.8358955          0.015625
#> 55: 2022-09-23 04:46:39        1   6.3087416   2.3786619          0.015625
#> 56: 2022-09-23 04:46:39        1   9.7359563   1.5101412          0.015625
#> 57: 2022-09-23 04:46:39        1   5.3046154  12.6187687          0.015625
#> 58: 2022-09-23 04:46:39        1  -4.7678293   1.6039519          0.015625
#> 59: 2022-09-23 04:46:39        1  -0.2220417   6.0154498          0.015625
#> 60: 2022-09-23 04:46:39        1  -4.0142319  12.3603229          0.015625
#> 61: 2022-09-23 04:46:39        1  -2.2544269   2.5503175          0.015625
#> 62: 2022-09-23 04:46:39        1   1.4200763   3.4565274          0.015625
#> 63: 2022-09-23 04:46:39        1   4.2849222   6.8172763          0.015625
#> 64: 2022-09-23 04:46:39        1  -1.5287534   6.5847851          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity