Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.338591 1.136996 0.015625 <list[3]> 0.5952303

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.338591 1.136996 0.015625 <list[3]> 0.5952303

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -0.3365986  4.0730902 0.015625     0       6          1  25.1449586
#>  2:  0.7558989  1.6313561 0.015625     0       6          1  27.1195463
#>  3:  6.8986883  8.7697947 0.015625     0       6          1 168.7879599
#>  4:  8.3640729 14.7065660 0.015625     0       6          1 400.0257221
#>  5: -0.6139298  3.2886104 0.015625     0       6          1  31.5392529
#>  6:  6.7934795 13.1646061 0.015625     0       6          1 292.5378341
#>  7: -1.9592651 10.5070550 0.015625     0       6          1   7.9778184
#>  8: -3.9698521 13.4562885 0.015625     0       6          1   3.9345167
#>  9: -2.8722341 14.0234687 0.015625     0       6          1  10.9745992
#> 10: -4.8135344  5.7190008 0.015625     0       6          1  85.8679980
#> 11: -3.3076250  6.3723896 0.015625     0       6          1  27.8644594
#> 12:  6.8216890  5.2867600 0.015625     0       6          1  94.1593823
#> 13:  3.1360472 12.4207290 0.015625     0       6          1 123.8193084
#> 14: -3.1615206  4.2877877 0.015625     0       6          1  50.1199403
#> 15:  2.1422825  6.4519505 0.015625     0       6          1  18.6476813
#> 16:  1.7406735  7.8494154 0.015625     0       6          1  28.8669575
#> 17:  3.5704576  7.1092447 0.015625     0       6          1  42.2253776
#> 18: -3.8543504  8.2427015 0.015625     0       6          1  21.6445664
#> 19:  8.6788466 12.3983810 0.015625     0       6          1 323.1878779
#> 20:  8.3230595  0.8201440 0.015625     0       6          1  40.9043173
#> 21:  3.3385907  1.1369956 0.015625     0       6          1   0.5952303
#> 22: -1.4460755  5.1999205 0.015625     0       6          1  21.2172535
#> 23:  8.4236147  1.8220794 0.015625     0       6          1  54.4809241
#> 24:  0.7617846 12.6602125 0.015625     0       6          1  78.6458256
#> 25: -2.7469950  1.8579411 0.015625     0       6          1  77.6292506
#> 26: -4.8942479  7.1571388 0.015625     0       6          1  66.0358660
#> 27: -0.2466818  9.6102296 0.015625     0       6          1  29.6525002
#> 28:  0.4991199  8.7732538 0.015625     0       6          1  31.1040673
#> 29:  0.7003703  4.4843065 0.015625     0       6          1  17.5149504
#> 30:  7.8153750  9.6907437 0.015625     0       6          1 213.4692915
#> 31:  5.3371332  6.0896941 0.015625     0       6          1  75.0321717
#> 32:  4.1095444 10.7674505 0.015625     0       6          1 120.9523513
#> 33: -1.0755331  7.5940707 0.015625     0       6          1  14.5870200
#> 34:  8.1143947 12.5132247 0.015625     0       6          1 310.3989257
#> 35:  1.4892968 14.1096794 0.015625     0       6          1 119.1866944
#> 36:  7.0649780  4.8798087 0.015625     0       6          1  90.5913773
#> 37:  8.0806584 10.2399985 0.015625     0       6          1 235.6429910
#> 38:  4.8546886  3.9594285 0.015625     0       6          1  35.9758704
#> 39: -0.2153684  0.1200171 0.015625     0       6          1  58.1206794
#> 40:  3.2555183  2.5019318 0.015625     0       6          1   2.3026166
#> 41:  2.1556586  5.5519924 0.015625     0       6          1  12.7641518
#> 42: -2.0413563  9.0282558 0.015625     0       6          1   5.7681969
#> 43:  6.3370439  9.9958366 0.015625     0       6          1 184.6287925
#> 44: -1.5742041  2.5292841 0.015625     0       6          1  46.5980276
#> 45: -0.5271964  1.5480458 0.015625     0       6          1  46.3836726
#> 46:  5.4006220 14.8894400 0.015625     0       6          1 291.2610008
#> 47: -0.5043797  6.4229923 0.015625     0       6          1  18.5566181
#> 48:  5.6315028  7.6632858 0.015625     0       6          1 110.7760166
#> 49:  5.2575440  5.6679231 0.015625     0       6          1  66.6129290
#> 50:  5.3873063  3.3535071 0.015625     0       6          1  41.3539481
#> 51: -4.0823376  6.3465654 0.015625     0       6          1  48.7393410
#> 52:  2.0496387  7.8160746 0.015625     0       6          1  30.0684783
#> 53:  4.7424627  0.2992322 0.015625     0       6          1  11.6240611
#> 54: -2.0433661 13.7281067 0.015625     0       6          1  24.5310549
#> 55:  2.9302201 13.8185640 0.015625     0       6          1 149.8948905
#> 56:  9.8576654  0.8643565 0.015625     0       6          1  58.5207206
#> 57: -4.5480243 13.3248906 0.015625     0       6          1   8.7308013
#> 58:  7.0921916  9.8447084 0.015625     0       6          1 201.1996765
#> 59:  6.8107129  4.5708281 0.015625     0       6          1  82.0440276
#> 60:  7.5634988 12.5290117 0.015625     0       6          1 295.2511416
#> 61:  0.3342960  4.8680418 0.015625     0       6          1  19.4345804
#> 62:  9.5841290  7.3720658 0.015625     0       6          1 190.9981289
#> 63:  0.8702350  6.6739635 0.015625     0       6          1  20.3340488
#> 64:  8.5931081  8.9118897 0.015625     0       6          1 208.5321641
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-11-01 04:43:17        1  -0.3365986   4.0730902          0.015625
#>  2: 2022-11-01 04:43:17        1   0.7558989   1.6313561          0.015625
#>  3: 2022-11-01 04:43:17        1   6.8986883   8.7697947          0.015625
#>  4: 2022-11-01 04:43:17        1   8.3640729  14.7065660          0.015625
#>  5: 2022-11-01 04:43:17        1  -0.6139298   3.2886104          0.015625
#>  6: 2022-11-01 04:43:17        1   6.7934795  13.1646061          0.015625
#>  7: 2022-11-01 04:43:17        1  -1.9592651  10.5070550          0.015625
#>  8: 2022-11-01 04:43:17        1  -3.9698521  13.4562885          0.015625
#>  9: 2022-11-01 04:43:17        1  -2.8722341  14.0234687          0.015625
#> 10: 2022-11-01 04:43:17        1  -4.8135344   5.7190008          0.015625
#> 11: 2022-11-01 04:43:17        1  -3.3076250   6.3723896          0.015625
#> 12: 2022-11-01 04:43:17        1   6.8216890   5.2867600          0.015625
#> 13: 2022-11-01 04:43:17        1   3.1360472  12.4207290          0.015625
#> 14: 2022-11-01 04:43:17        1  -3.1615206   4.2877877          0.015625
#> 15: 2022-11-01 04:43:17        1   2.1422825   6.4519505          0.015625
#> 16: 2022-11-01 04:43:17        1   1.7406735   7.8494154          0.015625
#> 17: 2022-11-01 04:43:17        1   3.5704576   7.1092447          0.015625
#> 18: 2022-11-01 04:43:17        1  -3.8543504   8.2427015          0.015625
#> 19: 2022-11-01 04:43:17        1   8.6788466  12.3983810          0.015625
#> 20: 2022-11-01 04:43:17        1   8.3230595   0.8201440          0.015625
#> 21: 2022-11-01 04:43:17        1   3.3385907   1.1369956          0.015625
#> 22: 2022-11-01 04:43:17        1  -1.4460755   5.1999205          0.015625
#> 23: 2022-11-01 04:43:17        1   8.4236147   1.8220794          0.015625
#> 24: 2022-11-01 04:43:17        1   0.7617846  12.6602125          0.015625
#> 25: 2022-11-01 04:43:17        1  -2.7469950   1.8579411          0.015625
#> 26: 2022-11-01 04:43:17        1  -4.8942479   7.1571388          0.015625
#> 27: 2022-11-01 04:43:17        1  -0.2466818   9.6102296          0.015625
#> 28: 2022-11-01 04:43:17        1   0.4991199   8.7732538          0.015625
#> 29: 2022-11-01 04:43:17        1   0.7003703   4.4843065          0.015625
#> 30: 2022-11-01 04:43:17        1   7.8153750   9.6907437          0.015625
#> 31: 2022-11-01 04:43:17        1   5.3371332   6.0896941          0.015625
#> 32: 2022-11-01 04:43:17        1   4.1095444  10.7674505          0.015625
#> 33: 2022-11-01 04:43:17        1  -1.0755331   7.5940707          0.015625
#> 34: 2022-11-01 04:43:17        1   8.1143947  12.5132247          0.015625
#> 35: 2022-11-01 04:43:17        1   1.4892968  14.1096794          0.015625
#> 36: 2022-11-01 04:43:17        1   7.0649780   4.8798087          0.015625
#> 37: 2022-11-01 04:43:17        1   8.0806584  10.2399985          0.015625
#> 38: 2022-11-01 04:43:17        1   4.8546886   3.9594285          0.015625
#> 39: 2022-11-01 04:43:17        1  -0.2153684   0.1200171          0.015625
#> 40: 2022-11-01 04:43:17        1   3.2555183   2.5019318          0.015625
#> 41: 2022-11-01 04:43:17        1   2.1556586   5.5519924          0.015625
#> 42: 2022-11-01 04:43:17        1  -2.0413563   9.0282558          0.015625
#> 43: 2022-11-01 04:43:17        1   6.3370439   9.9958366          0.015625
#> 44: 2022-11-01 04:43:17        1  -1.5742041   2.5292841          0.015625
#> 45: 2022-11-01 04:43:17        1  -0.5271964   1.5480458          0.015625
#> 46: 2022-11-01 04:43:17        1   5.4006220  14.8894400          0.015625
#> 47: 2022-11-01 04:43:17        1  -0.5043797   6.4229923          0.015625
#> 48: 2022-11-01 04:43:17        1   5.6315028   7.6632858          0.015625
#> 49: 2022-11-01 04:43:17        1   5.2575440   5.6679231          0.015625
#> 50: 2022-11-01 04:43:17        1   5.3873063   3.3535071          0.015625
#> 51: 2022-11-01 04:43:17        1  -4.0823376   6.3465654          0.015625
#> 52: 2022-11-01 04:43:17        1   2.0496387   7.8160746          0.015625
#> 53: 2022-11-01 04:43:17        1   4.7424627   0.2992322          0.015625
#> 54: 2022-11-01 04:43:17        1  -2.0433661  13.7281067          0.015625
#> 55: 2022-11-01 04:43:17        1   2.9302201  13.8185640          0.015625
#> 56: 2022-11-01 04:43:17        1   9.8576654   0.8643565          0.015625
#> 57: 2022-11-01 04:43:17        1  -4.5480243  13.3248906          0.015625
#> 58: 2022-11-01 04:43:17        1   7.0921916   9.8447084          0.015625
#> 59: 2022-11-01 04:43:17        1   6.8107129   4.5708281          0.015625
#> 60: 2022-11-01 04:43:17        1   7.5634988  12.5290117          0.015625
#> 61: 2022-11-01 04:43:17        1   0.3342960   4.8680418          0.015625
#> 62: 2022-11-01 04:43:17        1   9.5841290   7.3720658          0.015625
#> 63: 2022-11-01 04:43:17        1   0.8702350   6.6739635          0.015625
#> 64: 2022-11-01 04:43:17        1   8.5931081   8.9118897          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity