Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.515538 0.3031267 0.015625 <list[3]> 1.293514

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.515538 0.3031267 0.015625 <list[3]> 1.293514

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  3.51553791  0.3031267 0.015625     0       6          1   1.293514
#>  2: -0.21236154 14.2474931 0.015625     0       6          1  81.924813
#>  3:  0.25951436  6.4667082 0.015625     0       6          1  20.050883
#>  4:  9.32110404 13.9393141 0.015625     0       6          1 404.576841
#>  5: -1.17688680  2.1516995 0.015625     0       6          1  46.908546
#>  6: -2.69848284  1.1489754 0.015625     0       6          1  89.116273
#>  7:  6.25516344 13.1667465 0.015625     0       6          1 273.016342
#>  8:  0.87561772  6.6366628 0.015625     0       6          1  20.177126
#>  9:  0.72625564  3.3313275 0.015625     0       6          1  19.517071
#> 10:  5.58281120  7.2037037 0.015625     0       6          1 100.711608
#> 11:  2.68005978  2.7599394 0.015625     0       6          1   2.049832
#> 12:  7.34419568  4.9345126 0.015625     0       6          1  95.052575
#> 13:  5.54992971  0.1897964 0.015625     0       6          1  21.442990
#> 14:  0.19583425 13.1081828 0.015625     0       6          1  74.455440
#> 15:  3.25862226 13.8651583 0.015625     0       6          1 162.387307
#> 16:  5.25087901 14.1731143 0.015625     0       6          1 260.862005
#> 17:  8.98834347  2.4717905 0.015625     0       6          1  70.073831
#> 18: -4.00318398  0.9546287 0.015625     0       6          1 145.579954
#> 19: -3.60659714 10.3312596 0.015625     0       6          1   4.689061
#> 20:  8.13026528  1.5993932 0.015625     0       6          1  49.717830
#> 21:  9.22497766  4.1627196 0.015625     0       6          1 105.203972
#> 22:  1.81072637 11.4816872 0.015625     0       6          1  75.991020
#> 23:  1.38359555 10.3650844 0.015625     0       6          1  54.144816
#> 24: -3.79662594  8.9187824 0.015625     0       6          1  15.108278
#> 25:  2.37872761  0.6244461 0.015625     0       6          1   6.169551
#> 26: -1.89947491 13.5096858 0.015625     0       6          1  26.046754
#> 27: -1.76154428 14.6529711 0.015625     0       6          1  41.287756
#> 28:  8.38735290  3.5774021 0.015625     0       6          1  81.913472
#> 29:  9.80719523  8.0062676 0.015625     0       6          1 215.937409
#> 30:  4.84451690  5.8380616 0.015625     0       6          1  57.869365
#> 31:  0.37378496 12.6411393 0.015625     0       6          1  71.237179
#> 32:  9.54574682  2.5327453 0.015625     0       6          1  80.098094
#> 33:  5.79764786  3.7956564 0.015625     0       6          1  54.365360
#> 34:  6.83335198  0.3232082 0.015625     0       6          1  32.346146
#> 35:  1.56869535  7.2752674 0.015625     0       6          1  23.682521
#> 36:  8.29537837  3.5898828 0.015625     0       6          1  81.181066
#> 37: -2.01350832  0.5357935 0.015625     0       6          1  83.211465
#> 38: -1.07638991  9.2998556 0.015625     0       6          1  16.962214
#> 39: -2.94011467  2.2909879 0.015625     0       6          1  75.486182
#> 40:  6.17755446 10.3859440 0.015625     0       6          1 189.706383
#> 41: -0.02831846  1.1876749 0.015625     0       6          1  43.192792
#> 42: -1.73690497  3.1912674 0.015625     0       6          1  40.514318
#> 43:  8.70066344  8.1720943 0.015625     0       6          1 190.278990
#> 44:  8.25556767  9.4711300 0.015625     0       6          1 216.924129
#> 45: -2.17793817 10.1368224 0.015625     0       6          1   4.797069
#> 46:  1.81386446  4.1210851 0.015625     0       6          1   8.511196
#> 47: -0.74176994 14.5919981 0.015625     0       6          1  71.758268
#> 48:  1.33153884 11.4261704 0.015625     0       6          1  68.388594
#> 49:  8.98028622  0.6327267 0.015625     0       6          1  42.877811
#> 50: -4.07447959  7.8283038 0.015625     0       6          1  30.978371
#> 51:  4.29655346  3.4459975 0.015625     0       6          1  19.934103
#> 52:  7.82495196  1.9148701 0.015625     0       6          1  52.347324
#> 53:  5.13033181 14.5341110 0.015625     0       6          1 266.389959
#> 54: -1.00857741 11.6994342 0.015625     0       6          1  31.626166
#> 55:  5.15138613 14.9210179 0.015625     0       6          1 279.894347
#> 56:  7.35628451  6.9125705 0.015625     0       6          1 134.630921
#> 57:  1.00825803  9.9393707 0.015625     0       6          1  45.512176
#> 58:  9.18555539  6.6715586 0.015625     0       6          1 161.874037
#> 59:  1.97045233  0.3088656 0.015625     0       6          1  13.416449
#> 60:  2.33568111 11.9021669 0.015625     0       6          1  92.687152
#> 61: -1.27728143  8.7252842 0.015625     0       6          1  13.190575
#> 62: -1.37941419 13.7682977 0.015625     0       6          1  42.234939
#> 63: -1.05399191  8.4844077 0.015625     0       6          1  15.341621
#> 64:  4.11141000  6.0861664 0.015625     0       6          1  41.901858
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-11-04 04:30:06        1  3.51553791   0.3031267          0.015625
#>  2: 2022-11-04 04:30:06        1 -0.21236154  14.2474931          0.015625
#>  3: 2022-11-04 04:30:06        1  0.25951436   6.4667082          0.015625
#>  4: 2022-11-04 04:30:06        1  9.32110404  13.9393141          0.015625
#>  5: 2022-11-04 04:30:06        1 -1.17688680   2.1516995          0.015625
#>  6: 2022-11-04 04:30:06        1 -2.69848284   1.1489754          0.015625
#>  7: 2022-11-04 04:30:06        1  6.25516344  13.1667465          0.015625
#>  8: 2022-11-04 04:30:06        1  0.87561772   6.6366628          0.015625
#>  9: 2022-11-04 04:30:06        1  0.72625564   3.3313275          0.015625
#> 10: 2022-11-04 04:30:06        1  5.58281120   7.2037037          0.015625
#> 11: 2022-11-04 04:30:06        1  2.68005978   2.7599394          0.015625
#> 12: 2022-11-04 04:30:06        1  7.34419568   4.9345126          0.015625
#> 13: 2022-11-04 04:30:06        1  5.54992971   0.1897964          0.015625
#> 14: 2022-11-04 04:30:06        1  0.19583425  13.1081828          0.015625
#> 15: 2022-11-04 04:30:06        1  3.25862226  13.8651583          0.015625
#> 16: 2022-11-04 04:30:06        1  5.25087901  14.1731143          0.015625
#> 17: 2022-11-04 04:30:06        1  8.98834347   2.4717905          0.015625
#> 18: 2022-11-04 04:30:06        1 -4.00318398   0.9546287          0.015625
#> 19: 2022-11-04 04:30:06        1 -3.60659714  10.3312596          0.015625
#> 20: 2022-11-04 04:30:06        1  8.13026528   1.5993932          0.015625
#> 21: 2022-11-04 04:30:06        1  9.22497766   4.1627196          0.015625
#> 22: 2022-11-04 04:30:06        1  1.81072637  11.4816872          0.015625
#> 23: 2022-11-04 04:30:06        1  1.38359555  10.3650844          0.015625
#> 24: 2022-11-04 04:30:06        1 -3.79662594   8.9187824          0.015625
#> 25: 2022-11-04 04:30:06        1  2.37872761   0.6244461          0.015625
#> 26: 2022-11-04 04:30:06        1 -1.89947491  13.5096858          0.015625
#> 27: 2022-11-04 04:30:06        1 -1.76154428  14.6529711          0.015625
#> 28: 2022-11-04 04:30:06        1  8.38735290   3.5774021          0.015625
#> 29: 2022-11-04 04:30:06        1  9.80719523   8.0062676          0.015625
#> 30: 2022-11-04 04:30:06        1  4.84451690   5.8380616          0.015625
#> 31: 2022-11-04 04:30:06        1  0.37378496  12.6411393          0.015625
#> 32: 2022-11-04 04:30:06        1  9.54574682   2.5327453          0.015625
#> 33: 2022-11-04 04:30:06        1  5.79764786   3.7956564          0.015625
#> 34: 2022-11-04 04:30:06        1  6.83335198   0.3232082          0.015625
#> 35: 2022-11-04 04:30:06        1  1.56869535   7.2752674          0.015625
#> 36: 2022-11-04 04:30:06        1  8.29537837   3.5898828          0.015625
#> 37: 2022-11-04 04:30:06        1 -2.01350832   0.5357935          0.015625
#> 38: 2022-11-04 04:30:06        1 -1.07638991   9.2998556          0.015625
#> 39: 2022-11-04 04:30:06        1 -2.94011467   2.2909879          0.015625
#> 40: 2022-11-04 04:30:06        1  6.17755446  10.3859440          0.015625
#> 41: 2022-11-04 04:30:06        1 -0.02831846   1.1876749          0.015625
#> 42: 2022-11-04 04:30:06        1 -1.73690497   3.1912674          0.015625
#> 43: 2022-11-04 04:30:06        1  8.70066344   8.1720943          0.015625
#> 44: 2022-11-04 04:30:06        1  8.25556767   9.4711300          0.015625
#> 45: 2022-11-04 04:30:06        1 -2.17793817  10.1368224          0.015625
#> 46: 2022-11-04 04:30:06        1  1.81386446   4.1210851          0.015625
#> 47: 2022-11-04 04:30:06        1 -0.74176994  14.5919981          0.015625
#> 48: 2022-11-04 04:30:06        1  1.33153884  11.4261704          0.015625
#> 49: 2022-11-04 04:30:06        1  8.98028622   0.6327267          0.015625
#> 50: 2022-11-04 04:30:06        1 -4.07447959   7.8283038          0.015625
#> 51: 2022-11-04 04:30:06        1  4.29655346   3.4459975          0.015625
#> 52: 2022-11-04 04:30:06        1  7.82495196   1.9148701          0.015625
#> 53: 2022-11-04 04:30:06        1  5.13033181  14.5341110          0.015625
#> 54: 2022-11-04 04:30:06        1 -1.00857741  11.6994342          0.015625
#> 55: 2022-11-04 04:30:06        1  5.15138613  14.9210179          0.015625
#> 56: 2022-11-04 04:30:06        1  7.35628451   6.9125705          0.015625
#> 57: 2022-11-04 04:30:06        1  1.00825803   9.9393707          0.015625
#> 58: 2022-11-04 04:30:06        1  9.18555539   6.6715586          0.015625
#> 59: 2022-11-04 04:30:06        1  1.97045233   0.3088656          0.015625
#> 60: 2022-11-04 04:30:06        1  2.33568111  11.9021669          0.015625
#> 61: 2022-11-04 04:30:06        1 -1.27728143   8.7252842          0.015625
#> 62: 2022-11-04 04:30:06        1 -1.37941419  13.7682977          0.015625
#> 63: 2022-11-04 04:30:06        1 -1.05399191   8.4844077          0.015625
#> 64: 2022-11-04 04:30:06        1  4.11141000   6.0861664          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity