Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# Branin function with fidelity, see `bbotk::branin()`
fun = function(xs) branin_wu(xs[["x1"]], xs[["x2"]], xs[["fidelity"]])

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.080695 8.550277 0.015625 <list[3]> 6.113203

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.080695 8.550277 0.015625 <list[3]> 6.113203

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition          y
#>  1:  6.98515572 14.85574519 0.015625     0       6          1 358.574221
#>  2:  4.74915866  8.00079340 0.015625     0       6          1  88.955938
#>  3: -0.76126986 11.86665402 0.015625     0       6          1  38.455514
#>  4:  6.52476794  5.06757586 0.015625     0       6          1  85.633131
#>  5:  6.81585498  9.27501440 0.015625     0       6          1 179.419705
#>  6:  2.78805278  9.32539152 0.015625     0       6          1  57.597984
#>  7:  9.64241880  8.15076463 0.015625     0       6          1 214.907437
#>  8:  1.62414993  2.83747006 0.015625     0       6          1   9.921850
#>  9: -2.09849144 11.93959798 0.015625     0       6          1  11.237919
#> 10:  8.96523816  3.64699356 0.015625     0       6          1  90.588980
#> 11:  1.36805596  8.15028952 0.015625     0       6          1  30.166949
#> 12:  8.80121198  8.45768456 0.015625     0       6          1 200.551209
#> 13:  0.12001402 11.43783444 0.015625     0       6          1  51.211929
#> 14:  7.33374932 12.14151529 0.015625     0       6          1 275.913671
#> 15:  3.01437623 10.34781455 0.015625     0       6          1  79.080760
#> 16:  7.54778702 13.52315378 0.015625     0       6          1 329.172187
#> 17:  7.16008942  0.40835162 0.015625     0       6          1  34.014197
#> 18:  0.09474210  4.74385112 0.015625     0       6          1  20.781486
#> 19:  2.33441392  8.47677760 0.015625     0       6          1  39.655054
#> 20: -0.37179605  7.09093319 0.015625     0       6          1  19.191038
#> 21: -3.33121780  7.73823277 0.015625     0       6          1  15.817195
#> 22:  6.50600859 12.05200129 0.015625     0       6          1 247.530933
#> 23:  9.84473754  4.23384557 0.015625     0       6          1 120.528176
#> 24: -4.40398123 10.47230243 0.015625     0       6          1  16.902244
#> 25: -1.51373763  6.78451369 0.015625     0       6          1  13.421047
#> 26:  3.55317433  7.39361558 0.015625     0       6          1  45.561837
#> 27: -2.08069537  8.55027665 0.015625     0       6          1   6.113203
#> 28:  1.23382562  4.06840983 0.015625     0       6          1  13.174960
#> 29:  5.61461219  3.92410491 0.015625     0       6          1  52.236051
#> 30:  7.95790488  2.64839111 0.015625     0       6          1  63.271236
#> 31:  6.65190352  1.69111977 0.015625     0       6          1  43.138216
#> 32:  1.13408186  6.45866915 0.015625     0       6          1  19.007847
#> 33: -4.02339077  8.79334246 0.015625     0       6          1  20.769385
#> 34: -4.99201164  0.46577706 0.015625     0       6          1 215.583737
#> 35:  1.40190487  8.75897108 0.015625     0       6          1  35.916389
#> 36: -4.60327987  3.66566344 0.015625     0       6          1 115.296257
#> 37: -0.90199814  3.25786661 0.015625     0       6          1  33.616615
#> 38: -0.03907316 13.13001536 0.015625     0       6          1  69.548320
#> 39:  5.30501633  4.68912311 0.015625     0       6          1  54.638471
#> 40:  5.88066715  9.29606262 0.015625     0       6          1 153.212268
#> 41:  0.52282257  9.57751273 0.015625     0       6          1  37.690014
#> 42: -3.85509844  4.29533139 0.015625     0       6          1  71.583701
#> 43:  1.00895408  1.74451787 0.015625     0       6          1  22.303153
#> 44:  6.11301612 10.92192769 0.015625     0       6          1 201.770424
#> 45:  5.28873116 11.16190555 0.015625     0       6          1 177.010007
#> 46:  8.05983316  5.35987782 0.015625     0       6          1 111.876416
#> 47: -2.18771053  6.75517203 0.015625     0       6          1  12.703904
#> 48:  5.02095552 11.40356344 0.015625     0       6          1 172.168644
#> 49:  1.03228546 12.98572984 0.015625     0       6          1  88.813997
#> 50: -1.91137650  9.36517966 0.015625     0       6          1   6.837004
#> 51: -0.63410592 13.94035759 0.015625     0       6          1  65.605045
#> 52:  8.14674120  0.53343973 0.015625     0       6          1  37.026506
#> 53:  2.11246957  9.33447266 0.015625     0       6          1  48.074678
#> 54: -0.29947187 11.71713320 0.015625     0       6          1  46.608786
#> 55: -0.02366642  0.01606912 0.015625     0       6          1  55.859263
#> 56:  1.58151027 10.39997638 0.015625     0       6          1  56.684429
#> 57:  9.40385479 14.79281947 0.015625     0       6          1 443.101989
#> 58:  0.20356764  7.90648273 0.015625     0       6          1  24.373161
#> 59: -4.63031528  2.52501892 0.015625     0       6          1 141.544859
#> 60:  2.79196061  9.19695958 0.015625     0       6          1  55.751091
#> 61:  3.89955083  8.95362437 0.015625     0       6          1  78.584333
#> 62: -2.93504045  7.67099135 0.015625     0       6          1  11.263128
#> 63:  2.16217075  5.91988833 0.015625     0       6          1  14.998127
#> 64:  7.73003042  7.64487127 0.015625     0       6          1 157.847952
#>              x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2023-01-12 10:59:48        1  6.98515572 14.85574519          0.015625
#>  2: 2023-01-12 10:59:48        1  4.74915866  8.00079340          0.015625
#>  3: 2023-01-12 10:59:48        1 -0.76126986 11.86665402          0.015625
#>  4: 2023-01-12 10:59:48        1  6.52476794  5.06757586          0.015625
#>  5: 2023-01-12 10:59:48        1  6.81585498  9.27501440          0.015625
#>  6: 2023-01-12 10:59:48        1  2.78805278  9.32539152          0.015625
#>  7: 2023-01-12 10:59:48        1  9.64241880  8.15076463          0.015625
#>  8: 2023-01-12 10:59:48        1  1.62414993  2.83747006          0.015625
#>  9: 2023-01-12 10:59:48        1 -2.09849144 11.93959798          0.015625
#> 10: 2023-01-12 10:59:48        1  8.96523816  3.64699356          0.015625
#> 11: 2023-01-12 10:59:48        1  1.36805596  8.15028952          0.015625
#> 12: 2023-01-12 10:59:48        1  8.80121198  8.45768456          0.015625
#> 13: 2023-01-12 10:59:48        1  0.12001402 11.43783444          0.015625
#> 14: 2023-01-12 10:59:48        1  7.33374932 12.14151529          0.015625
#> 15: 2023-01-12 10:59:48        1  3.01437623 10.34781455          0.015625
#> 16: 2023-01-12 10:59:48        1  7.54778702 13.52315378          0.015625
#> 17: 2023-01-12 10:59:48        1  7.16008942  0.40835162          0.015625
#> 18: 2023-01-12 10:59:48        1  0.09474210  4.74385112          0.015625
#> 19: 2023-01-12 10:59:48        1  2.33441392  8.47677760          0.015625
#> 20: 2023-01-12 10:59:48        1 -0.37179605  7.09093319          0.015625
#> 21: 2023-01-12 10:59:48        1 -3.33121780  7.73823277          0.015625
#> 22: 2023-01-12 10:59:48        1  6.50600859 12.05200129          0.015625
#> 23: 2023-01-12 10:59:48        1  9.84473754  4.23384557          0.015625
#> 24: 2023-01-12 10:59:48        1 -4.40398123 10.47230243          0.015625
#> 25: 2023-01-12 10:59:48        1 -1.51373763  6.78451369          0.015625
#> 26: 2023-01-12 10:59:48        1  3.55317433  7.39361558          0.015625
#> 27: 2023-01-12 10:59:48        1 -2.08069537  8.55027665          0.015625
#> 28: 2023-01-12 10:59:48        1  1.23382562  4.06840983          0.015625
#> 29: 2023-01-12 10:59:48        1  5.61461219  3.92410491          0.015625
#> 30: 2023-01-12 10:59:48        1  7.95790488  2.64839111          0.015625
#> 31: 2023-01-12 10:59:48        1  6.65190352  1.69111977          0.015625
#> 32: 2023-01-12 10:59:48        1  1.13408186  6.45866915          0.015625
#> 33: 2023-01-12 10:59:48        1 -4.02339077  8.79334246          0.015625
#> 34: 2023-01-12 10:59:48        1 -4.99201164  0.46577706          0.015625
#> 35: 2023-01-12 10:59:48        1  1.40190487  8.75897108          0.015625
#> 36: 2023-01-12 10:59:48        1 -4.60327987  3.66566344          0.015625
#> 37: 2023-01-12 10:59:48        1 -0.90199814  3.25786661          0.015625
#> 38: 2023-01-12 10:59:48        1 -0.03907316 13.13001536          0.015625
#> 39: 2023-01-12 10:59:48        1  5.30501633  4.68912311          0.015625
#> 40: 2023-01-12 10:59:48        1  5.88066715  9.29606262          0.015625
#> 41: 2023-01-12 10:59:48        1  0.52282257  9.57751273          0.015625
#> 42: 2023-01-12 10:59:48        1 -3.85509844  4.29533139          0.015625
#> 43: 2023-01-12 10:59:48        1  1.00895408  1.74451787          0.015625
#> 44: 2023-01-12 10:59:48        1  6.11301612 10.92192769          0.015625
#> 45: 2023-01-12 10:59:48        1  5.28873116 11.16190555          0.015625
#> 46: 2023-01-12 10:59:48        1  8.05983316  5.35987782          0.015625
#> 47: 2023-01-12 10:59:48        1 -2.18771053  6.75517203          0.015625
#> 48: 2023-01-12 10:59:48        1  5.02095552 11.40356344          0.015625
#> 49: 2023-01-12 10:59:48        1  1.03228546 12.98572984          0.015625
#> 50: 2023-01-12 10:59:48        1 -1.91137650  9.36517966          0.015625
#> 51: 2023-01-12 10:59:48        1 -0.63410592 13.94035759          0.015625
#> 52: 2023-01-12 10:59:48        1  8.14674120  0.53343973          0.015625
#> 53: 2023-01-12 10:59:48        1  2.11246957  9.33447266          0.015625
#> 54: 2023-01-12 10:59:48        1 -0.29947187 11.71713320          0.015625
#> 55: 2023-01-12 10:59:48        1 -0.02366642  0.01606912          0.015625
#> 56: 2023-01-12 10:59:48        1  1.58151027 10.39997638          0.015625
#> 57: 2023-01-12 10:59:48        1  9.40385479 14.79281947          0.015625
#> 58: 2023-01-12 10:59:48        1  0.20356764  7.90648273          0.015625
#> 59: 2023-01-12 10:59:48        1 -4.63031528  2.52501892          0.015625
#> 60: 2023-01-12 10:59:48        1  2.79196061  9.19695958          0.015625
#> 61: 2023-01-12 10:59:48        1  3.89955083  8.95362437          0.015625
#> 62: 2023-01-12 10:59:48        1 -2.93504045  7.67099135          0.015625
#> 63: 2023-01-12 10:59:48        1  2.16217075  5.91988833          0.015625
#> 64: 2023-01-12 10:59:48        1  7.73003042  7.64487127          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity