Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# Branin function with fidelity, see `bbotk::branin()`
fun = function(xs) branin_wu(xs[["x1"]], xs[["x2"]], xs[["fidelity"]])

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>         x1       x2 fidelity  x_domain         y
#> 1: 3.24188 1.161331 0.015625 <list[3]> 0.4461382

# best scoring evaluation
instance$result
#>         x1       x2 fidelity  x_domain         y
#> 1: 3.24188 1.161331 0.015625 <list[3]> 0.4461382

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -1.9127142 13.8667383 0.015625     0       6          1  28.9652419
#>  2:  3.2418796  1.1613310 0.015625     0       6          1   0.4461382
#>  3:  7.2808561 13.9740118 0.015625     0       6          1 336.7609156
#>  4:  8.7636937 13.5175037 0.015625     0       6          1 367.3803076
#>  5: -1.0965087  1.4001536 0.015625     0       6          1  55.1147747
#>  6:  7.3605935  6.7516918 0.015625     0       6          1 131.2010388
#>  7:  6.6891002 10.7445478 0.015625     0       6          1 215.2374697
#>  8:  4.7568829  2.9642385 0.015625     0       6          1  25.1674066
#>  9:  4.1837505  9.0235943 0.015625     0       6          1  88.7707752
#> 10:  3.2872197  6.5017702 0.015625     0       6          1  29.6735271
#> 11: -4.4980273 10.9456895 0.015625     0       6          1  15.9958996
#> 12: -2.6349579 11.7141621 0.015625     0       6          1   3.3123821
#> 13: -1.5777934  4.2369510 0.015625     0       6          1  28.8616459
#> 14:  1.9227207  1.2564172 0.015625     0       6          1   9.9198354
#> 15:  5.8803714 12.4476205 0.015625     0       6          1 236.1993123
#> 16:  2.3426470 13.7355686 0.015625     0       6          1 130.8859565
#> 17:  7.0916967  3.6492058 0.015625     0       6          1  71.2380330
#> 18:  8.0718635 13.7760361 0.015625     0       6          1 354.6096152
#> 19:  2.3911306 14.2923545 0.015625     0       6          1 145.1151855
#> 20:  9.9432313 12.3263842 0.015625     0       6          1 366.9144280
#> 21: -4.7779934  2.1199661 0.015625     0       6          1 159.1372680
#> 22:  2.6497426  1.0732072 0.015625     0       6          1   2.3926203
#> 23:  7.9549236 11.9732667 0.015625     0       6          1 287.5292510
#> 24: -2.0761236  6.9550112 0.015625     0       6          1  11.5108619
#> 25:  7.7350601  0.4643276 0.015625     0       6          1  35.4976681
#> 26:  7.6482584  3.6225021 0.015625     0       6          1  75.9056582
#> 27:  7.0674336  8.7350110 0.015625     0       6          1 171.7355446
#> 28: -4.6873845  4.7343562 0.015625     0       6          1  98.1464049
#> 29: -4.3051207  8.9903496 0.015625     0       6          1  25.8332641
#> 30: -4.6144176 13.4526235 0.015625     0       6          1   9.3590430
#> 31: -2.2106287  6.7769302 0.015625     0       6          1  12.6286020
#> 32:  5.3502649 11.8364300 0.015625     0       6          1 197.1992152
#> 33: -0.5368170  4.7526389 0.015625     0       6          1  22.7060873
#> 34: -2.3873923  6.6819799 0.015625     0       6          1  13.8451163
#> 35: -3.3375519  3.7321019 0.015625     0       6          1  63.3441111
#> 36: -4.3174764 11.3198185 0.015625     0       6          1  10.8205837
#> 37:  8.7808074  9.5299218 0.015625     0       6          1 231.3691772
#> 38:  2.1716038  8.0707616 0.015625     0       6          1  33.5375256
#> 39: -4.1120170  9.3934803 0.015625     0       6          1  18.0507219
#> 40:  8.5901146  9.4095993 0.015625     0       6          1 222.9591078
#> 41:  6.4973970  7.1572237 0.015625     0       6          1 123.4253276
#> 42: -0.7662443  8.6036133 0.015625     0       6          1  18.7845945
#> 43:  1.7864127 13.7832349 0.015625     0       6          1 118.7902494
#> 44:  3.7573638  4.0685885 0.015625     0       6          1  15.2264176
#> 45: -2.5287962  8.8119498 0.015625     0       6          1   4.1314004
#> 46:  5.7987343 14.8341166 0.015625     0       6          1 308.4912268
#> 47: -4.7797827 13.1972844 0.015625     0       6          1  11.8841371
#> 48: -3.7501140  9.0645485 0.015625     0       6          1  13.2527495
#> 49: -0.6852485  3.5062563 0.015625     0       6          1  30.3858265
#> 50:  4.2839137  7.1981046 0.015625     0       6          1  61.5412621
#> 51:  5.4186977  1.0786240 0.015625     0       6          1  24.0717621
#> 52:  5.4589243  1.1361216 0.015625     0       6          1  24.9773310
#> 53:  3.2772582  2.6746115 0.015625     0       6          1   2.9206337
#> 54:  1.5031065  3.8329300 0.015625     0       6          1  10.6737209
#> 55:  4.7971608  3.0536418 0.015625     0       6          1  26.6612713
#> 56:  5.2650850  0.7412646 0.015625     0       6          1  20.1873450
#> 57:  4.9878228 11.8934141 0.015625     0       6          1 183.3538706
#> 58: -1.9996825  2.0088232 0.015625     0       6          1  59.2489965
#> 59:  3.5747652  7.0140347 0.015625     0       6          1  41.1076144
#> 60:  6.8730865  1.7222142 0.015625     0       6          1  45.1088631
#> 61:  6.2427126  1.0646100 0.015625     0       6          1  34.0489954
#> 62: -3.8397070  5.4851104 0.015625     0       6          1  52.7605523
#> 63: -1.0052929  4.8832263 0.015625     0       6          1  22.6957222
#> 64:  9.9822898  9.4245718 0.015625     0       6          1 265.8513348
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-11-26 04:14:14        1  -1.9127142  13.8667383          0.015625
#>  2: 2022-11-26 04:14:14        1   3.2418796   1.1613310          0.015625
#>  3: 2022-11-26 04:14:14        1   7.2808561  13.9740118          0.015625
#>  4: 2022-11-26 04:14:14        1   8.7636937  13.5175037          0.015625
#>  5: 2022-11-26 04:14:14        1  -1.0965087   1.4001536          0.015625
#>  6: 2022-11-26 04:14:14        1   7.3605935   6.7516918          0.015625
#>  7: 2022-11-26 04:14:14        1   6.6891002  10.7445478          0.015625
#>  8: 2022-11-26 04:14:14        1   4.7568829   2.9642385          0.015625
#>  9: 2022-11-26 04:14:14        1   4.1837505   9.0235943          0.015625
#> 10: 2022-11-26 04:14:14        1   3.2872197   6.5017702          0.015625
#> 11: 2022-11-26 04:14:14        1  -4.4980273  10.9456895          0.015625
#> 12: 2022-11-26 04:14:14        1  -2.6349579  11.7141621          0.015625
#> 13: 2022-11-26 04:14:14        1  -1.5777934   4.2369510          0.015625
#> 14: 2022-11-26 04:14:14        1   1.9227207   1.2564172          0.015625
#> 15: 2022-11-26 04:14:14        1   5.8803714  12.4476205          0.015625
#> 16: 2022-11-26 04:14:14        1   2.3426470  13.7355686          0.015625
#> 17: 2022-11-26 04:14:14        1   7.0916967   3.6492058          0.015625
#> 18: 2022-11-26 04:14:14        1   8.0718635  13.7760361          0.015625
#> 19: 2022-11-26 04:14:14        1   2.3911306  14.2923545          0.015625
#> 20: 2022-11-26 04:14:14        1   9.9432313  12.3263842          0.015625
#> 21: 2022-11-26 04:14:14        1  -4.7779934   2.1199661          0.015625
#> 22: 2022-11-26 04:14:14        1   2.6497426   1.0732072          0.015625
#> 23: 2022-11-26 04:14:14        1   7.9549236  11.9732667          0.015625
#> 24: 2022-11-26 04:14:14        1  -2.0761236   6.9550112          0.015625
#> 25: 2022-11-26 04:14:14        1   7.7350601   0.4643276          0.015625
#> 26: 2022-11-26 04:14:14        1   7.6482584   3.6225021          0.015625
#> 27: 2022-11-26 04:14:14        1   7.0674336   8.7350110          0.015625
#> 28: 2022-11-26 04:14:14        1  -4.6873845   4.7343562          0.015625
#> 29: 2022-11-26 04:14:14        1  -4.3051207   8.9903496          0.015625
#> 30: 2022-11-26 04:14:14        1  -4.6144176  13.4526235          0.015625
#> 31: 2022-11-26 04:14:14        1  -2.2106287   6.7769302          0.015625
#> 32: 2022-11-26 04:14:14        1   5.3502649  11.8364300          0.015625
#> 33: 2022-11-26 04:14:14        1  -0.5368170   4.7526389          0.015625
#> 34: 2022-11-26 04:14:14        1  -2.3873923   6.6819799          0.015625
#> 35: 2022-11-26 04:14:14        1  -3.3375519   3.7321019          0.015625
#> 36: 2022-11-26 04:14:14        1  -4.3174764  11.3198185          0.015625
#> 37: 2022-11-26 04:14:14        1   8.7808074   9.5299218          0.015625
#> 38: 2022-11-26 04:14:14        1   2.1716038   8.0707616          0.015625
#> 39: 2022-11-26 04:14:14        1  -4.1120170   9.3934803          0.015625
#> 40: 2022-11-26 04:14:14        1   8.5901146   9.4095993          0.015625
#> 41: 2022-11-26 04:14:14        1   6.4973970   7.1572237          0.015625
#> 42: 2022-11-26 04:14:14        1  -0.7662443   8.6036133          0.015625
#> 43: 2022-11-26 04:14:14        1   1.7864127  13.7832349          0.015625
#> 44: 2022-11-26 04:14:14        1   3.7573638   4.0685885          0.015625
#> 45: 2022-11-26 04:14:14        1  -2.5287962   8.8119498          0.015625
#> 46: 2022-11-26 04:14:14        1   5.7987343  14.8341166          0.015625
#> 47: 2022-11-26 04:14:14        1  -4.7797827  13.1972844          0.015625
#> 48: 2022-11-26 04:14:14        1  -3.7501140   9.0645485          0.015625
#> 49: 2022-11-26 04:14:14        1  -0.6852485   3.5062563          0.015625
#> 50: 2022-11-26 04:14:14        1   4.2839137   7.1981046          0.015625
#> 51: 2022-11-26 04:14:14        1   5.4186977   1.0786240          0.015625
#> 52: 2022-11-26 04:14:14        1   5.4589243   1.1361216          0.015625
#> 53: 2022-11-26 04:14:14        1   3.2772582   2.6746115          0.015625
#> 54: 2022-11-26 04:14:14        1   1.5031065   3.8329300          0.015625
#> 55: 2022-11-26 04:14:14        1   4.7971608   3.0536418          0.015625
#> 56: 2022-11-26 04:14:14        1   5.2650850   0.7412646          0.015625
#> 57: 2022-11-26 04:14:14        1   4.9878228  11.8934141          0.015625
#> 58: 2022-11-26 04:14:14        1  -1.9996825   2.0088232          0.015625
#> 59: 2022-11-26 04:14:14        1   3.5747652   7.0140347          0.015625
#> 60: 2022-11-26 04:14:14        1   6.8730865   1.7222142          0.015625
#> 61: 2022-11-26 04:14:14        1   6.2427126   1.0646100          0.015625
#> 62: 2022-11-26 04:14:14        1  -3.8397070   5.4851104          0.015625
#> 63: 2022-11-26 04:14:14        1  -1.0052929   4.8832263          0.015625
#> 64: 2022-11-26 04:14:14        1   9.9822898   9.4245718          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity