Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.276448 12.02967 0.015625 <list[3]> 0.7202636

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.276448 12.02967 0.015625 <list[3]> 0.7202636

# all evaluations
as.data.table(instance$archive)
#>               x1         x2 fidelity stage bracket repetition           y
#>  1:  9.521293273  6.7473803 0.015625     0       6          1 172.4095604
#>  2: -1.402801026  4.6923590 0.015625     0       6          1  24.5711040
#>  3: -1.260345916 11.9617172 0.015625     0       6          1  28.1977686
#>  4:  9.635964069  2.7702826 0.015625     0       6          1  86.2011079
#>  5: -3.086234658 12.6835791 0.015625     0       6          1   2.5995158
#>  6: -3.541556500 13.8707887 0.015625     0       6          1   4.5729752
#>  7: -1.444503331  0.6473427 0.015625     0       6          1  70.7432153
#>  8: -1.870407417  4.6176704 0.015625     0       6          1  27.1177294
#>  9:  4.856350630 14.0558894 0.015625     0       6          1 238.1772414
#> 10:  7.632550291 12.1507412 0.015625     0       6          1 284.5942134
#> 11: -4.001028027  0.4800631 0.015625     0       6          1 156.9949149
#> 12:  6.219617152  3.1326839 0.015625     0       6          1  53.7129201
#> 13: -4.407220950 12.8077513 0.015625     0       6          1   7.7610683
#> 14:  0.365258382  5.8014418 0.015625     0       6          1  19.1120640
#> 15: -3.616806761  4.2430172 0.015625     0       6          1  64.1173096
#> 16: -1.352118343 13.6795698 0.015625     0       6          1  42.0192301
#> 17:  5.830388335  6.7153602 0.015625     0       6          1  98.7283018
#> 18: -0.832467937  6.2730423 0.015625     0       6          1  17.6144402
#> 19:  7.335565586 13.3961821 0.015625     0       6          1 318.0953160
#> 20:  5.199170543  9.0375989 0.015625     0       6          1 124.3473045
#> 21:  1.213736291 12.2822916 0.015625     0       6          1  80.0840921
#> 22:  8.304732423  7.5554537 0.015625     0       6          1 165.8968133
#> 23: -3.366280738  1.9202804 0.015625     0       6          1  96.3999993
#> 24:  4.190880174  1.8615260 0.015625     0       6          1   9.1823844
#> 25:  5.502475898  9.1147452 0.015625     0       6          1 136.5329341
#> 26: -3.270356534  5.2617014 0.015625     0       6          1  39.8163561
#> 27:  3.247972066  7.3879698 0.015625     0       6          1  39.3014319
#> 28:  2.058953734 10.7247618 0.015625     0       6          1  67.4546549
#> 29: -3.123995831 13.2245659 0.015625     0       6          1   4.2116353
#> 30:  9.309087943  2.3405773 0.015625     0       6          1  72.5751560
#> 31:  8.820593095  8.2002586 0.015625     0       6          1 193.8223361
#> 32: -0.185434493  1.3476120 0.015625     0       6          1  43.9258753
#> 33:  9.822750076 13.0108680 0.015625     0       6          1 388.3559097
#> 34:  0.008142696  4.2329526 0.015625     0       6          1  22.6786258
#> 35: -0.898347091  1.5323100 0.015625     0       6          1  51.0544458
#> 36: -3.412294097 10.3191532 0.015625     0       6          1   2.9075568
#> 37:  0.790088159  6.6522476 0.015625     0       6          1  20.3318606
#> 38:  3.204195771  5.1412685 0.015625     0       6          1  15.8241193
#> 39:  9.825722236 14.0426732 0.015625     0       6          1 430.1600772
#> 40:  8.703206658  7.6745161 0.015625     0       6          1 176.9556113
#> 41:  3.002886394  3.8054462 0.015625     0       6          1   5.8143539
#> 42: -3.276448114 12.0296723 0.015625     0       6          1   0.7202636
#> 43:  7.918036687 11.0297500 0.015625     0       6          1 256.0010887
#> 44:  5.131878257 14.2778220 0.015625     0       6          1 258.3860394
#> 45: -3.304228699  9.0184678 0.015625     0       6          1   7.1607302
#> 46:  9.155302937  6.9117073 0.015625     0       6          1 167.2995335
#> 47: -1.395702518  2.8789196 0.015625     0       6          1  40.8575891
#> 48: -2.398135861  9.7370371 0.015625     0       6          1   2.9973794
#> 49:  6.817092146  8.9166498 0.015625     0       6          1 170.4794272
#> 50:  1.823612220  1.8722480 0.015625     0       6          1   9.3608231
#> 51:  5.237635366  1.0167905 0.015625     0       6          1  21.1111807
#> 52: -2.013728716  1.3999438 0.015625     0       6          1  68.7645647
#> 53: -1.703678269  1.0804960 0.015625     0       6          1  68.3298479
#> 54:  5.512223350  8.7086253 0.015625     0       6          1 128.1339446
#> 55:  5.429297305  1.7292302 0.015625     0       6          1  28.3076529
#> 56:  5.057535924  7.1044778 0.015625     0       6          1  83.2609193
#> 57:  7.521761522 13.0727878 0.015625     0       6          1 312.5764331
#> 58:  7.781770575  1.5138971 0.015625     0       6          1  47.1388117
#> 59: -4.695786138  3.2331007 0.015625     0       6          1 129.0533348
#> 60:  6.532586121 14.0244508 0.015625     0       6          1 312.0319997
#> 61: -1.781815286  4.0464470 0.015625     0       6          1  31.8717184
#> 62:  8.140259667 12.7444595 0.015625     0       6          1 319.2584554
#> 63:  8.904175645  0.4572329 0.015625     0       6          1  39.9973707
#> 64:  8.866077660 11.2808526 0.015625     0       6          1 289.9988236
#>               x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-03 04:31:22        1  9.521293273   6.7473803          0.015625
#>  2: 2022-10-03 04:31:22        1 -1.402801026   4.6923590          0.015625
#>  3: 2022-10-03 04:31:22        1 -1.260345916  11.9617172          0.015625
#>  4: 2022-10-03 04:31:22        1  9.635964069   2.7702826          0.015625
#>  5: 2022-10-03 04:31:22        1 -3.086234658  12.6835791          0.015625
#>  6: 2022-10-03 04:31:22        1 -3.541556500  13.8707887          0.015625
#>  7: 2022-10-03 04:31:22        1 -1.444503331   0.6473427          0.015625
#>  8: 2022-10-03 04:31:22        1 -1.870407417   4.6176704          0.015625
#>  9: 2022-10-03 04:31:22        1  4.856350630  14.0558894          0.015625
#> 10: 2022-10-03 04:31:22        1  7.632550291  12.1507412          0.015625
#> 11: 2022-10-03 04:31:22        1 -4.001028027   0.4800631          0.015625
#> 12: 2022-10-03 04:31:22        1  6.219617152   3.1326839          0.015625
#> 13: 2022-10-03 04:31:22        1 -4.407220950  12.8077513          0.015625
#> 14: 2022-10-03 04:31:22        1  0.365258382   5.8014418          0.015625
#> 15: 2022-10-03 04:31:22        1 -3.616806761   4.2430172          0.015625
#> 16: 2022-10-03 04:31:22        1 -1.352118343  13.6795698          0.015625
#> 17: 2022-10-03 04:31:22        1  5.830388335   6.7153602          0.015625
#> 18: 2022-10-03 04:31:22        1 -0.832467937   6.2730423          0.015625
#> 19: 2022-10-03 04:31:22        1  7.335565586  13.3961821          0.015625
#> 20: 2022-10-03 04:31:22        1  5.199170543   9.0375989          0.015625
#> 21: 2022-10-03 04:31:22        1  1.213736291  12.2822916          0.015625
#> 22: 2022-10-03 04:31:22        1  8.304732423   7.5554537          0.015625
#> 23: 2022-10-03 04:31:22        1 -3.366280738   1.9202804          0.015625
#> 24: 2022-10-03 04:31:22        1  4.190880174   1.8615260          0.015625
#> 25: 2022-10-03 04:31:22        1  5.502475898   9.1147452          0.015625
#> 26: 2022-10-03 04:31:22        1 -3.270356534   5.2617014          0.015625
#> 27: 2022-10-03 04:31:22        1  3.247972066   7.3879698          0.015625
#> 28: 2022-10-03 04:31:22        1  2.058953734  10.7247618          0.015625
#> 29: 2022-10-03 04:31:22        1 -3.123995831  13.2245659          0.015625
#> 30: 2022-10-03 04:31:22        1  9.309087943   2.3405773          0.015625
#> 31: 2022-10-03 04:31:22        1  8.820593095   8.2002586          0.015625
#> 32: 2022-10-03 04:31:22        1 -0.185434493   1.3476120          0.015625
#> 33: 2022-10-03 04:31:22        1  9.822750076  13.0108680          0.015625
#> 34: 2022-10-03 04:31:22        1  0.008142696   4.2329526          0.015625
#> 35: 2022-10-03 04:31:22        1 -0.898347091   1.5323100          0.015625
#> 36: 2022-10-03 04:31:22        1 -3.412294097  10.3191532          0.015625
#> 37: 2022-10-03 04:31:22        1  0.790088159   6.6522476          0.015625
#> 38: 2022-10-03 04:31:22        1  3.204195771   5.1412685          0.015625
#> 39: 2022-10-03 04:31:22        1  9.825722236  14.0426732          0.015625
#> 40: 2022-10-03 04:31:22        1  8.703206658   7.6745161          0.015625
#> 41: 2022-10-03 04:31:22        1  3.002886394   3.8054462          0.015625
#> 42: 2022-10-03 04:31:22        1 -3.276448114  12.0296723          0.015625
#> 43: 2022-10-03 04:31:22        1  7.918036687  11.0297500          0.015625
#> 44: 2022-10-03 04:31:22        1  5.131878257  14.2778220          0.015625
#> 45: 2022-10-03 04:31:22        1 -3.304228699   9.0184678          0.015625
#> 46: 2022-10-03 04:31:22        1  9.155302937   6.9117073          0.015625
#> 47: 2022-10-03 04:31:22        1 -1.395702518   2.8789196          0.015625
#> 48: 2022-10-03 04:31:22        1 -2.398135861   9.7370371          0.015625
#> 49: 2022-10-03 04:31:22        1  6.817092146   8.9166498          0.015625
#> 50: 2022-10-03 04:31:22        1  1.823612220   1.8722480          0.015625
#> 51: 2022-10-03 04:31:22        1  5.237635366   1.0167905          0.015625
#> 52: 2022-10-03 04:31:22        1 -2.013728716   1.3999438          0.015625
#> 53: 2022-10-03 04:31:22        1 -1.703678269   1.0804960          0.015625
#> 54: 2022-10-03 04:31:22        1  5.512223350   8.7086253          0.015625
#> 55: 2022-10-03 04:31:22        1  5.429297305   1.7292302          0.015625
#> 56: 2022-10-03 04:31:22        1  5.057535924   7.1044778          0.015625
#> 57: 2022-10-03 04:31:22        1  7.521761522  13.0727878          0.015625
#> 58: 2022-10-03 04:31:22        1  7.781770575   1.5138971          0.015625
#> 59: 2022-10-03 04:31:22        1 -4.695786138   3.2331007          0.015625
#> 60: 2022-10-03 04:31:22        1  6.532586121  14.0244508          0.015625
#> 61: 2022-10-03 04:31:22        1 -1.781815286   4.0464470          0.015625
#> 62: 2022-10-03 04:31:22        1  8.140259667  12.7444595          0.015625
#> 63: 2022-10-03 04:31:22        1  8.904175645   0.4572329          0.015625
#> 64: 2022-10-03 04:31:22        1  8.866077660  11.2808526          0.015625
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity