Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.141068 1.069783 0.015625 <list[3]> 0.4528374

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.141068 1.069783 0.015625 <list[3]> 0.4528374

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition           y
#>  1: -0.43673607  0.04746670 0.015625     0       6          1  62.9696925
#>  2: -2.51917123  9.99971310 0.015625     0       6          1   2.2405228
#>  3:  3.80144176  5.88396926 0.015625     0       6          1  32.5517283
#>  4:  0.46104006  5.89143431 0.015625     0       6          1  18.9823074
#>  5:  6.55734034 13.79896648 0.015625     0       6          1 305.3003894
#>  6: -0.03637844  6.29763302 0.015625     0       6          1  19.6532130
#>  7:  5.90506521 14.12131299 0.015625     0       6          1 289.4399420
#>  8: -3.71797823 13.83368977 0.015625     0       6          1   4.1732493
#>  9:  7.34742996  8.89502102 0.015625     0       6          1 181.8163581
#> 10: -1.19124405  9.37174116 0.015625     0       6          1  15.6087825
#> 11: -2.34738767 11.60134933 0.015625     0       6          1   6.1465612
#> 12: -4.92146129 12.63819184 0.015625     0       6          1  15.7537248
#> 13:  9.10361154 10.18156243 0.015625     0       6          1 260.8151309
#> 14:  2.90020096  7.22585262 0.015625     0       6          1  31.8467097
#> 15:  6.85702962 10.19314643 0.015625     0       6          1 204.6803808
#> 16:  2.42530125 10.70809712 0.015625     0       6          1  73.1032288
#> 17:  6.74428896 11.36830623 0.015625     0       6          1 234.7961529
#> 18:  0.27766556  7.43418835 0.015625     0       6          1  22.7452197
#> 19:  2.97663690  9.96590845 0.015625     0       6          1  71.6090370
#> 20: -1.03353799  6.77853041 0.015625     0       6          1  15.7228266
#> 21:  3.37066767 12.22713633 0.015625     0       6          1 127.0400903
#> 22:  9.64698973  7.79894730 0.015625     0       6          1 204.8711207
#> 23:  5.27965722 14.38046727 0.015625     0       6          1 268.8045166
#> 24:  9.01267562  6.93930235 0.015625     0       6          1 164.6808323
#> 25:  2.92397815  7.61800648 0.015625     0       6          1  36.7298614
#> 26: -3.72486864  2.70186007 0.015625     0       6          1  95.1669706
#> 27: -0.41145530 14.30823048 0.015625     0       6          1  77.2952791
#> 28: -3.90276396  0.06284585 0.015625     0       6          1 162.2344922
#> 29:  3.98794499  1.43300667 0.015625     0       6          1   5.3032463
#> 30: -3.54904817 13.10674684 0.015625     0       6          1   2.3309899
#> 31: -3.19810711  5.83540393 0.015625     0       6          1  31.4271857
#> 32:  7.88704562  3.71032770 0.015625     0       6          1  79.4103835
#> 33:  7.26364349 14.43558684 0.015625     0       6          1 352.9417008
#> 34: -2.20630934  2.73289818 0.015625     0       6          1  52.3005448
#> 35:  1.14531151  3.18702013 0.015625     0       6          1  15.0253050
#> 36:  7.83935485 10.07553017 0.015625     0       6          1 225.1345860
#> 37:  9.61869089  9.58363961 0.015625     0       6          1 258.1024354
#> 38:  8.39834225 14.67182348 0.015625     0       6          1 399.8262626
#> 39:  3.63755437  1.28365837 0.015625     0       6          1   1.9986112
#> 40:  8.18305135  3.88616829 0.015625     0       6          1  85.2372909
#> 41: -0.53414136 11.68058823 0.015625     0       6          1  41.5134215
#> 42: -1.81426068  2.46600437 0.015625     0       6          1  50.2306905
#> 43:  3.14106796  1.06978306 0.015625     0       6          1   0.4528374
#> 44: -4.43616822 11.84572080 0.015625     0       6          1  10.6927815
#> 45:  2.55662328  0.99746216 0.015625     0       6          1   3.2815711
#> 46: -1.50129418  1.81030317 0.015625     0       6          1  54.8677992
#> 47:  5.66748830 10.18878480 0.015625     0       6          1 167.1982205
#> 48: -2.22444627  8.34709836 0.015625     0       6          1   5.9710611
#> 49:  9.48817183  0.40767091 0.015625     0       6          1  45.8522361
#> 50:  8.43420121  6.61861241 0.015625     0       6          1 145.2734935
#> 51: -0.55851257  5.17961163 0.015625     0       6          1  21.0975646
#> 52: -0.83135833  4.55486344 0.015625     0       6          1  24.2521103
#> 53:  7.72867074  1.84130946 0.015625     0       6          1  50.9566564
#> 54:  3.35509248 13.14132685 0.015625     0       6          1 147.8744694
#> 55:  7.66408988  7.68913300 0.015625     0       6          1 157.7601975
#> 56: -1.02022580  2.53953155 0.015625     0       6          1  41.1991923
#> 57: -3.10306781 11.66488691 0.015625     0       6          1   0.5900301
#> 58: -3.25731187  9.30730288 0.015625     0       6          1   5.3157480
#> 59:  7.76446507 14.03582842 0.015625     0       6          1 354.5794690
#> 60: -0.99257916  6.68735852 0.015625     0       6          1  16.0991846
#> 61:  7.44341900 14.95510972 0.015625     0       6          1 378.5720367
#> 62:  0.16839319  1.77678557 0.015625     0       6          1  35.1168646
#> 63:  8.46196725  8.62150385 0.015625     0       6          1 197.3779726
#> 64: -2.69437839  1.83223611 0.015625     0       6          1  76.6709525
#>              x1          x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-05 04:29:22        1 -0.43673607  0.04746670          0.015625
#>  2: 2022-10-05 04:29:22        1 -2.51917123  9.99971310          0.015625
#>  3: 2022-10-05 04:29:22        1  3.80144176  5.88396926          0.015625
#>  4: 2022-10-05 04:29:22        1  0.46104006  5.89143431          0.015625
#>  5: 2022-10-05 04:29:22        1  6.55734034 13.79896648          0.015625
#>  6: 2022-10-05 04:29:22        1 -0.03637844  6.29763302          0.015625
#>  7: 2022-10-05 04:29:22        1  5.90506521 14.12131299          0.015625
#>  8: 2022-10-05 04:29:22        1 -3.71797823 13.83368977          0.015625
#>  9: 2022-10-05 04:29:22        1  7.34742996  8.89502102          0.015625
#> 10: 2022-10-05 04:29:22        1 -1.19124405  9.37174116          0.015625
#> 11: 2022-10-05 04:29:22        1 -2.34738767 11.60134933          0.015625
#> 12: 2022-10-05 04:29:22        1 -4.92146129 12.63819184          0.015625
#> 13: 2022-10-05 04:29:22        1  9.10361154 10.18156243          0.015625
#> 14: 2022-10-05 04:29:22        1  2.90020096  7.22585262          0.015625
#> 15: 2022-10-05 04:29:22        1  6.85702962 10.19314643          0.015625
#> 16: 2022-10-05 04:29:22        1  2.42530125 10.70809712          0.015625
#> 17: 2022-10-05 04:29:22        1  6.74428896 11.36830623          0.015625
#> 18: 2022-10-05 04:29:22        1  0.27766556  7.43418835          0.015625
#> 19: 2022-10-05 04:29:22        1  2.97663690  9.96590845          0.015625
#> 20: 2022-10-05 04:29:22        1 -1.03353799  6.77853041          0.015625
#> 21: 2022-10-05 04:29:22        1  3.37066767 12.22713633          0.015625
#> 22: 2022-10-05 04:29:22        1  9.64698973  7.79894730          0.015625
#> 23: 2022-10-05 04:29:22        1  5.27965722 14.38046727          0.015625
#> 24: 2022-10-05 04:29:22        1  9.01267562  6.93930235          0.015625
#> 25: 2022-10-05 04:29:22        1  2.92397815  7.61800648          0.015625
#> 26: 2022-10-05 04:29:22        1 -3.72486864  2.70186007          0.015625
#> 27: 2022-10-05 04:29:22        1 -0.41145530 14.30823048          0.015625
#> 28: 2022-10-05 04:29:22        1 -3.90276396  0.06284585          0.015625
#> 29: 2022-10-05 04:29:22        1  3.98794499  1.43300667          0.015625
#> 30: 2022-10-05 04:29:22        1 -3.54904817 13.10674684          0.015625
#> 31: 2022-10-05 04:29:22        1 -3.19810711  5.83540393          0.015625
#> 32: 2022-10-05 04:29:22        1  7.88704562  3.71032770          0.015625
#> 33: 2022-10-05 04:29:22        1  7.26364349 14.43558684          0.015625
#> 34: 2022-10-05 04:29:22        1 -2.20630934  2.73289818          0.015625
#> 35: 2022-10-05 04:29:22        1  1.14531151  3.18702013          0.015625
#> 36: 2022-10-05 04:29:22        1  7.83935485 10.07553017          0.015625
#> 37: 2022-10-05 04:29:22        1  9.61869089  9.58363961          0.015625
#> 38: 2022-10-05 04:29:22        1  8.39834225 14.67182348          0.015625
#> 39: 2022-10-05 04:29:22        1  3.63755437  1.28365837          0.015625
#> 40: 2022-10-05 04:29:22        1  8.18305135  3.88616829          0.015625
#> 41: 2022-10-05 04:29:22        1 -0.53414136 11.68058823          0.015625
#> 42: 2022-10-05 04:29:22        1 -1.81426068  2.46600437          0.015625
#> 43: 2022-10-05 04:29:22        1  3.14106796  1.06978306          0.015625
#> 44: 2022-10-05 04:29:22        1 -4.43616822 11.84572080          0.015625
#> 45: 2022-10-05 04:29:22        1  2.55662328  0.99746216          0.015625
#> 46: 2022-10-05 04:29:22        1 -1.50129418  1.81030317          0.015625
#> 47: 2022-10-05 04:29:22        1  5.66748830 10.18878480          0.015625
#> 48: 2022-10-05 04:29:22        1 -2.22444627  8.34709836          0.015625
#> 49: 2022-10-05 04:29:22        1  9.48817183  0.40767091          0.015625
#> 50: 2022-10-05 04:29:22        1  8.43420121  6.61861241          0.015625
#> 51: 2022-10-05 04:29:22        1 -0.55851257  5.17961163          0.015625
#> 52: 2022-10-05 04:29:22        1 -0.83135833  4.55486344          0.015625
#> 53: 2022-10-05 04:29:22        1  7.72867074  1.84130946          0.015625
#> 54: 2022-10-05 04:29:22        1  3.35509248 13.14132685          0.015625
#> 55: 2022-10-05 04:29:22        1  7.66408988  7.68913300          0.015625
#> 56: 2022-10-05 04:29:22        1 -1.02022580  2.53953155          0.015625
#> 57: 2022-10-05 04:29:22        1 -3.10306781 11.66488691          0.015625
#> 58: 2022-10-05 04:29:22        1 -3.25731187  9.30730288          0.015625
#> 59: 2022-10-05 04:29:22        1  7.76446507 14.03582842          0.015625
#> 60: 2022-10-05 04:29:22        1 -0.99257916  6.68735852          0.015625
#> 61: 2022-10-05 04:29:22        1  7.44341900 14.95510972          0.015625
#> 62: 2022-10-05 04:29:22        1  0.16839319  1.77678557          0.015625
#> 63: 2022-10-05 04:29:22        1  8.46196725  8.62150385          0.015625
#> 64: 2022-10-05 04:29:22        1 -2.69437839  1.83223611          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity