Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# Branin function with fidelity, see `bbotk::branin()`
fun = function(xs) branin_wu(xs[["x1"]], xs[["x2"]], xs[["fidelity"]])

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.628251 3.844061 0.015625 <list[3]> 4.928514

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.628251 3.844061 0.015625 <list[3]> 4.928514

# all evaluations
as.data.table(instance$archive)
#>             x1           x2 fidelity stage bracket repetition          y
#>  1:  2.8269008 14.699241497 0.015625     0       6          1 168.641433
#>  2:  5.5400676  3.659677848 0.015625     0       6          1  47.687730
#>  3:  6.4123808  1.361883815 0.015625     0       6          1  38.039876
#>  4:  1.4961299  0.009480813 0.015625     0       6          1  24.245287
#>  5:  2.2607158 12.287017156 0.015625     0       6          1  98.520829
#>  6:  2.4163902  4.123363694 0.015625     0       6          1   6.016918
#>  7: -1.9226525 14.714705350 0.015625     0       6          1  37.393950
#>  8: -3.4791748  4.484011101 0.015625     0       6          1  56.077149
#>  9: -2.3361855  4.929254375 0.015625     0       6          1  27.916463
#> 10:  6.4447438  0.339133468 0.015625     0       6          1  30.494116
#> 11:  3.4661265  4.170587438 0.015625     0       6          1  11.906279
#> 12:  8.4380440 10.459916128 0.015625     0       6          1 251.203868
#> 13:  9.6613648  3.329326540 0.015625     0       6          1  97.409898
#> 14:  5.8044098  6.294532494 0.015625     0       6          1  90.715251
#> 15: -4.3232968  6.722821528 0.015625     0       6          1  51.685465
#> 16: -1.3732614  5.912146317 0.015625     0       6          1  17.320106
#> 17:  3.5640238  5.739135335 0.015625     0       6          1  26.451385
#> 18: -2.1141952  8.930449826 0.015625     0       6          1   5.362234
#> 19: -4.2603434 13.569561567 0.015625     0       6          1   5.859068
#> 20:  2.6232766  5.583928601 0.015625     0       6          1  14.243233
#> 21: -3.0337208  3.130234476 0.015625     0       6          1  64.151036
#> 22:  1.3746738  8.072946633 0.015625     0       6          1  29.533866
#> 23: -2.1603800  6.450406403 0.015625     0       6          1  14.467053
#> 24: -3.9270164  5.857516275 0.015625     0       6          1  50.361861
#> 25:  1.8282175 13.372885636 0.015625     0       6          1 111.184076
#> 26: -0.0286592 13.465650505 0.015625     0       6          1  74.654758
#> 27:  2.9335615  7.346481183 0.015625     0       6          1  33.676468
#> 28: -2.1278737 13.722828784 0.015625     0       6          1  22.538055
#> 29: -1.8788135 13.377353864 0.015625     0       6          1  25.395299
#> 30:  4.9379735  9.593277440 0.015625     0       6          1 126.693069
#> 31:  5.6697087 11.312236539 0.015625     0       6          1 196.006331
#> 32:  5.4546288  1.109171938 0.015625     0       6          1  24.759923
#> 33:  6.5003870  9.124695832 0.015625     0       6          1 167.513817
#> 34:  1.7845942 13.537728771 0.015625     0       6          1 113.642662
#> 35:  4.1077551  8.207868022 0.015625     0       6          1  72.220974
#> 36:  1.4425173  6.364527466 0.015625     0       6          1  17.969589
#> 37:  9.2686107  0.349110186 0.015625     0       6          1  42.235674
#> 38:  9.9863815  2.449106304 0.015625     0       6          1  87.927948
#> 39: -0.3667847 11.939738696 0.015625     0       6          1  47.605689
#> 40:  9.5906124  2.425436365 0.015625     0       6          1  79.051635
#> 41:  1.9043226  1.906654666 0.015625     0       6          1   8.234822
#> 42:  2.1120327  0.371416323 0.015625     0       6          1  10.833841
#> 43: -1.1303725 10.202991829 0.015625     0       6          1  19.685224
#> 44: -3.4463588  9.176500957 0.015625     0       6          1   7.989285
#> 45:  0.3349995  5.761007478 0.015625     0       6          1  19.152857
#> 46: -4.7003295 14.715061253 0.015625     0       6          1  10.192185
#> 47:  7.3053715  1.905382234 0.015625     0       6          1  49.715192
#> 48: -0.5613532  7.268492606 0.015625     0       6          1  18.262034
#> 49:  2.9446318 14.645413198 0.015625     0       6          1 171.286578
#> 50:  6.4992242 12.777815396 0.015625     0       6          1 269.743098
#> 51:  2.6282506  3.844060746 0.015625     0       6          1   4.928514
#> 52: -2.0511593  1.351953149 0.015625     0       6          1  70.235471
#> 53:  7.8425336 12.692452418 0.015625     0       6          1 308.816411
#> 54:  9.5028346  7.425921854 0.015625     0       6          1 190.138606
#> 55: -0.7812058 13.183817687 0.015625     0       6          1  51.884958
#> 56:  5.8550244  8.905952963 0.015625     0       6          1 143.514649
#> 57: -4.6486152  3.833078505 0.015625     0       6          1 114.037957
#> 58:  3.0152520  5.577168093 0.015625     0       6          1  17.256115
#> 59:  5.8477586  8.310859414 0.015625     0       6          1 130.355435
#> 60:  6.5356623  3.921724381 0.015625     0       6          1  68.440585
#> 61: -2.6184449  5.713497922 0.015625     0       6          1  23.441549
#> 62: -0.4811158  4.505907559 0.015625     0       6          1  23.651040
#> 63:  5.2895369  2.891760607 0.015625     0       6          1  35.042144
#> 64: -2.5559288 12.879662466 0.015625     0       6          1   8.814987
#>             x1           x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1  x_domain_x2 x_domain_fidelity
#>  1: 2022-11-21 04:16:58        1   2.8269008 14.699241497          0.015625
#>  2: 2022-11-21 04:16:58        1   5.5400676  3.659677848          0.015625
#>  3: 2022-11-21 04:16:58        1   6.4123808  1.361883815          0.015625
#>  4: 2022-11-21 04:16:58        1   1.4961299  0.009480813          0.015625
#>  5: 2022-11-21 04:16:58        1   2.2607158 12.287017156          0.015625
#>  6: 2022-11-21 04:16:58        1   2.4163902  4.123363694          0.015625
#>  7: 2022-11-21 04:16:58        1  -1.9226525 14.714705350          0.015625
#>  8: 2022-11-21 04:16:58        1  -3.4791748  4.484011101          0.015625
#>  9: 2022-11-21 04:16:58        1  -2.3361855  4.929254375          0.015625
#> 10: 2022-11-21 04:16:58        1   6.4447438  0.339133468          0.015625
#> 11: 2022-11-21 04:16:58        1   3.4661265  4.170587438          0.015625
#> 12: 2022-11-21 04:16:58        1   8.4380440 10.459916128          0.015625
#> 13: 2022-11-21 04:16:58        1   9.6613648  3.329326540          0.015625
#> 14: 2022-11-21 04:16:58        1   5.8044098  6.294532494          0.015625
#> 15: 2022-11-21 04:16:58        1  -4.3232968  6.722821528          0.015625
#> 16: 2022-11-21 04:16:58        1  -1.3732614  5.912146317          0.015625
#> 17: 2022-11-21 04:16:58        1   3.5640238  5.739135335          0.015625
#> 18: 2022-11-21 04:16:58        1  -2.1141952  8.930449826          0.015625
#> 19: 2022-11-21 04:16:58        1  -4.2603434 13.569561567          0.015625
#> 20: 2022-11-21 04:16:58        1   2.6232766  5.583928601          0.015625
#> 21: 2022-11-21 04:16:58        1  -3.0337208  3.130234476          0.015625
#> 22: 2022-11-21 04:16:58        1   1.3746738  8.072946633          0.015625
#> 23: 2022-11-21 04:16:58        1  -2.1603800  6.450406403          0.015625
#> 24: 2022-11-21 04:16:58        1  -3.9270164  5.857516275          0.015625
#> 25: 2022-11-21 04:16:58        1   1.8282175 13.372885636          0.015625
#> 26: 2022-11-21 04:16:58        1  -0.0286592 13.465650505          0.015625
#> 27: 2022-11-21 04:16:58        1   2.9335615  7.346481183          0.015625
#> 28: 2022-11-21 04:16:58        1  -2.1278737 13.722828784          0.015625
#> 29: 2022-11-21 04:16:58        1  -1.8788135 13.377353864          0.015625
#> 30: 2022-11-21 04:16:58        1   4.9379735  9.593277440          0.015625
#> 31: 2022-11-21 04:16:58        1   5.6697087 11.312236539          0.015625
#> 32: 2022-11-21 04:16:58        1   5.4546288  1.109171938          0.015625
#> 33: 2022-11-21 04:16:58        1   6.5003870  9.124695832          0.015625
#> 34: 2022-11-21 04:16:58        1   1.7845942 13.537728771          0.015625
#> 35: 2022-11-21 04:16:58        1   4.1077551  8.207868022          0.015625
#> 36: 2022-11-21 04:16:58        1   1.4425173  6.364527466          0.015625
#> 37: 2022-11-21 04:16:58        1   9.2686107  0.349110186          0.015625
#> 38: 2022-11-21 04:16:58        1   9.9863815  2.449106304          0.015625
#> 39: 2022-11-21 04:16:58        1  -0.3667847 11.939738696          0.015625
#> 40: 2022-11-21 04:16:58        1   9.5906124  2.425436365          0.015625
#> 41: 2022-11-21 04:16:58        1   1.9043226  1.906654666          0.015625
#> 42: 2022-11-21 04:16:58        1   2.1120327  0.371416323          0.015625
#> 43: 2022-11-21 04:16:58        1  -1.1303725 10.202991829          0.015625
#> 44: 2022-11-21 04:16:58        1  -3.4463588  9.176500957          0.015625
#> 45: 2022-11-21 04:16:58        1   0.3349995  5.761007478          0.015625
#> 46: 2022-11-21 04:16:58        1  -4.7003295 14.715061253          0.015625
#> 47: 2022-11-21 04:16:58        1   7.3053715  1.905382234          0.015625
#> 48: 2022-11-21 04:16:58        1  -0.5613532  7.268492606          0.015625
#> 49: 2022-11-21 04:16:58        1   2.9446318 14.645413198          0.015625
#> 50: 2022-11-21 04:16:58        1   6.4992242 12.777815396          0.015625
#> 51: 2022-11-21 04:16:58        1   2.6282506  3.844060746          0.015625
#> 52: 2022-11-21 04:16:58        1  -2.0511593  1.351953149          0.015625
#> 53: 2022-11-21 04:16:58        1   7.8425336 12.692452418          0.015625
#> 54: 2022-11-21 04:16:58        1   9.5028346  7.425921854          0.015625
#> 55: 2022-11-21 04:16:58        1  -0.7812058 13.183817687          0.015625
#> 56: 2022-11-21 04:16:58        1   5.8550244  8.905952963          0.015625
#> 57: 2022-11-21 04:16:58        1  -4.6486152  3.833078505          0.015625
#> 58: 2022-11-21 04:16:58        1   3.0152520  5.577168093          0.015625
#> 59: 2022-11-21 04:16:58        1   5.8477586  8.310859414          0.015625
#> 60: 2022-11-21 04:16:58        1   6.5356623  3.921724381          0.015625
#> 61: 2022-11-21 04:16:58        1  -2.6184449  5.713497922          0.015625
#> 62: 2022-11-21 04:16:58        1  -0.4811158  4.505907559          0.015625
#> 63: 2022-11-21 04:16:58        1   5.2895369  2.891760607          0.015625
#> 64: 2022-11-21 04:16:58        1  -2.5559288 12.879662466          0.015625
#>               timestamp batch_nr x_domain_x1  x_domain_x2 x_domain_fidelity