Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# Branin function with fidelity, see `bbotk::branin()`
fun = function(xs) branin_wu(xs[["x1"]], xs[["x2"]], xs[["fidelity"]])

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 3.700593 1.575106 0.015625 <list[3]> 2.948822

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 3.700593 1.575106 0.015625 <list[3]> 2.948822

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1: -4.9854506  2.5335175 0.015625     0       6          1 160.583585
#>  2:  9.5140744  1.4676234 0.015625     0       6          1  61.691769
#>  3:  6.0541631  5.5238117 0.015625     0       6          1  83.869953
#>  4:  7.2912107 11.5328377 0.015625     0       6          1 255.453486
#>  5:  7.5689692 14.6489551 0.015625     0       6          1 371.191227
#>  6: -3.0386899  2.8916360 0.015625     0       6          1  68.156826
#>  7:  2.7210019 14.6479344 0.015625     0       6          1 163.820109
#>  8:  5.2807375  5.2242384 0.015625     0       6          1  61.019785
#>  9: -0.1697522  3.6819792 0.015625     0       6          1  26.167411
#> 10:  7.1071365 10.5452845 0.015625     0       6          1 221.115071
#> 11:  2.9814585  6.7539769 0.015625     0       6          1  27.829749
#> 12:  5.4676647  5.2949880 0.015625     0       6          1  66.678161
#> 13: -4.6290735 12.1938606 0.015625     0       6          1  12.558594
#> 14:  5.4284996  1.7301547 0.015625     0       6          1  28.301331
#> 15:  4.2231581  7.0067928 0.015625     0       6          1  57.036794
#> 16:  4.6910068  9.1371055 0.015625     0       6          1 108.329509
#> 17:  0.8336023  2.6017286 0.015625     0       6          1  20.834972
#> 18:  6.1405462  4.7245126 0.015625     0       6          1  73.352885
#> 19: -3.3377705  3.4757621 0.015625     0       6          1  67.478242
#> 20:  2.9394275 13.1009996 0.015625     0       6          1 133.156023
#> 21:  0.2561653  2.4129422 0.015625     0       6          1  29.409934
#> 22:  9.1255999  6.4309408 0.015625     0       6          1 154.442562
#> 23:  0.8897992  9.5961424 0.015625     0       6          1  40.924909
#> 24: -3.5206745  5.2824413 0.015625     0       6          1  45.996395
#> 25: -1.0585228 11.2125906 0.015625     0       6          1  26.910753
#> 26:  2.6644697  4.0323613 0.015625     0       6          1   5.692091
#> 27: -4.0835219  6.2875997 0.015625     0       6          1  49.567140
#> 28:  5.1389473 14.5767839 0.015625     0       6          1 268.173757
#> 29:  8.2516777  6.0543683 0.015625     0       6          1 129.352475
#> 30:  0.2715060  5.5364890 0.015625     0       6          1  19.251502
#> 31:  8.2086242 14.0315276 0.015625     0       6          1 368.584996
#> 32:  4.0446203 13.4016246 0.015625     0       6          1 181.898894
#> 33: -1.9097352  1.6016297 0.015625     0       6          1  63.809102
#> 34:  3.5008259  3.6248834 0.015625     0       6          1   8.962050
#> 35: -0.5006424  4.0186742 0.015625     0       6          1  26.184533
#> 36:  9.9888964 11.2356960 0.015625     0       6          1 328.253050
#> 37: -1.6844315 11.1249905 0.015625     0       6          1  14.466167
#> 38:  2.0867073  7.2625243 0.015625     0       6          1  25.063191
#> 39:  1.8978607  4.7981784 0.015625     0       6          1   9.832347
#> 40: -3.6010250  7.7067563 0.015625     0       6          1  20.957940
#> 41: -0.6663864  9.1537019 0.015625     0       6          1  21.872000
#> 42:  4.4573042 11.3780542 0.015625     0       6          1 148.265333
#> 43:  9.3488264  6.7073679 0.015625     0       6          1 166.814455
#> 44: -1.3234792 11.2368169 0.015625     0       6          1  21.815963
#> 45:  3.0922051 13.1096031 0.015625     0       6          1 138.166896
#> 46:  7.6118113 14.3971534 0.015625     0       6          1 363.125167
#> 47:  7.0484169  0.2891178 0.015625     0       6          1  32.761712
#> 48:  2.5034680  8.7135509 0.015625     0       6          1  44.605614
#> 49:  9.3519365 10.6216659 0.015625     0       6          1 283.222769
#> 50: -4.2198326 10.2224042 0.015625     0       6          1  14.708127
#> 51:  0.5122780 12.1588305 0.015625     0       6          1  66.895725
#> 52:  0.9531587  4.1839675 0.015625     0       6          1  15.667595
#> 53: -0.1176450 12.7903558 0.015625     0       6          1  63.131287
#> 54: -1.2749295 11.7845142 0.015625     0       6          1  26.529842
#> 55:  0.9590204  0.4748047 0.015625     0       6          1  31.732609
#> 56: -4.4784605 10.3357322 0.015625     0       6          1  19.393048
#> 57:  7.5303068  9.7463952 0.015625     0       6          1 208.710445
#> 58:  3.7005933  1.5751057 0.015625     0       6          1   2.948822
#> 59: -1.6446798  0.2802925 0.015625     0       6          1  80.195465
#> 60: -0.8074505 13.6282601 0.015625     0       6          1  56.620152
#> 61: -0.4530248  1.6649138 0.015625     0       6          1  44.261495
#> 62:  7.4112853 10.3326703 0.015625     0       6          1 222.605315
#> 63:  8.9769916 14.8233785 0.015625     0       6          1 427.061732
#> 64: -3.7974972  3.9117019 0.015625     0       6          1  75.931354
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-11-11 04:22:32        1  -4.9854506   2.5335175          0.015625
#>  2: 2022-11-11 04:22:32        1   9.5140744   1.4676234          0.015625
#>  3: 2022-11-11 04:22:32        1   6.0541631   5.5238117          0.015625
#>  4: 2022-11-11 04:22:32        1   7.2912107  11.5328377          0.015625
#>  5: 2022-11-11 04:22:32        1   7.5689692  14.6489551          0.015625
#>  6: 2022-11-11 04:22:32        1  -3.0386899   2.8916360          0.015625
#>  7: 2022-11-11 04:22:32        1   2.7210019  14.6479344          0.015625
#>  8: 2022-11-11 04:22:32        1   5.2807375   5.2242384          0.015625
#>  9: 2022-11-11 04:22:32        1  -0.1697522   3.6819792          0.015625
#> 10: 2022-11-11 04:22:32        1   7.1071365  10.5452845          0.015625
#> 11: 2022-11-11 04:22:32        1   2.9814585   6.7539769          0.015625
#> 12: 2022-11-11 04:22:32        1   5.4676647   5.2949880          0.015625
#> 13: 2022-11-11 04:22:32        1  -4.6290735  12.1938606          0.015625
#> 14: 2022-11-11 04:22:32        1   5.4284996   1.7301547          0.015625
#> 15: 2022-11-11 04:22:32        1   4.2231581   7.0067928          0.015625
#> 16: 2022-11-11 04:22:32        1   4.6910068   9.1371055          0.015625
#> 17: 2022-11-11 04:22:32        1   0.8336023   2.6017286          0.015625
#> 18: 2022-11-11 04:22:32        1   6.1405462   4.7245126          0.015625
#> 19: 2022-11-11 04:22:32        1  -3.3377705   3.4757621          0.015625
#> 20: 2022-11-11 04:22:32        1   2.9394275  13.1009996          0.015625
#> 21: 2022-11-11 04:22:32        1   0.2561653   2.4129422          0.015625
#> 22: 2022-11-11 04:22:32        1   9.1255999   6.4309408          0.015625
#> 23: 2022-11-11 04:22:32        1   0.8897992   9.5961424          0.015625
#> 24: 2022-11-11 04:22:32        1  -3.5206745   5.2824413          0.015625
#> 25: 2022-11-11 04:22:32        1  -1.0585228  11.2125906          0.015625
#> 26: 2022-11-11 04:22:32        1   2.6644697   4.0323613          0.015625
#> 27: 2022-11-11 04:22:32        1  -4.0835219   6.2875997          0.015625
#> 28: 2022-11-11 04:22:32        1   5.1389473  14.5767839          0.015625
#> 29: 2022-11-11 04:22:32        1   8.2516777   6.0543683          0.015625
#> 30: 2022-11-11 04:22:32        1   0.2715060   5.5364890          0.015625
#> 31: 2022-11-11 04:22:32        1   8.2086242  14.0315276          0.015625
#> 32: 2022-11-11 04:22:32        1   4.0446203  13.4016246          0.015625
#> 33: 2022-11-11 04:22:32        1  -1.9097352   1.6016297          0.015625
#> 34: 2022-11-11 04:22:32        1   3.5008259   3.6248834          0.015625
#> 35: 2022-11-11 04:22:32        1  -0.5006424   4.0186742          0.015625
#> 36: 2022-11-11 04:22:32        1   9.9888964  11.2356960          0.015625
#> 37: 2022-11-11 04:22:32        1  -1.6844315  11.1249905          0.015625
#> 38: 2022-11-11 04:22:32        1   2.0867073   7.2625243          0.015625
#> 39: 2022-11-11 04:22:32        1   1.8978607   4.7981784          0.015625
#> 40: 2022-11-11 04:22:32        1  -3.6010250   7.7067563          0.015625
#> 41: 2022-11-11 04:22:32        1  -0.6663864   9.1537019          0.015625
#> 42: 2022-11-11 04:22:32        1   4.4573042  11.3780542          0.015625
#> 43: 2022-11-11 04:22:32        1   9.3488264   6.7073679          0.015625
#> 44: 2022-11-11 04:22:32        1  -1.3234792  11.2368169          0.015625
#> 45: 2022-11-11 04:22:32        1   3.0922051  13.1096031          0.015625
#> 46: 2022-11-11 04:22:32        1   7.6118113  14.3971534          0.015625
#> 47: 2022-11-11 04:22:32        1   7.0484169   0.2891178          0.015625
#> 48: 2022-11-11 04:22:32        1   2.5034680   8.7135509          0.015625
#> 49: 2022-11-11 04:22:32        1   9.3519365  10.6216659          0.015625
#> 50: 2022-11-11 04:22:32        1  -4.2198326  10.2224042          0.015625
#> 51: 2022-11-11 04:22:32        1   0.5122780  12.1588305          0.015625
#> 52: 2022-11-11 04:22:32        1   0.9531587   4.1839675          0.015625
#> 53: 2022-11-11 04:22:32        1  -0.1176450  12.7903558          0.015625
#> 54: 2022-11-11 04:22:32        1  -1.2749295  11.7845142          0.015625
#> 55: 2022-11-11 04:22:32        1   0.9590204   0.4748047          0.015625
#> 56: 2022-11-11 04:22:32        1  -4.4784605  10.3357322          0.015625
#> 57: 2022-11-11 04:22:32        1   7.5303068   9.7463952          0.015625
#> 58: 2022-11-11 04:22:32        1   3.7005933   1.5751057          0.015625
#> 59: 2022-11-11 04:22:32        1  -1.6446798   0.2802925          0.015625
#> 60: 2022-11-11 04:22:32        1  -0.8074505  13.6282601          0.015625
#> 61: 2022-11-11 04:22:32        1  -0.4530248   1.6649138          0.015625
#> 62: 2022-11-11 04:22:32        1   7.4112853  10.3326703          0.015625
#> 63: 2022-11-11 04:22:32        1   8.9769916  14.8233785          0.015625
#> 64: 2022-11-11 04:22:32        1  -3.7974972   3.9117019          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity