Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1      x2 fidelity  x_domain        y
#> 1: 3.077783 2.32004 0.015625 <list[3]> 1.277176

# best scoring evaluation
instance$result
#>          x1      x2 fidelity  x_domain        y
#> 1: 3.077783 2.32004 0.015625 <list[3]> 1.277176

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition          y
#>  1:  1.44437163  7.52535081 0.015625     0       6          1  25.348275
#>  2:  0.72861221  4.45374300 0.015625     0       6          1  17.326503
#>  3: -3.50993026 13.75012935 0.015625     0       6          1   4.228548
#>  4:  8.88535010  3.07151801 0.015625     0       6          1  78.946944
#>  5:  3.93946157  2.51299128 0.015625     0       6          1   8.611572
#>  6: -2.39934159  7.02197660 0.015625     0       6          1  11.766625
#>  7:  9.50456538  4.48647031 0.015625     0       6          1 117.844528
#>  8:  7.36461176  4.21491613 0.015625     0       6          1  82.880411
#>  9:  3.07778332  2.32003973 0.015625     0       6          1   1.277176
#> 10:  5.40500926  5.88277413 0.015625     0       6          1  73.692030
#> 11: -4.85510665  7.16293339 0.015625     0       6          1  64.494961
#> 12:  6.39852762  6.83236550 0.015625     0       6          1 114.739710
#> 13:  8.76294173 14.64250877 0.015625     0       6          1 411.602353
#> 14: -1.30601774 13.85729884 0.015625     0       6          1  45.302912
#> 15:  4.35780186 13.59714806 0.015625     0       6          1 201.238094
#> 16:  0.93577239  3.75646849 0.015625     0       6          1  16.306103
#> 17:  6.75944073 14.76019772 0.015625     0       6          1 346.627007
#> 18:  9.98159326 10.37112562 0.015625     0       6          1 297.479670
#> 19:  2.57125822 10.88515158 0.015625     0       6          1  78.903525
#> 20:  4.76554690  1.55867647 0.015625     0       6          1  16.488222
#> 21:  1.48381466  2.92834358 0.015625     0       6          1  11.439109
#> 22:  5.01811915 10.77801351 0.015625     0       6          1 156.658431
#> 23:  8.10895119 11.91207461 0.015625     0       6          1 289.687133
#> 24: -0.22044789  1.85205205 0.015625     0       6          1  39.622401
#> 25:  6.08187703  4.85733338 0.015625     0       6          1  74.162777
#> 26:  5.10717414 14.00161078 0.015625     0       6          1 248.639107
#> 27:  3.75278879 13.27590949 0.015625     0       6          1 166.376738
#> 28:  8.08278278  1.94236010 0.015625     0       6          1  54.031739
#> 29:  0.99392993  3.93758409 0.015625     0       6          1  15.498014
#> 30: -2.55386364  5.57942548 0.015625     0       6          1  23.965013
#> 31: -1.58651360 14.13300516 0.015625     0       6          1  40.436616
#> 32:  7.05058312 11.82954839 0.015625     0       6          1 257.857095
#> 33: -3.38339590  7.22825686 0.015625     0       6          1  21.004323
#> 34:  2.88575142  9.83453505 0.015625     0       6          1  67.480629
#> 35:  7.37368906 10.53470220 0.015625     0       6          1 227.553867
#> 36:  8.95347234 11.07954659 0.015625     0       6          1 285.860041
#> 37:  1.52478324  3.73596136 0.015625     0       6          1  10.449992
#> 38: -3.85190879  2.86176395 0.015625     0       6          1  97.294539
#> 39:  7.41028233  6.81929813 0.015625     0       6          1 133.471978
#> 40:  5.80057826  4.26637859 0.015625     0       6          1  60.285500
#> 41: -4.60819088  4.38783561 0.015625     0       6          1 101.146960
#> 42:  4.86885683 14.83986624 0.015625     0       6          1 263.036252
#> 43:  4.85058477  8.77130778 0.015625     0       6          1 106.733299
#> 44: -1.71186790 10.23349197 0.015625     0       6          1  10.663083
#> 45:  5.92823156 12.94002299 0.015625     0       6          1 252.926062
#> 46:  4.46322463 11.55928374 0.015625     0       6          1 152.840325
#> 47: -3.59507559  8.72554113 0.015625     0       6          1  12.884866
#> 48: -1.59761792 13.02133461 0.015625     0       6          1  29.103970
#> 49:  8.24561834  9.88517465 0.015625     0       6          1 228.881219
#> 50:  5.55982402 10.51950169 0.015625     0       6          1 171.399255
#> 51:  0.77009414  3.56349218 0.015625     0       6          1  18.403515
#> 52: -2.04611837  8.09144030 0.015625     0       6          1   7.279711
#> 53:  9.97808371  7.98322292 0.015625     0       6          1 220.947504
#> 54:  3.03711988  0.06654749 0.015625     0       6          1   2.363869
#> 55:  8.94947635  2.28429959 0.015625     0       6          1  66.509840
#> 56:  0.06922344  5.78248521 0.015625     0       6          1  19.590670
#> 57: -1.27881477 14.63753620 0.015625     0       6          1  55.692108
#> 58:  1.93606465  9.51314375 0.015625     0       6          1  48.550624
#> 59:  2.86120659 12.37829200 0.015625     0       6          1 114.842380
#> 60:  3.48829763  5.80573237 0.015625     0       6          1  25.803454
#> 61: -4.09097927  9.89875363 0.015625     0       6          1  14.186874
#> 62: -1.92741237  1.39067304 0.015625     0       6          1  67.349429
#> 63: -4.45609218  3.17358602 0.015625     0       6          1 118.426565
#> 64: -1.44550711  9.51938135 0.015625     0       6          1  12.532858
#>              x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-29 06:26:55        1  1.44437163  7.52535081          0.015625
#>  2: 2022-08-29 06:26:55        1  0.72861221  4.45374300          0.015625
#>  3: 2022-08-29 06:26:55        1 -3.50993026 13.75012935          0.015625
#>  4: 2022-08-29 06:26:55        1  8.88535010  3.07151801          0.015625
#>  5: 2022-08-29 06:26:55        1  3.93946157  2.51299128          0.015625
#>  6: 2022-08-29 06:26:55        1 -2.39934159  7.02197660          0.015625
#>  7: 2022-08-29 06:26:55        1  9.50456538  4.48647031          0.015625
#>  8: 2022-08-29 06:26:55        1  7.36461176  4.21491613          0.015625
#>  9: 2022-08-29 06:26:55        1  3.07778332  2.32003973          0.015625
#> 10: 2022-08-29 06:26:55        1  5.40500926  5.88277413          0.015625
#> 11: 2022-08-29 06:26:55        1 -4.85510665  7.16293339          0.015625
#> 12: 2022-08-29 06:26:55        1  6.39852762  6.83236550          0.015625
#> 13: 2022-08-29 06:26:55        1  8.76294173 14.64250877          0.015625
#> 14: 2022-08-29 06:26:55        1 -1.30601774 13.85729884          0.015625
#> 15: 2022-08-29 06:26:55        1  4.35780186 13.59714806          0.015625
#> 16: 2022-08-29 06:26:55        1  0.93577239  3.75646849          0.015625
#> 17: 2022-08-29 06:26:55        1  6.75944073 14.76019772          0.015625
#> 18: 2022-08-29 06:26:55        1  9.98159326 10.37112562          0.015625
#> 19: 2022-08-29 06:26:55        1  2.57125822 10.88515158          0.015625
#> 20: 2022-08-29 06:26:55        1  4.76554690  1.55867647          0.015625
#> 21: 2022-08-29 06:26:55        1  1.48381466  2.92834358          0.015625
#> 22: 2022-08-29 06:26:55        1  5.01811915 10.77801351          0.015625
#> 23: 2022-08-29 06:26:55        1  8.10895119 11.91207461          0.015625
#> 24: 2022-08-29 06:26:55        1 -0.22044789  1.85205205          0.015625
#> 25: 2022-08-29 06:26:55        1  6.08187703  4.85733338          0.015625
#> 26: 2022-08-29 06:26:55        1  5.10717414 14.00161078          0.015625
#> 27: 2022-08-29 06:26:55        1  3.75278879 13.27590949          0.015625
#> 28: 2022-08-29 06:26:55        1  8.08278278  1.94236010          0.015625
#> 29: 2022-08-29 06:26:55        1  0.99392993  3.93758409          0.015625
#> 30: 2022-08-29 06:26:55        1 -2.55386364  5.57942548          0.015625
#> 31: 2022-08-29 06:26:55        1 -1.58651360 14.13300516          0.015625
#> 32: 2022-08-29 06:26:55        1  7.05058312 11.82954839          0.015625
#> 33: 2022-08-29 06:26:55        1 -3.38339590  7.22825686          0.015625
#> 34: 2022-08-29 06:26:55        1  2.88575142  9.83453505          0.015625
#> 35: 2022-08-29 06:26:55        1  7.37368906 10.53470220          0.015625
#> 36: 2022-08-29 06:26:55        1  8.95347234 11.07954659          0.015625
#> 37: 2022-08-29 06:26:55        1  1.52478324  3.73596136          0.015625
#> 38: 2022-08-29 06:26:55        1 -3.85190879  2.86176395          0.015625
#> 39: 2022-08-29 06:26:55        1  7.41028233  6.81929813          0.015625
#> 40: 2022-08-29 06:26:55        1  5.80057826  4.26637859          0.015625
#> 41: 2022-08-29 06:26:55        1 -4.60819088  4.38783561          0.015625
#> 42: 2022-08-29 06:26:55        1  4.86885683 14.83986624          0.015625
#> 43: 2022-08-29 06:26:55        1  4.85058477  8.77130778          0.015625
#> 44: 2022-08-29 06:26:55        1 -1.71186790 10.23349197          0.015625
#> 45: 2022-08-29 06:26:55        1  5.92823156 12.94002299          0.015625
#> 46: 2022-08-29 06:26:55        1  4.46322463 11.55928374          0.015625
#> 47: 2022-08-29 06:26:55        1 -3.59507559  8.72554113          0.015625
#> 48: 2022-08-29 06:26:55        1 -1.59761792 13.02133461          0.015625
#> 49: 2022-08-29 06:26:55        1  8.24561834  9.88517465          0.015625
#> 50: 2022-08-29 06:26:55        1  5.55982402 10.51950169          0.015625
#> 51: 2022-08-29 06:26:55        1  0.77009414  3.56349218          0.015625
#> 52: 2022-08-29 06:26:55        1 -2.04611837  8.09144030          0.015625
#> 53: 2022-08-29 06:26:55        1  9.97808371  7.98322292          0.015625
#> 54: 2022-08-29 06:26:55        1  3.03711988  0.06654749          0.015625
#> 55: 2022-08-29 06:26:55        1  8.94947635  2.28429959          0.015625
#> 56: 2022-08-29 06:26:55        1  0.06922344  5.78248521          0.015625
#> 57: 2022-08-29 06:26:55        1 -1.27881477 14.63753620          0.015625
#> 58: 2022-08-29 06:26:55        1  1.93606465  9.51314375          0.015625
#> 59: 2022-08-29 06:26:55        1  2.86120659 12.37829200          0.015625
#> 60: 2022-08-29 06:26:55        1  3.48829763  5.80573237          0.015625
#> 61: 2022-08-29 06:26:55        1 -4.09097927  9.89875363          0.015625
#> 62: 2022-08-29 06:26:55        1 -1.92741237  1.39067304          0.015625
#> 63: 2022-08-29 06:26:55        1 -4.45609218  3.17358602          0.015625
#> 64: 2022-08-29 06:26:55        1 -1.44550711  9.51938135          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity