Optimizer using the Hyperband algorithm

OptimizerHyperband class that implements hyperband optimization (HBX). HBX repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HBX is called a bracket. HBX considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repeats

integer(1)
If 1 (default), optimization is stopped once all stages are evaluated. Otherwise, optimization is stopped after repeats runs of SHA. The bbotk::Terminator might stop the optimization before all repeats are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

stage (integer(1))
The stages of each bracket. Starts counting at 0.

bracket (integer(1))
The bracket index. Counts down to 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.760003 10.58461 0.015625 <list[3]> 1.090317

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.760003 10.58461 0.015625 <list[3]> 1.090317

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition          y
#>  1: -1.01490188 11.51084266 0.015625     0       6          1  29.996835
#>  2:  1.59842969  5.06897488 0.015625     0       6          1  12.089069
#>  3: -1.27463354 14.10342933 0.015625     0       6          1  49.101007
#>  4:  4.12608673  6.25806163 0.015625     0       6          1  44.395915
#>  5: -1.79458804 12.79022029 0.015625     0       6          1  22.576417
#>  6:  9.70442330  0.03518726 0.015625     0       6          1  44.128269
#>  7:  3.90728131 11.78303244 0.015625     0       6          1 136.070666
#>  8: -2.00302989 10.60740420 0.015625     0       6          1   7.657608
#>  9:  5.88129096  8.37605034 0.015625     0       6          1 132.747600
#> 10:  3.98077407 10.64135874 0.015625     0       6          1 113.619386
#> 11:  3.12084344 10.34357445 0.015625     0       6          1  81.599616
#> 12:  2.21490354  4.83128280 0.015625     0       6          1   9.098628
#> 13:  7.71289637  3.23970389 0.015625     0       6          1  70.425714
#> 14:  7.60018707  2.20133179 0.015625     0       6          1  54.939160
#> 15: -4.45756647 12.50570765 0.015625     0       6          1   9.018768
#> 16:  6.57772273  7.78708292 0.015625     0       6          1 138.556043
#> 17:  7.66299722  3.42293562 0.015625     0       6          1  72.872933
#> 18:  3.04458581  0.51485600 0.015625     0       6          1   1.297814
#> 19: -1.83243173  9.19753892 0.015625     0       6          1   7.547956
#> 20:  3.42053765 12.22047013 0.015625     0       6          1 128.564889
#> 21:  4.97449965 14.75407506 0.015625     0       6          1 265.628368
#> 22: -1.12035337  1.17664863 0.015625     0       6          1  58.337008
#> 23: -2.03055647 10.47191432 0.015625     0       6          1   6.978903
#> 24:  0.01217091  1.51204528 0.015625     0       6          1  39.569686
#> 25: -1.95764518 11.10370287 0.015625     0       6          1   9.874964
#> 26: -0.06031522  5.87929621 0.015625     0       6          1  19.631659
#> 27:  7.21070781 13.97155315 0.015625     0       6          1 334.349236
#> 28:  6.56772673  3.36934724 0.015625     0       6          1  61.413197
#> 29:  7.88239473  0.98288236 0.015625     0       6          1  41.286051
#> 30: -2.77618502 10.32856730 0.015625     0       6          1   1.138662
#> 31: -3.11998215 13.55162119 0.015625     0       6          1   5.629195
#> 32: -2.43708675  3.31246402 0.015625     0       6          1  48.231478
#> 33:  5.76377802  8.15782095 0.015625     0       6          1 124.625802
#> 34:  8.70766602  9.72015087 0.015625     0       6          1 235.248697
#> 35: -3.73630092 10.68396542 0.015625     0       6          1   4.908536
#> 36:  2.93892765 14.22258876 0.015625     0       6          1 160.224148
#> 37: -1.39700914  5.54175597 0.015625     0       6          1  19.177037
#> 38:  3.25763789 14.80587779 0.015625     0       6          1 187.174906
#> 39:  5.34852044 12.42674187 0.015625     0       6          1 213.377048
#> 40:  7.09788704  8.52915360 0.015625     0       6          1 167.306641
#> 41:  8.49356672  1.62511479 0.015625     0       6          1  52.223598
#> 42: -1.50070293  9.43479249 0.015625     0       6          1  11.627226
#> 43: -2.76000348 10.58461021 0.015625     0       6          1   1.090317
#> 44:  2.51159711  2.47975478 0.015625     0       6          1   2.321366
#> 45: -4.27128651  3.03921130 0.015625     0       6          1 112.396659
#> 46:  7.90269998  7.81358866 0.015625     0       6          1 165.055573
#> 47:  1.25115679  0.97605047 0.015625     0       6          1  22.508564
#> 48:  2.51763289 10.69131791 0.015625     0       6          1  74.514433
#> 49: -1.66866031 10.39000885 0.015625     0       6          1  11.779814
#> 50: -2.60906555  5.84668881 0.015625     0       6          1  22.113377
#> 51:  5.79093508  3.72523595 0.015625     0       6          1  53.398404
#> 52:  9.76411550  1.22058005 0.015625     0       6          1  62.243315
#> 53: -3.68116854  6.17736625 0.015625     0       6          1  38.948284
#> 54: -1.10194391 11.63246693 0.015625     0       6          1  29.094679
#> 55:  8.26874023  2.80832730 0.015625     0       6          1  68.007742
#> 56:  7.41137220  2.63515383 0.015625     0       6          1  59.564898
#> 57:  8.37146770  0.15880925 0.015625     0       6          1  33.633381
#> 58:  8.26050287  1.70693346 0.015625     0       6          1  51.845108
#> 59:  1.70319101  7.35129520 0.015625     0       6          1  24.515702
#> 60:  8.97786873  9.28003251 0.015625     0       6          1 229.063526
#> 61:  5.17254079 10.85682968 0.015625     0       6          1 164.732206
#> 62:  0.74217954  3.72324664 0.015625     0       6          1  18.314329
#> 63:  2.04240011  9.95049191 0.015625     0       6          1  55.662317
#> 64:  3.23435288  7.83826159 0.015625     0       6          1  44.851375
#>              x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr    status x_domain_x1 x_domain_x2
#>  1: 2021-12-18 15:36:12        1 evaluated -1.01490188 11.51084266
#>  2: 2021-12-18 15:36:12        1 evaluated  1.59842969  5.06897488
#>  3: 2021-12-18 15:36:12        1 evaluated -1.27463354 14.10342933
#>  4: 2021-12-18 15:36:12        1 evaluated  4.12608673  6.25806163
#>  5: 2021-12-18 15:36:12        1 evaluated -1.79458804 12.79022029
#>  6: 2021-12-18 15:36:12        1 evaluated  9.70442330  0.03518726
#>  7: 2021-12-18 15:36:12        1 evaluated  3.90728131 11.78303244
#>  8: 2021-12-18 15:36:12        1 evaluated -2.00302989 10.60740420
#>  9: 2021-12-18 15:36:12        1 evaluated  5.88129096  8.37605034
#> 10: 2021-12-18 15:36:12        1 evaluated  3.98077407 10.64135874
#> 11: 2021-12-18 15:36:12        1 evaluated  3.12084344 10.34357445
#> 12: 2021-12-18 15:36:12        1 evaluated  2.21490354  4.83128280
#> 13: 2021-12-18 15:36:12        1 evaluated  7.71289637  3.23970389
#> 14: 2021-12-18 15:36:12        1 evaluated  7.60018707  2.20133179
#> 15: 2021-12-18 15:36:12        1 evaluated -4.45756647 12.50570765
#> 16: 2021-12-18 15:36:12        1 evaluated  6.57772273  7.78708292
#> 17: 2021-12-18 15:36:12        1 evaluated  7.66299722  3.42293562
#> 18: 2021-12-18 15:36:12        1 evaluated  3.04458581  0.51485600
#> 19: 2021-12-18 15:36:12        1 evaluated -1.83243173  9.19753892
#> 20: 2021-12-18 15:36:12        1 evaluated  3.42053765 12.22047013
#> 21: 2021-12-18 15:36:12        1 evaluated  4.97449965 14.75407506
#> 22: 2021-12-18 15:36:12        1 evaluated -1.12035337  1.17664863
#> 23: 2021-12-18 15:36:12        1 evaluated -2.03055647 10.47191432
#> 24: 2021-12-18 15:36:12        1 evaluated  0.01217091  1.51204528
#> 25: 2021-12-18 15:36:12        1 evaluated -1.95764518 11.10370287
#> 26: 2021-12-18 15:36:12        1 evaluated -0.06031522  5.87929621
#> 27: 2021-12-18 15:36:12        1 evaluated  7.21070781 13.97155315
#> 28: 2021-12-18 15:36:12        1 evaluated  6.56772673  3.36934724
#> 29: 2021-12-18 15:36:12        1 evaluated  7.88239473  0.98288236
#> 30: 2021-12-18 15:36:12        1 evaluated -2.77618502 10.32856730
#> 31: 2021-12-18 15:36:12        1 evaluated -3.11998215 13.55162119
#> 32: 2021-12-18 15:36:12        1 evaluated -2.43708675  3.31246402
#> 33: 2021-12-18 15:36:12        1 evaluated  5.76377802  8.15782095
#> 34: 2021-12-18 15:36:12        1 evaluated  8.70766602  9.72015087
#> 35: 2021-12-18 15:36:12        1 evaluated -3.73630092 10.68396542
#> 36: 2021-12-18 15:36:12        1 evaluated  2.93892765 14.22258876
#> 37: 2021-12-18 15:36:12        1 evaluated -1.39700914  5.54175597
#> 38: 2021-12-18 15:36:12        1 evaluated  3.25763789 14.80587779
#> 39: 2021-12-18 15:36:12        1 evaluated  5.34852044 12.42674187
#> 40: 2021-12-18 15:36:12        1 evaluated  7.09788704  8.52915360
#> 41: 2021-12-18 15:36:12        1 evaluated  8.49356672  1.62511479
#> 42: 2021-12-18 15:36:12        1 evaluated -1.50070293  9.43479249
#> 43: 2021-12-18 15:36:12        1 evaluated -2.76000348 10.58461021
#> 44: 2021-12-18 15:36:12        1 evaluated  2.51159711  2.47975478
#> 45: 2021-12-18 15:36:12        1 evaluated -4.27128651  3.03921130
#> 46: 2021-12-18 15:36:12        1 evaluated  7.90269998  7.81358866
#> 47: 2021-12-18 15:36:12        1 evaluated  1.25115679  0.97605047
#> 48: 2021-12-18 15:36:12        1 evaluated  2.51763289 10.69131791
#> 49: 2021-12-18 15:36:12        1 evaluated -1.66866031 10.39000885
#> 50: 2021-12-18 15:36:12        1 evaluated -2.60906555  5.84668881
#> 51: 2021-12-18 15:36:12        1 evaluated  5.79093508  3.72523595
#> 52: 2021-12-18 15:36:12        1 evaluated  9.76411550  1.22058005
#> 53: 2021-12-18 15:36:12        1 evaluated -3.68116854  6.17736625
#> 54: 2021-12-18 15:36:12        1 evaluated -1.10194391 11.63246693
#> 55: 2021-12-18 15:36:12        1 evaluated  8.26874023  2.80832730
#> 56: 2021-12-18 15:36:12        1 evaluated  7.41137220  2.63515383
#> 57: 2021-12-18 15:36:12        1 evaluated  8.37146770  0.15880925
#> 58: 2021-12-18 15:36:12        1 evaluated  8.26050287  1.70693346
#> 59: 2021-12-18 15:36:12        1 evaluated  1.70319101  7.35129520
#> 60: 2021-12-18 15:36:12        1 evaluated  8.97786873  9.28003251
#> 61: 2021-12-18 15:36:12        1 evaluated  5.17254079 10.85682968
#> 62: 2021-12-18 15:36:12        1 evaluated  0.74217954  3.72324664
#> 63: 2021-12-18 15:36:12        1 evaluated  2.04240011  9.95049191
#> 64: 2021-12-18 15:36:12        1 evaluated  3.23435288  7.83826159
#>               timestamp batch_nr    status x_domain_x1 x_domain_x2
#>     x_domain_fidelity
#>  1:          0.015625
#>  2:          0.015625
#>  3:          0.015625
#>  4:          0.015625
#>  5:          0.015625
#>  6:          0.015625
#>  7:          0.015625
#>  8:          0.015625
#>  9:          0.015625
#> 10:          0.015625
#> 11:          0.015625
#> 12:          0.015625
#> 13:          0.015625
#> 14:          0.015625
#> 15:          0.015625
#> 16:          0.015625
#> 17:          0.015625
#> 18:          0.015625
#> 19:          0.015625
#> 20:          0.015625
#> 21:          0.015625
#> 22:          0.015625
#> 23:          0.015625
#> 24:          0.015625
#> 25:          0.015625
#> 26:          0.015625
#> 27:          0.015625
#> 28:          0.015625
#> 29:          0.015625
#> 30:          0.015625
#> 31:          0.015625
#> 32:          0.015625
#> 33:          0.015625
#> 34:          0.015625
#> 35:          0.015625
#> 36:          0.015625
#> 37:          0.015625
#> 38:          0.015625
#> 39:          0.015625
#> 40:          0.015625
#> 41:          0.015625
#> 42:          0.015625
#> 43:          0.015625
#> 44:          0.015625
#> 45:          0.015625
#> 46:          0.015625
#> 47:          0.015625
#> 48:          0.015625
#> 49:          0.015625
#> 50:          0.015625
#> 51:          0.015625
#> 52:          0.015625
#> 53:          0.015625
#> 54:          0.015625
#> 55:          0.015625
#> 56:          0.015625
#> 57:          0.015625
#> 58:          0.015625
#> 59:          0.015625
#> 60:          0.015625
#> 61:          0.015625
#> 62:          0.015625
#> 63:          0.015625
#> 64:          0.015625
#>     x_domain_fidelity