Optimizer using the Hyperband algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.426219 1.189902 0.015625 <list[3]> 0.8637147

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.426219 1.189902 0.015625 <list[3]> 0.8637147

# all evaluations
as.data.table(instance$archive)
#>             x1          x2 fidelity stage bracket repetition           y
#>  1:  5.4835296  5.65702173 0.015625     0       6          1  72.3407734
#>  2:  8.8228216  3.46683501 0.015625     0       6          1  85.1757622
#>  3:  5.7873812  0.06519265 0.015625     0       6          1  23.4915822
#>  4: -0.3741965  8.24913737 0.015625     0       6          1  21.6577858
#>  5:  9.3136946  8.85776989 0.015625     0       6          1 225.8721235
#>  6:  4.3648621 13.08119472 0.015625     0       6          1 187.4250920
#>  7: -1.5891047 14.26019150 0.015625     0       6          1  41.7852424
#>  8: -3.3046171 14.15115645 0.015625     0       6          1   7.0579418
#>  9: -4.9038023 12.25845920 0.015625     0       6          1  17.0505825
#> 10: -0.6869677  8.27869451 0.015625     0       6          1  18.7949600
#> 11:  7.2008546  3.20633042 0.015625     0       6          1  65.8558087
#> 12:  4.6738674  3.80785378 0.015625     0       6          1  30.5596863
#> 13:  1.2088585 13.83012480 0.015625     0       6          1 107.6675993
#> 14:  6.5871775  3.76504382 0.015625     0       6          1  66.9752243
#> 15:  9.8567855 10.41393417 0.015625     0       6          1 294.1770104
#> 16: -0.5676410  3.06269847 0.015625     0       6          1  32.9236374
#> 17:  7.3914757  6.17661383 0.015625     0       6          1 119.5660100
#> 18:  0.4110144  0.67283313 0.015625     0       6          1  40.6880715
#> 19:  6.7680581 10.41555901 0.015625     0       6          1 208.3517969
#> 20:  9.9221928  8.33593645 0.015625     0       6          1 229.5882869
#> 21: -2.3239347  9.13417390 0.015625     0       6          1   3.9664958
#> 22: -3.9013581  6.98437942 0.015625     0       6          1  35.4466071
#> 23:  2.4754083  7.40672556 0.015625     0       6          1  29.0564741
#> 24:  4.1056463 13.96041232 0.015625     0       6          1 199.8666274
#> 25:  8.1587341 11.16806602 0.015625     0       6          1 266.5356397
#> 26:  5.9970284 14.94287555 0.015625     0       6          1 321.3332859
#> 27:  0.3499776  7.94198497 0.015625     0       6          1  25.2461871
#> 28:  8.8603516 10.38312846 0.015625     0       6          1 260.1641461
#> 29:  9.5181045  4.28405550 0.015625     0       6          1 113.8000956
#> 30:  9.6276573  3.64362278 0.015625     0       6          1 102.9388006
#> 31:  8.1686479 13.01800917 0.015625     0       6          1 329.8472199
#> 32:  9.7051158  4.44958863 0.015625     0       6          1 121.7666130
#> 33:  7.3705570 13.18285055 0.015625     0       6          1 311.7884425
#> 34:  9.4602468  5.73929405 0.015625     0       6          1 145.4619068
#> 35: -4.1829815 11.35667416 0.015625     0       6          1   8.5316798
#> 36:  0.3736270  3.54173839 0.015625     0       6          1  22.4287400
#> 37: -0.1257373 11.06940947 0.015625     0       6          1  43.2315825
#> 38:  6.5261757 10.02821296 0.015625     0       6          1 191.0715197
#> 39:  4.8907138  3.19643217 0.015625     0       6          1  29.7216021
#> 40: -3.5652889  8.31139764 0.015625     0       6          1  15.3377375
#> 41: -0.8138902 10.31889746 0.015625     0       6          1  25.6126486
#> 42: -0.8544270 13.06708524 0.015625     0       6          1  48.6219089
#> 43: -3.9114675  7.19209277 0.015625     0       6          1  33.3955075
#> 44:  6.3888812 11.89774652 0.015625     0       6          1 238.9125136
#> 45: -4.4984835 10.13510011 0.015625     0       6          1  21.2598571
#> 46:  9.1856469  3.55948169 0.015625     0       6          1  92.5352935
#> 47: -1.8498192 10.96826777 0.015625     0       6          1  11.0378967
#> 48:  3.4262191  1.18990194 0.015625     0       6          1   0.8637147
#> 49: -0.8693471  8.17963385 0.015625     0       6          1  16.7936883
#> 50: -2.9356640  2.78062269 0.015625     0       6          1  67.1311055
#> 51:  0.4834264 14.03989913 0.015625     0       6          1  95.9789312
#> 52:  6.4718493  6.00498055 0.015625     0       6          1 100.7455079
#> 53:  0.8892917 10.60652928 0.015625     0       6          1  52.0197544
#> 54: -1.4445606  3.79612323 0.015625     0       6          1  32.0675697
#> 55:  5.4130474  8.41398770 0.015625     0       6          1 118.7711122
#> 56:  7.3966621 11.73399322 0.015625     0       6          1 264.6374391
#> 57:  1.9000115  0.77599394 0.015625     0       6          1  12.2365473
#> 58:  3.5820533  7.67813752 0.015625     0       6          1  50.0994746
#> 59: -2.2443905  0.73343072 0.015625     0       6          1  84.8934188
#> 60:  6.2911496 11.61412507 0.015625     0       6          1 227.2463990
#> 61: -1.0353402  5.59811418 0.015625     0       6          1  19.2367059
#> 62:  8.7024683  1.75119584 0.015625     0       6          1  55.6928667
#> 63:  5.3320672  6.80212283 0.015625     0       6          1  86.3754942
#> 64:  1.6627142  0.22990506 0.015625     0       6          1  19.4150775
#>             x1          x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-02-03 18:27:20        1   5.4835296  5.65702173          0.015625
#>  2: 2022-02-03 18:27:20        1   8.8228216  3.46683501          0.015625
#>  3: 2022-02-03 18:27:20        1   5.7873812  0.06519265          0.015625
#>  4: 2022-02-03 18:27:20        1  -0.3741965  8.24913737          0.015625
#>  5: 2022-02-03 18:27:20        1   9.3136946  8.85776989          0.015625
#>  6: 2022-02-03 18:27:20        1   4.3648621 13.08119472          0.015625
#>  7: 2022-02-03 18:27:20        1  -1.5891047 14.26019150          0.015625
#>  8: 2022-02-03 18:27:20        1  -3.3046171 14.15115645          0.015625
#>  9: 2022-02-03 18:27:20        1  -4.9038023 12.25845920          0.015625
#> 10: 2022-02-03 18:27:20        1  -0.6869677  8.27869451          0.015625
#> 11: 2022-02-03 18:27:20        1   7.2008546  3.20633042          0.015625
#> 12: 2022-02-03 18:27:20        1   4.6738674  3.80785378          0.015625
#> 13: 2022-02-03 18:27:20        1   1.2088585 13.83012480          0.015625
#> 14: 2022-02-03 18:27:20        1   6.5871775  3.76504382          0.015625
#> 15: 2022-02-03 18:27:20        1   9.8567855 10.41393417          0.015625
#> 16: 2022-02-03 18:27:20        1  -0.5676410  3.06269847          0.015625
#> 17: 2022-02-03 18:27:20        1   7.3914757  6.17661383          0.015625
#> 18: 2022-02-03 18:27:20        1   0.4110144  0.67283313          0.015625
#> 19: 2022-02-03 18:27:20        1   6.7680581 10.41555901          0.015625
#> 20: 2022-02-03 18:27:20        1   9.9221928  8.33593645          0.015625
#> 21: 2022-02-03 18:27:20        1  -2.3239347  9.13417390          0.015625
#> 22: 2022-02-03 18:27:20        1  -3.9013581  6.98437942          0.015625
#> 23: 2022-02-03 18:27:20        1   2.4754083  7.40672556          0.015625
#> 24: 2022-02-03 18:27:20        1   4.1056463 13.96041232          0.015625
#> 25: 2022-02-03 18:27:20        1   8.1587341 11.16806602          0.015625
#> 26: 2022-02-03 18:27:20        1   5.9970284 14.94287555          0.015625
#> 27: 2022-02-03 18:27:20        1   0.3499776  7.94198497          0.015625
#> 28: 2022-02-03 18:27:20        1   8.8603516 10.38312846          0.015625
#> 29: 2022-02-03 18:27:20        1   9.5181045  4.28405550          0.015625
#> 30: 2022-02-03 18:27:20        1   9.6276573  3.64362278          0.015625
#> 31: 2022-02-03 18:27:20        1   8.1686479 13.01800917          0.015625
#> 32: 2022-02-03 18:27:20        1   9.7051158  4.44958863          0.015625
#> 33: 2022-02-03 18:27:20        1   7.3705570 13.18285055          0.015625
#> 34: 2022-02-03 18:27:20        1   9.4602468  5.73929405          0.015625
#> 35: 2022-02-03 18:27:20        1  -4.1829815 11.35667416          0.015625
#> 36: 2022-02-03 18:27:20        1   0.3736270  3.54173839          0.015625
#> 37: 2022-02-03 18:27:20        1  -0.1257373 11.06940947          0.015625
#> 38: 2022-02-03 18:27:20        1   6.5261757 10.02821296          0.015625
#> 39: 2022-02-03 18:27:20        1   4.8907138  3.19643217          0.015625
#> 40: 2022-02-03 18:27:20        1  -3.5652889  8.31139764          0.015625
#> 41: 2022-02-03 18:27:20        1  -0.8138902 10.31889746          0.015625
#> 42: 2022-02-03 18:27:20        1  -0.8544270 13.06708524          0.015625
#> 43: 2022-02-03 18:27:20        1  -3.9114675  7.19209277          0.015625
#> 44: 2022-02-03 18:27:20        1   6.3888812 11.89774652          0.015625
#> 45: 2022-02-03 18:27:20        1  -4.4984835 10.13510011          0.015625
#> 46: 2022-02-03 18:27:20        1   9.1856469  3.55948169          0.015625
#> 47: 2022-02-03 18:27:20        1  -1.8498192 10.96826777          0.015625
#> 48: 2022-02-03 18:27:20        1   3.4262191  1.18990194          0.015625
#> 49: 2022-02-03 18:27:20        1  -0.8693471  8.17963385          0.015625
#> 50: 2022-02-03 18:27:20        1  -2.9356640  2.78062269          0.015625
#> 51: 2022-02-03 18:27:20        1   0.4834264 14.03989913          0.015625
#> 52: 2022-02-03 18:27:20        1   6.4718493  6.00498055          0.015625
#> 53: 2022-02-03 18:27:20        1   0.8892917 10.60652928          0.015625
#> 54: 2022-02-03 18:27:20        1  -1.4445606  3.79612323          0.015625
#> 55: 2022-02-03 18:27:20        1   5.4130474  8.41398770          0.015625
#> 56: 2022-02-03 18:27:20        1   7.3966621 11.73399322          0.015625
#> 57: 2022-02-03 18:27:20        1   1.9000115  0.77599394          0.015625
#> 58: 2022-02-03 18:27:20        1   3.5820533  7.67813752          0.015625
#> 59: 2022-02-03 18:27:20        1  -2.2443905  0.73343072          0.015625
#> 60: 2022-02-03 18:27:20        1   6.2911496 11.61412507          0.015625
#> 61: 2022-02-03 18:27:20        1  -1.0353402  5.59811418          0.015625
#> 62: 2022-02-03 18:27:20        1   8.7024683  1.75119584          0.015625
#> 63: 2022-02-03 18:27:20        1   5.3320672  6.80212283          0.015625
#> 64: 2022-02-03 18:27:20        1   1.6627142  0.22990506          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity