Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# Branin function with fidelity, see `bbotk::branin()`
fun = function(xs) branin_wu(xs[["x1"]], xs[["x2"]], xs[["fidelity"]])

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.241373 0.463074 0.015625 <list[3]> 0.9372791

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.241373 0.463074 0.015625 <list[3]> 0.9372791

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1:  9.4305035  6.2236745 0.015625     0       6          1 156.6060553
#>  2:  9.8064752  8.7165618 0.015625     0       6          1 237.2399194
#>  3: -2.3648628  1.8722289 0.015625     0       6          1  68.1720723
#>  4: -3.2167325  8.3015346 0.015625     0       6          1  10.2607621
#>  5:  3.2413726  0.4630740 0.015625     0       6          1   0.9372791
#>  6: -1.5658496  1.3232825 0.015625     0       6          1  62.5264063
#>  7:  7.7809937  9.4695888 0.015625     0       6          1 206.4729835
#>  8:  8.1815633 12.9985230 0.015625     0       6          1 329.5346837
#>  9:  5.3605829 11.8858089 0.015625     0       6          1 198.9637068
#> 10:  7.0549675  6.9448195 0.015625     0       6          1 130.1505461
#> 11:  3.9864096  9.1553131 0.015625     0       6          1  84.8283789
#> 12:  6.4130736 12.5927559 0.015625     0       6          1 260.8553266
#> 13: -2.1085515 14.1825398 0.015625     0       6          1  27.0775731
#> 14: -3.4938317  1.5292108 0.015625     0       6          1 109.2872433
#> 15:  8.0256745  1.4366851 0.015625     0       6          1  47.1659354
#> 16:  0.1416923 11.4815734 0.015625     0       6          1  52.0696436
#> 17:  9.3418862  6.0351090 0.015625     0       6          1 149.7559675
#> 18: -4.9854670 10.3872175 0.015625     0       6          1  31.1796647
#> 19: -1.6549553  8.7498140 0.015625     0       6          1   9.1938523
#> 20: -4.5749876 14.4004761 0.015625     0       6          1   8.9110054
#> 21:  8.2966701  0.7672736 0.015625     0       6          1  40.1721205
#> 22:  3.9113425  0.3018862 0.015625     0       6          1   3.1080719
#> 23: -3.7144330  0.3302164 0.015625     0       6          1 146.0676548
#> 24:  6.0357935 12.0998401 0.015625     0       6          1 232.0601620
#> 25:  9.2899348 10.9649939 0.015625     0       6          1 292.7865998
#> 26:  3.0561391  3.8164885 0.015625     0       6          1   6.1608491
#> 27:  8.4806820 12.2239903 0.015625     0       6          1 310.9697299
#> 28:  6.7173772  2.8759993 0.015625     0       6          1  56.8990256
#> 29: -4.8558897  1.9595644 0.015625     0       6          1 167.4689279
#> 30: -2.4621494  0.7652608 0.015625     0       6          1  89.7615186
#> 31: -0.2915741 14.6156735 0.015625     0       6          1  85.6031139
#> 32:  5.2833721  7.8168456 0.015625     0       6          1 102.9362506
#> 33:  6.8747312  2.4553969 0.015625     0       6          1  53.2981889
#> 34: -2.1158644  5.1268320 0.015625     0       6          1  24.1912319
#> 35: -1.2814540  8.1124402 0.015625     0       6          1  12.7401977
#> 36:  4.9789526  9.8712817 0.015625     0       6          1 134.2633774
#> 37:  6.9757666  5.2665139 0.015625     0       6          1  96.1128421
#> 38:  6.0108348 10.1461588 0.015625     0       6          1 178.0537012
#> 39:  4.2591722  4.9747055 0.015625     0       6          1  32.7900977
#> 40:  3.1950145 12.3643555 0.015625     0       6          1 124.4111652
#> 41:  0.6077164  7.4439865 0.015625     0       6          1  23.6421339
#> 42: -3.8504457  1.0765158 0.015625     0       6          1 135.1337827
#> 43: -3.6716265  2.1164658 0.015625     0       6          1 104.5675437
#> 44:  4.5513642  4.8360648 0.015625     0       6          1  38.0852658
#> 45: -0.6189721  4.1928620 0.015625     0       6          1  25.6833388
#> 46:  8.2190416  6.2070666 0.015625     0       6          1 132.2593174
#> 47:  2.2939007 13.3906074 0.015625     0       6          1 122.0134311
#> 48: -2.0489426  6.0557197 0.015625     0       6          1  16.6996622
#> 49:  2.5625571 13.5684015 0.015625     0       6          1 132.9495751
#> 50:  6.4740672  0.7817307 0.015625     0       6          1  33.8434973
#> 51:  0.8290405  7.0997329 0.015625     0       6          1  22.2377168
#> 52:  9.2997426  3.3930439 0.015625     0       6          1  91.3868283
#> 53: -4.9330824  2.0606384 0.015625     0       6          1 169.3245668
#> 54: -1.9070116  6.8260691 0.015625     0       6          1  12.2184551
#> 55:  6.1343391  3.0221715 0.015625     0       6          1  51.1732722
#> 56: -1.3311436 12.3409751 0.015625     0       6          1  29.6507014
#> 57:  2.5712507  6.6046415 0.015625     0       6          1  22.1105033
#> 58:  9.0897687  9.9654114 0.015625     0       6          1 253.4806828
#> 59:  4.1524776  8.5336704 0.015625     0       6          1  79.0731448
#> 60: -3.3040081 13.0602407 0.015625     0       6          1   2.6737036
#> 61:  0.1566417 12.2631001 0.015625     0       6          1  61.8861202
#> 62:  2.4104397  4.0484007 0.015625     0       6          1   5.7628745
#> 63: -3.8099577  5.2234857 0.015625     0       6          1  55.5579776
#> 64:  7.9119933 13.4998747 0.015625     0       6          1 339.4995715
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-11-27 12:09:43        1   9.4305035   6.2236745          0.015625
#>  2: 2022-11-27 12:09:43        1   9.8064752   8.7165618          0.015625
#>  3: 2022-11-27 12:09:43        1  -2.3648628   1.8722289          0.015625
#>  4: 2022-11-27 12:09:43        1  -3.2167325   8.3015346          0.015625
#>  5: 2022-11-27 12:09:43        1   3.2413726   0.4630740          0.015625
#>  6: 2022-11-27 12:09:43        1  -1.5658496   1.3232825          0.015625
#>  7: 2022-11-27 12:09:43        1   7.7809937   9.4695888          0.015625
#>  8: 2022-11-27 12:09:43        1   8.1815633  12.9985230          0.015625
#>  9: 2022-11-27 12:09:43        1   5.3605829  11.8858089          0.015625
#> 10: 2022-11-27 12:09:43        1   7.0549675   6.9448195          0.015625
#> 11: 2022-11-27 12:09:43        1   3.9864096   9.1553131          0.015625
#> 12: 2022-11-27 12:09:43        1   6.4130736  12.5927559          0.015625
#> 13: 2022-11-27 12:09:43        1  -2.1085515  14.1825398          0.015625
#> 14: 2022-11-27 12:09:43        1  -3.4938317   1.5292108          0.015625
#> 15: 2022-11-27 12:09:43        1   8.0256745   1.4366851          0.015625
#> 16: 2022-11-27 12:09:43        1   0.1416923  11.4815734          0.015625
#> 17: 2022-11-27 12:09:43        1   9.3418862   6.0351090          0.015625
#> 18: 2022-11-27 12:09:43        1  -4.9854670  10.3872175          0.015625
#> 19: 2022-11-27 12:09:43        1  -1.6549553   8.7498140          0.015625
#> 20: 2022-11-27 12:09:43        1  -4.5749876  14.4004761          0.015625
#> 21: 2022-11-27 12:09:43        1   8.2966701   0.7672736          0.015625
#> 22: 2022-11-27 12:09:43        1   3.9113425   0.3018862          0.015625
#> 23: 2022-11-27 12:09:43        1  -3.7144330   0.3302164          0.015625
#> 24: 2022-11-27 12:09:43        1   6.0357935  12.0998401          0.015625
#> 25: 2022-11-27 12:09:43        1   9.2899348  10.9649939          0.015625
#> 26: 2022-11-27 12:09:43        1   3.0561391   3.8164885          0.015625
#> 27: 2022-11-27 12:09:43        1   8.4806820  12.2239903          0.015625
#> 28: 2022-11-27 12:09:43        1   6.7173772   2.8759993          0.015625
#> 29: 2022-11-27 12:09:43        1  -4.8558897   1.9595644          0.015625
#> 30: 2022-11-27 12:09:43        1  -2.4621494   0.7652608          0.015625
#> 31: 2022-11-27 12:09:43        1  -0.2915741  14.6156735          0.015625
#> 32: 2022-11-27 12:09:43        1   5.2833721   7.8168456          0.015625
#> 33: 2022-11-27 12:09:43        1   6.8747312   2.4553969          0.015625
#> 34: 2022-11-27 12:09:43        1  -2.1158644   5.1268320          0.015625
#> 35: 2022-11-27 12:09:43        1  -1.2814540   8.1124402          0.015625
#> 36: 2022-11-27 12:09:43        1   4.9789526   9.8712817          0.015625
#> 37: 2022-11-27 12:09:43        1   6.9757666   5.2665139          0.015625
#> 38: 2022-11-27 12:09:43        1   6.0108348  10.1461588          0.015625
#> 39: 2022-11-27 12:09:43        1   4.2591722   4.9747055          0.015625
#> 40: 2022-11-27 12:09:43        1   3.1950145  12.3643555          0.015625
#> 41: 2022-11-27 12:09:43        1   0.6077164   7.4439865          0.015625
#> 42: 2022-11-27 12:09:43        1  -3.8504457   1.0765158          0.015625
#> 43: 2022-11-27 12:09:43        1  -3.6716265   2.1164658          0.015625
#> 44: 2022-11-27 12:09:43        1   4.5513642   4.8360648          0.015625
#> 45: 2022-11-27 12:09:43        1  -0.6189721   4.1928620          0.015625
#> 46: 2022-11-27 12:09:43        1   8.2190416   6.2070666          0.015625
#> 47: 2022-11-27 12:09:43        1   2.2939007  13.3906074          0.015625
#> 48: 2022-11-27 12:09:43        1  -2.0489426   6.0557197          0.015625
#> 49: 2022-11-27 12:09:43        1   2.5625571  13.5684015          0.015625
#> 50: 2022-11-27 12:09:43        1   6.4740672   0.7817307          0.015625
#> 51: 2022-11-27 12:09:43        1   0.8290405   7.0997329          0.015625
#> 52: 2022-11-27 12:09:43        1   9.2997426   3.3930439          0.015625
#> 53: 2022-11-27 12:09:43        1  -4.9330824   2.0606384          0.015625
#> 54: 2022-11-27 12:09:43        1  -1.9070116   6.8260691          0.015625
#> 55: 2022-11-27 12:09:43        1   6.1343391   3.0221715          0.015625
#> 56: 2022-11-27 12:09:43        1  -1.3311436  12.3409751          0.015625
#> 57: 2022-11-27 12:09:43        1   2.5712507   6.6046415          0.015625
#> 58: 2022-11-27 12:09:43        1   9.0897687   9.9654114          0.015625
#> 59: 2022-11-27 12:09:43        1   4.1524776   8.5336704          0.015625
#> 60: 2022-11-27 12:09:43        1  -3.3040081  13.0602407          0.015625
#> 61: 2022-11-27 12:09:43        1   0.1566417  12.2631001          0.015625
#> 62: 2022-11-27 12:09:43        1   2.4104397   4.0484007          0.015625
#> 63: 2022-11-27 12:09:43        1  -3.8099577   5.2234857          0.015625
#> 64: 2022-11-27 12:09:43        1   7.9119933  13.4998747          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity