Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1      x2 fidelity  x_domain        y
#> 1: -3.563925 11.9293 0.015625 <list[3]> 1.259368

# best scoring evaluation
instance$result
#>           x1      x2 fidelity  x_domain        y
#> 1: -3.563925 11.9293 0.015625 <list[3]> 1.259368

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  2.61143772  5.2760241 0.015625     0       6          1  12.100963
#>  2: -2.55154648 13.3993690 0.015625     0       6          1  11.870272
#>  3: -1.76080138  9.5818728 0.015625     0       6          1   8.654558
#>  4:  6.70748046 13.1353704 0.015625     0       6          1 288.608187
#>  5:  5.70141532  7.2246863 0.015625     0       6          1 104.499540
#>  6: -2.24195847  3.0835398 0.015625     0       6          1  48.107377
#>  7:  2.20537971  9.1739304 0.015625     0       6          1  47.005286
#>  8: -0.29187804 11.6956023 0.015625     0       6          1  46.532626
#>  9:  8.37322733  2.3682803 0.015625     0       6          1  62.071466
#> 10:  1.69358456 11.6295048 0.015625     0       6          1  76.667801
#> 11: -2.45698581  2.2632501 0.015625     0       6          1  63.913951
#> 12: -1.41301993 10.0591169 0.015625     0       6          1  14.567135
#> 13: -4.64511715  0.4071126 0.015625     0       6          1 195.656633
#> 14:  1.66434331  6.3718726 0.015625     0       6          1  17.720731
#> 15:  9.25790003  3.5157802 0.015625     0       6          1  92.977564
#> 16:  3.06194744  4.4570226 0.015625     0       6          1   9.682054
#> 17: -4.85657309  3.5815816 0.015625     0       6          1 129.604045
#> 18:  6.21703496  3.6314845 0.015625     0       6          1  59.748620
#> 19:  3.37551511  4.5274198 0.015625     0       6          1  13.257534
#> 20:  6.26269920 12.0534974 0.015625     0       6          1 239.082901
#> 21: -4.37305455 14.2268460 0.015625     0       6          1   7.264777
#> 22:  2.30547449  1.6353809 0.015625     0       6          1   4.300724
#> 23:  4.22579476  6.9784876 0.015625     0       6          1  56.703746
#> 24:  9.59191561 14.1609771 0.015625     0       6          1 424.813876
#> 25:  1.61821049  5.2336509 0.015625     0       6          1  12.532948
#> 26: -2.10513715  1.7258575 0.015625     0       6          1  65.340391
#> 27: -0.57470714 14.8633940 0.015625     0       6          1  81.080343
#> 28:  4.15167850 12.3727821 0.015625     0       6          1 159.906542
#> 29: -2.57630753 10.5134073 0.015625     0       6          1   1.935315
#> 30:  1.87075879  0.9407849 0.015625     0       6          1  11.956271
#> 31: -0.11471011  2.3754188 0.015625     0       6          1  34.036464
#> 32:  0.05638359  0.4507534 0.015625     0       6          1  49.394163
#> 33: -3.12649853  7.1654322 0.015625     0       6          1  17.300096
#> 34:  1.36077028 11.5383207 0.015625     0       6          1  70.480340
#> 35:  1.01098108 10.3673477 0.015625     0       6          1  50.441414
#> 36:  6.23749243  1.3606484 0.015625     0       6          1  36.333880
#> 37:  8.27746620  3.3206714 0.015625     0       6          1  76.412525
#> 38: -2.42187955  0.8686659 0.015625     0       6          1  86.798881
#> 39: -4.60450085 13.1387673 0.015625     0       6          1   9.674015
#> 40: -3.57702205 14.5932060 0.015625     0       6          1   7.577856
#> 41: -3.38520327  1.4719854 0.015625     0       6          1 106.114901
#> 42:  6.15765413 11.9180846 0.015625     0       6          1 231.300947
#> 43:  8.79778821 14.0515189 0.015625     0       6          1 389.281938
#> 44: -2.23161306  7.8917610 0.015625     0       6          1   7.393865
#> 45:  7.49910586 14.0107568 0.015625     0       6          1 345.190019
#> 46:  5.99641589 13.7917349 0.015625     0       6          1 282.614996
#> 47:  5.59757876 11.3326648 0.015625     0       6          1 193.740192
#> 48:  6.85421897  5.0489103 0.015625     0       6          1  90.553827
#> 49:  5.26880707  6.6772867 0.015625     0       6          1  82.464051
#> 50:  3.10763952  4.2748845 0.015625     0       6          1   8.952669
#> 51: -1.44645402 11.5674611 0.015625     0       6          1  21.437456
#> 52:  1.25691244  0.7720866 0.015625     0       6          1  23.697206
#> 53: -0.76081594 12.7589941 0.015625     0       6          1  47.538999
#> 54: -3.70946153 13.9991755 0.015625     0       6          1   4.701542
#> 55:  3.14002814 12.3274849 0.015625     0       6          1 121.878773
#> 56:  9.18088999 11.4121943 0.015625     0       6          1 304.570892
#> 57: -3.87860680 12.7547498 0.015625     0       6          1   2.904052
#> 58: -2.48719570  7.4563148 0.015625     0       6          1   9.630478
#> 59: -3.56392485 11.9292969 0.015625     0       6          1   1.259368
#> 60: -4.15768915  6.7302875 0.015625     0       6          1  46.138343
#> 61: -1.46954704 14.3069720 0.015625     0       6          1  45.800787
#> 62: -3.25490174  2.8412366 0.015625     0       6          1  75.539020
#> 63:  3.28975640  8.4369008 0.015625     0       6          1  54.377975
#> 64:  1.41816351  5.8919047 0.015625     0       6          1  15.816084
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-29 04:57:11        1  2.61143772   5.2760241          0.015625
#>  2: 2022-08-29 04:57:11        1 -2.55154648  13.3993690          0.015625
#>  3: 2022-08-29 04:57:11        1 -1.76080138   9.5818728          0.015625
#>  4: 2022-08-29 04:57:11        1  6.70748046  13.1353704          0.015625
#>  5: 2022-08-29 04:57:11        1  5.70141532   7.2246863          0.015625
#>  6: 2022-08-29 04:57:11        1 -2.24195847   3.0835398          0.015625
#>  7: 2022-08-29 04:57:11        1  2.20537971   9.1739304          0.015625
#>  8: 2022-08-29 04:57:11        1 -0.29187804  11.6956023          0.015625
#>  9: 2022-08-29 04:57:11        1  8.37322733   2.3682803          0.015625
#> 10: 2022-08-29 04:57:11        1  1.69358456  11.6295048          0.015625
#> 11: 2022-08-29 04:57:11        1 -2.45698581   2.2632501          0.015625
#> 12: 2022-08-29 04:57:11        1 -1.41301993  10.0591169          0.015625
#> 13: 2022-08-29 04:57:11        1 -4.64511715   0.4071126          0.015625
#> 14: 2022-08-29 04:57:11        1  1.66434331   6.3718726          0.015625
#> 15: 2022-08-29 04:57:11        1  9.25790003   3.5157802          0.015625
#> 16: 2022-08-29 04:57:11        1  3.06194744   4.4570226          0.015625
#> 17: 2022-08-29 04:57:11        1 -4.85657309   3.5815816          0.015625
#> 18: 2022-08-29 04:57:11        1  6.21703496   3.6314845          0.015625
#> 19: 2022-08-29 04:57:11        1  3.37551511   4.5274198          0.015625
#> 20: 2022-08-29 04:57:11        1  6.26269920  12.0534974          0.015625
#> 21: 2022-08-29 04:57:11        1 -4.37305455  14.2268460          0.015625
#> 22: 2022-08-29 04:57:11        1  2.30547449   1.6353809          0.015625
#> 23: 2022-08-29 04:57:11        1  4.22579476   6.9784876          0.015625
#> 24: 2022-08-29 04:57:11        1  9.59191561  14.1609771          0.015625
#> 25: 2022-08-29 04:57:11        1  1.61821049   5.2336509          0.015625
#> 26: 2022-08-29 04:57:11        1 -2.10513715   1.7258575          0.015625
#> 27: 2022-08-29 04:57:11        1 -0.57470714  14.8633940          0.015625
#> 28: 2022-08-29 04:57:11        1  4.15167850  12.3727821          0.015625
#> 29: 2022-08-29 04:57:11        1 -2.57630753  10.5134073          0.015625
#> 30: 2022-08-29 04:57:11        1  1.87075879   0.9407849          0.015625
#> 31: 2022-08-29 04:57:11        1 -0.11471011   2.3754188          0.015625
#> 32: 2022-08-29 04:57:11        1  0.05638359   0.4507534          0.015625
#> 33: 2022-08-29 04:57:11        1 -3.12649853   7.1654322          0.015625
#> 34: 2022-08-29 04:57:11        1  1.36077028  11.5383207          0.015625
#> 35: 2022-08-29 04:57:11        1  1.01098108  10.3673477          0.015625
#> 36: 2022-08-29 04:57:11        1  6.23749243   1.3606484          0.015625
#> 37: 2022-08-29 04:57:11        1  8.27746620   3.3206714          0.015625
#> 38: 2022-08-29 04:57:11        1 -2.42187955   0.8686659          0.015625
#> 39: 2022-08-29 04:57:11        1 -4.60450085  13.1387673          0.015625
#> 40: 2022-08-29 04:57:11        1 -3.57702205  14.5932060          0.015625
#> 41: 2022-08-29 04:57:11        1 -3.38520327   1.4719854          0.015625
#> 42: 2022-08-29 04:57:11        1  6.15765413  11.9180846          0.015625
#> 43: 2022-08-29 04:57:11        1  8.79778821  14.0515189          0.015625
#> 44: 2022-08-29 04:57:11        1 -2.23161306   7.8917610          0.015625
#> 45: 2022-08-29 04:57:11        1  7.49910586  14.0107568          0.015625
#> 46: 2022-08-29 04:57:11        1  5.99641589  13.7917349          0.015625
#> 47: 2022-08-29 04:57:11        1  5.59757876  11.3326648          0.015625
#> 48: 2022-08-29 04:57:11        1  6.85421897   5.0489103          0.015625
#> 49: 2022-08-29 04:57:11        1  5.26880707   6.6772867          0.015625
#> 50: 2022-08-29 04:57:11        1  3.10763952   4.2748845          0.015625
#> 51: 2022-08-29 04:57:11        1 -1.44645402  11.5674611          0.015625
#> 52: 2022-08-29 04:57:11        1  1.25691244   0.7720866          0.015625
#> 53: 2022-08-29 04:57:11        1 -0.76081594  12.7589941          0.015625
#> 54: 2022-08-29 04:57:11        1 -3.70946153  13.9991755          0.015625
#> 55: 2022-08-29 04:57:11        1  3.14002814  12.3274849          0.015625
#> 56: 2022-08-29 04:57:11        1  9.18088999  11.4121943          0.015625
#> 57: 2022-08-29 04:57:11        1 -3.87860680  12.7547498          0.015625
#> 58: 2022-08-29 04:57:11        1 -2.48719570   7.4563148          0.015625
#> 59: 2022-08-29 04:57:11        1 -3.56392485  11.9292969          0.015625
#> 60: 2022-08-29 04:57:11        1 -4.15768915   6.7302875          0.015625
#> 61: 2022-08-29 04:57:11        1 -1.46954704  14.3069720          0.015625
#> 62: 2022-08-29 04:57:11        1 -3.25490174   2.8412366          0.015625
#> 63: 2022-08-29 04:57:11        1  3.28975640   8.4369008          0.015625
#> 64: 2022-08-29 04:57:11        1  1.41816351   5.8919047          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity