Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain       y
#> 1: -2.800312 9.497625 0.015625 <list[3]> 2.39244

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain       y
#> 1: -2.800312 9.497625 0.015625 <list[3]> 2.39244

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  1.91522746  4.3688297 0.015625     0       6          1   8.458744
#>  2: -4.10889017  0.1310194 0.015625     0       6          1 171.672960
#>  3:  2.07979130  4.5674772 0.015625     0       6          1   8.364419
#>  4:  8.17344198  6.9424468 0.015625     0       6          1 148.518861
#>  5:  1.82922544  6.1999324 0.015625     0       6          1  16.596255
#>  6:  5.26036947  1.1626663 0.015625     0       6          1  22.206156
#>  7:  0.04615186  1.4408229 0.015625     0       6          1  39.714197
#>  8:  3.69389830  8.1150149 0.015625     0       6          1  59.198565
#>  9:  3.04298169  9.3338347 0.015625     0       6          1  62.731061
#> 10:  8.79964240  5.5153080 0.015625     0       6          1 126.302685
#> 11:  9.74881069 10.9486226 0.015625     0       6          1 308.625045
#> 12:  5.24962948  2.7415167 0.015625     0       6          1  32.969888
#> 13: -1.78664148 13.8609985 0.015625     0       6          1  32.143227
#> 14:  8.22106209  9.7740147 0.015625     0       6          1 225.007550
#> 15:  3.70320094 13.2302953 0.015625     0       6          1 163.225454
#> 16:  9.80460813  4.4988378 0.015625     0       6          1 125.352038
#> 17:  5.75536020  2.7940976 0.015625     0       6          1  42.655146
#> 18:  6.26941058  5.6075747 0.015625     0       6          1  89.777420
#> 19:  5.03552090 14.8283594 0.015625     0       6          1 271.069139
#> 20:  2.67641467  5.7703178 0.015625     0       6          1  15.932137
#> 21:  8.60194141 12.1784548 0.015625     0       6          1 313.010767
#> 22: -0.98237995 10.4504509 0.015625     0       6          1  23.493600
#> 23: -2.80031175  9.4976251 0.015625     0       6          1   2.392440
#> 24:  1.87691279 12.8647260 0.015625     0       6          1 102.044260
#> 25:  9.92240152 12.7391628 0.015625     0       6          1 381.967972
#> 26:  3.14323820  8.3367310 0.015625     0       6          1  49.897162
#> 27:  6.44846740 11.4697415 0.015625     0       6          1 228.396641
#> 28: -2.11538282 14.9844425 0.015625     0       6          1  35.057181
#> 29:  7.54840611  6.3379548 0.015625     0       6          1 125.242355
#> 30:  8.54778763  6.6697543 0.015625     0       6          1 148.519567
#> 31:  2.76015583 13.5797892 0.015625     0       6          1 138.879633
#> 32:  8.69818582 14.8656422 0.015625     0       6          1 418.288202
#> 33: -2.49387750  7.8259709 0.015625     0       6          1   7.792020
#> 34:  1.47262250  7.8287462 0.015625     0       6          1  27.798914
#> 35: -4.22376321  1.4151089 0.015625     0       6          1 146.051479
#> 36: -1.66820977 12.7081064 0.015625     0       6          1  24.807172
#> 37: -4.18639117 11.5308651 0.015625     0       6          1   7.970753
#> 38:  0.02760440  6.8920249 0.015625     0       6          1  20.474429
#> 39:  7.56237533 11.5077223 0.015625     0       6          1 261.933370
#> 40: -0.65214053  4.7284982 0.015625     0       6          1  23.025599
#> 41: -0.45000276  0.4784128 0.015625     0       6          1  57.633909
#> 42: -0.12264175  2.5234598 0.015625     0       6          1  33.014993
#> 43: -2.58746973  3.9277707 0.015625     0       6          1  42.745671
#> 44:  9.02835904  5.2406578 0.015625     0       6          1 124.430466
#> 45: -2.27527440  7.6963380 0.015625     0       6          1   8.124576
#> 46: -1.72978286  2.3642499 0.015625     0       6          1  50.480386
#> 47:  3.65161074  8.1829130 0.015625     0       6          1  59.146709
#> 48:  3.19341975  6.7350289 0.015625     0       6          1  30.704339
#> 49:  1.73842624 13.9120814 0.015625     0       6          1 120.460293
#> 50: -0.07975188 13.6440496 0.015625     0       6          1  76.075754
#> 51: -3.87363712 12.7396926 0.015625     0       6          1   2.870683
#> 52:  0.73852941  1.2581673 0.015625     0       6          1  29.939671
#> 53:  4.30622012  9.3791434 0.015625     0       6          1  99.571097
#> 54:  7.81332981  5.7728848 0.015625     0       6          1 117.122654
#> 55: -2.41313470  3.7855628 0.015625     0       6          1  41.698975
#> 56:  3.78150375  5.9346948 0.015625     0       6          1  32.696026
#> 57:  2.81787170 12.2654089 0.015625     0       6          1 111.273669
#> 58: -1.67310167 12.2110208 0.015625     0       6          1  21.005693
#> 59:  5.64325332  4.5536374 0.015625     0       6          1  60.683016
#> 60: -0.13998547  8.7607431 0.015625     0       6          1  25.946314
#> 61:  3.76557955  5.7566003 0.015625     0       6          1  30.443040
#> 62: -3.48264415 10.1546950 0.015625     0       6          1   4.052162
#> 63: -0.56830197 13.0203395 0.015625     0       6          1  55.375179
#> 64:  6.69837898 11.0358789 0.015625     0       6          1 223.766074
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-30 04:45:19        1  1.91522746   4.3688297          0.015625
#>  2: 2022-08-30 04:45:19        1 -4.10889017   0.1310194          0.015625
#>  3: 2022-08-30 04:45:19        1  2.07979130   4.5674772          0.015625
#>  4: 2022-08-30 04:45:19        1  8.17344198   6.9424468          0.015625
#>  5: 2022-08-30 04:45:19        1  1.82922544   6.1999324          0.015625
#>  6: 2022-08-30 04:45:19        1  5.26036947   1.1626663          0.015625
#>  7: 2022-08-30 04:45:19        1  0.04615186   1.4408229          0.015625
#>  8: 2022-08-30 04:45:19        1  3.69389830   8.1150149          0.015625
#>  9: 2022-08-30 04:45:19        1  3.04298169   9.3338347          0.015625
#> 10: 2022-08-30 04:45:19        1  8.79964240   5.5153080          0.015625
#> 11: 2022-08-30 04:45:19        1  9.74881069  10.9486226          0.015625
#> 12: 2022-08-30 04:45:19        1  5.24962948   2.7415167          0.015625
#> 13: 2022-08-30 04:45:19        1 -1.78664148  13.8609985          0.015625
#> 14: 2022-08-30 04:45:19        1  8.22106209   9.7740147          0.015625
#> 15: 2022-08-30 04:45:19        1  3.70320094  13.2302953          0.015625
#> 16: 2022-08-30 04:45:19        1  9.80460813   4.4988378          0.015625
#> 17: 2022-08-30 04:45:19        1  5.75536020   2.7940976          0.015625
#> 18: 2022-08-30 04:45:19        1  6.26941058   5.6075747          0.015625
#> 19: 2022-08-30 04:45:19        1  5.03552090  14.8283594          0.015625
#> 20: 2022-08-30 04:45:19        1  2.67641467   5.7703178          0.015625
#> 21: 2022-08-30 04:45:19        1  8.60194141  12.1784548          0.015625
#> 22: 2022-08-30 04:45:19        1 -0.98237995  10.4504509          0.015625
#> 23: 2022-08-30 04:45:19        1 -2.80031175   9.4976251          0.015625
#> 24: 2022-08-30 04:45:19        1  1.87691279  12.8647260          0.015625
#> 25: 2022-08-30 04:45:19        1  9.92240152  12.7391628          0.015625
#> 26: 2022-08-30 04:45:19        1  3.14323820   8.3367310          0.015625
#> 27: 2022-08-30 04:45:19        1  6.44846740  11.4697415          0.015625
#> 28: 2022-08-30 04:45:19        1 -2.11538282  14.9844425          0.015625
#> 29: 2022-08-30 04:45:19        1  7.54840611   6.3379548          0.015625
#> 30: 2022-08-30 04:45:19        1  8.54778763   6.6697543          0.015625
#> 31: 2022-08-30 04:45:19        1  2.76015583  13.5797892          0.015625
#> 32: 2022-08-30 04:45:19        1  8.69818582  14.8656422          0.015625
#> 33: 2022-08-30 04:45:19        1 -2.49387750   7.8259709          0.015625
#> 34: 2022-08-30 04:45:19        1  1.47262250   7.8287462          0.015625
#> 35: 2022-08-30 04:45:19        1 -4.22376321   1.4151089          0.015625
#> 36: 2022-08-30 04:45:19        1 -1.66820977  12.7081064          0.015625
#> 37: 2022-08-30 04:45:19        1 -4.18639117  11.5308651          0.015625
#> 38: 2022-08-30 04:45:19        1  0.02760440   6.8920249          0.015625
#> 39: 2022-08-30 04:45:19        1  7.56237533  11.5077223          0.015625
#> 40: 2022-08-30 04:45:19        1 -0.65214053   4.7284982          0.015625
#> 41: 2022-08-30 04:45:19        1 -0.45000276   0.4784128          0.015625
#> 42: 2022-08-30 04:45:19        1 -0.12264175   2.5234598          0.015625
#> 43: 2022-08-30 04:45:19        1 -2.58746973   3.9277707          0.015625
#> 44: 2022-08-30 04:45:19        1  9.02835904   5.2406578          0.015625
#> 45: 2022-08-30 04:45:19        1 -2.27527440   7.6963380          0.015625
#> 46: 2022-08-30 04:45:19        1 -1.72978286   2.3642499          0.015625
#> 47: 2022-08-30 04:45:19        1  3.65161074   8.1829130          0.015625
#> 48: 2022-08-30 04:45:19        1  3.19341975   6.7350289          0.015625
#> 49: 2022-08-30 04:45:19        1  1.73842624  13.9120814          0.015625
#> 50: 2022-08-30 04:45:19        1 -0.07975188  13.6440496          0.015625
#> 51: 2022-08-30 04:45:19        1 -3.87363712  12.7396926          0.015625
#> 52: 2022-08-30 04:45:19        1  0.73852941   1.2581673          0.015625
#> 53: 2022-08-30 04:45:19        1  4.30622012   9.3791434          0.015625
#> 54: 2022-08-30 04:45:19        1  7.81332981   5.7728848          0.015625
#> 55: 2022-08-30 04:45:19        1 -2.41313470   3.7855628          0.015625
#> 56: 2022-08-30 04:45:19        1  3.78150375   5.9346948          0.015625
#> 57: 2022-08-30 04:45:19        1  2.81787170  12.2654089          0.015625
#> 58: 2022-08-30 04:45:19        1 -1.67310167  12.2110208          0.015625
#> 59: 2022-08-30 04:45:19        1  5.64325332   4.5536374          0.015625
#> 60: 2022-08-30 04:45:19        1 -0.13998547   8.7607431          0.015625
#> 61: 2022-08-30 04:45:19        1  3.76557955   5.7566003          0.015625
#> 62: 2022-08-30 04:45:19        1 -3.48264415  10.1546950          0.015625
#> 63: 2022-08-30 04:45:19        1 -0.56830197  13.0203395          0.015625
#> 64: 2022-08-30 04:45:19        1  6.69837898  11.0358789          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity