Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# Branin function with fidelity, see `bbotk::branin()`
fun = function(xs) branin_wu(xs[["x1"]], xs[["x2"]], xs[["fidelity"]])

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.175857 12.09256 0.015625 <list[3]> 0.9333776

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.175857 12.09256 0.015625 <list[3]> 0.9333776

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1:  4.7426139 10.6110706 0.015625     0       6          1 141.7960615
#>  2: -4.3972312 10.0049006 0.015625     0       6          1  19.8975795
#>  3:  0.7011925  9.5414604 0.015625     0       6          1  38.8879056
#>  4: -3.2574664  5.1853258 0.015625     0       6          1  40.4723572
#>  5:  3.5895523 13.7977285 0.015625     0       6          1 173.3355886
#>  6: -3.3200729  5.7129342 0.015625     0       6          1  35.4790721
#>  7:  4.2285674 12.3653566 0.015625     0       6          1 162.9242562
#>  8:  4.6639933  8.8994151 0.015625     0       6          1 102.7266313
#>  9:  4.6719056  3.7787052 0.015625     0       6          1  30.2517572
#> 10: -4.7241396  7.7921977 0.015625     0       6          1  51.2355534
#> 11:  6.3962402 13.4275196 0.015625     0       6          1 286.8480691
#> 12:  1.5543269  9.2177316 0.015625     0       6          1  41.7114855
#> 13: -1.2432884 10.5503219 0.015625     0       6          1  19.4596241
#> 14:  8.8974255  1.0727688 0.015625     0       6          1  47.9342403
#> 15:  7.4817825  5.2057113 0.015625     0       6          1 101.7055681
#> 16: -3.6634734  0.9645445 0.015625     0       6          1 128.8854166
#> 17: -3.1651364  1.4957466 0.015625     0       6          1  97.4181361
#> 18:  2.1966953  6.6713770 0.015625     0       6          1  20.5284581
#> 19: -1.8182767  5.6489783 0.015625     0       6          1  18.8472627
#> 20:  3.3635667 12.6702497 0.015625     0       6          1 136.9548325
#> 21:  9.5549490 11.4490008 0.015625     0       6          1 319.0681805
#> 22: -1.2485093 13.5163480 0.015625     0       6          1  43.0866051
#> 23:  5.2437654  8.0467749 0.015625     0       6          1 106.0114439
#> 24:  9.5466011  6.6004303 0.015625     0       6          1 169.2639598
#> 25: -1.1286183 14.1686103 0.015625     0       6          1  54.2181718
#> 26:  9.1053392 10.1510799 0.015625     0       6          1 259.8853131
#> 27:  7.8749147 13.3873212 0.015625     0       6          1 334.2990117
#> 28: -3.0172254  1.3627241 0.015625     0       6          1  94.9358547
#> 29:  2.0776262  2.7352313 0.015625     0       6          1   5.3473100
#> 30: -2.0851933  0.7163389 0.015625     0       6          1  81.5940287
#> 31:  0.5792977  5.9280234 0.015625     0       6          1  18.7405751
#> 32: -1.9416573  4.4916521 0.015625     0       6          1  28.7466137
#> 33:  0.7311709 12.9488301 0.015625     0       6          1  82.6943721
#> 34: -3.5376324  4.1626846 0.015625     0       6          1  62.8017289
#> 35: -1.3371301  8.9508974 0.015625     0       6          1  12.8128681
#> 36: -4.5404101  4.2397318 0.015625     0       6          1 100.9091498
#> 37:  3.0624594  2.2916725 0.015625     0       6          1   1.1977009
#> 38: -4.2554749 14.8950895 0.015625     0       6          1   8.2144946
#> 39:  8.8170923  9.5845655 0.015625     0       6          1 233.9814200
#> 40: -2.0294887 11.5318405 0.015625     0       6          1  10.4797684
#> 41:  3.1876285  7.7671443 0.015625     0       6          1  43.0227543
#> 42:  3.6341532 11.3784126 0.015625     0       6          1 117.2366650
#> 43:  4.3537139 11.9826514 0.015625     0       6          1 158.6335282
#> 44:  5.5351318  2.1273887 0.015625     0       6          1  32.9969715
#> 45:  9.7250078  4.6475890 0.015625     0       6          1 126.6596508
#> 46:  5.4099297  7.7787914 0.015625     0       6          1 106.2103866
#> 47:  3.7292341  8.9241865 0.015625     0       6          1  73.1046027
#> 48:  0.7401392 10.0419114 0.015625     0       6          1  44.1615415
#> 49: -2.1396354  5.9533277 0.015625     0       6          1  17.7357684
#> 50:  6.0255627  6.2131460 0.015625     0       6          1  94.7454236
#> 51:  9.2330286 14.2055515 0.015625     0       6          1 411.8209074
#> 52:  3.2737319  6.4510635 0.015625     0       6          1  28.9101567
#> 53: -1.1471548  7.5508345 0.015625     0       6          1  14.0467281
#> 54: -1.6615114  9.6787805 0.015625     0       6          1  10.0317297
#> 55: -0.6235262  9.7575777 0.015625     0       6          1  25.3756212
#> 56:  1.4757196  2.7346588 0.015625     0       6          1  11.8790702
#> 57:  7.1716417 13.1296673 0.015625     0       6          1 303.7746726
#> 58:  5.4989269 11.9808505 0.015625     0       6          1 207.3183403
#> 59: -4.6819812  9.7993044 0.015625     0       6          1  28.4249711
#> 60:  3.7881466  8.9263572 0.015625     0       6          1  74.8268290
#> 61: -2.2527749 14.4267316 0.015625     0       6          1  25.8994228
#> 62: -1.7817258  7.9409973 0.015625     0       6          1   8.9743062
#> 63: -0.1030907 10.1592463 0.015625     0       6          1  35.5099248
#> 64: -3.1758569 12.0925605 0.015625     0       6          1   0.9333776
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-11-18 04:16:37        1   4.7426139  10.6110706          0.015625
#>  2: 2022-11-18 04:16:37        1  -4.3972312  10.0049006          0.015625
#>  3: 2022-11-18 04:16:37        1   0.7011925   9.5414604          0.015625
#>  4: 2022-11-18 04:16:37        1  -3.2574664   5.1853258          0.015625
#>  5: 2022-11-18 04:16:37        1   3.5895523  13.7977285          0.015625
#>  6: 2022-11-18 04:16:37        1  -3.3200729   5.7129342          0.015625
#>  7: 2022-11-18 04:16:37        1   4.2285674  12.3653566          0.015625
#>  8: 2022-11-18 04:16:37        1   4.6639933   8.8994151          0.015625
#>  9: 2022-11-18 04:16:37        1   4.6719056   3.7787052          0.015625
#> 10: 2022-11-18 04:16:37        1  -4.7241396   7.7921977          0.015625
#> 11: 2022-11-18 04:16:37        1   6.3962402  13.4275196          0.015625
#> 12: 2022-11-18 04:16:37        1   1.5543269   9.2177316          0.015625
#> 13: 2022-11-18 04:16:37        1  -1.2432884  10.5503219          0.015625
#> 14: 2022-11-18 04:16:37        1   8.8974255   1.0727688          0.015625
#> 15: 2022-11-18 04:16:37        1   7.4817825   5.2057113          0.015625
#> 16: 2022-11-18 04:16:37        1  -3.6634734   0.9645445          0.015625
#> 17: 2022-11-18 04:16:37        1  -3.1651364   1.4957466          0.015625
#> 18: 2022-11-18 04:16:37        1   2.1966953   6.6713770          0.015625
#> 19: 2022-11-18 04:16:37        1  -1.8182767   5.6489783          0.015625
#> 20: 2022-11-18 04:16:37        1   3.3635667  12.6702497          0.015625
#> 21: 2022-11-18 04:16:37        1   9.5549490  11.4490008          0.015625
#> 22: 2022-11-18 04:16:37        1  -1.2485093  13.5163480          0.015625
#> 23: 2022-11-18 04:16:37        1   5.2437654   8.0467749          0.015625
#> 24: 2022-11-18 04:16:37        1   9.5466011   6.6004303          0.015625
#> 25: 2022-11-18 04:16:37        1  -1.1286183  14.1686103          0.015625
#> 26: 2022-11-18 04:16:37        1   9.1053392  10.1510799          0.015625
#> 27: 2022-11-18 04:16:37        1   7.8749147  13.3873212          0.015625
#> 28: 2022-11-18 04:16:37        1  -3.0172254   1.3627241          0.015625
#> 29: 2022-11-18 04:16:37        1   2.0776262   2.7352313          0.015625
#> 30: 2022-11-18 04:16:37        1  -2.0851933   0.7163389          0.015625
#> 31: 2022-11-18 04:16:37        1   0.5792977   5.9280234          0.015625
#> 32: 2022-11-18 04:16:37        1  -1.9416573   4.4916521          0.015625
#> 33: 2022-11-18 04:16:37        1   0.7311709  12.9488301          0.015625
#> 34: 2022-11-18 04:16:37        1  -3.5376324   4.1626846          0.015625
#> 35: 2022-11-18 04:16:37        1  -1.3371301   8.9508974          0.015625
#> 36: 2022-11-18 04:16:37        1  -4.5404101   4.2397318          0.015625
#> 37: 2022-11-18 04:16:37        1   3.0624594   2.2916725          0.015625
#> 38: 2022-11-18 04:16:37        1  -4.2554749  14.8950895          0.015625
#> 39: 2022-11-18 04:16:37        1   8.8170923   9.5845655          0.015625
#> 40: 2022-11-18 04:16:37        1  -2.0294887  11.5318405          0.015625
#> 41: 2022-11-18 04:16:37        1   3.1876285   7.7671443          0.015625
#> 42: 2022-11-18 04:16:37        1   3.6341532  11.3784126          0.015625
#> 43: 2022-11-18 04:16:37        1   4.3537139  11.9826514          0.015625
#> 44: 2022-11-18 04:16:37        1   5.5351318   2.1273887          0.015625
#> 45: 2022-11-18 04:16:37        1   9.7250078   4.6475890          0.015625
#> 46: 2022-11-18 04:16:37        1   5.4099297   7.7787914          0.015625
#> 47: 2022-11-18 04:16:37        1   3.7292341   8.9241865          0.015625
#> 48: 2022-11-18 04:16:37        1   0.7401392  10.0419114          0.015625
#> 49: 2022-11-18 04:16:37        1  -2.1396354   5.9533277          0.015625
#> 50: 2022-11-18 04:16:37        1   6.0255627   6.2131460          0.015625
#> 51: 2022-11-18 04:16:37        1   9.2330286  14.2055515          0.015625
#> 52: 2022-11-18 04:16:37        1   3.2737319   6.4510635          0.015625
#> 53: 2022-11-18 04:16:37        1  -1.1471548   7.5508345          0.015625
#> 54: 2022-11-18 04:16:37        1  -1.6615114   9.6787805          0.015625
#> 55: 2022-11-18 04:16:37        1  -0.6235262   9.7575777          0.015625
#> 56: 2022-11-18 04:16:37        1   1.4757196   2.7346588          0.015625
#> 57: 2022-11-18 04:16:37        1   7.1716417  13.1296673          0.015625
#> 58: 2022-11-18 04:16:37        1   5.4989269  11.9808505          0.015625
#> 59: 2022-11-18 04:16:37        1  -4.6819812   9.7993044          0.015625
#> 60: 2022-11-18 04:16:37        1   3.7881466   8.9263572          0.015625
#> 61: 2022-11-18 04:16:37        1  -2.2527749  14.4267316          0.015625
#> 62: 2022-11-18 04:16:37        1  -1.7817258   7.9409973          0.015625
#> 63: 2022-11-18 04:16:37        1  -0.1030907  10.1592463          0.015625
#> 64: 2022-11-18 04:16:37        1  -3.1758569  12.0925605          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity