Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# Branin function with fidelity, see `bbotk::branin()`
fun = function(xs) branin_wu(xs[["x1"]], xs[["x2"]], xs[["fidelity"]])

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.542877 11.82057 0.015625 <list[3]> 1.202307

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.542877 11.82057 0.015625 <list[3]> 1.202307

# all evaluations
as.data.table(instance$archive)
#>             x1          x2 fidelity stage bracket repetition          y
#>  1:  2.5012521  1.29821954 0.015625     0       6          1   3.134211
#>  2:  9.6341152 14.41953706 0.015625     0       6          1 437.371023
#>  3:  9.5412377  0.67432842 0.015625     0       6          1  50.315274
#>  4:  8.7727244  5.39165966 0.015625     0       6          1 123.094625
#>  5: -3.4939838 10.86910102 0.015625     0       6          1   2.126647
#>  6: -0.3958601 14.30931971 0.015625     0       6          1  77.757035
#>  7: -1.0147091  9.57360324 0.015625     0       6          1  18.781912
#>  8: -2.2427365  8.48093845 0.015625     0       6          1   5.568031
#>  9: -0.9134055  2.36803848 0.015625     0       6          1  41.993243
#> 10:  4.2881525 14.51762141 0.015625     0       6          1 224.408567
#> 11:  9.9234922  2.07479732 0.015625     0       6          1  79.725285
#> 12: -3.5428770 11.82056853 0.015625     0       6          1   1.202307
#> 13:  5.6551748  1.73380394 0.015625     0       6          1  31.839817
#> 14: -4.7046462 14.29426205 0.015625     0       6          1   9.941540
#> 15: -2.9017187 12.34312990 0.015625     0       6          1   2.822008
#> 16:  5.9333066  9.95937224 0.015625     0       6          1 170.805260
#> 17:  4.9347359  1.99654191 0.015625     0       6          1  21.737863
#> 18:  2.2116744  5.58255294 0.015625     0       6          1  12.974050
#> 19:  4.7050660  2.38942541 0.015625     0       6          1  20.151159
#> 20: -2.3873104  6.79499718 0.015625     0       6          1  13.113209
#> 21: -0.1472083  0.76900133 0.015625     0       6          1  49.374916
#> 22:  9.3798641  4.51735570 0.015625     0       6          1 115.770086
#> 23: -1.2009797  2.87260021 0.015625     0       6          1  39.309210
#> 24: -0.3600231  0.09075161 0.015625     0       6          1  61.057667
#> 25:  1.1613698  7.06850756 0.015625     0       6          1  22.090447
#> 26: -2.9428129 12.50098988 0.015625     0       6          1   2.992832
#> 27:  6.3343995  2.30553973 0.015625     0       6          1  46.146412
#> 28: -4.2729234  9.59427901 0.015625     0       6          1  20.110023
#> 29: -1.9886418  9.47592215 0.015625     0       6          1   6.139373
#> 30:  2.4156527  8.32812683 0.015625     0       6          1  38.738902
#> 31:  6.0230846 10.88447813 0.015625     0       6          1 197.637536
#> 32:  9.1982561  8.16466205 0.015625     0       6          1 202.359903
#> 33:  0.6784966 12.05515215 0.015625     0       6          1  68.182034
#> 34:  7.6002823 11.86647017 0.015625     0       6          1 274.416537
#> 35:  2.2320301 10.98848004 0.015625     0       6          1  74.456704
#> 36:  0.3616684  6.31471042 0.015625     0       6          1  19.766461
#> 37:  4.7045673  6.28749847 0.015625     0       6          1  60.257221
#> 38:  1.2997499 12.79999763 0.015625     0       6          1  90.304562
#> 39:  2.2240691  2.22144182 0.015625     0       6          1   4.316773
#> 40:  8.8822227 13.04889551 0.015625     0       6          1 353.700962
#> 41: -4.7888194  7.32009148 0.015625     0       6          1  59.826549
#> 42: -3.1801017  7.36580862 0.015625     0       6          1  16.456459
#> 43: -2.6838421 13.03726195 0.015625     0       6          1   7.860018
#> 44:  3.8263960  7.17543251 0.015625     0       6          1  49.009077
#> 45:  1.3349286  6.69059224 0.015625     0       6          1  19.863722
#> 46:  0.8621303 14.39411520 0.015625     0       6          1 111.182811
#> 47: -2.7465808 12.04998964 0.015625     0       6          1   3.230337
#> 48:  1.0755025 10.49343777 0.015625     0       6          1  52.627605
#> 49:  0.3564877 11.25075178 0.015625     0       6          1  52.803475
#> 50:  2.0778199 10.14125799 0.015625     0       6          1  58.853472
#> 51:  1.2155025 14.95447155 0.015625     0       6          1 130.923406
#> 52:  2.6178907  9.71360846 0.015625     0       6          1  60.504378
#> 53:  7.0178014  9.80023312 0.015625     0       6          1 198.166272
#> 54:  7.1698402 14.06494052 0.015625     0       6          1 336.317393
#> 55:  9.0430899  2.95420126 0.015625     0       6          1  79.098637
#> 56:  0.5565899  0.46762281 0.015625     0       6          1  39.831698
#> 57:  9.7342402 10.96585943 0.015625     0       6          1 308.678853
#> 58:  4.0869435  1.18228592 0.015625     0       6          1   5.754966
#> 59:  5.2010765 14.09513493 0.015625     0       6          1 256.035593
#> 60:  6.3383950  3.83415265 0.015625     0       6          1  64.300145
#> 61:  5.7765812  5.18551647 0.015625     0       6          1  72.466162
#> 62:  7.3847843  1.21182741 0.015625     0       6          1  42.307775
#> 63: -2.7459552 14.13277913 0.015625     0       6          1  13.604882
#> 64:  9.4636487 13.07301447 0.015625     0       6          1 376.034394
#>             x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2023-01-18 08:36:58        1   2.5012521  1.29821954          0.015625
#>  2: 2023-01-18 08:36:58        1   9.6341152 14.41953706          0.015625
#>  3: 2023-01-18 08:36:58        1   9.5412377  0.67432842          0.015625
#>  4: 2023-01-18 08:36:58        1   8.7727244  5.39165966          0.015625
#>  5: 2023-01-18 08:36:58        1  -3.4939838 10.86910102          0.015625
#>  6: 2023-01-18 08:36:58        1  -0.3958601 14.30931971          0.015625
#>  7: 2023-01-18 08:36:58        1  -1.0147091  9.57360324          0.015625
#>  8: 2023-01-18 08:36:58        1  -2.2427365  8.48093845          0.015625
#>  9: 2023-01-18 08:36:58        1  -0.9134055  2.36803848          0.015625
#> 10: 2023-01-18 08:36:58        1   4.2881525 14.51762141          0.015625
#> 11: 2023-01-18 08:36:58        1   9.9234922  2.07479732          0.015625
#> 12: 2023-01-18 08:36:58        1  -3.5428770 11.82056853          0.015625
#> 13: 2023-01-18 08:36:58        1   5.6551748  1.73380394          0.015625
#> 14: 2023-01-18 08:36:58        1  -4.7046462 14.29426205          0.015625
#> 15: 2023-01-18 08:36:58        1  -2.9017187 12.34312990          0.015625
#> 16: 2023-01-18 08:36:58        1   5.9333066  9.95937224          0.015625
#> 17: 2023-01-18 08:36:58        1   4.9347359  1.99654191          0.015625
#> 18: 2023-01-18 08:36:58        1   2.2116744  5.58255294          0.015625
#> 19: 2023-01-18 08:36:58        1   4.7050660  2.38942541          0.015625
#> 20: 2023-01-18 08:36:58        1  -2.3873104  6.79499718          0.015625
#> 21: 2023-01-18 08:36:58        1  -0.1472083  0.76900133          0.015625
#> 22: 2023-01-18 08:36:58        1   9.3798641  4.51735570          0.015625
#> 23: 2023-01-18 08:36:58        1  -1.2009797  2.87260021          0.015625
#> 24: 2023-01-18 08:36:58        1  -0.3600231  0.09075161          0.015625
#> 25: 2023-01-18 08:36:58        1   1.1613698  7.06850756          0.015625
#> 26: 2023-01-18 08:36:58        1  -2.9428129 12.50098988          0.015625
#> 27: 2023-01-18 08:36:58        1   6.3343995  2.30553973          0.015625
#> 28: 2023-01-18 08:36:58        1  -4.2729234  9.59427901          0.015625
#> 29: 2023-01-18 08:36:58        1  -1.9886418  9.47592215          0.015625
#> 30: 2023-01-18 08:36:58        1   2.4156527  8.32812683          0.015625
#> 31: 2023-01-18 08:36:58        1   6.0230846 10.88447813          0.015625
#> 32: 2023-01-18 08:36:58        1   9.1982561  8.16466205          0.015625
#> 33: 2023-01-18 08:36:58        1   0.6784966 12.05515215          0.015625
#> 34: 2023-01-18 08:36:58        1   7.6002823 11.86647017          0.015625
#> 35: 2023-01-18 08:36:58        1   2.2320301 10.98848004          0.015625
#> 36: 2023-01-18 08:36:58        1   0.3616684  6.31471042          0.015625
#> 37: 2023-01-18 08:36:58        1   4.7045673  6.28749847          0.015625
#> 38: 2023-01-18 08:36:58        1   1.2997499 12.79999763          0.015625
#> 39: 2023-01-18 08:36:58        1   2.2240691  2.22144182          0.015625
#> 40: 2023-01-18 08:36:58        1   8.8822227 13.04889551          0.015625
#> 41: 2023-01-18 08:36:58        1  -4.7888194  7.32009148          0.015625
#> 42: 2023-01-18 08:36:58        1  -3.1801017  7.36580862          0.015625
#> 43: 2023-01-18 08:36:58        1  -2.6838421 13.03726195          0.015625
#> 44: 2023-01-18 08:36:58        1   3.8263960  7.17543251          0.015625
#> 45: 2023-01-18 08:36:58        1   1.3349286  6.69059224          0.015625
#> 46: 2023-01-18 08:36:58        1   0.8621303 14.39411520          0.015625
#> 47: 2023-01-18 08:36:58        1  -2.7465808 12.04998964          0.015625
#> 48: 2023-01-18 08:36:58        1   1.0755025 10.49343777          0.015625
#> 49: 2023-01-18 08:36:58        1   0.3564877 11.25075178          0.015625
#> 50: 2023-01-18 08:36:58        1   2.0778199 10.14125799          0.015625
#> 51: 2023-01-18 08:36:58        1   1.2155025 14.95447155          0.015625
#> 52: 2023-01-18 08:36:58        1   2.6178907  9.71360846          0.015625
#> 53: 2023-01-18 08:36:58        1   7.0178014  9.80023312          0.015625
#> 54: 2023-01-18 08:36:58        1   7.1698402 14.06494052          0.015625
#> 55: 2023-01-18 08:36:58        1   9.0430899  2.95420126          0.015625
#> 56: 2023-01-18 08:36:58        1   0.5565899  0.46762281          0.015625
#> 57: 2023-01-18 08:36:58        1   9.7342402 10.96585943          0.015625
#> 58: 2023-01-18 08:36:58        1   4.0869435  1.18228592          0.015625
#> 59: 2023-01-18 08:36:58        1   5.2010765 14.09513493          0.015625
#> 60: 2023-01-18 08:36:58        1   6.3383950  3.83415265          0.015625
#> 61: 2023-01-18 08:36:58        1   5.7765812  5.18551647          0.015625
#> 62: 2023-01-18 08:36:58        1   7.3847843  1.21182741          0.015625
#> 63: 2023-01-18 08:36:58        1  -2.7459552 14.13277913          0.015625
#> 64: 2023-01-18 08:36:58        1   9.4636487 13.07301447          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity