Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1      x2 fidelity  x_domain        y
#> 1: 3.085123 2.21173 0.015625 <list[3]> 1.100776

# best scoring evaluation
instance$result
#>          x1      x2 fidelity  x_domain        y
#> 1: 3.085123 2.21173 0.015625 <list[3]> 1.100776

# all evaluations
as.data.table(instance$archive)
#>               x1         x2 fidelity stage bracket repetition          y
#>  1:  8.699336649  3.0618686 0.015625     0       6          1  76.439014
#>  2:  2.707476809 11.4557772 0.015625     0       6          1  92.290076
#>  3:  0.330769896  9.9607326 0.015625     0       6          1  39.186116
#>  4:  3.946582150  5.0855396 0.015625     0       6          1  27.235377
#>  5:  7.845559715 11.3848106 0.015625     0       6          1 265.404258
#>  6: -0.202115958  8.6027068 0.015625     0       6          1  24.604018
#>  7: -2.544226598  6.4969507 0.015625     0       6          1  16.133332
#>  8: -3.496957249  2.4081584 0.015625     0       6          1  91.883893
#>  9:  4.590252086 13.7738796 0.015625     0       6          1 217.102388
#> 10: -3.746623219 12.7051235 0.015625     0       6          1   2.198873
#> 11:  6.058681572  7.4629169 0.015625     0       6          1 118.900845
#> 12: -1.484428616  8.2381495 0.015625     0       6          1  10.865201
#> 13:  7.495361415 10.2642598 0.015625     0       6          1 222.638727
#> 14: -0.010489440  7.8381920 0.015625     0       6          1  22.919425
#> 15: -4.267004782 10.9650534 0.015625     0       6          1  11.555963
#> 16:  7.540410677  9.4778051 0.015625     0       6          1 201.489416
#> 17: -1.396829482  4.8405053 0.015625     0       6          1  23.513595
#> 18: -2.055296851  5.8759972 0.015625     0       6          1  17.953229
#> 19: -0.134841906  8.4408628 0.015625     0       6          1  24.468674
#> 20:  3.175885582 11.6836303 0.015625     0       6          1 109.148535
#> 21: -2.166683453  0.9543124 0.015625     0       6          1  79.233072
#> 22:  3.255388140  8.1492684 0.015625     0       6          1  49.523523
#> 23:  0.828454117 11.9480361 0.015625     0       6          1  68.987855
#> 24:  3.085122880  2.2117301 0.015625     0       6          1   1.100776
#> 25:  0.579433086  3.9044017 0.015625     0       6          1  19.435995
#> 26:  2.636541625  5.0071124 0.015625     0       6          1  10.534215
#> 27:  8.827508979 13.1004595 0.015625     0       6          1 353.769711
#> 28:  1.603082357  3.2024753 0.015625     0       6          1   9.795766
#> 29:  1.818162677  0.5287545 0.015625     0       6          1  14.826972
#> 30:  6.712094832  3.0859237 0.015625     0       6          1  59.479297
#> 31:  7.992673602  2.4991927 0.015625     0       6          1  61.318029
#> 32:  0.322158035  5.0650464 0.015625     0       6          1  19.289103
#> 33:  4.755326094  3.7697790 0.015625     0       6          1  31.968026
#> 34: -1.256102417  7.7866127 0.015625     0       6          1  13.040244
#> 35:  0.931258184  1.0346820 0.015625     0       6          1  28.049754
#> 36:  0.937008809  0.3737334 0.015625     0       6          1  33.008342
#> 37:  9.759694482  2.2611390 0.015625     0       6          1  79.528001
#> 38:  8.037929572 12.4799544 0.015625     0       6          1 307.056515
#> 39:  4.305426722  2.6108755 0.015625     0       6          1  14.570039
#> 40:  2.539484921  6.4626915 0.015625     0       6          1  20.629152
#> 41:  8.122085728  4.3309836 0.015625     0       6          1  92.637401
#> 42:  3.452329619  7.2276816 0.015625     0       6          1  41.253612
#> 43: -0.009261159  2.1259005 0.015625     0       6          1  34.724791
#> 44: -3.698238260  2.4569333 0.015625     0       6          1  98.860572
#> 45: -4.493551842  4.6787973 0.015625     0       6          1  90.611838
#> 46:  6.179093241  8.9177917 0.015625     0       6          1 153.604080
#> 47:  2.899468985 10.3211236 0.015625     0       6          1  75.973232
#> 48:  5.061437417  2.5694506 0.015625     0       6          1  28.008807
#> 49:  9.441265928  9.3251164 0.015625     0       6          1 244.091468
#> 50:  5.931964645  4.2013081 0.015625     0       6          1  62.054667
#> 51:  1.309084403  2.2658013 0.015625     0       6          1  15.386029
#> 52:  8.259936121 12.4505558 0.015625     0       6          1 312.419325
#> 53:  9.017674626  6.9317565 0.015625     0       6          1 164.601345
#> 54: -2.151383769  3.9239414 0.015625     0       6          1  36.569810
#> 55: -3.002820483  4.3132306 0.015625     0       6          1  45.960272
#> 56:  6.555658417  7.3800289 0.015625     0       6          1 129.335806
#> 57:  1.582611860  1.5295918 0.015625     0       6          1  14.001823
#> 58:  8.986708970 10.0717834 0.015625     0       6          1 253.841745
#> 59: -1.089781349 11.5452124 0.015625     0       6          1  28.687704
#> 60:  1.932439987  2.3921721 0.015625     0       6          1   7.021361
#> 61:  8.089381577  5.7271814 0.015625     0       6          1 119.904480
#> 62:  6.995261334  9.7451312 0.015625     0       6          1 196.129292
#> 63:  8.319206804 14.4983792 0.015625     0       6          1 390.276996
#> 64: -4.050101915  5.3095960 0.015625     0       6          1  62.475663
#>               x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-11-02 04:38:01        1  8.699336649   3.0618686          0.015625
#>  2: 2022-11-02 04:38:01        1  2.707476809  11.4557772          0.015625
#>  3: 2022-11-02 04:38:01        1  0.330769896   9.9607326          0.015625
#>  4: 2022-11-02 04:38:01        1  3.946582150   5.0855396          0.015625
#>  5: 2022-11-02 04:38:01        1  7.845559715  11.3848106          0.015625
#>  6: 2022-11-02 04:38:01        1 -0.202115958   8.6027068          0.015625
#>  7: 2022-11-02 04:38:01        1 -2.544226598   6.4969507          0.015625
#>  8: 2022-11-02 04:38:01        1 -3.496957249   2.4081584          0.015625
#>  9: 2022-11-02 04:38:01        1  4.590252086  13.7738796          0.015625
#> 10: 2022-11-02 04:38:01        1 -3.746623219  12.7051235          0.015625
#> 11: 2022-11-02 04:38:01        1  6.058681572   7.4629169          0.015625
#> 12: 2022-11-02 04:38:01        1 -1.484428616   8.2381495          0.015625
#> 13: 2022-11-02 04:38:01        1  7.495361415  10.2642598          0.015625
#> 14: 2022-11-02 04:38:01        1 -0.010489440   7.8381920          0.015625
#> 15: 2022-11-02 04:38:01        1 -4.267004782  10.9650534          0.015625
#> 16: 2022-11-02 04:38:01        1  7.540410677   9.4778051          0.015625
#> 17: 2022-11-02 04:38:01        1 -1.396829482   4.8405053          0.015625
#> 18: 2022-11-02 04:38:01        1 -2.055296851   5.8759972          0.015625
#> 19: 2022-11-02 04:38:01        1 -0.134841906   8.4408628          0.015625
#> 20: 2022-11-02 04:38:01        1  3.175885582  11.6836303          0.015625
#> 21: 2022-11-02 04:38:01        1 -2.166683453   0.9543124          0.015625
#> 22: 2022-11-02 04:38:01        1  3.255388140   8.1492684          0.015625
#> 23: 2022-11-02 04:38:01        1  0.828454117  11.9480361          0.015625
#> 24: 2022-11-02 04:38:01        1  3.085122880   2.2117301          0.015625
#> 25: 2022-11-02 04:38:01        1  0.579433086   3.9044017          0.015625
#> 26: 2022-11-02 04:38:01        1  2.636541625   5.0071124          0.015625
#> 27: 2022-11-02 04:38:01        1  8.827508979  13.1004595          0.015625
#> 28: 2022-11-02 04:38:01        1  1.603082357   3.2024753          0.015625
#> 29: 2022-11-02 04:38:01        1  1.818162677   0.5287545          0.015625
#> 30: 2022-11-02 04:38:01        1  6.712094832   3.0859237          0.015625
#> 31: 2022-11-02 04:38:01        1  7.992673602   2.4991927          0.015625
#> 32: 2022-11-02 04:38:01        1  0.322158035   5.0650464          0.015625
#> 33: 2022-11-02 04:38:01        1  4.755326094   3.7697790          0.015625
#> 34: 2022-11-02 04:38:01        1 -1.256102417   7.7866127          0.015625
#> 35: 2022-11-02 04:38:01        1  0.931258184   1.0346820          0.015625
#> 36: 2022-11-02 04:38:01        1  0.937008809   0.3737334          0.015625
#> 37: 2022-11-02 04:38:01        1  9.759694482   2.2611390          0.015625
#> 38: 2022-11-02 04:38:01        1  8.037929572  12.4799544          0.015625
#> 39: 2022-11-02 04:38:01        1  4.305426722   2.6108755          0.015625
#> 40: 2022-11-02 04:38:01        1  2.539484921   6.4626915          0.015625
#> 41: 2022-11-02 04:38:01        1  8.122085728   4.3309836          0.015625
#> 42: 2022-11-02 04:38:01        1  3.452329619   7.2276816          0.015625
#> 43: 2022-11-02 04:38:01        1 -0.009261159   2.1259005          0.015625
#> 44: 2022-11-02 04:38:01        1 -3.698238260   2.4569333          0.015625
#> 45: 2022-11-02 04:38:01        1 -4.493551842   4.6787973          0.015625
#> 46: 2022-11-02 04:38:01        1  6.179093241   8.9177917          0.015625
#> 47: 2022-11-02 04:38:01        1  2.899468985  10.3211236          0.015625
#> 48: 2022-11-02 04:38:01        1  5.061437417   2.5694506          0.015625
#> 49: 2022-11-02 04:38:01        1  9.441265928   9.3251164          0.015625
#> 50: 2022-11-02 04:38:01        1  5.931964645   4.2013081          0.015625
#> 51: 2022-11-02 04:38:01        1  1.309084403   2.2658013          0.015625
#> 52: 2022-11-02 04:38:01        1  8.259936121  12.4505558          0.015625
#> 53: 2022-11-02 04:38:01        1  9.017674626   6.9317565          0.015625
#> 54: 2022-11-02 04:38:01        1 -2.151383769   3.9239414          0.015625
#> 55: 2022-11-02 04:38:01        1 -3.002820483   4.3132306          0.015625
#> 56: 2022-11-02 04:38:01        1  6.555658417   7.3800289          0.015625
#> 57: 2022-11-02 04:38:01        1  1.582611860   1.5295918          0.015625
#> 58: 2022-11-02 04:38:01        1  8.986708970  10.0717834          0.015625
#> 59: 2022-11-02 04:38:01        1 -1.089781349  11.5452124          0.015625
#> 60: 2022-11-02 04:38:01        1  1.932439987   2.3921721          0.015625
#> 61: 2022-11-02 04:38:01        1  8.089381577   5.7271814          0.015625
#> 62: 2022-11-02 04:38:01        1  6.995261334   9.7451312          0.015625
#> 63: 2022-11-02 04:38:01        1  8.319206804  14.4983792          0.015625
#> 64: 2022-11-02 04:38:01        1 -4.050101915   5.3095960          0.015625
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity