Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# Branin function with fidelity, see `bbotk::branin()`
fun = function(xs) branin_wu(xs[["x1"]], xs[["x2"]], xs[["fidelity"]])

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: -3.09732 10.83145 0.015625 <list[3]> 0.5617878

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: -3.09732 10.83145 0.015625 <list[3]> 0.5617878

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1:  8.8351630  8.3740843 0.015625     0       6          1 199.0163327
#>  2: -3.9496467  3.8974821 0.015625     0       6          1  82.0111681
#>  3: -1.4774853  6.2215505 0.015625     0       6          1  15.7217559
#>  4: -0.6173775 13.1758619 0.015625     0       6          1  56.0411873
#>  5:  5.8436750 14.1999162 0.015625     0       6          1 289.3067807
#>  6:  3.4999829 14.1544463 0.015625     0       6          1 179.1822870
#>  7:  0.1652877 12.7606074 0.015625     0       6          1  68.7913998
#>  8: -0.1267217  4.8568034 0.015625     0       6          1  21.3351502
#>  9: -3.5167763  4.9928330 0.015625     0       6          1  49.8498939
#> 10: -4.8303933  7.4100301 0.015625     0       6          1  60.0631489
#> 11: -1.5920497  9.5104589 0.015625     0       6          1  10.6036012
#> 12:  6.8234364 10.6907481 0.015625     0       6          1 217.5814634
#> 13:  1.9109901 14.8138568 0.015625     0       6          1 144.6942903
#> 14: -3.8652373 14.2147355 0.015625     0       6          1   5.3759168
#> 15:  3.7937524  6.9066241 0.015625     0       6          1  44.6450115
#> 16:  6.9778693  8.2331520 0.015625     0       6          1 157.6023552
#> 17:  1.7625346 14.6908169 0.015625     0       6          1 138.1406882
#> 18: -1.9376191  6.4572706 0.015625     0       6          1  14.0746517
#> 19: -3.0973195 10.8314512 0.015625     0       6          1   0.5617878
#> 20: -4.7377421  9.6813168 0.015625     0       6          1  30.9385231
#> 21:  3.2988399  8.2722924 0.015625     0       6          1  52.1831211
#> 22: -2.9735650 14.9923831 0.015625     0       6          1  16.4367798
#> 23: -0.3064199 10.4458711 0.015625     0       6          1  34.7992539
#> 24: -4.7111058 12.7372983 0.015625     0       6          1  12.0701431
#> 25:  1.5452231  4.3551742 0.015625     0       6          1  10.7946972
#> 26:  9.2101318  1.1492596 0.015625     0       6          1  52.4507172
#> 27:  0.2621213  2.1533855 0.015625     0       6          1  31.0496507
#> 28:  1.0489754  6.2893594 0.015625     0       6          1  18.4919788
#> 29:  2.3336664 14.5657495 0.015625     0       6          1 150.0763166
#> 30: -1.9753610  2.9867349 0.015625     0       6          1  45.6227428
#> 31: -4.4203000  9.5578128 0.015625     0       6          1  23.8657801
#> 32:  9.9475785  1.5298554 0.015625     0       6          1  70.8923960
#> 33: -4.0270809  4.9198365 0.015625     0       6          1  67.7324203
#> 34:  4.3464877  9.4033140 0.015625     0       6          1 101.4338598
#> 35: -4.4151717  6.5441167 0.015625     0       6          1  57.3457800
#> 36:  5.5363287  8.6015355 0.015625     0       6          1 126.6766574
#> 37: -0.8077533 12.2993431 0.015625     0       6          1  41.5733148
#> 38:  5.2138077 13.1925514 0.015625     0       6          1 229.3775508
#> 39: -2.9170108  3.9745116 0.015625     0       6          1  48.6594897
#> 40:  7.4627538  1.5452057 0.015625     0       6          1  46.2674808
#> 41: -1.2283099  1.4138272 0.015625     0       6          1  56.6195469
#> 42: -4.1909360  5.9812861 0.015625     0       6          1  57.4728441
#> 43:  5.7172745  2.2073662 0.015625     0       6          1  36.6093992
#> 44:  5.4169323 12.2886614 0.015625     0       6          1 212.4365707
#> 45: -2.8055887 12.0839957 0.015625     0       6          1   2.8302609
#> 46:  2.4914839  2.1324349 0.015625     0       6          1   2.3652147
#> 47: -3.0101725 12.6443737 0.015625     0       6          1   2.9611005
#> 48:  4.8949789  7.5782275 0.015625     0       6          1  86.2567593
#> 49:  4.9551137  2.2032768 0.015625     0       6          1  23.4277373
#> 50:  5.8212116 14.9889893 0.015625     0       6          1 314.8425446
#> 51:  1.1399754  4.8913673 0.015625     0       6          1  14.4532104
#> 52:  3.3599367 12.4217263 0.015625     0       6          1 131.0907667
#> 53:  7.8201642  3.4891852 0.015625     0       6          1  75.2081770
#> 54:  2.2260843 13.1019919 0.015625     0       6          1 114.2421926
#> 55: -0.8607320 10.4781766 0.015625     0       6          1  25.7797597
#> 56:  1.4782349  8.3038412 0.015625     0       6          1  31.9495345
#> 57:  2.7090689  6.6068414 0.015625     0       6          1  23.3045527
#> 58:  5.5010684 14.5824074 0.015625     0       6          1 286.0072313
#> 59:  0.2600566 11.6737071 0.015625     0       6          1  56.3128070
#> 60:  7.9821765  9.4477874 0.015625     0       6          1 210.2070286
#> 61:  3.4648312  2.7094943 0.015625     0       6          1   4.3355397
#> 62:  3.0092363  9.3165375 0.015625     0       6          1  61.7509398
#> 63:  7.4080306 14.1470337 0.015625     0       6          1 347.2010511
#> 64: -2.0766475  0.9847308 0.015625     0       6          1  76.7996682
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-11-15 04:15:52        1   8.8351630   8.3740843          0.015625
#>  2: 2022-11-15 04:15:52        1  -3.9496467   3.8974821          0.015625
#>  3: 2022-11-15 04:15:52        1  -1.4774853   6.2215505          0.015625
#>  4: 2022-11-15 04:15:52        1  -0.6173775  13.1758619          0.015625
#>  5: 2022-11-15 04:15:52        1   5.8436750  14.1999162          0.015625
#>  6: 2022-11-15 04:15:52        1   3.4999829  14.1544463          0.015625
#>  7: 2022-11-15 04:15:52        1   0.1652877  12.7606074          0.015625
#>  8: 2022-11-15 04:15:52        1  -0.1267217   4.8568034          0.015625
#>  9: 2022-11-15 04:15:52        1  -3.5167763   4.9928330          0.015625
#> 10: 2022-11-15 04:15:52        1  -4.8303933   7.4100301          0.015625
#> 11: 2022-11-15 04:15:52        1  -1.5920497   9.5104589          0.015625
#> 12: 2022-11-15 04:15:52        1   6.8234364  10.6907481          0.015625
#> 13: 2022-11-15 04:15:52        1   1.9109901  14.8138568          0.015625
#> 14: 2022-11-15 04:15:52        1  -3.8652373  14.2147355          0.015625
#> 15: 2022-11-15 04:15:52        1   3.7937524   6.9066241          0.015625
#> 16: 2022-11-15 04:15:52        1   6.9778693   8.2331520          0.015625
#> 17: 2022-11-15 04:15:52        1   1.7625346  14.6908169          0.015625
#> 18: 2022-11-15 04:15:52        1  -1.9376191   6.4572706          0.015625
#> 19: 2022-11-15 04:15:52        1  -3.0973195  10.8314512          0.015625
#> 20: 2022-11-15 04:15:52        1  -4.7377421   9.6813168          0.015625
#> 21: 2022-11-15 04:15:52        1   3.2988399   8.2722924          0.015625
#> 22: 2022-11-15 04:15:52        1  -2.9735650  14.9923831          0.015625
#> 23: 2022-11-15 04:15:52        1  -0.3064199  10.4458711          0.015625
#> 24: 2022-11-15 04:15:52        1  -4.7111058  12.7372983          0.015625
#> 25: 2022-11-15 04:15:52        1   1.5452231   4.3551742          0.015625
#> 26: 2022-11-15 04:15:52        1   9.2101318   1.1492596          0.015625
#> 27: 2022-11-15 04:15:52        1   0.2621213   2.1533855          0.015625
#> 28: 2022-11-15 04:15:52        1   1.0489754   6.2893594          0.015625
#> 29: 2022-11-15 04:15:52        1   2.3336664  14.5657495          0.015625
#> 30: 2022-11-15 04:15:52        1  -1.9753610   2.9867349          0.015625
#> 31: 2022-11-15 04:15:52        1  -4.4203000   9.5578128          0.015625
#> 32: 2022-11-15 04:15:52        1   9.9475785   1.5298554          0.015625
#> 33: 2022-11-15 04:15:52        1  -4.0270809   4.9198365          0.015625
#> 34: 2022-11-15 04:15:52        1   4.3464877   9.4033140          0.015625
#> 35: 2022-11-15 04:15:52        1  -4.4151717   6.5441167          0.015625
#> 36: 2022-11-15 04:15:52        1   5.5363287   8.6015355          0.015625
#> 37: 2022-11-15 04:15:52        1  -0.8077533  12.2993431          0.015625
#> 38: 2022-11-15 04:15:52        1   5.2138077  13.1925514          0.015625
#> 39: 2022-11-15 04:15:52        1  -2.9170108   3.9745116          0.015625
#> 40: 2022-11-15 04:15:52        1   7.4627538   1.5452057          0.015625
#> 41: 2022-11-15 04:15:52        1  -1.2283099   1.4138272          0.015625
#> 42: 2022-11-15 04:15:52        1  -4.1909360   5.9812861          0.015625
#> 43: 2022-11-15 04:15:52        1   5.7172745   2.2073662          0.015625
#> 44: 2022-11-15 04:15:52        1   5.4169323  12.2886614          0.015625
#> 45: 2022-11-15 04:15:52        1  -2.8055887  12.0839957          0.015625
#> 46: 2022-11-15 04:15:52        1   2.4914839   2.1324349          0.015625
#> 47: 2022-11-15 04:15:52        1  -3.0101725  12.6443737          0.015625
#> 48: 2022-11-15 04:15:52        1   4.8949789   7.5782275          0.015625
#> 49: 2022-11-15 04:15:52        1   4.9551137   2.2032768          0.015625
#> 50: 2022-11-15 04:15:52        1   5.8212116  14.9889893          0.015625
#> 51: 2022-11-15 04:15:52        1   1.1399754   4.8913673          0.015625
#> 52: 2022-11-15 04:15:52        1   3.3599367  12.4217263          0.015625
#> 53: 2022-11-15 04:15:52        1   7.8201642   3.4891852          0.015625
#> 54: 2022-11-15 04:15:52        1   2.2260843  13.1019919          0.015625
#> 55: 2022-11-15 04:15:52        1  -0.8607320  10.4781766          0.015625
#> 56: 2022-11-15 04:15:52        1   1.4782349   8.3038412          0.015625
#> 57: 2022-11-15 04:15:52        1   2.7090689   6.6068414          0.015625
#> 58: 2022-11-15 04:15:52        1   5.5010684  14.5824074          0.015625
#> 59: 2022-11-15 04:15:52        1   0.2600566  11.6737071          0.015625
#> 60: 2022-11-15 04:15:52        1   7.9821765   9.4477874          0.015625
#> 61: 2022-11-15 04:15:52        1   3.4648312   2.7094943          0.015625
#> 62: 2022-11-15 04:15:52        1   3.0092363   9.3165375          0.015625
#> 63: 2022-11-15 04:15:52        1   7.4080306  14.1470337          0.015625
#> 64: 2022-11-15 04:15:52        1  -2.0766475   0.9847308          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity