Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.139168 12.22269 0.015625 <list[3]> 1.250876

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.139168 12.22269 0.015625 <list[3]> 1.250876

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1:  9.5657820  8.3651710 0.015625     0       6          1 218.826490
#>  2: -4.4986362 11.0706424 0.015625     0       6          1  15.314856
#>  3: -3.1097879  8.9054083 0.015625     0       6          1   5.884525
#>  4:  4.8969923 13.8539678 0.015625     0       6          1 234.083854
#>  5:  7.5272785  9.8570374 0.015625     0       6          1 211.749149
#>  6:  8.0427281 14.6839268 0.015625     0       6          1 388.270608
#>  7:  5.7211453  4.9778635 0.015625     0       6          1  68.208283
#>  8:  5.9578484  5.7918040 0.015625     0       6          1  86.053682
#>  9: -2.5077856 10.1733123 0.015625     0       6          1   2.262950
#> 10: -4.3845670 12.1719990 0.015625     0       6          1   8.860881
#> 11:  5.8121320  1.1146273 0.015625     0       6          1  29.620371
#> 12: -3.0055663 13.7201039 0.015625     0       6          1   7.556046
#> 13: -1.7707875 14.1020261 0.015625     0       6          1  35.000698
#> 14:  5.7740367  6.8916530 0.015625     0       6          1 100.399444
#> 15:  4.2203973 14.4271955 0.015625     0       6          1 218.522188
#> 16: -0.7693390  9.6023562 0.015625     0       6          1  22.466138
#> 17:  2.2960939 14.8595849 0.015625     0       6          1 156.198119
#> 18: -3.1391680 12.2226922 0.015625     0       6          1   1.250876
#> 19:  3.8684846  0.4142927 0.015625     0       6          1   2.837215
#> 20:  0.4582366  9.1098094 0.015625     0       6          1  33.300779
#> 21: -2.1599968  4.4055843 0.015625     0       6          1  31.451080
#> 22:  7.8507761 13.1708467 0.015625     0       6          1 325.827232
#> 23:  3.0722929  9.8397907 0.015625     0       6          1  71.642317
#> 24: -3.3212683  9.3375742 0.015625     0       6          1   5.785365
#> 25:  8.5949268 10.5968959 0.015625     0       6          1 259.671697
#> 26:  9.5598980 14.8521545 0.015625     0       6          1 452.353316
#> 27: -4.5928360  6.9572983 0.015625     0       6          1  57.868899
#> 28:  6.4123626  3.6526422 0.015625     0       6          1  63.002506
#> 29: -2.0270012 14.9188663 0.015625     0       6          1  36.755348
#> 30: -1.2114429  4.6127622 0.015625     0       6          1  24.669277
#> 31:  6.3954315 14.1564234 0.015625     0       6          1 311.181618
#> 32:  5.0139598  3.7577368 0.015625     0       6          1  37.500551
#> 33:  4.6015703  3.1928057 0.015625     0       6          1  23.879256
#> 34:  4.3626328 14.3333716 0.015625     0       6          1 222.550502
#> 35:  9.0477577 10.8614705 0.015625     0       6          1 281.447716
#> 36: -4.2378776 12.0092213 0.015625     0       6          1   7.271121
#> 37: -1.5656891 10.9179137 0.015625     0       6          1  15.574686
#> 38:  1.1547500 12.8697940 0.015625     0       6          1  88.991250
#> 39:  9.0986664 12.9597765 0.015625     0       6          1 357.937848
#> 40:  9.1172707 13.6320126 0.015625     0       6          1 384.489713
#> 41:  1.1780567  8.5313794 0.015625     0       6          1  32.716315
#> 42:  0.7426132  1.3781567 0.015625     0       6          1  29.024030
#> 43:  0.8099313  7.0977430 0.015625     0       6          1  22.222005
#> 44: -2.0157884  5.5014220 0.015625     0       6          1  20.549013
#> 45:  0.4491310 12.9231973 0.015625     0       6          1  76.894333
#> 46: -4.1111598  5.4929425 0.015625     0       6          1  61.871074
#> 47:  3.5698512  3.6004191 0.015625     0       6          1   9.618185
#> 48: -4.2013214  2.0286692 0.015625     0       6          1 130.758312
#> 49:  2.4599980 12.0364486 0.015625     0       6          1  97.910020
#> 50:  2.2009914  3.0750799 0.015625     0       6          1   4.525593
#> 51:  3.5687118  2.2188409 0.015625     0       6          1   3.531675
#> 52:  4.7247369  1.1544505 0.015625     0       6          1  14.069646
#> 53:  8.8913891  0.8617743 0.015625     0       6          1  45.055759
#> 54:  3.1910522 14.0094357 0.015625     0       6          1 163.611838
#> 55: -0.0747840  3.1163729 0.015625     0       6          1  28.592211
#> 56:  6.1847639 13.5487024 0.015625     0       6          1 282.512569
#> 57:  7.5705075  5.4822115 0.015625     0       6          1 108.116226
#> 58: -0.4273260  7.9131207 0.015625     0       6          1  20.245164
#> 59:  8.9585355  7.5840941 0.015625     0       6          1 180.298263
#> 60: -3.5751465  0.5338935 0.015625     0       6          1 134.668624
#> 61:  4.4094288 13.4495684 0.015625     0       6          1 199.500387
#> 62: -0.3358349  3.4988061 0.015625     0       6          1  28.302186
#> 63: -4.9411407  4.9083424 0.015625     0       6          1 106.391817
#> 64:  3.5708231 14.8615153 0.015625     0       6          1 201.565258
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-20 04:38:59        1   9.5657820   8.3651710          0.015625
#>  2: 2022-10-20 04:38:59        1  -4.4986362  11.0706424          0.015625
#>  3: 2022-10-20 04:38:59        1  -3.1097879   8.9054083          0.015625
#>  4: 2022-10-20 04:38:59        1   4.8969923  13.8539678          0.015625
#>  5: 2022-10-20 04:38:59        1   7.5272785   9.8570374          0.015625
#>  6: 2022-10-20 04:38:59        1   8.0427281  14.6839268          0.015625
#>  7: 2022-10-20 04:38:59        1   5.7211453   4.9778635          0.015625
#>  8: 2022-10-20 04:38:59        1   5.9578484   5.7918040          0.015625
#>  9: 2022-10-20 04:38:59        1  -2.5077856  10.1733123          0.015625
#> 10: 2022-10-20 04:38:59        1  -4.3845670  12.1719990          0.015625
#> 11: 2022-10-20 04:38:59        1   5.8121320   1.1146273          0.015625
#> 12: 2022-10-20 04:38:59        1  -3.0055663  13.7201039          0.015625
#> 13: 2022-10-20 04:38:59        1  -1.7707875  14.1020261          0.015625
#> 14: 2022-10-20 04:38:59        1   5.7740367   6.8916530          0.015625
#> 15: 2022-10-20 04:38:59        1   4.2203973  14.4271955          0.015625
#> 16: 2022-10-20 04:38:59        1  -0.7693390   9.6023562          0.015625
#> 17: 2022-10-20 04:38:59        1   2.2960939  14.8595849          0.015625
#> 18: 2022-10-20 04:38:59        1  -3.1391680  12.2226922          0.015625
#> 19: 2022-10-20 04:38:59        1   3.8684846   0.4142927          0.015625
#> 20: 2022-10-20 04:38:59        1   0.4582366   9.1098094          0.015625
#> 21: 2022-10-20 04:38:59        1  -2.1599968   4.4055843          0.015625
#> 22: 2022-10-20 04:38:59        1   7.8507761  13.1708467          0.015625
#> 23: 2022-10-20 04:38:59        1   3.0722929   9.8397907          0.015625
#> 24: 2022-10-20 04:38:59        1  -3.3212683   9.3375742          0.015625
#> 25: 2022-10-20 04:38:59        1   8.5949268  10.5968959          0.015625
#> 26: 2022-10-20 04:38:59        1   9.5598980  14.8521545          0.015625
#> 27: 2022-10-20 04:38:59        1  -4.5928360   6.9572983          0.015625
#> 28: 2022-10-20 04:38:59        1   6.4123626   3.6526422          0.015625
#> 29: 2022-10-20 04:38:59        1  -2.0270012  14.9188663          0.015625
#> 30: 2022-10-20 04:38:59        1  -1.2114429   4.6127622          0.015625
#> 31: 2022-10-20 04:38:59        1   6.3954315  14.1564234          0.015625
#> 32: 2022-10-20 04:38:59        1   5.0139598   3.7577368          0.015625
#> 33: 2022-10-20 04:38:59        1   4.6015703   3.1928057          0.015625
#> 34: 2022-10-20 04:38:59        1   4.3626328  14.3333716          0.015625
#> 35: 2022-10-20 04:38:59        1   9.0477577  10.8614705          0.015625
#> 36: 2022-10-20 04:38:59        1  -4.2378776  12.0092213          0.015625
#> 37: 2022-10-20 04:38:59        1  -1.5656891  10.9179137          0.015625
#> 38: 2022-10-20 04:38:59        1   1.1547500  12.8697940          0.015625
#> 39: 2022-10-20 04:38:59        1   9.0986664  12.9597765          0.015625
#> 40: 2022-10-20 04:38:59        1   9.1172707  13.6320126          0.015625
#> 41: 2022-10-20 04:38:59        1   1.1780567   8.5313794          0.015625
#> 42: 2022-10-20 04:38:59        1   0.7426132   1.3781567          0.015625
#> 43: 2022-10-20 04:38:59        1   0.8099313   7.0977430          0.015625
#> 44: 2022-10-20 04:38:59        1  -2.0157884   5.5014220          0.015625
#> 45: 2022-10-20 04:38:59        1   0.4491310  12.9231973          0.015625
#> 46: 2022-10-20 04:38:59        1  -4.1111598   5.4929425          0.015625
#> 47: 2022-10-20 04:38:59        1   3.5698512   3.6004191          0.015625
#> 48: 2022-10-20 04:38:59        1  -4.2013214   2.0286692          0.015625
#> 49: 2022-10-20 04:38:59        1   2.4599980  12.0364486          0.015625
#> 50: 2022-10-20 04:38:59        1   2.2009914   3.0750799          0.015625
#> 51: 2022-10-20 04:38:59        1   3.5687118   2.2188409          0.015625
#> 52: 2022-10-20 04:38:59        1   4.7247369   1.1544505          0.015625
#> 53: 2022-10-20 04:38:59        1   8.8913891   0.8617743          0.015625
#> 54: 2022-10-20 04:38:59        1   3.1910522  14.0094357          0.015625
#> 55: 2022-10-20 04:38:59        1  -0.0747840   3.1163729          0.015625
#> 56: 2022-10-20 04:38:59        1   6.1847639  13.5487024          0.015625
#> 57: 2022-10-20 04:38:59        1   7.5705075   5.4822115          0.015625
#> 58: 2022-10-20 04:38:59        1  -0.4273260   7.9131207          0.015625
#> 59: 2022-10-20 04:38:59        1   8.9585355   7.5840941          0.015625
#> 60: 2022-10-20 04:38:59        1  -3.5751465   0.5338935          0.015625
#> 61: 2022-10-20 04:38:59        1   4.4094288  13.4495684          0.015625
#> 62: 2022-10-20 04:38:59        1  -0.3358349   3.4988061          0.015625
#> 63: 2022-10-20 04:38:59        1  -4.9411407   4.9083424          0.015625
#> 64: 2022-10-20 04:38:59        1   3.5708231  14.8615153          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity