Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -2.989672 11.53125 0.015625 <list[3]> 0.7567043

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -2.989672 11.53125 0.015625 <list[3]> 0.7567043

# all evaluations
as.data.table(instance$archive)
#>                x1          x2 fidelity stage bracket repetition           y
#>  1:  5.3461124725  3.66798385 0.015625     0       6          1  43.7526580
#>  2: -4.8798778350  4.75901023 0.015625     0       6          1 106.4633022
#>  3:  5.0829232030 13.03586442 0.015625     0       6          1 218.8604880
#>  4: -4.1028850607  6.06596619 0.015625     0       6          1  53.2453901
#>  5:  6.9571386883 14.80668468 0.015625     0       6          1 355.7354503
#>  6:  1.4069087640  0.02376348 0.015625     0       6          1  25.9909198
#>  7:  2.6687318389 12.11748749 0.015625     0       6          1 104.3912688
#>  8:  6.9167602982  1.26920398 0.015625     0       6          1  40.8417534
#>  9:  2.0657992689 11.54325427 0.015625     0       6          1  81.1262708
#> 10: -1.2588954694  2.98152028 0.015625     0       6          1  38.6596222
#> 11: -4.0778184624 14.24976074 0.015625     0       6          1   5.8661439
#> 12:  3.9727236505 12.12366314 0.015625     0       6          1 146.5976960
#> 13:  0.0002133008  9.78933354 0.015625     0       6          1  33.9637340
#> 14:  6.3198769023  2.06611040 0.015625     0       6          1  43.5708272
#> 15: -4.1778679704  1.65073235 0.015625     0       6          1 138.1698726
#> 16: -4.5059488167 10.59199681 0.015625     0       6          1  18.2955951
#> 17:  2.0568136382  2.42592737 0.015625     0       6          1   5.7002147
#> 18:  3.2585932640 14.07117883 0.015625     0       6          1 167.6718678
#> 19: -2.3450383765 11.34928116 0.015625     0       6          1   5.3830155
#> 20:  4.2932655232  6.01952156 0.015625     0       6          1  45.6027645
#> 21: -3.7082996522  2.52109035 0.015625     0       6          1  98.0108773
#> 22:  7.8546547925  8.11992136 0.015625     0       6          1 171.8946664
#> 23: -2.9896718927 11.53125034 0.015625     0       6          1   0.7567043
#> 24:  6.4048103325  7.60793797 0.015625     0       6          1 130.6273929
#> 25:  7.4723406474  9.14422797 0.015625     0       6          1 190.9997490
#> 26: -4.3900776841 12.24815531 0.015625     0       6          1   8.7311989
#> 27:  2.2298161930  0.12108056 0.015625     0       6          1  10.2851786
#> 28:  3.0740226456  6.70831157 0.015625     0       6          1  28.6182692
#> 29:  2.5471422262  7.02391866 0.015625     0       6          1  25.8432388
#> 30:  4.8707297712 14.55037696 0.015625     0       6          1 254.0305786
#> 31: -4.2912612518 12.46479015 0.015625     0       6          1   6.9418272
#> 32:  8.3658144076  9.50129966 0.015625     0       6          1 220.3306125
#> 33:  6.0844649037  2.90757131 0.015625     0       6          1  49.1485893
#> 34:  6.5656433592 11.65814052 0.015625     0       6          1 237.7366590
#> 35:  5.1794745831  9.74472358 0.015625     0       6          1 138.9421757
#> 36:  6.7929750285  2.65109789 0.015625     0       6          1  54.9068549
#> 37:  3.5888274142  9.63371576 0.015625     0       6          1  81.4358445
#> 38:  9.8652440379  2.30104704 0.015625     0       6          1  82.4886806
#> 39:  8.4715826402 10.61113906 0.015625     0       6          1 256.8503194
#> 40:  7.5433316512 12.11830519 0.015625     0       6          1 281.0537880
#> 41:  5.3617616009  5.26895281 0.015625     0       6          1  63.6723483
#> 42:  2.8479050600  6.24432034 0.015625     0       6          1  21.3075179
#> 43:  0.9852579527 14.06300102 0.015625     0       6          1 107.4904111
#> 44: -3.7224504561 12.04354996 0.015625     0       6          1   2.0669352
#> 45:  7.2670213820 13.16795922 0.015625     0       6          1 308.0716277
#> 46:  4.5947564894 11.17617600 0.015625     0       6          1 149.0547462
#> 47: -3.7313697266  3.19929548 0.015625     0       6          1  86.0623129
#> 48:  1.4941874228 13.11574812 0.015625     0       6          1  99.5688233
#> 49: -0.1695384865  3.62991017 0.015625     0       6          1  26.4382824
#> 50:  3.5813677718  4.84265856 0.015625     0       6          1  18.5192429
#> 51:  3.7175793492  9.76061456 0.015625     0       6          1  87.5538583
#> 52: -2.1185971866 13.91804553 0.015625     0       6          1  24.4312240
#> 53:  9.9406235758  7.59839287 0.015625     0       6          1 208.4631116
#> 54: -3.9240827423  4.16542584 0.015625     0       6          1  76.3512392
#> 55:  0.0931792730  1.66262343 0.015625     0       6          1  37.1110623
#> 56:  4.8376810027 11.93330226 0.015625     0       6          1 177.9489425
#> 57:  8.2273070444 11.65693905 0.015625     0       6          1 284.3829920
#> 58:  4.3285670539  0.63222873 0.015625     0       6          1   7.2978532
#> 59:  7.3337893211  7.91901750 0.015625     0       6          1 157.2740832
#> 60: -0.0535212993  9.30548630 0.015625     0       6          1  29.9581570
#> 61:  5.8666002762  2.34001719 0.015625     0       6          1  40.1140473
#> 62:  1.8693242490 10.44656828 0.015625     0       6          1  60.6741226
#> 63: -0.9690201201 12.23211726 0.015625     0       6          1  37.1607718
#> 64:  1.0766328790  4.44013695 0.015625     0       6          1  14.5681640
#>                x1          x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr   x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-09-27 04:43:14        1  5.3461124725  3.66798385          0.015625
#>  2: 2022-09-27 04:43:14        1 -4.8798778350  4.75901023          0.015625
#>  3: 2022-09-27 04:43:14        1  5.0829232030 13.03586442          0.015625
#>  4: 2022-09-27 04:43:14        1 -4.1028850607  6.06596619          0.015625
#>  5: 2022-09-27 04:43:14        1  6.9571386883 14.80668468          0.015625
#>  6: 2022-09-27 04:43:14        1  1.4069087640  0.02376348          0.015625
#>  7: 2022-09-27 04:43:14        1  2.6687318389 12.11748749          0.015625
#>  8: 2022-09-27 04:43:14        1  6.9167602982  1.26920398          0.015625
#>  9: 2022-09-27 04:43:14        1  2.0657992689 11.54325427          0.015625
#> 10: 2022-09-27 04:43:14        1 -1.2588954694  2.98152028          0.015625
#> 11: 2022-09-27 04:43:14        1 -4.0778184624 14.24976074          0.015625
#> 12: 2022-09-27 04:43:14        1  3.9727236505 12.12366314          0.015625
#> 13: 2022-09-27 04:43:14        1  0.0002133008  9.78933354          0.015625
#> 14: 2022-09-27 04:43:14        1  6.3198769023  2.06611040          0.015625
#> 15: 2022-09-27 04:43:14        1 -4.1778679704  1.65073235          0.015625
#> 16: 2022-09-27 04:43:14        1 -4.5059488167 10.59199681          0.015625
#> 17: 2022-09-27 04:43:14        1  2.0568136382  2.42592737          0.015625
#> 18: 2022-09-27 04:43:14        1  3.2585932640 14.07117883          0.015625
#> 19: 2022-09-27 04:43:14        1 -2.3450383765 11.34928116          0.015625
#> 20: 2022-09-27 04:43:14        1  4.2932655232  6.01952156          0.015625
#> 21: 2022-09-27 04:43:14        1 -3.7082996522  2.52109035          0.015625
#> 22: 2022-09-27 04:43:14        1  7.8546547925  8.11992136          0.015625
#> 23: 2022-09-27 04:43:14        1 -2.9896718927 11.53125034          0.015625
#> 24: 2022-09-27 04:43:14        1  6.4048103325  7.60793797          0.015625
#> 25: 2022-09-27 04:43:14        1  7.4723406474  9.14422797          0.015625
#> 26: 2022-09-27 04:43:14        1 -4.3900776841 12.24815531          0.015625
#> 27: 2022-09-27 04:43:14        1  2.2298161930  0.12108056          0.015625
#> 28: 2022-09-27 04:43:14        1  3.0740226456  6.70831157          0.015625
#> 29: 2022-09-27 04:43:14        1  2.5471422262  7.02391866          0.015625
#> 30: 2022-09-27 04:43:14        1  4.8707297712 14.55037696          0.015625
#> 31: 2022-09-27 04:43:14        1 -4.2912612518 12.46479015          0.015625
#> 32: 2022-09-27 04:43:14        1  8.3658144076  9.50129966          0.015625
#> 33: 2022-09-27 04:43:14        1  6.0844649037  2.90757131          0.015625
#> 34: 2022-09-27 04:43:14        1  6.5656433592 11.65814052          0.015625
#> 35: 2022-09-27 04:43:14        1  5.1794745831  9.74472358          0.015625
#> 36: 2022-09-27 04:43:14        1  6.7929750285  2.65109789          0.015625
#> 37: 2022-09-27 04:43:14        1  3.5888274142  9.63371576          0.015625
#> 38: 2022-09-27 04:43:14        1  9.8652440379  2.30104704          0.015625
#> 39: 2022-09-27 04:43:14        1  8.4715826402 10.61113906          0.015625
#> 40: 2022-09-27 04:43:14        1  7.5433316512 12.11830519          0.015625
#> 41: 2022-09-27 04:43:14        1  5.3617616009  5.26895281          0.015625
#> 42: 2022-09-27 04:43:14        1  2.8479050600  6.24432034          0.015625
#> 43: 2022-09-27 04:43:14        1  0.9852579527 14.06300102          0.015625
#> 44: 2022-09-27 04:43:14        1 -3.7224504561 12.04354996          0.015625
#> 45: 2022-09-27 04:43:14        1  7.2670213820 13.16795922          0.015625
#> 46: 2022-09-27 04:43:14        1  4.5947564894 11.17617600          0.015625
#> 47: 2022-09-27 04:43:14        1 -3.7313697266  3.19929548          0.015625
#> 48: 2022-09-27 04:43:14        1  1.4941874228 13.11574812          0.015625
#> 49: 2022-09-27 04:43:14        1 -0.1695384865  3.62991017          0.015625
#> 50: 2022-09-27 04:43:14        1  3.5813677718  4.84265856          0.015625
#> 51: 2022-09-27 04:43:14        1  3.7175793492  9.76061456          0.015625
#> 52: 2022-09-27 04:43:14        1 -2.1185971866 13.91804553          0.015625
#> 53: 2022-09-27 04:43:14        1  9.9406235758  7.59839287          0.015625
#> 54: 2022-09-27 04:43:14        1 -3.9240827423  4.16542584          0.015625
#> 55: 2022-09-27 04:43:14        1  0.0931792730  1.66262343          0.015625
#> 56: 2022-09-27 04:43:14        1  4.8376810027 11.93330226          0.015625
#> 57: 2022-09-27 04:43:14        1  8.2273070444 11.65693905          0.015625
#> 58: 2022-09-27 04:43:14        1  4.3285670539  0.63222873          0.015625
#> 59: 2022-09-27 04:43:14        1  7.3337893211  7.91901750          0.015625
#> 60: 2022-09-27 04:43:14        1 -0.0535212993  9.30548630          0.015625
#> 61: 2022-09-27 04:43:14        1  5.8666002762  2.34001719          0.015625
#> 62: 2022-09-27 04:43:14        1  1.8693242490 10.44656828          0.015625
#> 63: 2022-09-27 04:43:14        1 -0.9690201201 12.23211726          0.015625
#> 64: 2022-09-27 04:43:14        1  1.0766328790  4.44013695          0.015625
#>               timestamp batch_nr   x_domain_x1 x_domain_x2 x_domain_fidelity