Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.764193 1.346362 0.015625 <list[3]> 1.312961

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.764193 1.346362 0.015625 <list[3]> 1.312961

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition          y
#>  1:  4.04942090 13.39963523 0.015625     0       6          1 182.054036
#>  2: -3.35588011  9.88297541 0.015625     0       6          1   3.873164
#>  3: -1.94380972  4.49003527 0.015625     0       6          1  28.777358
#>  4:  6.34449954  7.00270365 0.015625     0       6          1 116.855716
#>  5: -3.16757008 14.18470620 0.015625     0       6          1   8.437568
#>  6:  5.44003857 11.88902751 0.015625     0       6          1 202.359479
#>  7:  0.91337471 11.71689498 0.015625     0       6          1  66.917595
#>  8:  3.55231820  2.49874171 0.015625     0       6          1   4.309717
#>  9: -2.95242307  5.43864380 0.015625     0       6          1  31.131272
#> 10: -1.97948161  4.81013456 0.015625     0       6          1  26.082699
#> 11:  3.14579129 12.22164838 0.015625     0       6          1 119.733020
#> 12: -3.11013857  3.59270844 0.015625     0       6          1  58.996235
#> 13:  5.25549781 13.07753277 0.015625     0       6          1 227.908594
#> 14:  4.67847591 12.44878384 0.015625     0       6          1 184.490811
#> 15:  3.10372968 14.98673858 0.015625     0       6          1 186.189510
#> 16:  7.18922449 11.28965650 0.015625     0       6          1 245.218695
#> 17:  1.71267805  1.11221535 0.015625     0       6          1  13.714436
#> 18: -0.41984451  1.32910031 0.015625     0       6          1  47.332115
#> 19:  1.09409586  0.19675933 0.015625     0       6          1  31.205588
#> 20:  3.96516812  2.08250899 0.015625     0       6          1   7.121900
#> 21:  5.57265158  2.93771120 0.015625     0       6          1  40.820733
#> 22: -3.95254294  3.76592884 0.015625     0       6          1  84.477557
#> 23: -2.75184929  0.02915774 0.015625     0       6          1 113.125989
#> 24:  2.45350814  6.34838948 0.015625     0       6          1  19.132812
#> 25:  8.82661764  7.02715758 0.015625     0       6          1 162.839400
#> 26:  4.61406712 13.99040471 0.015625     0       6          1 224.540111
#> 27: -0.40507733 14.42471481 0.015625     0       6          1  79.275175
#> 28:  3.43435132  3.26180621 0.015625     0       6          1   6.400105
#> 29:  6.78268202  6.36791238 0.015625     0       6          1 113.459813
#> 30:  2.99463075  9.89215902 0.015625     0       6          1  70.768176
#> 31: -3.13981532  9.81947097 0.015625     0       6          1   2.590724
#> 32:  5.97890514  6.62484156 0.015625     0       6          1 100.908690
#> 33:  8.15669277  0.52587321 0.015625     0       6          1  36.971011
#> 34: -4.40945551  8.74714613 0.015625     0       6          1  30.838201
#> 35: -3.79625890  1.03403425 0.015625     0       6          1 133.508799
#> 36:  1.17077852 12.00978092 0.015625     0       6          1  75.063757
#> 37:  5.53885564 13.29794318 0.015625     0       6          1 247.192285
#> 38:  9.07634175  9.07317779 0.015625     0       6          1 225.545695
#> 39: -0.47768118 13.10858091 0.015625     0       6          1  58.739520
#> 40:  8.41583824  2.18248912 0.015625     0       6          1  59.629800
#> 41:  1.92100149  8.26474768 0.015625     0       6          1  33.835667
#> 42:  1.75943119 12.68584362 0.015625     0       6          1  96.388150
#> 43:  5.50247558 11.65950069 0.015625     0       6          1 198.694431
#> 44: -2.98725182  3.39344163 0.015625     0       6          1  58.809735
#> 45: -4.08199259 12.37561676 0.015625     0       6          1   4.741090
#> 46:  0.61623884  6.98887371 0.015625     0       6          1  21.669536
#> 47:  5.61807569  3.62959764 0.015625     0       6          1  48.921932
#> 48:  1.20935413 13.28352073 0.015625     0       6          1  97.361582
#> 49:  7.08214368  0.45851898 0.015625     0       6          1  34.235789
#> 50:  8.81199427  2.48179630 0.015625     0       6          1  68.062782
#> 51:  3.88359518  3.71777130 0.015625     0       6          1  14.721134
#> 52:  8.49163461  9.20874781 0.015625     0       6          1 214.722385
#> 53: -0.03658245  7.73150382 0.015625     0       6          1  22.395420
#> 54:  4.32532127  0.13318630 0.015625     0       6          1   6.570740
#> 55:  4.75018950 11.77093198 0.015625     0       6          1 170.064448
#> 56:  6.43332136 10.84672715 0.015625     0       6          1 210.296669
#> 57: -4.52616333 13.23147673 0.015625     0       6          1   8.584591
#> 58:  4.55132160  9.05922355 0.015625     0       6          1 101.891010
#> 59: -1.35581521  3.17316131 0.015625     0       6          1  37.462165
#> 60:  4.37028518 12.60013333 0.015625     0       6          1 174.958414
#> 61:  2.76419305  1.34636195 0.015625     0       6          1   1.312961
#> 62:  0.51748851 14.53798103 0.015625     0       6          1 105.830116
#> 63:  7.32537310 11.09248985 0.015625     0       6          1 242.890865
#> 64:  5.10798770 11.06234252 0.015625     0       6          1 167.205455
#>              x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-22 04:38:09        1  4.04942090 13.39963523          0.015625
#>  2: 2022-10-22 04:38:09        1 -3.35588011  9.88297541          0.015625
#>  3: 2022-10-22 04:38:09        1 -1.94380972  4.49003527          0.015625
#>  4: 2022-10-22 04:38:09        1  6.34449954  7.00270365          0.015625
#>  5: 2022-10-22 04:38:09        1 -3.16757008 14.18470620          0.015625
#>  6: 2022-10-22 04:38:09        1  5.44003857 11.88902751          0.015625
#>  7: 2022-10-22 04:38:09        1  0.91337471 11.71689498          0.015625
#>  8: 2022-10-22 04:38:09        1  3.55231820  2.49874171          0.015625
#>  9: 2022-10-22 04:38:09        1 -2.95242307  5.43864380          0.015625
#> 10: 2022-10-22 04:38:09        1 -1.97948161  4.81013456          0.015625
#> 11: 2022-10-22 04:38:09        1  3.14579129 12.22164838          0.015625
#> 12: 2022-10-22 04:38:09        1 -3.11013857  3.59270844          0.015625
#> 13: 2022-10-22 04:38:09        1  5.25549781 13.07753277          0.015625
#> 14: 2022-10-22 04:38:09        1  4.67847591 12.44878384          0.015625
#> 15: 2022-10-22 04:38:09        1  3.10372968 14.98673858          0.015625
#> 16: 2022-10-22 04:38:09        1  7.18922449 11.28965650          0.015625
#> 17: 2022-10-22 04:38:09        1  1.71267805  1.11221535          0.015625
#> 18: 2022-10-22 04:38:09        1 -0.41984451  1.32910031          0.015625
#> 19: 2022-10-22 04:38:09        1  1.09409586  0.19675933          0.015625
#> 20: 2022-10-22 04:38:09        1  3.96516812  2.08250899          0.015625
#> 21: 2022-10-22 04:38:09        1  5.57265158  2.93771120          0.015625
#> 22: 2022-10-22 04:38:09        1 -3.95254294  3.76592884          0.015625
#> 23: 2022-10-22 04:38:09        1 -2.75184929  0.02915774          0.015625
#> 24: 2022-10-22 04:38:09        1  2.45350814  6.34838948          0.015625
#> 25: 2022-10-22 04:38:09        1  8.82661764  7.02715758          0.015625
#> 26: 2022-10-22 04:38:09        1  4.61406712 13.99040471          0.015625
#> 27: 2022-10-22 04:38:09        1 -0.40507733 14.42471481          0.015625
#> 28: 2022-10-22 04:38:09        1  3.43435132  3.26180621          0.015625
#> 29: 2022-10-22 04:38:09        1  6.78268202  6.36791238          0.015625
#> 30: 2022-10-22 04:38:09        1  2.99463075  9.89215902          0.015625
#> 31: 2022-10-22 04:38:09        1 -3.13981532  9.81947097          0.015625
#> 32: 2022-10-22 04:38:09        1  5.97890514  6.62484156          0.015625
#> 33: 2022-10-22 04:38:09        1  8.15669277  0.52587321          0.015625
#> 34: 2022-10-22 04:38:09        1 -4.40945551  8.74714613          0.015625
#> 35: 2022-10-22 04:38:09        1 -3.79625890  1.03403425          0.015625
#> 36: 2022-10-22 04:38:09        1  1.17077852 12.00978092          0.015625
#> 37: 2022-10-22 04:38:09        1  5.53885564 13.29794318          0.015625
#> 38: 2022-10-22 04:38:09        1  9.07634175  9.07317779          0.015625
#> 39: 2022-10-22 04:38:09        1 -0.47768118 13.10858091          0.015625
#> 40: 2022-10-22 04:38:09        1  8.41583824  2.18248912          0.015625
#> 41: 2022-10-22 04:38:09        1  1.92100149  8.26474768          0.015625
#> 42: 2022-10-22 04:38:09        1  1.75943119 12.68584362          0.015625
#> 43: 2022-10-22 04:38:09        1  5.50247558 11.65950069          0.015625
#> 44: 2022-10-22 04:38:09        1 -2.98725182  3.39344163          0.015625
#> 45: 2022-10-22 04:38:09        1 -4.08199259 12.37561676          0.015625
#> 46: 2022-10-22 04:38:09        1  0.61623884  6.98887371          0.015625
#> 47: 2022-10-22 04:38:09        1  5.61807569  3.62959764          0.015625
#> 48: 2022-10-22 04:38:09        1  1.20935413 13.28352073          0.015625
#> 49: 2022-10-22 04:38:09        1  7.08214368  0.45851898          0.015625
#> 50: 2022-10-22 04:38:09        1  8.81199427  2.48179630          0.015625
#> 51: 2022-10-22 04:38:09        1  3.88359518  3.71777130          0.015625
#> 52: 2022-10-22 04:38:09        1  8.49163461  9.20874781          0.015625
#> 53: 2022-10-22 04:38:09        1 -0.03658245  7.73150382          0.015625
#> 54: 2022-10-22 04:38:09        1  4.32532127  0.13318630          0.015625
#> 55: 2022-10-22 04:38:09        1  4.75018950 11.77093198          0.015625
#> 56: 2022-10-22 04:38:09        1  6.43332136 10.84672715          0.015625
#> 57: 2022-10-22 04:38:09        1 -4.52616333 13.23147673          0.015625
#> 58: 2022-10-22 04:38:09        1  4.55132160  9.05922355          0.015625
#> 59: 2022-10-22 04:38:09        1 -1.35581521  3.17316131          0.015625
#> 60: 2022-10-22 04:38:09        1  4.37028518 12.60013333          0.015625
#> 61: 2022-10-22 04:38:09        1  2.76419305  1.34636195          0.015625
#> 62: 2022-10-22 04:38:09        1  0.51748851 14.53798103          0.015625
#> 63: 2022-10-22 04:38:09        1  7.32537310 11.09248985          0.015625
#> 64: 2022-10-22 04:38:09        1  5.10798770 11.06234252          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity