Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# Branin function with fidelity, see `bbotk::branin()`
fun = function(xs) branin_wu(xs[["x1"]], xs[["x2"]], xs[["fidelity"]])

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.010655 11.24379 0.015625 <list[3]> 0.5101816

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.010655 11.24379 0.015625 <list[3]> 0.5101816

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition           y
#>  1:  3.36503742 10.2203114 0.015625     0       6          1  85.7883683
#>  2:  2.78119829  9.7630461 0.015625     0       6          1  64.2431768
#>  3:  0.75662016 12.6813741 0.015625     0       6          1  78.8872372
#>  4:  4.63660766  9.5020445 0.015625     0       6          1 113.7302559
#>  5: -3.45133738 14.5918684 0.015625     0       6          1   8.3221832
#>  6:  6.37216394 10.1295596 0.015625     0       6          1 189.1550332
#>  7: -1.58958828 11.9240154 0.015625     0       6          1  20.8181821
#>  8:  7.25648321  4.4211717 0.015625     0       6          1  85.1443281
#>  9:  1.91266108 12.1922789 0.015625     0       6          1  90.0263614
#> 10: -1.38658654  9.5839720 0.015625     0       6          1  13.4960358
#> 11: -3.34485831  4.1620131 0.015625     0       6          1  56.9280122
#> 12:  6.82908494  0.6979208 0.015625     0       6          1  35.2867246
#> 13: -2.70584835 13.2198089 0.015625     0       6          1   8.5215638
#> 14:  8.31629496  6.8031030 0.015625     0       6          1 147.6228028
#> 15:  3.83527584  5.7111449 0.015625     0       6          1  31.3777508
#> 16: -0.37990226  8.0191927 0.015625     0       6          1  20.9059343
#> 17:  3.51928792  0.6984953 0.015625     0       6          1   1.0812657
#> 18:  0.06393774  3.3444965 0.015625     0       6          1  26.1047398
#> 19:  2.57334582  5.2300881 0.015625     0       6          1  11.6543158
#> 20:  2.91254894 11.0722358 0.015625     0       6          1  89.8921288
#> 21:  2.40211761 12.5168266 0.015625     0       6          1 106.1821511
#> 22:  9.52049944  4.5089125 0.015625     0       6          1 118.6933727
#> 23:  3.85510869  9.6823162 0.015625     0       6          1  90.3676435
#> 24: -1.84995370  7.8283430 0.015625     0       6          1   8.8454500
#> 25: -4.12978337 13.5672926 0.015625     0       6          1   4.9379385
#> 26:  5.90130543 13.9516995 0.015625     0       6          1 283.7255729
#> 27: -2.62074433 12.9202989 0.015625     0       6          1   8.1129780
#> 28:  4.59717468 10.3372569 0.015625     0       6          1 129.9859190
#> 29: -2.89477384  0.5787086 0.015625     0       6          1 106.4931128
#> 30: -0.14819331  4.6861399 0.015625     0       6          1  21.9005845
#> 31: -3.23977534 11.9409980 0.015625     0       6          1   0.6575852
#> 32:  6.73707951  9.2023321 0.015625     0       6          1 175.6100778
#> 33:  8.98107277  5.9032478 0.015625     0       6          1 138.6163731
#> 34:  6.90246270  4.1685406 0.015625     0       6          1  76.9434078
#> 35:  9.93869528 14.5807992 0.015625     0       6          1 457.9564597
#> 36:  3.78350973  4.1884080 0.015625     0       6          1  16.5214069
#> 37:  0.43240912  5.4521879 0.015625     0       6          1  18.7364506
#> 38:  8.67559354  4.0276064 0.015625     0       6          1  93.6191135
#> 39:  3.41751152  9.1210391 0.015625     0       6          1  68.0186570
#> 40: -3.06801870  6.6915925 0.015625     0       6          1  20.5007561
#> 41:  2.88610559  4.0307479 0.015625     0       6          1   6.3170743
#> 42:  4.52387410  4.8892706 0.015625     0       6          1  38.0120397
#> 43: -4.54177518  9.2818750 0.015625     0       6          1  29.3536449
#> 44:  3.47597396  8.4653055 0.015625     0       6          1  59.0854735
#> 45:  3.53277681  0.2288282 0.015625     0       6          1   1.4066141
#> 46: -4.60307855  2.1806832 0.015625     0       6          1 148.1174792
#> 47:  1.86524983  9.2893787 0.015625     0       6          1  45.0486263
#> 48:  1.14938670  8.2018734 0.015625     0       6          1  29.8522827
#> 49:  9.15976548  9.3095122 0.015625     0       6          1 235.0685080
#> 50: -3.01065543 11.2437923 0.015625     0       6          1   0.5101816
#> 51: -3.03504280 14.0602192 0.015625     0       6          1   9.1346339
#> 52: -4.12934878  6.6082030 0.015625     0       6          1  46.8094182
#> 53:  8.50074445  1.6057775 0.015625     0       6          1  52.0070371
#> 54:  9.75081422  1.1121492 0.015625     0       6          1  60.3120063
#> 55: -4.27046338 11.9909546 0.015625     0       6          1   7.7604850
#> 56:  7.95434417 13.8459109 0.015625     0       6          1 353.5200871
#> 57:  7.33848262 10.1354886 0.015625     0       6          1 215.2169915
#> 58:  9.09373020  0.1521360 0.015625     0       6          1  37.9173394
#> 59: -1.34796784  5.0376767 0.015625     0       6          1  22.1300016
#> 60:  1.03087054  7.1922346 0.015625     0       6          1  22.7775323
#> 61: -4.13360077 11.8335307 0.015625     0       6          1   6.3621331
#> 62:  6.92353311 10.3409986 0.015625     0       6          1 210.5284886
#> 63:  2.66616527  9.3258138 0.015625     0       6          1  55.4938801
#> 64:  1.63300602  9.1090358 0.015625     0       6          1  41.0555048
#>              x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-11-27 04:12:12        1  3.36503742  10.2203114          0.015625
#>  2: 2022-11-27 04:12:12        1  2.78119829   9.7630461          0.015625
#>  3: 2022-11-27 04:12:12        1  0.75662016  12.6813741          0.015625
#>  4: 2022-11-27 04:12:12        1  4.63660766   9.5020445          0.015625
#>  5: 2022-11-27 04:12:12        1 -3.45133738  14.5918684          0.015625
#>  6: 2022-11-27 04:12:12        1  6.37216394  10.1295596          0.015625
#>  7: 2022-11-27 04:12:12        1 -1.58958828  11.9240154          0.015625
#>  8: 2022-11-27 04:12:12        1  7.25648321   4.4211717          0.015625
#>  9: 2022-11-27 04:12:12        1  1.91266108  12.1922789          0.015625
#> 10: 2022-11-27 04:12:12        1 -1.38658654   9.5839720          0.015625
#> 11: 2022-11-27 04:12:12        1 -3.34485831   4.1620131          0.015625
#> 12: 2022-11-27 04:12:12        1  6.82908494   0.6979208          0.015625
#> 13: 2022-11-27 04:12:12        1 -2.70584835  13.2198089          0.015625
#> 14: 2022-11-27 04:12:12        1  8.31629496   6.8031030          0.015625
#> 15: 2022-11-27 04:12:12        1  3.83527584   5.7111449          0.015625
#> 16: 2022-11-27 04:12:12        1 -0.37990226   8.0191927          0.015625
#> 17: 2022-11-27 04:12:12        1  3.51928792   0.6984953          0.015625
#> 18: 2022-11-27 04:12:12        1  0.06393774   3.3444965          0.015625
#> 19: 2022-11-27 04:12:12        1  2.57334582   5.2300881          0.015625
#> 20: 2022-11-27 04:12:12        1  2.91254894  11.0722358          0.015625
#> 21: 2022-11-27 04:12:12        1  2.40211761  12.5168266          0.015625
#> 22: 2022-11-27 04:12:12        1  9.52049944   4.5089125          0.015625
#> 23: 2022-11-27 04:12:12        1  3.85510869   9.6823162          0.015625
#> 24: 2022-11-27 04:12:12        1 -1.84995370   7.8283430          0.015625
#> 25: 2022-11-27 04:12:12        1 -4.12978337  13.5672926          0.015625
#> 26: 2022-11-27 04:12:12        1  5.90130543  13.9516995          0.015625
#> 27: 2022-11-27 04:12:12        1 -2.62074433  12.9202989          0.015625
#> 28: 2022-11-27 04:12:12        1  4.59717468  10.3372569          0.015625
#> 29: 2022-11-27 04:12:12        1 -2.89477384   0.5787086          0.015625
#> 30: 2022-11-27 04:12:12        1 -0.14819331   4.6861399          0.015625
#> 31: 2022-11-27 04:12:12        1 -3.23977534  11.9409980          0.015625
#> 32: 2022-11-27 04:12:12        1  6.73707951   9.2023321          0.015625
#> 33: 2022-11-27 04:12:12        1  8.98107277   5.9032478          0.015625
#> 34: 2022-11-27 04:12:12        1  6.90246270   4.1685406          0.015625
#> 35: 2022-11-27 04:12:12        1  9.93869528  14.5807992          0.015625
#> 36: 2022-11-27 04:12:12        1  3.78350973   4.1884080          0.015625
#> 37: 2022-11-27 04:12:12        1  0.43240912   5.4521879          0.015625
#> 38: 2022-11-27 04:12:12        1  8.67559354   4.0276064          0.015625
#> 39: 2022-11-27 04:12:12        1  3.41751152   9.1210391          0.015625
#> 40: 2022-11-27 04:12:12        1 -3.06801870   6.6915925          0.015625
#> 41: 2022-11-27 04:12:12        1  2.88610559   4.0307479          0.015625
#> 42: 2022-11-27 04:12:12        1  4.52387410   4.8892706          0.015625
#> 43: 2022-11-27 04:12:12        1 -4.54177518   9.2818750          0.015625
#> 44: 2022-11-27 04:12:12        1  3.47597396   8.4653055          0.015625
#> 45: 2022-11-27 04:12:12        1  3.53277681   0.2288282          0.015625
#> 46: 2022-11-27 04:12:12        1 -4.60307855   2.1806832          0.015625
#> 47: 2022-11-27 04:12:12        1  1.86524983   9.2893787          0.015625
#> 48: 2022-11-27 04:12:12        1  1.14938670   8.2018734          0.015625
#> 49: 2022-11-27 04:12:12        1  9.15976548   9.3095122          0.015625
#> 50: 2022-11-27 04:12:12        1 -3.01065543  11.2437923          0.015625
#> 51: 2022-11-27 04:12:12        1 -3.03504280  14.0602192          0.015625
#> 52: 2022-11-27 04:12:12        1 -4.12934878   6.6082030          0.015625
#> 53: 2022-11-27 04:12:12        1  8.50074445   1.6057775          0.015625
#> 54: 2022-11-27 04:12:12        1  9.75081422   1.1121492          0.015625
#> 55: 2022-11-27 04:12:12        1 -4.27046338  11.9909546          0.015625
#> 56: 2022-11-27 04:12:12        1  7.95434417  13.8459109          0.015625
#> 57: 2022-11-27 04:12:12        1  7.33848262  10.1354886          0.015625
#> 58: 2022-11-27 04:12:12        1  9.09373020   0.1521360          0.015625
#> 59: 2022-11-27 04:12:12        1 -1.34796784   5.0376767          0.015625
#> 60: 2022-11-27 04:12:12        1  1.03087054   7.1922346          0.015625
#> 61: 2022-11-27 04:12:12        1 -4.13360077  11.8335307          0.015625
#> 62: 2022-11-27 04:12:12        1  6.92353311  10.3409986          0.015625
#> 63: 2022-11-27 04:12:12        1  2.66616527   9.3258138          0.015625
#> 64: 2022-11-27 04:12:12        1  1.63300602   9.1090358          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity