Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.250854 11.96573 0.015625 <list[3]> 0.6731388

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.250854 11.96573 0.015625 <list[3]> 0.6731388

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -3.4792044 10.2255352 0.015625     0       6          1   3.7757269
#>  2: -3.2508540 11.9657283 0.015625     0       6          1   0.6731388
#>  3:  5.6881474  2.0147711 0.015625     0       6          1  34.5404372
#>  4: -4.6220112  5.6014031 0.015625     0       6          1  79.8884491
#>  5:  0.4609327  5.5947952 0.015625     0       6          1  18.7036113
#>  6:  5.7051613  6.3186899 0.015625     0       6          1  88.5678533
#>  7: -1.8423720  8.1945876 0.015625     0       6          1   8.1332074
#>  8:  1.4597262  2.3044349 0.015625     0       6          1  13.1317418
#>  9: -3.4196182  5.9238222 0.015625     0       6          1  35.3200647
#> 10:  5.6336185  6.3219230 0.015625     0       6          1  86.7403914
#> 11:  7.3616120  0.8952380 0.015625     0       6          1  38.9953016
#> 12:  6.8641658  5.0948472 0.015625     0       6          1  91.4853695
#> 13:  6.4747661 12.8785853 0.015625     0       6          1 272.0616017
#> 14:  1.1636268  7.3450381 0.015625     0       6          1  23.7589443
#> 15:  2.5687513  1.4282571 0.015625     0       6          1   2.4017787
#> 16: -1.9844731  8.2536382 0.015625     0       6          1   7.1924920
#> 17: -3.5033401  9.8323065 0.015625     0       6          1   5.5171408
#> 18: -1.5391412  6.2840689 0.015625     0       6          1  15.3142895
#> 19: -2.1615652 10.2185569 0.015625     0       6          1   5.0544162
#> 20:  1.2829265  9.9310402 0.015625     0       6          1  47.7994755
#> 21:  4.7729638 10.8612546 0.015625     0       6          1 148.8132576
#> 22:  8.8866707 12.6311738 0.015625     0       6          1 338.3514306
#> 23:  8.9907702 12.8916661 0.015625     0       6          1 351.5587762
#> 24: -3.7145044 12.7613316 0.015625     0       6          1   2.1119576
#> 25: -1.6075483  6.2025285 0.015625     0       6          1  15.5784578
#> 26:  5.8714085  0.8568103 0.015625     0       6          1  28.6684549
#> 27:  9.9212350  7.0254491 0.015625     0       6          1 191.6971775
#> 28: -0.2616090 14.9573293 0.015625     0       6          1  92.1875536
#> 29:  4.2657814 13.7216597 0.015625     0       6          1 200.4932597
#> 30:  0.7471410  9.6980650 0.015625     0       6          1  40.7614719
#> 31: -4.8646591  1.4441528 0.015625     0       6          1 181.1284028
#> 32: -3.1148308  8.8040812 0.015625     0       6          1   6.4118580
#> 33:  1.9411230  3.5842854 0.015625     0       6          1   6.8359728
#> 34:  5.3492735 11.6907259 0.015625     0       6          1 193.2537307
#> 35:  4.1657662 10.7807644 0.015625     0       6          1 123.3225937
#> 36: -2.6125495  1.8041253 0.015625     0       6          1  75.0481874
#> 37: -1.6543075  0.1881108 0.015625     0       6          1  81.9419976
#> 38:  9.7405749  9.0827492 0.015625     0       6          1 246.3625891
#> 39:  3.4743639 11.5013432 0.015625     0       6          1 114.5562366
#> 40:  1.1399449  6.0352589 0.015625     0       6          1  17.2848593
#> 41:  6.4337148  1.4854669 0.015625     0       6          1  39.3168679
#> 42: -2.4813440 12.4643531 0.015625     0       6          1   7.8254952
#> 43:  4.1863109  6.8504938 0.015625     0       6          1  53.8201381
#> 44:  0.1840589  2.4905627 0.015625     0       6          1  29.7924866
#> 45:  6.2564232  3.0963130 0.015625     0       6          1  53.8234530
#> 46: -0.4261668  7.8732737 0.015625     0       6          1  20.1580002
#> 47:  7.8927022  0.2003759 0.015625     0       6          1  33.1180618
#> 48:  0.7422300 11.6078832 0.015625     0       6          1  62.9396405
#> 49: -4.6579831 11.1542304 0.015625     0       6          1  18.0410303
#> 50: -4.1871306  4.5065432 0.015625     0       6          1  80.8149661
#> 51:  3.8141603  7.0615275 0.015625     0       6          1  47.1735403
#> 52:  2.1848548 14.4219364 0.015625     0       6          1 142.5878512
#> 53:  1.6973264  2.4187367 0.015625     0       6          1   9.7262061
#> 54: -0.6193160  7.1468138 0.015625     0       6          1  17.8410718
#> 55:  6.4989040  3.1902218 0.015625     0       6          1  58.2538606
#> 56:  8.5010583  3.2930779 0.015625     0       6          1  78.1869175
#> 57:  3.6833163  4.2588300 0.015625     0       6          1  15.4913658
#> 58:  0.6370147 14.0381275 0.015625     0       6          1  99.4312981
#> 59:  2.7894049  0.3714153 0.015625     0       6          1   3.0274310
#> 60:  2.2199992 14.1831493 0.015625     0       6          1 137.9408806
#> 61:  3.3583763  4.1589096 0.015625     0       6          1  10.5901988
#> 62:  2.6774099  1.7808139 0.015625     0       6          1   1.4457230
#> 63:  2.1580536 11.8244407 0.015625     0       6          1  87.7792018
#> 64:  7.5213475  8.0069572 0.015625     0       6          1 162.9083331
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-05-04 04:36:05        1  -3.4792044  10.2255352          0.015625
#>  2: 2022-05-04 04:36:05        1  -3.2508540  11.9657283          0.015625
#>  3: 2022-05-04 04:36:05        1   5.6881474   2.0147711          0.015625
#>  4: 2022-05-04 04:36:05        1  -4.6220112   5.6014031          0.015625
#>  5: 2022-05-04 04:36:05        1   0.4609327   5.5947952          0.015625
#>  6: 2022-05-04 04:36:05        1   5.7051613   6.3186899          0.015625
#>  7: 2022-05-04 04:36:05        1  -1.8423720   8.1945876          0.015625
#>  8: 2022-05-04 04:36:05        1   1.4597262   2.3044349          0.015625
#>  9: 2022-05-04 04:36:05        1  -3.4196182   5.9238222          0.015625
#> 10: 2022-05-04 04:36:05        1   5.6336185   6.3219230          0.015625
#> 11: 2022-05-04 04:36:05        1   7.3616120   0.8952380          0.015625
#> 12: 2022-05-04 04:36:05        1   6.8641658   5.0948472          0.015625
#> 13: 2022-05-04 04:36:05        1   6.4747661  12.8785853          0.015625
#> 14: 2022-05-04 04:36:05        1   1.1636268   7.3450381          0.015625
#> 15: 2022-05-04 04:36:05        1   2.5687513   1.4282571          0.015625
#> 16: 2022-05-04 04:36:05        1  -1.9844731   8.2536382          0.015625
#> 17: 2022-05-04 04:36:05        1  -3.5033401   9.8323065          0.015625
#> 18: 2022-05-04 04:36:05        1  -1.5391412   6.2840689          0.015625
#> 19: 2022-05-04 04:36:05        1  -2.1615652  10.2185569          0.015625
#> 20: 2022-05-04 04:36:05        1   1.2829265   9.9310402          0.015625
#> 21: 2022-05-04 04:36:05        1   4.7729638  10.8612546          0.015625
#> 22: 2022-05-04 04:36:05        1   8.8866707  12.6311738          0.015625
#> 23: 2022-05-04 04:36:05        1   8.9907702  12.8916661          0.015625
#> 24: 2022-05-04 04:36:05        1  -3.7145044  12.7613316          0.015625
#> 25: 2022-05-04 04:36:05        1  -1.6075483   6.2025285          0.015625
#> 26: 2022-05-04 04:36:05        1   5.8714085   0.8568103          0.015625
#> 27: 2022-05-04 04:36:05        1   9.9212350   7.0254491          0.015625
#> 28: 2022-05-04 04:36:05        1  -0.2616090  14.9573293          0.015625
#> 29: 2022-05-04 04:36:05        1   4.2657814  13.7216597          0.015625
#> 30: 2022-05-04 04:36:05        1   0.7471410   9.6980650          0.015625
#> 31: 2022-05-04 04:36:05        1  -4.8646591   1.4441528          0.015625
#> 32: 2022-05-04 04:36:05        1  -3.1148308   8.8040812          0.015625
#> 33: 2022-05-04 04:36:05        1   1.9411230   3.5842854          0.015625
#> 34: 2022-05-04 04:36:05        1   5.3492735  11.6907259          0.015625
#> 35: 2022-05-04 04:36:05        1   4.1657662  10.7807644          0.015625
#> 36: 2022-05-04 04:36:05        1  -2.6125495   1.8041253          0.015625
#> 37: 2022-05-04 04:36:05        1  -1.6543075   0.1881108          0.015625
#> 38: 2022-05-04 04:36:05        1   9.7405749   9.0827492          0.015625
#> 39: 2022-05-04 04:36:05        1   3.4743639  11.5013432          0.015625
#> 40: 2022-05-04 04:36:05        1   1.1399449   6.0352589          0.015625
#> 41: 2022-05-04 04:36:05        1   6.4337148   1.4854669          0.015625
#> 42: 2022-05-04 04:36:05        1  -2.4813440  12.4643531          0.015625
#> 43: 2022-05-04 04:36:05        1   4.1863109   6.8504938          0.015625
#> 44: 2022-05-04 04:36:05        1   0.1840589   2.4905627          0.015625
#> 45: 2022-05-04 04:36:05        1   6.2564232   3.0963130          0.015625
#> 46: 2022-05-04 04:36:05        1  -0.4261668   7.8732737          0.015625
#> 47: 2022-05-04 04:36:05        1   7.8927022   0.2003759          0.015625
#> 48: 2022-05-04 04:36:05        1   0.7422300  11.6078832          0.015625
#> 49: 2022-05-04 04:36:05        1  -4.6579831  11.1542304          0.015625
#> 50: 2022-05-04 04:36:05        1  -4.1871306   4.5065432          0.015625
#> 51: 2022-05-04 04:36:05        1   3.8141603   7.0615275          0.015625
#> 52: 2022-05-04 04:36:05        1   2.1848548  14.4219364          0.015625
#> 53: 2022-05-04 04:36:05        1   1.6973264   2.4187367          0.015625
#> 54: 2022-05-04 04:36:05        1  -0.6193160   7.1468138          0.015625
#> 55: 2022-05-04 04:36:05        1   6.4989040   3.1902218          0.015625
#> 56: 2022-05-04 04:36:05        1   8.5010583   3.2930779          0.015625
#> 57: 2022-05-04 04:36:05        1   3.6833163   4.2588300          0.015625
#> 58: 2022-05-04 04:36:05        1   0.6370147  14.0381275          0.015625
#> 59: 2022-05-04 04:36:05        1   2.7894049   0.3714153          0.015625
#> 60: 2022-05-04 04:36:05        1   2.2199992  14.1831493          0.015625
#> 61: 2022-05-04 04:36:05        1   3.3583763   4.1589096          0.015625
#> 62: 2022-05-04 04:36:05        1   2.6774099   1.7808139          0.015625
#> 63: 2022-05-04 04:36:05        1   2.1580536  11.8244407          0.015625
#> 64: 2022-05-04 04:36:05        1   7.5213475   8.0069572          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity