Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.418206 0.4087827 0.015625 <list[3]> 1.023234

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.418206 0.4087827 0.015625 <list[3]> 1.023234

# all evaluations
as.data.table(instance$archive)
#>             x1          x2 fidelity stage bracket repetition          y
#>  1: -0.5426932 12.70342778 0.015625     0       6          1  52.218959
#>  2: -4.4836199 11.49878139 0.015625     0       6          1  12.908479
#>  3:  7.5100457  3.55013235 0.015625     0       6          1  73.588711
#>  4:  9.5451188  1.55118928 0.015625     0       6          1  63.532839
#>  5:  4.3020486 13.75058299 0.015625     0       6          1 202.966977
#>  6:  6.3688239  2.38548177 0.015625     0       6          1  47.388562
#>  7:  9.0795745  1.46743557 0.015625     0       6          1  55.477239
#>  8:  2.6356557  2.76587621 0.015625     0       6          1   2.158935
#>  9:  6.4815834  1.14802022 0.015625     0       6          1  36.819925
#> 10:  4.6255363 10.31989404 0.015625     0       6          1 130.691541
#> 11: -4.0773607  8.20912056 0.015625     0       6          1  27.261209
#> 12:  4.9374072  2.93356599 0.015625     0       6          1  28.481387
#> 13:  3.2029430 14.78267757 0.015625     0       6          1 184.422219
#> 14: -4.1797765  0.09276157 0.015625     0       6          1 176.648417
#> 15:  4.0169585  1.87749885 0.015625     0       6          1   6.996816
#> 16:  0.2410891  3.99067021 0.015625     0       6          1  21.972874
#> 17:  0.7160663  2.12675984 0.015625     0       6          1  24.802703
#> 18:  9.0366118  1.47767727 0.015625     0       6          1  55.121216
#> 19: -3.9306271 11.91778890 0.015625     0       6          1   3.896040
#> 20:  3.8301096  9.51415767 0.015625     0       6          1  86.471091
#> 21:  0.6696025 10.40666758 0.015625     0       6          1  47.324887
#> 22:  0.2972516  1.18124453 0.015625     0       6          1  38.089436
#> 23: -2.3643640  4.35996442 0.015625     0       6          1  34.234775
#> 24: -3.4159672  2.18148508 0.015625     0       6          1  93.185614
#> 25: -3.7090653 14.89448952 0.015625     0       6          1   8.499257
#> 26:  3.1558935  7.30232402 0.015625     0       6          1  36.625480
#> 27: -4.1882647 12.25642795 0.015625     0       6          1   6.094707
#> 28:  3.4182059  0.40878267 0.015625     0       6          1   1.023234
#> 29: -2.9954935 13.27093117 0.015625     0       6          1   5.462220
#> 30:  1.1013155 11.31825887 0.015625     0       6          1  63.818056
#> 31:  4.5702190 10.19831583 0.015625     0       6          1 125.924801
#> 32:  9.4037054 10.21804699 0.015625     0       6          1 271.514717
#> 33:  1.7747653 14.81995306 0.015625     0       6          1 141.405157
#> 34: -4.7452416 13.47063315 0.015625     0       6          1  10.914464
#> 35:  5.6876621  3.28444913 0.015625     0       6          1  46.486059
#> 36:  7.2873856  4.85849363 0.015625     0       6          1  93.014878
#> 37: -1.8624582 13.95549049 0.015625     0       6          1  31.098698
#> 38:  0.1211653  7.12993523 0.015625     0       6          1  21.280256
#> 39: -4.1360132  1.03538112 0.015625     0       6          1 150.530691
#> 40:  2.2367177  5.94894773 0.015625     0       6          1  15.323794
#> 41: -3.2994040  4.82263555 0.015625     0       6          1  46.258664
#> 42: -1.7571010  8.70659577 0.015625     0       6          1   8.255580
#> 43: -2.1079322  7.47001612 0.015625     0       6          1   9.173210
#> 44: -4.2120311 11.38136791 0.015625     0       6          1   8.882077
#> 45:  8.6430851 10.01167184 0.015625     0       6          1 242.527037
#> 46: -2.7920045  6.96048266 0.015625     0       6          1  14.838013
#> 47:  8.8443799 13.16011166 0.015625     0       6          1 356.586429
#> 48:  4.0865798  6.02280066 0.015625     0       6          1  40.535490
#> 49:  4.7410002 14.27603519 0.015625     0       6          1 239.205759
#> 50:  7.7321378  6.20367985 0.015625     0       6          1 125.048406
#> 51:  1.3592233  9.17881258 0.015625     0       6          1  39.950599
#> 52: -1.0325489  3.28380116 0.015625     0       6          1  34.214936
#> 53:  1.2442100  7.09060432 0.015625     0       6          1  22.220376
#> 54:  1.1824327  1.10076498 0.015625     0       6          1  23.001655
#> 55: -1.5991518 14.72673454 0.015625     0       6          1  46.974085
#> 56:  5.2364980 12.10030180 0.015625     0       6          1 199.529719
#> 57:  6.5680724  2.01882786 0.015625     0       6          1  45.694618
#> 58:  9.3119794  8.55234641 0.015625     0       6          1 216.744690
#> 59:  1.3340632  5.81337748 0.015625     0       6          1  15.793452
#> 60:  8.3477171  1.21606163 0.015625     0       6          1  45.889820
#> 61:  9.6847709  3.95317949 0.015625     0       6          1 110.615184
#> 62:  8.3978765 13.73234836 0.015625     0       6          1 363.359853
#> 63:  5.2634086  9.13842324 0.015625     0       6          1 128.739596
#> 64: -3.4558665  2.78194908 0.015625     0       6          1  83.413502
#>             x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-21 04:38:44        1  -0.5426932 12.70342778          0.015625
#>  2: 2022-10-21 04:38:44        1  -4.4836199 11.49878139          0.015625
#>  3: 2022-10-21 04:38:44        1   7.5100457  3.55013235          0.015625
#>  4: 2022-10-21 04:38:44        1   9.5451188  1.55118928          0.015625
#>  5: 2022-10-21 04:38:44        1   4.3020486 13.75058299          0.015625
#>  6: 2022-10-21 04:38:44        1   6.3688239  2.38548177          0.015625
#>  7: 2022-10-21 04:38:44        1   9.0795745  1.46743557          0.015625
#>  8: 2022-10-21 04:38:44        1   2.6356557  2.76587621          0.015625
#>  9: 2022-10-21 04:38:44        1   6.4815834  1.14802022          0.015625
#> 10: 2022-10-21 04:38:44        1   4.6255363 10.31989404          0.015625
#> 11: 2022-10-21 04:38:44        1  -4.0773607  8.20912056          0.015625
#> 12: 2022-10-21 04:38:44        1   4.9374072  2.93356599          0.015625
#> 13: 2022-10-21 04:38:44        1   3.2029430 14.78267757          0.015625
#> 14: 2022-10-21 04:38:44        1  -4.1797765  0.09276157          0.015625
#> 15: 2022-10-21 04:38:44        1   4.0169585  1.87749885          0.015625
#> 16: 2022-10-21 04:38:44        1   0.2410891  3.99067021          0.015625
#> 17: 2022-10-21 04:38:44        1   0.7160663  2.12675984          0.015625
#> 18: 2022-10-21 04:38:44        1   9.0366118  1.47767727          0.015625
#> 19: 2022-10-21 04:38:44        1  -3.9306271 11.91778890          0.015625
#> 20: 2022-10-21 04:38:44        1   3.8301096  9.51415767          0.015625
#> 21: 2022-10-21 04:38:44        1   0.6696025 10.40666758          0.015625
#> 22: 2022-10-21 04:38:44        1   0.2972516  1.18124453          0.015625
#> 23: 2022-10-21 04:38:44        1  -2.3643640  4.35996442          0.015625
#> 24: 2022-10-21 04:38:44        1  -3.4159672  2.18148508          0.015625
#> 25: 2022-10-21 04:38:44        1  -3.7090653 14.89448952          0.015625
#> 26: 2022-10-21 04:38:44        1   3.1558935  7.30232402          0.015625
#> 27: 2022-10-21 04:38:44        1  -4.1882647 12.25642795          0.015625
#> 28: 2022-10-21 04:38:44        1   3.4182059  0.40878267          0.015625
#> 29: 2022-10-21 04:38:44        1  -2.9954935 13.27093117          0.015625
#> 30: 2022-10-21 04:38:44        1   1.1013155 11.31825887          0.015625
#> 31: 2022-10-21 04:38:44        1   4.5702190 10.19831583          0.015625
#> 32: 2022-10-21 04:38:44        1   9.4037054 10.21804699          0.015625
#> 33: 2022-10-21 04:38:44        1   1.7747653 14.81995306          0.015625
#> 34: 2022-10-21 04:38:44        1  -4.7452416 13.47063315          0.015625
#> 35: 2022-10-21 04:38:44        1   5.6876621  3.28444913          0.015625
#> 36: 2022-10-21 04:38:44        1   7.2873856  4.85849363          0.015625
#> 37: 2022-10-21 04:38:44        1  -1.8624582 13.95549049          0.015625
#> 38: 2022-10-21 04:38:44        1   0.1211653  7.12993523          0.015625
#> 39: 2022-10-21 04:38:44        1  -4.1360132  1.03538112          0.015625
#> 40: 2022-10-21 04:38:44        1   2.2367177  5.94894773          0.015625
#> 41: 2022-10-21 04:38:44        1  -3.2994040  4.82263555          0.015625
#> 42: 2022-10-21 04:38:44        1  -1.7571010  8.70659577          0.015625
#> 43: 2022-10-21 04:38:44        1  -2.1079322  7.47001612          0.015625
#> 44: 2022-10-21 04:38:44        1  -4.2120311 11.38136791          0.015625
#> 45: 2022-10-21 04:38:44        1   8.6430851 10.01167184          0.015625
#> 46: 2022-10-21 04:38:44        1  -2.7920045  6.96048266          0.015625
#> 47: 2022-10-21 04:38:44        1   8.8443799 13.16011166          0.015625
#> 48: 2022-10-21 04:38:44        1   4.0865798  6.02280066          0.015625
#> 49: 2022-10-21 04:38:44        1   4.7410002 14.27603519          0.015625
#> 50: 2022-10-21 04:38:44        1   7.7321378  6.20367985          0.015625
#> 51: 2022-10-21 04:38:44        1   1.3592233  9.17881258          0.015625
#> 52: 2022-10-21 04:38:44        1  -1.0325489  3.28380116          0.015625
#> 53: 2022-10-21 04:38:44        1   1.2442100  7.09060432          0.015625
#> 54: 2022-10-21 04:38:44        1   1.1824327  1.10076498          0.015625
#> 55: 2022-10-21 04:38:44        1  -1.5991518 14.72673454          0.015625
#> 56: 2022-10-21 04:38:44        1   5.2364980 12.10030180          0.015625
#> 57: 2022-10-21 04:38:44        1   6.5680724  2.01882786          0.015625
#> 58: 2022-10-21 04:38:44        1   9.3119794  8.55234641          0.015625
#> 59: 2022-10-21 04:38:44        1   1.3340632  5.81337748          0.015625
#> 60: 2022-10-21 04:38:44        1   8.3477171  1.21606163          0.015625
#> 61: 2022-10-21 04:38:44        1   9.6847709  3.95317949          0.015625
#> 62: 2022-10-21 04:38:44        1   8.3978765 13.73234836          0.015625
#> 63: 2022-10-21 04:38:44        1   5.2634086  9.13842324          0.015625
#> 64: 2022-10-21 04:38:44        1  -3.4558665  2.78194908          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity