Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.404563 11.62506 0.015625 <list[3]> 0.7504466

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.404563 11.62506 0.015625 <list[3]> 0.7504466

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition           y
#>  1: -1.22221071  3.6110446 0.015625     0       6          1  32.4650061
#>  2: -1.36366741  3.0849998 0.015625     0       6          1  38.4202078
#>  3:  6.51673110  6.9943698 0.015625     0       6          1 120.5513138
#>  4:  9.91849206  0.4805701 0.015625     0       6          1  53.9847422
#>  5:  8.14125103  7.0513871 0.015625     0       6          1 150.5765623
#>  6:  1.66914777  2.1336556 0.015625     0       6          1  10.7353984
#>  7: -4.17004492  4.3925255 0.015625     0       6          1  82.1143245
#>  8: -2.87300743 10.6129850 0.015625     0       6          1   0.7876628
#>  9:  9.44355161  4.8178940 0.015625     0       6          1 123.7369172
#> 10:  4.25320901  9.8952033 0.015625     0       6          1 107.9196993
#> 11:  1.83786266  6.6914418 0.015625     0       6          1  19.8045841
#> 12:  1.58954071 10.4157528 0.015625     0       6          1  56.9880583
#> 13: -1.05747586 13.5768657 0.015625     0       6          1  49.0485760
#> 14: -4.22235875  1.8276750 0.015625     0       6          1 136.3677266
#> 15:  6.25636591  3.5216408 0.015625     0       6          1  58.9799761
#> 16:  1.60009758  5.2257592 0.015625     0       6          1  12.5872101
#> 17:  1.66074853  2.8719054 0.015625     0       6          1   9.4620301
#> 18:  3.73625853 12.8532189 0.015625     0       6          1 155.0740669
#> 19: -0.05454609 10.8574939 0.015625     0       6          1  42.3463571
#> 20:  7.74380313 13.9408814 0.015625     0       6          1 350.4161924
#> 21:  8.22219172  0.5380762 0.015625     0       6          1  37.2959560
#> 22:  9.46129042 13.2236095 0.015625     0       6          1 381.8006249
#> 23:  9.93087571  1.4291494 0.015625     0       6          1  68.8791168
#> 24:  9.73060964 11.2634107 0.015625     0       6          1 319.0692235
#> 25:  0.15986735  6.8883694 0.015625     0       6          1  20.7838809
#> 26: -4.99832039 10.9886959 0.015625     0       6          1  26.6550532
#> 27:  8.47344370  3.1329645 0.015625     0       6          1  75.1742816
#> 28: -4.81816179 11.5788767 0.015625     0       6          1  18.8719495
#> 29:  7.65746553  9.6754360 0.015625     0       6          1 209.5518483
#> 30:  5.89872495  2.3907215 0.015625     0       6          1  41.0757824
#> 31:  8.32392075  5.5461583 0.015625     0       6          1 119.3663846
#> 32: -0.74728892 11.1619886 0.015625     0       6          1  32.6892281
#> 33:  1.36234817  1.8541083 0.015625     0       6          1  16.1271382
#> 34:  2.56653569 14.6571296 0.015625     0       6          1 159.1779337
#> 35:  3.27887434  5.4988612 0.015625     0       6          1  19.7321552
#> 36: -2.12180509  3.6272665 0.015625     0       6          1  39.6427532
#> 37:  5.76578610  7.3691550 0.015625     0       6          1 109.0428269
#> 38: -3.53784589 14.9944103 0.015625     0       6          1  10.0158445
#> 39:  6.56294951  8.7553664 0.015625     0       6          1 160.2748235
#> 40: -0.92548435 14.0380218 0.015625     0       6          1  58.5301873
#> 41: -4.53454455 10.1003990 0.015625     0       6          1  22.3546244
#> 42:  3.27005874  9.3772606 0.015625     0       6          1  68.5879363
#> 43: -4.96937945  5.4771348 0.015625     0       6          1  96.9181364
#> 44: -3.66389025 13.3045468 0.015625     0       6          1   2.8028181
#> 45:  8.31032432 10.2444093 0.015625     0       6          1 241.3072582
#> 46:  1.82258858  2.2606475 0.015625     0       6          1   8.4927373
#> 47:  1.87581224 13.1217707 0.015625     0       6          1 107.0970061
#> 48: -0.83092114  1.0766579 0.015625     0       6          1  55.7492781
#> 49: -3.91209725  0.1850900 0.015625     0       6          1 159.6544019
#> 50:  7.99574158  6.5780263 0.015625     0       6          1 137.1920177
#> 51: -4.69107399  9.6269726 0.015625     0       6          1  30.1871428
#> 52:  4.88146096  5.6665557 0.015625     0       6          1  56.5456933
#> 53:  5.86791581  1.9313759 0.015625     0       6          1  36.5249250
#> 54: -1.95627562 13.4719088 0.015625     0       6          1  24.3733670
#> 55:  2.45165204 12.0612233 0.015625     0       6          1  98.2099940
#> 56: -3.39946198 14.6396985 0.015625     0       6          1   8.9750441
#> 57:  7.66004535 10.2536038 0.015625     0       6          1 226.2042765
#> 58:  4.94924477  6.9504278 0.015625     0       6          1  77.4465719
#> 59:  5.38903982  6.6529046 0.015625     0       6          1  85.5162840
#> 60: -2.47673752  2.1480953 0.015625     0       6          1  66.1612144
#> 61: -3.40456312 11.6250592 0.015625     0       6          1   0.7504466
#> 62:  9.14508079  4.4699849 0.015625     0       6          1 110.0442646
#> 63:  0.45996277  9.9594566 0.015625     0       6          1  40.5534287
#> 64:  2.89064770  1.3296662 0.015625     0       6          1   0.8053380
#>              x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-11-03 04:24:08        1 -1.22221071   3.6110446          0.015625
#>  2: 2022-11-03 04:24:08        1 -1.36366741   3.0849998          0.015625
#>  3: 2022-11-03 04:24:08        1  6.51673110   6.9943698          0.015625
#>  4: 2022-11-03 04:24:08        1  9.91849206   0.4805701          0.015625
#>  5: 2022-11-03 04:24:08        1  8.14125103   7.0513871          0.015625
#>  6: 2022-11-03 04:24:08        1  1.66914777   2.1336556          0.015625
#>  7: 2022-11-03 04:24:08        1 -4.17004492   4.3925255          0.015625
#>  8: 2022-11-03 04:24:08        1 -2.87300743  10.6129850          0.015625
#>  9: 2022-11-03 04:24:08        1  9.44355161   4.8178940          0.015625
#> 10: 2022-11-03 04:24:08        1  4.25320901   9.8952033          0.015625
#> 11: 2022-11-03 04:24:08        1  1.83786266   6.6914418          0.015625
#> 12: 2022-11-03 04:24:08        1  1.58954071  10.4157528          0.015625
#> 13: 2022-11-03 04:24:08        1 -1.05747586  13.5768657          0.015625
#> 14: 2022-11-03 04:24:08        1 -4.22235875   1.8276750          0.015625
#> 15: 2022-11-03 04:24:08        1  6.25636591   3.5216408          0.015625
#> 16: 2022-11-03 04:24:08        1  1.60009758   5.2257592          0.015625
#> 17: 2022-11-03 04:24:08        1  1.66074853   2.8719054          0.015625
#> 18: 2022-11-03 04:24:08        1  3.73625853  12.8532189          0.015625
#> 19: 2022-11-03 04:24:08        1 -0.05454609  10.8574939          0.015625
#> 20: 2022-11-03 04:24:08        1  7.74380313  13.9408814          0.015625
#> 21: 2022-11-03 04:24:08        1  8.22219172   0.5380762          0.015625
#> 22: 2022-11-03 04:24:08        1  9.46129042  13.2236095          0.015625
#> 23: 2022-11-03 04:24:08        1  9.93087571   1.4291494          0.015625
#> 24: 2022-11-03 04:24:08        1  9.73060964  11.2634107          0.015625
#> 25: 2022-11-03 04:24:08        1  0.15986735   6.8883694          0.015625
#> 26: 2022-11-03 04:24:08        1 -4.99832039  10.9886959          0.015625
#> 27: 2022-11-03 04:24:08        1  8.47344370   3.1329645          0.015625
#> 28: 2022-11-03 04:24:08        1 -4.81816179  11.5788767          0.015625
#> 29: 2022-11-03 04:24:08        1  7.65746553   9.6754360          0.015625
#> 30: 2022-11-03 04:24:08        1  5.89872495   2.3907215          0.015625
#> 31: 2022-11-03 04:24:08        1  8.32392075   5.5461583          0.015625
#> 32: 2022-11-03 04:24:08        1 -0.74728892  11.1619886          0.015625
#> 33: 2022-11-03 04:24:08        1  1.36234817   1.8541083          0.015625
#> 34: 2022-11-03 04:24:08        1  2.56653569  14.6571296          0.015625
#> 35: 2022-11-03 04:24:08        1  3.27887434   5.4988612          0.015625
#> 36: 2022-11-03 04:24:08        1 -2.12180509   3.6272665          0.015625
#> 37: 2022-11-03 04:24:08        1  5.76578610   7.3691550          0.015625
#> 38: 2022-11-03 04:24:08        1 -3.53784589  14.9944103          0.015625
#> 39: 2022-11-03 04:24:08        1  6.56294951   8.7553664          0.015625
#> 40: 2022-11-03 04:24:08        1 -0.92548435  14.0380218          0.015625
#> 41: 2022-11-03 04:24:08        1 -4.53454455  10.1003990          0.015625
#> 42: 2022-11-03 04:24:08        1  3.27005874   9.3772606          0.015625
#> 43: 2022-11-03 04:24:08        1 -4.96937945   5.4771348          0.015625
#> 44: 2022-11-03 04:24:08        1 -3.66389025  13.3045468          0.015625
#> 45: 2022-11-03 04:24:08        1  8.31032432  10.2444093          0.015625
#> 46: 2022-11-03 04:24:08        1  1.82258858   2.2606475          0.015625
#> 47: 2022-11-03 04:24:08        1  1.87581224  13.1217707          0.015625
#> 48: 2022-11-03 04:24:08        1 -0.83092114   1.0766579          0.015625
#> 49: 2022-11-03 04:24:08        1 -3.91209725   0.1850900          0.015625
#> 50: 2022-11-03 04:24:08        1  7.99574158   6.5780263          0.015625
#> 51: 2022-11-03 04:24:08        1 -4.69107399   9.6269726          0.015625
#> 52: 2022-11-03 04:24:08        1  4.88146096   5.6665557          0.015625
#> 53: 2022-11-03 04:24:08        1  5.86791581   1.9313759          0.015625
#> 54: 2022-11-03 04:24:08        1 -1.95627562  13.4719088          0.015625
#> 55: 2022-11-03 04:24:08        1  2.45165204  12.0612233          0.015625
#> 56: 2022-11-03 04:24:08        1 -3.39946198  14.6396985          0.015625
#> 57: 2022-11-03 04:24:08        1  7.66004535  10.2536038          0.015625
#> 58: 2022-11-03 04:24:08        1  4.94924477   6.9504278          0.015625
#> 59: 2022-11-03 04:24:08        1  5.38903982   6.6529046          0.015625
#> 60: 2022-11-03 04:24:08        1 -2.47673752   2.1480953          0.015625
#> 61: 2022-11-03 04:24:08        1 -3.40456312  11.6250592          0.015625
#> 62: 2022-11-03 04:24:08        1  9.14508079   4.4699849          0.015625
#> 63: 2022-11-03 04:24:08        1  0.45996277   9.9594566          0.015625
#> 64: 2022-11-03 04:24:08        1  2.89064770   1.3296662          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity