Optimizer using the Hyperband algorithm

OptimizerHyperband class that implements hyperband optimization. HyperbandX (HBX) repeatedly calls SH (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SH within HBX is called a bracket. HBX considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s		3			2			1			0

i	ni	ri		ni	ri		ni	ri		ni	ri

0	8	1		6	2		4	4		8	4

1	4	2		3	4		2	8

2	2	4		1	8

3	1	8

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repeats

logical(1)
If FALSE (default), hyperband terminates once all brackets are evaluated. Otherwise, hyperband starts over again once the last bracket is evaluated.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

Custom sampler

All algorithms support custom paradox::Samplers to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: -3.84525 11.34062 0.015625 <list[3]> 4.201115

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: -3.84525 11.34062 0.015625 <list[3]> 4.201115

# all evaluations
as.data.table(instance$archive)
#>               x1         x2 fidelity bracket stage          y
#>  1:  8.010253073 10.2985382 0.015625       6     0 235.742980
#>  2: -1.271919546  2.4543480 0.015625       6     0  44.408529
#>  3:  1.043218178 14.2923720 0.015625       6     0 113.225495
#>  4:  6.544452640  4.8278183 0.015625       6     0  82.109584
#>  5: -3.207719382  5.4230118 0.015625       6     0  36.402105
#>  6: -2.079575575 13.3158513 0.015625       6     0  20.323870
#>  7: -2.531461277 12.4202163 0.015625       6     0   6.945035
#>  8:  4.948098719  1.5098469 0.015625       6     0  19.170839
#>  9:  7.848625068 13.5907737 0.015625       6     0 340.862218
#> 10:  8.898196726 11.5909555 0.015625       6     0 301.625371
#> 11:  3.285663920  5.7505601 0.015625       6     0  22.100571
#> 12:  3.655985416 14.9947869 0.015625       6     0 209.072395
#> 13:  5.311716184  5.2394857 0.015625       6     0  62.008497
#> 14: -1.329226556 14.2097740 0.015625       6     0  48.777708
#> 15: -4.330742625  3.2414996 0.015625       6     0 111.030813
#> 16:  8.647818350  0.4813906 0.015625       6     0  38.501167
#> 17: -3.939781712  2.1797377 0.015625       6     0 114.977580
#> 18:  9.953372101 12.8157583 0.015625       6     0 386.297680
#> 19:  4.177786262  3.1972397 0.015625       6     0  16.062096
#> 20: -2.411617314  3.1546611 0.015625       6     0  49.936696
#> 21:  8.641614477  0.5928103 0.015625       6     0  39.800737
#> 22: -4.438232497 14.1716220 0.015625       6     0   7.652681
#> 23:  3.903306855  3.6739198 0.015625       6     0  14.732546
#> 24: -1.445336677 11.7168385 0.015625       6     0  22.439323
#> 25:  8.594459001  4.3235575 0.015625       6     0  98.213307
#> 26:  7.283094760 13.1303687 0.015625       6     0 307.285790
#> 27:  5.497440358  4.4362514 0.015625       6     0  55.930755
#> 28: -1.699995062 14.7528811 0.015625       6     0  44.265358
#> 29:  5.919864075  8.8475633 0.015625       6     0 144.231626
#> 30: -1.743733066 11.3873757 0.015625       6     0  14.691299
#> 31:  1.843452974 12.5411296 0.015625       6     0  95.222158
#> 32: -0.008003626 11.4422920 0.015625       6     0  49.081839
#> 33:  3.525290028  6.2590489 0.015625       6     0  31.209935
#> 34: -1.216914128  2.0711226 0.015625       6     0  48.269691
#> 35:  1.960203501  1.2126744 0.015625       6     0   9.543403
#> 36:  8.764907604  9.8397394 0.015625       6     0 240.419146
#> 37:  9.592663269  9.0300579 0.015625       6     0 239.790083
#> 38:  7.286236699  9.8549374 0.015625       6     0 206.128341
#> 39:  8.543856973  4.9397575 0.015625       6     0 109.840303
#> 40:  3.720490660 14.6921133 0.015625       6     0 203.257668
#> 41:  6.595127217 10.7277920 0.015625       6     0 211.984706
#> 42:  9.926845396 13.0894546 0.015625       6     0 395.948470
#> 43:  5.664568746 14.7492562 0.015625       6     0 299.327560
#> 44: -1.775861065  3.2784449 0.015625       6     0  39.909523
#> 45: -0.623635551  9.9679510 0.015625       6     0  26.576630
#> 46:  5.826395940  5.8434606 0.015625       6     0  83.786001
#> 47:  7.999235550  0.6909546 0.015625       6     0  38.363916
#> 48: -1.423203403  9.2537184 0.015625       6     0  12.270173
#> 49: -4.932555379  8.9771249 0.015625       6     0  43.696698
#> 50:  9.152746963  6.1028044 0.015625       6     0 147.018029
#> 51:  1.572058006 12.8749222 0.015625       6     0  96.495429
#> 52:  6.259049923  7.7652178 0.015625       6     0 130.317468
#> 53:  5.017236467 14.6894012 0.015625       6     0 265.711035
#> 54:  1.119598022  0.2552353 0.015625       6     0  30.198231
#> 55:  0.268732224 10.1017174 0.015625       6     0  39.752994
#> 56:  6.071373422  5.5690482 0.015625       6     0  84.974129
#> 57:  4.964282393 13.7701596 0.015625       6     0 234.800565
#> 58: -3.721629498 10.1696714 0.015625       6     0   6.717944
#> 59:  7.841982350  9.9772869 0.015625       6     0 222.322909
#> 60: -3.845250175 11.3406163 0.015625       6     0   4.201115
#> 61:  7.792672050  8.1425573 0.015625       6     0 171.316094
#> 62: -3.404795587  3.5893215 0.015625       6     0  67.739456
#> 63:  2.272042353  7.6334036 0.015625       6     0  29.720737
#> 64: -1.291713344  6.2589656 0.015625       6     0  16.060838
#>               x1         x2 fidelity bracket stage          y
#>               timestamp batch_nr    status  x_domain_x1 x_domain_x2
#>  1: 2021-11-26 14:30:24        1 evaluated  8.010253073  10.2985382
#>  2: 2021-11-26 14:30:24        1 evaluated -1.271919546   2.4543480
#>  3: 2021-11-26 14:30:24        1 evaluated  1.043218178  14.2923720
#>  4: 2021-11-26 14:30:24        1 evaluated  6.544452640   4.8278183
#>  5: 2021-11-26 14:30:24        1 evaluated -3.207719382   5.4230118
#>  6: 2021-11-26 14:30:24        1 evaluated -2.079575575  13.3158513
#>  7: 2021-11-26 14:30:24        1 evaluated -2.531461277  12.4202163
#>  8: 2021-11-26 14:30:24        1 evaluated  4.948098719   1.5098469
#>  9: 2021-11-26 14:30:24        1 evaluated  7.848625068  13.5907737
#> 10: 2021-11-26 14:30:24        1 evaluated  8.898196726  11.5909555
#> 11: 2021-11-26 14:30:24        1 evaluated  3.285663920   5.7505601
#> 12: 2021-11-26 14:30:24        1 evaluated  3.655985416  14.9947869
#> 13: 2021-11-26 14:30:24        1 evaluated  5.311716184   5.2394857
#> 14: 2021-11-26 14:30:24        1 evaluated -1.329226556  14.2097740
#> 15: 2021-11-26 14:30:24        1 evaluated -4.330742625   3.2414996
#> 16: 2021-11-26 14:30:24        1 evaluated  8.647818350   0.4813906
#> 17: 2021-11-26 14:30:24        1 evaluated -3.939781712   2.1797377
#> 18: 2021-11-26 14:30:24        1 evaluated  9.953372101  12.8157583
#> 19: 2021-11-26 14:30:24        1 evaluated  4.177786262   3.1972397
#> 20: 2021-11-26 14:30:24        1 evaluated -2.411617314   3.1546611
#> 21: 2021-11-26 14:30:24        1 evaluated  8.641614477   0.5928103
#> 22: 2021-11-26 14:30:24        1 evaluated -4.438232497  14.1716220
#> 23: 2021-11-26 14:30:24        1 evaluated  3.903306855   3.6739198
#> 24: 2021-11-26 14:30:24        1 evaluated -1.445336677  11.7168385
#> 25: 2021-11-26 14:30:24        1 evaluated  8.594459001   4.3235575
#> 26: 2021-11-26 14:30:24        1 evaluated  7.283094760  13.1303687
#> 27: 2021-11-26 14:30:24        1 evaluated  5.497440358   4.4362514
#> 28: 2021-11-26 14:30:24        1 evaluated -1.699995062  14.7528811
#> 29: 2021-11-26 14:30:24        1 evaluated  5.919864075   8.8475633
#> 30: 2021-11-26 14:30:24        1 evaluated -1.743733066  11.3873757
#> 31: 2021-11-26 14:30:24        1 evaluated  1.843452974  12.5411296
#> 32: 2021-11-26 14:30:24        1 evaluated -0.008003626  11.4422920
#> 33: 2021-11-26 14:30:24        1 evaluated  3.525290028   6.2590489
#> 34: 2021-11-26 14:30:24        1 evaluated -1.216914128   2.0711226
#> 35: 2021-11-26 14:30:24        1 evaluated  1.960203501   1.2126744
#> 36: 2021-11-26 14:30:24        1 evaluated  8.764907604   9.8397394
#> 37: 2021-11-26 14:30:24        1 evaluated  9.592663269   9.0300579
#> 38: 2021-11-26 14:30:24        1 evaluated  7.286236699   9.8549374
#> 39: 2021-11-26 14:30:24        1 evaluated  8.543856973   4.9397575
#> 40: 2021-11-26 14:30:24        1 evaluated  3.720490660  14.6921133
#> 41: 2021-11-26 14:30:24        1 evaluated  6.595127217  10.7277920
#> 42: 2021-11-26 14:30:24        1 evaluated  9.926845396  13.0894546
#> 43: 2021-11-26 14:30:24        1 evaluated  5.664568746  14.7492562
#> 44: 2021-11-26 14:30:24        1 evaluated -1.775861065   3.2784449
#> 45: 2021-11-26 14:30:24        1 evaluated -0.623635551   9.9679510
#> 46: 2021-11-26 14:30:24        1 evaluated  5.826395940   5.8434606
#> 47: 2021-11-26 14:30:24        1 evaluated  7.999235550   0.6909546
#> 48: 2021-11-26 14:30:24        1 evaluated -1.423203403   9.2537184
#> 49: 2021-11-26 14:30:24        1 evaluated -4.932555379   8.9771249
#> 50: 2021-11-26 14:30:24        1 evaluated  9.152746963   6.1028044
#> 51: 2021-11-26 14:30:24        1 evaluated  1.572058006  12.8749222
#> 52: 2021-11-26 14:30:24        1 evaluated  6.259049923   7.7652178
#> 53: 2021-11-26 14:30:24        1 evaluated  5.017236467  14.6894012
#> 54: 2021-11-26 14:30:24        1 evaluated  1.119598022   0.2552353
#> 55: 2021-11-26 14:30:24        1 evaluated  0.268732224  10.1017174
#> 56: 2021-11-26 14:30:24        1 evaluated  6.071373422   5.5690482
#> 57: 2021-11-26 14:30:24        1 evaluated  4.964282393  13.7701596
#> 58: 2021-11-26 14:30:24        1 evaluated -3.721629498  10.1696714
#> 59: 2021-11-26 14:30:24        1 evaluated  7.841982350   9.9772869
#> 60: 2021-11-26 14:30:24        1 evaluated -3.845250175  11.3406163
#> 61: 2021-11-26 14:30:24        1 evaluated  7.792672050   8.1425573
#> 62: 2021-11-26 14:30:24        1 evaluated -3.404795587   3.5893215
#> 63: 2021-11-26 14:30:24        1 evaluated  2.272042353   7.6334036
#> 64: 2021-11-26 14:30:24        1 evaluated -1.291713344   6.2589656
#>               timestamp batch_nr    status  x_domain_x1 x_domain_x2
#>     x_domain_fidelity
#>  1:          0.015625
#>  2:          0.015625
#>  3:          0.015625
#>  4:          0.015625
#>  5:          0.015625
#>  6:          0.015625
#>  7:          0.015625
#>  8:          0.015625
#>  9:          0.015625
#> 10:          0.015625
#> 11:          0.015625
#> 12:          0.015625
#> 13:          0.015625
#> 14:          0.015625
#> 15:          0.015625
#> 16:          0.015625
#> 17:          0.015625
#> 18:          0.015625
#> 19:          0.015625
#> 20:          0.015625
#> 21:          0.015625
#> 22:          0.015625
#> 23:          0.015625
#> 24:          0.015625
#> 25:          0.015625
#> 26:          0.015625
#> 27:          0.015625
#> 28:          0.015625
#> 29:          0.015625
#> 30:          0.015625
#> 31:          0.015625
#> 32:          0.015625
#> 33:          0.015625
#> 34:          0.015625
#> 35:          0.015625
#> 36:          0.015625
#> 37:          0.015625
#> 38:          0.015625
#> 39:          0.015625
#> 40:          0.015625
#> 41:          0.015625
#> 42:          0.015625
#> 43:          0.015625
#> 44:          0.015625
#> 45:          0.015625
#> 46:          0.015625
#> 47:          0.015625
#> 48:          0.015625
#> 49:          0.015625
#> 50:          0.015625
#> 51:          0.015625
#> 52:          0.015625
#> 53:          0.015625
#> 54:          0.015625
#> 55:          0.015625
#> 56:          0.015625
#> 57:          0.015625
#> 58:          0.015625
#> 59:          0.015625
#> 60:          0.015625
#> 61:          0.015625
#> 62:          0.015625
#> 63:          0.015625
#> 64:          0.015625
#>     x_domain_fidelity