Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.105565 10.26358 0.015625 <list[3]> 1.355957

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.105565 10.26358 0.015625 <list[3]> 1.355957

# all evaluations
as.data.table(instance$archive)
#>               x1          x2 fidelity stage bracket repetition          y
#>  1: -0.309853938 10.49781881 0.015625     0       6          1  35.158596
#>  2:  1.090192186  1.72365425 0.015625     0       6          1  21.084224
#>  3:  1.438197696 11.51088770 0.015625     0       6          1  71.119108
#>  4: -3.989443796  8.70393545 0.015625     0       6          1  20.744022
#>  5: -2.570256294  2.44885681 0.015625     0       6          1  63.466216
#>  6:  2.432474256  4.36742868 0.015625     0       6          1   6.943462
#>  7: -0.001833239 11.98700797 0.015625     0       6          1  55.411432
#>  8: -2.558112607 13.01574975 0.015625     0       6          1   9.511592
#>  9:  4.608651719  5.21005075 0.015625     0       6          1  43.720073
#> 10: -3.105565133 10.26357890 0.015625     0       6          1   1.355957
#> 11:  1.851540065  7.02644542 0.015625     0       6          1  22.299850
#> 12:  1.239349976  8.41460197 0.015625     0       6          1  31.959032
#> 13: -1.020097402  6.76495149 0.015625     0       6          1  15.817753
#> 14:  2.728715263  3.45468141 0.015625     0       6          1   3.665346
#> 15: -4.296657298  9.29327465 0.015625     0       6          1  23.036369
#> 16: -1.305783315  2.94754536 0.015625     0       6          1  39.379520
#> 17: -2.213431454  6.31429474 0.015625     0       6          1  15.529134
#> 18: -0.559464227  4.74507393 0.015625     0       6          1  22.782044
#> 19:  5.531950550  2.80470909 0.015625     0       6          1  38.809270
#> 20: -1.101927569 10.54678153 0.015625     0       6          1  21.932722
#> 21:  3.920551297  6.59415050 0.015625     0       6          1  43.632811
#> 22: -3.861520259  2.64023043 0.015625     0       6          1 102.062891
#> 23: -4.281919230  6.30266670 0.015625     0       6          1  56.062279
#> 24:  3.999765406 14.44159855 0.015625     0       6          1 208.656291
#> 25:  1.769140651  3.70447394 0.015625     0       6          1   8.287642
#> 26:  8.373367812  0.01723869 0.015625     0       6          1  32.150331
#> 27: -2.752253001  6.42601326 0.015625     0       6          1  18.649480
#> 28: -4.828269696  5.27160497 0.015625     0       6          1  94.459898
#> 29:  5.767956368  1.14775871 0.015625     0       6          1  29.277363
#> 30:  8.475646428  7.31013199 0.015625     0       6          1 162.935628
#> 31: -2.008560246 12.58877524 0.015625     0       6          1  16.609402
#> 32:  7.224520436  0.07378700 0.015625     0       6          1  31.391278
#> 33:  4.882737022  5.99436181 0.015625     0       6          1  61.082957
#> 34:  5.558177165 10.50099728 0.015625     0       6          1 170.878519
#> 35: -3.972643808 11.88529901 0.015625     0       6          1   4.378421
#> 36:  2.517368679 14.16837127 0.015625     0       6          1 145.730034
#> 37:  0.202317785 14.02698902 0.015625     0       6          1  89.090846
#> 38:  8.127690305 14.13666958 0.015625     0       6          1 369.970401
#> 39:  8.987472492 10.29664943 0.015625     0       6          1 261.061651
#> 40:  0.182358273 12.71187723 0.015625     0       6          1  68.458116
#> 41:  8.876365765  0.83874267 0.015625     0       6          1  44.621627
#> 42:  7.052028353  8.42319615 0.015625     0       6          1 163.741420
#> 43: -3.119183260  4.81525464 0.015625     0       6          1  41.979926
#> 44: -0.298931019  1.81549207 0.015625     0       6          1  40.920023
#> 45:  6.473501677  7.15666177 0.015625     0       6          1 122.879378
#> 46: -4.419435122  1.32810913 0.015625     0       6          1 158.668973
#> 47:  7.021919512  5.14930480 0.015625     0       6          1  94.697232
#> 48:  4.961487154  5.79461480 0.015625     0       6          1  60.450167
#> 49: -2.432939772  8.10839123 0.015625     0       6          1   6.495643
#> 50:  6.712148745  9.96570257 0.015625     0       6          1 194.643971
#> 51:  4.087060476 14.97545571 0.015625     0       6          1 228.378822
#> 52:  2.201727150  0.82757932 0.015625     0       6          1   7.638370
#> 53: -0.909214225  6.10123110 0.015625     0       6          1  17.779526
#> 54: -4.443012165  5.86585299 0.015625     0       6          1  68.477824
#> 55:  8.843989694 13.48866840 0.015625     0       6          1 369.055041
#> 56:  9.217465841  4.50205426 0.015625     0       6          1 112.112641
#> 57: -0.718039169 14.15738329 0.015625     0       6          1  66.213629
#> 58: -3.580762681  7.99297163 0.015625     0       6          1  18.120910
#> 59:  5.012580012  7.81132601 0.015625     0       6          1  94.137487
#> 60: -2.345509918  4.55433693 0.015625     0       6          1  31.882286
#> 61:  9.875968935 13.05577844 0.015625     0       6          1 392.408183
#> 62: -4.145185760  8.36398045 0.015625     0       6          1  27.513857
#> 63: -2.074520666  9.25487650 0.015625     0       6          1   5.397245
#> 64: -4.603780026 11.47992219 0.015625     0       6          1  15.203663
#>               x1          x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-19 04:50:40        1 -0.309853938 10.49781881          0.015625
#>  2: 2022-10-19 04:50:40        1  1.090192186  1.72365425          0.015625
#>  3: 2022-10-19 04:50:40        1  1.438197696 11.51088770          0.015625
#>  4: 2022-10-19 04:50:40        1 -3.989443796  8.70393545          0.015625
#>  5: 2022-10-19 04:50:40        1 -2.570256294  2.44885681          0.015625
#>  6: 2022-10-19 04:50:40        1  2.432474256  4.36742868          0.015625
#>  7: 2022-10-19 04:50:40        1 -0.001833239 11.98700797          0.015625
#>  8: 2022-10-19 04:50:40        1 -2.558112607 13.01574975          0.015625
#>  9: 2022-10-19 04:50:40        1  4.608651719  5.21005075          0.015625
#> 10: 2022-10-19 04:50:40        1 -3.105565133 10.26357890          0.015625
#> 11: 2022-10-19 04:50:40        1  1.851540065  7.02644542          0.015625
#> 12: 2022-10-19 04:50:40        1  1.239349976  8.41460197          0.015625
#> 13: 2022-10-19 04:50:40        1 -1.020097402  6.76495149          0.015625
#> 14: 2022-10-19 04:50:40        1  2.728715263  3.45468141          0.015625
#> 15: 2022-10-19 04:50:40        1 -4.296657298  9.29327465          0.015625
#> 16: 2022-10-19 04:50:40        1 -1.305783315  2.94754536          0.015625
#> 17: 2022-10-19 04:50:40        1 -2.213431454  6.31429474          0.015625
#> 18: 2022-10-19 04:50:40        1 -0.559464227  4.74507393          0.015625
#> 19: 2022-10-19 04:50:40        1  5.531950550  2.80470909          0.015625
#> 20: 2022-10-19 04:50:40        1 -1.101927569 10.54678153          0.015625
#> 21: 2022-10-19 04:50:40        1  3.920551297  6.59415050          0.015625
#> 22: 2022-10-19 04:50:40        1 -3.861520259  2.64023043          0.015625
#> 23: 2022-10-19 04:50:40        1 -4.281919230  6.30266670          0.015625
#> 24: 2022-10-19 04:50:40        1  3.999765406 14.44159855          0.015625
#> 25: 2022-10-19 04:50:40        1  1.769140651  3.70447394          0.015625
#> 26: 2022-10-19 04:50:40        1  8.373367812  0.01723869          0.015625
#> 27: 2022-10-19 04:50:40        1 -2.752253001  6.42601326          0.015625
#> 28: 2022-10-19 04:50:40        1 -4.828269696  5.27160497          0.015625
#> 29: 2022-10-19 04:50:40        1  5.767956368  1.14775871          0.015625
#> 30: 2022-10-19 04:50:40        1  8.475646428  7.31013199          0.015625
#> 31: 2022-10-19 04:50:40        1 -2.008560246 12.58877524          0.015625
#> 32: 2022-10-19 04:50:40        1  7.224520436  0.07378700          0.015625
#> 33: 2022-10-19 04:50:40        1  4.882737022  5.99436181          0.015625
#> 34: 2022-10-19 04:50:40        1  5.558177165 10.50099728          0.015625
#> 35: 2022-10-19 04:50:40        1 -3.972643808 11.88529901          0.015625
#> 36: 2022-10-19 04:50:40        1  2.517368679 14.16837127          0.015625
#> 37: 2022-10-19 04:50:40        1  0.202317785 14.02698902          0.015625
#> 38: 2022-10-19 04:50:40        1  8.127690305 14.13666958          0.015625
#> 39: 2022-10-19 04:50:40        1  8.987472492 10.29664943          0.015625
#> 40: 2022-10-19 04:50:40        1  0.182358273 12.71187723          0.015625
#> 41: 2022-10-19 04:50:40        1  8.876365765  0.83874267          0.015625
#> 42: 2022-10-19 04:50:40        1  7.052028353  8.42319615          0.015625
#> 43: 2022-10-19 04:50:40        1 -3.119183260  4.81525464          0.015625
#> 44: 2022-10-19 04:50:40        1 -0.298931019  1.81549207          0.015625
#> 45: 2022-10-19 04:50:40        1  6.473501677  7.15666177          0.015625
#> 46: 2022-10-19 04:50:40        1 -4.419435122  1.32810913          0.015625
#> 47: 2022-10-19 04:50:40        1  7.021919512  5.14930480          0.015625
#> 48: 2022-10-19 04:50:40        1  4.961487154  5.79461480          0.015625
#> 49: 2022-10-19 04:50:40        1 -2.432939772  8.10839123          0.015625
#> 50: 2022-10-19 04:50:40        1  6.712148745  9.96570257          0.015625
#> 51: 2022-10-19 04:50:40        1  4.087060476 14.97545571          0.015625
#> 52: 2022-10-19 04:50:40        1  2.201727150  0.82757932          0.015625
#> 53: 2022-10-19 04:50:40        1 -0.909214225  6.10123110          0.015625
#> 54: 2022-10-19 04:50:40        1 -4.443012165  5.86585299          0.015625
#> 55: 2022-10-19 04:50:40        1  8.843989694 13.48866840          0.015625
#> 56: 2022-10-19 04:50:40        1  9.217465841  4.50205426          0.015625
#> 57: 2022-10-19 04:50:40        1 -0.718039169 14.15738329          0.015625
#> 58: 2022-10-19 04:50:40        1 -3.580762681  7.99297163          0.015625
#> 59: 2022-10-19 04:50:40        1  5.012580012  7.81132601          0.015625
#> 60: 2022-10-19 04:50:40        1 -2.345509918  4.55433693          0.015625
#> 61: 2022-10-19 04:50:40        1  9.875968935 13.05577844          0.015625
#> 62: 2022-10-19 04:50:40        1 -4.145185760  8.36398045          0.015625
#> 63: 2022-10-19 04:50:40        1 -2.074520666  9.25487650          0.015625
#> 64: 2022-10-19 04:50:40        1 -4.603780026 11.47992219          0.015625
#>               timestamp batch_nr  x_domain_x1 x_domain_x2 x_domain_fidelity