Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.356162 10.97098 0.015625 <list[3]> 1.131956

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.356162 10.97098 0.015625 <list[3]> 1.131956

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  7.46568020 10.8750887 0.015625     0       6          1 239.938224
#>  2: -2.70707998  1.2392707 0.015625     0       6          1  87.677959
#>  3: -3.35616210 10.9709761 0.015625     0       6          1   1.131956
#>  4: -2.38063878  5.1381651 0.015625     0       6          1  26.326943
#>  5:  2.77480248  1.1478916 0.015625     0       6          1   1.488994
#>  6:  2.37772667 10.5967598 0.015625     0       6          1  70.423776
#>  7: -0.04825100 13.8338799 0.015625     0       6          1  79.762210
#>  8:  2.04723665  2.1001836 0.015625     0       6          1   6.189813
#>  9: -3.65086913  1.5988517 0.015625     0       6          1 114.432874
#> 10: -2.10231190  6.2092696 0.015625     0       6          1  15.842889
#> 11:  1.19150169  1.5119024 0.015625     0       6          1  20.500739
#> 12: -0.76046166 12.9848231 0.015625     0       6          1  50.096839
#> 13: -3.75637143  8.3074476 0.015625     0       6          1  19.005846
#> 14: -3.45978385 10.6993445 0.015625     0       6          1   2.260787
#> 15:  2.69891400  4.3422474 0.015625     0       6          1   7.149586
#> 16:  0.22693638  3.5592661 0.015625     0       6          1  23.687048
#> 17:  4.98321170  6.4392581 0.015625     0       6          1  70.431654
#> 18:  3.83668915  5.1233368 0.015625     0       6          1  25.445533
#> 19:  1.91921435  4.2965714 0.015625     0       6          1   8.253985
#> 20:  1.31730517 12.6072246 0.015625     0       6          1  87.237933
#> 21: -3.53431181 13.1160634 0.015625     0       6          1   2.354238
#> 22: -2.88143983  6.5883533 0.015625     0       6          1  18.808011
#> 23:  3.54540281  8.1825696 0.015625     0       6          1  56.505437
#> 24:  1.97678295  8.6294760 0.015625     0       6          1  38.192275
#> 25:  3.60171444  1.0545676 0.015625     0       6          1   1.547077
#> 26:  8.48805220  9.8797229 0.015625     0       6          1 234.550990
#> 27:  7.91482590 11.2687711 0.015625     0       6          1 263.483077
#> 28:  3.03682444  8.2898277 0.015625     0       6          1  47.229649
#> 29: -3.90844985 13.3247963 0.015625     0       6          1   3.488303
#> 30:  0.73556209  7.8579216 0.015625     0       6          1  26.191481
#> 31:  0.86744905  9.8745109 0.015625     0       6          1  43.583834
#> 32: -0.66628111  4.7460994 0.015625     0       6          1  22.967908
#> 33:  5.18101666  3.8206769 0.015625     0       6          1  41.806892
#> 34:  7.18488390  3.1702742 0.015625     0       6          1  65.210173
#> 35:  2.93819110  8.8414430 0.015625     0       6          1  53.191420
#> 36:  2.52206280  1.2606917 0.015625     0       6          1   3.030475
#> 37: -1.82325833 14.3543130 0.015625     0       6          1  36.227167
#> 38: -3.45959696  7.5876434 0.015625     0       6          1  19.253247
#> 39: -4.74755835  3.5615503 0.015625     0       6          1 124.559001
#> 40:  8.83253093  1.3167243 0.015625     0       6          1  50.690126
#> 41:  9.82188175  9.1444453 0.015625     0       6          1 251.111018
#> 42: -1.18581510 13.3656193 0.015625     0       6          1  43.146315
#> 43:  7.86256865  0.5309434 0.015625     0       6          1  36.376538
#> 44: -2.26794056  5.9590588 0.015625     0       6          1  18.340807
#> 45: -2.15750871  5.9553852 0.015625     0       6          1  17.799419
#> 46:  5.57692393 11.2484280 0.015625     0       6          1 190.703613
#> 47:  1.54302106  5.8031953 0.015625     0       6          1  15.044317
#> 48:  6.32800346  0.4458122 0.015625     0       6          1  30.389745
#> 49:  1.05000077  2.8884313 0.015625     0       6          1  16.951403
#> 50: -1.06816340 14.5657044 0.015625     0       6          1  61.282601
#> 51:  6.06893227 11.7865288 0.015625     0       6          1 224.246325
#> 52: -3.58151109 13.4781458 0.015625     0       6          1   3.226475
#> 53: -0.71223042  9.8071291 0.015625     0       6          1  24.332767
#> 54: -2.80668961  4.6393503 0.015625     0       6          1  37.774374
#> 55:  6.91574896  7.5245717 0.015625     0       6          1 140.084939
#> 56:  0.04112513  1.6839857 0.015625     0       6          1  37.661710
#> 57:  9.49147440 10.5718414 0.015625     0       6          1 286.301489
#> 58:  2.83079902 14.4113417 0.015625     0       6          1 161.394655
#> 59:  3.34819993  2.7279875 0.015625     0       6          1   3.533474
#> 60:  5.70589675  7.3720573 0.015625     0       6          1 107.390808
#> 61: -3.73081089 14.2851845 0.015625     0       6          1   5.701322
#> 62:  6.23626257  9.6573124 0.015625     0       6          1 173.025634
#> 63:  6.13473229 11.6050385 0.015625     0       6          1 221.466256
#> 64: -0.87458023  7.2359570 0.015625     0       6          1  16.190237
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-09-21 04:48:07        1  7.46568020  10.8750887          0.015625
#>  2: 2022-09-21 04:48:07        1 -2.70707998   1.2392707          0.015625
#>  3: 2022-09-21 04:48:07        1 -3.35616210  10.9709761          0.015625
#>  4: 2022-09-21 04:48:07        1 -2.38063878   5.1381651          0.015625
#>  5: 2022-09-21 04:48:07        1  2.77480248   1.1478916          0.015625
#>  6: 2022-09-21 04:48:07        1  2.37772667  10.5967598          0.015625
#>  7: 2022-09-21 04:48:07        1 -0.04825100  13.8338799          0.015625
#>  8: 2022-09-21 04:48:07        1  2.04723665   2.1001836          0.015625
#>  9: 2022-09-21 04:48:07        1 -3.65086913   1.5988517          0.015625
#> 10: 2022-09-21 04:48:07        1 -2.10231190   6.2092696          0.015625
#> 11: 2022-09-21 04:48:07        1  1.19150169   1.5119024          0.015625
#> 12: 2022-09-21 04:48:07        1 -0.76046166  12.9848231          0.015625
#> 13: 2022-09-21 04:48:07        1 -3.75637143   8.3074476          0.015625
#> 14: 2022-09-21 04:48:07        1 -3.45978385  10.6993445          0.015625
#> 15: 2022-09-21 04:48:07        1  2.69891400   4.3422474          0.015625
#> 16: 2022-09-21 04:48:07        1  0.22693638   3.5592661          0.015625
#> 17: 2022-09-21 04:48:07        1  4.98321170   6.4392581          0.015625
#> 18: 2022-09-21 04:48:07        1  3.83668915   5.1233368          0.015625
#> 19: 2022-09-21 04:48:07        1  1.91921435   4.2965714          0.015625
#> 20: 2022-09-21 04:48:07        1  1.31730517  12.6072246          0.015625
#> 21: 2022-09-21 04:48:07        1 -3.53431181  13.1160634          0.015625
#> 22: 2022-09-21 04:48:07        1 -2.88143983   6.5883533          0.015625
#> 23: 2022-09-21 04:48:07        1  3.54540281   8.1825696          0.015625
#> 24: 2022-09-21 04:48:07        1  1.97678295   8.6294760          0.015625
#> 25: 2022-09-21 04:48:07        1  3.60171444   1.0545676          0.015625
#> 26: 2022-09-21 04:48:07        1  8.48805220   9.8797229          0.015625
#> 27: 2022-09-21 04:48:07        1  7.91482590  11.2687711          0.015625
#> 28: 2022-09-21 04:48:07        1  3.03682444   8.2898277          0.015625
#> 29: 2022-09-21 04:48:07        1 -3.90844985  13.3247963          0.015625
#> 30: 2022-09-21 04:48:07        1  0.73556209   7.8579216          0.015625
#> 31: 2022-09-21 04:48:07        1  0.86744905   9.8745109          0.015625
#> 32: 2022-09-21 04:48:07        1 -0.66628111   4.7460994          0.015625
#> 33: 2022-09-21 04:48:07        1  5.18101666   3.8206769          0.015625
#> 34: 2022-09-21 04:48:07        1  7.18488390   3.1702742          0.015625
#> 35: 2022-09-21 04:48:07        1  2.93819110   8.8414430          0.015625
#> 36: 2022-09-21 04:48:07        1  2.52206280   1.2606917          0.015625
#> 37: 2022-09-21 04:48:07        1 -1.82325833  14.3543130          0.015625
#> 38: 2022-09-21 04:48:07        1 -3.45959696   7.5876434          0.015625
#> 39: 2022-09-21 04:48:07        1 -4.74755835   3.5615503          0.015625
#> 40: 2022-09-21 04:48:07        1  8.83253093   1.3167243          0.015625
#> 41: 2022-09-21 04:48:07        1  9.82188175   9.1444453          0.015625
#> 42: 2022-09-21 04:48:07        1 -1.18581510  13.3656193          0.015625
#> 43: 2022-09-21 04:48:07        1  7.86256865   0.5309434          0.015625
#> 44: 2022-09-21 04:48:07        1 -2.26794056   5.9590588          0.015625
#> 45: 2022-09-21 04:48:07        1 -2.15750871   5.9553852          0.015625
#> 46: 2022-09-21 04:48:07        1  5.57692393  11.2484280          0.015625
#> 47: 2022-09-21 04:48:07        1  1.54302106   5.8031953          0.015625
#> 48: 2022-09-21 04:48:07        1  6.32800346   0.4458122          0.015625
#> 49: 2022-09-21 04:48:07        1  1.05000077   2.8884313          0.015625
#> 50: 2022-09-21 04:48:07        1 -1.06816340  14.5657044          0.015625
#> 51: 2022-09-21 04:48:07        1  6.06893227  11.7865288          0.015625
#> 52: 2022-09-21 04:48:07        1 -3.58151109  13.4781458          0.015625
#> 53: 2022-09-21 04:48:07        1 -0.71223042   9.8071291          0.015625
#> 54: 2022-09-21 04:48:07        1 -2.80668961   4.6393503          0.015625
#> 55: 2022-09-21 04:48:07        1  6.91574896   7.5245717          0.015625
#> 56: 2022-09-21 04:48:07        1  0.04112513   1.6839857          0.015625
#> 57: 2022-09-21 04:48:07        1  9.49147440  10.5718414          0.015625
#> 58: 2022-09-21 04:48:07        1  2.83079902  14.4113417          0.015625
#> 59: 2022-09-21 04:48:07        1  3.34819993   2.7279875          0.015625
#> 60: 2022-09-21 04:48:07        1  5.70589675   7.3720573          0.015625
#> 61: 2022-09-21 04:48:07        1 -3.73081089  14.2851845          0.015625
#> 62: 2022-09-21 04:48:07        1  6.23626257   9.6573124          0.015625
#> 63: 2022-09-21 04:48:07        1  6.13473229  11.6050385          0.015625
#> 64: 2022-09-21 04:48:07        1 -0.87458023   7.2359570          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity