Optimizer using the Hyperband algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: -3.54707 12.40877 0.015625 <list[3]> 1.318297

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: -3.54707 12.40877 0.015625 <list[3]> 1.318297

# all evaluations
as.data.table(instance$archive)
#>              x1           x2 fidelity stage bracket repetition          y
#>  1:  5.79068800  6.210624995 0.015625     0       6          1  88.950076
#>  2:  0.35166461  1.285725703 0.015625     0       6          1  36.306636
#>  3:  3.89683254 11.731116618 0.015625     0       6          1 134.482732
#>  4:  0.01039146  5.290930637 0.015625     0       6          1  20.081198
#>  5:  9.27810869  6.285872298 0.015625     0       6          1 154.400808
#>  6:  8.66851466  5.008579027 0.015625     0       6          1 113.150356
#>  7:  1.46928011 10.709315561 0.015625     0       6          1  59.712642
#>  8: -4.18083883  2.218507805 0.015625     0       6          1 125.538465
#>  9:  2.43256118  9.522155360 0.015625     0       6          1  54.721487
#> 10:  8.09830409  0.176528759 0.015625     0       6          1  33.168785
#> 11:  8.36796284 10.877852384 0.015625     0       6          1 262.653369
#> 12:  7.91093205  4.449142328 0.015625     0       6          1  92.546637
#> 13:  6.24248750 12.771023379 0.015625     0       6          1 260.094749
#> 14: -4.74552211  3.747523113 0.015625     0       6          1 120.518366
#> 15:  1.49791166  0.624797009 0.015625     0       6          1  20.063995
#> 16: -3.33021235 10.516651380 0.015625     0       6          1   1.832781
#> 17:  1.36084190  0.003812662 0.015625     0       6          1  27.112176
#> 18:  7.37027321  0.669664369 0.015625     0       6          1  36.834844
#> 19:  0.76685583  4.006257125 0.015625     0       6          1  17.540653
#> 20:  7.73684984  3.128307040 0.015625     0       6          1  68.903509
#> 21:  5.82966417  1.798315564 0.015625     0       6          1  34.885053
#> 22:  4.46773857  4.233939996 0.015625     0       6          1  30.055006
#> 23:  1.28082084  3.211729304 0.015625     0       6          1  13.385874
#> 24:  4.78001184  2.024458396 0.015625     0       6          1  19.231150
#> 25:  5.30654712 12.773489436 0.015625     0       6          1 221.392675
#> 26: -1.45196733  1.900389078 0.015625     0       6          1  53.067963
#> 27:  9.06901089 10.298318330 0.015625     0       6          1 263.544469
#> 28:  2.16363482  8.497726939 0.015625     0       6          1  38.244078
#> 29:  7.43600040 13.317970085 0.015625     0       6          1 318.491042
#> 30:  3.40813955 12.195322104 0.015625     0       6          1 127.578750
#> 31: -1.05576421  8.226409941 0.015625     0       6          1  14.991621
#> 32: -1.17574973  0.359591696 0.015625     0       6          1  70.760966
#> 33: -3.76731904 10.480151366 0.015625     0       6          1   6.027841
#> 34:  0.96745583 10.421337846 0.015625     0       6          1  50.640545
#> 35:  7.49075663  0.257748066 0.015625     0       6          1  33.253209
#> 36: -0.22906689 12.341437179 0.015625     0       6          1  55.054936
#> 37:  0.81875979  3.018467937 0.015625     0       6          1  19.446228
#> 38:  8.75869496  1.178394102 0.015625     0       6          1  48.141757
#> 39:  4.39250134 14.904127324 0.015625     0       6          1 241.124853
#> 40: -4.68668486  4.950274795 0.015625     0       6          1  94.102272
#> 41: -0.89471450  1.704197223 0.015625     0       6          1  49.006555
#> 42:  4.05905555 10.817514810 0.015625     0       6          1 120.180439
#> 43:  6.02158354  1.928906076 0.015625     0       6          1  38.615042
#> 44:  1.05448504  5.872688735 0.015625     0       6          1  17.040905
#> 45:  9.54243602 12.814689875 0.015625     0       6          1 369.188172
#> 46:  6.54947900  2.693682552 0.015625     0       6          1  52.887293
#> 47: -2.65992726 14.062447945 0.015625     0       6          1  14.533308
#> 48:  2.91519919  5.903630010 0.015625     0       6          1  18.978552
#> 49: -3.44808561  7.578564071 0.015625     0       6          1  19.119240
#> 50:  9.55701181  9.850029191 0.015625     0       6          1 264.614500
#> 51:  4.54075490  7.181429935 0.015625     0       6          1  68.799925
#> 52: -2.81369086  0.372968089 0.015625     0       6          1 108.002587
#> 53: -3.54707024 12.408769489 0.015625     0       6          1   1.318297
#> 54:  2.35939463  1.789030993 0.015625     0       6          1   3.581763
#> 55: -4.67210484  6.715658898 0.015625     0       6          1  64.245937
#> 56:  5.32511101  0.231763901 0.015625     0       6          1  18.889565
#> 57:  3.42167194 11.404966449 0.015625     0       6          1 110.827788
#> 58: -4.36391177  0.042692634 0.015625     0       6          1 188.653444
#> 59:  7.66161628  4.511395523 0.015625     0       6          1  91.052435
#> 60: -1.25082179  4.474747258 0.015625     0       6          1  25.723101
#> 61:  6.47533758  2.086317310 0.015625     0       6          1  45.465134
#> 62:  6.55122936  9.004056004 0.015625     0       6          1 165.936581
#> 63:  0.22777529 14.946493814 0.015625     0       6          1 105.982061
#> 64:  8.58332346  3.368439674 0.015625     0       6          1  80.408791
#>              x1           x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1  x_domain_x2 x_domain_fidelity
#>  1: 2022-02-06 04:23:46        1  5.79068800  6.210624995          0.015625
#>  2: 2022-02-06 04:23:46        1  0.35166461  1.285725703          0.015625
#>  3: 2022-02-06 04:23:46        1  3.89683254 11.731116618          0.015625
#>  4: 2022-02-06 04:23:46        1  0.01039146  5.290930637          0.015625
#>  5: 2022-02-06 04:23:46        1  9.27810869  6.285872298          0.015625
#>  6: 2022-02-06 04:23:46        1  8.66851466  5.008579027          0.015625
#>  7: 2022-02-06 04:23:46        1  1.46928011 10.709315561          0.015625
#>  8: 2022-02-06 04:23:46        1 -4.18083883  2.218507805          0.015625
#>  9: 2022-02-06 04:23:46        1  2.43256118  9.522155360          0.015625
#> 10: 2022-02-06 04:23:46        1  8.09830409  0.176528759          0.015625
#> 11: 2022-02-06 04:23:46        1  8.36796284 10.877852384          0.015625
#> 12: 2022-02-06 04:23:46        1  7.91093205  4.449142328          0.015625
#> 13: 2022-02-06 04:23:46        1  6.24248750 12.771023379          0.015625
#> 14: 2022-02-06 04:23:46        1 -4.74552211  3.747523113          0.015625
#> 15: 2022-02-06 04:23:46        1  1.49791166  0.624797009          0.015625
#> 16: 2022-02-06 04:23:46        1 -3.33021235 10.516651380          0.015625
#> 17: 2022-02-06 04:23:46        1  1.36084190  0.003812662          0.015625
#> 18: 2022-02-06 04:23:46        1  7.37027321  0.669664369          0.015625
#> 19: 2022-02-06 04:23:46        1  0.76685583  4.006257125          0.015625
#> 20: 2022-02-06 04:23:46        1  7.73684984  3.128307040          0.015625
#> 21: 2022-02-06 04:23:46        1  5.82966417  1.798315564          0.015625
#> 22: 2022-02-06 04:23:46        1  4.46773857  4.233939996          0.015625
#> 23: 2022-02-06 04:23:46        1  1.28082084  3.211729304          0.015625
#> 24: 2022-02-06 04:23:46        1  4.78001184  2.024458396          0.015625
#> 25: 2022-02-06 04:23:46        1  5.30654712 12.773489436          0.015625
#> 26: 2022-02-06 04:23:46        1 -1.45196733  1.900389078          0.015625
#> 27: 2022-02-06 04:23:46        1  9.06901089 10.298318330          0.015625
#> 28: 2022-02-06 04:23:46        1  2.16363482  8.497726939          0.015625
#> 29: 2022-02-06 04:23:46        1  7.43600040 13.317970085          0.015625
#> 30: 2022-02-06 04:23:46        1  3.40813955 12.195322104          0.015625
#> 31: 2022-02-06 04:23:46        1 -1.05576421  8.226409941          0.015625
#> 32: 2022-02-06 04:23:46        1 -1.17574973  0.359591696          0.015625
#> 33: 2022-02-06 04:23:46        1 -3.76731904 10.480151366          0.015625
#> 34: 2022-02-06 04:23:46        1  0.96745583 10.421337846          0.015625
#> 35: 2022-02-06 04:23:46        1  7.49075663  0.257748066          0.015625
#> 36: 2022-02-06 04:23:46        1 -0.22906689 12.341437179          0.015625
#> 37: 2022-02-06 04:23:46        1  0.81875979  3.018467937          0.015625
#> 38: 2022-02-06 04:23:46        1  8.75869496  1.178394102          0.015625
#> 39: 2022-02-06 04:23:46        1  4.39250134 14.904127324          0.015625
#> 40: 2022-02-06 04:23:46        1 -4.68668486  4.950274795          0.015625
#> 41: 2022-02-06 04:23:46        1 -0.89471450  1.704197223          0.015625
#> 42: 2022-02-06 04:23:46        1  4.05905555 10.817514810          0.015625
#> 43: 2022-02-06 04:23:46        1  6.02158354  1.928906076          0.015625
#> 44: 2022-02-06 04:23:46        1  1.05448504  5.872688735          0.015625
#> 45: 2022-02-06 04:23:46        1  9.54243602 12.814689875          0.015625
#> 46: 2022-02-06 04:23:46        1  6.54947900  2.693682552          0.015625
#> 47: 2022-02-06 04:23:46        1 -2.65992726 14.062447945          0.015625
#> 48: 2022-02-06 04:23:46        1  2.91519919  5.903630010          0.015625
#> 49: 2022-02-06 04:23:46        1 -3.44808561  7.578564071          0.015625
#> 50: 2022-02-06 04:23:46        1  9.55701181  9.850029191          0.015625
#> 51: 2022-02-06 04:23:46        1  4.54075490  7.181429935          0.015625
#> 52: 2022-02-06 04:23:46        1 -2.81369086  0.372968089          0.015625
#> 53: 2022-02-06 04:23:46        1 -3.54707024 12.408769489          0.015625
#> 54: 2022-02-06 04:23:46        1  2.35939463  1.789030993          0.015625
#> 55: 2022-02-06 04:23:46        1 -4.67210484  6.715658898          0.015625
#> 56: 2022-02-06 04:23:46        1  5.32511101  0.231763901          0.015625
#> 57: 2022-02-06 04:23:46        1  3.42167194 11.404966449          0.015625
#> 58: 2022-02-06 04:23:46        1 -4.36391177  0.042692634          0.015625
#> 59: 2022-02-06 04:23:46        1  7.66161628  4.511395523          0.015625
#> 60: 2022-02-06 04:23:46        1 -1.25082179  4.474747258          0.015625
#> 61: 2022-02-06 04:23:46        1  6.47533758  2.086317310          0.015625
#> 62: 2022-02-06 04:23:46        1  6.55122936  9.004056004          0.015625
#> 63: 2022-02-06 04:23:46        1  0.22777529 14.946493814          0.015625
#> 64: 2022-02-06 04:23:46        1  8.58332346  3.368439674          0.015625
#>               timestamp batch_nr x_domain_x1  x_domain_x2 x_domain_fidelity