Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain         y
#> 1: 3.350841 0.9368847 0.015625 <list[3]> 0.6130082

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain         y
#> 1: 3.350841 0.9368847 0.015625 <list[3]> 0.6130082

# all evaluations
as.data.table(instance$archive)
#>             x1          x2 fidelity stage bracket repetition           y
#>  1: -0.3371085  1.73739510 0.015625     0       6          1  42.1268539
#>  2:  3.3145398  2.62262219 0.015625     0       6          1   2.9749955
#>  3:  2.0272586 13.70940026 0.015625     0       6          1 122.6133291
#>  4:  9.5467574  4.75141253 0.015625     0       6          1 124.6412240
#>  5: -0.2420394  9.15788583 0.015625     0       6          1  26.9999260
#>  6:  4.8924753  1.67914493 0.015625     0       6          1  19.1716376
#>  7:  8.3625587  3.64039272 0.015625     0       6          1  82.7574700
#>  8:  1.5875011  7.74804704 0.015625     0       6          1  27.4556420
#>  9:  7.1345248 14.83927223 0.015625     0       6          1 363.3729407
#> 10: -2.9495589 12.50399224 0.015625     0       6          1   2.9525498
#> 11:  6.6110071  0.63295175 0.015625     0       6          1  33.6136095
#> 12:  1.6836983  0.05213711 0.015625     0       6          1  20.1764892
#> 13:  6.2361121 12.88154183 0.015625     0       6          1 263.2916497
#> 14:  3.9073881 11.94435236 0.015625     0       6          1 139.8215389
#> 15:  8.2008584  0.37048786 0.015625     0       6          1  35.4084307
#> 16:  8.6116583 14.83808471 0.015625     0       6          1 414.0204937
#> 17:  9.5791792  9.24630466 0.015625     0       6          1 246.0821888
#> 18:  7.9536918  7.39217461 0.015625     0       6          1 155.5939495
#> 19:  6.4273046 11.69683228 0.015625     0       6          1 234.3020379
#> 20:  2.2015993  0.35859582 0.015625     0       6          1   9.5647380
#> 21:  6.3482668 14.10127098 0.015625     0       6          1 307.4178420
#> 22:  6.7137813  2.27099248 0.015625     0       6          1  49.7548513
#> 23:  4.4440970  9.32556406 0.015625     0       6          1 103.3243915
#> 24: -1.8001023  4.49209732 0.015625     0       6          1  27.8205553
#> 25:  1.3995644 10.37441981 0.015625     0       6          1  54.4295998
#> 26:  1.0025984 12.89494431 0.015625     0       6          1  86.7339305
#> 27:  4.1796963  7.49563951 0.015625     0       6          1  63.0460745
#> 28:  1.7593455 14.82014895 0.015625     0       6          1 141.0273185
#> 29: -4.5211597 12.62205781 0.015625     0       6          1   9.6199799
#> 30: -2.8408717  8.11421959 0.015625     0       6          1   7.8794890
#> 31: -4.9173085  1.23060891 0.015625     0       6          1 189.8825960
#> 32:  7.9068374  5.72134406 0.015625     0       6          1 117.3040235
#> 33:  9.3742257  8.97599268 0.015625     0       6          1 231.2559025
#> 34: -3.6470188  2.51346198 0.015625     0       6          1  95.6866504
#> 35: -1.8122965 12.13505678 0.015625     0       6          1  17.6242433
#> 36:  0.5760318  8.57772167 0.015625     0       6          1  30.1929905
#> 37: -4.7007677  8.96430324 0.015625     0       6          1  36.8933164
#> 38: -4.9585807  8.90869065 0.015625     0       6          1  45.2777483
#> 39:  3.5583084  1.31170838 0.015625     0       6          1   1.5625645
#> 40: -2.1769902  1.08129675 0.015625     0       6          1  77.2766736
#> 41:  1.7949153 12.03883529 0.015625     0       6          1  85.2438991
#> 42: -0.6244570 11.92362214 0.015625     0       6          1  41.9745518
#> 43: -1.4190410  2.56709563 0.015625     0       6          1  44.5519637
#> 44: -2.7160965  4.18234153 0.015625     0       6          1  41.7964073
#> 45:  0.3077906  5.04535230 0.015625     0       6          1  19.3696038
#> 46:  8.9462517  7.20931703 0.015625     0       6          1 170.1352180
#> 47:  9.3364265 10.87986430 0.015625     0       6          1 291.4471028
#> 48:  6.8615669  4.77945789 0.015625     0       6          1  86.1420537
#> 49:  2.0553867  6.80736529 0.015625     0       6          1  21.1192889
#> 50: -4.3078378  3.47990438 0.015625     0       6          1 105.1597565
#> 51:  3.3508412  0.93688473 0.015625     0       6          1   0.6130082
#> 52: -0.0026195  5.32487024 0.015625     0       6          1  20.0635269
#> 53: -3.4145291 10.69969181 0.015625     0       6          1   1.9483658
#> 54: -4.3182206  9.25628379 0.015625     0       6          1  23.8661390
#> 55:  2.7085897  7.02471215 0.015625     0       6          1  27.3960904
#> 56:  3.6746204 12.42335472 0.015625     0       6          1 142.3071013
#> 57:  3.7267515  0.83886682 0.015625     0       6          1   2.1131907
#> 58:  5.8138687  5.67823045 0.015625     0       6          1  80.8477003
#> 59:  4.7122464  0.12434219 0.015625     0       6          1  10.8848076
#> 60: -0.5176163  3.09652786 0.015625     0       6          1  32.2983711
#> 61:  4.4704603 10.27110045 0.015625     0       6          1 123.7264947
#> 62: -0.6476997  7.33540638 0.015625     0       6          1  17.7424994
#> 63:  4.6167570 13.16717007 0.015625     0       6          1 201.1719623
#> 64:  7.3227518 14.28384470 0.015625     0       6          1 349.3748919
#>             x1          x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-09-16 04:48:20        1  -0.3371085  1.73739510          0.015625
#>  2: 2022-09-16 04:48:20        1   3.3145398  2.62262219          0.015625
#>  3: 2022-09-16 04:48:20        1   2.0272586 13.70940026          0.015625
#>  4: 2022-09-16 04:48:20        1   9.5467574  4.75141253          0.015625
#>  5: 2022-09-16 04:48:20        1  -0.2420394  9.15788583          0.015625
#>  6: 2022-09-16 04:48:20        1   4.8924753  1.67914493          0.015625
#>  7: 2022-09-16 04:48:20        1   8.3625587  3.64039272          0.015625
#>  8: 2022-09-16 04:48:20        1   1.5875011  7.74804704          0.015625
#>  9: 2022-09-16 04:48:20        1   7.1345248 14.83927223          0.015625
#> 10: 2022-09-16 04:48:20        1  -2.9495589 12.50399224          0.015625
#> 11: 2022-09-16 04:48:20        1   6.6110071  0.63295175          0.015625
#> 12: 2022-09-16 04:48:20        1   1.6836983  0.05213711          0.015625
#> 13: 2022-09-16 04:48:20        1   6.2361121 12.88154183          0.015625
#> 14: 2022-09-16 04:48:20        1   3.9073881 11.94435236          0.015625
#> 15: 2022-09-16 04:48:20        1   8.2008584  0.37048786          0.015625
#> 16: 2022-09-16 04:48:20        1   8.6116583 14.83808471          0.015625
#> 17: 2022-09-16 04:48:20        1   9.5791792  9.24630466          0.015625
#> 18: 2022-09-16 04:48:20        1   7.9536918  7.39217461          0.015625
#> 19: 2022-09-16 04:48:20        1   6.4273046 11.69683228          0.015625
#> 20: 2022-09-16 04:48:20        1   2.2015993  0.35859582          0.015625
#> 21: 2022-09-16 04:48:20        1   6.3482668 14.10127098          0.015625
#> 22: 2022-09-16 04:48:20        1   6.7137813  2.27099248          0.015625
#> 23: 2022-09-16 04:48:20        1   4.4440970  9.32556406          0.015625
#> 24: 2022-09-16 04:48:20        1  -1.8001023  4.49209732          0.015625
#> 25: 2022-09-16 04:48:20        1   1.3995644 10.37441981          0.015625
#> 26: 2022-09-16 04:48:20        1   1.0025984 12.89494431          0.015625
#> 27: 2022-09-16 04:48:20        1   4.1796963  7.49563951          0.015625
#> 28: 2022-09-16 04:48:20        1   1.7593455 14.82014895          0.015625
#> 29: 2022-09-16 04:48:20        1  -4.5211597 12.62205781          0.015625
#> 30: 2022-09-16 04:48:20        1  -2.8408717  8.11421959          0.015625
#> 31: 2022-09-16 04:48:20        1  -4.9173085  1.23060891          0.015625
#> 32: 2022-09-16 04:48:20        1   7.9068374  5.72134406          0.015625
#> 33: 2022-09-16 04:48:20        1   9.3742257  8.97599268          0.015625
#> 34: 2022-09-16 04:48:20        1  -3.6470188  2.51346198          0.015625
#> 35: 2022-09-16 04:48:20        1  -1.8122965 12.13505678          0.015625
#> 36: 2022-09-16 04:48:20        1   0.5760318  8.57772167          0.015625
#> 37: 2022-09-16 04:48:20        1  -4.7007677  8.96430324          0.015625
#> 38: 2022-09-16 04:48:20        1  -4.9585807  8.90869065          0.015625
#> 39: 2022-09-16 04:48:20        1   3.5583084  1.31170838          0.015625
#> 40: 2022-09-16 04:48:20        1  -2.1769902  1.08129675          0.015625
#> 41: 2022-09-16 04:48:20        1   1.7949153 12.03883529          0.015625
#> 42: 2022-09-16 04:48:20        1  -0.6244570 11.92362214          0.015625
#> 43: 2022-09-16 04:48:20        1  -1.4190410  2.56709563          0.015625
#> 44: 2022-09-16 04:48:20        1  -2.7160965  4.18234153          0.015625
#> 45: 2022-09-16 04:48:20        1   0.3077906  5.04535230          0.015625
#> 46: 2022-09-16 04:48:20        1   8.9462517  7.20931703          0.015625
#> 47: 2022-09-16 04:48:20        1   9.3364265 10.87986430          0.015625
#> 48: 2022-09-16 04:48:20        1   6.8615669  4.77945789          0.015625
#> 49: 2022-09-16 04:48:20        1   2.0553867  6.80736529          0.015625
#> 50: 2022-09-16 04:48:20        1  -4.3078378  3.47990438          0.015625
#> 51: 2022-09-16 04:48:20        1   3.3508412  0.93688473          0.015625
#> 52: 2022-09-16 04:48:20        1  -0.0026195  5.32487024          0.015625
#> 53: 2022-09-16 04:48:20        1  -3.4145291 10.69969181          0.015625
#> 54: 2022-09-16 04:48:20        1  -4.3182206  9.25628379          0.015625
#> 55: 2022-09-16 04:48:20        1   2.7085897  7.02471215          0.015625
#> 56: 2022-09-16 04:48:20        1   3.6746204 12.42335472          0.015625
#> 57: 2022-09-16 04:48:20        1   3.7267515  0.83886682          0.015625
#> 58: 2022-09-16 04:48:20        1   5.8138687  5.67823045          0.015625
#> 59: 2022-09-16 04:48:20        1   4.7122464  0.12434219          0.015625
#> 60: 2022-09-16 04:48:20        1  -0.5176163  3.09652786          0.015625
#> 61: 2022-09-16 04:48:20        1   4.4704603 10.27110045          0.015625
#> 62: 2022-09-16 04:48:20        1  -0.6476997  7.33540638          0.015625
#> 63: 2022-09-16 04:48:20        1   4.6167570 13.16717007          0.015625
#> 64: 2022-09-16 04:48:20        1   7.3227518 14.28384470          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity