Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.474387 11.50428 0.015625 <list[3]> 1.081966

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.474387 11.50428 0.015625 <list[3]> 1.081966

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1:  6.8967377  2.7878709 0.015625     0       6          1  57.564563
#>  2: -4.3585789  6.6249290 0.015625     0       6          1  54.228924
#>  3:  5.8941509  0.5100592 0.015625     0       6          1  26.852291
#>  4:  3.8264013 10.3262424 0.015625     0       6          1 101.883369
#>  5:  4.5034126  6.8528294 0.015625     0       6          1  62.718559
#>  6:  1.6678360 14.8866678 0.015625     0       6          1 140.300057
#>  7:  9.3064341 14.2042010 0.015625     0       6          1 414.704010
#>  8:  6.8485071  8.3006556 0.015625     0       6          1 156.365692
#>  9:  4.0757356  5.3395490 0.015625     0       6          1  32.546287
#> 10:  1.1539661  6.0549229 0.015625     0       6          1  17.312172
#> 11:  8.9476712  9.8853613 0.015625     0       6          1 246.843344
#> 12: -4.7511529 10.0017401 0.015625     0       6          1  28.468814
#> 13:  0.3695095  4.4507351 0.015625     0       6          1  19.885954
#> 14:  4.9901993  4.2822216 0.015625     0       6          1  42.430852
#> 15: -4.3299494  2.8449947 0.015625     0       6          1 119.260363
#> 16: -3.1023620  0.8806206 0.015625     0       6          1 107.587215
#> 17:  5.3942163  1.2796762 0.015625     0       6          1  24.873313
#> 18: -4.0443690  6.0225444 0.015625     0       6          1  51.899726
#> 19:  0.6357801 13.8010293 0.015625     0       6          1  95.174326
#> 20:  9.6830347  3.4056823 0.015625     0       6          1  99.397410
#> 21:  5.5478584  8.0530911 0.015625     0       6          1 115.853638
#> 22:  8.6932372  3.3491329 0.015625     0       6          1  81.375938
#> 23:  5.8070713  5.4121988 0.015625     0       6          1  76.561820
#> 24: -3.4743870 11.5042808 0.015625     0       6          1   1.081966
#> 25:  3.9043889 10.2636829 0.015625     0       6          1 103.238563
#> 26:  2.4102638 11.0979883 0.015625     0       6          1  79.510672
#> 27:  8.9345821 12.0325255 0.015625     0       6          1 318.294599
#> 28:  2.1609440  7.6564736 0.015625     0       6          1  29.180312
#> 29:  9.3003185 11.5066214 0.015625     0       6          1 311.961020
#> 30:  2.3688733  6.1475463 0.015625     0       6          1  17.151166
#> 31:  1.8674318  9.4981132 0.015625     0       6          1  47.681019
#> 32:  8.7626039  8.4447140 0.015625     0       6          1 199.267469
#> 33:  3.1323204  0.1790448 0.015625     0       6          1   1.691944
#> 34:  3.3313860 12.9927436 0.015625     0       6          1 143.458259
#> 35:  5.2412701  7.3323432 0.015625     0       6          1  92.803842
#> 36: -2.1254883  9.9366560 0.015625     0       6          1   5.114912
#> 37: -2.5430589 12.8014548 0.015625     0       6          1   8.596168
#> 38:  2.8176557  8.3381204 0.015625     0       6          1  44.173335
#> 39:  8.8161160  2.2157808 0.015625     0       6          1  63.859264
#> 40:  0.8999302 13.5999483 0.015625     0       6          1  97.101296
#> 41: -4.1781735  4.9519521 0.015625     0       6          1  72.919110
#> 42:  5.0362126 11.9559593 0.015625     0       6          1 187.070761
#> 43: -2.7644742  2.9030688 0.015625     0       6          1  60.851925
#> 44:  9.3964162  8.7376364 0.015625     0       6          1 224.735124
#> 45: -2.5345751 14.4329080 0.015625     0       6          1  19.765756
#> 46:  5.6393863 14.4824522 0.015625     0       6          1 289.269617
#> 47:  6.3870329  6.7311618 0.015625     0       6          1 112.521266
#> 48:  7.5256034  6.0444024 0.015625     0       6          1 118.783886
#> 49:  7.6397664 14.6572151 0.015625     0       6          1 373.866298
#> 50:  5.3954733 12.1578125 0.015625     0       6          1 207.880325
#> 51:  5.6545382  7.1856375 0.015625     0       6          1 102.443510
#> 52:  3.7903913  2.8667914 0.015625     0       6          1   8.388944
#> 53: -1.2374365  9.8362257 0.015625     0       6          1  16.453320
#> 54: -1.8111654  4.3650007 0.015625     0       6          1  29.043901
#> 55: -4.4945346 10.8123896 0.015625     0       6          1  16.698089
#> 56:  3.5490280 14.5642326 0.015625     0       6          1 192.325850
#> 57:  7.7765698  4.2929658 0.015625     0       6          1  88.364569
#> 58:  2.2209581 12.3213662 0.015625     0       6          1  98.364363
#> 59:  9.4848660 12.6965154 0.015625     0       6          1 362.406369
#> 60:  1.7711554 10.9529761 0.015625     0       6          1  67.000807
#> 61: -4.8143366  5.5741473 0.015625     0       6          1  88.430532
#> 62: -3.0142200  4.4169080 0.015625     0       6          1  44.827608
#> 63: -3.9906687  3.8303774 0.015625     0       6          1  84.847533
#> 64:  2.2075126 10.1598186 0.015625     0       6          1  60.891784
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-31 04:34:45        1   6.8967377   2.7878709          0.015625
#>  2: 2022-10-31 04:34:45        1  -4.3585789   6.6249290          0.015625
#>  3: 2022-10-31 04:34:45        1   5.8941509   0.5100592          0.015625
#>  4: 2022-10-31 04:34:45        1   3.8264013  10.3262424          0.015625
#>  5: 2022-10-31 04:34:45        1   4.5034126   6.8528294          0.015625
#>  6: 2022-10-31 04:34:45        1   1.6678360  14.8866678          0.015625
#>  7: 2022-10-31 04:34:45        1   9.3064341  14.2042010          0.015625
#>  8: 2022-10-31 04:34:45        1   6.8485071   8.3006556          0.015625
#>  9: 2022-10-31 04:34:45        1   4.0757356   5.3395490          0.015625
#> 10: 2022-10-31 04:34:45        1   1.1539661   6.0549229          0.015625
#> 11: 2022-10-31 04:34:45        1   8.9476712   9.8853613          0.015625
#> 12: 2022-10-31 04:34:45        1  -4.7511529  10.0017401          0.015625
#> 13: 2022-10-31 04:34:45        1   0.3695095   4.4507351          0.015625
#> 14: 2022-10-31 04:34:45        1   4.9901993   4.2822216          0.015625
#> 15: 2022-10-31 04:34:45        1  -4.3299494   2.8449947          0.015625
#> 16: 2022-10-31 04:34:45        1  -3.1023620   0.8806206          0.015625
#> 17: 2022-10-31 04:34:45        1   5.3942163   1.2796762          0.015625
#> 18: 2022-10-31 04:34:45        1  -4.0443690   6.0225444          0.015625
#> 19: 2022-10-31 04:34:45        1   0.6357801  13.8010293          0.015625
#> 20: 2022-10-31 04:34:45        1   9.6830347   3.4056823          0.015625
#> 21: 2022-10-31 04:34:45        1   5.5478584   8.0530911          0.015625
#> 22: 2022-10-31 04:34:45        1   8.6932372   3.3491329          0.015625
#> 23: 2022-10-31 04:34:45        1   5.8070713   5.4121988          0.015625
#> 24: 2022-10-31 04:34:45        1  -3.4743870  11.5042808          0.015625
#> 25: 2022-10-31 04:34:45        1   3.9043889  10.2636829          0.015625
#> 26: 2022-10-31 04:34:45        1   2.4102638  11.0979883          0.015625
#> 27: 2022-10-31 04:34:45        1   8.9345821  12.0325255          0.015625
#> 28: 2022-10-31 04:34:45        1   2.1609440   7.6564736          0.015625
#> 29: 2022-10-31 04:34:45        1   9.3003185  11.5066214          0.015625
#> 30: 2022-10-31 04:34:45        1   2.3688733   6.1475463          0.015625
#> 31: 2022-10-31 04:34:45        1   1.8674318   9.4981132          0.015625
#> 32: 2022-10-31 04:34:45        1   8.7626039   8.4447140          0.015625
#> 33: 2022-10-31 04:34:45        1   3.1323204   0.1790448          0.015625
#> 34: 2022-10-31 04:34:45        1   3.3313860  12.9927436          0.015625
#> 35: 2022-10-31 04:34:45        1   5.2412701   7.3323432          0.015625
#> 36: 2022-10-31 04:34:45        1  -2.1254883   9.9366560          0.015625
#> 37: 2022-10-31 04:34:45        1  -2.5430589  12.8014548          0.015625
#> 38: 2022-10-31 04:34:45        1   2.8176557   8.3381204          0.015625
#> 39: 2022-10-31 04:34:45        1   8.8161160   2.2157808          0.015625
#> 40: 2022-10-31 04:34:45        1   0.8999302  13.5999483          0.015625
#> 41: 2022-10-31 04:34:45        1  -4.1781735   4.9519521          0.015625
#> 42: 2022-10-31 04:34:45        1   5.0362126  11.9559593          0.015625
#> 43: 2022-10-31 04:34:45        1  -2.7644742   2.9030688          0.015625
#> 44: 2022-10-31 04:34:45        1   9.3964162   8.7376364          0.015625
#> 45: 2022-10-31 04:34:45        1  -2.5345751  14.4329080          0.015625
#> 46: 2022-10-31 04:34:45        1   5.6393863  14.4824522          0.015625
#> 47: 2022-10-31 04:34:45        1   6.3870329   6.7311618          0.015625
#> 48: 2022-10-31 04:34:45        1   7.5256034   6.0444024          0.015625
#> 49: 2022-10-31 04:34:45        1   7.6397664  14.6572151          0.015625
#> 50: 2022-10-31 04:34:45        1   5.3954733  12.1578125          0.015625
#> 51: 2022-10-31 04:34:45        1   5.6545382   7.1856375          0.015625
#> 52: 2022-10-31 04:34:45        1   3.7903913   2.8667914          0.015625
#> 53: 2022-10-31 04:34:45        1  -1.2374365   9.8362257          0.015625
#> 54: 2022-10-31 04:34:45        1  -1.8111654   4.3650007          0.015625
#> 55: 2022-10-31 04:34:45        1  -4.4945346  10.8123896          0.015625
#> 56: 2022-10-31 04:34:45        1   3.5490280  14.5642326          0.015625
#> 57: 2022-10-31 04:34:45        1   7.7765698   4.2929658          0.015625
#> 58: 2022-10-31 04:34:45        1   2.2209581  12.3213662          0.015625
#> 59: 2022-10-31 04:34:45        1   9.4848660  12.6965154          0.015625
#> 60: 2022-10-31 04:34:45        1   1.7711554  10.9529761          0.015625
#> 61: 2022-10-31 04:34:45        1  -4.8143366   5.5741473          0.015625
#> 62: 2022-10-31 04:34:45        1  -3.0142200   4.4169080          0.015625
#> 63: 2022-10-31 04:34:45        1  -3.9906687   3.8303774          0.015625
#> 64: 2022-10-31 04:34:45        1   2.2075126  10.1598186          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity