Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.038104 10.81987 0.015625 <list[3]> 0.5387953

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain         y
#> 1: -3.038104 10.81987 0.015625 <list[3]> 0.5387953

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -3.9199009  4.1569733 0.015625     0       6          1  76.3364735
#>  2: -3.0059236 10.0980784 0.015625     0       6          1   1.4150591
#>  3:  0.7638661  7.6728540 0.015625     0       6          1  25.1749278
#>  4:  5.6706946  0.3515372 0.015625     0       6          1  23.5591867
#>  5: -4.4768881 14.7354080 0.015625     0       6          1   8.7475179
#>  6:  8.7066174  5.9360250 0.015625     0       6          1 134.1527150
#>  7:  4.1271732  7.4429557 0.015625     0       6          1  60.7635143
#>  8: -1.7647694  2.9309561 0.015625     0       6          1  43.8320532
#>  9:  8.9512684  1.0485723 0.015625     0       6          1  48.1216914
#> 10:  5.8930907  5.2248724 0.015625     0       6          1  75.6753396
#> 11:  0.2911545  6.6800843 0.015625     0       6          1  20.4995616
#> 12:  3.1363513  4.6493067 0.015625     0       6          1  11.5437063
#> 13:  7.8495553  0.4069383 0.015625     0       6          1  35.0965377
#> 14: -3.5517740  3.2938590 0.015625     0       6          1  77.7016323
#> 15:  4.6152708  0.3065038 0.015625     0       6          1  10.0629381
#> 16:  9.4703441 11.0968777 0.015625     0       6          1 303.5778877
#> 17: -3.0381042 10.8198660 0.015625     0       6          1   0.5387953
#> 18:  9.4742032 10.8057271 0.015625     0       6          1 293.6588689
#> 19:  8.5848470 14.5704720 0.015625     0       6          1 402.2970283
#> 20:  1.6793370 11.4278093 0.015625     0       6          1  73.1815414
#> 21:  8.0782968  2.9144998 0.015625     0       6          1  68.1592362
#> 22:  0.8547054  5.3049184 0.015625     0       6          1  16.7163623
#> 23:  7.1616642  5.1680134 0.015625     0       6          1  96.9342328
#> 24:  1.9081398 11.7146496 0.015625     0       6          1  81.4645438
#> 25: -2.8628972  3.3903185 0.015625     0       6          1  55.7970640
#> 26: -1.0041278 14.8493765 0.015625     0       6          1  67.2867442
#> 27:  5.3302587  6.8022405 0.015625     0       6          1  86.3248737
#> 28:  2.9732219 13.4708670 0.015625     0       6          1 142.8846158
#> 29:  5.3900827 14.6509631 0.015625     0       6          1 282.8936848
#> 30:  9.5165969  8.5292278 0.015625     0       6          1 222.1724832
#> 31: -2.9223112  8.4892897 0.015625     0       6          1   6.5049993
#> 32:  7.0911782  4.8034513 0.015625     0       6          1  89.6226388
#> 33:  5.1243829  6.7549028 0.015625     0       6          1  79.5071617
#> 34: -3.8940719  5.3422007 0.015625     0       6          1  56.5970274
#> 35: -3.2222620 10.1567670 0.015625     0       6          1   2.0954524
#> 36:  8.9041416 12.4698849 0.015625     0       6          1 333.0385939
#> 37: -1.8690100 13.5807177 0.015625     0       6          1  27.4169874
#> 38:  7.7689195  2.6051447 0.015625     0       6          1  61.4247053
#> 39:  2.4130773  9.6162703 0.015625     0       6          1  55.8011024
#> 40:  9.7908364 13.5585577 0.015625     0       6          1 408.8209980
#> 41:  1.8092072  7.8372236 0.015625     0       6          1  29.0400564
#> 42: -1.2014962  1.6409461 0.015625     0       6          1  53.3538276
#> 43:  2.1889329  1.6228667 0.015625     0       6          1   5.5183809
#> 44:  2.0803012 12.3277521 0.015625     0       6          1  95.6728931
#> 45: -1.5170085  0.9897558 0.015625     0       6          1  66.6971997
#> 46: -3.1509050 14.9204202 0.015625     0       6          1  13.3607279
#> 47:  4.1691359  9.4141915 0.015625     0       6          1  95.5743737
#> 48: -0.9748325  5.1589392 0.015625     0       6          1  21.2547181
#> 49:  2.9292660  4.4772581 0.015625     0       6          1   8.8820312
#> 50: -1.3149721  3.3095864 0.015625     0       6          1  35.8207332
#> 51:  1.0147393  7.8365443 0.015625     0       6          1  26.7640535
#> 52:  2.1267458  4.2644648 0.015625     0       6          1   7.2132362
#> 53:  7.4447281  9.9623857 0.015625     0       6          1 212.8260216
#> 54: -3.5408898 13.6729953 0.015625     0       6          1   3.8823174
#> 55:  8.7029802  4.5460340 0.015625     0       6          1 104.1655808
#> 56:  5.7880297 10.3631441 0.015625     0       6          1 175.8262508
#> 57:  3.2089635 12.6770440 0.015625     0       6          1 131.9267198
#> 58: -2.4719106  5.0377733 0.015625     0       6          1  28.3215548
#> 59:  1.0677913 13.9057780 0.015625     0       6          1 106.2168304
#> 60:  7.4660345 12.4785628 0.015625     0       6          1 290.7628418
#> 61:  1.1204291  7.5407100 0.015625     0       6          1  24.9731542
#> 62: -2.9793451  0.4741533 0.015625     0       6          1 111.6271340
#> 63:  6.1481760  1.5513505 0.015625     0       6          1  36.9389924
#> 64:  0.7821292 14.4127909 0.015625     0       6          1 109.7179469
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-07 04:32:46        1  -3.9199009   4.1569733          0.015625
#>  2: 2022-10-07 04:32:46        1  -3.0059236  10.0980784          0.015625
#>  3: 2022-10-07 04:32:46        1   0.7638661   7.6728540          0.015625
#>  4: 2022-10-07 04:32:46        1   5.6706946   0.3515372          0.015625
#>  5: 2022-10-07 04:32:46        1  -4.4768881  14.7354080          0.015625
#>  6: 2022-10-07 04:32:46        1   8.7066174   5.9360250          0.015625
#>  7: 2022-10-07 04:32:46        1   4.1271732   7.4429557          0.015625
#>  8: 2022-10-07 04:32:46        1  -1.7647694   2.9309561          0.015625
#>  9: 2022-10-07 04:32:46        1   8.9512684   1.0485723          0.015625
#> 10: 2022-10-07 04:32:46        1   5.8930907   5.2248724          0.015625
#> 11: 2022-10-07 04:32:46        1   0.2911545   6.6800843          0.015625
#> 12: 2022-10-07 04:32:46        1   3.1363513   4.6493067          0.015625
#> 13: 2022-10-07 04:32:46        1   7.8495553   0.4069383          0.015625
#> 14: 2022-10-07 04:32:46        1  -3.5517740   3.2938590          0.015625
#> 15: 2022-10-07 04:32:46        1   4.6152708   0.3065038          0.015625
#> 16: 2022-10-07 04:32:46        1   9.4703441  11.0968777          0.015625
#> 17: 2022-10-07 04:32:46        1  -3.0381042  10.8198660          0.015625
#> 18: 2022-10-07 04:32:46        1   9.4742032  10.8057271          0.015625
#> 19: 2022-10-07 04:32:46        1   8.5848470  14.5704720          0.015625
#> 20: 2022-10-07 04:32:46        1   1.6793370  11.4278093          0.015625
#> 21: 2022-10-07 04:32:46        1   8.0782968   2.9144998          0.015625
#> 22: 2022-10-07 04:32:46        1   0.8547054   5.3049184          0.015625
#> 23: 2022-10-07 04:32:46        1   7.1616642   5.1680134          0.015625
#> 24: 2022-10-07 04:32:46        1   1.9081398  11.7146496          0.015625
#> 25: 2022-10-07 04:32:46        1  -2.8628972   3.3903185          0.015625
#> 26: 2022-10-07 04:32:46        1  -1.0041278  14.8493765          0.015625
#> 27: 2022-10-07 04:32:46        1   5.3302587   6.8022405          0.015625
#> 28: 2022-10-07 04:32:46        1   2.9732219  13.4708670          0.015625
#> 29: 2022-10-07 04:32:46        1   5.3900827  14.6509631          0.015625
#> 30: 2022-10-07 04:32:46        1   9.5165969   8.5292278          0.015625
#> 31: 2022-10-07 04:32:46        1  -2.9223112   8.4892897          0.015625
#> 32: 2022-10-07 04:32:46        1   7.0911782   4.8034513          0.015625
#> 33: 2022-10-07 04:32:46        1   5.1243829   6.7549028          0.015625
#> 34: 2022-10-07 04:32:46        1  -3.8940719   5.3422007          0.015625
#> 35: 2022-10-07 04:32:46        1  -3.2222620  10.1567670          0.015625
#> 36: 2022-10-07 04:32:46        1   8.9041416  12.4698849          0.015625
#> 37: 2022-10-07 04:32:46        1  -1.8690100  13.5807177          0.015625
#> 38: 2022-10-07 04:32:46        1   7.7689195   2.6051447          0.015625
#> 39: 2022-10-07 04:32:46        1   2.4130773   9.6162703          0.015625
#> 40: 2022-10-07 04:32:46        1   9.7908364  13.5585577          0.015625
#> 41: 2022-10-07 04:32:46        1   1.8092072   7.8372236          0.015625
#> 42: 2022-10-07 04:32:46        1  -1.2014962   1.6409461          0.015625
#> 43: 2022-10-07 04:32:46        1   2.1889329   1.6228667          0.015625
#> 44: 2022-10-07 04:32:46        1   2.0803012  12.3277521          0.015625
#> 45: 2022-10-07 04:32:46        1  -1.5170085   0.9897558          0.015625
#> 46: 2022-10-07 04:32:46        1  -3.1509050  14.9204202          0.015625
#> 47: 2022-10-07 04:32:46        1   4.1691359   9.4141915          0.015625
#> 48: 2022-10-07 04:32:46        1  -0.9748325   5.1589392          0.015625
#> 49: 2022-10-07 04:32:46        1   2.9292660   4.4772581          0.015625
#> 50: 2022-10-07 04:32:46        1  -1.3149721   3.3095864          0.015625
#> 51: 2022-10-07 04:32:46        1   1.0147393   7.8365443          0.015625
#> 52: 2022-10-07 04:32:46        1   2.1267458   4.2644648          0.015625
#> 53: 2022-10-07 04:32:46        1   7.4447281   9.9623857          0.015625
#> 54: 2022-10-07 04:32:46        1  -3.5408898  13.6729953          0.015625
#> 55: 2022-10-07 04:32:46        1   8.7029802   4.5460340          0.015625
#> 56: 2022-10-07 04:32:46        1   5.7880297  10.3631441          0.015625
#> 57: 2022-10-07 04:32:46        1   3.2089635  12.6770440          0.015625
#> 58: 2022-10-07 04:32:46        1  -2.4719106   5.0377733          0.015625
#> 59: 2022-10-07 04:32:46        1   1.0677913  13.9057780          0.015625
#> 60: 2022-10-07 04:32:46        1   7.4660345  12.4785628          0.015625
#> 61: 2022-10-07 04:32:46        1   1.1204291   7.5407100          0.015625
#> 62: 2022-10-07 04:32:46        1  -2.9793451   0.4741533          0.015625
#> 63: 2022-10-07 04:32:46        1   6.1481760   1.5513505          0.015625
#> 64: 2022-10-07 04:32:46        1   0.7821292  14.4127909          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity