Optimizer using the Hyperband algorithm

OptimizerHyperband class that implements hyperband optimization (HBX). HBX repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HBX is called a bracket. HBX considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repeats

integer(1)
If 1 (default), optimization is stopped once all stages are evaluated. Otherwise, optimization is stopped after repeats runs of SHA. The bbotk::Terminator might stop the optimization before all repeats are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

stage (integer(1))
The stages of each bracket. Starts counting at 0.

bracket (integer(1))
The bracket index. Counts down to 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.109547 1.943446 0.015625 <list[3]> 0.7570115

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.109547 1.943446 0.015625 <list[3]> 0.7570115

# all evaluations
as.data.table(instance$archive)
#>              x1          x2 fidelity stage bracket repetition           y
#>  1:  0.60029387  2.71412858 0.015625     0       6          1  23.4062462
#>  2:  7.28014941 10.64362232 0.015625     0       6          1 228.3938890
#>  3: -0.16527571 10.79922443 0.015625     0       6          1  40.0405747
#>  4:  0.98195657 11.33410909 0.015625     0       6          1  62.4927080
#>  5:  7.25638117  6.50858339 0.015625     0       6          1 124.3647369
#>  6:  5.01003155  7.05016062 0.015625     0       6          1  80.9133170
#>  7:  8.08661647  9.47486580 0.015625     0       6          1 213.2632011
#>  8:  4.15784989 13.66879608 0.015625     0       6          1 194.1350541
#>  9:  5.07307697  9.91971202 0.015625     0       6          1 138.8837878
#> 10: -0.51891366  8.51139241 0.015625     0       6          1  21.1511975
#> 11:  0.32684476 11.82484075 0.015625     0       6          1  59.3115161
#> 12:  7.76650687  9.73681052 0.015625     0       6          1 213.7011424
#> 13:  3.10954738  1.94344550 0.015625     0       6          1   0.7570115
#> 14:  3.59831233 13.54460569 0.015625     0       6          1 167.1062336
#> 15: -3.76539262  8.14490774 0.015625     0       6          1  20.5575674
#> 16:  2.71818989  2.95577192 0.015625     0       6          1   2.3582452
#> 17: -1.71254097  2.70163964 0.015625     0       6          1  46.0260746
#> 18: -0.96229005  2.76844408 0.015625     0       6          1  38.4480219
#> 19: -0.71737305  5.62874402 0.015625     0       6          1  19.5728025
#> 20:  0.34559300  9.42449156 0.015625     0       6          1  34.8020163
#> 21:  2.44134113 14.67616649 0.015625     0       6          1 155.8828583
#> 22:  2.50180684  6.97408132 0.015625     0       6          1  24.9867842
#> 23:  1.42349532  7.34421966 0.015625     0       6          1  23.9938986
#> 24:  5.29593339  0.58500282 0.015625     0       6          1  19.9190755
#> 25:  1.01100002  8.03189565 0.015625     0       6          1  28.1275232
#> 26:  7.20510665 14.03763581 0.015625     0       6          1 336.5251624
#> 27:  2.82011850  8.67960980 0.015625     0       6          1  48.8237455
#> 28:  1.40105968  2.42071156 0.015625     0       6          1  13.6095181
#> 29:  5.26733247  9.73103130 0.015625     0       6          1 141.8735263
#> 30:  6.28301309  6.03682386 0.015625     0       6          1  97.4434367
#> 31:  3.47676189 10.64109383 0.015625     0       6          1  97.0283876
#> 32:  5.15831342  0.01815935 0.015625     0       6          1  16.1287039
#> 33:  4.83658424  4.73573195 0.015625     0       6          1  43.8409151
#> 34: -0.52631719  2.11981241 0.015625     0       6          1  40.6411142
#> 35:  1.88160852 11.05388736 0.015625     0       6          1  70.1022354
#> 36:  1.65947233  1.23510764 0.015625     0       6          1  14.0268137
#> 37:  2.21483790 12.26420561 0.015625     0       6          1  97.1333351
#> 38: -0.63345126  4.26657840 0.015625     0       6          1  25.3233214
#> 39: -0.71448907  9.07214214 0.015625     0       6          1  20.9374232
#> 40: -0.75189699  0.83087350 0.015625     0       6          1  57.7584514
#> 41:  0.06055089  1.51293023 0.015625     0       6          1  38.8637522
#> 42:  5.58876558  5.39917816 0.015625     0       6          1  71.1603921
#> 43:  9.71025032  5.69859504 0.015625     0       6          1 150.9429657
#> 44: -4.04529703 14.41156738 0.015625     0       6          1   6.2203875
#> 45: -1.13365483  5.79447986 0.015625     0       6          1  18.2647118
#> 46:  6.74153785  3.15148765 0.015625     0       6          1  60.6477930
#> 47: -4.48746045  6.40513146 0.015625     0       6          1  61.9697811
#> 48: -2.21045497 12.38000768 0.015625     0       6          1  11.6217496
#> 49:  0.43359157 11.62693950 0.015625     0       6          1  58.5453271
#> 50: -4.60562176  3.67481628 0.015625     0       6          1 115.2204117
#> 51:  9.25042220 12.91655407 0.015625     0       6          1 361.8485648
#> 52:  2.42341690 13.21789210 0.015625     0       6          1 121.4554038
#> 53:  5.49073918  1.04607892 0.015625     0       6          1  24.9092566
#> 54:  9.25504066  0.13925588 0.015625     0       6          1  39.4167343
#> 55:  4.94751627  7.48591686 0.015625     0       6          1  86.3262157
#> 56:  8.23263628 11.80114075 0.015625     0       6          1 289.3584006
#> 57:  8.63681380  2.20391481 0.015625     0       6          1  61.8462192
#> 58:  3.74136911 12.76654311 0.015625     0       6          1 153.1357483
#> 59: -2.23493658  1.40279242 0.015625     0       6          1  73.1009536
#> 60:  4.38170488  8.10992661 0.015625     0       6          1  79.0185068
#> 61:  6.24885772 14.73659031 0.015625     0       6          1 325.1937106
#> 62: -1.59923596  8.56684149 0.015625     0       6          1   9.7302119
#> 63: -4.12101047  6.73632250 0.015625     0       6          1  44.9011892
#> 64: -0.63897261  2.04381795 0.015625     0       6          1  42.5648271
#>              x1          x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr    status x_domain_x1 x_domain_x2
#>  1: 2021-12-06 14:19:12        1 evaluated  0.60029387  2.71412858
#>  2: 2021-12-06 14:19:12        1 evaluated  7.28014941 10.64362232
#>  3: 2021-12-06 14:19:12        1 evaluated -0.16527571 10.79922443
#>  4: 2021-12-06 14:19:12        1 evaluated  0.98195657 11.33410909
#>  5: 2021-12-06 14:19:12        1 evaluated  7.25638117  6.50858339
#>  6: 2021-12-06 14:19:12        1 evaluated  5.01003155  7.05016062
#>  7: 2021-12-06 14:19:12        1 evaluated  8.08661647  9.47486580
#>  8: 2021-12-06 14:19:12        1 evaluated  4.15784989 13.66879608
#>  9: 2021-12-06 14:19:12        1 evaluated  5.07307697  9.91971202
#> 10: 2021-12-06 14:19:12        1 evaluated -0.51891366  8.51139241
#> 11: 2021-12-06 14:19:12        1 evaluated  0.32684476 11.82484075
#> 12: 2021-12-06 14:19:12        1 evaluated  7.76650687  9.73681052
#> 13: 2021-12-06 14:19:12        1 evaluated  3.10954738  1.94344550
#> 14: 2021-12-06 14:19:12        1 evaluated  3.59831233 13.54460569
#> 15: 2021-12-06 14:19:12        1 evaluated -3.76539262  8.14490774
#> 16: 2021-12-06 14:19:12        1 evaluated  2.71818989  2.95577192
#> 17: 2021-12-06 14:19:12        1 evaluated -1.71254097  2.70163964
#> 18: 2021-12-06 14:19:12        1 evaluated -0.96229005  2.76844408
#> 19: 2021-12-06 14:19:12        1 evaluated -0.71737305  5.62874402
#> 20: 2021-12-06 14:19:12        1 evaluated  0.34559300  9.42449156
#> 21: 2021-12-06 14:19:12        1 evaluated  2.44134113 14.67616649
#> 22: 2021-12-06 14:19:12        1 evaluated  2.50180684  6.97408132
#> 23: 2021-12-06 14:19:12        1 evaluated  1.42349532  7.34421966
#> 24: 2021-12-06 14:19:12        1 evaluated  5.29593339  0.58500282
#> 25: 2021-12-06 14:19:12        1 evaluated  1.01100002  8.03189565
#> 26: 2021-12-06 14:19:12        1 evaluated  7.20510665 14.03763581
#> 27: 2021-12-06 14:19:12        1 evaluated  2.82011850  8.67960980
#> 28: 2021-12-06 14:19:12        1 evaluated  1.40105968  2.42071156
#> 29: 2021-12-06 14:19:12        1 evaluated  5.26733247  9.73103130
#> 30: 2021-12-06 14:19:12        1 evaluated  6.28301309  6.03682386
#> 31: 2021-12-06 14:19:12        1 evaluated  3.47676189 10.64109383
#> 32: 2021-12-06 14:19:12        1 evaluated  5.15831342  0.01815935
#> 33: 2021-12-06 14:19:12        1 evaluated  4.83658424  4.73573195
#> 34: 2021-12-06 14:19:12        1 evaluated -0.52631719  2.11981241
#> 35: 2021-12-06 14:19:12        1 evaluated  1.88160852 11.05388736
#> 36: 2021-12-06 14:19:12        1 evaluated  1.65947233  1.23510764
#> 37: 2021-12-06 14:19:12        1 evaluated  2.21483790 12.26420561
#> 38: 2021-12-06 14:19:12        1 evaluated -0.63345126  4.26657840
#> 39: 2021-12-06 14:19:12        1 evaluated -0.71448907  9.07214214
#> 40: 2021-12-06 14:19:12        1 evaluated -0.75189699  0.83087350
#> 41: 2021-12-06 14:19:12        1 evaluated  0.06055089  1.51293023
#> 42: 2021-12-06 14:19:12        1 evaluated  5.58876558  5.39917816
#> 43: 2021-12-06 14:19:12        1 evaluated  9.71025032  5.69859504
#> 44: 2021-12-06 14:19:12        1 evaluated -4.04529703 14.41156738
#> 45: 2021-12-06 14:19:12        1 evaluated -1.13365483  5.79447986
#> 46: 2021-12-06 14:19:12        1 evaluated  6.74153785  3.15148765
#> 47: 2021-12-06 14:19:12        1 evaluated -4.48746045  6.40513146
#> 48: 2021-12-06 14:19:12        1 evaluated -2.21045497 12.38000768
#> 49: 2021-12-06 14:19:12        1 evaluated  0.43359157 11.62693950
#> 50: 2021-12-06 14:19:12        1 evaluated -4.60562176  3.67481628
#> 51: 2021-12-06 14:19:12        1 evaluated  9.25042220 12.91655407
#> 52: 2021-12-06 14:19:12        1 evaluated  2.42341690 13.21789210
#> 53: 2021-12-06 14:19:12        1 evaluated  5.49073918  1.04607892
#> 54: 2021-12-06 14:19:12        1 evaluated  9.25504066  0.13925588
#> 55: 2021-12-06 14:19:12        1 evaluated  4.94751627  7.48591686
#> 56: 2021-12-06 14:19:12        1 evaluated  8.23263628 11.80114075
#> 57: 2021-12-06 14:19:12        1 evaluated  8.63681380  2.20391481
#> 58: 2021-12-06 14:19:12        1 evaluated  3.74136911 12.76654311
#> 59: 2021-12-06 14:19:12        1 evaluated -2.23493658  1.40279242
#> 60: 2021-12-06 14:19:12        1 evaluated  4.38170488  8.10992661
#> 61: 2021-12-06 14:19:12        1 evaluated  6.24885772 14.73659031
#> 62: 2021-12-06 14:19:12        1 evaluated -1.59923596  8.56684149
#> 63: 2021-12-06 14:19:12        1 evaluated -4.12101047  6.73632250
#> 64: 2021-12-06 14:19:12        1 evaluated -0.63897261  2.04381795
#>               timestamp batch_nr    status x_domain_x1 x_domain_x2
#>     x_domain_fidelity
#>  1:          0.015625
#>  2:          0.015625
#>  3:          0.015625
#>  4:          0.015625
#>  5:          0.015625
#>  6:          0.015625
#>  7:          0.015625
#>  8:          0.015625
#>  9:          0.015625
#> 10:          0.015625
#> 11:          0.015625
#> 12:          0.015625
#> 13:          0.015625
#> 14:          0.015625
#> 15:          0.015625
#> 16:          0.015625
#> 17:          0.015625
#> 18:          0.015625
#> 19:          0.015625
#> 20:          0.015625
#> 21:          0.015625
#> 22:          0.015625
#> 23:          0.015625
#> 24:          0.015625
#> 25:          0.015625
#> 26:          0.015625
#> 27:          0.015625
#> 28:          0.015625
#> 29:          0.015625
#> 30:          0.015625
#> 31:          0.015625
#> 32:          0.015625
#> 33:          0.015625
#> 34:          0.015625
#> 35:          0.015625
#> 36:          0.015625
#> 37:          0.015625
#> 38:          0.015625
#> 39:          0.015625
#> 40:          0.015625
#> 41:          0.015625
#> 42:          0.015625
#> 43:          0.015625
#> 44:          0.015625
#> 45:          0.015625
#> 46:          0.015625
#> 47:          0.015625
#> 48:          0.015625
#> 49:          0.015625
#> 50:          0.015625
#> 51:          0.015625
#> 52:          0.015625
#> 53:          0.015625
#> 54:          0.015625
#> 55:          0.015625
#> 56:          0.015625
#> 57:          0.015625
#> 58:          0.015625
#> 59:          0.015625
#> 60:          0.015625
#> 61:          0.015625
#> 62:          0.015625
#> 63:          0.015625
#> 64:          0.015625
#>     x_domain_fidelity