Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.358481 0.874195 0.015625 <list[3]> 0.6390848

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.358481 0.874195 0.015625 <list[3]> 0.6390848

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1:  9.4522623  4.3035357 0.015625     0       6          1 112.7653036
#>  2:  8.9716990  1.0922513 0.015625     0       6          1  48.9254459
#>  3: -2.4919490  8.3307461 0.015625     0       6          1   5.6890213
#>  4: -1.9287325 12.3729144 0.015625     0       6          1  16.8048194
#>  5:  9.0074162 10.8034859 0.015625     0       6          1 278.2654954
#>  6:  3.3406337  4.0145053 0.015625     0       6          1   9.5165510
#>  7:  7.6355003  0.4610040 0.015625     0       6          1  35.3204088
#>  8:  7.1473590  0.1669384 0.015625     0       6          1  32.0080427
#>  9:  5.0233254  1.0856497 0.015625     0       6          1  18.2492107
#> 10:  5.4487840 13.4781528 0.015625     0       6          1 248.6244871
#> 11: -4.3037824 13.9404254 0.015625     0       6          1   6.4564639
#> 12:  6.4644144 11.3789880 0.015625     0       6          1 226.3025151
#> 13:  5.4909290  9.8219514 0.015625     0       6          1 152.0930583
#> 14:  9.5656089 12.1282737 0.015625     0       6          1 344.1887671
#> 15: -2.1588742 10.3349463 0.015625     0       6          1   5.2441686
#> 16:  1.8845261 12.0932640 0.015625     0       6          1  87.7379004
#> 17: -2.4499074  2.3108797 0.015625     0       6          1  63.0213565
#> 18: -1.3536660  5.5407123 0.015625     0       6          1  19.1977609
#> 19:  9.8240416  2.5343703 0.015625     0       6          1  85.8367145
#> 20:  0.6869774 12.2545233 0.015625     0       6          1  71.2023631
#> 21:  1.7423648 10.5635390 0.015625     0       6          1  60.8254059
#> 22:  1.7739484  5.4639056 0.015625     0       6          1  12.8608803
#> 23:  8.6473616  5.1115992 0.015625     0       6          1 114.9898426
#> 24:  5.4016942  1.8830519 0.015625     0       6          1  28.9446275
#> 25: -1.0660472  8.2447838 0.015625     0       6          1  14.9068130
#> 26:  3.3132982  8.5489993 0.015625     0       6          1  56.5603940
#> 27: -4.8415636  9.7259090 0.015625     0       6          1  33.3307687
#> 28: -0.6647938  2.1886958 0.015625     0       6          1  41.4004382
#> 29:  3.0279133  8.4674942 0.015625     0       6          1  49.5252854
#> 30: -0.5062020 13.6033666 0.015625     0       6          1  64.4998946
#> 31:  4.4997428  9.6156229 0.015625     0       6          1 111.0900162
#> 32:  2.0401477 13.9516418 0.015625     0       6          1 128.2164542
#> 33:  6.1563567  3.5067019 0.015625     0       6          1  57.2186205
#> 34:  8.2996735 13.2331832 0.015625     0       6          1 341.6490630
#> 35: -3.5398482  9.8511160 0.015625     0       6          1   5.8495979
#> 36:  7.3766082  9.7277369 0.015625     0       6          1 204.7107973
#> 37:  4.9603018  8.1346629 0.015625     0       6          1  98.3393420
#> 38:  1.3080452  5.7681745 0.015625     0       6          1  15.7246306
#> 39:  0.7312325 11.3313723 0.015625     0       6          1  59.1212220
#> 40:  3.3584811  0.8741950 0.015625     0       6          1   0.6390848
#> 41: -2.7105378  6.5370819 0.015625     0       6          1  17.2984178
#> 42: -2.3421178  5.4852995 0.015625     0       6          1  22.7630955
#> 43:  7.1311255  4.2313251 0.015625     0       6          1  80.6289807
#> 44:  1.5652037  8.2027149 0.015625     0       6          1  31.3841324
#> 45:  9.8139271  3.4766782 0.015625     0       6          1 103.8276881
#> 46:  6.9448552  1.4024943 0.015625     0       6          1  42.3026547
#> 47:  7.9565627 13.2572917 0.015625     0       6          1 332.0834568
#> 48:  1.8545801  3.8351337 0.015625     0       6          1   7.7753138
#> 49: -0.7135192  5.7442033 0.015625     0       6          1  19.2395980
#> 50:  3.9951622  0.4008311 0.015625     0       6          1   3.7606859
#> 51:  2.6975920 10.4004483 0.015625     0       6          1  73.0707055
#> 52: -3.8451713 13.4764506 0.015625     0       6          1   3.4918049
#> 53:  4.9925910 10.5402268 0.015625     0       6          1 150.0087816
#> 54: -3.5642572  2.3560546 0.015625     0       6          1  95.4734830
#> 55:  3.2156439 11.5303720 0.015625     0       6          1 107.1391540
#> 56: -2.3192774 14.7268862 0.015625     0       6          1  27.1848594
#> 57: -1.0884476  6.2749540 0.015625     0       6          1  16.6854583
#> 58:  1.4621647  1.6213699 0.015625     0       6          1  15.5238230
#> 59:  8.5501179  8.4401615 0.015625     0       6          1 194.2922759
#> 60: -1.7153918  4.2504829 0.015625     0       6          1  29.5024364
#> 61:  1.2734627 14.9575320 0.015625     0       6          1 132.3753057
#> 62: -3.9177494 11.9046133 0.015625     0       6          1   3.7920030
#> 63:  8.2771523  9.4762630 0.015625     0       6          1 217.5628037
#> 64:  9.1199057  6.2931858 0.015625     0       6          1 150.9192444
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-28 04:52:18        1   9.4522623   4.3035357          0.015625
#>  2: 2022-08-28 04:52:18        1   8.9716990   1.0922513          0.015625
#>  3: 2022-08-28 04:52:18        1  -2.4919490   8.3307461          0.015625
#>  4: 2022-08-28 04:52:18        1  -1.9287325  12.3729144          0.015625
#>  5: 2022-08-28 04:52:18        1   9.0074162  10.8034859          0.015625
#>  6: 2022-08-28 04:52:18        1   3.3406337   4.0145053          0.015625
#>  7: 2022-08-28 04:52:18        1   7.6355003   0.4610040          0.015625
#>  8: 2022-08-28 04:52:18        1   7.1473590   0.1669384          0.015625
#>  9: 2022-08-28 04:52:18        1   5.0233254   1.0856497          0.015625
#> 10: 2022-08-28 04:52:18        1   5.4487840  13.4781528          0.015625
#> 11: 2022-08-28 04:52:18        1  -4.3037824  13.9404254          0.015625
#> 12: 2022-08-28 04:52:18        1   6.4644144  11.3789880          0.015625
#> 13: 2022-08-28 04:52:18        1   5.4909290   9.8219514          0.015625
#> 14: 2022-08-28 04:52:18        1   9.5656089  12.1282737          0.015625
#> 15: 2022-08-28 04:52:18        1  -2.1588742  10.3349463          0.015625
#> 16: 2022-08-28 04:52:18        1   1.8845261  12.0932640          0.015625
#> 17: 2022-08-28 04:52:18        1  -2.4499074   2.3108797          0.015625
#> 18: 2022-08-28 04:52:18        1  -1.3536660   5.5407123          0.015625
#> 19: 2022-08-28 04:52:18        1   9.8240416   2.5343703          0.015625
#> 20: 2022-08-28 04:52:18        1   0.6869774  12.2545233          0.015625
#> 21: 2022-08-28 04:52:18        1   1.7423648  10.5635390          0.015625
#> 22: 2022-08-28 04:52:18        1   1.7739484   5.4639056          0.015625
#> 23: 2022-08-28 04:52:18        1   8.6473616   5.1115992          0.015625
#> 24: 2022-08-28 04:52:18        1   5.4016942   1.8830519          0.015625
#> 25: 2022-08-28 04:52:18        1  -1.0660472   8.2447838          0.015625
#> 26: 2022-08-28 04:52:18        1   3.3132982   8.5489993          0.015625
#> 27: 2022-08-28 04:52:18        1  -4.8415636   9.7259090          0.015625
#> 28: 2022-08-28 04:52:18        1  -0.6647938   2.1886958          0.015625
#> 29: 2022-08-28 04:52:18        1   3.0279133   8.4674942          0.015625
#> 30: 2022-08-28 04:52:18        1  -0.5062020  13.6033666          0.015625
#> 31: 2022-08-28 04:52:18        1   4.4997428   9.6156229          0.015625
#> 32: 2022-08-28 04:52:18        1   2.0401477  13.9516418          0.015625
#> 33: 2022-08-28 04:52:18        1   6.1563567   3.5067019          0.015625
#> 34: 2022-08-28 04:52:18        1   8.2996735  13.2331832          0.015625
#> 35: 2022-08-28 04:52:18        1  -3.5398482   9.8511160          0.015625
#> 36: 2022-08-28 04:52:18        1   7.3766082   9.7277369          0.015625
#> 37: 2022-08-28 04:52:18        1   4.9603018   8.1346629          0.015625
#> 38: 2022-08-28 04:52:18        1   1.3080452   5.7681745          0.015625
#> 39: 2022-08-28 04:52:18        1   0.7312325  11.3313723          0.015625
#> 40: 2022-08-28 04:52:18        1   3.3584811   0.8741950          0.015625
#> 41: 2022-08-28 04:52:18        1  -2.7105378   6.5370819          0.015625
#> 42: 2022-08-28 04:52:18        1  -2.3421178   5.4852995          0.015625
#> 43: 2022-08-28 04:52:18        1   7.1311255   4.2313251          0.015625
#> 44: 2022-08-28 04:52:18        1   1.5652037   8.2027149          0.015625
#> 45: 2022-08-28 04:52:18        1   9.8139271   3.4766782          0.015625
#> 46: 2022-08-28 04:52:18        1   6.9448552   1.4024943          0.015625
#> 47: 2022-08-28 04:52:18        1   7.9565627  13.2572917          0.015625
#> 48: 2022-08-28 04:52:18        1   1.8545801   3.8351337          0.015625
#> 49: 2022-08-28 04:52:18        1  -0.7135192   5.7442033          0.015625
#> 50: 2022-08-28 04:52:18        1   3.9951622   0.4008311          0.015625
#> 51: 2022-08-28 04:52:18        1   2.6975920  10.4004483          0.015625
#> 52: 2022-08-28 04:52:18        1  -3.8451713  13.4764506          0.015625
#> 53: 2022-08-28 04:52:18        1   4.9925910  10.5402268          0.015625
#> 54: 2022-08-28 04:52:18        1  -3.5642572   2.3560546          0.015625
#> 55: 2022-08-28 04:52:18        1   3.2156439  11.5303720          0.015625
#> 56: 2022-08-28 04:52:18        1  -2.3192774  14.7268862          0.015625
#> 57: 2022-08-28 04:52:18        1  -1.0884476   6.2749540          0.015625
#> 58: 2022-08-28 04:52:18        1   1.4621647   1.6213699          0.015625
#> 59: 2022-08-28 04:52:18        1   8.5501179   8.4401615          0.015625
#> 60: 2022-08-28 04:52:18        1  -1.7153918   4.2504829          0.015625
#> 61: 2022-08-28 04:52:18        1   1.2734627  14.9575320          0.015625
#> 62: 2022-08-28 04:52:18        1  -3.9177494  11.9046133          0.015625
#> 63: 2022-08-28 04:52:18        1   8.2771523   9.4762630          0.015625
#> 64: 2022-08-28 04:52:18        1   9.1199057   6.2931858          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity