Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.110696 0.5333152 0.015625 <list[3]> 1.064057

# best scoring evaluation
instance$result
#>          x1        x2 fidelity  x_domain        y
#> 1: 3.110696 0.5333152 0.015625 <list[3]> 1.064057

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1:  1.3695682  5.1139350 0.015625     0       6          1  13.446892
#>  2:  7.4495283  9.1712436 0.015625     0       6          1 191.233393
#>  3: -3.9767661 12.1082216 0.015625     0       6          1   4.056716
#>  4: -2.5392985  7.4452547 0.015625     0       6          1   9.896289
#>  5: -0.5202771 10.7655053 0.015625     0       6          1  33.769683
#>  6:  3.3937858 12.9917873 0.015625     0       6          1 145.639815
#>  7:  5.7683965  4.2636541 0.015625     0       6          1  59.590152
#>  8:  9.8119810  6.4794319 0.015625     0       6          1 173.650555
#>  9:  9.8617409 11.9589794 0.015625     0       6          1 349.650958
#> 10: -3.8381789  0.3819379 0.015625     0       6          1 150.979109
#> 11:  0.8602201  2.9779643 0.015625     0       6          1  19.071164
#> 12:  0.2977161 13.7507743 0.015625     0       6          1  86.779000
#> 13: -2.0522089 11.8972827 0.015625     0       6          1  11.811910
#> 14: -4.5471538  0.5519071 0.015625     0       6          1 185.865826
#> 15: -3.4071772  1.1876342 0.015625     0       6          1 112.924891
#> 16:  9.0377318  5.0387511 0.015625     0       6          1 120.164948
#> 17:  5.4514013  5.7700605 0.015625     0       6          1  73.206173
#> 18:  6.9109749  2.2761330 0.015625     0       6          1  51.489784
#> 19:  7.7769598 11.1327693 0.015625     0       6          1 255.679252
#> 20:  7.1902274 13.2971705 0.015625     0       6          1 310.078347
#> 21:  3.3581931 10.5239122 0.015625     0       6          1  91.288724
#> 22: -0.4855749  6.4641938 0.015625     0       6          1  18.591949
#> 23: -0.5728776  4.5603064 0.015625     0       6          1  23.645990
#> 24:  6.7231602  4.7129068 0.015625     0       6          1  83.061882
#> 25:  2.6827911 11.6297532 0.015625     0       6          1  95.059430
#> 26:  8.0582852  8.4716938 0.015625     0       6          1 184.948960
#> 27:  3.8445844 10.6853651 0.015625     0       6          1 109.791659
#> 28:  7.1379336  7.4059648 0.015625     0       6          1 141.737888
#> 29:  7.0892680 10.0531302 0.015625     0       6          1 206.833338
#> 30:  3.6877784  1.3677167 0.015625     0       6          1   2.465392
#> 31:  4.5005634  0.1509351 0.015625     0       6          1   8.458714
#> 32: -3.2628960  1.3087275 0.015625     0       6          1 104.746865
#> 33:  0.3891195  2.3999178 0.015625     0       6          1  27.797114
#> 34: -4.3023582 13.2064501 0.015625     0       6          1   6.216380
#> 35:  1.0321822 13.4348665 0.015625     0       6          1  96.735185
#> 36: -4.3641051 12.6623126 0.015625     0       6          1   7.478042
#> 37:  4.1656343 14.0951992 0.015625     0       6          1 206.405636
#> 38: -0.7142399  8.0453873 0.015625     0       6          1  18.052641
#> 39:  2.9796545  9.4799088 0.015625     0       6          1  63.713197
#> 40: -1.1956040 10.4035311 0.015625     0       6          1  19.554164
#> 41:  7.1187806  6.7345796 0.015625     0       6          1 126.822899
#> 42:  6.4778894  2.8544445 0.015625     0       6          1  53.925582
#> 43:  0.9280907  6.7184100 0.015625     0       6          1  20.459839
#> 44:  0.7104071  9.6640379 0.015625     0       6          1  40.119779
#> 45:  9.6258934  5.6233193 0.015625     0       6          1 146.867231
#> 46:  6.0052213  2.5960330 0.015625     0       6          1  44.683753
#> 47:  7.1120887  4.9482042 0.015625     0       6          1  92.390526
#> 48:  1.5244139 12.1024363 0.015625     0       6          1  81.968855
#> 49: -3.8929698  1.1506977 0.015625     0       6          1 135.489462
#> 50: -0.6816983  6.7035162 0.015625     0       6          1  17.612678
#> 51:  3.2483697  3.1335127 0.015625     0       6          1   4.369071
#> 52: -0.5378417  3.8711773 0.015625     0       6          1  27.208802
#> 53:  9.4734142  8.5624779 0.015625     0       6          1 221.837831
#> 54: -2.2133391  6.6532618 0.015625     0       6          1  13.366511
#> 55:  3.1106957  0.5333152 0.015625     0       6          1   1.064057
#> 56:  8.7322601  0.7017872 0.015625     0       6          1  41.735829
#> 57: -1.7366938  4.7380207 0.015625     0       6          1  25.378429
#> 58:  8.0822049  2.8131968 0.015625     0       6          1  66.625284
#> 59:  0.5356455  3.5936691 0.015625     0       6          1  20.699099
#> 60: -0.5006984  4.0217847 0.015625     0       6          1  26.167460
#> 61: -4.8338618  5.3473515 0.015625     0       6          1  93.327236
#> 62:  5.8137696 14.8765934 0.015625     0       6          1 310.640081
#> 63:  1.8013914 13.0659210 0.015625     0       6          1 104.496213
#> 64:  0.2725603  7.6117256 0.015625     0       6          1  23.422457
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-16 04:41:57        1   1.3695682   5.1139350          0.015625
#>  2: 2022-10-16 04:41:57        1   7.4495283   9.1712436          0.015625
#>  3: 2022-10-16 04:41:57        1  -3.9767661  12.1082216          0.015625
#>  4: 2022-10-16 04:41:57        1  -2.5392985   7.4452547          0.015625
#>  5: 2022-10-16 04:41:57        1  -0.5202771  10.7655053          0.015625
#>  6: 2022-10-16 04:41:57        1   3.3937858  12.9917873          0.015625
#>  7: 2022-10-16 04:41:57        1   5.7683965   4.2636541          0.015625
#>  8: 2022-10-16 04:41:57        1   9.8119810   6.4794319          0.015625
#>  9: 2022-10-16 04:41:57        1   9.8617409  11.9589794          0.015625
#> 10: 2022-10-16 04:41:57        1  -3.8381789   0.3819379          0.015625
#> 11: 2022-10-16 04:41:57        1   0.8602201   2.9779643          0.015625
#> 12: 2022-10-16 04:41:57        1   0.2977161  13.7507743          0.015625
#> 13: 2022-10-16 04:41:57        1  -2.0522089  11.8972827          0.015625
#> 14: 2022-10-16 04:41:57        1  -4.5471538   0.5519071          0.015625
#> 15: 2022-10-16 04:41:57        1  -3.4071772   1.1876342          0.015625
#> 16: 2022-10-16 04:41:57        1   9.0377318   5.0387511          0.015625
#> 17: 2022-10-16 04:41:57        1   5.4514013   5.7700605          0.015625
#> 18: 2022-10-16 04:41:57        1   6.9109749   2.2761330          0.015625
#> 19: 2022-10-16 04:41:57        1   7.7769598  11.1327693          0.015625
#> 20: 2022-10-16 04:41:57        1   7.1902274  13.2971705          0.015625
#> 21: 2022-10-16 04:41:57        1   3.3581931  10.5239122          0.015625
#> 22: 2022-10-16 04:41:57        1  -0.4855749   6.4641938          0.015625
#> 23: 2022-10-16 04:41:57        1  -0.5728776   4.5603064          0.015625
#> 24: 2022-10-16 04:41:57        1   6.7231602   4.7129068          0.015625
#> 25: 2022-10-16 04:41:57        1   2.6827911  11.6297532          0.015625
#> 26: 2022-10-16 04:41:57        1   8.0582852   8.4716938          0.015625
#> 27: 2022-10-16 04:41:57        1   3.8445844  10.6853651          0.015625
#> 28: 2022-10-16 04:41:57        1   7.1379336   7.4059648          0.015625
#> 29: 2022-10-16 04:41:57        1   7.0892680  10.0531302          0.015625
#> 30: 2022-10-16 04:41:57        1   3.6877784   1.3677167          0.015625
#> 31: 2022-10-16 04:41:57        1   4.5005634   0.1509351          0.015625
#> 32: 2022-10-16 04:41:57        1  -3.2628960   1.3087275          0.015625
#> 33: 2022-10-16 04:41:57        1   0.3891195   2.3999178          0.015625
#> 34: 2022-10-16 04:41:57        1  -4.3023582  13.2064501          0.015625
#> 35: 2022-10-16 04:41:57        1   1.0321822  13.4348665          0.015625
#> 36: 2022-10-16 04:41:57        1  -4.3641051  12.6623126          0.015625
#> 37: 2022-10-16 04:41:57        1   4.1656343  14.0951992          0.015625
#> 38: 2022-10-16 04:41:57        1  -0.7142399   8.0453873          0.015625
#> 39: 2022-10-16 04:41:57        1   2.9796545   9.4799088          0.015625
#> 40: 2022-10-16 04:41:57        1  -1.1956040  10.4035311          0.015625
#> 41: 2022-10-16 04:41:57        1   7.1187806   6.7345796          0.015625
#> 42: 2022-10-16 04:41:57        1   6.4778894   2.8544445          0.015625
#> 43: 2022-10-16 04:41:57        1   0.9280907   6.7184100          0.015625
#> 44: 2022-10-16 04:41:57        1   0.7104071   9.6640379          0.015625
#> 45: 2022-10-16 04:41:57        1   9.6258934   5.6233193          0.015625
#> 46: 2022-10-16 04:41:57        1   6.0052213   2.5960330          0.015625
#> 47: 2022-10-16 04:41:57        1   7.1120887   4.9482042          0.015625
#> 48: 2022-10-16 04:41:57        1   1.5244139  12.1024363          0.015625
#> 49: 2022-10-16 04:41:57        1  -3.8929698   1.1506977          0.015625
#> 50: 2022-10-16 04:41:57        1  -0.6816983   6.7035162          0.015625
#> 51: 2022-10-16 04:41:57        1   3.2483697   3.1335127          0.015625
#> 52: 2022-10-16 04:41:57        1  -0.5378417   3.8711773          0.015625
#> 53: 2022-10-16 04:41:57        1   9.4734142   8.5624779          0.015625
#> 54: 2022-10-16 04:41:57        1  -2.2133391   6.6532618          0.015625
#> 55: 2022-10-16 04:41:57        1   3.1106957   0.5333152          0.015625
#> 56: 2022-10-16 04:41:57        1   8.7322601   0.7017872          0.015625
#> 57: 2022-10-16 04:41:57        1  -1.7366938   4.7380207          0.015625
#> 58: 2022-10-16 04:41:57        1   8.0822049   2.8131968          0.015625
#> 59: 2022-10-16 04:41:57        1   0.5356455   3.5936691          0.015625
#> 60: 2022-10-16 04:41:57        1  -0.5006984   4.0217847          0.015625
#> 61: 2022-10-16 04:41:57        1  -4.8338618   5.3473515          0.015625
#> 62: 2022-10-16 04:41:57        1   5.8137696  14.8765934          0.015625
#> 63: 2022-10-16 04:41:57        1   1.8013914  13.0659210          0.015625
#> 64: 2022-10-16 04:41:57        1   0.2725603   7.6117256          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity