Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 2.884923 2.184435 0.015625 <list[3]> 0.9828786

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 2.884923 2.184435 0.015625 <list[3]> 0.9828786

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition           y
#>  1: -4.8267048  3.2977533 0.015625     0       6          1 134.3164142
#>  2: -1.1266152 10.2849693 0.015625     0       6          1  20.1428240
#>  3: -2.4954915  0.4867084 0.015625     0       6          1  95.9673097
#>  4:  2.0391604  2.6461243 0.015625     0       6          1   5.7211844
#>  5: -3.6246844  7.1586532 0.015625     0       6          1  26.6386835
#>  6: -2.3431919  8.5858844 0.015625     0       6          1   5.0211081
#>  7: -1.2896500 11.7213105 0.015625     0       6          1  25.7514252
#>  8:  7.9730436 11.7732647 0.015625     0       6          1 281.3797244
#>  9: -0.2607682 12.7138026 0.015625     0       6          1  58.9257450
#> 10:  1.5869711  7.3032568 0.015625     0       6          1  23.9189279
#> 11: -1.1065029 11.5973655 0.015625     0       6          1  28.7296078
#> 12:  0.9457572  9.6108515 0.015625     0       6          1  41.5120079
#> 13: -1.3568510  4.4308027 0.015625     0       6          1  26.3671822
#> 14: -4.0499975 13.5191359 0.015625     0       6          1   4.4184625
#> 15: -2.4384866 14.7088961 0.015625     0       6          1  24.2520618
#> 16:  2.8849228  2.1844352 0.015625     0       6          1   0.9828786
#> 17:  8.0192415  6.9167187 0.015625     0       6          1 145.3677549
#> 18: -4.8644831  8.9611891 0.015625     0       6          1  41.7978092
#> 19:  2.5246702 13.0306868 0.015625     0       6          1 119.9521375
#> 20: -4.5450700  6.2634706 0.015625     0       6          1  66.2428918
#> 21:  7.1513462 14.9658195 0.015625     0       6          1 368.7089828
#> 22:  9.9303780  7.4028605 0.015625     0       6          1 202.5443888
#> 23:  3.7504404  9.6125899 0.015625     0       6          1  85.8297090
#> 24:  5.9904710  3.0583145 0.015625     0       6          1  49.3235247
#> 25: -1.9281135  8.2020777 0.015625     0       6          1   7.6037408
#> 26:  8.4414824  0.9085672 0.015625     0       6          1  42.5323762
#> 27:  5.0792871  1.0106720 0.015625     0       6          1  18.7407510
#> 28:  5.8003902  1.9020956 0.015625     0       6          1  35.3083244
#> 29:  8.4625236  9.9520815 0.015625     0       6          1 236.1189776
#> 30:  4.6796124  7.9603935 0.015625     0       6          1  85.9839181
#> 31:  5.8459243 13.2908116 0.015625     0       6          1 260.3182170
#> 32:  5.1926789 10.9684302 0.015625     0       6          1 168.2901743
#> 33:  1.7117563  7.7326702 0.015625     0       6          1  27.7209799
#> 34: -1.3505491 12.0095083 0.015625     0       6          1  26.5678984
#> 35: -2.2124270 12.9662624 0.015625     0       6          1  15.1073279
#> 36:  5.9771698  9.2410843 0.015625     0       6          1 155.0081071
#> 37:  8.0606008  4.9316097 0.015625     0       6          1 103.3408096
#> 38:  7.7074014  9.1686062 0.015625     0       6          1 196.6024284
#> 39:  8.9165177  9.7857092 0.015625     0       6          1 242.8630398
#> 40:  4.0050175  7.4277902 0.015625     0       6          1  57.1785378
#> 41:  2.6736281  9.9716562 0.015625     0       6          1  65.5435422
#> 42:  8.9729224  2.8980186 0.015625     0       6          1  77.1094163
#> 43:  4.4435463 10.2480118 0.015625     0       6          1 122.2185911
#> 44: -3.9929622 12.6977271 0.015625     0       6          1   3.6944069
#> 45:  8.3870230 14.2936176 0.015625     0       6          1 384.5586415
#> 46:  0.3121966 10.3990510 0.015625     0       6          1  43.0787271
#> 47: -1.0640878  7.2482123 0.015625     0       6          1  14.8904689
#> 48:  8.5925520  8.7553531 0.015625     0       6          1 204.0613398
#> 49: -1.9758474  9.4801166 0.015625     0       6          1   6.2625437
#> 50:  8.6027935  4.6271720 0.015625     0       6          1 104.3334010
#> 51:  6.2225714  5.1958366 0.015625     0       6          1  82.1340748
#> 52: -1.1476077  4.2292097 0.015625     0       6          1  27.1765859
#> 53:  7.4676870  8.8431469 0.015625     0       6          1 182.9738466
#> 54:  6.4804420  7.0495165 0.015625     0       6          1 120.8652202
#> 55:  7.2303759 11.2639814 0.015625     0       6          1 245.5566184
#> 56: -4.7352641  0.6225647 0.015625     0       6          1 195.2686953
#> 57:  8.9897687  3.2054662 0.015625     0       6          1  82.8019901
#> 58:  4.5813341 14.6492341 0.015625     0       6          1 242.6918989
#> 59:  9.6067363 12.0740751 0.015625     0       6          1 343.7708726
#> 60: -0.6609563 13.5845037 0.015625     0       6          1  60.0789754
#> 61: -1.5382152  4.1741954 0.015625     0       6          1  29.2066005
#> 62: -2.3409529 13.1416479 0.015625     0       6          1  13.8602250
#> 63: -4.9942628 14.6638007 0.015625     0       6          1  12.6735612
#> 64:  9.7252416  4.3083357 0.015625     0       6          1 119.1693167
#>             x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-18 04:49:33        1  -4.8267048   3.2977533          0.015625
#>  2: 2022-10-18 04:49:33        1  -1.1266152  10.2849693          0.015625
#>  3: 2022-10-18 04:49:33        1  -2.4954915   0.4867084          0.015625
#>  4: 2022-10-18 04:49:33        1   2.0391604   2.6461243          0.015625
#>  5: 2022-10-18 04:49:33        1  -3.6246844   7.1586532          0.015625
#>  6: 2022-10-18 04:49:33        1  -2.3431919   8.5858844          0.015625
#>  7: 2022-10-18 04:49:33        1  -1.2896500  11.7213105          0.015625
#>  8: 2022-10-18 04:49:33        1   7.9730436  11.7732647          0.015625
#>  9: 2022-10-18 04:49:33        1  -0.2607682  12.7138026          0.015625
#> 10: 2022-10-18 04:49:33        1   1.5869711   7.3032568          0.015625
#> 11: 2022-10-18 04:49:33        1  -1.1065029  11.5973655          0.015625
#> 12: 2022-10-18 04:49:33        1   0.9457572   9.6108515          0.015625
#> 13: 2022-10-18 04:49:33        1  -1.3568510   4.4308027          0.015625
#> 14: 2022-10-18 04:49:33        1  -4.0499975  13.5191359          0.015625
#> 15: 2022-10-18 04:49:33        1  -2.4384866  14.7088961          0.015625
#> 16: 2022-10-18 04:49:33        1   2.8849228   2.1844352          0.015625
#> 17: 2022-10-18 04:49:33        1   8.0192415   6.9167187          0.015625
#> 18: 2022-10-18 04:49:33        1  -4.8644831   8.9611891          0.015625
#> 19: 2022-10-18 04:49:33        1   2.5246702  13.0306868          0.015625
#> 20: 2022-10-18 04:49:33        1  -4.5450700   6.2634706          0.015625
#> 21: 2022-10-18 04:49:33        1   7.1513462  14.9658195          0.015625
#> 22: 2022-10-18 04:49:33        1   9.9303780   7.4028605          0.015625
#> 23: 2022-10-18 04:49:33        1   3.7504404   9.6125899          0.015625
#> 24: 2022-10-18 04:49:33        1   5.9904710   3.0583145          0.015625
#> 25: 2022-10-18 04:49:33        1  -1.9281135   8.2020777          0.015625
#> 26: 2022-10-18 04:49:33        1   8.4414824   0.9085672          0.015625
#> 27: 2022-10-18 04:49:33        1   5.0792871   1.0106720          0.015625
#> 28: 2022-10-18 04:49:33        1   5.8003902   1.9020956          0.015625
#> 29: 2022-10-18 04:49:33        1   8.4625236   9.9520815          0.015625
#> 30: 2022-10-18 04:49:33        1   4.6796124   7.9603935          0.015625
#> 31: 2022-10-18 04:49:33        1   5.8459243  13.2908116          0.015625
#> 32: 2022-10-18 04:49:33        1   5.1926789  10.9684302          0.015625
#> 33: 2022-10-18 04:49:33        1   1.7117563   7.7326702          0.015625
#> 34: 2022-10-18 04:49:33        1  -1.3505491  12.0095083          0.015625
#> 35: 2022-10-18 04:49:33        1  -2.2124270  12.9662624          0.015625
#> 36: 2022-10-18 04:49:33        1   5.9771698   9.2410843          0.015625
#> 37: 2022-10-18 04:49:33        1   8.0606008   4.9316097          0.015625
#> 38: 2022-10-18 04:49:33        1   7.7074014   9.1686062          0.015625
#> 39: 2022-10-18 04:49:33        1   8.9165177   9.7857092          0.015625
#> 40: 2022-10-18 04:49:33        1   4.0050175   7.4277902          0.015625
#> 41: 2022-10-18 04:49:33        1   2.6736281   9.9716562          0.015625
#> 42: 2022-10-18 04:49:33        1   8.9729224   2.8980186          0.015625
#> 43: 2022-10-18 04:49:33        1   4.4435463  10.2480118          0.015625
#> 44: 2022-10-18 04:49:33        1  -3.9929622  12.6977271          0.015625
#> 45: 2022-10-18 04:49:33        1   8.3870230  14.2936176          0.015625
#> 46: 2022-10-18 04:49:33        1   0.3121966  10.3990510          0.015625
#> 47: 2022-10-18 04:49:33        1  -1.0640878   7.2482123          0.015625
#> 48: 2022-10-18 04:49:33        1   8.5925520   8.7553531          0.015625
#> 49: 2022-10-18 04:49:33        1  -1.9758474   9.4801166          0.015625
#> 50: 2022-10-18 04:49:33        1   8.6027935   4.6271720          0.015625
#> 51: 2022-10-18 04:49:33        1   6.2225714   5.1958366          0.015625
#> 52: 2022-10-18 04:49:33        1  -1.1476077   4.2292097          0.015625
#> 53: 2022-10-18 04:49:33        1   7.4676870   8.8431469          0.015625
#> 54: 2022-10-18 04:49:33        1   6.4804420   7.0495165          0.015625
#> 55: 2022-10-18 04:49:33        1   7.2303759  11.2639814          0.015625
#> 56: 2022-10-18 04:49:33        1  -4.7352641   0.6225647          0.015625
#> 57: 2022-10-18 04:49:33        1   8.9897687   3.2054662          0.015625
#> 58: 2022-10-18 04:49:33        1   4.5813341  14.6492341          0.015625
#> 59: 2022-10-18 04:49:33        1   9.6067363  12.0740751          0.015625
#> 60: 2022-10-18 04:49:33        1  -0.6609563  13.5845037          0.015625
#> 61: 2022-10-18 04:49:33        1  -1.5382152   4.1741954          0.015625
#> 62: 2022-10-18 04:49:33        1  -2.3409529  13.1416479          0.015625
#> 63: 2022-10-18 04:49:33        1  -4.9942628  14.6638007          0.015625
#> 64: 2022-10-18 04:49:33        1   9.7252416   4.3083357          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity