Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: -3.24893 11.00569 0.015625 <list[3]> 0.692946

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: -3.24893 11.00569 0.015625 <list[3]> 0.692946

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1: -4.4257974  1.6397443 0.015625     0       6          1 151.439121
#>  2: -3.5009404  8.5850527 0.015625     0       6          1  12.325822
#>  3:  6.3920502 14.8873001 0.015625     0       6          1 336.538127
#>  4: -3.3157472  7.3957408 0.015625     0       6          1  18.347086
#>  5: -4.2370296  8.9399284 0.015625     0       6          1  24.575868
#>  6:  6.1565022 14.7415984 0.015625     0       6          1 321.401432
#>  7: -4.4515936  8.2087615 0.015625     0       6          1  37.614523
#>  8:  3.4274195  2.0751288 0.015625     0       6          1   2.153654
#>  9: -3.0251011  3.4031554 0.015625     0       6          1  59.642380
#> 10:  9.6700066  8.0063819 0.015625     0       6          1 211.560653
#> 11:  8.8642514  1.1418030 0.015625     0       6          1  48.567421
#> 12:  8.8415715  6.5351390 0.015625     0       6          1 150.906596
#> 13:  3.8007629  5.0532395 0.015625     0       6          1  24.108130
#> 14: -2.2119137  5.7320709 0.015625     0       6          1  19.770660
#> 15:  9.1360739  7.8055238 0.015625     0       6          1 190.674048
#> 16: -0.5990881 11.0567836 0.015625     0       6          1  34.676579
#> 17:  4.6052836 10.6200398 0.015625     0       6          1 136.606450
#> 18:  4.0156613  7.1786913 0.015625     0       6          1  53.879729
#> 19: -0.2378530 10.4554841 0.015625     0       6          1  35.938948
#> 20:  7.3873097 13.9603224 0.015625     0       6          1 339.749290
#> 21:  8.3670964  9.0472454 0.015625     0       6          1 207.248773
#> 22:  6.8712085  8.0110330 0.015625     0       6          1 150.129572
#> 23:  4.1292789  5.5824183 0.015625     0       6          1  36.410933
#> 24:  3.5690254 13.6427005 0.015625     0       6          1 168.480985
#> 25: -4.2808247  4.2205611 0.015625     0       6          1  89.816509
#> 26: -1.2706528  1.7315260 0.015625     0       6          1  53.039905
#> 27: -3.0498207  0.4219407 0.015625     0       6          1 115.313607
#> 28: -0.9221884 10.4543247 0.015625     0       6          1  24.564792
#> 29:  6.9462947  4.8754737 0.015625     0       6          1  88.923637
#> 30:  6.7173420  9.1387850 0.015625     0       6          1 173.524081
#> 31: -3.0380158 13.7513201 0.015625     0       6          1   7.378810
#> 32:  8.4141499 14.1662928 0.015625     0       6          1 380.521438
#> 33:  1.7857933  5.8556037 0.015625     0       6          1  14.710028
#> 34:  3.1764544  1.8148723 0.015625     0       6          1   0.717460
#> 35: -0.3424180 14.5173364 0.015625     0       6          1  82.545742
#> 36:  7.5272966 12.0849400 0.015625     0       6          1 279.517554
#> 37: -4.1401888 11.0166544 0.015625     0       6          1   9.209360
#> 38: -1.3431862  7.0101757 0.015625     0       6          1  13.566309
#> 39:  8.9399488  9.6345265 0.015625     0       6          1 238.834333
#> 40: -2.0791802  4.1707445 0.015625     0       6          1  33.112543
#> 41:  4.9305089  8.4888123 0.015625     0       6          1 104.017247
#> 42:  1.6293889  3.9127219 0.015625     0       6          1   9.617777
#> 43:  4.3151668  7.6484955 0.015625     0       6          1  69.388777
#> 44: -4.5566545  3.7597080 0.015625     0       6          1 111.144496
#> 45: -2.1243235  3.2681359 0.015625     0       6          1  44.034553
#> 46: -3.2489301 11.0056934 0.015625     0       6          1   0.692946
#> 47: -0.5528837 13.5765526 0.015625     0       6          1  62.890334
#> 48:  1.4658040  9.3990127 0.015625     0       6          1  43.108150
#> 49:  8.9477459 13.6460676 0.015625     0       6          1 378.807301
#> 50:  4.5847366  7.8247417 0.015625     0       6          1  80.607970
#> 51:  9.4360164 11.9363906 0.015625     0       6          1 332.243463
#> 52:  6.0055231  1.6050331 0.015625     0       6          1  35.670788
#> 53:  1.0062277  9.9338085 0.015625     0       6          1  45.433165
#> 54:  4.3064161  4.2225374 0.015625     0       6          1  26.513854
#> 55:  6.8122560  4.4819734 0.015625     0       6          1  80.654165
#> 56: -0.4519616  8.3617104 0.015625     0       6          1  21.314843
#> 57: -0.1760435  8.1994393 0.015625     0       6          1  23.133598
#> 58:  8.8109950 11.1467388 0.015625     0       6          1 283.815268
#> 59:  2.2269214  2.0667608 0.015625     0       6          1   4.435403
#> 60: -1.0821184  5.9192920 0.015625     0       6          1  17.889561
#> 61: -1.7426138  7.9417500 0.015625     0       6          1   9.214060
#> 62: -4.5673336  9.3957321 0.015625     0       6          1  28.995498
#> 63:  2.0016786 12.6662564 0.015625     0       6          1 100.639639
#> 64:  0.1894860  1.4515830 0.015625     0       6          1  37.475281
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-08 04:25:43        1  -4.4257974   1.6397443          0.015625
#>  2: 2022-10-08 04:25:43        1  -3.5009404   8.5850527          0.015625
#>  3: 2022-10-08 04:25:43        1   6.3920502  14.8873001          0.015625
#>  4: 2022-10-08 04:25:43        1  -3.3157472   7.3957408          0.015625
#>  5: 2022-10-08 04:25:43        1  -4.2370296   8.9399284          0.015625
#>  6: 2022-10-08 04:25:43        1   6.1565022  14.7415984          0.015625
#>  7: 2022-10-08 04:25:43        1  -4.4515936   8.2087615          0.015625
#>  8: 2022-10-08 04:25:43        1   3.4274195   2.0751288          0.015625
#>  9: 2022-10-08 04:25:43        1  -3.0251011   3.4031554          0.015625
#> 10: 2022-10-08 04:25:43        1   9.6700066   8.0063819          0.015625
#> 11: 2022-10-08 04:25:43        1   8.8642514   1.1418030          0.015625
#> 12: 2022-10-08 04:25:43        1   8.8415715   6.5351390          0.015625
#> 13: 2022-10-08 04:25:43        1   3.8007629   5.0532395          0.015625
#> 14: 2022-10-08 04:25:43        1  -2.2119137   5.7320709          0.015625
#> 15: 2022-10-08 04:25:43        1   9.1360739   7.8055238          0.015625
#> 16: 2022-10-08 04:25:43        1  -0.5990881  11.0567836          0.015625
#> 17: 2022-10-08 04:25:43        1   4.6052836  10.6200398          0.015625
#> 18: 2022-10-08 04:25:43        1   4.0156613   7.1786913          0.015625
#> 19: 2022-10-08 04:25:43        1  -0.2378530  10.4554841          0.015625
#> 20: 2022-10-08 04:25:43        1   7.3873097  13.9603224          0.015625
#> 21: 2022-10-08 04:25:43        1   8.3670964   9.0472454          0.015625
#> 22: 2022-10-08 04:25:43        1   6.8712085   8.0110330          0.015625
#> 23: 2022-10-08 04:25:43        1   4.1292789   5.5824183          0.015625
#> 24: 2022-10-08 04:25:43        1   3.5690254  13.6427005          0.015625
#> 25: 2022-10-08 04:25:43        1  -4.2808247   4.2205611          0.015625
#> 26: 2022-10-08 04:25:43        1  -1.2706528   1.7315260          0.015625
#> 27: 2022-10-08 04:25:43        1  -3.0498207   0.4219407          0.015625
#> 28: 2022-10-08 04:25:43        1  -0.9221884  10.4543247          0.015625
#> 29: 2022-10-08 04:25:43        1   6.9462947   4.8754737          0.015625
#> 30: 2022-10-08 04:25:43        1   6.7173420   9.1387850          0.015625
#> 31: 2022-10-08 04:25:43        1  -3.0380158  13.7513201          0.015625
#> 32: 2022-10-08 04:25:43        1   8.4141499  14.1662928          0.015625
#> 33: 2022-10-08 04:25:43        1   1.7857933   5.8556037          0.015625
#> 34: 2022-10-08 04:25:43        1   3.1764544   1.8148723          0.015625
#> 35: 2022-10-08 04:25:43        1  -0.3424180  14.5173364          0.015625
#> 36: 2022-10-08 04:25:43        1   7.5272966  12.0849400          0.015625
#> 37: 2022-10-08 04:25:43        1  -4.1401888  11.0166544          0.015625
#> 38: 2022-10-08 04:25:43        1  -1.3431862   7.0101757          0.015625
#> 39: 2022-10-08 04:25:43        1   8.9399488   9.6345265          0.015625
#> 40: 2022-10-08 04:25:43        1  -2.0791802   4.1707445          0.015625
#> 41: 2022-10-08 04:25:43        1   4.9305089   8.4888123          0.015625
#> 42: 2022-10-08 04:25:43        1   1.6293889   3.9127219          0.015625
#> 43: 2022-10-08 04:25:43        1   4.3151668   7.6484955          0.015625
#> 44: 2022-10-08 04:25:43        1  -4.5566545   3.7597080          0.015625
#> 45: 2022-10-08 04:25:43        1  -2.1243235   3.2681359          0.015625
#> 46: 2022-10-08 04:25:43        1  -3.2489301  11.0056934          0.015625
#> 47: 2022-10-08 04:25:43        1  -0.5528837  13.5765526          0.015625
#> 48: 2022-10-08 04:25:43        1   1.4658040   9.3990127          0.015625
#> 49: 2022-10-08 04:25:43        1   8.9477459  13.6460676          0.015625
#> 50: 2022-10-08 04:25:43        1   4.5847366   7.8247417          0.015625
#> 51: 2022-10-08 04:25:43        1   9.4360164  11.9363906          0.015625
#> 52: 2022-10-08 04:25:43        1   6.0055231   1.6050331          0.015625
#> 53: 2022-10-08 04:25:43        1   1.0062277   9.9338085          0.015625
#> 54: 2022-10-08 04:25:43        1   4.3064161   4.2225374          0.015625
#> 55: 2022-10-08 04:25:43        1   6.8122560   4.4819734          0.015625
#> 56: 2022-10-08 04:25:43        1  -0.4519616   8.3617104          0.015625
#> 57: 2022-10-08 04:25:43        1  -0.1760435   8.1994393          0.015625
#> 58: 2022-10-08 04:25:43        1   8.8109950  11.1467388          0.015625
#> 59: 2022-10-08 04:25:43        1   2.2269214   2.0667608          0.015625
#> 60: 2022-10-08 04:25:43        1  -1.0821184   5.9192920          0.015625
#> 61: 2022-10-08 04:25:43        1  -1.7426138   7.9417500          0.015625
#> 62: 2022-10-08 04:25:43        1  -4.5673336   9.3957321          0.015625
#> 63: 2022-10-08 04:25:43        1   2.0016786  12.6662564          0.015625
#> 64: 2022-10-08 04:25:43        1   0.1894860   1.4515830          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity