Optimizer using the Hyperband algorithm

OptimizerHyperband class that implements hyperband optimization (HBX). HBX repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HBX is called a bracket. HBX considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Usage,NULL

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repeats

integer(1)
If 1 (default), optimization is stopped once all stages are evaluated. Otherwise, optimization is stopped after repeats runs of SHA. The bbotk::Terminator might stop the optimization before all repeats are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

stage (integer(1))
The stages of each bracket. Starts counting at 0.

bracket (integer(1))
The bracket index. Counts down to 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.811486 12.03275 0.015625 <list[3]> 2.703509

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.811486 12.03275 0.015625 <list[3]> 2.703509

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage bracket repetition          y
#>  1:  0.1618267  5.0412279 0.015625     0       6          1  19.969492
#>  2:  0.0613847  1.4823089 0.015625     0       6          1  39.121402
#>  3:  0.6708171  4.8680836 0.015625     0       6          1  17.527571
#>  4:  8.2731658  4.9262137 0.015625     0       6          1 105.869488
#>  5: -2.7111118  5.0463776 0.015625     0       6          1  31.463303
#>  6:  1.1521835  6.7645834 0.015625     0       6          1  20.444119
#>  7:  3.6828906  7.1235067 0.015625     0       6          1  44.908783
#>  8:  5.8506289  9.4542688 0.015625     0       6          1 155.920724
#>  9: -3.8114863 12.0327529 0.015625     0       6          1   2.703509
#> 10:  3.2250740  9.0633296 0.015625     0       6          1  62.468870
#> 11: -3.7284934  2.5370827 0.015625     0       6          1  98.524186
#> 12:  8.2075086 11.8088976 0.015625     0       6          1 288.926967
#> 13:  8.1400478  8.3685333 0.015625     0       6          1 183.821907
#> 14: -0.5377112  6.8136303 0.015625     0       6          1  18.249699
#> 15:  7.0090698  6.3972303 0.015625     0       6          1 118.023329
#> 16:  0.9309608  2.1622981 0.015625     0       6          1  21.410234
#> 17: -2.8669590  9.2557969 0.015625     0       6          1   3.190811
#> 18:  2.1669193  8.7707105 0.015625     0       6          1  41.515795
#> 19:  2.1919240  5.7435397 0.015625     0       6          1  13.925365
#> 20: -1.3941886 12.9839835 0.015625     0       6          1  33.826841
#> 21:  2.7503149  8.3950922 0.015625     0       6          1  43.892271
#> 22:  0.7077839  8.0811534 0.015625     0       6          1  27.486020
#> 23: -0.1053913 10.6119867 0.015625     0       6          1  39.297169
#> 24:  1.9549179  5.9583236 0.015625     0       6          1  15.116942
#> 25:  9.7432173  8.5826999 0.015625     0       6          1 231.030453
#> 26:  7.1513225  2.1446509 0.015625     0       6          1  51.654148
#> 27:  2.1604260 11.7825019 0.015625     0       6          1  87.060210
#> 28:  3.9221444  9.3250681 0.015625     0       6          1  85.884965
#> 29: -1.7414043  1.6622377 0.015625     0       6          1  60.246223
#> 30:  6.4515128 14.5503263 0.015625     0       6          1 327.064513
#> 31:  3.7623406 13.8972208 0.015625     0       6          1 183.090152
#> 32:  6.5061076 13.8332049 0.015625     0       6          1 304.518167
#> 33:  3.3861444  4.1544836 0.015625     0       6          1  10.867062
#> 34:  6.3416196 13.2752762 0.015625     0       6          1 279.819122
#> 35:  3.8310593 12.6482600 0.015625     0       6          1 153.741108
#> 36: -2.0476703 14.1323742 0.015625     0       6          1  28.102728
#> 37:  4.7476297 11.2972259 0.015625     0       6          1 158.210627
#> 38: -1.2013400  6.9891763 0.015625     0       6          1  14.402862
#> 39:  2.6223184 10.3234675 0.015625     0       6          1  70.314534
#> 40: -4.3440091  9.7587779 0.015625     0       6          1  20.493477
#> 41: -0.8625183 12.8530085 0.015625     0       6          1  46.029583
#> 42: -1.6690802  1.1252435 0.015625     0       6          1  67.073977
#> 43: -1.2500028  9.0071851 0.015625     0       6          1  13.968057
#> 44:  2.5526341  6.3344895 0.015625     0       6          1  19.628677
#> 45: -4.4263995 12.5615935 0.015625     0       6          1   8.469849
#> 46:  7.5648691  6.3250899 0.015625     0       6          1 125.211799
#> 47:  2.2921974 10.6273152 0.015625     0       6          1  69.493953
#> 48:  5.5243825  7.9629195 0.015625     0       6          1 113.339112
#> 49:  6.1704873  9.4200734 0.015625     0       6          1 165.226564
#> 50:  5.4174114 13.2616490 0.015625     0       6          1 240.663598
#> 51: -2.3978923  2.9768443 0.015625     0       6          1  52.161777
#> 52:  3.1116903  8.0510128 0.015625     0       6          1  45.368730
#> 53:  1.1416711 12.8359279 0.015625     0       6          1  88.176870
#> 54:  3.3325539  2.9746808 0.015625     0       6          1   4.324910
#> 55: -2.3003639  6.8584964 0.015625     0       6          1  12.393036
#> 56:  1.6719916  1.5688385 0.015625     0       6          1  12.474926
#> 57:  2.7450744 12.5138203 0.015625     0       6          1 114.587767
#> 58:  7.0894608  5.9089132 0.015625     0       6          1 109.706900
#> 59:  3.5813040  7.2694230 0.015625     0       6          1  44.540500
#> 60:  0.2704535  6.6255357 0.015625     0       6          1  20.363415
#> 61: -1.1674056  9.2720357 0.015625     0       6          1  15.651998
#> 62:  6.4942419  9.4896205 0.015625     0       6          1 176.359043
#> 63: -4.6178751  0.2442206 0.015625     0       6          1 198.459732
#> 64: -2.4662010  6.5008003 0.015625     0       6          1  15.547318
#>             x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr    status x_domain_x1 x_domain_x2
#>  1: 2021-12-01 16:59:37        1 evaluated   0.1618267   5.0412279
#>  2: 2021-12-01 16:59:37        1 evaluated   0.0613847   1.4823089
#>  3: 2021-12-01 16:59:37        1 evaluated   0.6708171   4.8680836
#>  4: 2021-12-01 16:59:37        1 evaluated   8.2731658   4.9262137
#>  5: 2021-12-01 16:59:37        1 evaluated  -2.7111118   5.0463776
#>  6: 2021-12-01 16:59:37        1 evaluated   1.1521835   6.7645834
#>  7: 2021-12-01 16:59:37        1 evaluated   3.6828906   7.1235067
#>  8: 2021-12-01 16:59:37        1 evaluated   5.8506289   9.4542688
#>  9: 2021-12-01 16:59:37        1 evaluated  -3.8114863  12.0327529
#> 10: 2021-12-01 16:59:37        1 evaluated   3.2250740   9.0633296
#> 11: 2021-12-01 16:59:37        1 evaluated  -3.7284934   2.5370827
#> 12: 2021-12-01 16:59:37        1 evaluated   8.2075086  11.8088976
#> 13: 2021-12-01 16:59:37        1 evaluated   8.1400478   8.3685333
#> 14: 2021-12-01 16:59:37        1 evaluated  -0.5377112   6.8136303
#> 15: 2021-12-01 16:59:37        1 evaluated   7.0090698   6.3972303
#> 16: 2021-12-01 16:59:37        1 evaluated   0.9309608   2.1622981
#> 17: 2021-12-01 16:59:37        1 evaluated  -2.8669590   9.2557969
#> 18: 2021-12-01 16:59:37        1 evaluated   2.1669193   8.7707105
#> 19: 2021-12-01 16:59:37        1 evaluated   2.1919240   5.7435397
#> 20: 2021-12-01 16:59:37        1 evaluated  -1.3941886  12.9839835
#> 21: 2021-12-01 16:59:37        1 evaluated   2.7503149   8.3950922
#> 22: 2021-12-01 16:59:37        1 evaluated   0.7077839   8.0811534
#> 23: 2021-12-01 16:59:37        1 evaluated  -0.1053913  10.6119867
#> 24: 2021-12-01 16:59:37        1 evaluated   1.9549179   5.9583236
#> 25: 2021-12-01 16:59:37        1 evaluated   9.7432173   8.5826999
#> 26: 2021-12-01 16:59:37        1 evaluated   7.1513225   2.1446509
#> 27: 2021-12-01 16:59:37        1 evaluated   2.1604260  11.7825019
#> 28: 2021-12-01 16:59:37        1 evaluated   3.9221444   9.3250681
#> 29: 2021-12-01 16:59:37        1 evaluated  -1.7414043   1.6622377
#> 30: 2021-12-01 16:59:37        1 evaluated   6.4515128  14.5503263
#> 31: 2021-12-01 16:59:37        1 evaluated   3.7623406  13.8972208
#> 32: 2021-12-01 16:59:37        1 evaluated   6.5061076  13.8332049
#> 33: 2021-12-01 16:59:37        1 evaluated   3.3861444   4.1544836
#> 34: 2021-12-01 16:59:37        1 evaluated   6.3416196  13.2752762
#> 35: 2021-12-01 16:59:37        1 evaluated   3.8310593  12.6482600
#> 36: 2021-12-01 16:59:37        1 evaluated  -2.0476703  14.1323742
#> 37: 2021-12-01 16:59:37        1 evaluated   4.7476297  11.2972259
#> 38: 2021-12-01 16:59:37        1 evaluated  -1.2013400   6.9891763
#> 39: 2021-12-01 16:59:37        1 evaluated   2.6223184  10.3234675
#> 40: 2021-12-01 16:59:37        1 evaluated  -4.3440091   9.7587779
#> 41: 2021-12-01 16:59:37        1 evaluated  -0.8625183  12.8530085
#> 42: 2021-12-01 16:59:37        1 evaluated  -1.6690802   1.1252435
#> 43: 2021-12-01 16:59:37        1 evaluated  -1.2500028   9.0071851
#> 44: 2021-12-01 16:59:37        1 evaluated   2.5526341   6.3344895
#> 45: 2021-12-01 16:59:37        1 evaluated  -4.4263995  12.5615935
#> 46: 2021-12-01 16:59:37        1 evaluated   7.5648691   6.3250899
#> 47: 2021-12-01 16:59:37        1 evaluated   2.2921974  10.6273152
#> 48: 2021-12-01 16:59:37        1 evaluated   5.5243825   7.9629195
#> 49: 2021-12-01 16:59:37        1 evaluated   6.1704873   9.4200734
#> 50: 2021-12-01 16:59:37        1 evaluated   5.4174114  13.2616490
#> 51: 2021-12-01 16:59:37        1 evaluated  -2.3978923   2.9768443
#> 52: 2021-12-01 16:59:37        1 evaluated   3.1116903   8.0510128
#> 53: 2021-12-01 16:59:37        1 evaluated   1.1416711  12.8359279
#> 54: 2021-12-01 16:59:37        1 evaluated   3.3325539   2.9746808
#> 55: 2021-12-01 16:59:37        1 evaluated  -2.3003639   6.8584964
#> 56: 2021-12-01 16:59:37        1 evaluated   1.6719916   1.5688385
#> 57: 2021-12-01 16:59:37        1 evaluated   2.7450744  12.5138203
#> 58: 2021-12-01 16:59:37        1 evaluated   7.0894608   5.9089132
#> 59: 2021-12-01 16:59:37        1 evaluated   3.5813040   7.2694230
#> 60: 2021-12-01 16:59:37        1 evaluated   0.2704535   6.6255357
#> 61: 2021-12-01 16:59:37        1 evaluated  -1.1674056   9.2720357
#> 62: 2021-12-01 16:59:37        1 evaluated   6.4942419   9.4896205
#> 63: 2021-12-01 16:59:37        1 evaluated  -4.6178751   0.2442206
#> 64: 2021-12-01 16:59:37        1 evaluated  -2.4662010   6.5008003
#>               timestamp batch_nr    status x_domain_x1 x_domain_x2
#>     x_domain_fidelity
#>  1:          0.015625
#>  2:          0.015625
#>  3:          0.015625
#>  4:          0.015625
#>  5:          0.015625
#>  6:          0.015625
#>  7:          0.015625
#>  8:          0.015625
#>  9:          0.015625
#> 10:          0.015625
#> 11:          0.015625
#> 12:          0.015625
#> 13:          0.015625
#> 14:          0.015625
#> 15:          0.015625
#> 16:          0.015625
#> 17:          0.015625
#> 18:          0.015625
#> 19:          0.015625
#> 20:          0.015625
#> 21:          0.015625
#> 22:          0.015625
#> 23:          0.015625
#> 24:          0.015625
#> 25:          0.015625
#> 26:          0.015625
#> 27:          0.015625
#> 28:          0.015625
#> 29:          0.015625
#> 30:          0.015625
#> 31:          0.015625
#> 32:          0.015625
#> 33:          0.015625
#> 34:          0.015625
#> 35:          0.015625
#> 36:          0.015625
#> 37:          0.015625
#> 38:          0.015625
#> 39:          0.015625
#> 40:          0.015625
#> 41:          0.015625
#> 42:          0.015625
#> 43:          0.015625
#> 44:          0.015625
#> 45:          0.015625
#> 46:          0.015625
#> 47:          0.015625
#> 48:          0.015625
#> 49:          0.015625
#> 50:          0.015625
#> 51:          0.015625
#> 52:          0.015625
#> 53:          0.015625
#> 54:          0.015625
#> 55:          0.015625
#> 56:          0.015625
#> 57:          0.015625
#> 58:          0.015625
#> 59:          0.015625
#> 60:          0.015625
#> 61:          0.015625
#> 62:          0.015625
#> 63:          0.015625
#> 64:          0.015625
#>     x_domain_fidelity