Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 2.935719 2.050617 0.015625 <list[3]> 0.8103913

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 2.935719 2.050617 0.015625 <list[3]> 0.8103913

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition           y
#>  1:  0.12465288 10.1626043 0.015625     0       6          1  38.5417241
#>  2: -4.44805305 13.3208043 0.015625     0       6          1   7.6258341
#>  3:  2.16254134  9.7973333 0.015625     0       6          1  54.9876047
#>  4: -4.80978674  0.8729496 0.015625     0       6          1 193.0046883
#>  5: -0.44496685  3.5116939 0.015625     0       6          1  28.9236290
#>  6: -3.70671367  1.2559888 0.015625     0       6          1 124.3445930
#>  7: -2.67764719  0.3473501 0.015625     0       6          1 104.1251851
#>  8:  0.59678287 14.5061841 0.015625     0       6          1 107.1512042
#>  9:  3.41906727  7.1615051 0.015625     0       6          1  39.7487865
#> 10:  5.21899606  9.3995130 0.015625     0       6          1 132.7794233
#> 11:  2.56500461  9.3325357 0.015625     0       6          1  53.9715232
#> 12:  4.38194007  7.3054194 0.015625     0       6          1  66.0068071
#> 13:  4.35517075 14.7300858 0.015625     0       6          1 233.9994059
#> 14:  8.31910110 11.3512192 0.015625     0       6          1 276.7419090
#> 15:  1.27570892  7.9438685 0.015625     0       6          1  28.1917369
#> 16: -2.42364263 13.9380762 0.015625     0       6          1  17.9790475
#> 17:  1.82495470  9.9982223 0.015625     0       6          1  53.8301695
#> 18:  2.51653760  5.6301949 0.015625     0       6          1  14.0515554
#> 19:  9.74824253  8.7172662 0.015625     0       6          1 235.2993676
#> 20: -4.35755110 12.8060730 0.015625     0       6          1   7.1722510
#> 21:  8.13635834  1.4337832 0.015625     0       6          1  47.6182498
#> 22:  0.60923673  1.3334938 0.015625     0       6          1  31.6259503
#> 23:  4.75364429 11.5373314 0.015625     0       6          1 164.3593791
#> 24:  2.86359259  5.4983980 0.015625     0       6          1  15.2355704
#> 25:  0.65032159 13.0337862 0.015625     0       6          1  82.5381571
#> 26:  2.86024626  9.5409836 0.015625     0       6          1  62.2671338
#> 27:  8.02593534  8.8578701 0.015625     0       6          1 194.7055699
#> 28: -2.95338608  6.2774277 0.015625     0       6          1  22.5750182
#> 29:  9.41135945  7.2930776 0.015625     0       6          1 183.9564162
#> 30:  6.20712591  4.0660639 0.015625     0       6          1  65.2770653
#> 31:  4.96372553  9.9117340 0.015625     0       6          1 134.5830642
#> 32:  2.93571921  2.0506168 0.015625     0       6          1   0.8103913
#> 33:  6.51327711  2.1802695 0.015625     0       6          1  46.8286489
#> 34:  3.56220577 14.8166654 0.015625     0       6          1 199.9301642
#> 35:  7.66272316  3.8080933 0.015625     0       6          1  79.0376587
#> 36: -0.69021955  2.0538915 0.015625     0       6          1  43.0005124
#> 37:  9.97412904 12.6491632 0.015625     0       6          1 380.6840953
#> 38: -4.59361829  4.9831421 0.015625     0       6          1  89.4421390
#> 39:  9.41579622 13.4022265 0.015625     0       6          1 386.9920189
#> 40:  4.68137460 13.4534970 0.015625     0       6          1 212.2038854
#> 41:  5.11768395  3.3691486 0.015625     0       6          1  35.9598510
#> 42:  8.34973978  0.9754679 0.015625     0       6          1  42.8972973
#> 43:  4.59761981 10.4559182 0.015625     0       6          1 132.6287234
#> 44:  6.59913432  4.4161257 0.015625     0       6          1  76.5830114
#> 45:  8.48363780 14.0996128 0.015625     0       6          1 380.2717652
#> 46: -4.29396109  7.8263444 0.015625     0       6          1  37.1748533
#> 47:  9.21471587  8.2593641 0.015625     0       6          1 205.5076002
#> 48:  9.49928115  4.7436805 0.015625     0       6          1 123.3629103
#> 49:  6.96916544 14.2922245 0.015625     0       6          1 337.5049616
#> 50: -2.34718652 12.3561287 0.015625     0       6          1   9.2794680
#> 51:  4.80077233  9.7808277 0.015625     0       6          1 125.6128270
#> 52: -4.33414410 12.6555853 0.015625     0       6          1   7.1264291
#> 53: -1.01865005  1.2123188 0.015625     0       6          1  56.5205917
#> 54: -2.25414186  5.2561261 0.015625     0       6          1  24.0766071
#> 55: -2.18231235 10.8191848 0.015625     0       6          1   5.9261240
#> 56: -4.54232818 12.0761493 0.015625     0       6          1  11.5703794
#> 57: -4.81124954 12.4282790 0.015625     0       6          1  14.7144215
#> 58: -0.01592695  3.4913151 0.015625     0       6          1  26.0222597
#> 59:  0.30155249 12.6865669 0.015625     0       6          1  70.4875256
#> 60: -3.17309499  5.3450904 0.015625     0       6          1  36.5783579
#> 61:  6.37112000  2.4576116 0.015625     0       6          1  48.1822369
#> 62:  1.21297867  0.3595077 0.015625     0       6          1  27.4645997
#> 63:  2.03682923  1.2179793 0.015625     0       6          1   8.4671068
#> 64:  6.32211928  0.4870055 0.015625     0       6          1  30.6174864
#>              x1         x2 fidelity stage bracket repetition           y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-01 04:48:50        1  0.12465288  10.1626043          0.015625
#>  2: 2022-10-01 04:48:50        1 -4.44805305  13.3208043          0.015625
#>  3: 2022-10-01 04:48:50        1  2.16254134   9.7973333          0.015625
#>  4: 2022-10-01 04:48:50        1 -4.80978674   0.8729496          0.015625
#>  5: 2022-10-01 04:48:50        1 -0.44496685   3.5116939          0.015625
#>  6: 2022-10-01 04:48:50        1 -3.70671367   1.2559888          0.015625
#>  7: 2022-10-01 04:48:50        1 -2.67764719   0.3473501          0.015625
#>  8: 2022-10-01 04:48:50        1  0.59678287  14.5061841          0.015625
#>  9: 2022-10-01 04:48:50        1  3.41906727   7.1615051          0.015625
#> 10: 2022-10-01 04:48:50        1  5.21899606   9.3995130          0.015625
#> 11: 2022-10-01 04:48:50        1  2.56500461   9.3325357          0.015625
#> 12: 2022-10-01 04:48:50        1  4.38194007   7.3054194          0.015625
#> 13: 2022-10-01 04:48:50        1  4.35517075  14.7300858          0.015625
#> 14: 2022-10-01 04:48:50        1  8.31910110  11.3512192          0.015625
#> 15: 2022-10-01 04:48:50        1  1.27570892   7.9438685          0.015625
#> 16: 2022-10-01 04:48:50        1 -2.42364263  13.9380762          0.015625
#> 17: 2022-10-01 04:48:50        1  1.82495470   9.9982223          0.015625
#> 18: 2022-10-01 04:48:50        1  2.51653760   5.6301949          0.015625
#> 19: 2022-10-01 04:48:50        1  9.74824253   8.7172662          0.015625
#> 20: 2022-10-01 04:48:50        1 -4.35755110  12.8060730          0.015625
#> 21: 2022-10-01 04:48:50        1  8.13635834   1.4337832          0.015625
#> 22: 2022-10-01 04:48:50        1  0.60923673   1.3334938          0.015625
#> 23: 2022-10-01 04:48:50        1  4.75364429  11.5373314          0.015625
#> 24: 2022-10-01 04:48:50        1  2.86359259   5.4983980          0.015625
#> 25: 2022-10-01 04:48:50        1  0.65032159  13.0337862          0.015625
#> 26: 2022-10-01 04:48:50        1  2.86024626   9.5409836          0.015625
#> 27: 2022-10-01 04:48:50        1  8.02593534   8.8578701          0.015625
#> 28: 2022-10-01 04:48:50        1 -2.95338608   6.2774277          0.015625
#> 29: 2022-10-01 04:48:50        1  9.41135945   7.2930776          0.015625
#> 30: 2022-10-01 04:48:50        1  6.20712591   4.0660639          0.015625
#> 31: 2022-10-01 04:48:50        1  4.96372553   9.9117340          0.015625
#> 32: 2022-10-01 04:48:50        1  2.93571921   2.0506168          0.015625
#> 33: 2022-10-01 04:48:50        1  6.51327711   2.1802695          0.015625
#> 34: 2022-10-01 04:48:50        1  3.56220577  14.8166654          0.015625
#> 35: 2022-10-01 04:48:50        1  7.66272316   3.8080933          0.015625
#> 36: 2022-10-01 04:48:50        1 -0.69021955   2.0538915          0.015625
#> 37: 2022-10-01 04:48:50        1  9.97412904  12.6491632          0.015625
#> 38: 2022-10-01 04:48:50        1 -4.59361829   4.9831421          0.015625
#> 39: 2022-10-01 04:48:50        1  9.41579622  13.4022265          0.015625
#> 40: 2022-10-01 04:48:50        1  4.68137460  13.4534970          0.015625
#> 41: 2022-10-01 04:48:50        1  5.11768395   3.3691486          0.015625
#> 42: 2022-10-01 04:48:50        1  8.34973978   0.9754679          0.015625
#> 43: 2022-10-01 04:48:50        1  4.59761981  10.4559182          0.015625
#> 44: 2022-10-01 04:48:50        1  6.59913432   4.4161257          0.015625
#> 45: 2022-10-01 04:48:50        1  8.48363780  14.0996128          0.015625
#> 46: 2022-10-01 04:48:50        1 -4.29396109   7.8263444          0.015625
#> 47: 2022-10-01 04:48:50        1  9.21471587   8.2593641          0.015625
#> 48: 2022-10-01 04:48:50        1  9.49928115   4.7436805          0.015625
#> 49: 2022-10-01 04:48:50        1  6.96916544  14.2922245          0.015625
#> 50: 2022-10-01 04:48:50        1 -2.34718652  12.3561287          0.015625
#> 51: 2022-10-01 04:48:50        1  4.80077233   9.7808277          0.015625
#> 52: 2022-10-01 04:48:50        1 -4.33414410  12.6555853          0.015625
#> 53: 2022-10-01 04:48:50        1 -1.01865005   1.2123188          0.015625
#> 54: 2022-10-01 04:48:50        1 -2.25414186   5.2561261          0.015625
#> 55: 2022-10-01 04:48:50        1 -2.18231235  10.8191848          0.015625
#> 56: 2022-10-01 04:48:50        1 -4.54232818  12.0761493          0.015625
#> 57: 2022-10-01 04:48:50        1 -4.81124954  12.4282790          0.015625
#> 58: 2022-10-01 04:48:50        1 -0.01592695   3.4913151          0.015625
#> 59: 2022-10-01 04:48:50        1  0.30155249  12.6865669          0.015625
#> 60: 2022-10-01 04:48:50        1 -3.17309499   5.3450904          0.015625
#> 61: 2022-10-01 04:48:50        1  6.37112000   2.4576116          0.015625
#> 62: 2022-10-01 04:48:50        1  1.21297867   0.3595077          0.015625
#> 63: 2022-10-01 04:48:50        1  2.03682923   1.2179793          0.015625
#> 64: 2022-10-01 04:48:50        1  6.32211928   0.4870055          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity