Optimizer using the Hyperband algorithm

OptimizerHyperband class that implements hyperband optimization. Hyperband is a budget oriented-procedure, weeding out suboptimal performing points early in a sequential training process, increasing optimization efficiency as a consequence.

For this, several brackets are constructed with an associated set of points for each bracket. Each bracket as several stages. Different brackets are initialized with different amounts of points and different budget sizes. To get an idea of how the bracket layout looks like for a given argument set, use hyperband_schedule().

To identify the budget for evaluating hyperband, the user has to specify explicitly which parameter of the objective function influences the budget by tagging a single parameter in the paradox::ParamSet with "budget".

Naturally, hyperband terminates once all of its brackets are evaluated, so a bbotk::Terminator in the bbotk::OptimInstanceSingleCrit | bbotk::OptimInstanceMultiCrit acts as an upper bound and should be only set to a low value if one is unsure of how long hyperband will take to finish under the given settings.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
Fraction parameter of the successive halving algorithm: With every stage the configuration budget is increased by a factor of eta and only the best 1/eta points are used for the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn during the initialization of each bracket. The default is uniform sampling.

repeats

logical(1)
If FALSE (default), hyperband terminates once all brackets are evaluated. Otherwise, hyperband starts over again once the last bracket is evaluated.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

Custom sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Runtime

The calculation of each bracket currently assumes a linear runtime in the chosen budget parameter is always given. Hyperband is designed so each bracket requires approximately the same runtime as the sum of the budget over all configurations in each bracket is roughly the same. This will not hold true once the scaling in the budget parameter is not linear anymore, even though the sum of the budgets in each bracket remains the same. A possible adaption would be to introduce a trafo, like it is shown in the examples section.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Parallelization

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. The points of one stage in a bracket are evaluated in one batch. Parallelization is supported via package future (see mlr3::benchmark()'s section on parallelization for more details).

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# modified branin function
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("none")
)

optimizer = opt("hyperband")

# modifies the instance by reference
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: -3.24339 11.97415      0.5 <list[3]> 0.4480316

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: -3.24339 11.97415      0.5 <list[3]> 0.4480316

# all evaluations
as.data.table(instance$archive)
#>              x1           x2 fidelity bracket stage         y
#>   1: -4.5024954  7.812216488 0.015625       6     0 43.724414
#>   2:  7.8963974  1.807365033 0.015625       6     0 51.294740
#>   3:  0.5375425  5.955704696 0.015625       6     0 18.891681
#>   4:  7.3959100  0.005407889 0.015625       6     0 31.011323
#>   5:  8.2615344  4.008464915 0.015625       6     0 88.251058
#>  ---                                                         
#> 297: -2.4128401  9.872479438 1.000000       2     2  3.354825
#> 298: -2.1771277  9.276767902 1.000000       2     2  5.169066
#> 299: -2.7912454 10.861012812 1.000000       1     1  1.326779
#> 300: -4.6011178 12.034890665 1.000000       1     1 25.117397
#> 301: -2.0763616 14.405963172 1.000000       1     1 26.001074
#>                timestamp batch_nr x_domain_x1  x_domain_x2 x_domain_fidelity
#>   1: 2022-01-30 04:25:59        1  -4.5024954  7.812216488          0.015625
#>   2: 2022-01-30 04:25:59        1   7.8963974  1.807365033          0.015625
#>   3: 2022-01-30 04:25:59        1   0.5375425  5.955704696          0.015625
#>   4: 2022-01-30 04:25:59        1   7.3959100  0.005407889          0.015625
#>   5: 2022-01-30 04:25:59        1   8.2615344  4.008464915          0.015625
#>  ---                                                                        
#> 297: 2022-01-30 04:26:00        7  -2.4128401  9.872479438          1.000000
#> 298: 2022-01-30 04:26:00        7  -2.1771277  9.276767902          1.000000
#> 299: 2022-01-30 04:26:00        7  -2.7912454 10.861012812          1.000000
#> 300: 2022-01-30 04:26:00        7  -4.6011178 12.034890665          1.000000
#> 301: 2022-01-30 04:26:00        7  -2.0763616 14.405963172          1.000000