Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.981926 11.97182 0.015625 <list[3]> 1.427356

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.981926 11.97182 0.015625 <list[3]> 1.427356

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  5.89752323  8.4602300 0.015625     0       6          1 135.041096
#>  2:  3.04304291  4.8570020 0.015625     0       6          1  12.109670
#>  3:  3.43431900  7.1125688 0.015625     0       6          1  39.442651
#>  4:  7.25900501  8.4297920 0.015625     0       6          1 168.218244
#>  5:  0.51497418  9.3868354 0.015625     0       6          1  35.982395
#>  6: -3.01170200 10.0019889 0.015625     0       6          1   1.624038
#>  7:  5.55599923 11.9860711 0.015625     0       6          1 209.816123
#>  8: -3.24931358  6.5283632 0.015625     0       6          1  25.131667
#>  9: -2.83042051  2.9223321 0.015625     0       6          1  62.148274
#> 10:  1.98665321  3.6537409 0.015625     0       6          1   6.602974
#> 11:  7.92477041  7.0434754 0.015625     0       6          1 146.800364
#> 12:  4.44790790 13.8913299 0.015625     0       6          1 213.759840
#> 13:  5.78340638 13.0216406 0.015625     0       6          1 249.400737
#> 14:  3.54047981  7.9400032 0.015625     0       6          1  52.839733
#> 15:  8.04562949  5.7566070 0.015625     0       6          1 119.923487
#> 16:  6.63668834  1.3846999 0.015625     0       6          1  40.104445
#> 17:  4.49587176  8.0839953 0.015625     0       6          1  82.205597
#> 18: -2.97819763  0.7574753 0.015625     0       6          1 105.694683
#> 19: -2.28137257 14.5799951 0.015625     0       6          1  26.671824
#> 20:  9.10597508  1.2246644 0.015625     0       6          1  52.258642
#> 21:  4.09428535 10.8281771 0.015625     0       6          1 121.703278
#> 22:  6.99652924  5.0816329 0.015625     0       6          1  93.157153
#> 23:  7.58269542 11.5563039 0.015625     0       6          1 264.007049
#> 24: -2.57251059  4.6313395 0.015625     0       6          1  34.019492
#> 25:  0.05369033 10.6425778 0.015625     0       6          1  41.941693
#> 26:  9.05867182  0.6023125 0.015625     0       6          1  43.239413
#> 27:  2.57396974  0.8743489 0.015625     0       6          1   3.423374
#> 28: -2.51313866  9.0541133 0.015625     0       6          1   3.531785
#> 29:  9.83551246  6.1722285 0.015625     0       6          1 166.358853
#> 30:  1.61626758  5.1987687 0.015625     0       6          1  12.422394
#> 31:  4.51351116  2.4796833 0.015625     0       6          1  17.324998
#> 32:  5.78625834  0.2495910 0.015625     0       6          1  24.342121
#> 33: -2.42557120 13.4039772 0.015625     0       6          1  14.063453
#> 34:  8.80629164  3.5008641 0.015625     0       6          1  85.571099
#> 35: -2.42815857  4.4350144 0.015625     0       6          1  34.220770
#> 36:  9.06806182 10.9796581 0.015625     0       6          1 286.058572
#> 37:  5.56427056 14.0262715 0.015625     0       6          1 270.994569
#> 38:  9.09824877  8.6227535 0.015625     0       6          1 212.837302
#> 39: -3.45769557 14.2372322 0.015625     0       6          1   6.474135
#> 40:  9.08461619 14.9959935 0.015625     0       6          1 438.471097
#> 41:  7.24474265 12.2638406 0.015625     0       6          1 277.300414
#> 42: -0.71411987 12.2772699 0.015625     0       6          1  43.521990
#> 43:  4.90278053  5.9569011 0.015625     0       6          1  61.109399
#> 44:  4.90930980 10.6405333 0.015625     0       6          1 149.070661
#> 45: -2.81078202  5.3010452 0.015625     0       6          1  30.244752
#> 46:  2.46799362 12.5653811 0.015625     0       6          1 108.709525
#> 47: -0.71001482 14.7544129 0.015625     0       6          1  75.176975
#> 48: -1.47967521  6.2680382 0.015625     0       6          1  15.514567
#> 49: -4.24252021  9.7076627 0.015625     0       6          1  18.597578
#> 50:  4.33441109  1.3809886 0.015625     0       6          1   9.352441
#> 51:  9.01665614  9.4020976 0.015625     0       6          1 233.834956
#> 52:  3.75617328  9.0020348 0.015625     0       6          1  75.195293
#> 53: -1.39207794  7.5699638 0.015625     0       6          1  12.204234
#> 54:  2.19640466  9.7239109 0.015625     0       6          1  54.379924
#> 55: -2.98192595 11.9718216 0.015625     0       6          1   1.427356
#> 56:  7.11670916  8.0526245 0.015625     0       6          1 156.213920
#> 57:  8.73109354  7.0722555 0.015625     0       6          1 161.990350
#> 58:  4.00446208  8.3717482 0.015625     0       6          1  71.851586
#> 59: -3.44301586  9.0777176 0.015625     0       6          1   8.484313
#> 60:  8.83622338 14.1695619 0.015625     0       6          1 395.327961
#> 61: -1.51347833 14.8267050 0.015625     0       6          1  50.840826
#> 62: -1.41789249  8.4304814 0.015625     0       6          1  11.475034
#> 63:  7.52998207 13.1169938 0.015625     0       6          1 314.355628
#> 64: -3.31099422  9.4048548 0.015625     0       6          1   5.383371
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-09-10 04:41:51        1  5.89752323   8.4602300          0.015625
#>  2: 2022-09-10 04:41:51        1  3.04304291   4.8570020          0.015625
#>  3: 2022-09-10 04:41:51        1  3.43431900   7.1125688          0.015625
#>  4: 2022-09-10 04:41:51        1  7.25900501   8.4297920          0.015625
#>  5: 2022-09-10 04:41:51        1  0.51497418   9.3868354          0.015625
#>  6: 2022-09-10 04:41:51        1 -3.01170200  10.0019889          0.015625
#>  7: 2022-09-10 04:41:51        1  5.55599923  11.9860711          0.015625
#>  8: 2022-09-10 04:41:51        1 -3.24931358   6.5283632          0.015625
#>  9: 2022-09-10 04:41:51        1 -2.83042051   2.9223321          0.015625
#> 10: 2022-09-10 04:41:51        1  1.98665321   3.6537409          0.015625
#> 11: 2022-09-10 04:41:51        1  7.92477041   7.0434754          0.015625
#> 12: 2022-09-10 04:41:51        1  4.44790790  13.8913299          0.015625
#> 13: 2022-09-10 04:41:51        1  5.78340638  13.0216406          0.015625
#> 14: 2022-09-10 04:41:51        1  3.54047981   7.9400032          0.015625
#> 15: 2022-09-10 04:41:51        1  8.04562949   5.7566070          0.015625
#> 16: 2022-09-10 04:41:51        1  6.63668834   1.3846999          0.015625
#> 17: 2022-09-10 04:41:51        1  4.49587176   8.0839953          0.015625
#> 18: 2022-09-10 04:41:51        1 -2.97819763   0.7574753          0.015625
#> 19: 2022-09-10 04:41:51        1 -2.28137257  14.5799951          0.015625
#> 20: 2022-09-10 04:41:51        1  9.10597508   1.2246644          0.015625
#> 21: 2022-09-10 04:41:51        1  4.09428535  10.8281771          0.015625
#> 22: 2022-09-10 04:41:51        1  6.99652924   5.0816329          0.015625
#> 23: 2022-09-10 04:41:51        1  7.58269542  11.5563039          0.015625
#> 24: 2022-09-10 04:41:51        1 -2.57251059   4.6313395          0.015625
#> 25: 2022-09-10 04:41:51        1  0.05369033  10.6425778          0.015625
#> 26: 2022-09-10 04:41:51        1  9.05867182   0.6023125          0.015625
#> 27: 2022-09-10 04:41:51        1  2.57396974   0.8743489          0.015625
#> 28: 2022-09-10 04:41:51        1 -2.51313866   9.0541133          0.015625
#> 29: 2022-09-10 04:41:51        1  9.83551246   6.1722285          0.015625
#> 30: 2022-09-10 04:41:51        1  1.61626758   5.1987687          0.015625
#> 31: 2022-09-10 04:41:51        1  4.51351116   2.4796833          0.015625
#> 32: 2022-09-10 04:41:51        1  5.78625834   0.2495910          0.015625
#> 33: 2022-09-10 04:41:51        1 -2.42557120  13.4039772          0.015625
#> 34: 2022-09-10 04:41:51        1  8.80629164   3.5008641          0.015625
#> 35: 2022-09-10 04:41:51        1 -2.42815857   4.4350144          0.015625
#> 36: 2022-09-10 04:41:51        1  9.06806182  10.9796581          0.015625
#> 37: 2022-09-10 04:41:51        1  5.56427056  14.0262715          0.015625
#> 38: 2022-09-10 04:41:51        1  9.09824877   8.6227535          0.015625
#> 39: 2022-09-10 04:41:51        1 -3.45769557  14.2372322          0.015625
#> 40: 2022-09-10 04:41:51        1  9.08461619  14.9959935          0.015625
#> 41: 2022-09-10 04:41:51        1  7.24474265  12.2638406          0.015625
#> 42: 2022-09-10 04:41:51        1 -0.71411987  12.2772699          0.015625
#> 43: 2022-09-10 04:41:51        1  4.90278053   5.9569011          0.015625
#> 44: 2022-09-10 04:41:51        1  4.90930980  10.6405333          0.015625
#> 45: 2022-09-10 04:41:51        1 -2.81078202   5.3010452          0.015625
#> 46: 2022-09-10 04:41:51        1  2.46799362  12.5653811          0.015625
#> 47: 2022-09-10 04:41:51        1 -0.71001482  14.7544129          0.015625
#> 48: 2022-09-10 04:41:51        1 -1.47967521   6.2680382          0.015625
#> 49: 2022-09-10 04:41:51        1 -4.24252021   9.7076627          0.015625
#> 50: 2022-09-10 04:41:51        1  4.33441109   1.3809886          0.015625
#> 51: 2022-09-10 04:41:51        1  9.01665614   9.4020976          0.015625
#> 52: 2022-09-10 04:41:51        1  3.75617328   9.0020348          0.015625
#> 53: 2022-09-10 04:41:51        1 -1.39207794   7.5699638          0.015625
#> 54: 2022-09-10 04:41:51        1  2.19640466   9.7239109          0.015625
#> 55: 2022-09-10 04:41:51        1 -2.98192595  11.9718216          0.015625
#> 56: 2022-09-10 04:41:51        1  7.11670916   8.0526245          0.015625
#> 57: 2022-09-10 04:41:51        1  8.73109354   7.0722555          0.015625
#> 58: 2022-09-10 04:41:51        1  4.00446208   8.3717482          0.015625
#> 59: 2022-09-10 04:41:51        1 -3.44301586   9.0777176          0.015625
#> 60: 2022-09-10 04:41:51        1  8.83622338  14.1695619          0.015625
#> 61: 2022-09-10 04:41:51        1 -1.51347833  14.8267050          0.015625
#> 62: 2022-09-10 04:41:51        1 -1.41789249   8.4304814          0.015625
#> 63: 2022-09-10 04:41:51        1  7.52998207  13.1169938          0.015625
#> 64: 2022-09-10 04:41:51        1 -3.31099422   9.4048548          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity