Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.411169 12.21385 0.015625 <list[3]> 0.927039

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.411169 12.21385 0.015625 <list[3]> 0.927039

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  3.29308585  8.2852104 0.015625     0       6          1  52.245470
#>  2:  2.63984643 11.0646342 0.015625     0       6          1  83.516685
#>  3:  8.68723651  7.4597405 0.015625     0       6          1 170.997153
#>  4:  6.82654386 14.3702109 0.015625     0       6          1 335.135020
#>  5:  0.81123339  8.3718039 0.015625     0       6          1  29.881230
#>  6: -4.67662190  4.2557411 0.015625     0       6          1 106.872197
#>  7: -2.75737986  8.3880308 0.015625     0       6          1   6.089812
#>  8:  1.28747637  0.9012054 0.015625     0       6          1  22.298427
#>  9: -3.04392826  4.2983384 0.015625     0       6          1  47.107713
#> 10:  1.47235664 10.4498720 0.015625     0       6          1  56.190123
#> 11:  3.72042590 12.9314976 0.015625     0       6          1 156.396175
#> 12:  8.32663322  9.8968960 0.015625     0       6          1 231.149976
#> 13:  8.85133818  2.8200384 0.015625     0       6          1  74.157831
#> 14:  3.93756073 10.6828954 0.015625     0       6          1 112.966248
#> 15:  4.10233942  9.1405397 0.015625     0       6          1  88.261060
#> 16:  9.75428422  8.4558200 0.015625     0       6          1 227.561122
#> 17:  0.29049764 13.3604071 0.015625     0       6          1  80.354603
#> 18:  8.84988530  5.3254023 0.015625     0       6          1 122.992291
#> 19:  0.02294468 12.1157673 0.015625     0       6          1  57.449995
#> 20:  8.55273287  7.2270568 0.015625     0       6          1 162.332270
#> 21:  7.22430388 13.9323897 0.015625     0       6          1 333.403789
#> 22:  5.52552289 12.5415770 0.015625     0       6          1 224.248160
#> 23: -1.41021681  7.5970049 0.015625     0       6          1  12.037359
#> 24:  1.82124996  3.7314659 0.015625     0       6          1   7.899057
#> 25: -0.12213099  9.9805314 0.015625     0       6          1  33.862078
#> 26:  3.64968943  5.4266160 0.015625     0       6          1  24.898474
#> 27: -1.84461257  8.0421971 0.015625     0       6          1   8.399949
#> 28:  3.25686774  9.2791540 0.015625     0       6          1  66.663934
#> 29:  8.04408101  4.4541494 0.015625     0       6          1  94.065648
#> 30:  7.64307565 10.3241686 0.015625     0       6          1 227.875765
#> 31: -1.29501817 11.6476910 0.015625     0       6          1  25.111127
#> 32: -3.31812832  0.5788265 0.015625     0       6          1 122.443353
#> 33:  9.93798790 12.7481496 0.015625     0       6          1 382.987961
#> 34:  3.44485345  4.0439519 0.015625     0       6          1  10.832597
#> 35:  5.04300141  7.2085182 0.015625     0       6          1  84.566051
#> 36: -3.31526328  8.9690489 0.015625     0       6          1   7.539921
#> 37: -1.41003653  6.9208042 0.015625     0       6          1  13.453750
#> 38: -1.37194440 11.3450442 0.015625     0       6          1  21.529507
#> 39: -0.82730858 12.7035162 0.015625     0       6          1  45.290760
#> 40:  3.74310295  6.1851839 0.015625     0       6          1  34.707052
#> 41: -3.41116881 12.2138549 0.015625     0       6          1   0.927039
#> 42: -2.69583394 12.6266987 0.015625     0       6          1   5.799633
#> 43:  4.94818405  6.8957333 0.015625     0       6          1  76.534529
#> 44:  2.59198747 12.9578620 0.015625     0       6          1 120.111703
#> 45: -2.10413170 14.2061235 0.015625     0       6          1  27.407610
#> 46:  5.15944276 12.0320448 0.015625     0       6          1 194.383661
#> 47:  4.57681442  5.4446457 0.015625     0       6          1  45.727063
#> 48:  1.90492736  0.3684373 0.015625     0       6          1  14.202420
#> 49: -1.21529972 14.9952923 0.015625     0       6          1  62.561710
#> 50:  8.64575309  1.1404870 0.015625     0       6          1  46.757786
#> 51:  6.34368026  7.8679066 0.015625     0       6          1 134.650021
#> 52:  3.34265713  3.7749548 0.015625     0       6          1   8.161592
#> 53:  8.53340161  8.3215511 0.015625     0       6          1 190.669213
#> 54: -0.03014706 14.7396836 0.015625     0       6          1  95.142966
#> 55:  6.42288812  6.4224185 0.015625     0       6          1 107.424333
#> 56: -0.96087017  0.8820906 0.015625     0       6          1  60.063386
#> 57:  2.05625653  1.2011261 0.015625     0       6          1   8.262651
#> 58:  6.84858230  6.3669593 0.015625     0       6          1 114.632271
#> 59: -2.20960763  4.9395788 0.015625     0       6          1  26.621611
#> 60: -0.88562939  1.8252719 0.015625     0       6          1  47.530020
#> 61:  9.72080826 11.7823317 0.015625     0       6          1 337.467129
#> 62:  7.06429474  9.5478487 0.015625     0       6          1 192.556649
#> 63:  3.45222479  1.6007178 0.015625     0       6          1   1.388393
#> 64:  2.68345057 14.7786562 0.015625     0       6          1 165.948094
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-10-04 04:36:42        1  3.29308585   8.2852104          0.015625
#>  2: 2022-10-04 04:36:42        1  2.63984643  11.0646342          0.015625
#>  3: 2022-10-04 04:36:42        1  8.68723651   7.4597405          0.015625
#>  4: 2022-10-04 04:36:42        1  6.82654386  14.3702109          0.015625
#>  5: 2022-10-04 04:36:42        1  0.81123339   8.3718039          0.015625
#>  6: 2022-10-04 04:36:42        1 -4.67662190   4.2557411          0.015625
#>  7: 2022-10-04 04:36:42        1 -2.75737986   8.3880308          0.015625
#>  8: 2022-10-04 04:36:42        1  1.28747637   0.9012054          0.015625
#>  9: 2022-10-04 04:36:42        1 -3.04392826   4.2983384          0.015625
#> 10: 2022-10-04 04:36:42        1  1.47235664  10.4498720          0.015625
#> 11: 2022-10-04 04:36:42        1  3.72042590  12.9314976          0.015625
#> 12: 2022-10-04 04:36:42        1  8.32663322   9.8968960          0.015625
#> 13: 2022-10-04 04:36:42        1  8.85133818   2.8200384          0.015625
#> 14: 2022-10-04 04:36:42        1  3.93756073  10.6828954          0.015625
#> 15: 2022-10-04 04:36:42        1  4.10233942   9.1405397          0.015625
#> 16: 2022-10-04 04:36:42        1  9.75428422   8.4558200          0.015625
#> 17: 2022-10-04 04:36:42        1  0.29049764  13.3604071          0.015625
#> 18: 2022-10-04 04:36:42        1  8.84988530   5.3254023          0.015625
#> 19: 2022-10-04 04:36:42        1  0.02294468  12.1157673          0.015625
#> 20: 2022-10-04 04:36:42        1  8.55273287   7.2270568          0.015625
#> 21: 2022-10-04 04:36:42        1  7.22430388  13.9323897          0.015625
#> 22: 2022-10-04 04:36:42        1  5.52552289  12.5415770          0.015625
#> 23: 2022-10-04 04:36:42        1 -1.41021681   7.5970049          0.015625
#> 24: 2022-10-04 04:36:42        1  1.82124996   3.7314659          0.015625
#> 25: 2022-10-04 04:36:42        1 -0.12213099   9.9805314          0.015625
#> 26: 2022-10-04 04:36:42        1  3.64968943   5.4266160          0.015625
#> 27: 2022-10-04 04:36:42        1 -1.84461257   8.0421971          0.015625
#> 28: 2022-10-04 04:36:42        1  3.25686774   9.2791540          0.015625
#> 29: 2022-10-04 04:36:42        1  8.04408101   4.4541494          0.015625
#> 30: 2022-10-04 04:36:42        1  7.64307565  10.3241686          0.015625
#> 31: 2022-10-04 04:36:42        1 -1.29501817  11.6476910          0.015625
#> 32: 2022-10-04 04:36:42        1 -3.31812832   0.5788265          0.015625
#> 33: 2022-10-04 04:36:42        1  9.93798790  12.7481496          0.015625
#> 34: 2022-10-04 04:36:42        1  3.44485345   4.0439519          0.015625
#> 35: 2022-10-04 04:36:42        1  5.04300141   7.2085182          0.015625
#> 36: 2022-10-04 04:36:42        1 -3.31526328   8.9690489          0.015625
#> 37: 2022-10-04 04:36:42        1 -1.41003653   6.9208042          0.015625
#> 38: 2022-10-04 04:36:42        1 -1.37194440  11.3450442          0.015625
#> 39: 2022-10-04 04:36:42        1 -0.82730858  12.7035162          0.015625
#> 40: 2022-10-04 04:36:42        1  3.74310295   6.1851839          0.015625
#> 41: 2022-10-04 04:36:42        1 -3.41116881  12.2138549          0.015625
#> 42: 2022-10-04 04:36:42        1 -2.69583394  12.6266987          0.015625
#> 43: 2022-10-04 04:36:42        1  4.94818405   6.8957333          0.015625
#> 44: 2022-10-04 04:36:42        1  2.59198747  12.9578620          0.015625
#> 45: 2022-10-04 04:36:42        1 -2.10413170  14.2061235          0.015625
#> 46: 2022-10-04 04:36:42        1  5.15944276  12.0320448          0.015625
#> 47: 2022-10-04 04:36:42        1  4.57681442   5.4446457          0.015625
#> 48: 2022-10-04 04:36:42        1  1.90492736   0.3684373          0.015625
#> 49: 2022-10-04 04:36:42        1 -1.21529972  14.9952923          0.015625
#> 50: 2022-10-04 04:36:42        1  8.64575309   1.1404870          0.015625
#> 51: 2022-10-04 04:36:42        1  6.34368026   7.8679066          0.015625
#> 52: 2022-10-04 04:36:42        1  3.34265713   3.7749548          0.015625
#> 53: 2022-10-04 04:36:42        1  8.53340161   8.3215511          0.015625
#> 54: 2022-10-04 04:36:42        1 -0.03014706  14.7396836          0.015625
#> 55: 2022-10-04 04:36:42        1  6.42288812   6.4224185          0.015625
#> 56: 2022-10-04 04:36:42        1 -0.96087017   0.8820906          0.015625
#> 57: 2022-10-04 04:36:42        1  2.05625653   1.2011261          0.015625
#> 58: 2022-10-04 04:36:42        1  6.84858230   6.3669593          0.015625
#> 59: 2022-10-04 04:36:42        1 -2.20960763   4.9395788          0.015625
#> 60: 2022-10-04 04:36:42        1 -0.88562939   1.8252719          0.015625
#> 61: 2022-10-04 04:36:42        1  9.72080826  11.7823317          0.015625
#> 62: 2022-10-04 04:36:42        1  7.06429474   9.5478487          0.015625
#> 63: 2022-10-04 04:36:42        1  3.45222479   1.6007178          0.015625
#> 64: 2022-10-04 04:36:42        1  2.68345057  14.7786562          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity