Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.894542 10.25382 0.015625 <list[3]> 1.062251

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -2.894542 10.25382 0.015625 <list[3]> 1.062251

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  7.28344731  9.5472085 0.015625     0       6          1 197.654395
#>  2:  7.60606641  7.9321231 0.015625     0       6          1 162.633647
#>  3: -0.04717513 13.9011731 0.015625     0       6          1  80.838068
#>  4:  2.59637829 14.9693867 0.015625     0       6          1 168.055073
#>  5: -0.28771886 12.8296325 0.015625     0       6          1  59.773712
#>  6:  3.39199848  3.0356466 0.015625     0       6          1   5.025487
#>  7:  9.53848868  1.3518448 0.015625     0       6          1  60.295896
#>  8: -1.52012102  8.2380976 0.015625     0       6          1  10.550037
#>  9:  2.89742431  0.4200557 0.015625     0       6          1   2.187426
#> 10:  5.91125472  0.5563958 0.015625     0       6          1  27.298034
#> 11:  0.40821126  4.9131524 0.015625     0       6          1  19.008744
#> 12: -3.50553972  0.3446773 0.015625     0       6          1 135.874821
#> 13: -4.43766963  6.0403609 0.015625     0       6          1  65.580147
#> 14:  4.82696308  0.6391389 0.015625     0       6          1  13.674084
#> 15: -4.17685764  8.7585947 0.015625     0       6          1  24.685087
#> 16: -4.02205496  8.2124555 0.015625     0       6          1  25.846232
#> 17: -2.15427553 14.8674730 0.015625     0       6          1  32.759032
#> 18: -3.66345704  8.3062884 0.015625     0       6          1  17.175490
#> 19:  7.63552537  8.3718364 0.015625     0       6          1 174.173706
#> 20:  8.05915737 13.5993427 0.015625     0       6          1 347.666968
#> 21:  9.14674709  3.8529596 0.015625     0       6          1  97.554314
#> 22:  8.98783645 13.0440217 0.015625     0       6          1 357.181710
#> 23:  3.34894467 14.7511239 0.015625     0       6          1 189.289340
#> 24:  1.76777455 13.0220512 0.015625     0       6          1 102.978010
#> 25:  4.63937341  6.9582361 0.015625     0       6          1  68.285547
#> 26:  5.44814723 13.2889976 0.015625     0       6          1 242.867227
#> 27:  4.14275167  3.1785960 0.015625     0       6          1  15.346793
#> 28:  3.97047058  2.5414076 0.015625     0       6          1   9.156685
#> 29:  9.80635695  2.1013753 0.015625     0       6          1  77.684284
#> 30: -2.89454151 10.2538241 0.015625     0       6          1   1.062251
#> 31: -0.43863457  5.0784473 0.015625     0       6          1  21.335605
#> 32:  6.92235178  1.3642585 0.015625     0       6          1  41.796615
#> 33:  7.42968140  6.8396759 0.015625     0       6          1 134.231211
#> 34: -0.90301199  5.3361685 0.015625     0       6          1  20.466353
#> 35:  9.18461771  9.0082939 0.015625     0       6          1 226.644574
#> 36:  8.78812517  7.6272575 0.015625     0       6          1 177.559859
#> 37:  3.41292819 13.4467315 0.015625     0       6          1 157.510405
#> 38:  3.00169390  5.0895786 0.015625     0       6          1  13.378988
#> 39:  0.29724356  1.3967701 0.015625     0       6          1  36.261639
#> 40: -2.15778702  6.3710793 0.015625     0       6          1  14.962184
#> 41:  7.20347477  4.4300511 0.015625     0       6          1  84.692882
#> 42: -1.08797674 14.9937817 0.015625     0       6          1  66.670496
#> 43: -4.86149337  6.7541238 0.015625     0       6          1  70.868385
#> 44:  6.73681363 13.1728514 0.015625     0       6          1 290.860549
#> 45: -1.40231964  3.0900854 0.015625     0       6          1  38.673387
#> 46:  8.33297835 14.4643910 0.015625     0       6          1 389.409627
#> 47:  9.03234518  5.0581179 0.015625     0       6          1 120.485647
#> 48:  9.57479992  4.7869278 0.015625     0       6          1 126.100526
#> 49:  5.10988220 10.6050344 0.015625     0       6          1 156.157283
#> 50:  6.61355401  8.7239953 0.015625     0       6          1 160.810622
#> 51:  6.15415617 14.7355171 0.015625     0       6          1 321.088433
#> 52:  0.60951564  1.6864259 0.015625     0       6          1  29.128522
#> 53: -2.30082887 13.2041787 0.015625     0       6          1  15.017612
#> 54:  4.00250633 13.8607403 0.015625     0       6          1 192.484357
#> 55:  8.08429372  1.0941615 0.015625     0       6          1  43.225024
#> 56:  5.16805484  4.7623058 0.015625     0       6          1  52.248828
#> 57: -4.73294261  6.0307251 0.015625     0       6          1  77.285614
#> 58:  6.99515665  7.5254188 0.015625     0       6          1 141.682066
#> 59: -1.04507217  9.0262444 0.015625     0       6          1  16.585969
#> 60:  7.53566299  3.9220634 0.015625     0       6          1  79.744878
#> 61: -2.00179875 10.3104260 0.015625     0       6          1   6.990924
#> 62: -4.69122871 14.0681486 0.015625     0       6          1   9.802433
#> 63:  9.40498934  4.6794656 0.015625     0       6          1 119.827082
#> 64:  8.72935638  9.7158153 0.015625     0       6          1 235.679761
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-09-18 04:42:19        1  7.28344731   9.5472085          0.015625
#>  2: 2022-09-18 04:42:19        1  7.60606641   7.9321231          0.015625
#>  3: 2022-09-18 04:42:19        1 -0.04717513  13.9011731          0.015625
#>  4: 2022-09-18 04:42:19        1  2.59637829  14.9693867          0.015625
#>  5: 2022-09-18 04:42:19        1 -0.28771886  12.8296325          0.015625
#>  6: 2022-09-18 04:42:19        1  3.39199848   3.0356466          0.015625
#>  7: 2022-09-18 04:42:19        1  9.53848868   1.3518448          0.015625
#>  8: 2022-09-18 04:42:19        1 -1.52012102   8.2380976          0.015625
#>  9: 2022-09-18 04:42:19        1  2.89742431   0.4200557          0.015625
#> 10: 2022-09-18 04:42:19        1  5.91125472   0.5563958          0.015625
#> 11: 2022-09-18 04:42:19        1  0.40821126   4.9131524          0.015625
#> 12: 2022-09-18 04:42:19        1 -3.50553972   0.3446773          0.015625
#> 13: 2022-09-18 04:42:19        1 -4.43766963   6.0403609          0.015625
#> 14: 2022-09-18 04:42:19        1  4.82696308   0.6391389          0.015625
#> 15: 2022-09-18 04:42:19        1 -4.17685764   8.7585947          0.015625
#> 16: 2022-09-18 04:42:19        1 -4.02205496   8.2124555          0.015625
#> 17: 2022-09-18 04:42:19        1 -2.15427553  14.8674730          0.015625
#> 18: 2022-09-18 04:42:19        1 -3.66345704   8.3062884          0.015625
#> 19: 2022-09-18 04:42:19        1  7.63552537   8.3718364          0.015625
#> 20: 2022-09-18 04:42:19        1  8.05915737  13.5993427          0.015625
#> 21: 2022-09-18 04:42:19        1  9.14674709   3.8529596          0.015625
#> 22: 2022-09-18 04:42:19        1  8.98783645  13.0440217          0.015625
#> 23: 2022-09-18 04:42:19        1  3.34894467  14.7511239          0.015625
#> 24: 2022-09-18 04:42:19        1  1.76777455  13.0220512          0.015625
#> 25: 2022-09-18 04:42:19        1  4.63937341   6.9582361          0.015625
#> 26: 2022-09-18 04:42:19        1  5.44814723  13.2889976          0.015625
#> 27: 2022-09-18 04:42:19        1  4.14275167   3.1785960          0.015625
#> 28: 2022-09-18 04:42:19        1  3.97047058   2.5414076          0.015625
#> 29: 2022-09-18 04:42:19        1  9.80635695   2.1013753          0.015625
#> 30: 2022-09-18 04:42:19        1 -2.89454151  10.2538241          0.015625
#> 31: 2022-09-18 04:42:19        1 -0.43863457   5.0784473          0.015625
#> 32: 2022-09-18 04:42:19        1  6.92235178   1.3642585          0.015625
#> 33: 2022-09-18 04:42:19        1  7.42968140   6.8396759          0.015625
#> 34: 2022-09-18 04:42:19        1 -0.90301199   5.3361685          0.015625
#> 35: 2022-09-18 04:42:19        1  9.18461771   9.0082939          0.015625
#> 36: 2022-09-18 04:42:19        1  8.78812517   7.6272575          0.015625
#> 37: 2022-09-18 04:42:19        1  3.41292819  13.4467315          0.015625
#> 38: 2022-09-18 04:42:19        1  3.00169390   5.0895786          0.015625
#> 39: 2022-09-18 04:42:19        1  0.29724356   1.3967701          0.015625
#> 40: 2022-09-18 04:42:19        1 -2.15778702   6.3710793          0.015625
#> 41: 2022-09-18 04:42:19        1  7.20347477   4.4300511          0.015625
#> 42: 2022-09-18 04:42:19        1 -1.08797674  14.9937817          0.015625
#> 43: 2022-09-18 04:42:19        1 -4.86149337   6.7541238          0.015625
#> 44: 2022-09-18 04:42:19        1  6.73681363  13.1728514          0.015625
#> 45: 2022-09-18 04:42:19        1 -1.40231964   3.0900854          0.015625
#> 46: 2022-09-18 04:42:19        1  8.33297835  14.4643910          0.015625
#> 47: 2022-09-18 04:42:19        1  9.03234518   5.0581179          0.015625
#> 48: 2022-09-18 04:42:19        1  9.57479992   4.7869278          0.015625
#> 49: 2022-09-18 04:42:19        1  5.10988220  10.6050344          0.015625
#> 50: 2022-09-18 04:42:19        1  6.61355401   8.7239953          0.015625
#> 51: 2022-09-18 04:42:19        1  6.15415617  14.7355171          0.015625
#> 52: 2022-09-18 04:42:19        1  0.60951564   1.6864259          0.015625
#> 53: 2022-09-18 04:42:19        1 -2.30082887  13.2041787          0.015625
#> 54: 2022-09-18 04:42:19        1  4.00250633  13.8607403          0.015625
#> 55: 2022-09-18 04:42:19        1  8.08429372   1.0941615          0.015625
#> 56: 2022-09-18 04:42:19        1  5.16805484   4.7623058          0.015625
#> 57: 2022-09-18 04:42:19        1 -4.73294261   6.0307251          0.015625
#> 58: 2022-09-18 04:42:19        1  6.99515665   7.5254188          0.015625
#> 59: 2022-09-18 04:42:19        1 -1.04507217   9.0262444          0.015625
#> 60: 2022-09-18 04:42:19        1  7.53566299   3.9220634          0.015625
#> 61: 2022-09-18 04:42:19        1 -2.00179875  10.3104260          0.015625
#> 62: 2022-09-18 04:42:19        1 -4.69122871  14.0681486          0.015625
#> 63: 2022-09-18 04:42:19        1  9.40498934   4.6794656          0.015625
#> 64: 2022-09-18 04:42:19        1  8.72935638   9.7158153          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity