Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.540321 14.34193 0.015625 <list[3]> 6.542581

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.540321 14.34193 0.015625 <list[3]> 6.542581

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1: -0.23426022 10.9190335 0.015625     0       6          1  39.992410
#>  2: -2.17312764 13.5391463 0.015625     0       6          1  20.046396
#>  3: -4.43218681  2.7419959 0.015625     0       6          1 126.504715
#>  4: -4.25177823  4.2636400 0.015625     0       6          1  87.799343
#>  5:  9.12641623 12.4599018 0.015625     0       6          1 340.268864
#>  6:  8.87869500  6.4676804 0.015625     0       6          1 150.018611
#>  7:  8.52117030  2.8922695 0.015625     0       6          1  71.652822
#>  8:  1.64358377 12.9899250 0.015625     0       6          1  99.983726
#>  9:  2.33928712 10.3720577 0.015625     0       6          1  66.161676
#> 10:  1.81282533 11.1611247 0.015625     0       6          1  70.826120
#> 11:  6.80159374  5.6325391 0.015625     0       6          1  99.975382
#> 12:  1.22195013  5.0485665 0.015625     0       6          1  14.179796
#> 13:  8.99146628  8.6530748 0.015625     0       6          1 210.887687
#> 14:  0.39203270  1.7094962 0.015625     0       6          1  32.352011
#> 15:  5.20691172  8.4422893 0.015625     0       6          1 112.482865
#> 16:  2.65364194 10.5347083 0.015625     0       6          1  74.477358
#> 17:  5.48478964  7.6846424 0.015625     0       6          1 106.741953
#> 18:  4.97978637  7.6212837 0.015625     0       6          1  89.702550
#> 19:  5.18001351  8.3540491 0.015625     0       6          1 109.845246
#> 20:  7.04704977 12.0748954 0.015625     0       6          1 265.428548
#> 21:  6.28145886 12.1595372 0.015625     0       6          1 242.913674
#> 22:  9.55100341  6.5173211 0.015625     0       6          1 167.230742
#> 23:  5.64377558  8.6313143 0.015625     0       6          1 130.793580
#> 24:  2.73786274  5.5280497 0.015625     0       6          1  14.529028
#> 25:  6.23250948  2.6759397 0.015625     0       6          1  48.759987
#> 26:  4.20173946 12.3544237 0.015625     0       6          1 161.528855
#> 27:  8.98588419  2.6050225 0.015625     0       6          1  72.268357
#> 28: -1.51828005  6.7325092 0.015625     0       6          1  13.583309
#> 29: -4.72156407 10.8991566 0.015625     0       6          1  20.984002
#> 30:  7.88488139  1.1543608 0.015625     0       6          1  43.250100
#> 31:  4.40498235  7.4348905 0.015625     0       6          1  68.701737
#> 32: -4.51516978 10.6130131 0.015625     0       6          1  18.358113
#> 33:  6.79397969  0.5105539 0.015625     0       6          1  33.619891
#> 34:  5.98083322 13.9441638 0.015625     0       6          1 286.918367
#> 35: -3.54065442  5.9558454 0.015625     0       6          1  37.933349
#> 36: -0.59559770 12.3068221 0.015625     0       6          1  46.549773
#> 37: -4.75340321 14.5065227 0.015625     0       6          1  10.454486
#> 38:  5.61928763 12.6185003 0.015625     0       6          1 230.459715
#> 39:  9.56124952  5.6569136 0.015625     0       6          1 146.010381
#> 40:  9.69785025  9.2125990 0.015625     0       6          1 248.989553
#> 41:  7.51177010 13.6831767 0.015625     0       6          1 333.760304
#> 42: -3.54032108 14.3419268 0.015625     0       6          1   6.542581
#> 43:  3.76107132 13.8881366 0.015625     0       6          1 182.792378
#> 44: -2.36502078  6.0686665 0.015625     0       6          1  18.107083
#> 45: -1.62693404  4.2991726 0.015625     0       6          1  28.571762
#> 46:  4.30905274  9.2991480 0.015625     0       6          1  98.128614
#> 47: -0.20885674  1.7557008 0.015625     0       6          1  40.351952
#> 48:  1.59283153  2.0512357 0.015625     0       6          1  12.013610
#> 49:  6.64917112  5.9683817 0.015625     0       6          1 103.449768
#> 50:  4.54913531  6.9464673 0.015625     0       6          1  65.447053
#> 51:  0.64445519  7.0923021 0.015625     0       6          1  22.108118
#> 52:  0.09017823  7.0619135 0.015625     0       6          1  21.015571
#> 53:  0.57362135 12.6374525 0.015625     0       6          1  74.921066
#> 54: -2.26150527 13.8727641 0.015625     0       6          1  20.825868
#> 55:  8.28190262  9.6856557 0.015625     0       6          1 223.807339
#> 56:  9.10145457  6.3531492 0.015625     0       6          1 151.978613
#> 57:  5.24479960  4.1512155 0.015625     0       6          1  46.828146
#> 58: -2.92384879 14.7143823 0.015625     0       6          1  15.050027
#> 59: -4.06779368  0.4466569 0.015625     0       6          1 161.386813
#> 60:  6.55198392 14.6068278 0.015625     0       6          1 333.068304
#> 61:  9.80431429 12.4245752 0.015625     0       6          1 364.863691
#> 62:  8.74698614 11.0641968 0.015625     0       6          1 279.177402
#> 63: -3.76627126  4.7999612 0.015625     0       6          1  60.434023
#> 64:  3.08291368  7.2156371 0.015625     0       6          1  34.403501
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-27 04:49:17        1 -0.23426022  10.9190335          0.015625
#>  2: 2022-08-27 04:49:17        1 -2.17312764  13.5391463          0.015625
#>  3: 2022-08-27 04:49:17        1 -4.43218681   2.7419959          0.015625
#>  4: 2022-08-27 04:49:17        1 -4.25177823   4.2636400          0.015625
#>  5: 2022-08-27 04:49:17        1  9.12641623  12.4599018          0.015625
#>  6: 2022-08-27 04:49:17        1  8.87869500   6.4676804          0.015625
#>  7: 2022-08-27 04:49:17        1  8.52117030   2.8922695          0.015625
#>  8: 2022-08-27 04:49:17        1  1.64358377  12.9899250          0.015625
#>  9: 2022-08-27 04:49:17        1  2.33928712  10.3720577          0.015625
#> 10: 2022-08-27 04:49:17        1  1.81282533  11.1611247          0.015625
#> 11: 2022-08-27 04:49:17        1  6.80159374   5.6325391          0.015625
#> 12: 2022-08-27 04:49:17        1  1.22195013   5.0485665          0.015625
#> 13: 2022-08-27 04:49:17        1  8.99146628   8.6530748          0.015625
#> 14: 2022-08-27 04:49:17        1  0.39203270   1.7094962          0.015625
#> 15: 2022-08-27 04:49:17        1  5.20691172   8.4422893          0.015625
#> 16: 2022-08-27 04:49:17        1  2.65364194  10.5347083          0.015625
#> 17: 2022-08-27 04:49:17        1  5.48478964   7.6846424          0.015625
#> 18: 2022-08-27 04:49:17        1  4.97978637   7.6212837          0.015625
#> 19: 2022-08-27 04:49:17        1  5.18001351   8.3540491          0.015625
#> 20: 2022-08-27 04:49:17        1  7.04704977  12.0748954          0.015625
#> 21: 2022-08-27 04:49:17        1  6.28145886  12.1595372          0.015625
#> 22: 2022-08-27 04:49:17        1  9.55100341   6.5173211          0.015625
#> 23: 2022-08-27 04:49:17        1  5.64377558   8.6313143          0.015625
#> 24: 2022-08-27 04:49:17        1  2.73786274   5.5280497          0.015625
#> 25: 2022-08-27 04:49:17        1  6.23250948   2.6759397          0.015625
#> 26: 2022-08-27 04:49:17        1  4.20173946  12.3544237          0.015625
#> 27: 2022-08-27 04:49:17        1  8.98588419   2.6050225          0.015625
#> 28: 2022-08-27 04:49:17        1 -1.51828005   6.7325092          0.015625
#> 29: 2022-08-27 04:49:17        1 -4.72156407  10.8991566          0.015625
#> 30: 2022-08-27 04:49:17        1  7.88488139   1.1543608          0.015625
#> 31: 2022-08-27 04:49:17        1  4.40498235   7.4348905          0.015625
#> 32: 2022-08-27 04:49:17        1 -4.51516978  10.6130131          0.015625
#> 33: 2022-08-27 04:49:17        1  6.79397969   0.5105539          0.015625
#> 34: 2022-08-27 04:49:17        1  5.98083322  13.9441638          0.015625
#> 35: 2022-08-27 04:49:17        1 -3.54065442   5.9558454          0.015625
#> 36: 2022-08-27 04:49:17        1 -0.59559770  12.3068221          0.015625
#> 37: 2022-08-27 04:49:17        1 -4.75340321  14.5065227          0.015625
#> 38: 2022-08-27 04:49:17        1  5.61928763  12.6185003          0.015625
#> 39: 2022-08-27 04:49:17        1  9.56124952   5.6569136          0.015625
#> 40: 2022-08-27 04:49:17        1  9.69785025   9.2125990          0.015625
#> 41: 2022-08-27 04:49:17        1  7.51177010  13.6831767          0.015625
#> 42: 2022-08-27 04:49:17        1 -3.54032108  14.3419268          0.015625
#> 43: 2022-08-27 04:49:17        1  3.76107132  13.8881366          0.015625
#> 44: 2022-08-27 04:49:17        1 -2.36502078   6.0686665          0.015625
#> 45: 2022-08-27 04:49:17        1 -1.62693404   4.2991726          0.015625
#> 46: 2022-08-27 04:49:17        1  4.30905274   9.2991480          0.015625
#> 47: 2022-08-27 04:49:17        1 -0.20885674   1.7557008          0.015625
#> 48: 2022-08-27 04:49:17        1  1.59283153   2.0512357          0.015625
#> 49: 2022-08-27 04:49:17        1  6.64917112   5.9683817          0.015625
#> 50: 2022-08-27 04:49:17        1  4.54913531   6.9464673          0.015625
#> 51: 2022-08-27 04:49:17        1  0.64445519   7.0923021          0.015625
#> 52: 2022-08-27 04:49:17        1  0.09017823   7.0619135          0.015625
#> 53: 2022-08-27 04:49:17        1  0.57362135  12.6374525          0.015625
#> 54: 2022-08-27 04:49:17        1 -2.26150527  13.8727641          0.015625
#> 55: 2022-08-27 04:49:17        1  8.28190262   9.6856557          0.015625
#> 56: 2022-08-27 04:49:17        1  9.10145457   6.3531492          0.015625
#> 57: 2022-08-27 04:49:17        1  5.24479960   4.1512155          0.015625
#> 58: 2022-08-27 04:49:17        1 -2.92384879  14.7143823          0.015625
#> 59: 2022-08-27 04:49:17        1 -4.06779368   0.4466569          0.015625
#> 60: 2022-08-27 04:49:17        1  6.55198392  14.6068278          0.015625
#> 61: 2022-08-27 04:49:17        1  9.80431429  12.4245752          0.015625
#> 62: 2022-08-27 04:49:17        1  8.74698614  11.0641968          0.015625
#> 63: 2022-08-27 04:49:17        1 -3.76627126   4.7999612          0.015625
#> 64: 2022-08-27 04:49:17        1  3.08291368   7.2156371          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity