Optimizer Using the Hyperband Algorithm

OptimizerHyperband class that implements hyperband optimization (HB). HB repeatedly calls SHA (OptimizerSuccessiveHalving) with different numbers of starting points. A larger number of starting points corresponds to a smaller budget allocated in the base stage. Each run of SHA within HB is called a bracket. HB considers s_max + 1 brackets with s_max = floor(log(r_max / r_min, eta). The most explorative bracket s = s_max constructs s_max + 1 stages and allocates the minimum budget (r_min) in the base stage. The minimum budget is increased in each bracket by a factor of eta and the number of starting points is computed so that each bracket approximately spends the same budget. Use hyperband_schedule() to get a preview of the bracket layout.

s			3			2			1			0

i		n_i	r_i		n_i	r_i		n_i	r_i		n_i	r_i

0		8	1		6	2		4	4		8	4

1		4	2		3	4		2	8

2		2	4		1	8

3		1	8

s is the bracket number, i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage of the most explorative bracket, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which which is allocated to the candidates in the last stages.

Source

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. https://jmlr.org/papers/v18/16-558.html.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("hyperband")
opt("hyperband")

Parameters

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn in the base stage of each bracket. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all brackets are evaluated. Otherwise, optimization is stopped after repetitions runs of hyperband. The bbotk::Terminator might stop the optimization before all repetitions are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the hyperband algorithm:

bracket (integer(1))
The bracket index. Counts down to 0.

stage (integer(1))
The stages of each bracket. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerHyperband

Methods

Public methods

OptimizerHyperband$new()

OptimizerHyperband$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("hyperband")

# optimize branin function
optimizer$optimize(instance)
#>         x1        x2 fidelity  x_domain        y
#> 1: 2.85747 0.5966956 0.015625 <list[3]> 2.007313

# best scoring evaluation
instance$result
#>         x1        x2 fidelity  x_domain        y
#> 1: 2.85747 0.5966956 0.015625 <list[3]> 2.007313

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage bracket repetition          y
#>  1:  3.11147958  4.4088171 0.015625     0       6          1   9.785554
#>  2:  7.82916898  6.3642579 0.015625     0       6          1 129.924064
#>  3: -0.60371992  3.0099720 0.015625     0       6          1  33.602857
#>  4:  8.09414941 12.7278366 0.015625     0       6          1 317.323851
#>  5:  3.66605036  1.7992885 0.015625     0       6          1   3.178687
#>  6:  6.02785320  4.0947871 0.015625     0       6          1  62.471744
#>  7:  8.58957148 14.3574943 0.015625     0       6          1 394.001893
#>  8:  8.76368576  0.4282266 0.015625     0       6          1  38.596592
#>  9:  6.66809246  7.7857043 0.015625     0       6          1 140.586594
#> 10:  2.85747040  0.5966956 0.015625     0       6          1   2.007313
#> 11:  3.19318432 14.4871853 0.015625     0       6          1 176.126536
#> 12:  6.36562056  6.3006823 0.015625     0       6          1 103.951709
#> 13:  1.50038714 11.3349974 0.015625     0       6          1  69.254967
#> 14: -4.24099406 10.9068632 0.015625     0       6          1  11.379759
#> 15: -0.45604564  4.4198234 0.015625     0       6          1  23.967936
#> 16:  1.57642495  4.1022572 0.015625     0       6          1  10.231971
#> 17: -2.74156070  4.5097872 0.015625     0       6          1  38.178850
#> 18:  5.78315495 10.6825407 0.015625     0       6          1 183.764638
#> 19:  6.16760341  0.7611290 0.015625     0       6          1  31.149641
#> 20:  2.96690325  3.3939715 0.015625     0       6          1   3.949135
#> 21: -0.66576944  2.8758359 0.015625     0       6          1  35.169650
#> 22:  8.47100778 12.1554645 0.015625     0       6          1 308.289058
#> 23:  8.57174813  9.9553382 0.015625     0       6          1 238.956280
#> 24: -2.66773877  5.1591924 0.015625     0       6          1  29.603863
#> 25:  3.91358961 10.0147146 0.015625     0       6          1  98.620859
#> 26: -1.67339980  9.3331612 0.015625     0       6          1   9.357297
#> 27:  4.03664199  0.9124823 0.015625     0       6          1   4.692948
#> 28:  5.15690079  2.2407383 0.015625     0       6          1  27.309795
#> 29:  0.01118836  1.3612137 0.015625     0       6          1  40.954999
#> 30:  6.91944231 10.3845942 0.015625     0       6          1 211.631701
#> 31: -4.23496781  9.3223744 0.015625     0       6          1  21.342728
#> 32:  3.91044511 12.3149152 0.015625     0       6          1 148.745432
#> 33: -1.43990484  4.7467479 0.015625     0       6          1  24.275823
#> 34: -0.54290408 11.5898029 0.015625     0       6          1  40.468530
#> 35:  3.70705596  0.9009409 0.015625     0       6          1   2.035712
#> 36:  2.97778654  0.1900722 0.015625     0       6          1   2.330807
#> 37:  5.06375887  4.9744406 0.015625     0       6          1  52.308156
#> 38:  9.32945091 13.3893373 0.015625     0       6          1 383.091469
#> 39:  6.09084132  6.4672130 0.015625     0       6          1 100.793102
#> 40:  6.01044758  0.5300204 0.015625     0       6          1  28.158576
#> 41:  1.62082413 14.6787628 0.015625     0       6          1 134.458792
#> 42: -0.89235527  7.0028152 0.015625     0       6          1  16.221367
#> 43:  5.07496099  1.7751677 0.015625     0       6          1  22.771254
#> 44:  6.22451667 10.2513398 0.015625     0       6          1 187.720512
#> 45:  3.02440936  5.9576281 0.015625     0       6          1  20.622767
#> 46:  7.68713159  1.7103122 0.015625     0       6          1  49.145359
#> 47: -3.91351732 13.8547128 0.015625     0       6          1   4.454015
#> 48:  2.34885462  6.4701195 0.015625     0       6          1  19.572227
#> 49:  7.35839779 14.9912379 0.015625     0       6          1 376.998875
#> 50: -0.36235374  7.0251279 0.015625     0       6          1  19.176080
#> 51:  2.98887837  8.2061259 0.015625     0       6          1  45.244311
#> 52:  4.56957569 10.1450740 0.015625     0       6          1 124.750148
#> 53:  6.37397777 13.3658458 0.015625     0       6          1 283.988199
#> 54:  6.66210800  8.4060259 0.015625     0       6          1 154.513831
#> 55:  7.23746810  6.0765712 0.015625     0       6          1 115.248156
#> 56:  3.58342917  3.4143817 0.015625     0       6          1   8.733448
#> 57:  3.76238101  6.5851149 0.015625     0       6          1  39.863150
#> 58:  9.24848150  9.0070599 0.015625     0       6          1 228.452484
#> 59:  7.85350697 11.1732736 0.015625     0       6          1 258.890436
#> 60:  4.04055672  3.4528130 0.015625     0       6          1  15.458526
#> 61: -0.63358014  4.4974267 0.015625     0       6          1  24.105461
#> 62: -0.23749198  7.7700227 0.015625     0       6          1  21.265549
#> 63:  5.45432173 13.1540918 0.015625     0       6          1 239.100722
#> 64:  1.28923036  6.6012783 0.015625     0       6          1  19.438695
#>              x1         x2 fidelity stage bracket repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-09-30 04:56:06        1  3.11147958   4.4088171          0.015625
#>  2: 2022-09-30 04:56:06        1  7.82916898   6.3642579          0.015625
#>  3: 2022-09-30 04:56:06        1 -0.60371992   3.0099720          0.015625
#>  4: 2022-09-30 04:56:06        1  8.09414941  12.7278366          0.015625
#>  5: 2022-09-30 04:56:06        1  3.66605036   1.7992885          0.015625
#>  6: 2022-09-30 04:56:06        1  6.02785320   4.0947871          0.015625
#>  7: 2022-09-30 04:56:06        1  8.58957148  14.3574943          0.015625
#>  8: 2022-09-30 04:56:06        1  8.76368576   0.4282266          0.015625
#>  9: 2022-09-30 04:56:06        1  6.66809246   7.7857043          0.015625
#> 10: 2022-09-30 04:56:06        1  2.85747040   0.5966956          0.015625
#> 11: 2022-09-30 04:56:06        1  3.19318432  14.4871853          0.015625
#> 12: 2022-09-30 04:56:06        1  6.36562056   6.3006823          0.015625
#> 13: 2022-09-30 04:56:06        1  1.50038714  11.3349974          0.015625
#> 14: 2022-09-30 04:56:06        1 -4.24099406  10.9068632          0.015625
#> 15: 2022-09-30 04:56:06        1 -0.45604564   4.4198234          0.015625
#> 16: 2022-09-30 04:56:06        1  1.57642495   4.1022572          0.015625
#> 17: 2022-09-30 04:56:06        1 -2.74156070   4.5097872          0.015625
#> 18: 2022-09-30 04:56:06        1  5.78315495  10.6825407          0.015625
#> 19: 2022-09-30 04:56:06        1  6.16760341   0.7611290          0.015625
#> 20: 2022-09-30 04:56:06        1  2.96690325   3.3939715          0.015625
#> 21: 2022-09-30 04:56:06        1 -0.66576944   2.8758359          0.015625
#> 22: 2022-09-30 04:56:06        1  8.47100778  12.1554645          0.015625
#> 23: 2022-09-30 04:56:06        1  8.57174813   9.9553382          0.015625
#> 24: 2022-09-30 04:56:06        1 -2.66773877   5.1591924          0.015625
#> 25: 2022-09-30 04:56:06        1  3.91358961  10.0147146          0.015625
#> 26: 2022-09-30 04:56:06        1 -1.67339980   9.3331612          0.015625
#> 27: 2022-09-30 04:56:06        1  4.03664199   0.9124823          0.015625
#> 28: 2022-09-30 04:56:06        1  5.15690079   2.2407383          0.015625
#> 29: 2022-09-30 04:56:06        1  0.01118836   1.3612137          0.015625
#> 30: 2022-09-30 04:56:06        1  6.91944231  10.3845942          0.015625
#> 31: 2022-09-30 04:56:06        1 -4.23496781   9.3223744          0.015625
#> 32: 2022-09-30 04:56:06        1  3.91044511  12.3149152          0.015625
#> 33: 2022-09-30 04:56:06        1 -1.43990484   4.7467479          0.015625
#> 34: 2022-09-30 04:56:06        1 -0.54290408  11.5898029          0.015625
#> 35: 2022-09-30 04:56:06        1  3.70705596   0.9009409          0.015625
#> 36: 2022-09-30 04:56:06        1  2.97778654   0.1900722          0.015625
#> 37: 2022-09-30 04:56:06        1  5.06375887   4.9744406          0.015625
#> 38: 2022-09-30 04:56:06        1  9.32945091  13.3893373          0.015625
#> 39: 2022-09-30 04:56:06        1  6.09084132   6.4672130          0.015625
#> 40: 2022-09-30 04:56:06        1  6.01044758   0.5300204          0.015625
#> 41: 2022-09-30 04:56:06        1  1.62082413  14.6787628          0.015625
#> 42: 2022-09-30 04:56:06        1 -0.89235527   7.0028152          0.015625
#> 43: 2022-09-30 04:56:06        1  5.07496099   1.7751677          0.015625
#> 44: 2022-09-30 04:56:06        1  6.22451667  10.2513398          0.015625
#> 45: 2022-09-30 04:56:06        1  3.02440936   5.9576281          0.015625
#> 46: 2022-09-30 04:56:06        1  7.68713159   1.7103122          0.015625
#> 47: 2022-09-30 04:56:06        1 -3.91351732  13.8547128          0.015625
#> 48: 2022-09-30 04:56:06        1  2.34885462   6.4701195          0.015625
#> 49: 2022-09-30 04:56:06        1  7.35839779  14.9912379          0.015625
#> 50: 2022-09-30 04:56:06        1 -0.36235374   7.0251279          0.015625
#> 51: 2022-09-30 04:56:06        1  2.98887837   8.2061259          0.015625
#> 52: 2022-09-30 04:56:06        1  4.56957569  10.1450740          0.015625
#> 53: 2022-09-30 04:56:06        1  6.37397777  13.3658458          0.015625
#> 54: 2022-09-30 04:56:06        1  6.66210800   8.4060259          0.015625
#> 55: 2022-09-30 04:56:06        1  7.23746810   6.0765712          0.015625
#> 56: 2022-09-30 04:56:06        1  3.58342917   3.4143817          0.015625
#> 57: 2022-09-30 04:56:06        1  3.76238101   6.5851149          0.015625
#> 58: 2022-09-30 04:56:06        1  9.24848150   9.0070599          0.015625
#> 59: 2022-09-30 04:56:06        1  7.85350697  11.1732736          0.015625
#> 60: 2022-09-30 04:56:06        1  4.04055672   3.4528130          0.015625
#> 61: 2022-09-30 04:56:06        1 -0.63358014   4.4974267          0.015625
#> 62: 2022-09-30 04:56:06        1 -0.23749198   7.7700227          0.015625
#> 63: 2022-09-30 04:56:06        1  5.45432173  13.1540918          0.015625
#> 64: 2022-09-30 04:56:06        1  1.28923036   6.6012783          0.015625
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity