Hyperparameter Optimization with Successive Halving

OptimizerSuccessiveHalving class that implements the successive halving algorithm. The algorithm samples n points and evaluates them with the smallest budget (lower bound of the budget parameter). With every stage the budget is increased by a factor of eta and only the best 1/eta points are promoted to the next stage. The optimization terminates when the maximum budget is reached (upper bound of the budget parameter).

To identify the budget, the user has to specify explicitly which parameter of the objective function influences the budget by tagging a single parameter in the search space (paradox::ParamSet) with "budget".

Source

Jamieson K, Talwalkar A (2016). “Non-stochastic Best Arm Identification and Hyperparameter Optimization.” In Gretton A, Robert CC (eds.), Proceedings of the 19th International Conference on Artificial Intelligence and Statistics, volume 51 series Proceedings of Machine Learning Research, 240-248. http://proceedings.mlr.press/v51/jamieson16.html.

Parameters

n

integer(1)
Number of points in first stage.

eta

numeric(1)
With every stage, the point budget is increased by a factor of eta and only the best 1/eta points are used for the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn during the initialization of each bracket. The default is uniform sampling.

repeats

logical(1)
If FALSE (default), successive halving terminates once all stages are evaluated. Otherwise, successive halving starts over again once the last stage is evaluated.

Archive

The bbotk::Archive holds the following additional column that is specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

Custom sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Runtime

The calculation of each bracket currently assumes a linear runtime in the chosen budget parameter is always given. Hyperband is designed so each bracket requires approximately the same runtime as the sum of the budget over all configurations in each bracket is roughly the same. This will not hold true once the scaling in the budget parameter is not linear anymore, even though the sum of the budgets in each bracket remains the same. A possible adaption would be to introduce a trafo, like it is shown in the examples section.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Parallelization

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. The points of one stage in a bracket are evaluated in one batch. Parallelization is supported via package future (see mlr3::benchmark()'s section on parallelization for more details).

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerSuccessiveHalving

Methods

Public methods

OptimizerSuccessiveHalving$new()

OptimizerSuccessiveHalving$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerSuccessiveHalving$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerSuccessiveHalving$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# modified branin function
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("none")
)

optimizer = opt("successive_halving")

# modifies the instance by reference
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.125237 5.843612     0.16 <list[3]> 14.07698

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.125237 5.843612     0.16 <list[3]> 14.07698

# all evaluations
as.data.table(instance$archive)
#>            x1        x2 fidelity stage         y           timestamp batch_nr
#>  1:  9.980292  6.536968     0.01     0 181.79140 2022-01-20 20:33:09        1
#>  2:  7.498364 12.108867     0.01     0 280.52657 2022-01-20 20:33:09        1
#>  3:  2.125237  5.843612     0.01     0  14.49104 2022-01-20 20:33:09        1
#>  4: -1.484592 12.806111     0.01     0  29.98293 2022-01-20 20:33:09        1
#>  5:  6.078157  2.428200     0.01     0  44.26869 2022-01-20 20:33:09        1
#>  6:  8.012933  8.118735     0.01     0 175.75955 2022-01-20 20:33:09        1
#>  7:  8.073664 12.310203     0.01     0 303.48234 2022-01-20 20:33:09        1
#>  8:  6.060838 11.485350     0.01     0 216.00136 2022-01-20 20:33:09        1
#>  9:  9.925451  2.615521     0.01     0  90.66650 2022-01-20 20:33:09        1
#> 10: -3.481303  8.063647     0.01     0  15.71404 2022-01-20 20:33:09        1
#> 11:  9.836293  4.271996     0.01     0 122.34279 2022-01-20 20:33:09        1
#> 12: -4.140259  2.609289     0.01     0 114.99979 2022-01-20 20:33:09        1
#> 13:  2.391306  9.194557     0.01     0  49.59542 2022-01-20 20:33:09        1
#> 14: -4.654399  6.398599     0.01     0  68.16515 2022-01-20 20:33:09        1
#> 15:  5.396373  9.719005     0.01     0 146.68037 2022-01-20 20:33:09        1
#> 16: -4.155979  4.943253     0.01     0  72.04651 2022-01-20 20:33:09        1
#> 17:  2.125237  5.843612     0.02     1  14.46315 2022-01-20 20:33:09        2
#> 18: -3.481303  8.063647     0.02     1  15.80733 2022-01-20 20:33:09        2
#> 19: -1.484592 12.806111     0.02     1  29.96365 2022-01-20 20:33:09        2
#> 20:  6.078157  2.428200     0.02     1  43.90159 2022-01-20 20:33:09        2
#> 21:  2.391306  9.194557     0.02     1  49.51736 2022-01-20 20:33:09        2
#> 22: -4.654399  6.398599     0.02     1  68.49763 2022-01-20 20:33:09        2
#> 23: -4.155979  4.943253     0.02     1  72.32981 2022-01-20 20:33:09        2
#> 24:  9.925451  2.615521     0.02     1  88.81650 2022-01-20 20:33:09        2
#> 25:  2.125237  5.843612     0.04     2  14.40749 2022-01-20 20:33:09        3
#> 26: -3.481303  8.063647     0.04     2  15.99480 2022-01-20 20:33:09        3
#> 27: -1.484592 12.806111     0.04     2  29.92510 2022-01-20 20:33:09        3
#> 28:  6.078157  2.428200     0.04     2  43.17559 2022-01-20 20:33:09        3
#> 29:  2.125237  5.843612     0.08     3  14.29667 2022-01-20 20:33:09        4
#> 30: -3.481303  8.063647     0.08     3  16.37326 2022-01-20 20:33:09        4
#> 31:  2.125237  5.843612     0.16     4  14.07698 2022-01-20 20:33:09        5
#>            x1        x2 fidelity stage         y           timestamp batch_nr
#>     x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1:    9.980292    6.536968              0.01
#>  2:    7.498364   12.108867              0.01
#>  3:    2.125237    5.843612              0.01
#>  4:   -1.484592   12.806111              0.01
#>  5:    6.078157    2.428200              0.01
#>  6:    8.012933    8.118735              0.01
#>  7:    8.073664   12.310203              0.01
#>  8:    6.060838   11.485350              0.01
#>  9:    9.925451    2.615521              0.01
#> 10:   -3.481303    8.063647              0.01
#> 11:    9.836293    4.271996              0.01
#> 12:   -4.140259    2.609289              0.01
#> 13:    2.391306    9.194557              0.01
#> 14:   -4.654399    6.398599              0.01
#> 15:    5.396373    9.719005              0.01
#> 16:   -4.155979    4.943253              0.01
#> 17:    2.125237    5.843612              0.02
#> 18:   -3.481303    8.063647              0.02
#> 19:   -1.484592   12.806111              0.02
#> 20:    6.078157    2.428200              0.02
#> 21:    2.391306    9.194557              0.02
#> 22:   -4.654399    6.398599              0.02
#> 23:   -4.155979    4.943253              0.02
#> 24:    9.925451    2.615521              0.02
#> 25:    2.125237    5.843612              0.04
#> 26:   -3.481303    8.063647              0.04
#> 27:   -1.484592   12.806111              0.04
#> 28:    6.078157    2.428200              0.04
#> 29:    2.125237    5.843612              0.08
#> 30:   -3.481303    8.063647              0.08
#> 31:    2.125237    5.843612              0.16
#>     x_domain_x1 x_domain_x2 x_domain_fidelity