Hyperparameter Optimization with Successive Halving

OptimizerSuccessiveHalving class that implements the successive halving algorithm (SHA). SHA randomly samples n candidate points and allocates a minimum budget (r_min) to all candidates. The candidates are raced down in stages to a single best candidate by repeatedly increasing the budget by a factor of eta and promoting only the best 1 / eta fraction to the next stage. This means promising points are allocated a higher budget overall and lower performing ones are discarded early on.

#' The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which is allocated to the candidates in the last stage. The number of stages is computed so that each candidate in base stage is allocated the minimum budget and the candidates in the last stage are not evaluated on more than the maximum budget. The following table is the stage layout for eta = 2, r_min = 1 and r_max = 8.

i	n_i	r_i

0	8	1

1	4	2

2	2	4

3	1	8

i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

Source

Jamieson K, Talwalkar A (2016). “Non-stochastic Best Arm Identification and Hyperparameter Optimization.” In Gretton A, Robert CC (eds.), Proceedings of the 19th International Conference on Artificial Intelligence and Statistics, volume 51 series Proceedings of Machine Learning Research, 240-248. http://proceedings.mlr.press/v51/jamieson16.html.

Parameters

n

integer(1)
Number of points in base stage.

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all stages are evaluated. Otherwise, optimization is stopped after repetitions runs of SHA. The bbotk::Terminator might stop the optimization before all repetitions are executed.

adjust_minimum_budget

logical(1)
If TRUE, minimum budget is increased so that the last stage uses the maximum budget defined in the search space.

Archive

The bbotk::Archive holds the following additional columns that are specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerSuccessiveHalving

Methods

Public methods

OptimizerSuccessiveHalving$new()

OptimizerSuccessiveHalving$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerSuccessiveHalving$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerSuccessiveHalving$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("successive_halving")

# optimize branin function
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.368083 2.949161     0.16 <list[3]> 3.345986

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain        y
#> 1: 2.368083 2.949161     0.16 <list[3]> 3.345986

# all evaluations
as.data.table(instance$archive)
#>              x1         x2 fidelity stage repetition          y
#>  1:  0.38248816  2.9970397     0.01     0          1  24.661666
#>  2:  2.36808299  2.9491614     0.01     0          1   3.431239
#>  3:  6.86676708  5.7145324     0.01     0          1 103.022357
#>  4: -3.22315061  0.4630643     0.01     0          1 120.997166
#>  5: -1.14882134  4.1178524     0.01     0          1  27.998101
#>  6:  1.80846470 11.2577406     0.01     0          1  72.337204
#>  7:  9.02484403 13.0281185     0.01     0          1 359.619478
#>  8:  1.61091373  5.2987824     0.01     0          1  12.798622
#>  9:  5.00245168 14.8406801     0.01     0          1 270.251769
#> 10: -4.81772171  9.4901503     0.01     0          1  34.805288
#> 11: -4.01997006  7.7986161     0.01     0          1  29.749113
#> 12: -3.80050994  9.5968172     0.01     0          1  10.747792
#> 13: -3.15685359  6.7636920     0.01     0          1  21.205440
#> 14:  0.06358683  3.5554350     0.01     0          1  25.074631
#> 15:  2.44580404  7.9153996     0.01     0          1  34.298165
#> 16: -2.42433184  2.6243788     0.01     0          1  57.693665
#> 17:  2.36808299  2.9491614     0.02     1          1   3.425115
#> 18: -3.80050994  9.5968172     0.02     1          1  10.831424
#> 19:  1.61091373  5.2987824     0.02     1          1  12.789368
#> 20: -3.15685359  6.7636920     0.02     1          1  21.296454
#> 21:  0.38248816  2.9970397     0.02     1          1  24.662368
#> 22:  0.06358683  3.5554350     0.02     1          1  25.074650
#> 23: -1.14882134  4.1178524     0.02     1          1  28.008002
#> 24: -4.01997006  7.7986161     0.02     1          1  29.913792
#> 25:  2.36808299  2.9491614     0.04     2          1   3.413056
#> 26: -3.80050994  9.5968172     0.04     2          1  10.999941
#> 27:  1.61091373  5.2987824     0.04     2          1  12.770901
#> 28: -3.15685359  6.7636920     0.04     2          1  21.479080
#> 29:  2.36808299  2.9491614     0.08     3          1   3.389693
#> 30: -3.80050994  9.5968172     0.08     3          1  11.341981
#> 31:  2.36808299  2.9491614     0.16     4          1   3.345986
#>              x1         x2 fidelity stage repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2022-08-05 04:51:35        1  0.38248816   2.9970397              0.01
#>  2: 2022-08-05 04:51:35        1  2.36808299   2.9491614              0.01
#>  3: 2022-08-05 04:51:35        1  6.86676708   5.7145324              0.01
#>  4: 2022-08-05 04:51:35        1 -3.22315061   0.4630643              0.01
#>  5: 2022-08-05 04:51:35        1 -1.14882134   4.1178524              0.01
#>  6: 2022-08-05 04:51:35        1  1.80846470  11.2577406              0.01
#>  7: 2022-08-05 04:51:35        1  9.02484403  13.0281185              0.01
#>  8: 2022-08-05 04:51:35        1  1.61091373   5.2987824              0.01
#>  9: 2022-08-05 04:51:35        1  5.00245168  14.8406801              0.01
#> 10: 2022-08-05 04:51:35        1 -4.81772171   9.4901503              0.01
#> 11: 2022-08-05 04:51:35        1 -4.01997006   7.7986161              0.01
#> 12: 2022-08-05 04:51:35        1 -3.80050994   9.5968172              0.01
#> 13: 2022-08-05 04:51:35        1 -3.15685359   6.7636920              0.01
#> 14: 2022-08-05 04:51:35        1  0.06358683   3.5554350              0.01
#> 15: 2022-08-05 04:51:35        1  2.44580404   7.9153996              0.01
#> 16: 2022-08-05 04:51:35        1 -2.42433184   2.6243788              0.01
#> 17: 2022-08-05 04:51:35        2  2.36808299   2.9491614              0.02
#> 18: 2022-08-05 04:51:35        2 -3.80050994   9.5968172              0.02
#> 19: 2022-08-05 04:51:35        2  1.61091373   5.2987824              0.02
#> 20: 2022-08-05 04:51:35        2 -3.15685359   6.7636920              0.02
#> 21: 2022-08-05 04:51:35        2  0.38248816   2.9970397              0.02
#> 22: 2022-08-05 04:51:35        2  0.06358683   3.5554350              0.02
#> 23: 2022-08-05 04:51:35        2 -1.14882134   4.1178524              0.02
#> 24: 2022-08-05 04:51:35        2 -4.01997006   7.7986161              0.02
#> 25: 2022-08-05 04:51:35        3  2.36808299   2.9491614              0.04
#> 26: 2022-08-05 04:51:35        3 -3.80050994   9.5968172              0.04
#> 27: 2022-08-05 04:51:35        3  1.61091373   5.2987824              0.04
#> 28: 2022-08-05 04:51:35        3 -3.15685359   6.7636920              0.04
#> 29: 2022-08-05 04:51:35        4  2.36808299   2.9491614              0.08
#> 30: 2022-08-05 04:51:35        4 -3.80050994   9.5968172              0.08
#> 31: 2022-08-05 04:51:35        5  2.36808299   2.9491614              0.16
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity