Hyperparameter Optimization with Successive Halving

OptimizerSuccessiveHalving class that implements the successive halving algorithm (SHA). SHA randomly samples n candidate points and allocates a minimum budget (r_min) to all candidates. The candidates are raced down in stages to a single best candidate by repeatedly increasing the budget by a factor of eta and promoting only the best 1 / eta fraction to the next stage. This means promising points are allocated a higher budget overall and lower performing ones are discarded early on.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which is allocated to the candidates in the last stage. The number of stages is computed so that each candidate in base stage is allocated the minimum budget and the candidates in the last stage are not evaluated on more than the maximum budget. The following table is the stage layout for eta = 2, r_min = 1 and r_max = 8.

i	n_i	r_i

0	8	1

1	4	2

2	2	4

3	1	8

i is stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

Source

Jamieson K, Talwalkar A (2016). “Non-stochastic Best Arm Identification and Hyperparameter Optimization.” In Gretton A, Robert CC (eds.), Proceedings of the 19th International Conference on Artificial Intelligence and Statistics, volume 51 series Proceedings of Machine Learning Research, 240-248. http://proceedings.mlr.press/v51/jamieson16.html.

Parameters

n

integer(1)
Number of points in base stage.

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all stages are evaluated. Otherwise, optimization is stopped after repetitions runs of SHA. The bbotk::Terminator might stop the optimization before all repetition are executed.

Archive

The bbotk::Archive holds the following additional columns that are specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

mlr3hyperband supports custom paradox::Sampler to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

The $optimize() method supports progress bars via the package progressr. Simply wrap the method call in progressr::with_progress() to enable them. Alternatively, call progressr::handlers() with global = TRUE to enable progress bars globally. We recommend the progress package as backend which can be enabled with progressr::handlers("progress").

Logging

mlr3hyperband uses the lgr package for logging. lgr supports multiple log levels which can be queried with getOption("lgr.log_levels"). Use lgr::get_logger("bbotk") to access and control the logger.

To suppress output and reduce verbosity, you can lower the log from the default level "info" to "warn":

lgr::get_logger("bbotk")$set_threshold("warn")

To log to a file or a data base, see the documentation of lgr::lgr-package.

Super class

bbotk::Optimizer -> OptimizerSuccessiveHalving

Methods

Public methods

OptimizerSuccessiveHalving$new()

OptimizerSuccessiveHalving$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerSuccessiveHalving$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerSuccessiveHalving$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("successive_halving")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.070961 11.68456     0.16 <list[3]> 0.559356

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.070961 11.68456     0.16 <list[3]> 0.559356

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage repetition           y
#>  1: -4.4498089 13.4711606     0.01     0          1   7.5510670
#>  2:  5.9936600  8.9170842     0.01     0          1 148.5235618
#>  3: -1.7709101 12.4753488     0.01     0          1  20.7805203
#>  4: -4.7599528  8.9011255     0.01     0          1  39.1697392
#>  5: -3.0709613 11.6845595     0.01     0          1   0.6842889
#>  6:  5.2909466  5.9665747     0.01     0          1  72.1382296
#>  7:  4.6291190 12.7482416     0.01     0          1 190.6125474
#>  8: -0.0899222 11.1276837     0.01     0          1  44.4068035
#>  9:  0.8122297  4.7668533     0.01     0          1  16.6066898
#> 10:  5.6125059  1.6752034     0.01     0          1  30.8956858
#> 11:  5.4648314  1.5164305     0.01     0          1  27.5352210
#> 12:  8.8848515 12.0014662     0.01     0          1 317.1588213
#> 13:  1.8852659  5.6992011     0.01     0          1  13.7505524
#> 14:  3.9393130  0.7905146     0.01     0          1   3.6445879
#> 15: -2.5248352 14.7976305     0.01     0          1  23.2058289
#> 16:  3.5036468  9.0626095     0.01     0          1  69.3852538
#> 17: -3.0709613 11.6845595     0.02     1          1   0.6747149
#> 18:  3.9393130  0.7905146     0.02     1          1   3.6264640
#> 19: -4.4498089 13.4711606     0.02     1          1   7.5597203
#> 20:  1.8852659  5.6992011     0.02     1          1  13.7321370
#> 21:  0.8122297  4.7668533     0.02     1          1  16.6066379
#> 22: -1.7709101 12.4753488     0.02     1          1  20.7581872
#> 23: -2.5248352 14.7976305     0.02     1          1  23.1473896
#> 24:  5.4648314  1.5164305     0.02     1          1  27.3382591
#> 25: -3.0709613 11.6845595     0.04     2          1   0.6561005
#> 26:  3.9393130  0.7905146     0.04     2          1   3.5916610
#> 27: -4.4498089 13.4711606     0.04     2          1   7.5793793
#> 28:  1.8852659  5.6992011     0.04     2          1  13.6953820
#> 29: -3.0709613 11.6845595     0.08     3          1   0.6210062
#> 30:  3.9393130  0.7905146     0.08     3          1   3.5278345
#> 31: -3.0709613 11.6845595     0.16     4          1   0.5593560
#>             x1         x2 fidelity stage repetition           y
#>               timestamp batch_nr    status x_domain_x1 x_domain_x2
#>  1: 2021-11-27 11:24:41        1 evaluated  -4.4498089  13.4711606
#>  2: 2021-11-27 11:24:41        1 evaluated   5.9936600   8.9170842
#>  3: 2021-11-27 11:24:41        1 evaluated  -1.7709101  12.4753488
#>  4: 2021-11-27 11:24:41        1 evaluated  -4.7599528   8.9011255
#>  5: 2021-11-27 11:24:41        1 evaluated  -3.0709613  11.6845595
#>  6: 2021-11-27 11:24:41        1 evaluated   5.2909466   5.9665747
#>  7: 2021-11-27 11:24:41        1 evaluated   4.6291190  12.7482416
#>  8: 2021-11-27 11:24:41        1 evaluated  -0.0899222  11.1276837
#>  9: 2021-11-27 11:24:41        1 evaluated   0.8122297   4.7668533
#> 10: 2021-11-27 11:24:41        1 evaluated   5.6125059   1.6752034
#> 11: 2021-11-27 11:24:41        1 evaluated   5.4648314   1.5164305
#> 12: 2021-11-27 11:24:41        1 evaluated   8.8848515  12.0014662
#> 13: 2021-11-27 11:24:41        1 evaluated   1.8852659   5.6992011
#> 14: 2021-11-27 11:24:41        1 evaluated   3.9393130   0.7905146
#> 15: 2021-11-27 11:24:41        1 evaluated  -2.5248352  14.7976305
#> 16: 2021-11-27 11:24:41        1 evaluated   3.5036468   9.0626095
#> 17: 2021-11-27 11:24:41        2 evaluated  -3.0709613  11.6845595
#> 18: 2021-11-27 11:24:41        2 evaluated   3.9393130   0.7905146
#> 19: 2021-11-27 11:24:41        2 evaluated  -4.4498089  13.4711606
#> 20: 2021-11-27 11:24:41        2 evaluated   1.8852659   5.6992011
#> 21: 2021-11-27 11:24:41        2 evaluated   0.8122297   4.7668533
#> 22: 2021-11-27 11:24:41        2 evaluated  -1.7709101  12.4753488
#> 23: 2021-11-27 11:24:41        2 evaluated  -2.5248352  14.7976305
#> 24: 2021-11-27 11:24:41        2 evaluated   5.4648314   1.5164305
#> 25: 2021-11-27 11:24:41        3 evaluated  -3.0709613  11.6845595
#> 26: 2021-11-27 11:24:41        3 evaluated   3.9393130   0.7905146
#> 27: 2021-11-27 11:24:41        3 evaluated  -4.4498089  13.4711606
#> 28: 2021-11-27 11:24:41        3 evaluated   1.8852659   5.6992011
#> 29: 2021-11-27 11:24:41        4 evaluated  -3.0709613  11.6845595
#> 30: 2021-11-27 11:24:41        4 evaluated   3.9393130   0.7905146
#> 31: 2021-11-27 11:24:41        5 evaluated  -3.0709613  11.6845595
#>               timestamp batch_nr    status x_domain_x1 x_domain_x2
#>     x_domain_fidelity
#>  1:              0.01
#>  2:              0.01
#>  3:              0.01
#>  4:              0.01
#>  5:              0.01
#>  6:              0.01
#>  7:              0.01
#>  8:              0.01
#>  9:              0.01
#> 10:              0.01
#> 11:              0.01
#> 12:              0.01
#> 13:              0.01
#> 14:              0.01
#> 15:              0.01
#> 16:              0.01
#> 17:              0.02
#> 18:              0.02
#> 19:              0.02
#> 20:              0.02
#> 21:              0.02
#> 22:              0.02
#> 23:              0.02
#> 24:              0.02
#> 25:              0.04
#> 26:              0.04
#> 27:              0.04
#> 28:              0.04
#> 29:              0.08
#> 30:              0.08
#> 31:              0.16
#>     x_domain_fidelity