Hyperparameter Optimization with Successive Halving

OptimizerSuccessiveHalving class that implements the successive halving algorithm. Successive halving (SHA) randomly samples n candidate points and allocates a minimum budget (r_min) to all candidates. The candidates are raced down in stages to a single best candidate by repeatedly increasing the budget by a factor of eta and promoting only the best 1 / eta fraction to the next stage. This means promising points are allocated a higher budget overall and lower performing ones are discarded early on.

The budget hyperparameter must be tagged with "budget" in the search space. The minimum budget (r_min) which is allocated in the base stage, is set by the lower bound of the budget parameter. The upper bound defines the maximum budget (r_max) which is allocated to the candidates in the last stage. The number of stages is computed so that each candidate in base stage is allocated the minimum budget and the candidates in the last stage are not evaluated on more than the maximum budget. The following table is the stage layout for eta = 2, r_min = 1 and r_max = 8.

i	ni	ri

0	8	1

1	4	2

2	2	4

3	1	8

Source

Jamieson K, Talwalkar A (2016). “Non-stochastic Best Arm Identification and Hyperparameter Optimization.” In Gretton A, Robert CC (eds.), Proceedings of the 19th International Conference on Artificial Intelligence and Statistics, volume 51 series Proceedings of Machine Learning Research, 240-248. http://proceedings.mlr.press/v51/jamieson16.html.

Parameters

n

integer(1)
Number of points in base stage.

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta points are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn. The default is uniform sampling.

repeats

logical(1)
If FALSE (default), SHA terminates once all stages are evaluated. Otherwise, SHA starts over again once the last stage is evaluated.

Archive

The bbotk::Archive holds the following additional column that is specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

Custom sampler

All algorithms support custom paradox::Samplers to draw candidates in the base stage.

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerSuccessiveHalving

Methods

Public methods

OptimizerSuccessiveHalving$new()

OptimizerSuccessiveHalving$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerSuccessiveHalving$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerSuccessiveHalving$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# objective with modified branin function, see `bbotk::branin()`
objective = ObjectiveRFun$new(
  fun = branin,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("successive_halving")

# optimize branin function
optimizer$optimize(instance)
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.070961 11.68456     0.16 <list[3]> 0.559356

# best scoring evaluation
instance$result
#>           x1       x2 fidelity  x_domain        y
#> 1: -3.070961 11.68456     0.16 <list[3]> 0.559356

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage           y           timestamp
#>  1: -4.4498089 13.4711606     0.01     0   7.5510670 2021-11-26 14:30:25
#>  2:  5.9936600  8.9170842     0.01     0 148.5235618 2021-11-26 14:30:25
#>  3: -1.7709101 12.4753488     0.01     0  20.7805203 2021-11-26 14:30:25
#>  4: -4.7599528  8.9011255     0.01     0  39.1697392 2021-11-26 14:30:25
#>  5: -3.0709613 11.6845595     0.01     0   0.6842889 2021-11-26 14:30:25
#>  6:  5.2909466  5.9665747     0.01     0  72.1382296 2021-11-26 14:30:25
#>  7:  4.6291190 12.7482416     0.01     0 190.6125474 2021-11-26 14:30:25
#>  8: -0.0899222 11.1276837     0.01     0  44.4068035 2021-11-26 14:30:25
#>  9:  0.8122297  4.7668533     0.01     0  16.6066898 2021-11-26 14:30:25
#> 10:  5.6125059  1.6752034     0.01     0  30.8956858 2021-11-26 14:30:25
#> 11:  5.4648314  1.5164305     0.01     0  27.5352210 2021-11-26 14:30:25
#> 12:  8.8848515 12.0014662     0.01     0 317.1588213 2021-11-26 14:30:25
#> 13:  1.8852659  5.6992011     0.01     0  13.7505524 2021-11-26 14:30:25
#> 14:  3.9393130  0.7905146     0.01     0   3.6445879 2021-11-26 14:30:25
#> 15: -2.5248352 14.7976305     0.01     0  23.2058289 2021-11-26 14:30:25
#> 16:  3.5036468  9.0626095     0.01     0  69.3852538 2021-11-26 14:30:25
#> 17: -3.0709613 11.6845595     0.02     1   0.6747149 2021-11-26 14:30:25
#> 18:  3.9393130  0.7905146     0.02     1   3.6264640 2021-11-26 14:30:25
#> 19: -4.4498089 13.4711606     0.02     1   7.5597203 2021-11-26 14:30:25
#> 20:  1.8852659  5.6992011     0.02     1  13.7321370 2021-11-26 14:30:25
#> 21:  0.8122297  4.7668533     0.02     1  16.6066379 2021-11-26 14:30:25
#> 22: -1.7709101 12.4753488     0.02     1  20.7581872 2021-11-26 14:30:25
#> 23: -2.5248352 14.7976305     0.02     1  23.1473896 2021-11-26 14:30:25
#> 24:  5.4648314  1.5164305     0.02     1  27.3382591 2021-11-26 14:30:25
#> 25: -3.0709613 11.6845595     0.04     2   0.6561005 2021-11-26 14:30:25
#> 26:  3.9393130  0.7905146     0.04     2   3.5916610 2021-11-26 14:30:25
#> 27: -4.4498089 13.4711606     0.04     2   7.5793793 2021-11-26 14:30:25
#> 28:  1.8852659  5.6992011     0.04     2  13.6953820 2021-11-26 14:30:25
#> 29: -3.0709613 11.6845595     0.08     3   0.6210062 2021-11-26 14:30:25
#> 30:  3.9393130  0.7905146     0.08     3   3.5278345 2021-11-26 14:30:25
#> 31: -3.0709613 11.6845595     0.16     4   0.5593560 2021-11-26 14:30:25
#>             x1         x2 fidelity stage           y           timestamp
#>     batch_nr    status x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1:        1 evaluated  -4.4498089  13.4711606              0.01
#>  2:        1 evaluated   5.9936600   8.9170842              0.01
#>  3:        1 evaluated  -1.7709101  12.4753488              0.01
#>  4:        1 evaluated  -4.7599528   8.9011255              0.01
#>  5:        1 evaluated  -3.0709613  11.6845595              0.01
#>  6:        1 evaluated   5.2909466   5.9665747              0.01
#>  7:        1 evaluated   4.6291190  12.7482416              0.01
#>  8:        1 evaluated  -0.0899222  11.1276837              0.01
#>  9:        1 evaluated   0.8122297   4.7668533              0.01
#> 10:        1 evaluated   5.6125059   1.6752034              0.01
#> 11:        1 evaluated   5.4648314   1.5164305              0.01
#> 12:        1 evaluated   8.8848515  12.0014662              0.01
#> 13:        1 evaluated   1.8852659   5.6992011              0.01
#> 14:        1 evaluated   3.9393130   0.7905146              0.01
#> 15:        1 evaluated  -2.5248352  14.7976305              0.01
#> 16:        1 evaluated   3.5036468   9.0626095              0.01
#> 17:        2 evaluated  -3.0709613  11.6845595              0.02
#> 18:        2 evaluated   3.9393130   0.7905146              0.02
#> 19:        2 evaluated  -4.4498089  13.4711606              0.02
#> 20:        2 evaluated   1.8852659   5.6992011              0.02
#> 21:        2 evaluated   0.8122297   4.7668533              0.02
#> 22:        2 evaluated  -1.7709101  12.4753488              0.02
#> 23:        2 evaluated  -2.5248352  14.7976305              0.02
#> 24:        2 evaluated   5.4648314   1.5164305              0.02
#> 25:        3 evaluated  -3.0709613  11.6845595              0.04
#> 26:        3 evaluated   3.9393130   0.7905146              0.04
#> 27:        3 evaluated  -4.4498089  13.4711606              0.04
#> 28:        3 evaluated   1.8852659   5.6992011              0.04
#> 29:        4 evaluated  -3.0709613  11.6845595              0.08
#> 30:        4 evaluated   3.9393130   0.7905146              0.08
#> 31:        5 evaluated  -3.0709613  11.6845595              0.16
#>     batch_nr    status x_domain_x1 x_domain_x2 x_domain_fidelity