Hyperparameter Optimization with Successive Halving

OptimizerSuccessiveHalving class that implements the successive halving algorithm. The algorithm samples n points and evaluates them with the smallest budget (lower bound of the budget parameter). With every stage the budget is increased by a factor of eta and only the best 1/eta points are promoted to the next stage. The optimization terminates when the maximum budget is reached (upper bound of the budget parameter).

To identify the budget, the user has to specify explicitly which parameter of the objective function influences the budget by tagging a single parameter in the search space (paradox::ParamSet) with "budget".

Source

Jamieson K, Talwalkar A (2016). “Non-stochastic Best Arm Identification and Hyperparameter Optimization.” In Gretton A, Robert CC (eds.), Proceedings of the 19th International Conference on Artificial Intelligence and Statistics, volume 51 series Proceedings of Machine Learning Research, 240-248. http://proceedings.mlr.press/v51/jamieson16.html.

Parameters

n

integer(1)
Number of points in first stage.

eta

numeric(1)
With every stage, the point budget is increased by a factor of eta and only the best 1/eta points are used for the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn during the initialization of each bracket. The default is uniform sampling.

repeats

logical(1)
If FALSE (default), successive halving terminates once all stages are evaluated. Otherwise, successive halving starts over again once the last stage is evaluated.

Archive

The bbotk::Archive holds the following additional column that is specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

Custom sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Runtime

The calculation of each bracket currently assumes a linear runtime in the chosen budget parameter is always given. Hyperband is designed so each bracket requires approximately the same runtime as the sum of the budget over all configurations in each bracket is roughly the same. This will not hold true once the scaling in the budget parameter is not linear anymore, even though the sum of the budgets in each bracket remains the same. A possible adaption would be to introduce a trafo, like it is shown in the examples section.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Parallelization

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. The points of one stage in a bracket are evaluated in one batch. Parallelization is supported via package future (see mlr3::benchmark()'s section on parallelization for more details).

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerSuccessiveHalving

Methods

Public methods

OptimizerSuccessiveHalving$new()

OptimizerSuccessiveHalving$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerSuccessiveHalving$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerSuccessiveHalving$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# modified branin function
objective = ObjectiveRFunDt$new(
  fun = function(xdt) {
    a = 1
    b = 5.1 / (4 * (pi ^ 2))
    c = 5 / pi
    r = 6
    s = 10
    t = 1 / (8 * pi)
    data.table(y =
      (a * ((xdt[["x2"]] -
      b * (xdt[["x1"]] ^ 2L) +
      c * xdt[["x1"]] - r) ^ 2) +
      ((s * (1 - t)) * cos(xdt[["x1"]])) +
      s - (5 * xdt[["fidelity"]] * xdt[["x1"]])))
  },
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("none")
)

optimizer = opt("successive_halving")

# modifies the instance by reference
optimizer$optimize(instance)
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.526259 2.049459     0.16 <list[3]> -1.718372

# best scoring evaluation
instance$result
#>          x1       x2 fidelity  x_domain         y
#> 1: 3.526259 2.049459     0.16 <list[3]> -1.718372

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage           y           timestamp
#>  1:  4.8067706 12.3361057     0.01     0 131.6976916 2022-01-03 04:29:59
#>  2:  3.5546516  0.4229391     0.01     0   3.4363433 2022-01-03 04:29:59
#>  3:  3.5262589  2.0494585     0.01     0   0.9263224 2022-01-03 04:29:59
#>  4:  1.7415065  5.4600249     0.01     0  11.6670091 2022-01-03 04:29:59
#>  5: -2.9615141 14.8561522     0.01     0   9.7597375 2022-01-03 04:29:59
#>  6:  4.8570208  6.5519708     0.01     0  38.5422543 2022-01-03 04:29:59
#>  7:  4.6638320  7.7851660     0.01     0  50.2345117 2022-01-03 04:29:59
#>  8: -3.1012960  1.4324203     0.01     0 116.0361203 2022-01-03 04:29:59
#>  9:  0.3248212  0.7425342     0.01     0  41.6854792 2022-01-03 04:29:59
#> 10:  2.3366996 13.1616956     0.01     0 106.7638187 2022-01-03 04:29:59
#> 11:  1.8934755 12.2549459     0.01     0  84.3945516 2022-01-03 04:29:59
#> 12:  3.5739485 12.8238126     0.01     0 119.0823460 2022-01-03 04:29:59
#> 13:  8.6513194  6.6077093     0.01     0  24.8610328 2022-01-03 04:29:59
#> 14: -1.7437859  0.5591148     0.01     0  82.5497851 2022-01-03 04:29:59
#> 15: -0.6033094  5.9057558     0.01     0  19.1503630 2022-01-03 04:29:59
#> 16:  2.7974630 13.0680755     0.01     0 111.2685667 2022-01-03 04:29:59
#> 17:  3.5262589  2.0494585     0.02     1   0.7500095 2022-01-03 04:29:59
#> 18:  3.5546516  0.4229391     0.02     1   3.2586107 2022-01-03 04:29:59
#> 19: -2.9615141 14.8561522     0.02     1   9.9078132 2022-01-03 04:29:59
#> 20:  1.7415065  5.4600249     0.02     1  11.5799338 2022-01-03 04:29:59
#> 21: -0.6033094  5.9057558     0.02     1  19.1805284 2022-01-03 04:29:59
#> 22:  8.6513194  6.6077093     0.02     1  24.4284669 2022-01-03 04:29:59
#> 23:  4.8570208  6.5519708     0.02     1  38.2994033 2022-01-03 04:29:59
#> 24:  0.3248212  0.7425342     0.02     1  41.6692381 2022-01-03 04:29:59
#> 25:  3.5262589  2.0494585     0.04     2   0.3973836 2022-01-03 04:29:59
#> 26:  3.5546516  0.4229391     0.04     2   2.9031455 2022-01-03 04:29:59
#> 27: -2.9615141 14.8561522     0.04     2  10.2039646 2022-01-03 04:29:59
#> 28:  1.7415065  5.4600249     0.04     2  11.4057831 2022-01-03 04:29:59
#> 29:  3.5262589  2.0494585     0.08     3  -0.3078682 2022-01-03 04:29:59
#> 30:  3.5546516  0.4229391     0.08     3   2.1922152 2022-01-03 04:29:59
#> 31:  3.5262589  2.0494585     0.16     4  -1.7183718 2022-01-03 04:29:59
#>             x1         x2 fidelity stage           y           timestamp
#>     batch_nr
#>  1:        1
#>  2:        1
#>  3:        1
#>  4:        1
#>  5:        1
#>  6:        1
#>  7:        1
#>  8:        1
#>  9:        1
#> 10:        1
#> 11:        1
#> 12:        1
#> 13:        1
#> 14:        1
#> 15:        1
#> 16:        1
#> 17:        2
#> 18:        2
#> 19:        2
#> 20:        2
#> 21:        2
#> 22:        2
#> 23:        2
#> 24:        2
#> 25:        3
#> 26:        3
#> 27:        3
#> 28:        3
#> 29:        4
#> 30:        4
#> 31:        5
#>     batch_nr