Hyperparameter Optimization with Successive Halving

Optimizer using the Successive Halving Algorithm (SHA). SHA is initialized with the number of starting configurations n, the proportion of configurations discarded in each stage eta, and the minimum r_min and maximum _max budget of a single evaluation. The algorithm starts by sampling n random configurations and allocating the minimum budget r_min to them. The configurations are evaluated and 1 / eta of the worst-performing configurations are discarded. The remaining configurations are promoted to the next stage and evaluated on a larger budget. The following table is the stage layout for eta = 2, r_min = 1 and r_max = 8.

i	n_i	r_i

0	8	1

1	4	2

2	2	4

3	1	8

i is the stage number, n_i is the number of configurations and r_i is the budget allocated to a single configuration.

The number of stages is calculated so that each stage consumes approximately the same budget. This sometimes results in the minimum budget having to be slightly adjusted by the algorithm.

Source

Jamieson K, Talwalkar A (2016). “Non-stochastic Best Arm Identification and Hyperparameter Optimization.” In Gretton A, Robert CC (eds.), Proceedings of the 19th International Conference on Artificial Intelligence and Statistics, volume 51 series Proceedings of Machine Learning Research, 240-248. http://proceedings.mlr.press/v51/jamieson16.html.

Resources

The gallery features a collection of case studies and demos about optimization.

Tune the hyperparameters of XGBoost with Hyperband (Hyperband can be easily swapped with SHA).

Use data subsampling and Hyperband to optimize a support vector machine.

Dictionary

This Optimizer can be instantiated via the dictionary mlr_optimizers or with the associated sugar function opt():

mlr_optimizers$get("successive_halving")
opt("successive_halving")

Parameters

n

integer(1)
Number of configurations in the base stage.

eta

numeric(1)
With every stage, the budget is increased by a factor of eta and only the best 1 / eta configurations are promoted to the next stage. Non-integer values are supported, but eta is not allowed to be less or equal to 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn. The default is uniform sampling.

repetitions

integer(1)
If 1 (default), optimization is stopped once all stages are evaluated. Otherwise, optimization is stopped after repetitions runs of SHA. The bbotk::Terminator might stop the optimization before all repetitions are executed.

adjust_minimum_budget

logical(1)
If TRUE, the minimum budget is increased so that the last stage uses the maximum budget defined in the search space.

Archive

The bbotk::Archive holds the following additional columns that are specific to SHA:

stage (integer(1))
Stage index. Starts counting at 0.

repetition (integer(1))
Repetition index. Start counting at 1.

Custom Sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerSuccessiveHalving

Methods

Public methods

OptimizerSuccessiveHalving$new()

OptimizerSuccessiveHalving$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$help()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerSuccessiveHalving$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerSuccessiveHalving$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

# set search space
search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# Branin function with fidelity, see `bbotk::branin()`
fun = function(xs) branin_wu(xs[["x1"]], xs[["x2"]], xs[["fidelity"]])

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

# initialize instance and optimizer
instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("evals", n_evals = 50)
)

optimizer = opt("successive_halving")

# optimize branin function
optimizer$optimize(instance)
#>           x1      x2 fidelity  x_domain        y
#> 1: -3.918878 13.2493     0.16 <list[3]> 3.256725

# best scoring evaluation
instance$result
#>           x1      x2 fidelity  x_domain        y
#> 1: -3.918878 13.2493     0.16 <list[3]> 3.256725

# all evaluations
as.data.table(instance$archive)
#>             x1         x2 fidelity stage repetition          y
#>  1: -0.3372068  2.5626176     0.01     0          1  34.881820
#>  2:  8.3970241 11.0706666     0.01     0          1 270.944531
#>  3: -2.1577893  8.4204167     0.01     0          1   6.014305
#>  4:  1.2884819 13.2046450     0.01     0          1  97.410945
#>  5:  6.7828500  7.4134317     0.01     0          1 135.499994
#>  6:  5.3998678  0.1145088     0.01     0          1  19.436927
#>  7: -3.9188779 13.2492969     0.01     0          1   3.456434
#>  8:  9.2577396 14.8941033     0.01     0          1 443.266485
#>  9:  6.4151535 11.5905662     0.01     0          1 231.465076
#> 10:  2.0551525  8.5388451     0.01     0          1  37.816658
#> 11:  7.3654601 14.1208567     0.01     0          1 345.959112
#> 12:  0.5776038 14.7411405     0.01     0          1 111.173750
#> 13: -4.8482821  3.8263493     0.01     0          1 123.649053
#> 14: -2.3414694 11.6942187     0.01     0          1   6.558804
#> 15:  4.1944802  2.2417142     0.01     0          1  10.941150
#> 16: -3.4531016 14.7766259     0.01     0          1   9.391829
#> 17: -3.9188779 13.2492969     0.02     1          1   3.439818
#> 18: -2.1577893  8.4204167     0.02     1          1   6.025076
#> 19: -2.3414694 11.6942187     0.02     1          1   6.539073
#> 20: -3.4531016 14.7766259     0.02     1          1   9.322314
#> 21:  4.1944802  2.2417142     0.02     1          1  10.857489
#> 22:  5.3998678  0.1145088     0.02     1          1  19.331142
#> 23: -0.3372068  2.5626176     0.02     1          1  34.882724
#> 24:  2.0551525  8.5388451     0.02     1          1  37.768676
#> 25: -3.9188779 13.2492969     0.04     2          1   3.408002
#> 26: -2.1577893  8.4204167     0.04     2          1   6.046748
#> 27: -2.3414694 11.6942187     0.04     2          1   6.499793
#> 28: -3.4531016 14.7766259     0.04     2          1   9.184136
#> 29: -3.9188779 13.2492969     0.08     3          1   3.350029
#> 30: -2.1577893  8.4204167     0.08     3          1   6.090613
#> 31: -3.9188779 13.2492969     0.16     4          1   3.256725
#>             x1         x2 fidelity stage repetition          y
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity
#>  1: 2023-03-02 10:25:48        1  -0.3372068   2.5626176              0.01
#>  2: 2023-03-02 10:25:48        1   8.3970241  11.0706666              0.01
#>  3: 2023-03-02 10:25:48        1  -2.1577893   8.4204167              0.01
#>  4: 2023-03-02 10:25:48        1   1.2884819  13.2046450              0.01
#>  5: 2023-03-02 10:25:48        1   6.7828500   7.4134317              0.01
#>  6: 2023-03-02 10:25:48        1   5.3998678   0.1145088              0.01
#>  7: 2023-03-02 10:25:48        1  -3.9188779  13.2492969              0.01
#>  8: 2023-03-02 10:25:48        1   9.2577396  14.8941033              0.01
#>  9: 2023-03-02 10:25:48        1   6.4151535  11.5905662              0.01
#> 10: 2023-03-02 10:25:48        1   2.0551525   8.5388451              0.01
#> 11: 2023-03-02 10:25:48        1   7.3654601  14.1208567              0.01
#> 12: 2023-03-02 10:25:48        1   0.5776038  14.7411405              0.01
#> 13: 2023-03-02 10:25:48        1  -4.8482821   3.8263493              0.01
#> 14: 2023-03-02 10:25:48        1  -2.3414694  11.6942187              0.01
#> 15: 2023-03-02 10:25:48        1   4.1944802   2.2417142              0.01
#> 16: 2023-03-02 10:25:48        1  -3.4531016  14.7766259              0.01
#> 17: 2023-03-02 10:25:48        2  -3.9188779  13.2492969              0.02
#> 18: 2023-03-02 10:25:48        2  -2.1577893   8.4204167              0.02
#> 19: 2023-03-02 10:25:48        2  -2.3414694  11.6942187              0.02
#> 20: 2023-03-02 10:25:48        2  -3.4531016  14.7766259              0.02
#> 21: 2023-03-02 10:25:48        2   4.1944802   2.2417142              0.02
#> 22: 2023-03-02 10:25:48        2   5.3998678   0.1145088              0.02
#> 23: 2023-03-02 10:25:48        2  -0.3372068   2.5626176              0.02
#> 24: 2023-03-02 10:25:48        2   2.0551525   8.5388451              0.02
#> 25: 2023-03-02 10:25:48        3  -3.9188779  13.2492969              0.04
#> 26: 2023-03-02 10:25:48        3  -2.1577893   8.4204167              0.04
#> 27: 2023-03-02 10:25:48        3  -2.3414694  11.6942187              0.04
#> 28: 2023-03-02 10:25:48        3  -3.4531016  14.7766259              0.04
#> 29: 2023-03-02 10:25:48        4  -3.9188779  13.2492969              0.08
#> 30: 2023-03-02 10:25:48        4  -2.1577893   8.4204167              0.08
#> 31: 2023-03-02 10:25:48        5  -3.9188779  13.2492969              0.16
#>               timestamp batch_nr x_domain_x1 x_domain_x2 x_domain_fidelity