Hyperparameter Optimization with Successive Halving

OptimizerSuccessiveHalving class that implements the successive halving algorithm. The algorithm samples n points and evaluates them with the smallest budget (lower bound of the budget parameter). With every stage the budget is increased by a factor of eta and only the best 1/eta points are promoted to the next stage. The optimization terminates when the maximum budget is reached (upper bound of the budget parameter).

To identify the budget, the user has to specify explicitly which parameter of the objective function influences the budget by tagging a single parameter in the search_space (paradox::ParamSet) with "budget".

Source

Jamieson K, Talwalkar A (2016). “Non-stochastic Best Arm Identification and Hyperparameter Optimization.” In Gretton A, Robert CC (eds.), Proceedings of the 19th International Conference on Artificial Intelligence and Statistics, volume 51 series Proceedings of Machine Learning Research, 240-248. http://proceedings.mlr.press/v51/jamieson16.html.

Parameters

n

integer(1)
Number of points in first stage.

eta

numeric(1)
With every stage, the point budget is increased by a factor of eta and only the best 1/eta points are used for the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn during the initialization of each bracket. The default is uniform sampling.

Archive

The bbotk::Archive holds the following additional column that is specific to the successive halving algorithm:

stage (integer(1))
Stage index. Starts counting at 0.

Custom sampler

Hyperband supports custom paradox::Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Runtime

The calculation of each bracket currently assumes a linear runtime in the chosen budget parameter is always given. Hyperband is designed so each bracket requires approximately the same runtime as the sum of the budget over all configurations in each bracket is roughly the same. This will not hold true once the scaling in the budget parameter is not linear anymore, even though the sum of the budgets in each bracket remains the same. A possible adaption would be to introduce a trafo, like it is shown in the examples section.

Progress Bars

$optimize() supports progress bars via the package progressr combined with a Terminator. Simply wrap the function in progressr::with_progress() to enable them. We recommend to use package progress as backend; enable with progressr::handlers("progress").

Parallelization

In order to support general termination criteria and parallelization, we evaluate points in a batch-fashion of size batch_size. The points of one stage in a bracket are evaluated in one batch. Parallelization is supported via package future (see mlr3::benchmark()'s section on parallelization for more details).

Logging

Hyperband uses a logger (as implemented in lgr) from package bbotk. Use lgr::get_logger("bbotk") to access and control the logger.

Super class

bbotk::Optimizer -> OptimizerSuccessiveHalving

Methods

Public methods

OptimizerSuccessiveHalving$new()

OptimizerSuccessiveHalving$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class.

Usage

OptimizerSuccessiveHalving$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerSuccessiveHalving$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)
library(data.table)

search_space = domain = ps(
  x1 = p_dbl(-5, 10),
  x2 = p_dbl(0, 15),
  fidelity = p_dbl(1e-2, 1, tags = "budget")
)

# modified branin function
objective = ObjectiveRFunDt$new(
  fun = function(xdt) {
    a = 1
    b = 5.1 / (4 * (pi ^ 2))
    c = 5 / pi
    r = 6
    s = 10
    t = 1 / (8 * pi)
    data.table(y =
      (a * ((xdt[["x2"]] -
      b * (xdt[["x1"]] ^ 2L) +
      c * xdt[["x1"]] - r) ^ 2) +
      ((s * (1 - t)) * cos(xdt[["x1"]])) +
      s - (5 * xdt[["fidelity"]] * xdt[["x1"]])))
  },
  domain = domain,
  codomain = ps(y = p_dbl(tags = "minimize"))
)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  search_space = search_space,
  terminator = trm("none")
)

optimizer = opt("successive_halving")

# modifies the instance by reference
optimizer$optimize(instance)

#>           x1       x2 fidelity  x_domain         y
#> 1: -3.070961 11.68456     0.01 <list[3]> 0.7529051

# best scoring evaluation
instance$result

#>           x1       x2 fidelity  x_domain         y
#> 1: -3.070961 11.68456     0.01 <list[3]> 0.7529051

# all evaluations
as.data.table(instance$archive)

#>             x1         x2 fidelity stage           y           timestamp
#>  1:  4.0562560  2.4751648     0.01     0   4.5880116 2021-08-01 04:58:20
#>  2:  7.0819560  8.5036468     0.01     0  69.5728737 2021-08-01 04:58:20
#>  3: -4.4498089 13.4711606     0.01     0  12.4341098 2021-08-01 04:58:20
#>  4:  5.9936600  8.9170842     0.01     0  79.9843753 2021-08-01 04:58:20
#>  5: -1.7709101 12.4753488     0.01     0  18.7535078 2021-08-01 04:58:20
#>  6: -4.7599528  8.9011255     0.01     0  68.4777621 2021-08-01 04:58:20
#>  7: -3.0709613 11.6845595     0.01     0   0.7529051 2021-08-01 04:58:20
#>  8:  5.2909466  5.9665747     0.01     0  37.7482198 2021-08-01 04:58:20
#>  9:  4.6291190 12.7482416     0.01     0 137.7345598 2021-08-01 04:58:20
#> 10: -0.0899222 11.1276837     0.01     0  44.4033202 2021-08-01 04:58:20
#> 11:  0.8122297  4.7668533     0.01     0  16.5651656 2021-08-01 04:58:20
#> 12:  5.6125059  1.6752034     0.01     0  17.5315870 2021-08-01 04:58:20
#> 13:  5.4648314  1.5164305     0.01     0  16.4157815 2021-08-01 04:58:20
#> 14:  8.8848515 12.0014662     0.01     0 100.2076016 2021-08-01 04:58:20
#> 15:  1.8852659  5.6992011     0.01     0  11.9557240 2021-08-01 04:58:20
#> 16:  3.9393130  0.7905146     0.01     0   3.9897309 2021-08-01 04:58:20
#> 17: -3.0709613 11.6845595     0.02     1   0.9064532 2021-08-01 04:58:20
#> 18:  3.9393130  0.7905146     0.02     1   3.7927653 2021-08-01 04:58:20
#> 19:  4.0562560  2.4751648     0.02     1   4.3851988 2021-08-01 04:58:20
#> 20:  1.8852659  5.6992011     0.02     1  11.8614607 2021-08-01 04:58:20
#> 21: -4.4498089 13.4711606     0.02     1  12.6566002 2021-08-01 04:58:20
#> 22:  5.4648314  1.5164305     0.02     1  16.1425399 2021-08-01 04:58:20
#> 23:  0.8122297  4.7668533     0.02     1  16.5245541 2021-08-01 04:58:20
#> 24:  5.6125059  1.6752034     0.02     1  17.2509617 2021-08-01 04:58:20
#> 25: -3.0709613 11.6845595     0.04     2   1.2135493 2021-08-01 04:58:20
#> 26:  3.9393130  0.7905146     0.04     2   3.3988339 2021-08-01 04:58:20
#> 27:  4.0562560  2.4751648     0.04     2   3.9795732 2021-08-01 04:58:20
#> 28:  1.8852659  5.6992011     0.04     2  11.6729342 2021-08-01 04:58:20
#> 29: -3.0709613 11.6845595     0.08     3   1.8277416 2021-08-01 04:58:20
#> 30:  3.9393130  0.7905146     0.08     3   2.6109713 2021-08-01 04:58:20
#> 31: -3.0709613 11.6845595     0.16     4   3.0561261 2021-08-01 04:58:20
#>             x1         x2 fidelity stage           y           timestamp
#>     batch_nr
#>  1:        1
#>  2:        1
#>  3:        1
#>  4:        1
#>  5:        1
#>  6:        1
#>  7:        1
#>  8:        1
#>  9:        1
#> 10:        1
#> 11:        1
#> 12:        1
#> 13:        1
#> 14:        1
#> 15:        1
#> 16:        1
#> 17:        2
#> 18:        2
#> 19:        2
#> 20:        2
#> 21:        2
#> 22:        2
#> 23:        2
#> 24:        2
#> 25:        3
#> 26:        3
#> 27:        3
#> 28:        3
#> 29:        4
#> 30:        4
#> 31:        5
#>     batch_nr