Tuner using the Hyperband algorithm

TunerHyperband class that implements hyperband tuning. Hyperband is a budget oriented-procedure, weeding out suboptimal performing configurations early in a sequential training process, increasing tuning efficiency as a consequence.

For this, several brackets are constructed with an associated set of configurations for each bracket. Each bracket as several stages. Different brackets are initialized with different amounts of configurations and different budget sizes. To get an idea of how the bracket layout looks like for a given argument set, please have a look in the details.

To identify the budget for evaluating hyperband, the user has to specify explicitly which hyperparameter of the learner influences the budget by tagging a single hyperparameter in the paradox::ParamSet with "budget". An alternative approach using subsampling and pipelines is described below.

Naturally, hyperband terminates once all of its brackets are evaluated, so a bbotk::Terminator in the tuning instance acts as an upper bound and should be only set to a low value if one is unsure of how long hyperband will take to finish under the given settings.

Details

This sections explains the calculation of the constants for each bracket. A small overview will be given here, but for more details please check out the original paper (see references). To keep things uniform with the notation in the paper (and to safe space in the formulas), R is used for the upper budget that last remaining configuration should reach. The formula to calculate the amount of brackets is floor(log(R, eta)) + 1. To calculate the starting budget in each bracket, use R * eta^(-s), where s is the maximum bracket minus the current bracket index. For the starting configurations in each bracket it is ceiling((B/R) * ((eta^s)/(s+1))), with B = (bracket amount) * R. To receive a table with the full brackets layout, load the following function and execute it for the desired R and eta.

hyperband_brackets = function(R, eta) {

  result = data.frame()
  smax = floor(log(R, eta))
  B = (smax + 1) * R

  # outer loop - iterate over brackets
  for (s in smax:0) {

    n = ceiling((B/R) * ((eta^s)/(s+1)))
    r = R * eta^(-s)

    # inner loop - iterate over bracket stages
    for (i in 0:s) {

      ni = floor(n * eta^(-i))
      ri = r * eta^i
      result = rbind(result, c(smax - s + 1, i + 1, ri, ni))
    }
  }

  names(result) = c("bracket", "bracket_stage", "budget", "n_configs")
  return(result)
}

hyperband_brackets(R = 81L, eta = 3L)

Parameters

eta

numeric(1)
Fraction parameter of the successive halving algorithm: With every step the configuration budget is increased by a factor of eta and only the best 1/eta configurations are used for the next stage. Non-integer values are supported, but eta is not allowed to be less or equal 1.

sampler

paradox::Sampler
Object defining how the samples of the parameter space should be drawn during the initialization of each bracket. The default is uniform sampling.

Archive

The mlr3tuning::ArchiveTuning holds the following additional columns that are specific to the hyperband tuner:

bracket (integer())
The console logs about the bracket index are actually not matching with the original hyperband algorithm, which counts down the brackets and stops after evaluating bracket 0. The true bracket indices are given in this column.

bracket_stage (integer())
The bracket stage of each bracket. Hyperband starts counting at 0.

budget_scaled (integer())
The intermediate budget in each bracket stage calculated by hyperband. Because hyperband is originally only considered for budgets starting at 1, some rescaling is done to allow budgets starting at different values. For this, budgets are internally divided by the lower budget bound to get a lower budget of 1. Before the learner receives its budgets for evaluation, the budget is transformed back to match the original scale again.

budget_real (integer())
The real budget values the learner uses for evaluation after hyperband calculated its scaled budget.

n_configs (integer())
The amount of evaluated configurations in each stage. These correspond to the r_i in the original paper.

Hyperband without learner budget

Thanks to mlr3pipelines, it is possible to use hyperband in combination with learners lacking a natural budget parameter. For example, any mlr3::Learner can be augmented with a PipeOp operator such as PipeOpSubsample. With the subsampling rate as budget parameter, the resulting GraphLearner is fitted on small proportions of the Task in the first brackets, and on the complete Task in last brackets. See examples for some code.

Custom sampler

Hyperband supports custom Sampler object for initial configurations in each bracket. A custom sampler may look like this (the full example is given in the examples section):

# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

Runtime scaling w.r.t. the chosen budget

The calculation of each bracket currently assumes a linear runtime in the chosen budget parameter is always given. Hyperband is designed so each bracket requires approximately the same runtime as the sum of the budget over all configurations in each bracket is roughly the same. This will not hold true once the scaling in the budget parameter is not linear anymore, even though the sum of the budgets in each bracket remains the same. A basic example can be viewed by calling the function hyperband_brackets below with the arguments R = 2 and eta = 2. If we run a learner with O(budget^2) time complexity, the runtime of the last bracket will be 33% longer than the first bracket (time of bracket 1 = 2 * 1^2 + 2^2 = 6; time of bracket 2 = 2 * 2^2 = 8). Of course, this won't break anything, but it should be kept in mind when applying hyperband. A possible adaption would be to introduce a trafo, like it is shown in the examples.

Logging

When loading the mlr3hyperband package, two loggers based on the lgr package are made available. One is called mlr3, the other bbotk. All mlr3 methods log into the mlr3 logger. All optimization methods form the packags bbotk, mlr3tuning and mlr3hyperband log into the bbotk logger. To hide the mlr3 logging messages run:

lgr::get_logger("mlr3")$set_threshold("warn")

References

Li L, Jamieson K, DeSalvo G, Rostamizadeh A, Talwalkar A (2018). “Hyperband: A Novel Bandit-Based Approach to Hyperparameter Optimization.” Journal of Machine Learning Research, 18(185), 1-52. http://jmlr.org/papers/v18/16-558.html.

Super class

mlr3tuning::Tuner -> TunerHyperband

Methods

Public methods

TunerHyperband$new()

TunerHyperband$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$optimize()

mlr3tuning::Tuner$print()

Method new()

Creates a new instance of this R6 class.

Usage

TunerHyperband$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerHyperband$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3hyperband)
library(mlr3learners)
library(mlr3tuning)
library(mlr3)
library(paradox)

set.seed(123)

# define hyperparameter and budget parameter for tuning with hyperband
params = list(
  ParamInt$new("nrounds", lower = 1, upper = 16, tag = "budget"),
  ParamDbl$new("eta", lower = 0, upper = 1),
  ParamFct$new("booster", levels = c("gbtree", "gblinear", "dart"))
)

inst = TuningInstanceSingleCrit$new(
  tsk("iris"),
  lrn("classif.xgboost"),
  rsmp("holdout"),
  msr("classif.ce"),
  ParamSet$new(params),
  trm("evals", n_evals = 100000)
)

# create custom sampler (optional):
# - beta distribution with alpha = 2 and beta = 5
# - categorical distribution with custom probabilities
sampler = SamplerJointIndep$new(list(
  Sampler1DRfun$new(params[[2]], function(n) rbeta(n, 2, 5)),
  Sampler1DCateg$new(params[[3]], prob = c(0.2, 0.3, 0.5))
))

tuner = tnr("hyperband", eta = 2L, sampler = sampler)
# \donttest{
tuner$optimize(inst)

#>    nrounds      eta booster learner_param_vals  x_domain classif.ce
#> 1:       1 0.231625    dart          <list[4]> <list[3]>       0.04

# return the best evaluation
inst$result

#>    nrounds      eta booster learner_param_vals  x_domain classif.ce
#> 1:       1 0.231625    dart          <list[4]> <list[3]>       0.04

# print all evaluations
print(inst$archive$data())

#>            eta  booster nrounds bracket bracket_stage budget_scaled budget_real
#>  1: 0.16633231 gblinear       1       4             0             1           1
#>  2: 0.53672392 gblinear       1       4             0             1           1
#>  3: 0.23162500     dart       1       4             0             1           1
#>  4: 0.09921319     dart       1       4             0             1           1
#>  5: 0.32374930     dart       1       4             0             1           1
#>  6: 0.25848344 gblinear       1       4             0             1           1
#>  7: 0.28687990 gblinear       1       4             0             1           1
#>  8: 0.36995279   gbtree       1       4             0             1           1
#>  9: 0.21663062 gblinear       1       4             0             1           1
#> 10: 0.43375908     dart       1       4             0             1           1
#> 11: 0.24324338 gblinear       1       4             0             1           1
#> 12: 0.35748527     dart       1       4             0             1           1
#> 13: 0.38180400     dart       1       4             0             1           1
#> 14: 0.22436454     dart       1       4             0             1           1
#> 15: 0.57168085     dart       1       4             0             1           1
#> 16: 0.52772844   gbtree       1       4             0             1           1
#> 17: 0.23162500     dart       2       4             1             2           2
#> 18: 0.09921319     dart       2       4             1             2           2
#> 19: 0.32374930     dart       2       4             1             2           2
#> 20: 0.36995279   gbtree       2       4             1             2           2
#> 21: 0.43375908     dart       2       4             1             2           2
#> 22: 0.35748527     dart       2       4             1             2           2
#> 23: 0.38180400     dart       2       4             1             2           2
#> 24: 0.22436454     dart       2       4             1             2           2
#> 25: 0.23162500     dart       4       4             2             4           4
#> 26: 0.09921319     dart       4       4             2             4           4
#> 27: 0.32374930     dart       4       4             2             4           4
#> 28: 0.36995279   gbtree       4       4             2             4           4
#> 29: 0.23162500     dart       8       4             3             8           8
#> 30: 0.09921319     dart       8       4             3             8           8
#> 31: 0.23162500     dart      16       4             4            16          16
#> 32: 0.12152038     dart       2       3             0             2           2
#> 33: 0.31007089     dart       2       3             0             2           2
#> 34: 0.13099165     dart       2       3             0             2           2
#> 35: 0.19669320     dart       2       3             0             2           2
#> 36: 0.23676637     dart       2       3             0             2           2
#> 37: 0.17165036 gblinear       2       3             0             2           2
#> 38: 0.33564743     dart       2       3             0             2           2
#> 39: 0.30172118     dart       2       3             0             2           2
#> 40: 0.12918127     dart       2       3             0             2           2
#> 41: 0.27152929   gbtree       2       3             0             2           2
#> 42: 0.12152038     dart       4       3             1             4           4
#> 43: 0.31007089     dart       4       3             1             4           4
#> 44: 0.13099165     dart       4       3             1             4           4
#> 45: 0.19669320     dart       4       3             1             4           4
#> 46: 0.23676637     dart       4       3             1             4           4
#> 47: 0.12152038     dart       8       3             2             8           8
#> 48: 0.31007089     dart       8       3             2             8           8
#> 49: 0.12152038     dart      16       3             3            16          16
#> 50: 0.44501551     dart       4       2             0             4           4
#> 51: 0.09232408     dart       4       2             0             4           4
#> 52: 0.47135712   gbtree       4       2             0             4           4
#> 53: 0.46636871     dart       4       2             0             4           4
#> 54: 0.35758233     dart       4       2             0             4           4
#> 55: 0.12343101   gbtree       4       2             0             4           4
#> 56: 0.26219631 gblinear       4       2             0             4           4
#> 57: 0.44501551     dart       8       2             1             8           8
#> 58: 0.09232408     dart       8       2             1             8           8
#> 59: 0.47135712   gbtree       8       2             1             8           8
#> 60: 0.44501551     dart      16       2             2            16          16
#> 61: 0.18500249   gbtree       8       1             0             8           8
#> 62: 0.52125349 gblinear       8       1             0             8           8
#> 63: 0.28196564     dart       8       1             0             8           8
#> 64: 0.36458346     dart       8       1             0             8           8
#> 65: 0.10382335   gbtree       8       1             0             8           8
#> 66: 0.18500249   gbtree      16       1             1            16          16
#> 67: 0.28196564     dart      16       1             1            16          16
#> 68: 0.11684580 gblinear      16       0             0            16          16
#> 69: 0.60449159 gblinear      16       0             0            16          16
#> 70: 0.22024316     dart      16       0             0            16          16
#> 71: 0.14160471 gblinear      16       0             0            16          16
#> 72: 0.24035756 gblinear      16       0             0            16          16
#>            eta  booster nrounds bracket bracket_stage budget_scaled budget_real
#>     n_configs classif.ce                                uhash  x_domain
#>  1:        16       0.74 b99465ad-a300-4bc7-a4f7-5371a511949a <list[3]>
#>  2:        16       0.50 6c1f621c-7558-4cd0-8ed0-9c9f9f5d64ab <list[3]>
#>  3:        16       0.04 546b6dc4-7c72-40a5-8fe3-6bcfad8bdbc0 <list[3]>
#>  4:        16       0.04 a2445435-b1e4-48dd-8d79-c57227a897bc <list[3]>
#>  5:        16       0.04 3eeeedc5-00dc-498b-a6c5-43e55ed6933a <list[3]>
#>  6:        16       0.74 36fd51cd-d591-49db-996a-dd7cc824c70c <list[3]>
#>  7:        16       0.74 9c8e8b8b-65c9-4e5c-9a7a-80e6b824e204 <list[3]>
#>  8:        16       0.04 0bdb2f0c-1f8f-4756-9de2-5d59b9575e43 <list[3]>
#>  9:        16       0.74 edaed2bb-bd6b-4abe-a7fd-3a8b17c0ccb6 <list[3]>
#> 10:        16       0.04 21dafce2-83f6-4038-b6b8-520b90ff451c <list[3]>
#> 11:        16       0.72 a6666f44-f605-4eeb-8376-0fae1327c323 <list[3]>
#> 12:        16       0.04 1c1edaa6-ab7e-40d2-9361-d28e47fa06c7 <list[3]>
#> 13:        16       0.04 3f47b974-be25-43f5-b5a0-f26fdd1bd355 <list[3]>
#> 14:        16       0.04 e38cc5ce-9f01-4ab8-824f-676db478fe6c <list[3]>
#> 15:        16       0.04 98a92056-0afa-447f-97c8-c3c9aa0aab92 <list[3]>
#> 16:        16       0.04 465493f9-2052-4558-ade4-2d455cbeae9f <list[3]>
#> 17:         8       0.04 e2febe48-6842-4c21-9e28-90528ff32197 <list[3]>
#> 18:         8       0.04 85b386dc-ec39-4511-b8bd-8ea703c27d32 <list[3]>
#> 19:         8       0.04 d906b470-6070-479a-a788-b41c777a7f21 <list[3]>
#> 20:         8       0.04 5092e1bb-85df-45c7-af3b-985bba31d4b6 <list[3]>
#> 21:         8       0.04 39dd5ab5-f42f-407c-81fb-9239b2a542a0 <list[3]>
#> 22:         8       0.04 327e85fa-9228-4a6e-bf40-a7504b279b28 <list[3]>
#> 23:         8       0.04 d956d04c-d2ea-41c4-ae71-191c0d0baa48 <list[3]>
#> 24:         8       0.04 05b185b3-f667-40ed-b053-cde1a4e6477b <list[3]>
#> 25:         4       0.04 daf77529-9768-4e5a-81db-e8ecbc732223 <list[3]>
#> 26:         4       0.04 ad5c3117-52c2-431e-ba58-3d12334543d9 <list[3]>
#> 27:         4       0.04 127f26f3-b15c-47b9-9b90-87a60d157170 <list[3]>
#> 28:         4       0.04 7197f9a5-af8d-4fef-aedf-633abe6260a7 <list[3]>
#> 29:         2       0.04 94b07f05-1d04-4ee6-8137-0f7c481287cb <list[3]>
#> 30:         2       0.04 de937a08-1638-40ba-8eb6-7fa6acc46b8d <list[3]>
#> 31:         1       0.04 6526ce3e-d9b2-47c5-8dbe-71c9336886e5 <list[3]>
#> 32:        10       0.04 aba179b8-797f-49dc-bd84-88553ed61e78 <list[3]>
#> 33:        10       0.04 c992f4c2-a2c5-4232-a2fc-b2526262da13 <list[3]>
#> 34:        10       0.04 bb0c7a51-c7dc-4fa3-ad87-300ee55b0b63 <list[3]>
#> 35:        10       0.04 19f586f9-5495-4c7b-9f88-43e7f2a39885 <list[3]>
#> 36:        10       0.04 7f393669-c945-4229-8e62-6693192eb7cd <list[3]>
#> 37:        10       0.74 69f92290-001e-4d5e-b3cd-79ae8aa81b65 <list[3]>
#> 38:        10       0.04 47360d7d-1575-4178-9774-a726a6ea1823 <list[3]>
#> 39:        10       0.04 6db4734c-803f-4f81-852c-e2a9a32578ff <list[3]>
#> 40:        10       0.04 9c36631b-bfb2-4328-a3c3-f7d2d59e36d1 <list[3]>
#> 41:        10       0.04 e167e030-f912-4df0-a3e7-d5e0fbc1f4bf <list[3]>
#> 42:         5       0.04 aad3094b-32c1-48a7-b960-e203295d7e93 <list[3]>
#> 43:         5       0.04 af45801d-e6a4-45d0-a01d-2f2617866b48 <list[3]>
#> 44:         5       0.04 c481f2e6-e813-4535-8393-fae817b566b2 <list[3]>
#> 45:         5       0.04 749c5266-06db-4806-bea7-eb5719b75c17 <list[3]>
#> 46:         5       0.04 c01c3387-bebe-40b0-8dc3-d3a3c53803fd <list[3]>
#> 47:         2       0.04 11477025-b95e-4012-890d-66008afc2365 <list[3]>
#> 48:         2       0.04 c952e14f-43a3-4879-b54a-d18a4970241d <list[3]>
#> 49:         1       0.04 14d996f0-0533-4959-85cf-61d38e3b9884 <list[3]>
#> 50:         7       0.04 f71cb78b-9794-4236-9fb6-0d7af867ae4a <list[3]>
#> 51:         7       0.04 519092ed-d545-4c97-8aae-38ccbbf5f75b <list[3]>
#> 52:         7       0.04 9f91e31d-0551-4673-a04f-240428040f40 <list[3]>
#> 53:         7       0.04 446cc403-c6b9-4899-b744-c66d76fffb38 <list[3]>
#> 54:         7       0.04 798494cb-cfaa-4428-9bed-8680997b0f09 <list[3]>
#> 55:         7       0.04 c17b063e-56fa-4db3-ba5e-b9f114dd089b <list[3]>
#> 56:         7       0.42 b17a2625-e69c-4700-a08b-a47d54224f35 <list[3]>
#> 57:         3       0.04 340ed9ba-d78f-4876-a174-8455152c5280 <list[3]>
#> 58:         3       0.04 e1b7183f-a7a0-4a82-8ba6-2667e792ce3d <list[3]>
#> 59:         3       0.04 ce869037-30bf-4706-a8b1-4788e1d06a03 <list[3]>
#> 60:         1       0.04 c6f24c9e-62c7-4dcf-9f03-e6ae32521fce <list[3]>
#> 61:         5       0.04 c44e6236-4110-4a63-a6a2-cdfd8efacae2 <list[3]>
#> 62:         5       0.42 37f7255d-db72-40e4-8555-f8617ff09c90 <list[3]>
#> 63:         5       0.04 45f07b55-70d5-4295-888c-f32d44973cbc <list[3]>
#> 64:         5       0.04 821235d5-d490-486a-ab74-ed8e934c81cf <list[3]>
#> 65:         5       0.04 73150d7a-758f-407a-a53b-929ed66088cd <list[3]>
#> 66:         2       0.04 944d687e-b4cb-4fe9-a205-c72682e0e3e4 <list[3]>
#> 67:         2       0.04 37514c53-b95c-434e-9bed-2310f9fcd253 <list[3]>
#> 68:         5       0.42 1da7e84e-0104-443e-8499-c001b94442e7 <list[3]>
#> 69:         5       0.18 fe4ab311-1284-4109-8dc8-2135bc0c83cb <list[3]>
#> 70:         5       0.04 f4e380ec-a713-4809-a867-742ec9c30fbe <list[3]>
#> 71:         5       0.42 05dfba13-dc51-42af-b831-823fdd3bb5ac <list[3]>
#> 72:         5       0.42 92b44b40-7aa5-464b-b7fa-0e53d2382210 <list[3]>
#>     n_configs classif.ce                                uhash  x_domain
#>               timestamp batch_nr
#>  1: 2020-09-08 22:36:17        1
#>  2: 2020-09-08 22:36:17        1
#>  3: 2020-09-08 22:36:17        1
#>  4: 2020-09-08 22:36:17        1
#>  5: 2020-09-08 22:36:17        1
#>  6: 2020-09-08 22:36:17        1
#>  7: 2020-09-08 22:36:17        1
#>  8: 2020-09-08 22:36:17        1
#>  9: 2020-09-08 22:36:17        1
#> 10: 2020-09-08 22:36:17        1
#> 11: 2020-09-08 22:36:17        1
#> 12: 2020-09-08 22:36:17        1
#> 13: 2020-09-08 22:36:17        1
#> 14: 2020-09-08 22:36:17        1
#> 15: 2020-09-08 22:36:17        1
#> 16: 2020-09-08 22:36:17        1
#> 17: 2020-09-08 22:36:17        2
#> 18: 2020-09-08 22:36:17        2
#> 19: 2020-09-08 22:36:17        2
#> 20: 2020-09-08 22:36:17        2
#> 21: 2020-09-08 22:36:17        2
#> 22: 2020-09-08 22:36:17        2
#> 23: 2020-09-08 22:36:17        2
#> 24: 2020-09-08 22:36:17        2
#> 25: 2020-09-08 22:36:18        3
#> 26: 2020-09-08 22:36:18        3
#> 27: 2020-09-08 22:36:18        3
#> 28: 2020-09-08 22:36:18        3
#> 29: 2020-09-08 22:36:18        4
#> 30: 2020-09-08 22:36:18        4
#> 31: 2020-09-08 22:36:18        5
#> 32: 2020-09-08 22:36:18        6
#> 33: 2020-09-08 22:36:18        6
#> 34: 2020-09-08 22:36:18        6
#> 35: 2020-09-08 22:36:18        6
#> 36: 2020-09-08 22:36:18        6
#> 37: 2020-09-08 22:36:18        6
#> 38: 2020-09-08 22:36:18        6
#> 39: 2020-09-08 22:36:18        6
#> 40: 2020-09-08 22:36:18        6
#> 41: 2020-09-08 22:36:18        6
#> 42: 2020-09-08 22:36:19        7
#> 43: 2020-09-08 22:36:19        7
#> 44: 2020-09-08 22:36:19        7
#> 45: 2020-09-08 22:36:19        7
#> 46: 2020-09-08 22:36:19        7
#> 47: 2020-09-08 22:36:19        8
#> 48: 2020-09-08 22:36:19        8
#> 49: 2020-09-08 22:36:19        9
#> 50: 2020-09-08 22:36:19       10
#> 51: 2020-09-08 22:36:19       10
#> 52: 2020-09-08 22:36:19       10
#> 53: 2020-09-08 22:36:19       10
#> 54: 2020-09-08 22:36:19       10
#> 55: 2020-09-08 22:36:19       10
#> 56: 2020-09-08 22:36:19       10
#> 57: 2020-09-08 22:36:20       11
#> 58: 2020-09-08 22:36:20       11
#> 59: 2020-09-08 22:36:20       11
#> 60: 2020-09-08 22:36:20       12
#> 61: 2020-09-08 22:36:20       13
#> 62: 2020-09-08 22:36:20       13
#> 63: 2020-09-08 22:36:20       13
#> 64: 2020-09-08 22:36:20       13
#> 65: 2020-09-08 22:36:20       13
#> 66: 2020-09-08 22:36:20       14
#> 67: 2020-09-08 22:36:20       14
#> 68: 2020-09-08 22:36:21       15
#> 69: 2020-09-08 22:36:21       15
#> 70: 2020-09-08 22:36:21       15
#> 71: 2020-09-08 22:36:21       15
#> 72: 2020-09-08 22:36:21       15
#>               timestamp batch_nr

# }

### use parameter trafo to convert budget parameter

set.seed(123)

# define hyperparameter and budget parameter for tuning with hyperband
ps = ParamSet$new(list(
  ParamInt$new("nrounds", lower = 1, upper = 10, tags = "budget"),
  # ParamDbl$new("eta",     lower = 0, upper = 1),
  ParamFct$new("booster", levels = c("gbtree", "gblinear", "dart"))
))

ps$trafo = function(x, param_set) {
  x$nrounds = round(log(x$nrounds)) + 1L
  return(x)
}

inst = TuningInstanceSingleCrit$new(
  tsk("iris"),
  lrn("classif.xgboost"),
  rsmp("holdout"),
  msr("classif.ce"),
  ps,
  trm("evals", n_evals = 100000)
)

# eta can be a double
tuner = tnr("hyperband", eta = 1.9)
# \donttest{
tuner$optimize(inst)

#>    nrounds booster learner_param_vals  x_domain classif.ce
#> 1:       1  gbtree          <list[3]> <list[2]>       0.04

# return the best evaluation
inst$result

#>    nrounds booster learner_param_vals  x_domain classif.ce
#> 1:       1  gbtree          <list[3]> <list[2]>       0.04

# print all evaluations
print(inst$archive$data())

#>      booster nrounds bracket bracket_stage budget_scaled budget_real n_configs
#>  1:   gbtree       1       3             0      1.457938           1         7
#>  2:     dart       1       3             0      1.457938           1         7
#>  3:     dart       1       3             0      1.457938           1         7
#>  4: gblinear       1       3             0      1.457938           1         7
#>  5: gblinear       1       3             0      1.457938           1         7
#>  6:   gbtree       1       3             0      1.457938           1         7
#>  7:   gbtree       1       3             0      1.457938           1         7
#>  8:   gbtree       3       3             1      2.770083           3         3
#>  9:     dart       3       3             1      2.770083           3         3
#> 10:     dart       3       3             1      2.770083           3         3
#> 11:   gbtree       5       3             2      5.263158           5         1
#> 12:   gbtree      10       3             3     10.000000          10         1
#> 13: gblinear       3       2             0      2.770083           3         5
#> 14: gblinear       3       2             0      2.770083           3         5
#> 15:   gbtree       3       2             0      2.770083           3         5
#> 16: gblinear       3       2             0      2.770083           3         5
#> 17:   gbtree       3       2             0      2.770083           3         5
#> 18:   gbtree       5       2             1      5.263158           5         2
#> 19:   gbtree       5       2             1      5.263158           5         2
#> 20:   gbtree      10       2             2     10.000000          10         1
#> 21: gblinear       5       1             0      5.263158           5         4
#> 22: gblinear       5       1             0      5.263158           5         4
#> 23: gblinear       5       1             0      5.263158           5         4
#> 24: gblinear       5       1             0      5.263158           5         4
#> 25: gblinear      10       1             1     10.000000          10         2
#> 26: gblinear      10       1             1     10.000000          10         2
#> 27: gblinear      10       0             0     10.000000          10         4
#> 28: gblinear      10       0             0     10.000000          10         4
#> 29:     dart      10       0             0     10.000000          10         4
#> 30:   gbtree      10       0             0     10.000000          10         4
#>      booster nrounds bracket bracket_stage budget_scaled budget_real n_configs
#>     classif.ce                                uhash  x_domain
#>  1:       0.04 e3134741-ef1f-4179-acce-65354bac7eca <list[2]>
#>  2:       0.04 e19bb0f9-7d4b-46f1-b01a-68208059a7d4 <list[2]>
#>  3:       0.04 a428ce01-acdc-4223-9955-c7c9c26b0378 <list[2]>
#>  4:       0.42 72224c44-3124-4bb3-995e-ccd12333cb3f <list[2]>
#>  5:       0.44 c0b6136b-12f4-4014-8dd9-3bb3a01d7989 <list[2]>
#>  6:       0.04 03be17c7-40dc-4d2a-8ce2-9df2e7505746 <list[2]>
#>  7:       0.04 f8301d10-bd46-4d45-a530-170d4cfcc61c <list[2]>
#>  8:       0.04 95caccd6-53a0-48bb-9cd2-92d652231f7b <list[2]>
#>  9:       0.04 6d2297c6-62ec-4fc1-86d3-22726bc35684 <list[2]>
#> 10:       0.04 5d286ca3-c14a-49ed-99de-297cc8820b90 <list[2]>
#> 11:       0.04 dda769fd-b638-4deb-99f8-0c1ec6bca518 <list[2]>
#> 12:       0.04 52414a16-9c6a-4aaa-ba4c-39b4c06b8be4 <list[2]>
#> 13:       0.42 e5fb97e1-6c40-471f-afa3-f92f67268345 <list[2]>
#> 14:       0.42 72e30e1f-97e2-486d-a404-17c8086d6495 <list[2]>
#> 15:       0.04 385712d4-dc4a-4b64-bff0-f3394fcb5fd8 <list[2]>
#> 16:       0.42 beee7328-ba2b-49ee-910d-506053fbab85 <list[2]>
#> 17:       0.04 3572ef46-919e-4157-b8a0-916f8a7174af <list[2]>
#> 18:       0.04 b4143999-2eea-426f-9edf-448d53629e43 <list[2]>
#> 19:       0.04 f8b27ae4-a684-4e76-a0ba-42fea8ef9c69 <list[2]>
#> 20:       0.04 e88a7ccb-b00f-4ae6-b73c-d1700fe24a30 <list[2]>
#> 21:       0.42 a5edba13-8c5f-4d1b-9e70-c23a9030e090 <list[2]>
#> 22:       0.42 2a447906-bd8f-4b7d-aecd-499d5d95278f <list[2]>
#> 23:       0.42 e97a7b12-be30-4e87-bf75-ceaac9b2b0fd <list[2]>
#> 24:       0.42 8d065679-5136-48d3-bb79-843ea162abba <list[2]>
#> 25:       0.42 fbaa5b6d-cdeb-43a6-8671-f0b2487bf874 <list[2]>
#> 26:       0.42 02579ec9-fe7f-4908-88c0-d06bc0630279 <list[2]>
#> 27:       0.42 30182a3f-46b7-461c-98a3-efe6f9b0def9 <list[2]>
#> 28:       0.42 25638be0-896b-48ad-a250-906f08a6037f <list[2]>
#> 29:       0.04 2432ac15-11a0-4f21-a979-004fde9ebd52 <list[2]>
#> 30:       0.04 9083567a-f1f9-4438-b6bb-45446494edbc <list[2]>
#>     classif.ce                                uhash  x_domain
#>               timestamp batch_nr
#>  1: 2020-09-08 22:36:21        1
#>  2: 2020-09-08 22:36:21        1
#>  3: 2020-09-08 22:36:21        1
#>  4: 2020-09-08 22:36:21        1
#>  5: 2020-09-08 22:36:21        1
#>  6: 2020-09-08 22:36:21        1
#>  7: 2020-09-08 22:36:21        1
#>  8: 2020-09-08 22:36:22        2
#>  9: 2020-09-08 22:36:22        2
#> 10: 2020-09-08 22:36:22        2
#> 11: 2020-09-08 22:36:22        3
#> 12: 2020-09-08 22:36:22        4
#> 13: 2020-09-08 22:36:22        5
#> 14: 2020-09-08 22:36:22        5
#> 15: 2020-09-08 22:36:22        5
#> 16: 2020-09-08 22:36:22        5
#> 17: 2020-09-08 22:36:22        5
#> 18: 2020-09-08 22:36:22        6
#> 19: 2020-09-08 22:36:22        6
#> 20: 2020-09-08 22:36:22        7
#> 21: 2020-09-08 22:36:23        8
#> 22: 2020-09-08 22:36:23        8
#> 23: 2020-09-08 22:36:23        8
#> 24: 2020-09-08 22:36:23        8
#> 25: 2020-09-08 22:36:23        9
#> 26: 2020-09-08 22:36:23        9
#> 27: 2020-09-08 22:36:23       10
#> 28: 2020-09-08 22:36:23       10
#> 29: 2020-09-08 22:36:23       10
#> 30: 2020-09-08 22:36:23       10
#>               timestamp batch_nr

# }

### use subsampling for budget

library(mlr3pipelines)
set.seed(123)
ll = po("subsample") %>>% lrn("classif.rpart")

# define with extended hyperparameters with subsampling fraction as budget
# ==> no learner budget is required
params = list(
  ParamDbl$new("classif.rpart.cp", lower = 0.001, upper = 0.1),
  ParamInt$new("classif.rpart.minsplit", lower = 1, upper = 10),
  ParamDbl$new("subsample.frac", lower = 0.1, upper = 1, tags = "budget")
)

# define TuningInstanceSingleCrit with the Graph Learner and the extended hyperparams
inst = TuningInstanceSingleCrit$new(
  tsk("iris"),
  ll,
  rsmp("holdout"),
  msr("classif.ce"),
  ParamSet$new(params),
  trm("evals", n_evals = 100000)
)

# define and call hyperband as usual
tuner = tnr("hyperband", eta = 4L)
# \donttest{
tuner$optimize(inst)

#>    classif.rpart.cp classif.rpart.minsplit subsample.frac learner_param_vals
#> 1:       0.07348139                      3           0.25          <list[6]>
#>     x_domain classif.ce
#> 1: <list[3]>       0.02

# return the best evaluation
inst$result

#>    classif.rpart.cp classif.rpart.minsplit subsample.frac learner_param_vals
#> 1:       0.07348139                      3           0.25          <list[6]>
#>     x_domain classif.ce
#> 1: <list[3]>       0.02

# print all evaluations
print(inst$archive$data())

#>    classif.rpart.cp classif.rpart.minsplit subsample.frac bracket bracket_stage
#> 1:       0.02532664                      4           0.25       1             0
#> 2:       0.07348139                      3           0.25       1             0
#> 3:       0.08489786                      2           0.25       1             0
#> 4:       0.05025520                      4           0.25       1             0
#> 5:       0.07348139                      3           1.00       1             1
#> 6:       0.05762160                      5           1.00       0             0
#> 7:       0.02247238                      3           1.00       0             0
#>    budget_scaled budget_real n_configs classif.ce
#> 1:           2.5        0.25         4       0.04
#> 2:           2.5        0.25         4       0.02
#> 3:           2.5        0.25         4       0.08
#> 4:           2.5        0.25         4       0.02
#> 5:          10.0        1.00         1       0.04
#> 6:          10.0        1.00         2       0.04
#> 7:          10.0        1.00         2       0.04
#>                                   uhash  x_domain           timestamp batch_nr
#> 1: 7c61e5fd-b936-4c11-bccb-f7bef1cbf6bd <list[3]> 2020-09-08 22:36:25        1
#> 2: 963dd916-6766-424c-a2b2-6ce8be89e6b5 <list[3]> 2020-09-08 22:36:25        1
#> 3: b32154f3-b48c-4b5d-bf0e-1b5c11a12ec6 <list[3]> 2020-09-08 22:36:25        1
#> 4: 1bafd488-c253-4650-bde0-be40e2e9cde5 <list[3]> 2020-09-08 22:36:25        1
#> 5: 4dd2e49e-44df-434d-bee1-de215e0c0c49 <list[3]> 2020-09-08 22:36:25        2
#> 6: fd12c6b2-7758-46a9-ade9-b55e0f4ce1af <list[3]> 2020-09-08 22:36:25        3
#> 7: 6241bd11-2d7c-44cf-b674-21489d98bb7f <list[3]> 2020-09-08 22:36:25        3

# }